Perform a fluorescence curve analysis on all samples in the provided dataset.

flFit performs all computational fluorescence fitting operations based on the user input.

Usage

flFit(
  fl_data,
  time = NULL,
  growth = NULL,
  control = fl.control(),
  parallelize = TRUE,
  ...
)

Arguments

fl_data

Either...

a grodata object created with read_data or parse_data,

a list containing a 'time' matrix (for x_type == "time") or 'growth' dataframe (for x_type == "growth") and a 'fluorescence' dataframes, or

a dataframe containing (normalized) fluorescence values (if a time matrix or growth dataframe is provided as separate argument).

time

(optional) A matrix containing time values for each sample.

growth

(optional) A dataframe containing growth values for each sample and sample identifiers in the first three columns.

control

A fl.control object created with fl.control, defining relevant fitting options.

parallelize

Run linear fits and bootstrapping operations in parallel using all but one available processor cores

...

Further arguments passed to the shiny app.

Value

An flFit object that contains all fluorescence fitting results, compatible with various plotting functions of the QurvE package.

raw.x

Raw x matrix passed to the function as time (for x_type = 'time') or growth (for x_type = 'growth').

raw.fl

Raw growth dataframe passed to the function as data.

flTable

Table with fluorescence parameters and related statistics for each fluorescence curve evaluation performed by the function. This table, which is also returned by the generic summary.flFit method applied to a flFit object, is used as an input for fl.drFit.

flFittedLinear

List of all flFitLinear objects, generated by the call of flFitLinear. Note: access to each object in the list via double brace: flFittedLinear[[#n]].

flFittedSplines

List of all flFitSpline objects, generated by the call of flFitSpline. Note: access to each object via double brace: flFittedSplines[[#n]].

flBootSplines

List of all flBootSpline objects, generated by the call of flBootSpline. Note: access to each object via double brace: flFittedSplines[[#n]].

control

Object of class fl.control containing list of options passed to the function as control.

See also

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters.xlsx", package = "QurvE"),
                   data.fl = system.file("lac_promoters.xlsx", package = "QurvE"),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Error in 1:nrow(dat): argument of length 0

# Define fit controls
control <- fl.control(fit.opt = "s",
             x_type = "time", norm_fl = TRUE,
             dr.parameter = "max_slope.spline",
             dr.method = "model",
             suppress.messages = TRUE)

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             control = control,
             parallelize = FALSE)
#> Error in class(fl_data) %in% c("list", "grodata"): object 'input' not found

summary(res)
#> Error in summary(res): object 'res' not found