Calculate parameter standard deviations for several samples

Internal function used within the table_group set of functions.

Usage

get_sd_param(table = data.frame(), ndx.rep = list(), param1, param2 = NULL)

Arguments

table

A dataframe with class gcTable or flTable, generated with growth.gcFit or flFit, respectively.

ndx.rep

A list with each element being a vector of row indices for samples (replicates) for which to calculate standard deviations.

param1

Name of the parameter for which to calculate standard deviations

param2

(optional) Name of a second parameter for which to calculate standard deviations The results for the second parameter will be attached in parentheses after the first parameter.

Value

A vector of strings containing parameter standard deviations.

Examples

if (FALSE) {
# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

conc <- unique(res$expdesign$concentration)
ndx.rep <- lapply(1:length(conc), function(x) grep(paste0("Test1.+", conc[x]), res$expdesign$label))

get_sd_param(res$gcFit$gcTable, ndx.rep, "mu.spline")
}