Perform a growth curve analysis on all samples in the provided dataset.

growth.gcFit performs all computational growth fitting operations based on the user input.

Usage

growth.gcFit(time, data, control = growth.control(), parallelize = TRUE, ...)

Arguments

time

(optional) A matrix containing time values for each sample.

data

Either...

a grodata object created with read_data or parse_data,

a list containing a 'time' matrix as well as 'growth' and, if appropriate, a 'fluorescence' dataframes, or

a dataframe containing growth values (if a time matrix is provided as separate argument).

control

A grofit.control object created with growth.control, defining relevant fitting options.

parallelize

Run linear fits and bootstrapping operations in parallel using all but one available processor cores

...

Further arguments passed to the shiny app.

Value

A gcFit object that contains all growth fitting results, compatible with various plotting functions of the QurvE package.

raw.time

Raw time matrix passed to the function as time.

raw.data

Raw growth dataframe passed to the function as data.

gcTable

Table with growth parameters and related statistics for each growth curve evaluation performed by the function. This table, which is also returned by the generic summary.gcFit method applied to a gcFit object, is used as an input for growth.drFit.

gcFittedLinear

List of all gcFitLinear objects, generated by the call of growth.gcFitLinear. Note: access to each object in the list via double brace: gcFittedLinear[[#n]].

gcFittedModels

List of all gcFitModel objects, generated by the call of growth.gcFitModel. Note: access to each object in the list via double brace: gcFittedModels[[#n]].

gcFittedSplines

List of all gcFitSpline objects, generated by the call of growth.gcFitSpline. Note: access to each object via double brace: gcFittedSplines[[#n]].

gcBootSplines

List of all gcBootSpline objects, generated by the call of growth.gcBootSpline. Note: access to each object via double brace: gcFittedSplines[[#n]].

control

Object of class grofit.control containing list of options passed to the function as control.

References

Matthias Kahm, Guido Hasenbrink, Hella Lichtenberg-Frate, Jost Ludwig, Maik Kschischo (2010). grofit: Fitting Biological Growth Curves with R. Journal of Statistical Software, 33(7), 1-21. DOI: 10.18637/jss.v033.i07

See also

Examples

# Create random growth data set
  rnd.data1 <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')
  rnd.data2 <- rdm.data(d = 35, mu = 0.6, A = 4.5, label = 'Test2')

  rnd.data <- list()
  rnd.data[['time']] <- rbind(rnd.data1$time, rnd.data2$time)
  rnd.data[['data']] <- rbind(rnd.data1$data, rnd.data2$data)

# Run growth curve analysis workflow
  res <- growth.gcFit(time = rnd.data$time,
                      data = rnd.data$data,
                      parallelize = FALSE,
                      control = growth.control(suppress.messages = TRUE,
                                               fit.opt = 's'))