Adjusted modalDialog() function to include classes

Adjusted modalDialog() function to include classes

Usage

help_modal(
  ...,
  title = NULL,
  footer = NULL,
  size = c("m", "s", "l"),
  easyClose = TRUE,
  fade = TRUE,
  idcss = ""
)

Arguments

...

UI elements for the body of the modal dialog box.

title

An optional title for the dialog.

footer

UI for footer. Use NULL for no footer.

size

One of "s" for small, "m" (the default) for medium, or "l" for large.

easyClose

If TRUE, the modal dialog can be dismissed by clicking outside the dialog box, or be pressing the Escape key. If FALSE (the default), the modal dialog can't be dismissed in those ways; instead it must be dismissed by clicking on a modalButton(), or from a call to removeModal() on the server.

fade

If FALSE, the modal dialog will have no fade-in animation (it will simply appear rather than fade in to view).

idcss

CSS class of the modal

Value

div HTML tag for a modal dialog

Author

Stackoverflow user mfindinge

Examples

if (FALSE) {
require(shiny)

shinyApp(
  ui = basicPage(
    tags$style(
      type = 'text/css',
      '.modal-dialog.test{ width: fit-content !important; }'
    ),
    actionButton("show", "Show modal dialog"),
    actionButton("shows", "Show modal dialog2")
  ),

  server = function(input, output) {

    # Render DT
    output$dt <- DT::renderDT(cbind(iris, iris))

    # Modal management
    observeEvent(input$show, {
      showModal(
        help_modal(easyClose = T,

                DT::DTOutput("dt"),

                footer = tagList(
                  modalButton("Cancel"),
                  actionButton("ok", "OK")
                ),
                idcss = "test"
        )
      )
    })

    observeEvent(input$shows, {
      showModal(
        help_modal(easyClose = T,

                DT::DTOutput("dt"),

                footer = tagList(
                  modalButton("Cancel"),
                  actionButton("ok", "OK")
                ),
                idcss = "tests"
        )
      )
    })
  }
)
}