Generic summary function for flBootSpline objects

Generic summary function for flBootSpline objects

Usage

# S3 method for class 'flBootSpline'
summary(object, ...)

Arguments

object

object of class flBootSpline

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with statistical parameters extracted from a dose-response bootstrapping analysis.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Extract time and normalized fluorescence data for single sample
time <- input$time[4,]
data <- input$norm.fluorescence[4,-(1:3)] # Remove identifier columns

# Perform linear fit
TestFit <- flBootSpline(time = time,
                       fl_data = data,
                       ID = 'TestFit',
                       control = fl.control(fit.opt = 's', x_type = 'time',
                       nboot.fl = 50))

summary(TestFit)
#>       max_slope.bt        lambda.bt             A.bt           dY.bt
#> 1 243.151248472904 15.0881823776243 7522.68865015791 5985.1651922914
#>        integral.bt  stdmax_slope.bt      stdlambda.bt          stdA.bt
#> 1 147330.897515745 2.77246812801132 0.361549860782086 108.104526325862
#>           stddY.bt   stdintegral.bt reliable_fit.bt    ci90.mu.bt.lo
#> 1 142.067776666943 5392.99776303107            TRUE 238.590538402325
#>      ci90.mu.bt.up ci90.lambda.bt.lo ci90.lambda.bt.up     ci90.A.bt.lo
#> 1 247.711958543483  14.4934328566378  15.6829318986109 7344.85670435187
#>       ci90.A.bt.up ci90.integral.bt.lo ci90.integral.bt.up    ci95.mu.bt.lo
#> 1 7700.52059596395    138459.416195559    156202.378835931 237.717210942002
#>      ci95.mu.bt.up ci95.lambda.bt.lo ci95.lambda.bt.up     ci95.A.bt.lo
#> 1 248.585286003806  14.3795446504915  15.7968201047572 7310.80377855922
#>      ci95.A.bt.up ci95.integral.bt.lo ci95.integral.bt.up
#> 1 7734.5735217566    136760.621900204    157901.173131286