Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  91.992881542864  15.5516734210176
#> 2                 FALSE          0.75 33.2207777396463  30.5913363219176
#> 3                 FALSE          0.75 64.8163372893604   19.359773535155
#> 4                 FALSE          0.75 242.695863966074  14.9524921229994
#> 5                 FALSE          0.75 65.9771740634183   22.449934232483
#> 6                 FALSE          0.75 205.138036003799  12.4099915956992
#> 7                 FALSE          0.75  506.73172749568  6.99791480151584
#> 8                 FALSE          0.75 202.138513172581  13.4659168987332
#> 9                 FALSE          0.75 289.194680912532  4.77284839078049
#> 10                FALSE          0.75 670.918796254112 0.651711802031005
#> 11                FALSE          0.75 320.370563112088  10.0674216137795
#> 12                FALSE          0.75 334.182717991725  3.77005557331064
#> 13                FALSE          0.75 734.748720592447  0.81767409877198
#> 14                FALSE          0.75 346.758328668932  9.86506359662965
#> 15                FALSE          0.75 321.016290501053  3.42105351554778
#> 16                FALSE          0.75 776.460570318316 0.461031983929361
#> 17                FALSE          0.75 340.253400845953  8.34019834457536
#> 18                FALSE          0.75 349.529623665254  4.98824364903674
#> 19                FALSE          0.75 783.510532981664 0.683693974253554
#> 20                FALSE          0.75 349.943406706516   9.5653483131896
#> 21                FALSE          0.75 339.376510164544  5.09487064167539
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3403.57107518525  2110.1050182132 81119.2091698852  3.3147223688886
#> 2   1450.1687964348   282.3227464813  42237.845442756  3.1867196737289
#> 3  2193.88223048839 746.969372126361 65774.3948802402 1.31896356204661
#> 4  7552.35473503123 6038.56375360337   148301.7796206  2.7993621616228
#> 5  2367.49581985356 1113.55087185985 57212.1775761106 1.94762658066863
#> 6  4820.80900679125 3226.33205794699 130699.125080329 1.98694523829821
#> 7  17034.5966393232 15829.5573362138 361736.488248346 14.1014353133995
#> 8  5182.27821519625 3753.97212010151 118475.934503599 10.4130602564588
#> 9  7761.80172550337  6261.7761108906 221205.540172009 6.52926997575025
#> 10 20074.2063777765 18674.8763579327 485054.207564594 29.9236617845812
#> 11  6793.8888115572 5498.93167565936 168294.200570494 9.17901189020162
#> 12 9184.21298667822 7658.64351599286 267809.117963805 5.78211829610609
#> 13 20485.6370323319 19427.1898074497 492840.528863661 16.2507239500454
#> 14 6984.46577888401 5659.50392912821 177522.430761975 11.7173639696246
#> 15 9007.03428506906 7404.79279135419 265177.029461108  5.9740681405862
#> 16  20652.993875906 19349.7564345431 512563.089457842 16.9343662806549
#> 17 6879.66799313243   5953.351024591 175546.446411927 11.7458246304604
#> 18 9321.47065714299 7630.79417000419 272287.388908242 6.73278869222639
#> 19 19793.1785920189 18610.3965192276 505935.838431572  25.154756817298
#> 20 7159.25755113916 5806.56026974969 181360.981324817 13.7586950838821
#> 21 9254.54687433091 7506.48461045938 269722.027223431 9.89142933273389
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.702046342167667  28.084686623587 49.8877826031727 1632.32677915177
#> 2   2.21407501017153 20.9325623345054  70.274082206579  963.61040401262
#> 3   2.69124500043939 6.45187094381038  183.64018750062 1607.32195408774
#> 4  0.345807135649424 64.9976973112251 78.5475229993215   3566.277297995
#> 5   1.84471598822677 26.5949020056057 107.662576966225 1237.82180501041
#> 6  0.415693603632874 10.8868732437234 76.4453197287918 2399.85249347755
#> 7  0.443289516186625 174.067079240579 180.904967420104 9017.19506643186
#> 8  0.879238287465053 58.1551732124528 156.172924319288 2921.70152650911
#> 9  0.336361063378248 20.0882124062479 74.1255034357243 3748.19654125886
#> 10 0.529383441619844 59.0749591988619 300.058124867692   7553.407391891
#> 11 0.541570519769356 78.2094478732745 181.843276826805 2781.67489108947
#> 12 0.322743859821402 12.9948894854742 93.8675193374873 2340.21975619342
#> 13 0.327559704215368 63.0523254490087 275.536650568696 6470.21542159762
#> 14 0.528925579519595 47.7301501999838 174.988977165282  2920.3841126799
#> 15  0.35868352287116   11.84550832619  71.469149542686 2790.89451578302
#> 16 0.341881920189143 99.5693476215542 249.397488747303 8481.56936555086
#> 17 0.406621909087365 73.2610107507939  139.56983284389 2315.54146535818
#> 18 0.526961749131909 11.3187067415345 133.460172847999 2006.89843895416
#> 19 0.380839003753339 74.5483610745269 255.057194202641 6709.60646487219
#> 20 0.527980950905697 61.3783494846296 160.131833232907 2439.82463916475
#> 21 0.588245681035898 13.1260696840399 89.6415906559238 2600.60291131526
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 86.5401632460423 97.4455998396858   14.3968071881518
#> 2             TRUE 27.9786238763623 38.4629316029304   26.9491829301855
#> 3             TRUE 62.6466422297937 66.9860323489271   14.9326755094323
#> 4             TRUE 238.090913210204 247.300814721943   14.3836393848561
#> 5             TRUE 62.7733283382184 69.1810197886182     19.41537643185
#> 6             TRUE 201.869511086798 208.406560920799   11.7261756177232
#> 7             TRUE 483.534866405137 529.928588586222   6.26870354738884
#> 8             TRUE 185.009029050706 219.267997294455   12.0195699158532
#> 9             TRUE 278.454031802422 299.935330022641   4.21953444152327
#> 10            TRUE 621.694372618476 720.143219889748 -0.219123959433639
#> 11            TRUE 305.271088552706  335.47003767147   9.17653810875889
#> 12            TRUE 324.671133394631  343.69430258882   3.23914192390443
#> 13            TRUE 708.016279694622 761.481161490271    0.2788383853377
#> 14            TRUE   327.4832649389 366.033392398965   8.99498101831992
#> 15            TRUE 311.188948409789 330.843632592318   2.83101912042473
#> 16            TRUE 748.603537786639 804.317602849993 -0.101363774781779
#> 17            TRUE 320.931519328846 359.575282363061   7.67130530412665
#> 18            TRUE 338.454186266541 360.605061063966   4.12139157171475
#> 19            TRUE 742.130958017209 824.890107946119 0.0572138130793108
#> 20            TRUE  327.31035329353 372.576460119502   8.69681964894972
#> 21            TRUE 323.105108912196 355.647911416891   4.12720649637134
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7065396538834 3357.37176568945 3449.77038468105    78434.0316181806
#> 2   34.2334897136498 1415.73473139454 1484.60286147506    40652.7063281552
#> 3   23.7868715608778 2183.26890278582 2204.49555819095    63130.3502657658
#> 4   15.5213448611427 7445.43352295427  7659.2759471082    142435.253465399
#> 5   25.4844920331161 2323.74720605434 2411.24443365278    55175.9607068685
#> 6   13.0938075736753 4802.90010030532 4838.71791327717    126751.367728559
#> 7   7.72712605564284 16748.2562939725  17320.936984674    346903.202364066
#> 8   14.9122638816132 5086.61295526176 5277.94347513073    113669.735492492
#> 9    5.3261623400377  7728.7566160951 7794.84683491165    215039.756861638
#> 10  1.52254756349565 19977.0280698943 20171.3846856586    472628.852404934
#> 11  10.9583051188001 6665.23426980566 6922.54335330873    163718.345374652
#> 12  4.30096922271684 9162.83639347461 9205.58957988183    263959.456464867
#> 13  1.35650981220626 20381.9159569683 20589.3581076956    482197.024495133
#> 14  10.7351461749394 6905.94968180504 7062.98187596298    172718.398896617
#> 15  4.01108791067084 8987.54842387247 9026.52014626564    260586.007982645
#> 16   1.0234277426405 20489.2022990686 20816.7854527435    498610.907851511
#> 17  9.00909138502408 6759.15363044737 7000.18235581748    171737.380701413
#> 18  5.85509572635873 9302.85138455317 9340.08992973281    268986.040976163
#> 19   1.3101741354278 19670.5465380513 19915.8106459865    494898.535796858
#> 20  10.4338769774295 7058.29016623695 7260.22493604138    177347.469793391
#> 21  6.06253478697945 9232.95448970067 9276.13925896116    265444.035434318
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83804.3867215899 85.4960256998424 98.4897373858857    14.175662590369
#> 2     43822.9845573568 26.9748071791377  39.466748300155   26.2517493019814
#> 3     68418.4394947145  62.231168707749 67.4015058709717   14.0849333342938
#> 4     154168.305775802 237.209114129293 248.182613802855   14.2747101371265
#> 5     59248.3944453528 62.1598259653078 69.7945221615288   18.8342908955586
#> 6       134646.8824321 201.243623336734 209.032448670863   11.5952321325788
#> 7     376569.774132626 479.092914281417 534.370540709943   6.12906734979006
#> 8     123282.133514707 181.728915069921  222.54811127524   11.7426098553017
#> 9      227371.32348238 276.397311760061 301.992050065002   4.11358070655912
#> 10    497479.562724255 612.268419156333 729.569173351891  -0.38587974354389
#> 11    172870.055766337 302.379699807293 338.361426416883   9.00594339503155
#> 12    271658.779462743 322.849766131357 345.515669852093   3.13747760806069
#> 13    503484.033232189 702.897301650358 766.600139534536  0.175657078509859
#> 14    182326.462627334 323.792295288468 369.724362049396   8.82836946077125
#> 15    269768.050939572 309.307116945504 332.725464056602   2.71803381072031
#> 16    526515.271064173 743.269212408232 809.651928228399 -0.209056579641359
#> 17    179355.512122441 317.231584570251 363.275217121656   7.54321940276413
#> 18    275588.736840322  336.33335782849 362.725889502018    3.9553986207382
#> 19    516973.141066287  734.20720961976 832.813856343568 -0.062750473102991
#> 20    185374.492856243 322.976364342107 376.910449070925   8.53050564941443
#> 21    274000.019012545 319.989308672385 358.763711656702   3.94190910684503
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9276842516662 3348.52508940302 3458.61706096748    77919.8486827478
#> 2   34.9309233418538 1409.14097425917 1491.19661861043    40349.1690508913
#> 3   24.6346137360162 2181.23656343852 2206.52789753825    62624.0438502282
#> 4   15.6302741088723 7424.95924830123 7679.75022176123     141311.87611653
#> 5   26.0655775694075 2315.36981192257 2419.62182778455    54786.0468382902
#> 6   13.2247510588197 4799.47073523355 4842.14727834895    125995.414193113
#> 7   7.86676225324163 16693.4251640117 17375.7681146348    344062.785918139
#> 8   15.1892239421647 5068.29407569984 5296.26235469266    112749.399511641
#> 9   5.43211607500185 7722.42882918713 7801.17462181962    213859.074951142
#> 10   1.6893033476059 19958.4194577467 20189.9932978062    470249.529076488
#> 11  11.1288998325274 6640.59829372558 6947.17932938881    162842.117783959
#> 12  4.40263353856058 9158.74300328669 9209.68297006975    263222.287241666
#> 13   1.4596911190341 20362.0544744519  20609.219590212     480158.90663733
#> 14  10.9017577324881 6890.91468449204 7078.01687327598    171798.477901122
#> 15  4.12407322037526 8983.81708874973 9030.25148138839    259706.876210174
#> 16  1.13112054750008 20457.8379545678 20848.1497972443    495939.213501363
#> 17   9.1371772863866 6736.07641206087 7023.25957420398    171007.985139825
#> 18  6.02108867733528 9299.28599192958  9343.6553223564    268353.867967892
#> 19   1.4301384216101 19647.0638043129  19939.293379725    492785.009760423
#> 20  10.6001909769648 7038.95598614929 7279.55911612904    176578.925032054
#> 21  6.24783217650575  9228.8197777502 9280.27397091163    264624.845517253
#>    ci95.integral.bt.up
#> 1     84318.5696570227
#> 2     44126.5218346207
#> 3     68924.7459102522
#> 4     155291.683124671
#> 5      59638.308313931
#> 6     135402.835967545
#> 7     379410.190578552
#> 8     124202.469495557
#> 9     228552.005392876
#> 10    499858.886052701
#> 11     173746.28335703
#> 12    272395.948685944
#> 13    505522.151089992
#> 14    183246.383622828
#> 15    270647.182712043
#> 16    529186.965414322
#> 17    180084.907684029
#> 18    276220.909848593
#> 19    519086.667102722
#> 20     186143.03761758
#> 21    274819.208929609