Generic summary function for gcBootSpline objects

Generic summary function for gcBootSpline objects

Usage

# S3 method for class 'gcBootSpline'
summary(object, ...)

Arguments

object

object of class gcBootSpline

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with statistical parameters extracted from the spline fit bootstrapping computation.

Examples

# Create random growth dataset
rnd.dataset <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')

# Extract time and growth data for single sample
time <- rnd.dataset$time[1,]
data <- rnd.dataset$data[1,-(1:3)] # Remove identifier columns

# Introduce some noise into the measurements
data <- data + stats::runif(97, -0.01, 0.09)

# Perform bootstrapping spline fit
TestFit <- growth.gcBootSpline(time, data, gcID = 'TestFit',
              control = growth.control(fit.opt = 's', nboot.gc = 50))

summary(TestFit)
#>               mu.bt        lambda.bt             A.bt            dY.bt
#> 1 0.618548510373017 4.91482659059296 6.10101886387317 6.01977637198868
#>        integral.bt           stdmu.bt      stdlambda.bt            stdA.bt
#> 1 64.9061316828034 0.0181012132969828 0.186213969804493 0.0188061524062477
#>             stddY.bt   stdintegral.bt reliable_fit.bt    ci90.mu.bt.lo
#> 1 0.0201988801706744 1.14913643257375            TRUE 0.58877201449948
#>       ci90.mu.bt.up ci90.lambda.bt.lo ci90.lambda.bt.up     ci90.A.bt.lo
#> 1 0.648325006246554  4.60850461026457  5.22114857092135 6.07008274316489
#>       ci90.A.bt.up ci90.integral.bt.lo ci90.integral.bt.up     ci95.mu.bt.lo
#> 1 6.13195498458144    63.0158022512196    66.7964611143873 0.583070132310931
#>       ci95.mu.bt.up ci95.lambda.bt.lo ci95.lambda.bt.up     ci95.A.bt.lo
#> 1 0.654026888435103  4.54984720977615  5.27980597140977 6.06415880515692
#>       ci95.A.bt.up ci95.integral.bt.lo ci95.integral.bt.up
#> 1 6.13787892258941    62.6538242749589     67.158439090648