Generic summary function for gcFitModel objects

Generic summary function for gcFitModel objects

Usage

# S3 method for class 'gcFitModel'
summary(object, ...)

Arguments

object

object of class gcFitModel

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from the growth model fit.

Examples

# Create random growth dataset
rnd.dataset <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')

# Extract time and growth data for single sample
time <- rnd.dataset$time[1,]
data <- rnd.dataset$data[1,-(1:3)] # Remove identifier columns

# Perform parametric fit
TestFit <- growth.gcFitModel(time, data, gcID = 'TestFit',
                 control = growth.control(fit.opt = 'm'))
#> --> Try to fit model logistic
#> ....... OK
#> --> Try to fit model richards
#> ....... OK
#> --> Try to fit model gompertz
#> ....... OK
#> --> Try to fit model gompertz.exp
#> ... ERROR in nls(). For further information see help(growth.gcFitModel)
#> --> Try to fit model huang
#> .......... OK
#> --> Try to fit model baranyi
#> ........ OK
#> 
#> Best fitting model: ~baranyi

summary(TestFit)
#>            mu.model          tD.model     lambda.model          A.model
#> 1 0.801840034144145 0.864445713663806 5.23371514974094 4.80645397688004
#>           dY.model     A.orig.model    dY.orig.model   integral.model
#> 1 4.81750044867184 6.76962725346751 6.84482230533835 75.5725304147142
#>   parameter_nu.model parameter_alpha.model parameter_t_shift.model
#> 1               <NA>                  <NA>                    <NA>
#>    parameter_y0.model         stdmu.model    stdlambda.model
#> 1 -0.0110822504137362 0.00259447308407248 0.0160999910815046
#>            stdA.model         RMSE.model reliable_fit.model  ci90.mu.model.lo
#> 1 0.00253431286818392 0.0157867305162551               TRUE 0.797572125920845
#>    ci90.mu.model.up ci90.lambda.model.lo ci90.lambda.model.up  ci90.A.model.lo
#> 1 0.806107942367444     5.20723066441187     5.26019963507002 4.80228503221188
#>   ci90.A.model.up  ci95.mu.model.lo  ci95.mu.model.up ci95.lambda.model.lo
#> 1 4.8106229215482 0.796754866899363 0.806925201388927      5.2021591672212
#>   ci95.lambda.model.up ci95.A.model.lo  ci95.A.model.up
#> 1     5.26527113226069 4.8014867236584 4.81142123010168