Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime            dY
#> 1      Test2 | 0 0.465 ± 0.001 1.489 ± 0.002 3.941 ± 0.154 3.625 ± 0.189
#> 2  Test2 | 0.039 0.393 ± 0.006 1.766 ± 0.025 3.968 ± 0.181 1.859 ± 0.183
#> 3  Test2 | 0.059 0.344 ± 0.005  2.013 ± 0.03 3.801 ± 0.083 1.499 ± 0.041
#> 4  Test2 | 0.088 0.291 ± 0.008 2.382 ± 0.064 3.763 ± 0.102 1.138 ± 0.134
#> 5   Test2 | 0.13 0.211 ± 0.005  3.279 ± 0.08 3.335 ± 0.143    0.8 ± 0.11
#> 6    Test2 | 0.2 0.121 ± 0.011 5.761 ± 0.513 2.407 ± 0.277 0.427 ± 0.032
#> 7    Test2 | 0.3                                                        
#> 8   Test2 | 0.44                                                        
#> 9   Test2 | 0.67                                                        
#> 10     Test2 | 1                                                        
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1  3.675 ± 0.192   6.75 ± 0.433 10.167 ± 0.382
#> 2  1.912 ± 0.186   6.75 ± 0.433   9.75 ± 0.661
#> 3  1.552 ± 0.043       7 ± 0.25 10.333 ± 0.144
#> 4  1.189 ± 0.135       7 ± 0.25  10.333 ± 1.01
#> 5  0.853 ± 0.113   7.75 ± 1.146    10.25 ± 1.5
#> 6   0.48 ± 0.033 10.417 ± 2.155   12.5 ± 2.385
#> 7                                             
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }