Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.796 ± 0.005 0.871 ± 0.005 5.151 ± 0.145  4.85 ± 0.052
#> 2  Test1 | 0.017 0.782 ± 0.004 0.886 ± 0.005  5.13 ± 0.092 4.318 ± 0.017
#> 3  Test1 | 0.026 0.782 ± 0.009  0.886 ± 0.01  5.12 ± 0.077 4.078 ± 0.046
#> 4  Test1 | 0.039  0.765 ± 0.01 0.906 ± 0.012 5.324 ± 0.069 3.905 ± 0.039
#> 5  Test1 | 0.059 0.728 ± 0.007 0.952 ± 0.009  5.331 ± 0.13 3.698 ± 0.027
#> 6  Test1 | 0.088 0.676 ± 0.007  1.025 ± 0.01  5.506 ± 0.23 3.493 ± 0.011
#> 7   Test1 | 0.13 0.582 ± 0.005  1.192 ± 0.01 5.761 ± 0.264 3.146 ± 0.058
#> 8    Test1 | 0.2 0.385 ± 0.103 1.909 ± 0.602  5.433 ± 0.96 2.655 ± 0.018
#> 9    Test1 | 0.3 0.236 ± 0.071 3.176 ± 1.161 5.265 ± 1.733  2.238 ± 0.08
#> 10  Test1 | 0.44 0.131 ± 0.009  5.31 ± 0.347 6.367 ± 0.438 1.632 ± 0.114
#> 11  Test1 | 0.67 0.091 ± 0.011 7.686 ± 0.908 8.232 ± 0.503 0.823 ± 0.061
#> 12     Test1 | 1                                                        
#>               dY
#> 1  4.845 ± 0.042
#> 2  4.319 ± 0.015
#> 3  4.071 ± 0.043
#> 4  3.887 ± 0.036
#> 5  3.692 ± 0.027
#> 6  3.488 ± 0.023
#> 7  3.147 ± 0.054
#> 8  2.637 ± 0.027
#> 9   2.13 ± 0.071
#> 10   1.3 ± 0.054
#> 11 0.573 ± 0.033
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities