Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.668 ± 0.006  1.038 ± 0.009 4.491 ± 0.053 6.122 ± 0.524
#> 2  Test1 | 0.017 0.619 ± 0.001  1.119 ± 0.002 4.489 ± 0.243  3.835 ± 0.12
#> 3  Test1 | 0.026 0.598 ± 0.005   1.16 ± 0.009 4.475 ± 0.219 3.112 ± 0.025
#> 4  Test1 | 0.039  0.579 ± 0.01   1.198 ± 0.02 4.571 ± 0.216 2.985 ± 0.185
#> 5  Test1 | 0.059 0.539 ± 0.007  1.286 ± 0.016 4.546 ± 0.152 2.246 ± 0.096
#> 6  Test1 | 0.088 0.483 ± 0.012  1.437 ± 0.036 4.423 ± 0.146 1.737 ± 0.207
#> 7   Test1 | 0.13 0.385 ± 0.007  1.801 ± 0.031 4.333 ± 0.113 1.201 ± 0.097
#> 8    Test1 | 0.2 0.265 ± 0.007   2.617 ± 0.07 3.957 ± 0.192 0.793 ± 0.093
#> 9    Test1 | 0.3 0.155 ± 0.004  4.481 ± 0.104 3.213 ± 0.311 0.499 ± 0.025
#> 10  Test1 | 0.44 0.074 ± 0.004   9.36 ± 0.557 2.192 ± 0.474 0.209 ± 0.001
#> 11  Test1 | 0.67 0.039 ± 0.001 17.905 ± 0.419 1.838 ± 3.151 0.096 ± 0.002
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.071 ± 0.523              8
#> 2  3.783 ± 0.116  7.875 ± 0.177
#> 3   3.06 ± 0.025  7.917 ± 0.289
#> 4   2.93 ± 0.183  8.083 ± 0.144
#> 5  2.193 ± 0.093  8.083 ± 0.144
#> 6  1.685 ± 0.203              8
#> 7  1.148 ± 0.093  8.417 ± 0.382
#> 8   0.738 ± 0.09  8.917 ± 0.144
#> 9  0.446 ± 0.024 10.417 ± 1.041
#> 10 0.154 ± 0.001  10.083 ± 4.26
#> 11 0.045 ± 0.002  5.583 ± 8.596
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities