Sets the default breaks for log10 (from R package 'xgxr')

xgx_breaks_log10 sets nice breaks for log10 scale. it's better than the default function because it ensures there is at least 2 breaks and also, it will try to go by 3s (i.e. 1,3,10,30,100) if it makes sense

Usage

xgx_breaks_log10(data_range)

Arguments

data_range

range of the data

Value

numeric vector of breaks

Details

for the extended breaks function, weights is a set of 4 weights for

simplicity - how early in the Q order are you

coverage - labelings that don't extend outside the data: range(data) / range(labels)

growth (previously granularity) - how close to the number of ticks do you get (default is 5)

legibility - has to do with fontsize and formatting to prevent label overlap

References

Talbot, Justin, Sharon Lin, and Pat Hanrahan. 'An extension of Wilkinson’s algorithm for positioning tick labels on axes.' IEEE Transactions on visualization and computer graphics 16.6 (2010): 1036-1043.

Author

Andrew Stein

Examples

if (FALSE) {
xgx_breaks_log10(c(1, 1000))
xgx_breaks_log10(c(0.001, 100))
xgx_breaks_log10(c(1e-4, 1e4))
xgx_breaks_log10(c(1e-9, 1e9))
xgx_breaks_log10(c(1, 2))
xgx_breaks_log10(c(1, 5))
xgx_breaks_log10(c(1, 10))
xgx_breaks_log10(c(1, 100))
xgx_breaks_log10(c(1, 1.01))
xgx_breaks_log10(c(1, 1.0001))
print(xgx_breaks_log10(c(1, 1.000001)), digits = 10)
}