Generic summary function for flBootSpline objects

Generic summary function for flBootSpline objects

Usage

# S3 method for flBootSpline
summary(object, ...)

Arguments

object

object of class flBootSpline

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with statistical parameters extracted from a dose-response bootstrapping analysis.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Extract time and normalized fluorescence data for single sample
time <- input$time[4,]
data <- input$norm.fluorescence[4,-(1:3)] # Remove identifier columns

# Perform linear fit
TestFit <- flBootSpline(time = time,
                       fl_data = data,
                       ID = 'TestFit',
                       control = fl.control(fit.opt = 's', x_type = 'time',
                       nboot.fl = 50))

summary(TestFit)
#>       max_slope.bt        lambda.bt             A.bt            dY.bt
#> 1 243.056837634431 15.0672723060166 7526.34201390689 5992.87036255117
#>        integral.bt  stdmax_slope.bt     stdlambda.bt          stdA.bt
#> 1 147317.952491311 2.48828620899619 0.31409865007714 101.292987773993
#>           stddY.bt   stdintegral.bt reliable_fit.bt    ci90.mu.bt.lo
#> 1 125.321286350174 5234.58993426645            TRUE 238.963606820632
#>     ci90.mu.bt.up ci90.lambda.bt.lo ci90.lambda.bt.up     ci90.A.bt.lo
#> 1 247.15006844823  14.5505800266397  15.5839645853935 7359.71504901867
#>      ci90.A.bt.up ci90.integral.bt.lo ci90.integral.bt.up    ci95.mu.bt.lo
#> 1 7692.9689787951    138707.052049443     155928.85293318 238.179796664799
#>      ci95.mu.bt.up ci95.lambda.bt.lo ci95.lambda.bt.up     ci95.A.bt.lo
#> 1 247.933878604064  14.4516389518654  15.6829056601678 7327.80775786986
#>       ci95.A.bt.up ci95.integral.bt.lo ci95.integral.bt.up
#> 1 7724.87626994391    137058.156220149    157577.748762474