Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0346089877435  15.5133508344193
#> 2                 FALSE          0.75 33.3476229223515  30.9641140679101
#> 3                 FALSE          0.75 64.7578743034275   20.643724156624
#> 4                 FALSE          0.75 243.387623478634  15.0400927438039
#> 5                 FALSE          0.75 65.2523689481015  22.7386468017277
#> 6                 FALSE          0.75 204.495349514479  12.5958012697642
#> 7                 FALSE          0.75 510.696438602944  7.16965005242649
#> 8                 FALSE          0.75 199.563481234261  13.4163155979467
#> 9                 FALSE          0.75 290.942686033944  4.68150333114371
#> 10                FALSE          0.75 673.870156745321 0.736252731460194
#> 11                FALSE          0.75 324.682553551998  9.84965312088084
#> 12                FALSE          0.75 334.147814634789  3.89644925783771
#> 13                FALSE          0.75 726.745204801102 0.667238988194173
#> 14                FALSE          0.75 342.649290650193  9.71021810790619
#> 15                FALSE          0.75 317.657425538545  3.39586679743252
#> 16                FALSE          0.75  776.44826880574 0.383868984043919
#> 17                FALSE          0.75 336.216681851235  8.25338911623603
#> 18                FALSE          0.75 349.178915148402  4.81577012352079
#> 19                FALSE          0.75 782.277835273261 0.580443625025115
#> 20                FALSE          0.75 354.757689402395  9.47259047737971
#> 21                FALSE          0.75 341.677455869963  5.46501320704081
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.72494683499 2117.78043336639 81409.6578117145 2.06169569792236
#> 2   1464.0642288117 297.408080405933 42196.3572341797 1.95446196117551
#> 3  2195.38452308953 655.578384585179 66178.5118737627 1.79442316795535
#> 4  7572.62952878834  6049.7181119787 148969.007998997 2.17846600860318
#> 5  2355.12571268956 1121.52032190184 56963.3526965495 2.18778955430057
#> 6   4816.6391414254 3185.85782591607  130261.63194046    2.43574105204
#> 7  17018.6388930333 15803.6179922946  361421.59077402 14.4539522295293
#> 8  5198.69313535804 3775.54326526943 119494.556588687 7.55722077187709
#> 9  7764.26497120933 6289.36075790898 221296.698070766 4.98951888209077
#> 10 20109.2670505143 18674.9791057292 484364.828557741 31.8059968841132
#> 11 6817.45765732756 5620.17083354123 168333.211354865 10.9769853302219
#> 12 9188.99972896314 7626.70023615797 265152.579806807 5.50971267802172
#> 13  20447.521948883 19418.2989780615 491689.801108499 28.6550746502811
#> 14 6928.23194189977 5635.36860515001 175004.699505067 11.8262148454287
#> 15 8997.58964343255  7347.2769846795 264749.179102605 5.88648041151694
#> 16 20641.6389256412 19405.4492358955 514584.375553971 18.6094102742192
#> 17  6897.5757593337 5947.80882719065 176193.398644863 10.2967170077716
#> 18 9319.44247079129 7691.05450825715 271267.513209119 6.71159388478717
#> 19 19804.2750264898 18697.3712683447 500440.367469706 20.7544961642354
#> 20 7107.97124812545 5806.73912258554 179206.886288288 9.76568807204803
#> 21 9258.49444590491 7428.92526502603 268568.027914286 10.1165444739782
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.757160303815387  45.022542224757 76.0656080771256 2408.49150172626
#> 2   2.77889981686763 28.1735252485352 72.3457897506128 616.972692546791
#> 3   1.92916750024338 6.01388727305336 124.762369477038 1367.38374337548
#> 4  0.303949465443553 55.2237108406036 94.7307828976851 3111.42613006698
#> 5   2.86872306194597 43.7078822424219 165.598041744796 1616.94296390863
#> 6   0.21811922592166 8.17614648376829 42.0884541691355 3076.07739668837
#> 7  0.460429605283565 174.321432582564  167.97934681702 9069.35845996384
#> 8   0.71920162917619 25.1261788475954 126.257702006738 1509.99445446181
#> 9  0.328719841738715 23.2728620314397 66.9183918731029 4267.35958181377
#> 10 0.616259615102101 85.4523266345713  402.48877540959  7555.4656509555
#> 11  0.66389844358002 61.1791615966002 169.777511877912  1880.8927990926
#> 12 0.455008556680865 12.3903845663527 135.531388098273 4978.14397377844
#> 13 0.305386567708232 72.9200189187017 217.718338773488 7716.01501922719
#> 14 0.573298257771922 94.4958385662336  221.99179237044 5488.74253656058
#> 15 0.598668063696284 20.8473930878446 148.167925917507 4037.00385135304
#> 16 0.237760464969248 80.7679501193822 214.312840523684 9386.82048901095
#> 17 0.528716458988347 80.0624580635324  145.96954189365 3140.34498609895
#> 18 0.437984840721922 11.4712741842516 72.7670968503692 4179.76102527074
#> 19 0.280423170506405   77.85762827035 188.952248384672 12451.6633942499
#> 20 0.237644726549713 123.643747927736 133.011502476037 5008.35877868984
#> 21  0.76383729093603 6.91603786145475   174.4285749528 3059.46392909007
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE 88.6431195646612 95.4260984108257     14.2678221346429
#> 2             TRUE 30.1325329962178 36.5627128484852     26.3928238691629
#> 3             TRUE  61.806048192141  67.709700414714     17.4702436187237
#> 4             TRUE 239.804046894482 246.971200062786     14.5400958731492
#> 5             TRUE  61.653455131277 68.8512827649259     18.0195973648266
#> 6             TRUE 200.488555483873 208.502143545085     12.2369951431231
#> 7             TRUE 486.919687185368  534.47319002052     6.41224335173502
#> 8             TRUE 187.131853064523 211.995109403999     12.2332289179518
#> 9             TRUE 282.734927472905 299.150444594984     4.14075919148352
#> 10            TRUE 621.549291870955 726.191021619687   -0.277494335382763
#> 11            TRUE 306.625412683783 342.739694420213      8.7575401811917
#> 12            TRUE 325.084337279443 343.211291990135     3.14796018209769
#> 13            TRUE 679.607607001389 773.882802600814    0.164878084314131
#> 14            TRUE 323.195167229462 362.103414070923     8.76714247387138
#> 15            TRUE   307.9741652616 327.340685815491     2.41105783265213
#> 16            TRUE  745.83578890465 807.060748706831 -0.00724698083049446
#> 17            TRUE  319.27858237345 353.154781329019      7.3836505412002
#> 18            TRUE 338.138343207927 360.219487088877     4.09528506053322
#> 19            TRUE 748.136689083093 816.418981463428     0.11914750954208
#> 20            TRUE 338.693132523876 370.822246280914     9.08166490220544
#> 21            TRUE 325.035740210269 358.319171529657     4.20850086345104
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7588795341956 3333.66286487527 3481.78702879472    77447.6892913748
#> 2   35.5354042666574 1417.71877977786 1510.40967784554    41181.4371549402
#> 3   23.8172046945244 2185.49167852536 2205.27736765371      63929.16561591
#> 4   15.5400896144585 7481.78652445555 7663.47253312113    143850.712015037
#> 5   27.4576962386288 2283.22624640077 2427.02517897834    54303.4815209198
#> 6   12.9546073964053  4803.1893804596  4830.0889023912    125201.484622908
#> 7   7.92705675311795  16731.880136435 17305.3976496316    346502.496107379
#> 8   14.5994022779415 5157.36057115375 5240.02569956234    117010.615711098
#> 9   5.22224747080389 7725.98111316761 7802.54882925105    214276.891558682
#> 10  1.74999979830315 19968.6979732005 20249.8361278282    471936.087561919
#> 11    10.94176606057 6716.81793650115 6918.09737815397    165239.142700357
#> 12  4.64493833357773 9168.61754635149 9209.38191157479    256963.532969941
#> 13  1.16959989207421 20327.5685177617 20567.4753800042     478996.95640187
#> 14   10.653293741941 6772.78628745832 7083.67759634123    165975.718032424
#> 15  4.38067576221291 8963.29568180304 9031.88360506205     258108.30776713
#> 16 0.774984948918332 20508.7756476948 20774.5022035876    499143.055849548
#> 17  9.12312769127186 6765.87301581919 7029.27850284821    171027.531142731
#> 18  5.53625518650835  9300.5722247582 9338.31271682438    264391.806322549
#> 19  1.04173974050815 19676.1992279851 19932.3508249945    479957.381186165
#> 20  9.86351605255399 6904.57728278432 7311.36521346657    170968.136097343
#> 21  6.72152555063058 9247.11756362281   9269.871328187    263535.209750933
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     85371.6263320542 87.9936854198156 96.0755325556713    14.0293166389411
#> 2     43211.2773134192 29.5168774784475 37.1783683662555    25.5174704268496
#> 3     68427.8581316154  61.240804894235   68.27494371262     16.862555856147
#> 4     154087.303982957 239.117830101772 247.657416855496    14.4443517915345
#> 5     59623.2238721792 60.9643014216724 69.5404364745306    17.1159496003136
#> 6     135321.779258012 199.721297052481 209.269401976477    12.1682875869577
#> 7      376340.68544066 482.366692233067 539.026184972822     6.2672080260707
#> 8     121978.497466277 184.751328521382  214.37563394714    12.0066804047613
#> 9      228316.50458285 281.163229025046 300.722143042842    4.03721244133582
#> 10    496793.569553563 611.530402852459 736.209910638183  -0.471616114139925
#> 11    171427.280009372 303.167662304763 346.197444799233      8.548412171464
#> 12    273341.626643672 323.348777785866 344.946851483711    3.00463248674321
#> 13    504382.645815128 670.581258486551 782.909151115653  0.0686813154860382
#> 14    184033.680977709 319.469909553152 365.828671747233    8.58655352267322
#> 15    271390.050438081 306.119923931972 329.194927145118     2.2224773925878
#> 16    530025.695258394  739.97382466827  812.92271294321 -0.0821415272958076
#> 17    181359.266146996 316.035116516002 356.398247186467    7.21710485661887
#> 18     278143.22009569 336.024191134219 362.333639162585    3.95731983570582
#> 19    520923.353753247 741.599022791359 822.956647755162  0.0308142108325624
#> 20    187445.636479232 335.616940781181 373.898438023609    9.00680681334228
#> 21    273600.846077639 321.849028700966 361.505883038961    3.96789211680619
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9973850298974 3319.48076407447 3495.96912959551     76689.014468331
#> 2   36.4107577089707 1408.84411932457 1519.28433829883     40987.090756788
#> 3   24.4248924571011 2183.59730403435 2207.17174214472    63498.4397367468
#> 4   15.6358336960732 7464.39105554076 7680.86800203592    142870.612784066
#> 5   28.3613440031418 2269.45826349441  2440.7931618847    53794.1444872886
#> 6   13.0233149525706 4800.61389431721 4832.66438853358    124232.520242951
#> 7   8.07209207878228 16676.9688851715 17360.3089008952    343645.648192491
#> 8    14.825950791132 5149.44582481676 5247.94044589933    116534.967457942
#> 9   5.32579422095159 7718.65016162771 7809.87978079095    212932.673290411
#> 10  1.94412157706031 19941.7804903106 20276.7536107181    469556.115881868
#> 11  11.1508940702977 6697.54650059822  6937.3688140569    164646.661468643
#> 12  4.78826602893221 9164.71457521309 9213.28488271319    255395.417618201
#> 13  1.26579666090231 20304.5987118023 20590.4451859636    476566.411670814
#> 14  10.8338826931392 6743.02009830995 7113.44378548959    164246.764133408
#> 15  4.56925620227724 8956.72875298037 9038.45053388472    256836.651553954
#> 16 0.849879495383645 20483.3337434072 20799.9441078752     496186.20739551
#> 17  9.28967337585319 6740.65334152918 7054.49817713822     170038.32247211
#> 18  5.67422041133575 9296.95877339016 9341.92616819242    263075.181599589
#> 19  1.13007303921767 19651.6740750799 19956.8759778997    476035.107216976
#> 20  9.93837414141715 6865.62950218708 7350.31299406381    169390.503082056
#> 21  6.96213429727543 9244.93901169646 9272.04988011336    262571.478613269
#>    ci95.integral.bt.up
#> 1     86130.3011550979
#> 2     43405.6237115714
#> 3     68858.5840107786
#> 4     155067.403213928
#> 5     60132.5609058104
#> 6     136290.743637969
#> 7     379197.533355549
#> 8     122454.145719432
#> 9     229660.722851121
#> 10    499173.541233614
#> 11    172019.761241086
#> 12    274909.741995412
#> 13    506813.190546184
#> 14    185762.634876725
#> 15    272661.706651257
#> 16    532982.543712433
#> 17    182348.474817617
#> 18     279459.84481865
#> 19    524845.627722436
#> 20     189023.26949452
#> 21    274564.577215302