Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0433901016413  15.6366340313971
#> 2                 FALSE          0.75 33.1626363397438  30.8614877328151
#> 3                 FALSE          0.75 64.7533357833252  21.1723595061699
#> 4                 FALSE          0.75 241.891207232636  14.9078952874504
#> 5                 FALSE          0.75 66.3738764745747   22.704555745581
#> 6                 FALSE          0.75 205.123046935813  12.5532719119236
#> 7                 FALSE          0.75 504.728005388258  6.92415611513709
#> 8                 FALSE          0.75  200.42793432541  13.0750242088062
#> 9                 FALSE          0.75 289.114417350432  4.68674950465877
#> 10                FALSE          0.75 682.596810860578 0.783607880863724
#> 11                FALSE          0.75 330.364389439123  9.89570915132679
#> 12                FALSE          0.75 333.435779007769  3.69750703540161
#> 13                FALSE          0.75 728.341208631939 0.732380889187032
#> 14                FALSE          0.75 340.415051661906  9.74817975509215
#> 15                FALSE          0.75 318.400874827775  3.27062304443834
#> 16                FALSE          0.75 781.027290906754  0.55438544526614
#> 17                FALSE          0.75 338.361624256848  7.92277001000469
#> 18                FALSE          0.75 350.421369983317  5.09026342092182
#> 19                FALSE          0.75  778.90328735138 0.668668693237452
#> 20                FALSE          0.75  347.01365630194  9.47501543083159
#> 21                FALSE          0.75 337.772283407086   5.2059779343523
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3395.54282878718  2093.8041005343 80656.0750514596 2.16231944677274
#> 2  1463.58597050531 295.404616866455 42520.7960547818 2.54044261772991
#> 3  2196.74222406944 623.554134370879 66198.3889425551 1.35610453735161
#> 4  7567.78321563526 6057.10598030632 149046.295876161 2.17977936839529
#> 5  2368.81730577593 1088.11520610673 57358.4783797091 1.89137021494296
#> 6  4821.58657409604 3198.92148576333 130201.656078118   3.311241873422
#> 7  16987.8279174899  15791.343233634 359158.862236829 8.34251652763677
#> 8  5176.45388076916 3826.33463768237  117924.80726042 6.31154339316816
#> 9  7763.82655606755  6278.7388662968 222438.891362938 6.17074747054224
#> 10 20078.5407917643 18604.1416560023 486851.992261465 30.0239169863857
#> 11 6780.24530418607 5590.04142933137 167513.333625581 12.6987934100619
#> 12 9192.66080322925 7672.04211865988 265658.434001864 5.63846025491734
#> 13 20401.7151889024 19359.6992919159 487006.270357094  28.194148552125
#> 14 6965.13163913714 5641.25195780267 176945.408711411 13.2662165470752
#> 15 9007.13591360421  7414.2337890415 264745.202573201 5.73041363794079
#> 16 20688.4302501905  19328.979108488 513199.926715514 29.0845184551726
#> 17 6899.14520395783 6068.51317270959 175643.376216981 15.3794680512818
#> 18  9321.7748682783 7616.70390018253 270927.722194133 5.05945082972495
#> 19 19784.1298694011 18586.7811761495 500373.740225176 27.6363313058397
#> 20 7107.52949379754 5768.67471478397  180004.01859551 12.1382230299029
#> 21 9255.07854466234 7461.09614264685 269071.696038755 6.31577443767282
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.489065137786951 22.6039370570129 40.0089500324216 1683.66749204105
#> 2   2.55786449146082 21.9165294947149 73.8181305857623 903.483657341665
#> 3   1.41111767728385 5.75090442304485 84.2119753678662 1378.53814764906
#> 4  0.235669527372238  43.950364144395  49.939566184232 2344.15448658453
#> 5   2.86829292081264 35.0592351482093 165.592772650212 1318.16339713231
#> 6  0.410849339623542 13.2762924951592 81.0690650852731 3172.18726066018
#> 7  0.322766468592482 253.242780205705 271.785621149772 14026.6092188627
#> 8  0.512372278182815 48.9227677253973 116.287508388569 2606.25447504584
#> 9  0.330876184632628 29.3390329970129 84.9886625070281  2470.1249173368
#> 10 0.592806059433562 91.4830103145328 340.556057240939 10709.7276265034
#> 11 0.750037680642929 83.1697312822105 234.388559105409 3428.49964413752
#> 12 0.321021000973993 14.5140262727669 72.4507999305719 6669.61123234076
#> 13 0.400719470682424 107.418009233282  228.96265285904 11942.6146785434
#> 14 0.641878495013773 66.9803259615161 176.542748552944 3648.39924284996
#> 15 0.352663662513097 13.2303990310558 66.1989985026232 4530.72041516607
#> 16   0.5884011556798 95.5158876325806 404.486584891885 9216.72140364888
#> 17  0.73746856625994 92.3360658945216 174.498064513409 3655.35767525522
#> 18 0.532078114797169 10.0968861351813 152.501188756983 3548.05686417151
#> 19 0.387316893167193 87.6680802215487 266.018529783041 11880.4364761126
#> 20 0.394926607052454 71.2345855028012 116.329243921027 2148.48543733581
#> 21 0.414611616234396 12.9038144166604 113.206508987839 3650.59424990127
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.4863746117001 95.6004055915824   14.8321218797376
#> 2             TRUE 28.9836082335781 37.3416644459096   26.6538006443621
#> 3             TRUE 62.5225438193818 66.9841277472686   18.8510709270379
#> 4             TRUE 238.305470171625 245.476944293646   14.5202189149231
#> 5             TRUE 63.2625724709935 69.4851804781558   17.9862138908442
#> 6             TRUE 199.676054054033 210.570039817592   11.8774247482429
#> 7             TRUE 491.004565700295  518.45144507622   6.39320527430245
#> 8             TRUE 190.045445443648 210.810423207171   12.2321718111955
#> 9             TRUE  278.96353776139 299.265296939474    4.1424581809381
#> 10            TRUE 633.207467417973 731.986154303182 -0.191558086904486
#> 11            TRUE 309.474874279571 351.253904598675   8.66189716666918
#> 12            TRUE 324.160511888429 342.711046127108    3.1694274887994
#> 13            TRUE 681.961834263694 774.720583000185 0.0731973599144451
#> 14            TRUE 318.592125441967 362.237977881845   8.69228963079449
#> 15            TRUE 308.974344393363 327.827405262188    2.6904913196043
#> 16            TRUE 733.183258047995 828.871323765513 -0.413534455827132
#> 17            TRUE 313.062399312489 363.660849201206   6.70963421850709
#> 18            TRUE 342.098573368419 358.744166598214   4.21499492208048
#> 19            TRUE 733.441522353274 824.365052349486 0.0315324039774193
#> 20            TRUE 327.046279417749  366.98103318613    8.8253611622303
#> 21            TRUE 327.382834457114 348.161732357058   4.52394182564672
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4411461830567  3358.3593523284 3432.72630524597     77886.442027052
#> 2   35.0691748212682 1427.53327948651 1499.63866152412    41034.5654384548
#> 3   23.4936480853018 2187.28198629353 2206.20246184535    63930.6936896724
#> 4   15.2955716599777 7495.48486661773 7640.08156465279    145190.161745729
#> 5   27.4228976003178 2311.14486395713 2426.48974759474    55190.0995914265
#> 6   13.2291190756043  4799.7470729415 4843.42607525058    124983.408034332
#> 7   7.45510695597172 16571.2435440515 17404.4122909283    336085.090071799
#> 8    13.917876606417 5095.97592786088 5256.93183367744     113637.51864897
#> 9   5.23104082837945 7715.56384678746 7812.08926534764    218375.535873919
#> 10  1.75877384863193 19928.0512397969 20229.0303437317    469234.490315866
#> 11  11.1295211359844 6643.43109622684 6917.05951214531    161873.451710975
#> 12  4.22558658200383 9168.78523001055 9216.53637644795    254686.923524663
#> 13  1.39156441845962 20225.0125637137 20578.4178140912     467360.66921089
#> 14  10.8040698793898 6854.94900293045 7075.31427534384    170943.791956922
#> 15  3.85075476927239 8985.37190719812  9028.8999200103    257292.167490253
#> 16  1.52230534635941 20531.3066150349 20845.5538853461    498038.420006512
#> 17  9.13590580150229 6747.25237556134 7051.03803235432    169630.312841186
#> 18  5.96553191976316 9305.16549058593 9338.38424597068    265091.168652571
#> 19  1.30580498249748 19639.9158774366 19928.3438613655    480830.422221971
#> 20  10.1246696994329 6990.34860064543 7224.71038694965    176469.760051093
#> 21  5.88801404305789 9233.85176994694 9276.30531937775    263066.468497668
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83425.7080758671 87.8052439859667 96.2815362173158    14.6780663613347
#> 2     44007.0266711089 28.1833688089932 38.1419038704945    25.8480733295519
#> 3     68466.0841954379  62.095370890116 67.4113006765343    18.4065688586935
#> 4     152902.430006592 237.618839670581  246.16357479469    14.4459830138008
#> 5     59526.8571679918 62.6667908532865 70.0809620958629    17.0827016207882
#> 6     135419.904121904 198.633012863905  211.61308100772    11.7480072062615
#> 7     382232.634401858  488.37667299409 521.079337782426    6.29153383669582
#> 8     122212.095871871   188.0573092748 212.798559376019    12.0707745435679
#> 9     226502.246851957  277.01975230817 301.209082392695    4.03823218277882
#> 10    504469.494207063 623.749933567262 741.443688153894  -0.378291995626058
#> 11    173153.215540187 305.474754355402 355.254024522845    8.42563529726665
#> 12    276629.944479064 322.384396908131 344.487161107407    3.06830587349259
#> 13    506651.871503298 673.080677469774 783.601739794104 -0.0530292733505184
#> 14    182947.025465899 314.413267229638 366.416836094173    8.49009790486515
#> 15    272198.237656149 307.169264097411 329.632485558139    2.57940226591267
#> 16    528361.433424517 724.021634734615 838.032947078892  -0.598880819866269
#> 17    181656.439592776 308.217866876335  368.50538163736     6.4773316201352
#> 18    276764.275735695 340.504846357056 360.337893609578    4.04739031591937
#> 19    519917.058228381 724.736077991934 833.070496710826 -0.0904724173702465
#> 20    183538.277139928  323.22273916333 370.804573440549    8.70095928100878
#> 21    275076.923579843 325.393365509247 350.151201304925    4.39333916653289
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5952017014596 3351.23911215544 3439.84654541893    77356.0867670591
#> 2   35.8749021360783 1420.62957269567 1506.54236831495    40749.9680863922
#> 3   23.9381501536462 2185.47045140027  2208.0139967386     63496.454173163
#> 4      15.3698075611 7481.64050191224 7653.92592935827    144451.753082455
#> 5   28.3264098703738 2300.10120488544 2437.53340666642    54774.8781213298
#> 6   13.3585366175857 4795.56504080553 4847.60810738655    123984.169047224
#> 7   7.55677839357835 16491.4720682867 17484.1837666931    331666.708167858
#> 8   14.0792738740446 5080.56525602738 5272.34250551094     112816.54848933
#> 9   5.33526682653873  7706.3220513934 7821.33106074169    217597.446524958
#> 10   1.9455077573535 19899.2340915478 20257.8474919808    465860.926113518
#> 11  11.3657830053869 6617.23263087294 6943.25797749921    160793.474323071
#> 12  4.32670819731064 9164.21331173463 9221.10829472387    252585.995986476
#> 13  1.51779105172458 20191.1758908052 20612.2544869997    463598.745587149
#> 14  11.0062616053191 6833.85020025257 7096.41307802171    169794.546195425
#> 15  3.96184382296401 8981.20433150334 9033.06749570508    255864.990559476
#> 16  1.70765171039855 20501.2191104307 20875.6413899504    495135.152764363
#> 17  9.36820839987417 6718.16651480457 7080.12389311109    168478.875173481
#> 18  6.13313652592427 9301.98497145335 9341.56476510326    263973.530740357
#> 19  1.42780980384515 19612.3004321668 19955.9593066353    477088.084731995
#> 20  10.2490715806544 6967.90970621205 7247.14928138303    175792.987138332
#> 21  6.01861670217172 9229.78706840569   9280.370020919    261916.531308949
#>    ci95.integral.bt.up
#> 1       83956.06333586
#> 2     44291.6240231715
#> 3     68900.3237119473
#> 4     153640.838669866
#> 5     59942.0786380884
#> 6     136419.143109012
#> 7       386651.0163058
#> 8      123033.06603151
#> 9     227280.336200918
#> 10    507843.058409411
#> 11     174233.19292809
#> 12    278730.872017252
#> 13    510413.795127039
#> 14    184096.271227397
#> 15    273625.414586927
#> 16    531264.700666666
#> 17    182807.877260481
#> 18    277881.913647909
#> 19    523659.395718356
#> 20    184215.050052688
#> 21    276226.860768562