Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.7434002961597  15.4336592730603
#> 2                 FALSE          0.75 31.6915644838492   30.850939263052
#> 3                 FALSE          0.75 64.9263364029009  20.2672606556944
#> 4                 FALSE          0.75  242.23608839207  15.0115102230976
#> 5                 FALSE          0.75 65.5245971834789  22.6290130801862
#> 6                 FALSE          0.75  205.07689447713  12.4214768822799
#> 7                 FALSE          0.75 506.524675777174  6.92326326657107
#> 8                 FALSE          0.75 203.697440933569  13.2330151220709
#> 9                 FALSE          0.75 290.948139938075  4.54010053806154
#> 10                FALSE          0.75 672.825626091064 0.572009562650181
#> 11                FALSE          0.75 327.500479436463  9.88452751434742
#> 12                FALSE          0.75 334.787338472809  4.01126189916401
#> 13                FALSE          0.75 732.490463248234 0.923656035726564
#> 14                FALSE          0.75  336.50773000809  9.59536481945083
#> 15                FALSE          0.75 320.537204107974   3.2901066824612
#> 16                FALSE          0.75 785.488236316456 0.704943757997859
#> 17                FALSE          0.75 337.279512389036  8.38577168912467
#> 18                FALSE          0.75 349.978125776288  4.94617227315734
#> 19                FALSE          0.75 773.898152972253 0.652553860653867
#> 20                FALSE          0.75 345.665330315812  9.48672911479883
#> 21                FALSE          0.75 338.940309972423  5.35389883876667
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3409.66942564951 2124.52176656899 81502.8188359948 2.62307703082024
#> 2  1453.27777811241 282.487332940407 42490.6112409582 1.86580233459362
#> 3  2195.51949210496  670.61927675484  66182.602561029 2.19158223632972
#> 4  7495.46271178334 5963.23427542794 145924.875729385  2.1948846030809
#> 5   2361.0494698857 1076.42492105315 57336.5864813169 2.84455237983951
#> 6  4816.24635109381 3216.85577434693 130183.207751531 2.84473481321996
#> 7  17043.4826491543 15879.3677864921 362962.384772924 14.0843268787284
#> 8  5175.32223733836 3807.71041094872 117794.724972922 8.75354136879888
#> 9  7766.38912355449   6325.259296657 221116.792372151 7.43279706166011
#> 10 20089.6375878601 18759.2661544031 484239.318137611 18.3628188374146
#> 11 6797.20522258251 5584.91162921625 168563.663341125  12.626725270412
#> 12 9187.26487278778 7592.78371902088 266183.830901084 6.47696407948987
#> 13 20429.1034616122 19278.2968251703  489966.58378549 25.8544871125068
#> 14 6956.04604806732 5663.11229445779 177131.225317906 14.3147950797224
#> 15 8999.71298832388 7430.98348811815 265361.722346937 5.65098026254287
#> 16 20664.7560289701 19214.2203701877 512697.346082452 30.0930962537194
#> 17 6903.34506231351 5943.11577947853 176261.137173261 11.6101368925518
#> 18  9318.9561280993 7649.76484892116 270884.220763949 6.67786787795868
#> 19 19774.0824114588 18594.7221796243 501547.007506122 20.8971652003845
#> 20 7139.83495884476 5806.51793409055 179665.960222305 17.8776574397113
#> 21   9253.518911654 7411.73746225917 267244.237321088 8.55454731812017
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.702240137464027 36.7500580804763 66.4475249876785 1743.09501874994
#> 2   2.53217178516437 18.0459689280388 68.1680031872812 994.619320814832
#> 3   1.61595442574573   4.786884177385 113.310280478084 1320.74234003429
#> 4   0.29354672767758 140.200885298972 142.516826135742 6837.64266060731
#> 5   2.62617810924528  29.850471484934 127.535201434319 1136.87013395368
#> 6  0.502701673355527 10.5945083606438 99.8057425728992 3051.23680414408
#> 7  0.489820550488808 150.626053874203 169.263478355613 7890.82589565156
#> 8  0.524550674530684 60.7774950901225 110.846261500696 2907.04150514642
#> 9  0.362528319366134 22.5854953902731 71.0547606716483 3488.76196103053
#> 10 0.308533574167447   65.97841321844  241.99494670644 7170.99871921171
#> 11 0.737558174964394 68.4319229430759 236.756668479801 2522.26296784072
#> 12 0.593314337071904 10.2210845046957 175.689267538783  4207.5004727048
#> 13 0.427283688683783 97.3805174229348 266.127996419561 9724.40302451549
#> 14 0.595700480189748 43.1803046540771 197.986480856825 2293.01779300227
#> 15 0.293161619759273   19.42120302694 57.7967673053785 3512.40704175365
#> 16 0.714994213650699 91.1352634583299 418.114978823559 11260.7469874955
#> 17  0.45928410571304 78.8793314409253 146.660092625906 2412.00462103474
#> 18 0.382348282046487 8.16194270354665 67.7454810571941 4480.43569796976
#> 19  0.28960335744278 66.4022966648205 214.743891424889 9510.47734938186
#> 20 0.527111288938065 96.5434873941875 150.839147080613  3601.3564846805
#> 21 0.956765062440253 11.8726176782027 253.995246769348 5598.25413724637
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.4284385804604  96.058362011859    14.278474246932
#> 2             TRUE 28.6223196434427 34.7608093242557   26.6855166764566
#> 3             TRUE 61.3211836241385 68.5314891816633   17.6090156253426
#> 4             TRUE 238.625503220002 245.846673564138    14.528625856068
#> 5             TRUE 60.8453085186429 70.2038858483149   18.3089500904777
#> 6             TRUE 200.397305709383 209.756483244876     11.59453262961
#> 7             TRUE 483.355958061665 529.693393492682   6.11750846101698
#> 8             TRUE 189.297865381895 218.097016485243   12.3701292624679
#> 9             TRUE 278.721188771644 303.175091104505   3.94374145270425
#> 10            TRUE 642.618789103517 703.032463078611   0.06447183314473
#> 11            TRUE 306.729516366635  348.27144250629   8.67124431653099
#> 12            TRUE 324.132732562048  345.44194438357   3.03525981468073
#> 13            TRUE  689.95983194816 775.021094548308  0.220774367841742
#> 14            TRUE 312.959892101947 360.055567914233   8.61543752953869
#> 15            TRUE 311.241341576091 329.833066639857   2.80785581795719
#> 16            TRUE 735.985092979088 834.991379653825 -0.471221723457541
#> 17            TRUE 318.180837200788 356.378187577283   7.63024933522672
#> 18            TRUE 338.993033117046  360.96321843553   4.31720934919087
#> 19            TRUE  739.52231621762 808.273989726885  0.176156337660493
#> 20            TRUE 316.256583827487 375.074076804137   8.61963104449572
#> 21            TRUE 324.868079634115  353.01254031073   3.78002031105245
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5888442991886 3349.21558010713 3470.12327119189    78635.4275301512
#> 2   35.0163618496474 1423.59215922579 1482.96339699904    40854.4624582178
#> 3   22.9255056860461 2187.64506763316 2203.39391657675    64009.9814116726
#> 4   15.4943945901272 7264.83225546653 7726.09316810015    134676.953552686
#> 5   26.9490760698947 2311.94544429298 2410.15349547842    55466.4351109631
#> 6   13.2484211349497 4798.81838484055 4833.67431734707    125163.923208714
#> 7   7.72901807212516 16795.7027905313 17291.2625077774    349981.976174577
#> 8   14.0959009816739 5075.34325791511 5275.30121676161    113012.641696956
#> 9   5.13645962341883  7729.2359836375 7803.54226347149    215377.778946256
#> 10  1.07954729215563 19981.1030981158 20198.1720776045    472443.025244508
#> 11  11.0978107121638 6684.63470934115 6909.77573582387    164414.540759027
#> 12  4.98726398364729 9170.45118877755   9204.078556798    259262.492623485
#> 13  1.62653770361139 20268.9125104515 20589.2944127729    473969.940810162
#> 14   10.575292109363 6885.01444691136 7027.07764922327    173359.211048417
#> 15   3.7723575469652 8967.76510934457  9031.6608673032    259583.812763252
#> 16  1.88110923945326 20514.8385205811 20814.6735373591    494173.417288022
#> 17  9.14129404302262 6773.58856209319 7033.10156253383    172293.389571659
#> 18  5.57513519712381 9305.52973235197 9332.38252384663    263513.904040789
#> 19  1.12895138364724 19664.8506334452 19883.3141894724    485902.272266389
#> 20  10.3538271851019 6981.02092208132  7298.6489956082    173741.728805005
#> 21  6.92777736648088 9233.98845557336 9273.04936773465    258035.109265318
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84370.2101418385  86.602169315752 96.8846312765674    14.0572686036308
#> 2     44126.7600236986 28.0345919080457 35.3485370596527    25.8878825641298
#> 3     68355.2237103854 60.6308352196946 69.2218375861071    17.0999899812327
#> 4     157172.797906084 237.934114570032 246.538062214109    14.4361586368495
#> 5     59206.7378516707 59.9492745189934 71.0999198479643    17.4817039860655
#> 6     135202.492294348 199.501214243219 210.652574711041     11.436181602503
#> 7     375942.793371271 478.919395094866 534.129956459481    5.96321498761301
#> 8     122576.808248888 186.540499850723 220.854382016415    12.2048957999908
#> 9     226855.805798046 276.379857697221 305.516422178928    3.82954503210392
#> 10    496035.611030715 636.834501169732 708.816751012397 -0.0327162427180159
#> 11    172712.785923223 302.752097906455  352.24886096647    8.43891349141721
#> 12    273105.169178684 322.092488877009 347.482188068609    2.84836579850308
#> 13    505963.226760818 681.815668507721 783.165257988747  0.0861800059063501
#> 14    180903.239587394 308.450731651834 364.564728364346    8.42779187827892
#> 15    271139.631930622  309.46128279339 331.613125422558    2.71550990773302
#> 16    531221.274876882 726.505767659166 844.470704973746  -0.696444900757512
#> 17    180228.884774864 314.523644079634 360.035380698437    7.48557484192711
#> 18    278254.537487109 336.889504735489 363.066746817087    4.19676964034623
#> 19    517191.742745855 732.939709179499 814.856596765007  0.0849312800660176
#> 20    185590.191639604 310.625121733977 380.705538897646    8.45359098848022
#> 21    276453.365376859 322.173397228907 355.707222715938    3.47863931638377
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8100499424898 3337.63931181178 3481.69953948724    78086.3525992449
#> 2   35.8139959619741 1417.90767901346 1488.64787721137    40541.1573721611
#> 3    23.434531330156 2186.13719911728 2204.90178509263    63593.9475745618
#> 4   15.5868618093456 7220.66897659736 7770.25644696933    132523.096114595
#> 5   27.7763221743069 2302.54254577523 2419.55639399617    55108.3210187677
#> 6   13.4067721620567 4795.48111470695 4837.01158748067    124202.783615408
#> 7   7.88331154552913 16748.2555835609 17338.7097147478    347496.366017447
#> 8   14.2611344441511 5056.19834696172   5294.446127715    112096.923622835
#> 9   5.25065604401917 7722.12155258956 7810.65669451943    214278.818928531
#> 10  1.17673536801838  19960.319897952 20218.9552777683    470184.160647956
#> 11  11.3301415372776 6663.07865361408 6931.33179155094    163620.027924158
#> 12  5.17415799982494 9167.23154715858 9207.29819841698    257937.129974583
#> 13  1.76113206554678 20238.2376474633 20619.9692757612     470906.75385744
#> 14  10.7629377606227 6871.41265094532 7040.67944518931    172636.910443621
#> 15  3.86470345718938 8961.64743039108 9037.77854625669      258477.4045451
#> 16  2.10633241675323 20486.1309125918 20843.3811453484    490626.281986961
#> 17  9.28596853632223 6748.74157268929 7057.94855193772    171533.608116033
#> 18  5.69557490596845 9302.95872040035 9334.95353579825    262102.566795928
#> 19  1.22017644124172 19643.9339099958 19904.2309129219    482906.471901333
#> 20  10.5198672411174 6950.60972355215 7329.06019413737    172607.301512331
#> 21  7.22915836114956 9230.24858100473 9276.78924230328    256271.659212085
#>    ci95.integral.bt.up
#> 1     84919.2850727447
#> 2     44440.0651097553
#> 3     68771.2575474963
#> 4     159326.655344175
#> 5     59564.8519438661
#> 6     136163.631887653
#> 7     378428.403528401
#> 8     123492.526323009
#> 9     227954.765815771
#> 10    498294.475627266
#> 11    173507.298758093
#> 12    274430.531827586
#> 13     509026.41371354
#> 14     181625.54019219
#> 15    272246.040148774
#> 16    534768.410177943
#> 17     180988.66623049
#> 18     279665.87473197
#> 19     520187.54311091
#> 20    186724.618932278
#> 21    278216.815430091