Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1024318053689  15.6208156951123
#> 2                 FALSE          0.75 32.5377038594393  31.6429554190237
#> 3                 FALSE          0.75 64.8859349319735  20.4866381979579
#> 4                 FALSE          0.75 242.933565396714  14.9637386404216
#> 5                 FALSE          0.75 65.8711281348927  22.4398102979516
#> 6                 FALSE          0.75 204.939056486916  12.2500802563704
#> 7                 FALSE          0.75 512.920803408998  7.20571593809955
#> 8                 FALSE          0.75 204.405269794702  13.4582424121269
#> 9                 FALSE          0.75 292.632024298815  4.82095341149229
#> 10                FALSE          0.75 673.657869499585 0.556111723508714
#> 11                FALSE          0.75 327.124218408121  9.86047576194664
#> 12                FALSE          0.75 336.506275304342  3.75476061052825
#> 13                FALSE          0.75 730.458296562206 0.961432696484819
#> 14                FALSE          0.75 344.371624126206  9.47418054305843
#> 15                FALSE          0.75 321.625384760607  3.83919367076763
#> 16                FALSE          0.75 765.926697552461 0.389945487323059
#> 17                FALSE          0.75 328.539269744701   7.8557992113048
#> 18                FALSE          0.75 348.261534321702  4.95366498995933
#> 19                FALSE          0.75  775.24655684695 0.524390671292178
#> 20                FALSE          0.75 347.988820339392  9.40623652392337
#> 21                FALSE          0.75 339.209322571833  5.31518509498639
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3403.24906820507 2104.88821767724 81077.1579021274 2.55443840961582
#> 2  1455.36104561211 259.116123862563 42337.7280165797  1.9985292779779
#> 3  2196.92748032759 658.191582389714 66433.8850995097   1.594885728327
#> 4  7545.61307898915 6034.41192561848 148770.460731011 2.79586708883817
#> 5  2369.55512387162  1118.0400808102 57181.0163267015 2.36394362662429
#> 6  4820.95235390561 3262.14247732845 130559.750288033 2.29380583851249
#> 7  17061.9921322798 15862.5311058943 362791.509193176 13.2057236833376
#> 8    5184.824182283   3784.185286647 118409.209228059 7.18943061364712
#> 9  7764.41751332275 6255.37830607128 221357.648845641 6.28845135046988
#> 10 20063.2063228143 18733.7872913241 481918.723087544 22.5198141479373
#> 11 6759.41901493466  5551.7779544475 166762.460032687 12.1584048943559
#> 12 9190.56869597029 7700.43589433514 266943.373108838 5.65668383167055
#> 13 20417.2299067302 19203.9662056922 489265.696040219 26.5859313132135
#> 14 6945.33349979345 5722.76029276044 175192.690371422 15.9823128826912
#> 15 9007.40544105335 7289.28530325031 263857.209204534  6.5353152044825
#> 16  20668.030425934  19393.587607866 513118.413287616 27.6175676898758
#> 17 6891.29781741024 6029.90651984806 175609.849712655 13.0655821584185
#> 18 9323.64354202114 7640.75183724873 270402.889138143 6.11165820441469
#> 19 19791.5212541935 18679.9793226638 503884.300627067 25.1509535419168
#> 20 7136.51518058267 5832.29321779602 179546.471260433 11.7512865132867
#> 21 9259.07024786278 7459.60081838029 267870.186099149 7.73747939436778
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.87198647496798 35.9692144572505 71.1391546884032 1804.66449299496
#> 2   2.73504528297735 19.1529612702108 73.3452442288338  969.25000134454
#> 3   1.49477631554376 3.66173221764137 95.7915123018894 881.575238986197
#> 4  0.240029481044714 71.3045866263724 90.7384611211553 3700.84902493624
#> 5   2.36773715894179 27.4791407280504 123.035982685624 1213.05776454782
#> 6  0.584436395981319 10.2502403866043 107.521531534598 1630.43246595024
#> 7  0.488930526462719 95.5330228426931 139.285391176519 5453.50742498249
#> 8  0.558611111397257 61.2343486925422 102.085391704035 3319.92604081418
#> 9  0.329662329918152 9.54307693269257 59.8524722301062 3391.78954130912
#> 10 0.333826867772944  71.426373752527 190.728225371594 9596.91456373946
#> 11 0.593814385397561 96.1066905658413 192.967947951645 3656.11765055117
#> 12 0.299973134338826 13.3956295247946 84.2453247743443 3604.06757190992
#> 13 0.819495404150503 93.9129734571294 576.845736482267 9976.18557152121
#> 14  0.66263808252007 83.0242372328016 216.202549258673 4142.59919581742
#> 15 0.808182988440113 14.0920290291094 239.051271114673  3103.2770916323
#> 16 0.346213295706961 118.397534230649 238.840049946173 8888.62475852799
#> 17 0.747693148890753  105.61582853638 192.233950935168 3812.93439341765
#> 18 0.403996440214572 8.60291830466706 122.711600867593 4129.55697360751
#> 19 0.264276060869566  80.225347957981 230.489869993754 7105.15803073299
#> 20 0.347825740113705 104.859139585645 139.776543012136 4502.70140694528
#> 21 0.428543747482221 9.27591055550206 143.734168639507 3645.54874272399
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.9003806215508 96.3044829891869      14.18639794379
#> 2             TRUE 29.2501231971656 35.8252845217129    27.1438059285259
#> 3             TRUE 62.2623479088755 67.5095219550714    18.0277311588884
#> 4             TRUE 238.334364035575 247.532766757852     14.568890144103
#> 5             TRUE 61.9824408690958 69.7598154006897    18.5448826714924
#> 6             TRUE 201.165745882563 208.712367091269    11.2886823849811
#> 7             TRUE 491.197387949908 534.644218868089    6.40142522206838
#> 8             TRUE 192.578656435252 216.231883154151    12.5393271338784
#> 9             TRUE 282.287521827292 302.976526770338    4.27865887877693
#> 10            TRUE 636.612775226228 710.702963772942 0.00696652602222114
#> 11            TRUE 307.123642356906 347.124794459336    8.88365109796765
#> 12            TRUE 327.201030401244  345.81152020744    3.26130480454088
#> 13            TRUE 686.724439551969 774.192153572442  -0.386637243342759
#> 14            TRUE 318.080719434179 370.662528818233    8.38414089731292
#> 15            TRUE 310.874791249234 332.375978271981    2.50973265478364
#> 16            TRUE 720.495798702615 811.357596402306  -0.179575384114892
#> 17            TRUE 307.046387094102 350.032152395299    6.62584398137951
#> 18            TRUE  338.20785657544 358.315212067964    4.28909084580635
#> 19            TRUE 733.873238270497 816.619875423403  0.0896565511617419
#> 20            TRUE 328.657954025035 367.319686653748    8.83406318143633
#> 21            TRUE 326.481168968098 351.937476175568    4.61023063037813
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   17.0552334464346 3344.07971042289 3462.41842598725    78108.4848111507
#> 2   36.1421049095214 1423.85442432261  1486.8676669016    40743.3117643679
#> 3   22.9455452370274 2190.90393082957 2202.95102982561    64983.6938313774
#> 4   15.3585871367402 7428.31703398877 7662.90912398953    142682.564084991
#> 5   26.3347379244108 2324.35193737398 2414.75831036927    55185.5363040203
#> 6   13.2114781277596 4804.09070846965 4837.81399934158    127877.688881545
#> 7   8.01000665413073 16904.8403097035  17219.143954856     353820.48947908
#> 8   14.3771576903754 5084.09367868377 5285.55468588223    112947.930890919
#> 9   5.36324794420765 7748.71915176847 7780.11587487703    215778.155050188
#> 10  1.10525692099521 19945.7099379914 20180.7027076372    466131.798630192
#> 11  10.8373004259256 6601.32350895385 6917.51452091547    160748.146497531
#> 12  4.24821641651562   9168.532885402 9212.60450653858    261014.681953047
#> 13   2.3095026363124 20262.7430653932 20571.7167480671    472854.870775066
#> 14  10.5642201888039 6808.75862954549 7081.90837004141    168378.114694302
#> 15  5.16865468675161 8984.22405330047 9030.58682880624    258752.318388799
#> 16  0.95946635876101 20473.2664821246 20862.7943697435    498496.625559838
#> 17  9.08575444123009  6717.5597794679 7065.03585535259    169337.572635483
#> 18   5.6182391341123 9309.49174140996 9337.79534263232    263609.767916559
#> 19 0.959124791422614 19659.5505568026 19923.4919515844    492196.315666512
#> 20  9.97840986641042 6964.02189596429 7309.00846520106    172139.527446008
#> 21  6.02013955959464 9243.81137499898 9274.32912072658    261873.258417368
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84045.8309931041 87.0957325225218 97.1091310882159    13.9117222041751
#> 2     43932.1442687915 28.6205864746026  36.454821244276    26.2822666643881
#> 3      67884.076367642 61.7599589044525 68.0119109594944    17.5568766194921
#> 4     154858.357377031 237.453665902591 248.413464890836     14.493280857574
#> 5     59176.4963493827 61.2377986267091 70.5044576430763    17.7990454664257
#> 6     133241.811694521 200.443197043431   209.4349159304     11.104584920247
#> 7     371762.528907273 487.037584989657  538.80402182834    6.24741210623263
#> 8     123870.487565198 190.313985791953  218.49655379745    12.3633646337883
#> 9     226937.142641095 280.306659651894 304.957388945736    4.17481524485271
#> 10    497705.647544895 629.519033769628 717.796705229542 -0.0981889373262562
#> 11    172776.773567844 303.293744815184 350.954692001059    8.69659956656742
#> 12     272872.06426463 325.419174994268 347.593375614417    3.16681326722415
#> 13    505676.521305371 678.349871188307 782.566721936104  -0.644778295650168
#> 14    182007.266048541 313.046290876131 375.696957376281    8.17540990131909
#> 15    268962.100020269 308.816166959822 334.434602561393    2.25515501342501
#> 16    527740.201015395 711.796264880304 820.057130224617  -0.288632572262585
#> 17    181882.126789827 302.930728714201 354.147810775201    6.39032063947893
#> 18    277196.010359728 336.282684241049 360.240384402355    4.16183196713876
#> 19    515572.285587623 725.950687904793 824.542425789107 0.00640959198782864
#> 20    186953.415074858  324.95629877335 371.021341905433    8.72449807330051
#> 21     273867.11378093 324.043862958873 354.374782184794    4.47523934992123
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.3299091860496 3332.74940786886 3473.74872854128    77540.0154958573
#> 2   37.0036441736593 1417.82124152249 1492.90084970172    40437.9980139444
#> 3   23.4163997764236 2189.75048518102 2204.10447547417    64705.9976310968
#> 4   15.4341964232692 7405.85608920146 7685.37006877684    141516.796642136
#> 5   27.0805751294775 2315.69600804465  2423.4142396986    54803.4231081878
#> 6   13.3955755924938 4800.86188274787 4841.04282506336     127364.10265477
#> 7   8.16401976996648 16874.7474075081 17249.2368570514    352102.634640211
#> 8   14.5531201904655 5064.80485884562 5304.84350572039    111902.154188063
#> 9   5.46709157813187 7745.71308253467 7783.12194411083    214709.741344675
#> 10  1.21041238434368 19923.2106302594 20203.2020153693    463108.770542614
#> 11  11.0243519573259 6571.04990142561 6947.78812844371    159596.469437607
#> 12  4.34270795383235 9164.31326210169 9216.82412983889    259879.400667895
#> 13  2.56764368861981 20233.1604787542 20601.2993347061    469712.372320037
#> 14  10.7729511847978 6782.60599481716 7108.06100476974     167073.19594762
#> 15  5.42323232811025  8979.7850641563 9035.02581795041    257774.786104934
#> 16   1.0685235469087  20435.971258842 20900.0895930261    495696.708760901
#> 17  9.32127778313068 6684.29079347894 7098.30484134155    168136.498301556
#> 18  5.74549801277989 9306.78182214399 9340.50526189829    262308.957469873
#> 19  1.04237175059653 19634.2795721959 19948.7629361912    489958.190886831
#> 20  10.0879749745462 6930.99126699481 7342.03909417054     170721.17650282
#> 21  6.15513084005154   9240.889463174 9277.25103255157     260724.91056341
#>    ci95.integral.bt.up
#> 1     84614.3003083975
#> 2      44237.458019215
#> 3     68161.7725679226
#> 4     156024.124819886
#> 5     59558.6095452152
#> 6     133755.397921295
#> 7     373480.383746142
#> 8     124916.264268055
#> 9     228005.556346607
#> 10    500728.675632473
#> 11    173928.450627768
#> 12    274007.345549782
#> 13      508819.0197604
#> 14    183312.184795224
#> 15    269939.632304133
#> 16    530540.117814331
#> 17    183083.201123753
#> 18    278496.820806414
#> 19    517810.410367304
#> 20    188371.766018045
#> 21    275015.461634888