Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.5999885727832  15.5976037422648
#> 2                 FALSE          0.75 32.2357620394473  31.0072365889726
#> 3                 FALSE          0.75 65.0094690804434  20.4855752088817
#> 4                 FALSE          0.75 242.568298792107  14.9993915506635
#> 5                 FALSE          0.75 66.1213457946198  22.8747305356392
#> 6                 FALSE          0.75 205.515168217794  12.4300008185976
#> 7                 FALSE          0.75 505.764631540124  7.01506405845191
#> 8                 FALSE          0.75 204.478676047568  13.5059942985269
#> 9                 FALSE          0.75 289.207865957914  4.65562807189475
#> 10                FALSE          0.75  676.23691764429 0.625433258111603
#> 11                FALSE          0.75  328.74919122584  9.97214620497822
#> 12                FALSE          0.75 333.680221589089  3.80857988525739
#> 13                FALSE          0.75 737.544136331948 0.955792838267323
#> 14                FALSE          0.75 343.200624341336  9.75545416748919
#> 15                FALSE          0.75 319.643635501032   3.5267853034797
#> 16                FALSE          0.75  777.56542026882 0.459390064242502
#> 17                FALSE          0.75 335.826856792416  8.23406562594613
#> 18                FALSE          0.75  347.93201749299  4.80491364527305
#> 19                FALSE          0.75 778.872554388381 0.547660973045274
#> 20                FALSE          0.75 351.547262527367  9.54346800682757
#> 21                FALSE          0.75 338.236424046776  5.39729147080241
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.72042930883 2109.60910049579 81356.5206272844 1.97593954893864
#> 2  1456.52985247654 288.855295962447 42138.8657949546 2.15075844528644
#> 3  2194.71157073034 657.047413021063 66328.3400505358 1.42560727561886
#> 4  7556.01456090761 6035.48464356364 149183.171562688 2.41208707826947
#> 5  2362.74802981128  1078.3162561437 57049.6464577174 3.01748751722345
#> 6  4822.86487320694 3224.24750950249 131327.242290363  3.2255351610642
#> 7  16993.1645360394 15774.4659316326 360052.780845467 12.9232206922524
#> 8  5174.83160532446 3762.77139168746 118644.040892236 11.2652810105952
#> 9  7765.19014772231 6291.31016057644 220407.184586244 6.44840903924243
#> 10 20049.0529745928 18703.3999127015 481353.614286571 19.2294546489208
#> 11 6820.16405192783 5597.48679634559 168669.750845867 11.2300840528026
#> 12 9194.05353816605   7643.737707608 265523.230305632  6.9602901418074
#> 13 20439.4577264129 19282.6229068485 490218.586445598 31.0406096990361
#> 14 6961.04283835734 5652.16093335668 176808.482611618 10.2682888952916
#> 15 9006.55362218447 7362.54643514749 264595.792162245 3.59721189581374
#> 16 20691.8975244359 19407.6481865383 517312.628904075 19.0834408506074
#> 17 6907.65708162978 5992.78769741291 175574.790959978 14.4741565184162
#> 18 9318.43946236474 7687.93098101044  271199.88316529 5.98660601477385
#> 19  19796.266000997  18663.241940729 500317.435716758 25.3197279192919
#> 20  7102.4231197661 5775.05353360715 179207.984794279 15.7709017380546
#> 21 9254.54384322228 7409.57120164853 269146.361611064 8.64558734388744
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.600329916296805 16.6921916753989 58.1132423800816 1608.71363754812
#> 2   2.29867412013888  19.340531113896 51.6324996311365 1059.64614449173
#> 3   2.13665196195106 5.03393200316971 136.671676013656 1289.07611201504
#> 4  0.350931235818644 54.3123802410016 103.309980434332 2529.05292214373
#> 5   2.62767176243566 36.4881956270165 134.912559274861 1393.68752775141
#> 6  0.296510760882114 8.23833119983287 61.3165112920123 1758.60146593767
#> 7  0.426786547820517 192.289521989073 196.799803488419 10009.6257389523
#> 8  0.555808453920438  40.211582547931 106.598733477221 2369.78124006206
#> 9  0.339279412735011 29.3599391570246 68.1686263347247 5677.07608799205
#> 10 0.323525802192346 95.3887744550905 195.988073635281   12436.64106433
#> 11 0.501724084209706  66.474677241355 165.777805202397 2481.41674409121
#> 12 0.491074691265536 14.0798568742697 117.505196038958 4602.64991673824
#> 13 0.615284663173444 87.7300936730626 352.319890521274 9244.89090953447
#> 14 0.591625006469541 58.0394115871343  183.01112202625 3274.61278919218
#> 15 0.524576275753853 18.8295638153607 147.925722373394 3056.10673499539
#> 16 0.334976587781981 72.5036028839365 245.060309146198 4296.26287539758
#> 17 0.453849761249536 84.9197158488155   141.5458468361 3282.32450894533
#> 18 0.279154135575724  10.776261634845 53.2656404368663 3446.08777478695
#> 19 0.222373049961103 81.1373091583169  158.69948547199 11131.0235701997
#> 20 0.416722622126496 91.6715358001028 130.110507658423 3509.24933931704
#> 21 0.837721485660235 7.24169313351051 236.537095741997 3205.99538076658
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.3495680147792 94.8504091307873    14.6100610299566
#> 2             TRUE 28.6977643969511 35.7737596819435    27.2259176613441
#> 3             TRUE 62.6643451120504 67.3545930488365    16.9707827314722
#> 4             TRUE 238.600415548354 246.536182035861    14.4221096677418
#> 5             TRUE 61.1575788287872 71.0851127604524    18.5522104864326
#> 6             TRUE 200.209162877843 210.821173557745    11.9422406169465
#> 7             TRUE 484.505933501369  527.02332957888    6.31300018728716
#> 8             TRUE 185.947288785139 223.010063309997    12.5916893918278
#> 9             TRUE  278.60023308836 299.815498827468    4.09751343794566
#> 10            TRUE 644.604464746816 707.869370541765  0.0932333135051931
#> 11            TRUE 310.275702958979   347.2226794927    9.14681008645326
#> 12            TRUE 322.230544305816 345.129898872362    3.00076201812559
#> 13            TRUE 686.482333377033 788.605939286862 -0.0563504326529921
#> 14            TRUE 326.309289108582 360.091959574091    8.78223103184679
#> 15            TRUE 313.726221932418 325.561049069645    2.66385732986461
#> 16            TRUE 746.173160069571 808.957680468069 -0.0916464226588559
#> 17            TRUE 312.016869319621 359.636844265211    7.48748276869064
#> 18            TRUE 338.084050598687 357.779984387293    4.34570509225099
#> 19            TRUE 737.221601961146 820.523506815616    0.18185730585926
#> 20            TRUE 325.604129168267 377.490395886467    8.85795929342949
#> 21            TRUE 324.014432866082 352.458415227471    4.01923962689132
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5851464545731 3384.26177400279 3439.17908461486    78710.1866935178
#> 2    34.788555516601 1424.71467879418  1488.3450261589    40395.7478872657
#> 3   24.0003676862912 2186.43075258513 2202.99238887556     64207.809846271
#> 4   15.5766734335851 7466.67069541116 7645.35842640406    145022.879505761
#> 5   27.1972505848459 2302.72494800484 2422.77111161772    54757.0304745663
#> 6   12.9177610202486 4809.31281838321 4836.41692803067    128434.342878896
#> 7   7.71712792961667 16676.8482723674 17309.4807997115     343586.94650489
#> 8    14.420299205226 5108.68355203312 5240.97965861581    114745.750752334
#> 9   5.21374270584384   7716.893047809 7813.48724763561    211068.394421497
#> 10  1.15763320271801 19892.1384406142 20205.9675085714    460895.339735748
#> 11  10.7974823235032  6710.8132078658 6929.51489598986    164587.820301837
#> 12   4.6163977523892 9170.89217360788 9217.21490272422    257951.871192598
#> 13  1.96793610918764 20295.1417223207 20583.7737305051    475010.740899414
#> 14  10.7286773031316  6865.5680062965 7056.51767041817    171421.744573396
#> 15  4.38971327709478  8975.5789897082 9037.52825466074    259568.496583178
#> 16  1.01042655114386 20572.6290976918   20811.16595118    510245.276474046
#> 17  8.98064848320162 6767.96414905848 7047.35001420109    170175.367142763
#> 18  5.26412219829512 9300.71251197542 9336.16641275406    265531.068775765
#> 19 0.913464640231287 19662.7951274316 19929.7368745625    482006.901943779
#> 20  10.2289767202257 6951.62344337493 7253.22279615727    173435.269631102
#> 21   6.7753433147135 9242.63125801765  9266.4564284269    263872.499209703
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84002.8545610511 87.7271470568635  95.472830088703     14.4209571063231
#> 2     43881.9837026435 28.0202754866859 36.4512485922087     26.5018353135004
#> 3     68448.8702548005 62.2152788202305 67.8036593406564     16.2977373634576
#> 4     153343.463619614 237.840608118699 247.295989465515     14.3115663284589
#> 5     59342.2624408685 60.2070702608619 72.0356213283778     17.7244938812653
#> 6     134220.141701831 199.193119302108  211.83721713348     11.8488397272686
#> 7     376518.615186043  480.43511898331 531.094144096939      6.1785624247237
#> 8     122542.331032138 182.398725266801 226.558626828335     12.4166097288428
#> 9     229745.974750991 276.568984240999 301.846747674829     3.99064042293413
#> 10    501811.888837393 638.547186532406 713.926648756175 -0.00867731418539597
#> 11    172751.681389897 306.738226482347 350.760155969333      8.9887669999272
#> 12    273094.589418667 320.038052911146 347.322390267031     2.84607349037694
#> 13    505426.431991782 676.704541321837 798.383731342059   -0.250165101552627
#> 14    182195.220649839 323.074778106565 363.326470576108     8.59586915480889
#> 15    269623.087741313 312.593100185237 326.694170816827     2.49861580300215
#> 16    524379.981334104 740.161876201629  814.96896433601    -0.19716404781018
#> 17    180974.214777193  307.45751001632 364.196203568512     7.34452009389704
#> 18    276868.697554814 336.198269704033 359.665765281947     4.25777153954463
#> 19    518627.969489736 729.245887666569 828.499221110193    0.111809795121512
#> 20    184980.699957455  320.63629512078 382.458229933954     8.72669166745964
#> 21    274420.224012425 321.291072852757 355.181775240796     3.75535735890835
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7742503782066 3379.00373362504 3444.43712499261    78203.4418976901
#> 2   35.5126378644448 1418.62241149331 1494.43729345978    40061.9593517508
#> 3   24.6734130543057 2184.84506400413 2204.57807745655    63801.7508709863
#> 4    15.687216772868 7449.56229563525 7662.46682617998    144226.227835286
#> 5   28.0249671900131 2291.23116638233 2434.26489324023    54318.0189033246
#> 6   13.0111619099265 4806.71774405527 4839.01200235861    127880.383417125
#> 7   7.85156569218013 16616.2770729409  17370.051999138     340433.91439712
#> 8    14.595378868211 5096.01690353052 5253.64630711841    113999.269661715
#> 9   5.32061572085537 7707.64466697454 7822.73562847007     209280.11545378
#> 10   1.2595438304086 19862.0909766608 20236.0149725248    456977.797800484
#> 11  10.9555254100292 6689.87368453478 6950.45441932089    163806.174027448
#> 12  4.77108628013784 9166.45701869248 9221.65005763962    256502.036468825
#> 13  2.16175077808727 20267.5067428137 20611.4087100121     472098.60026291
#> 14  10.9150391801695 6847.28559164656 7074.80008506812    170390.241544801
#> 15  4.55495480395725 8969.64767710637 9043.45956726258    258605.822961654
#> 16  1.11594417629518 20549.7904627834 20834.0045860884    508891.953668295
#> 17  9.12361115799522 6741.21443856611 7074.09972469346    169141.434922445
#> 18  5.35205575100147 9297.31798956045 9339.56093516904    264445.551126708
#> 19 0.983512150969035 19637.2368750467 19955.2951269473    478500.629519167
#> 20  10.3602443461955 6922.74690959789  7282.0993299343    172329.856089218
#> 21  7.03922558269647  9240.3501246806 9268.73756176396    262862.610664761
#>    ci95.integral.bt.up
#> 1     84509.5993568787
#> 2     44215.7722381584
#> 3     68854.9292300853
#> 4     154140.115290089
#> 5     59781.2740121102
#> 6     134774.101163601
#> 7     379671.647293813
#> 8     123288.812122758
#> 9     231534.253718708
#> 10    505729.430772657
#> 11    173533.327664286
#> 12    274544.424142439
#> 13    508338.572628286
#> 14    183226.723678434
#> 15    270585.761362836
#> 16    525733.304139854
#> 17    182008.146997511
#> 18    277954.215203872
#> 19    522134.241914349
#> 20     186086.11349934
#> 21    275430.112557366