Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8620226192984  15.5805702500875
#> 2                 FALSE          0.75 33.7649970858381  31.1472664276962
#> 3                 FALSE          0.75 64.2941118358788  20.8320723468337
#> 4                 FALSE          0.75 242.339063293004  14.9566391313524
#> 5                 FALSE          0.75  64.960671331992  22.5241063620226
#> 6                 FALSE          0.75 205.099963262042  12.3736984419008
#> 7                 FALSE          0.75 507.231464201146    7.106401779582
#> 8                 FALSE          0.75 199.505128819403  13.2648764574194
#> 9                 FALSE          0.75 289.646115368085   4.7403520250618
#> 10                FALSE          0.75 685.838339291421 0.679181783858329
#> 11                FALSE          0.75 323.053858747783  9.52895700129131
#> 12                FALSE          0.75 334.411863904544  3.76976993282846
#> 13                FALSE          0.75 720.827416078831  0.64388208139099
#> 14                FALSE          0.75 342.192103357271  9.70120381693994
#> 15                FALSE          0.75 320.269174754335  3.53500555281501
#> 16                FALSE          0.75  769.40922775322 0.414537312238414
#> 17                FALSE          0.75 336.634106509223  8.11169544320693
#> 18                FALSE          0.75 348.247914499653  4.94472770155503
#> 19                FALSE          0.75 778.967545050464 0.710033062629067
#> 20                FALSE          0.75 350.798512051335   9.5342939376667
#> 21                FALSE          0.75 338.765536781124  5.24209690499403
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.20919446534 2107.38994246048 81361.0626440768 2.40657698306762
#> 2  1456.59733123734 270.496511950548 42183.7789514558 2.33467656599078
#> 3  2192.89793845545  639.40950915531 66618.3936891991 1.42532997592648
#> 4  7580.21469879111 6066.11967316245 150218.793062585 2.45133381142814
#> 5  2362.75382202881 1106.93781478082 57015.0188589482 2.10037495961819
#> 6   4819.3821232707 3233.16761535871 130905.120034944 2.59940756300692
#> 7   17022.212524846 15778.9932986229 361108.577681444 10.8438530093409
#> 8  5184.74611842831 3773.49231732927 119600.921341194 9.72464544422312
#> 9  7771.80995548125 6262.67685460459  221768.45683256 4.94552598426711
#> 10 20088.6572233225 18749.4921468729 482336.324527141 29.7372544396251
#> 11 6798.80079568806 5667.76310163229 168290.278289196 14.9884263013602
#> 12 9190.43797639564 7670.10449777724 264651.236578049 5.12715077442464
#> 13 20436.1983677525 19425.4021469462 489399.016421336 20.3044817491101
#> 14  6953.2332440526 5667.47810246096 176430.838840073 17.5143239074652
#> 15 9005.02278904307 7356.42832672738 264579.604012318  5.7467283972866
#> 16 20673.3957185831 19387.6016845066 513805.242071588 16.4550486153648
#> 17  6901.8806716136 6005.19131092595 176320.631135102 14.9099323684611
#> 18 9320.11737213086 7640.43046512887 271179.325347757 7.58539746306826
#> 19 19784.6292142287 18571.5093635869 500193.443460435 24.3324944023003
#> 20 7106.46430599868 5760.53174931781  179444.30177579 13.5232197889023
#> 21  9256.6863931621 7470.90228907713 267984.668599289 6.94997525571583
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.688022652693968 23.3186174851238 63.1583306317143 1696.95153236499
#> 2   2.78348540710603 27.3964480170462 95.0129697420116 1237.73916205547
#> 3   1.60528238260003 6.68208540033174 97.2531418736938 809.401453091748
#> 4  0.297956761934668 40.2557470292338 86.1559506329561 2103.76415861991
#> 5    2.1276576654827 33.1522541932986 138.439594582816 1452.20153461264
#> 6  0.399262795658124 9.40441673905486 78.0674557191444 2321.21686468499
#> 7  0.469695740940891 123.490774599948 174.230293090143 7313.29214967314
#> 8   0.91215392782788 57.9666625556403 160.332353514746 3613.47839376714
#> 9  0.340787858196612 20.4376590067589 76.0957102140564 2708.98297927289
#> 10  0.36932788014326 79.0753569047316  237.64222847482 7932.67112162471
#> 11 0.509784953669711  64.176790346527 114.499314980382 2525.85414555056
#> 12 0.277204251756972 10.5504224030337 78.7978574468051 6396.27581456329
#> 13 0.279763690600967 103.182082310427 205.700046766583 10624.6889477537
#> 14 0.608897261247472 74.8270173615211  158.81518528283 4402.34823747556
#> 15 0.655677879138524 16.2344455484174 207.917068616905 4409.76980675034
#> 16  0.28544293319599 105.223571328725 222.036182659328 11584.3769230319
#> 17  0.63766298086787  87.425639574832 199.319133431853 3041.62502179401
#> 18  0.48205969369639 8.18003726263898 114.158273279917 3944.25123844033
#> 19 0.516378414379365 64.5216803807085 340.983225532366 11949.2182327257
#> 20 0.380270503302756 103.431519242185  145.84567290572 4091.32070334946
#> 21 0.534052040060026 12.4443713080881 132.912107825581 4909.45514064365
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.9032034821521 95.8208417564446    14.4487729864059
#> 2             TRUE 29.9244541347832 37.6055400368929    26.5684329330067
#> 3             TRUE 61.9494440254798 66.6387796462779    18.1913828274567
#> 4             TRUE 238.306619173205 246.371507412803    14.4665002579699
#> 5             TRUE   61.50555452342 68.4157881405639    19.0241095023036
#> 6             TRUE 200.823937820896 209.375988703189    11.7169111430432
#> 7             TRUE  489.39332600078 525.069602401512    6.33375228573423
#> 8             TRUE 183.508087063656  215.50217057515    11.7643832461426
#> 9             TRUE 281.510725123965 297.781505612204    4.17975599832837
#> 10            TRUE 636.920555738238 734.756122844604  0.0716374210226665
#> 11            TRUE 298.397897482046 347.709820013521    8.69036075250464
#> 12            TRUE 325.977700880616 342.846026928473    3.31376893868824
#> 13            TRUE 687.426543601545 754.228288556117     0.1836708103524
#> 14            TRUE 313.381040529491 371.003166185052    8.69956782218785
#> 15            TRUE 310.815806540799 329.722542967872    2.45641544163214
#> 16            TRUE 742.340672780945 796.477782725495 -0.0550163128689898
#> 17            TRUE 312.107267763105 361.160945255342    7.06273983967929
#> 18            TRUE 335.769935672906 360.725893326401    4.15173950542447
#> 19            TRUE  738.94059175868 818.994498342248  -0.139409429024988
#> 20            TRUE 328.552815498591  373.04420860408    8.90874895973366
#> 21            TRUE 327.332827485471 350.198246076776    4.36358129909529
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7123675137691 3368.85006870231 3445.56832022836    78569.5773733364
#> 2   35.7260999223856  1411.5301742493 1501.66448822538    40147.6980298745
#> 3   23.4727618662107  2181.9059079719   2203.889968939    65286.9282988631
#> 4    15.446778004735 7513.99399492802  7646.4354026542    146758.101021655
#> 5   26.0241032217417 2308.21836388083 2417.28928017678    54626.1473345104
#> 6   13.0304857407584 4803.91185773496 4834.85238880645    127086.718292537
#> 7   7.87905127342976 16819.0702006291 17225.3548490629    349078.212095232
#> 8   14.7653696686963 5089.39095852429 5280.10127833234    113656.749383447
#> 9   5.30094805179523 7738.19000641513 7805.42990454737    217312.179831656
#> 10  1.28672614669399 19958.5782612143 20218.7361854308    469287.080532068
#> 11   10.367553250078 6693.22997556802 6904.37161580809    164135.248219765
#> 12  4.22577092696868 9173.08253154265 9207.79342124863    254129.362863092
#> 13  1.10409335242958 20266.4638423518 20605.9328931531    471921.403102281
#> 14   10.702839811692  6830.1428004929  7076.3236876123    169188.975989426
#> 15  4.61359566399789 8978.31712611592 9031.72845197022    257325.532680213
#> 16 0.884090937345817 20500.3029437474 20846.4884934189    494748.942033201
#> 17  9.16065104673458   6758.065494513  7045.6958487142     171317.15797425
#> 18   5.7377158976856 9306.66121083382  9333.5735334279    264691.032060523
#> 19  1.55947555428312 19678.4910500025  19890.767378455    480536.979467601
#> 20  10.1598389155997 6936.31945684528 7276.60915515207     172714.07921878
#> 21  6.12061251089277  9236.2154023603 9277.15738396391     259908.61489293
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84152.5479148172 87.1451317324858 96.5789135061109    14.2320458508073
#> 2      44219.859873037 29.1890310164961   38.34096315518    25.6916350297683
#> 3      67949.859079535 61.5004650830629 67.0877585886947    17.6857188769376
#> 4     153679.485103514 237.534449022605 247.143677563403    14.3726438779605
#> 5      59403.890383386 60.8439364111403 69.0774062528436    18.3538973376766
#> 6     134723.521777351 200.005124438549 210.194802085536    11.5911433624108
#> 7     373138.943267656 485.977512302838 528.485416099454    6.18579812733785
#> 8      125545.09329894 180.444823748725  218.56543389008    11.4770547588768
#> 9     226224.733833464 279.952884438921 299.339346297248    4.07240782299644
#> 10    495385.568522214 627.553320589756 744.123357993086 -0.0447008612224602
#> 11    172445.308358627 293.676543197117 352.431174298449    8.52977849209868
#> 12    275173.110293005 324.362648386672 344.461079422417     3.2264495993848
#> 13    506876.629740391 681.030631850576 760.624200307087  0.0955452478130956
#> 14    183672.701690721  307.86402849864 376.520178215903    8.50776518489489
#> 15    271833.675344422 309.005587095653 331.532762413017    2.24987690970351
#> 16    532861.542109976 737.157332467105 801.661123039335  -0.144930836825727
#> 17    181324.104295953 307.410639067039 365.857573951407    6.86187600070591
#> 18    277667.618634991  333.38053547204 363.115293527267    3.99989070191011
#> 19    519849.907453269 731.275856021955 826.659234078972  -0.302068629554488
#> 20      186174.5243328 324.293001265087 377.304022837584    8.78896375119329
#> 21    276060.722305648 325.143585279921 352.387488282327    4.19535490647638
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9290946493677 3361.50470419449 3452.91368473618    78035.0376406414
#> 2    36.602897825624 1402.90029312393 1510.29436935075     39757.810193827
#> 3   23.9784258167298  2179.8010510708  2205.9948258401    65031.9668411392
#> 4   15.5406343847444 7501.31343461381 7659.11596296841     146095.41531169
#> 5   26.6943153863687 2297.77540380994 2427.73224024767    54168.7038511074
#> 6   13.1562535213907 4800.94946646216 4837.81478007925    126355.534980161
#> 7   8.02700543182615 16780.1706066301 17264.2544430619    346774.525068084
#> 8   15.0526981559621 5071.13145981926 5298.36077703737     112518.50368941
#> 9   5.40829622712716   7731.752143828 7811.86776713449    216458.850193185
#> 10  1.40306442893912 19933.6695237893 20243.6449228558    466788.289128757
#> 11  10.5281355104839 6673.01428660886 6924.58730476725    163339.604163917
#> 12  4.31309026627213 9169.75914848569 9211.11680430558    252114.535981505
#> 13  1.19221891496888 20233.9614864241 20638.4352490809    468574.626083739
#> 14   10.894642448985 6806.57229002402 7099.89419808118    167802.236294621
#> 15  4.82013419592652 8973.20327576817 9036.84230231797    255936.455191087
#> 16 0.974005461302554 20467.1575187788 20879.6339183875    491099.863302446
#> 17  9.36151488570796 6730.52641804693 7073.23492518027    170359.046092385
#> 18  5.88956470119996 9304.08449909609 9336.15024516563    263448.592920414
#> 19  1.72213475481262 19658.1667206825 19911.0917077749    476772.975724293
#> 20  10.2796241241401 6903.73852828399 7309.19008371336    171425.313197225
#> 21  6.28883890351168 9232.29542539825 9281.07736092595    258362.136523627
#>    ci95.integral.bt.up
#> 1     84687.0876475121
#> 2     44609.7477090845
#> 3     68204.8205372589
#> 4      154342.17081348
#> 5     59861.3338667889
#> 6     135454.705089726
#> 7     375442.630294803
#> 8     126683.338992977
#> 9     227078.063471934
#> 10    497884.359925525
#> 11    173240.952414475
#> 12    277187.937174593
#> 13    510223.406758933
#> 14    185059.441385525
#> 15    273222.752833548
#> 16    536510.620840731
#> 17    182282.216177818
#> 18      278910.0577751
#> 19    523613.911196578
#> 20    187463.290354355
#> 21     277607.20067495