Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8919803076301  15.5371711843619
#> 2                 FALSE          0.75  31.549865379587  31.6168461022111
#> 3                 FALSE          0.75 64.9438400026954  20.6681990646001
#> 4                 FALSE          0.75 243.635224034923  15.0104803695125
#> 5                 FALSE          0.75 65.3142473692497   21.224741410984
#> 6                 FALSE          0.75 205.526830674366  12.4475498219542
#> 7                 FALSE          0.75 510.562399674141  7.17077433430381
#> 8                 FALSE          0.75 202.682908257938  13.5799374540986
#> 9                 FALSE          0.75 289.427325537429  4.70838592720971
#> 10                FALSE          0.75 679.594820878211  0.62445746862544
#> 11                FALSE          0.75 323.326516849153  9.89976879438175
#> 12                FALSE          0.75 334.098674975654  3.72571092751165
#> 13                FALSE          0.75 727.834477180297   0.6934253108145
#> 14                FALSE          0.75 344.956817111561  9.86527673067974
#> 15                FALSE          0.75 318.874047029542  3.50772356706906
#> 16                FALSE          0.75 781.484968540894 0.566571187604151
#> 17                FALSE          0.75 335.111602296707  8.12393909993182
#> 18                FALSE          0.75 350.667469539192  5.07242719623331
#> 19                FALSE          0.75 775.893131720148 0.660854384134927
#> 20                FALSE          0.75 347.708266209484  9.51390847179005
#> 21                FALSE          0.75 339.402916321247  5.16700891949799
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.00874543741 2120.99789239114  81567.229474202 1.79580652461582
#> 2  1464.37212027794 274.570160308108 42522.9107548973 1.43019682475352
#> 3  2197.63826813208 656.866431310269 66260.4163657481 1.69477978498503
#> 4  7547.33895318357 6029.83876849953 148184.838934246 2.42495290202226
#> 5  2360.25041982289 1145.51672369512 56851.3577896292 2.91240927855699
#> 6  4818.13846243443 3215.07660969148   130675.2873629 2.51211400117983
#> 7  17008.3993485449 15788.5555632589 360151.073964388 13.4194982693691
#> 8  5183.85905439165 3744.24846523589 119311.929031795 8.90323315911516
#> 9  7767.10544430576 6287.11908432608 221082.095919769 9.87474025271063
#> 10 20105.5158330884 18741.3233563165 484880.809309889 35.0629834243892
#> 11  6791.5445137902 5564.05907604569 167670.973883511 11.9987196103681
#> 12 9194.34653301576  7691.5324326669 267737.288833555 5.80527626189297
#> 13 20462.0801906624 19474.7770414721 492202.635115148 16.0403862988303
#> 14 6963.90797047914  5604.6865142383 177382.323629155 16.6264772012295
#> 15  9000.8611740402 7343.17142092012 264805.148716869 4.26840314939853
#> 16 20685.3186250571 19344.4981526917 514884.985628415 18.7340428574386
#> 17 6907.21128569209 6013.44066482301 176113.176503243 13.6312435599795
#> 18 9325.23845999575 7630.97381390273 270146.060978571 5.91850077394663
#> 19 19813.2513779644 18632.3054732791 501278.600310083 19.2452476894591
#> 20 7111.49019324832   5776.907373493 179492.391466727 16.5993067603633
#> 21 9257.33509216836 7492.48822487348 268757.506416303  7.3917858894945
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.770247049456615 17.3906315144141 71.1052947783679 1155.02554888347
#> 2    1.9265679249717 14.9544079366658 72.7921026835589 843.534612248689
#> 3   1.76993592040565 5.21090673684261  109.50152460004   1021.345598582
#> 4  0.366238868605176 44.9914683768702 85.0004632312661 2971.66998003701
#> 5   2.50848154553145 31.7604285075854  140.73964576998 1510.72742367923
#> 6  0.486482268023033 8.70803587400276 93.4296148714006 2020.90595273221
#> 7  0.432242817784891 274.914410549125 314.435266203568 14691.4546061287
#> 8  0.662035273326267 34.3534557047571 143.749849846732 1736.96596614663
#> 9  0.307034314710783 19.2937580586351 54.3010515873309 3053.68925667299
#> 10 0.517524919620392 96.8986595069088 328.668162506624  7569.1964921726
#> 11 0.838551416148442 92.1512043863415 277.116210340507 3653.76871591062
#> 12 0.321314393030857  12.366565209306 87.7794904850792 2828.49235470682
#> 13  0.25748331932096 60.4591412305614 210.671746337272 6076.80200209009
#> 14 0.758364041970535 55.2769637088209 188.081406271048 3304.95416786093
#> 15 0.506375092758412 16.6168543175247  138.47848042527 2640.89588197963
#> 16 0.333358994360738 100.124801001205 238.513899139276 8067.81554814949
#> 17 0.532439545323182 61.9490667239983 111.674232856036 2312.56378744275
#> 18 0.571355178863671 8.42494770658019 156.566447361301  4455.3861694043
#> 19   0.2914608082657 66.3994504746829 195.720884522039 8562.79369131956
#> 20  0.39149422533288 103.530955868457 115.154143399691 3193.70899480927
#> 21 0.420194270944165 6.35760609774763  123.75509795767 5264.37157302491
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.9378785746371 94.8460820406232   14.2701147880057
#> 2             TRUE 29.1971916028675 33.9025391563066   28.4476418656326
#> 3             TRUE  62.155927256395 67.7317527489957   17.7566544755328
#> 4             TRUE 239.646176511097  247.62427155875    14.408017430657
#> 5             TRUE 60.5233341060234 70.1051606324759   17.0982892685847
#> 6             TRUE 201.394403142425 209.659258206307   11.6472864910563
#> 7             TRUE 488.487325021029 532.637474327253   6.45973489904767
#> 8             TRUE 188.037089711194 217.328726804683   12.4908894294769
#> 9             TRUE  273.18337782172 305.671273253138   4.20331447951048
#> 10            TRUE 621.916213145091 737.273428611331 -0.226871024150105
#> 11            TRUE 303.588623090097 343.064410608208   8.52035171481756
#> 12            TRUE  324.54899552484 343.648354426468   3.19714875097589
#> 13            TRUE 701.448041718721 754.220912641873   0.26986525053152
#> 14            TRUE 317.606262115538 372.307372107583   8.61776788163822
#> 15            TRUE 311.852523848781 325.895570210302   2.67473653948147
#> 16            TRUE 750.667468040408 812.302469041381 0.0181956418807371
#> 17            TRUE 312.688206640541 357.534997952874   7.24807604787519
#> 18            TRUE 340.931535766049 360.403403312334   4.13254792700257
#> 19            TRUE 744.234699270987 807.551564169308  0.181401354537851
#> 20            TRUE 320.402406588686 375.014125830281   8.86990047111746
#> 21            TRUE 327.243428533028 351.562404109465   4.47578934379484
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.804227580718  3385.4011565962 3442.61633427863    79667.2124462887
#> 2   34.7860503387895 1439.77211922213 1488.97212133376    41135.2963177482
#> 3   23.5797436536674 2189.06632654998 2206.21020971419    64580.3028560807
#> 4    15.612943308368 7473.32798770361 7621.34991866352    143296.441817085
#> 5   25.3511935533832 2308.00451492792 2412.49632471787    54366.2111776769
#> 6   13.2478131528521 4803.81374342169 4832.46318144716    127350.897070655
#> 7   7.88181376955996 16556.1651431916 17460.6335538982    335983.631137306
#> 8   14.6689854787203 5127.34761975733 5240.37048902598    116454.620017484
#> 9   5.21345737490895 7735.36721229931 7798.84367631221    216058.777092542
#> 10  1.47578596140099 19946.1175381996 20264.9141279773    472429.481080265
#> 11  11.2791858739459 6639.95578257467 6943.13324500573    161660.524345838
#> 12  4.25427310404741 9174.00353324645 9214.68953278507    263084.418910062
#> 13  1.11698537109748 20362.6249033381 20561.5354779867     482206.29582171
#> 14  11.1127855797213 6872.97736517813 7054.83857578015    171945.674023024
#> 15  4.34071059465665 8973.52644868787 9028.19589939253    260460.874991012
#> 16  1.11494673332756 20520.6133274101 20850.0239227041    501613.429051709
#> 17  8.99980215198846 6805.30507093111 7009.11750045307    172309.009072899
#> 18  6.01230646546405 9311.37942101842 9339.09749897307    262816.950729901
#> 19    1.140307413732 19704.0242819336 19922.4784739953    487192.804687863
#> 20  10.1579164724626 6941.18177084471 7281.79861565193    174238.740170265
#> 21  5.85822849520114 9246.87683013756 9267.79335419915    260097.615178677
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83467.2465021154 88.3721995193831 95.4117610958772    14.0274869674269
#> 2     43910.5251920464 28.7466796030701 34.3530511561039    27.8407729692665
#> 3     67940.5298754155 61.6220716241247  68.265608381266     17.199124660605
#> 4     153073.236051406  238.88231634696 248.388131722887    14.2926521870464
#> 5     59336.5044015816  59.605925183278 71.0225695552214    16.3081175817423
#> 6     133999.677655144 200.603087232054 210.450574116679     11.494044576629
#> 7      384318.51679147 484.260183066177 536.864616282104    6.32357841144543
#> 8     122169.238046106 185.232571266073 220.133245249804    12.2823483183791
#> 9     226105.414746996 270.072834642116 308.781816432742    4.10659867037658
#> 10    497332.137539513 610.871373366408 748.318268390014  -0.389891373830529
#> 11    173681.423421184 299.809026412831 346.844007285474     8.2562080187308
#> 12    272390.158757048 322.720333502344 345.477016448965    3.09593471717117
#> 13    502198.974408586  696.39532003459 759.273634326004   0.188758004945418
#> 14    182818.973235286 312.368921797151  377.54471242597     8.3788832084175
#> 15    269149.422442725 310.507976856721 327.240117202363    2.51522838526257
#> 16    528156.542205121 744.766244540315 818.203692541474 -0.0868124413428952
#> 17    179917.343933586 308.394364919147 361.828839674267    7.08035759109839
#> 18    277475.171227241 339.067208022256 362.267731056127    3.95257104566052
#> 19    515364.395932304 738.172446248808 813.613817191488  0.0895911999341558
#> 20    184746.042763188 315.173624959172 380.242907459796    8.74657979013761
#> 21    277417.397653929 324.915015977838 353.890816664656    4.34342814844743
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0468554012968 3379.92310766916 3448.09438320567    79303.3793983904
#> 2   35.3929192351556 1435.06148072208 1493.68275983381    40869.5829148899
#> 3   24.1372734685952 2187.42489092787 2207.85164533629    64258.5789925274
#> 4   15.7283085519787  7459.1556751649 7635.52223120223    142360.365773373
#> 5   26.1413652402256 2297.99997994803 2422.50085969776     53890.332039218
#> 6   13.4010550672793 4801.07071212138 4835.20621274747    126714.311695545
#> 7    8.0179702571622 16469.5671038687 17547.2315932212    331355.822936376
#> 8   14.8775265898181 5116.52628121033 5251.19182757298    115907.475738148
#> 9   5.31017318404285 7729.28967851084 7804.92121010069     215096.86497669
#> 10  1.63880631108141 19915.5944604549  20295.437205722    470045.184185231
#> 11  11.5433295700327 6610.92815319297 6972.16087438743    160509.587200326
#> 12  4.35548713785213 9170.10806520552   9218.585000826     262193.44381833
#> 13  1.19809261668358 20343.5802738505 20580.5801074743    480292.103191051
#> 14   11.351670252942 6855.56512160985 7072.25081934843    170904.613460148
#> 15  4.50021874887554 8968.29213957785 9033.43020850255    259628.992788189
#> 16   1.2199548165512 20489.0740150948 20881.5632350195    499072.067154042
#> 17  9.16752060876526 6785.79111491306 7028.63145647113    171580.551479855
#> 18  6.19228334680611 9308.72556249085 9341.75135750065    261413.504086539
#> 19   1.2321175683357  19683.108455034 19943.3943008948    484495.524675097
#> 20  10.2812371534425 6908.56951974615  7314.4108667505      173232.7218369
#> 21  5.99058969054856 9244.87418421677 9269.79600011994    258439.338133174
#>    ci95.integral.bt.up
#> 1     83831.0795500136
#> 2     44176.2385949047
#> 3     68262.2537389689
#> 4     154009.312095118
#> 5     59812.3835400405
#> 6     134636.263030255
#> 7       388946.3249924
#> 8     122716.382325442
#> 9     227067.326862848
#> 10    499716.434434547
#> 11    174832.360566695
#> 12     273281.13384878
#> 13    504113.167039245
#> 14    183860.033798163
#> 15    269981.304645549
#> 16    530697.904102788
#> 17    180645.801526631
#> 18    278878.617870604
#> 19     518061.67594507
#> 20    185752.061096553
#> 21    279075.674699431