Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8273982157265  15.4701512495673
#> 2                 FALSE          0.75 32.3984064136218  30.9389651070999
#> 3                 FALSE          0.75 64.6847807880339  20.0531314350553
#> 4                 FALSE          0.75 242.846223125439  15.0369742230897
#> 5                 FALSE          0.75 65.2822924757835  22.3180842456036
#> 6                 FALSE          0.75 204.210645584497  12.2680250535649
#> 7                 FALSE          0.75 507.694909021724  6.94436045973639
#> 8                 FALSE          0.75 201.827718452685  13.2631034374695
#> 9                 FALSE          0.75 288.776261489024  4.54628152897749
#> 10                FALSE          0.75 670.680353709245 0.596054322030962
#> 11                FALSE          0.75 320.114758685482  9.67328765321278
#> 12                FALSE          0.75 334.802364135033  3.87206229162074
#> 13                FALSE          0.75 736.379442928143 0.854812949787861
#> 14                FALSE          0.75 342.631287217319  9.79256246008608
#> 15                FALSE          0.75 320.070417461823  3.44471552068869
#> 16                FALSE          0.75 774.033914093925 0.324119195246095
#> 17                FALSE          0.75 334.496115164791  8.23188168655317
#> 18                FALSE          0.75 353.252094364524  5.03494575038258
#> 19                FALSE          0.75  787.71450767113 0.767158436290572
#> 20                FALSE          0.75 355.794012689227   9.5888220948155
#> 21                FALSE          0.75 339.874354103982  5.35494690364009
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3415.16434619717 2136.15864473478 81414.0452614881 2.72437282435302
#> 2  1458.53536413536 278.818589799357 42500.4091315617 2.01820804221753
#> 3   2196.7998498976 687.421148262861 65766.3996552538 1.49767270650606
#> 4  7575.96869267508 6050.02363407739 150264.543735418 2.73888236921494
#> 5  2370.82033190749 1131.19352414774  57661.744228183  2.6028504261547
#> 6  4816.81556283886 3247.20340459819 130681.315557448 2.33732163953349
#> 7  16992.2202338604 15833.2086826737 359618.254515896 12.3595069592306
#> 8  5178.84889366504 3808.71013443846 118036.657089255 8.28397431894555
#> 9  7765.78854264317 6296.61694636712  221249.80538742 6.03746023865098
#> 10 20093.7539125006 18717.1253888488 487548.637095694 19.2444594685886
#> 11 6753.16357416317 5580.31952009726 166241.233294934 12.3478109474991
#> 12 9191.30012087952 7646.12194292132 266848.420606484 4.42540327284557
#> 13 20434.7435012262 19333.7619804787 492839.223773669  23.645623062365
#> 14 6972.13927753512 5640.42265393609 177196.409460389 17.1830635488974
#> 15 9001.30711226735 7375.54556680178 264832.021848876 4.12057077413477
#> 16 20663.2538598008 19445.2684965466 512328.288050602 24.7445735766414
#> 17 6902.58330327356 5979.87266239231 175575.425781351 15.8424140273807
#> 18 9323.28264182053 7671.89400414785 272123.917260282 6.80652386486165
#> 19 19819.4853359058 18604.2072887238 505942.279270053 27.6989696917511
#> 20 7134.65240039861 5822.33038068968 180124.836780722 11.3600110801676
#> 21 9255.80836011858 7439.96743682372  267706.33015081 9.37027649643248
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.749409579197673   23.35147696173 65.8751828287485 1595.13265469694
#> 2   2.37773667446135 17.7489412391476 71.4506120744351 774.548401040039
#> 3   1.64405616187761 5.12293829687156 108.066085454531 1480.49197010679
#> 4  0.310139950121719 33.4138321550024 78.2602695623082 1682.13720130169
#> 5   2.34929053620092 27.8552136931064 142.183784520592 1242.09628495823
#> 6  0.368376937629426 9.71821918022018 75.9151944639461 2584.59564039675
#> 7  0.430648183747367 143.926136201342 192.153583044515 7638.74985686646
#> 8  0.825912880807458  44.129292148308 156.528075849685 2265.06402124125
#> 9  0.403516453536967 22.0313491987366 69.8203712309798 3971.97689824526
#> 10 0.342774108801299 91.9201139671946 255.973560316927 5880.59913708169
#> 11 0.521511210909026 82.8444172557789 146.383031359595 2958.52867730567
#> 12 0.490440322188033 11.9527392657173 145.062854876712 3700.33536583029
#> 13 0.386805042968037 71.9380658246024 193.135518655302 6564.97512311197
#> 14 0.670283257000413 35.9824582997913 192.550903483744 1715.48021346064
#> 15 0.378760666392427 13.9636930438092 122.899368135322 4255.47468999503
#> 16 0.168356115720396 97.0451140927087 214.083358978756 8195.40915933721
#> 17 0.635334277103811 87.5442513352099 159.262204231553 3636.37962780197
#> 18 0.388843902124423 8.66641284050748 119.333872096133 2515.99406622126
#> 19 0.525898575428594 81.9411866605447 341.035308322779 6933.43195855988
#> 20 0.388787328013966 85.3480960769386 141.730022354934  3384.8679221906
#> 21 0.853575477650418  15.986028052189 212.510160107357 3854.21125982476
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.3458049196658 97.3089915117873    14.2373724917871
#> 2             TRUE  29.078454184174 35.7183586430697    27.0275882776109
#> 3             TRUE 62.2211091858315 67.1484523902364    17.3486590487666
#> 4             TRUE  238.34076162808 247.351684622797    14.5267940051394
#> 5             TRUE  61.000603524759  69.563981426808     18.453501313553
#> 6             TRUE 200.365751487464 208.055539681529    11.6620449911645
#> 7             TRUE  487.36352007379 528.026297969658    6.23594419747197
#> 8             TRUE  188.20058069802  215.45485620735    11.9044767485412
#> 9             TRUE 278.844639396443 298.707883581605    3.88249696290918
#> 10            TRUE 639.023217883417 702.337489535073  0.0321909130528262
#> 11            TRUE 299.802609676846 340.426907694118    8.81540171126743
#> 12            TRUE 327.522575751202 342.082152518864    3.06528796162142
#> 13            TRUE 697.482392990552 775.276492865733    0.21851865410544
#> 14            TRUE 314.365147679383 370.897426755255     8.6899465023204
#> 15            TRUE 313.292078538371 326.848756385275    2.82165422447315
#> 16            TRUE  733.32909056035 814.738737627501   0.047173384886043
#> 17            TRUE  308.43534408975 360.556886239832     7.1867568007174
#> 18            TRUE 342.055362606827 364.448826122222     4.3952975313879
#> 19            TRUE   742.1497025282 833.279312814061 -0.0979447202894653
#> 20            TRUE 337.106794462351 374.481230916103    8.94926694023252
#> 21            TRUE  324.46024926735 355.288458940613    3.95081524290516
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7029300073474 3376.75116659513 3453.57752579922    78790.0520445117
#> 2   34.8503419365888 1429.33835579697 1487.73237247376    41226.2770118509
#> 3   22.7576038213439 2188.37261639924 2205.22708339595    63330.9903644282
#> 4   15.5471544410399  7521.0029387801 7630.93444657006    147497.428039276
#> 5   26.1826671776541 2324.99850538233 2416.64215843265    55618.4958394267
#> 6   12.8740051159653 4800.82909228739 4832.80203339032    126429.655728996
#> 7   7.65277672200081 16755.4617398092 17228.9787279116    347052.511001351
#> 8   14.6217301263978 5106.25620808107 5251.44157924901    114310.626774313
#> 9    5.2100660950458 7729.54697321124 7802.03011207509    214715.903389807
#> 10   1.1599177310091 19942.5453250245 20244.9624999766    477875.051515195
#> 11  10.5311735951581 6616.88450777741 6889.44264054892    161374.453620767
#> 12  4.67883662162005 9171.63786478742 9210.96237697163    260761.368929693
#> 13  1.49110724547028 20316.4053829448 20553.0816195077     482039.83969615
#> 14  10.8951784178518 6912.94813363196 7031.33042143827    174374.444509246
#> 15  4.06777681690423 8978.33683721029 9024.27738732442    257831.765983834
#> 16 0.601065005606148 20503.6146471183 20822.8930724833    498846.839983493
#> 17  9.27700657238894 6758.57300982714 7046.59359671998    169593.581293617
#> 18  5.67459396937725 9309.02639269789 9337.53889094316    267985.107021348
#> 19  1.63226159287061 19684.6920838492 19954.2785879624    494536.783698222
#> 20  10.2283772493985 6994.25478235204 7275.05001844517    174556.729048718
#> 21  6.75907856437503 9229.51134397273 9282.10537626443    261366.152628398
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84038.0384784646 87.4876274799946 98.1671689514585     14.0013084743398
#> 2     43774.5412512726 28.4427186508755 36.3540941763682     26.2786012251556
#> 3     68201.8089460795 61.7493422832821 67.6202192927858     16.8307813577751
#> 4     153031.659431559 237.478013681777   248.2144325691     14.4290999208511
#> 5     59704.9926169393 60.1807056405203 70.3838793110467     17.7134747946498
#> 6     134932.975385901 199.629495171011 208.791795997982     11.5460062558112
#> 7     372183.998030441 483.470275381632 531.919542661816     6.10029001959155
#> 8     121762.687404197 185.591128787552 218.064308117818     11.6443141910869
#> 9     227783.707385034 276.942839421268  300.60968355678     3.75538928004503
#> 10    497222.222676194 632.961213150811 708.399494267679  -0.0757829312195829
#> 11    171108.012969102 295.913049228384  344.31646814258     8.65112567983109
#> 12    272935.472283275 326.128573720256 343.476154549811     2.91079926013219
#> 13    503638.607851189 690.034021725907 782.724864130378   0.0966750655705085
#> 14    180018.374411532  308.95248266148 376.310091773158     8.47880727636527
#> 15    271832.277713918 311.994098744519 328.146736179127     2.70234461455953
#> 16    525809.736117712 725.534549883708 822.533278304143 -0.00585879156588187
#> 17    181557.270269085 303.444983671125 365.547246658457      6.9866265034297
#> 18    276262.727499216 339.911307589395 366.592881139653     4.27281170221871
#> 19    517347.774841884 733.424527075298 842.004488266963   -0.263602771549472
#> 20    185692.944512726 333.528390972099 378.059634406355     8.82679893190812
#> 21    274046.507673221 321.508612170974 358.240096036989     3.68193896744527
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9389940247947 3369.39545135218 3460.93324104217    78287.5852582821
#> 2   35.5993289890441 1423.74743930663 1493.32328896409    40982.2942655232
#> 3   23.2754815123354 2186.75889083573 2206.84080895947    62864.6353938445
#> 4   15.6448485253282 7510.47758165128 7641.45980369888    146967.554820866
#> 5   26.9226936965574   2316.224113069 2425.41655074597    55227.2355096648
#> 6   12.9900438513186 4797.76785324562 4835.86327243209    125615.508102271
#> 7   7.78843089988123 16710.1250069058  17274.315460815    344646.304796438
#> 8   14.8818926838521 5092.35548105436 5265.34230627572    113597.131607622
#> 9   5.33717377790994 7722.60709821364 7808.96998707269     213464.73066686
#> 10  1.26789157528151 19913.5904891249 20273.9173358763    476022.662787014
#> 11  10.6954496265945 6590.78851634184 6915.53863198449    160442.517087415
#> 12  4.83332532310928 9167.87275191872 9214.72748984033    259595.763289456
#> 13  1.61295083400521   20293.74489221 20575.7421102424     479971.87253237
#> 14  11.1063176438069 6901.61365926753 7042.66489580271    173834.068242006
#> 15  4.18708642681785 8973.93827390149 9028.67595063322    256491.291456486
#> 16 0.654097182058072 20473.0454361791 20853.4622834225    496265.286098301
#> 17  9.47713686967664 6730.99657065655 7074.17003589057    168448.121710859
#> 18  5.79707979854644 9306.29647265313 9340.26881098792    267192.568890489
#> 19  1.79791964413062 19658.8806100512 19980.0900617605    492352.752631276
#> 20  10.3508452577229 6967.37013208781 7301.93466870941    173490.495653228
#> 21  7.02795483983491 9224.47574513629 9287.14097510087    260152.076081553
#>    ci95.integral.bt.up
#> 1     84540.5052646941
#> 2     44018.5239976002
#> 3     68668.1639166631
#> 4     153561.532649969
#> 5     60096.2529467011
#> 6     135747.123012626
#> 7     374590.204235354
#> 8     122476.182570888
#> 9     229034.880107981
#> 10    499074.611404374
#> 11    172039.949502453
#> 12    274101.077923511
#> 13    505706.575014969
#> 14    180558.750678772
#> 15    273172.752241267
#> 16    528391.290002903
#> 17    182702.729851843
#> 18    277055.265630076
#> 19    519531.805908831
#> 20    186759.177908216
#> 21    275260.584220066