Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2116913748967  15.3818689621358
#> 2                 FALSE          0.75 32.5359607532792   32.954378533242
#> 3                 FALSE          0.75 64.4416950649734  20.1019446677126
#> 4                 FALSE          0.75 241.817461654022  15.0823374951427
#> 5                 FALSE          0.75 65.7525337231776  22.6430379031423
#> 6                 FALSE          0.75 206.496730478843  12.4106491122695
#> 7                 FALSE          0.75 507.283506708382  6.98866398944651
#> 8                 FALSE          0.75 203.270093276092  13.3775065830596
#> 9                 FALSE          0.75 289.831018345361  4.71012943912245
#> 10                FALSE          0.75 669.849170880801 0.583394628134013
#> 11                FALSE          0.75 326.893050659234  9.88328193581883
#> 12                FALSE          0.75 335.957221538554  3.79341632136829
#> 13                FALSE          0.75 726.147707179302 0.760185835713928
#> 14                FALSE          0.75  342.87102335172  9.87042830290654
#> 15                FALSE          0.75 319.464952991597   3.5243493146553
#> 16                FALSE          0.75 771.009876196366  0.47244978638248
#> 17                FALSE          0.75 337.798656890808  8.07516095694729
#> 18                FALSE          0.75 350.212304220307  4.98016114420482
#> 19                FALSE          0.75 779.808857785397  0.76645311756426
#> 20                FALSE          0.75  351.15449645586   9.4745195028475
#> 21                FALSE          0.75 337.948174380548  5.24631769920762
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.73527618929 2136.08140959435 81735.7440325057 2.14415074919216
#> 2  1466.44769646827  253.61372889529 42700.2921905527 1.77401550118396
#> 3  2197.32474278183 697.099303711382 65669.0221558521 1.56421764903263
#> 4  7557.88171846014 6007.53574464902 149281.292209533 2.41235663464178
#> 5  2365.76441708295 1097.64010434789 56952.2341179624 2.34922569745422
#> 6  4821.68482526439 3237.03269943837 130941.407954426 2.19386086728064
#> 7  17012.5748014232 15820.5568827707 361195.588305575 10.3354241776194
#> 8  5185.13663082398 3788.50317612178 119165.974534911 7.37844159679157
#> 9  7767.15844280545 6276.41972802362 221639.368356609 6.41736998316528
#> 10 20064.0282349572 18686.9124730276  486789.00560872 34.0495455195798
#> 11 6759.19507594763 5551.12089662936 166951.077172668 12.1016673307137
#> 12 9186.73371419901 7683.44708359175 266043.270817562 4.93642290934432
#> 13 20406.9778096487 19344.2826213265 487869.758922267  19.856411125543
#> 14 6978.21911028982 5648.43062806311 177084.991552125 10.7085995353761
#> 15 9006.11322603729 7352.48587954809 262350.027097626 6.32991244663372
#> 16 20634.3650922059 19312.9951790746 513199.563634335 20.1559435480126
#> 17  6866.4161273346 5982.59658680038 175346.130010907 14.8001200569845
#> 18 9319.51244193731 7649.52264511683 270701.963924708 6.28842735416713
#> 19 19786.6660507463 18536.7407427591  499457.99168504 26.9854434586411
#> 20 7103.98018237525 5792.43486819718 178634.708475244 11.8332929287206
#> 21 9254.71279902983   7453.567816063 267860.609924889 5.51583606393384
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.643184010549031 20.9923919083989 52.7278752678955  1244.3052873635
#> 2   2.97982256540641 21.3415611301209 87.4799290592906 837.472751084079
#> 3   1.77517215926774 5.05999840312198 131.020595213446  1457.0466129212
#> 4  0.437649314405946 69.7665682389605 137.883253098257 3477.76670391504
#> 5   2.33031050852063 32.7426330662641 144.609250741051 1402.87526495088
#> 6  0.454397562053362 8.87223835191084 94.6708125478032 2063.56196202532
#> 7  0.372885732728176 154.961662617457 179.249525115205 7212.04544800265
#> 8  0.597823435425602 49.0298243867738 130.747782865237 2163.39834681183
#> 9  0.388220826005423 16.0464305780367 66.4545694386999 3269.12369697807
#> 10 0.492615014936641 84.6890975335536 292.440517223125 6286.60488693389
#> 11 0.702521055704354 111.134361818978 219.787783127219 4091.68583877563
#> 12 0.267786403275561 10.7628992031628 65.0561982612229 6427.20783606031
#> 13 0.292681148149484 154.962627101329 270.246921351577 15854.1274297478
#> 14 0.672490989173861 29.7413621752439 223.265893376514 2273.71294898136
#> 15 0.506882544893551 13.2342823609509 133.356709626685 8296.58072918879
#> 16 0.368355795657449 72.6647176578779  317.10912474906 8513.79657173232
#> 17 0.542361423253821 78.4087410873391 151.914925206305 2997.69019772525
#> 18 0.349631299267287 10.8301867578475 79.8034231404566 3834.18791962738
#> 19 0.389478311351219  99.808659860739  297.47205514942 15680.7633642776
#> 20 0.338339258933268 161.060090902619 187.816569728357 6979.48384084267
#> 21 0.627723217600585 10.6448934317583 201.024959913998 4681.91288599411
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.6845633924756 95.7388193573179   14.3238312647827
#> 2             TRUE 29.6177052538315 35.4542162527268   28.0525704131484
#> 3             TRUE 61.8685570323147 67.0148330976321   17.1817864657172
#> 4             TRUE 237.849134990036 245.785788318007    14.362404372945
#> 5             TRUE 61.8880574508654 69.6170099954898   18.8096771166259
#> 6             TRUE 202.887829352167  210.10563160552   11.6631651226918
#> 7             TRUE 490.281733936198 524.285279480566   6.37526695910866
#> 8             TRUE  191.13255684937 215.407629702815   12.3940870317845
#> 9             TRUE 279.274444723054 300.387591967668   4.07150618034353
#> 10            TRUE 613.837668501092  725.86067326051 -0.226957071436761
#> 11            TRUE  306.98580790021 346.800293418258   8.72763479918516
#> 12            TRUE 327.836805852683 344.077637224426      3.35290768798
#> 13            TRUE 693.483910877784  758.81150348082  0.278725347008026
#> 14            TRUE 325.255377116026 360.486669587413   8.76418062571554
#> 15            TRUE 309.052247016884 329.877658966309   2.69052752830541
#> 16            TRUE 737.853349059885 804.166403332847 -0.133495497474024
#> 17            TRUE 313.452459397068 362.144854384547   7.18297641569475
#> 18            TRUE 339.867841222703 360.556767217912   4.40501765691013
#> 19            TRUE 735.417803295932 824.199912274861  0.125761295391505
#> 20            TRUE 331.688729588115 370.620263323605   8.91795142190228
#> 21            TRUE 328.874624055377  347.02172470572   4.21371300625465
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.439906659489 3380.20279149997  3449.2677608786    79688.8618347927
#> 2   37.8561866533355 1431.34082840922 1501.55456452732    41322.6495150194
#> 3    23.022102869708 2189.00104540869 2205.64844015496    63272.1804775967
#> 4   15.8022706173405 7443.11571370705 7672.64772321323    143560.365981593
#> 5   26.4763986896587 2311.90278568894 2419.62604847695    54644.5043071182
#> 6   13.1581331018473 4807.08999317549 4836.27965735328    127546.848526894
#> 7   7.60206101978436 16757.6628664174 17267.4867364289    349331.773543611
#> 8   14.3609261343347 5104.48256970774 5265.79069194023    115607.184254406
#> 9   5.34875269790138 7740.76206450458 7793.55482110632     216261.65987508
#> 10  1.39374632770479 19924.7146695145 20203.3418003999    476447.540569714
#> 11  11.0389290724525 6576.37905075541 6942.01110113985    160220.253967882
#> 12  4.23392495475659  9169.0287450098 9204.43868338821    255470.513927243
#> 13  1.24164632441983 20152.0642880671 20661.8913312304    461789.719300332
#> 14  10.9766759800975 6929.29456951154 7027.14365106809    173344.733751051
#> 15   4.3581711010052 8984.34283155352 9027.88362052105    248702.151798111
#> 16  1.07839507023898 20514.8316316587 20753.8985527531    499194.368273835
#> 17  8.96734549819982 6737.43374824593 6995.39850642327    170414.929635649
#> 18  5.55530463149951 9301.69678472065 9337.32809915397    264394.724796921
#> 19  1.40714493973701 19622.4808052754 19950.8512962173    473663.135950804
#> 20  10.0310875837927 6839.03633284044 7368.92403191006    167153.457557058
#> 21  6.27892239216058 9237.20194933459 9272.22364872507    260158.863227429
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83782.6262302187 88.0091559064801 96.4142268433134   14.1212283014597
#> 2      44077.934866086 29.0588903709586 36.0130311355997   27.1139263050454
#> 3     68065.8638341074 61.3758284728694 67.5075616570774   16.6226072355478
#> 4     155002.218437474 237.089242650124 246.545680657919   14.2245448389071
#> 5     59259.9639288066 61.1480513561673 70.3570160901879   18.0756293064419
#> 6     134335.967381957 202.196763178973 210.796697778713    11.520029890645
#> 7     373059.403067539 487.026075320248 527.540938096516   6.25780795329929
#> 8     122724.764815417 188.808347746381 217.731838805804   12.2057726496254
#> 9     227017.076838138 277.252973178357 302.409063512365   3.94921662015183
#> 10    497130.470647726 603.112061662425 736.586280099177 -0.382130801141803
#> 11    173681.900377454 303.173782691035 350.612318627432   8.50634066663829
#> 12    276616.027707882  326.28183263624 345.632610440869    3.2685549709482
#> 13    513949.798544203 687.229141373238 765.066272985366  0.186530785340938
#> 14      180825.2493532 321.882168262383 363.859878441057   8.55234596412577
#> 15    275997.902397142 307.058324596195 331.871581386999   2.53085952666394
#> 16    527204.758994835 731.504226842261 810.515525550471  -0.24952757310612
#> 17    180277.330386165 308.790421579118 366.806892202497    7.0121325673698
#> 18    277009.203052495  337.88698660614 362.537621834475   4.29488379764094
#> 19    525252.847419277  726.91738860646 832.700326964333 0.0030756273158713
#> 20    190115.959393431 327.961242315568 374.347750596152    8.8113745553383
#> 21     275562.35662235 327.137135695238 348.759213065859   4.01598019271047
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6425096228119 3373.59018804883 3455.88036432975    79296.9056692732
#> 2   38.7948307614385 1424.61823665323 1508.27715628331    41058.8455984279
#> 3   23.5812820998774 2187.40714591171 2207.24233965195    62813.2107945265
#> 4   15.9401301513784 7421.13924471178  7694.6241922085     142464.86946986
#> 5   27.2104464998427 2301.58885627307 2429.93997789282    54202.5985986587
#> 6   13.3012683338941 4804.29523809464 4839.07441243413    126896.826508856
#> 7   7.71952002559374 16708.8499426929 17316.2996601534     347059.97922749
#> 8   14.5492405164937 5089.03817502591 5281.23508662206     114925.71377516
#> 9   5.47104225809308  7735.7074388725  7798.6094467384    215231.885910532
#> 10  1.54892005740983 19898.0376037914 20230.0188661229     474467.26003033
#> 11  11.2602232049994 6541.37172678243 6977.01842511283    158931.372928668
#> 12  4.31827767178839 9165.63843176081  9207.8289966372    253445.943458884
#> 13  1.33384088608692 20103.2510605301 20710.7045587673    456795.669159962
#> 14  11.1885106416873 6919.92604042634 7036.51218015329    172628.514172122
#> 15  4.51783910264666 8980.17403260982 9032.05241946475    246088.728868416
#> 16  1.19442714587108 20491.9422455965 20776.7879388154     496512.52235374
#> 17  9.13818934652478 6712.73499480342 7020.09725986579    169470.657223365
#> 18   5.6654384907687 9298.28527589193 9340.73960798269    263186.955602238
#> 19  1.52983060781265 19591.0410774193 19982.2910240734    468723.695491056
#> 20  10.1376644503567 6788.30240420611 7419.65796054438    164954.920147193
#> 21  6.47665520570476 9233.84880790358 9275.57679015607    258684.060668341
#>    ci95.integral.bt.up
#> 1     84174.5823957382
#> 2     44341.7387826775
#> 3     68524.8335171776
#> 4     156097.714949207
#> 5     59701.8696372661
#> 6     134985.989399995
#> 7      375331.19738366
#> 8     123406.235294662
#> 9     228046.850802686
#> 10     499110.75118711
#> 11    174970.781416668
#> 12    278640.598176241
#> 13    518943.848684573
#> 14    181541.468932129
#> 15    278611.325326836
#> 16     529886.60491493
#> 17    181221.602798448
#> 18    278216.972247178
#> 19    530192.287879024
#> 20    192314.496803296
#> 21    277037.159181438