Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2228825661817  15.6366941846254
#> 2                 FALSE          0.75 33.2355583920808  31.3900477590536
#> 3                 FALSE          0.75 65.0070558650204  20.3255908702362
#> 4                 FALSE          0.75 242.214898054442  15.0019707898619
#> 5                 FALSE          0.75  66.039902779214  21.6303763664423
#> 6                 FALSE          0.75 205.565850057535  12.3433281270225
#> 7                 FALSE          0.75 511.570874969784  7.16838033791017
#> 8                 FALSE          0.75 203.309514122811  13.2173114510051
#> 9                 FALSE          0.75 289.575453364985   4.7450203158189
#> 10                FALSE          0.75 673.632751890702 0.602626834166842
#> 11                FALSE          0.75 322.953477390413  9.93653539371711
#> 12                FALSE          0.75 333.789818847861  3.76272639773158
#> 13                FALSE          0.75 734.255811122036 0.788168016104145
#> 14                FALSE          0.75 340.421515494784  9.54690064844286
#> 15                FALSE          0.75 321.593442952618  3.49972613992735
#> 16                FALSE          0.75 776.017080173714 0.489587386369803
#> 17                FALSE          0.75 330.731565936872  7.96734830104239
#> 18                FALSE          0.75 350.566984111616  5.05260394206602
#> 19                FALSE          0.75 781.721387289301 0.560439326496302
#> 20                FALSE          0.75 347.186742577837    9.620060717127
#> 21                FALSE          0.75 339.219220989678  5.47656355607466
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3410.85713240142 2109.32822026219 81302.1280757426 2.38183517395174
#> 2  1464.56986923937 279.371472487031 42536.9674341542  2.6067182888861
#> 3  2195.10793859589 675.499536524577 65770.0911737276 1.77984282184487
#> 4  7540.77144862331 6011.92001557134 147981.353130003 2.31209537981998
#> 5  2362.31157294487 1154.21480139965  57128.390841614 2.48388495485319
#> 6  4819.80687234215  3243.0284037931  130410.94555838 3.04660926440059
#> 7  17087.3270516566  15875.383259309 364121.773665205 15.4723707344577
#> 8  5174.75249977169 3804.63271706016 118116.300847915 5.99111911283227
#> 9   7763.9693074037 6267.02664834526 222054.059100627 5.37101628572123
#> 10 20086.2312680666 18740.4230426922 486109.135541598 21.7599647014472
#> 11 6784.79149804719 5533.13054615539  167523.48710332 12.7994816279383
#> 12 9190.56538870197 7665.83547397768 266013.554020919 7.14284817549055
#> 13 20452.0371164389 19428.8844062299 488568.122817008 18.2467088502535
#> 14 6955.79221675736  5702.1194253451 176555.893336734 9.98776070282402
#> 15 8995.81165665105 7370.29113460311 264604.160480736 5.18147630195215
#> 16 20697.7496158531 19393.9921250538 516575.072746101 18.0787904267145
#> 17  6887.1622330667 6017.71523132269 175149.028196066 15.6861873571993
#> 18 9321.41434598771 7638.22229904588 270969.673855393 6.56942540788413
#> 19 19774.4018847819  18678.752044656 506599.647856901 14.9398285478143
#> 20 7093.49059763706 5709.32647413072 178706.560330582 15.6247746171132
#> 21 9256.65334931132 7392.87249480272 268697.933809971 8.58439883839035
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.803030143451484 21.0583783950002 61.8727875469232 1337.08404115458
#> 2   3.11941565785272  21.445482522815 108.813169030204 710.994463721197
#> 3   1.64028808599836  4.9545988123284 98.4349166271327 1402.88556543971
#> 4  0.237529476822022 69.6407506599175 72.0373122164509 3639.48063210505
#> 5   2.79090777951323 30.9134615740436 140.996311851941 1356.15104419988
#> 6   0.49102440610284 10.7049111615204 106.803805143423 1904.84220376873
#> 7  0.544891574084032  81.374254853381   114.3305642002 4885.03772606847
#> 8  0.718240748285364 41.0016843504178 147.503615031878 2544.92622764281
#> 9  0.307597979254765 21.1979537524455  64.021390976529 2531.37458189919
#> 10 0.334591466614955 71.1782711926968 182.485109408749 7912.01963024568
#> 11 0.720880079357332 107.089358733048  237.18053038761 4155.75267982234
#> 12 0.492357364860583 13.4158323540329 95.1280135680145 3644.11103008735
#> 13 0.392848501643831 150.312256881946  358.25945865317  13580.113736439
#> 14 0.473904223268596 54.0938823603072 132.317861175508 3141.91437082773
#> 15 0.511382034584772 28.3925229293243 126.065711866279 3280.72440726004
#> 16 0.298639401235517 80.9655362166621 222.370022812161 6089.92942776486
#> 17 0.498007228125891 94.2208173080388 186.647255335714 3157.44739076458
#> 18 0.301768948503432 8.73180293915736 83.0694279885436 4573.98312075899
#> 19 0.248000932417967 77.3709796266247 183.386569706258 7312.29916883882
#> 20  0.44115764942593 95.1058351175442 134.555710878857 3801.66789405306
#> 21 0.752706805812524 7.95839553541475 211.206260350863  3289.6033749048
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE  88.304763705031 96.1410014273323     14.3157095986477
#> 2             TRUE 28.9475068068632 37.5236099772984     26.2586090018858
#> 3             TRUE 62.0792144230856 67.9348973069552     17.6273169687689
#> 4             TRUE 238.411501154638 246.018294954246     14.6112348004897
#> 5             TRUE 61.9539120284805 70.1258935299475      17.039333069143
#> 6             TRUE 200.554177817596 210.577522297474     11.5355929789834
#> 7             TRUE 486.118825111601 537.022924827967     6.27203369854194
#> 8             TRUE 193.454123182202  213.16490506342     12.0358054200756
#> 9             TRUE 280.740131574973 298.410775154996     4.23902163994482
#> 10            TRUE 637.837609956821 709.427893824583   0.0522238715852417
#> 11            TRUE 301.898330112455 344.008624668372      8.7506876631743
#> 12            TRUE 322.039833599179 345.539804096543     2.95279853253592
#> 13            TRUE 704.239975063369 764.271647180703    0.141932230900042
#> 14            TRUE 323.991649138639  356.85138185093     8.76732820116602
#> 15            TRUE 313.069914435906 330.116971469329      2.6585026930354
#> 16            TRUE 746.277469921768 805.756690425659 -0.00167442866262202
#> 17            TRUE 304.927787734279 356.535344139465      7.1481264107753
#> 18            TRUE 339.760279315647 361.373688907586     4.55619402177788
#> 19            TRUE 757.145369328146 806.297405250456    0.152477792668746
#> 20            TRUE 321.483988332686 372.889496822988     8.89435638382135
#> 21            TRUE 325.097884900526  353.34055707883     4.23836086051306
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.9576787706031 3376.21609994165  3445.4981648612    79102.6248280433
#> 2   36.5214865162213 1429.29205048934  1499.8476879894    41367.3815413328
#> 3   23.0238647717035 2186.95762354961 2203.25825364217    63462.3444185793
#> 4   15.3927067792341 7426.21241378775 7655.33048345888     141994.40749019
#> 5   26.2214196637415 2311.45892865557 2413.16421723417    54897.5223739052
#> 6   13.1510632750617 4802.19729348145 4837.41645120285     127277.48013318
#> 7   8.06472697727841 16953.4664024228 17221.1877008905    356085.886605823
#> 8   14.3988174819345 5107.30472901525 5242.20027052812    113929.897203442
#> 9   5.25101899169299 7729.09867348093 7798.83994132647    217889.947913402
#> 10  1.15302979674844 19969.1430119546 20203.3195241786    473093.863249844
#> 11  11.1223831242599 6608.62950293132 6960.95349316305    160687.273945013
#> 12  4.57265426292724 9168.49634447959 9212.63443292435    260018.991376425
#> 13  1.43440380130825 20204.7734538681 20699.3007790097    466228.835720566
#> 14  10.3264730957197 6866.80778027466 7044.77665324007    171387.444196723
#> 15   4.3409495868193 8949.10595643232 9042.51735686979    259207.368830793
#> 16 0.980849201402228 20564.5613087767 20830.9379229295    506557.138837427
#> 17  8.78657019130948 6732.16898859497 7042.15547753842    169955.027238258
#> 18  5.54901386235417  9307.0505301528 9335.77816182262    263445.471621745
#> 19 0.968400860323858 19647.1266232961 19901.6771462677    494570.915724162
#> 20  10.3457650504327  6937.0414988687 7249.93969640542    172452.816644865
#> 21  6.71476625163626 9243.56178865556 9269.74490996708    263286.536258252
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83501.6313234419 87.5544856252363 96.8912795071271    14.0627551034605
#> 2     43706.5533269756 28.1263905458641 38.3447262382976    25.2759930696622
#> 3     68077.8379288759 61.5185639342044 68.4955477958363    17.1106262216794
#> 4     153968.298769815 237.683191109995 246.746604998889    14.5364130152907
#> 5     59359.2593093228 61.1714882677017 70.9083172907262    16.1601971185963
#> 6     133544.410983579  199.59449589931  211.53720421576     11.380920291061
#> 7     372157.660724588 481.245028330247 541.896721609322    6.10039285270547
#> 8     122302.704492387  191.56692066166 215.052107583962    11.8095595843658
#> 9     226218.170287851 279.048261444971 300.102645284998    4.14212827647956
#> 10    499124.407833352 630.983221075866 716.282282705539  -0.053172440398469
#> 11    174359.700261628 297.866493399654 348.040461381172    8.52361043817674
#> 12    272008.116665412   319.7898364239 347.789801271823    2.79770596260484
#> 13     510907.40991345 698.492261775539 770.019360468533  0.0181849528822351
#> 14    181724.342476746 320.845504517249 359.997526472319    8.61804837083642
#> 15    270000.952130679 311.437749400791 331.749136504444     2.4974173521412
#> 16    526593.006654774 740.582650937353 811.451509410074 -0.0957458400518097
#> 17    180343.029153873 299.986638716761 361.476493156983    6.99125413391564
#> 18    278493.876089042 337.690910312163 363.443057911069     4.4611368029993
#> 19    518628.379989641 752.439323335585 811.003451243017   0.074357498957086
#> 20      184960.3040163 316.562184328295 377.811300827379    8.75539172425218
#> 21    274109.331361689 322.393799266433 356.044642712923    4.00125821668211
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.2106332657903 3369.58271074722 3452.13155405562    78681.4433550796
#> 2   37.5041024484449 1422.53672349466 1506.60301498409    41143.4182852607
#> 3   23.5405555187929 2185.39692492373 2204.81895226806    63020.4354654658
#> 4   15.4675285644331 7404.27557732987 7677.26731991675    140847.971091077
#> 5   27.1005556142882 2301.72118825974    2422.90195763    54470.3347949822
#> 6   13.3057359629841 4798.82524646557 4840.78849821873    126677.454838993
#> 7   8.23636782311488  16927.833512144 17246.8205911693    354547.099722111
#> 8   14.6250633176444 5094.38919844487 5255.11580109851    113128.245441735
#> 9   5.34791235515824  7722.4213180489 7805.51729675849    217092.564920104
#> 10  1.25842610873215 19946.7218565289 20225.7406796043    470601.577066316
#> 11  11.3494603492575 6574.89635493041 6994.68664116396    159378.211850869
#> 12  4.72774683285833 9164.27035728807 9216.86042011588    258871.096401947
#> 13  1.55815107932605 20157.4250929503 20746.6491399275    461951.099893588
#> 14  10.4757529260493 6849.76820733116 7061.81622618356    170397.741169912
#> 15   4.5020349277135 8940.16231170958 9051.46100159253    258173.940642506
#> 16  1.07492061279142 20539.0571648684 20856.4420668377    504638.811067681
#> 17  8.94344246816913 6702.48943114294 7071.83503499045    168960.431310167
#> 18  5.64407108113275 9304.30001222696 9338.52867974846    262004.666938705
#> 19  1.04652115403552 19622.7547647137 19926.0490048501    492267.541485977
#> 20  10.4847297100018 6907.08316080668 7279.89803446745    171255.291258238
#> 21   6.9518688954672  9241.0548940619 9272.25180456073    262250.311195157
#>    ci95.integral.bt.up
#> 1     83922.8127964056
#> 2     43930.5165830477
#> 3     68519.7468819894
#> 4     155114.735168928
#> 5     59786.4468882458
#> 6     134144.436277766
#> 7       373696.4476083
#> 8     123104.356254094
#> 9     227015.553281149
#> 10     501616.69401688
#> 11    175668.762355772
#> 12     273156.01163989
#> 13    515185.145740429
#> 14    182714.045503557
#> 15    271034.380318966
#> 16     528511.33442452
#> 17    181337.625081964
#> 18    279934.680772081
#> 19    520931.754227825
#> 20    186157.829402926
#> 21    275145.556424784