Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1215663539382   15.613098933741
#> 2                 FALSE          0.75 34.0059503250876  31.0402111989617
#> 3                 FALSE          0.75 64.8584686861083   20.130278182232
#> 4                 FALSE          0.75 243.159622563938  14.9913620546507
#> 5                 FALSE          0.75 66.4345140038358  22.4361634379728
#> 6                 FALSE          0.75 205.516056477544  12.5479685347606
#> 7                 FALSE          0.75 505.113554668054  6.99599603863004
#> 8                 FALSE          0.75 203.089872207643  13.3272759090404
#> 9                 FALSE          0.75 290.458428146749  4.62559437261585
#> 10                FALSE          0.75 672.674283172791 0.770340399491107
#> 11                FALSE          0.75 328.557673562033  10.0009769249784
#> 12                FALSE          0.75 335.827842894845  3.95801502840134
#> 13                FALSE          0.75 731.661542084829 0.786132811802124
#> 14                FALSE          0.75 342.908335168782  9.49360456062447
#> 15                FALSE          0.75 320.239775246782  3.42261081986715
#> 16                FALSE          0.75 777.357114538266 0.575076038782317
#> 17                FALSE          0.75 333.958851530863  8.09157008657096
#> 18                FALSE          0.75 350.717671198219  5.03667038945991
#> 19                FALSE          0.75 789.123453954429 0.807989884566651
#> 20                FALSE          0.75 348.665483017703  9.57556235872938
#> 21                FALSE          0.75 340.521884172532  5.33111518782689
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.06036207955 2108.84533706407 81279.8995751513 2.30016196973166
#> 2  1474.25149718717 310.431690756494 42795.2048240137 2.47817622621173
#> 3  2195.74507477131 686.434872190217  65687.812832576 1.44013131468832
#> 4  7576.20734942565 6058.41075000321  150208.40177621 2.33812612556452
#> 5  2363.69956383044 1132.23903890907 57059.1128264473  1.5323476523641
#> 6  4821.04805230099 3202.06705715634 131088.525020289 2.83283572705057
#> 7  16917.1921639388 15703.9011754889 356110.527450842 11.3464064056193
#> 8  5190.72888331582 3789.52768335303  118909.07323744 9.75450465410602
#> 9  7766.70099673278 6294.84552699626 221476.635509889 7.15102097394398
#> 10 20055.7501346834 18560.1679630688 482974.832151226 30.9546955521932
#> 11 6778.62052436696 5549.21150623092  167355.21574853 12.7369278915486
#> 12 9191.85217421987 7629.62944226221 266745.806630666 6.83374576524781
#> 13 20395.8229323849 19369.3493646427 487209.496948447 23.1149171011162
#> 14 6962.23602542774 5724.63006567345 176816.064735412 12.7286021023337
#> 15  9000.6412121182 7394.09496322869 264540.183022897 6.87079850560387
#> 16 20665.1173254002  19280.570227519 513295.396375802 26.7927753520559
#> 17 6902.26546748866 5998.31307060578 175912.720244788 11.6627591277357
#> 18 9322.88661103618 7650.71148546443 270602.583082021 6.02821078125575
#> 19 19770.2644710336 18521.3378474724 501244.271833464 24.5842014720195
#> 20 7146.85154509176  5801.3706250886 180383.781780986 15.2348158158479
#> 21 9253.87199732393 7461.43555785835 267656.661911362 7.07957176432011
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.603043361497865 31.3423409021254  57.973347782735 1988.45275109982
#> 2   2.47800543897957 22.4840126984665 52.4140333617452  554.02935448655
#> 3   1.96794731145304 7.36301844416753 132.112974069254 1931.28271491074
#> 4   0.36107194039069 35.6758598597259 80.6446895452708 1826.64988290366
#> 5   2.57751920135758 40.3720137128255 138.700818915073  1490.4552240655
#> 6  0.228134755117995 9.56604153319819 51.1884134508478 1620.63103835239
#> 7  0.395801056268701 260.317248981326 299.685095703231 14222.4697096112
#> 8  0.502228413534429 45.4404618824468 103.409576053211 2247.45893439057
#> 9   0.36794856780678 26.7729718034647 84.9457262521067 3922.14970899108
#> 10 0.471297785412321 76.0653099396765 276.509319177907 10837.9375706002
#> 11 0.814024904913557 95.1198383884986  267.96208449631 3280.59905352386
#> 12 0.659316574011936 11.8197978192943 158.743435257609 4473.08714424654
#> 13 0.332291736220596  111.99373160249 253.092715812468 11290.3697546123
#> 14 0.660175737033849 69.7671184143682 168.430103589232 3751.36604462602
#> 15 0.572013152118441 14.3940530039883 116.315533394997 4536.02729523579
#> 16 0.378440274214745 79.3053684483116  249.53044562282 8768.87455074247
#> 17   0.5685488069551  101.42767722883 150.851869569546 4493.07617194761
#> 18 0.368342076319339 10.5234893603285 76.2620117283203 4076.29360289814
#> 19  0.34519560294634 83.4976402700639 262.306830944839 12562.5944329969
#> 20 0.372485465255014 75.4286636533145 121.381271017214 2966.48082361402
#> 21 0.471678426551026 8.64169097646343 151.280310293827 4798.65023415796
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE 88.3377999137296 95.9053327941468      14.621092604077
#> 2             TRUE 29.9293504329693 38.0825502172059     26.9638922518403
#> 3             TRUE 62.4894526734461 67.2274846987706     16.8930048548917
#> 4             TRUE 239.313405087384 247.005840040491      14.397398712708
#> 5             TRUE 63.9138021156969 68.9552258919748     18.1961443517396
#> 6             TRUE 200.856041706545 210.176071248542     12.1726868625915
#> 7             TRUE  486.44871613081 523.778393205298     6.34490330106803
#> 8             TRUE 187.043712051639 219.136032363647     12.5011101687763
#> 9             TRUE 278.694998644611 302.221857648887     4.02031897857369
#> 10            TRUE 621.753808989433 723.594757356149 -0.00494445751216144
#> 11            TRUE 307.605427180435  349.50991994363     8.66190595639561
#> 12            TRUE 324.586331111012 347.069354678677     2.87343926415171
#> 13            TRUE 693.637503453493 769.685580716165    0.239512905719244
#> 14            TRUE 321.969784710443 363.846885627121     8.40761547320379
#> 15            TRUE 308.937311705064 331.542238788501     2.48164918463232
#> 16            TRUE 733.282999084134 821.431229992398  -0.0474582123009384
#> 17            TRUE 314.773612765737 353.144090295988     7.15630729912982
#> 18            TRUE 340.801264463053 360.634077933385     4.43074767391459
#> 19            TRUE 748.682442532957 829.564465375901    0.240143117719922
#> 20            TRUE 323.604211000634 373.726755034773     8.96282376838488
#> 21            TRUE 328.875988620226 352.167779724839     4.55520417615046
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.605105263405 3355.50221129555 3458.61851286355    78008.8947995921
#> 2   35.1165301460831  1437.2652962982 1511.23769807615    41883.8265358833
#> 3   23.3675515095722 2183.63290943066 2207.85724011197    62510.8527665479
#> 4   15.5853253965934 7517.52055995641  7634.8941388949    147203.562718834
#> 5    26.676182524206 2297.28760127284 2430.11152638804    54607.3139828596
#> 6   12.9232502069297 4805.31191397888 4836.78419062311    128422.586962199
#> 7   7.64708877619205 16488.9702893645  17345.414038513    332714.564778532
#> 8   14.1534416493046  5115.9793235192 5265.47844311245    115212.003290368
#> 9     5.230869766658 7722.65945811608 7810.74253534948    215024.699238599
#> 10  1.54562525649438 19930.6226998326 20180.8775695342    465146.424847588
#> 11  11.3400478935612 6622.14839021788 6935.09265851604    161958.630305483
#> 12  5.04259079265098 9172.40860680713 9211.29574163261     259387.57827838
#> 13    1.332752717885 20211.5932438988  20580.052620871     468636.83870211
#> 14  10.5795936480451 6847.46911563611 7077.00293521938    170645.067592002
#> 15  4.36357245510199 8976.96299492664 9024.31942930976    257078.418122234
#> 16  1.19761028986557 20534.6599943027 20795.5746564976     498870.59773983
#> 17   9.0268328740121 6735.41693844724 7069.11399653009    168521.609941934
#> 18  5.64259310500522 9305.57547103844 9340.19775103392    263897.080105254
#> 19  1.37583665141338 19632.9108527893 19907.6180892778    480578.803991184
#> 20  10.1883009490739 7022.77139338206 7270.93169680147    175503.920826141
#> 21  6.10702619950333 9239.65641566765 9268.08757898021    259762.882276173
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84550.9043507105 87.6132488932642 96.6298838146123   14.4311339452052
#> 2     43706.5831121441 29.1487249217126 38.8631757284626   26.1833205385617
#> 3     68864.7728986042 62.0358113093192 67.6811260628974    16.273101451784
#> 4     153213.240833587 238.576895357831 247.742349770044   14.2836610514849
#> 5     59510.9116700351 63.4311126052022 69.4379154024695   17.3842258033119
#> 6     133754.463078378 199.963698452524 211.068414502563   12.1008244147294
#> 7     379506.490123153  482.87459811304 527.352511223068   6.22022596834339
#> 8     122606.143184513 183.971043085595 222.208701329691    12.342908218513
#> 9      227928.57178118 276.442427037819 304.474429255679   3.90441517971456
#> 10    500803.239454863 612.003079890492  733.34548645509 -0.153403259917043
#> 11    172751.801191577 303.593294894598 353.522052229468   8.40548811134784
#> 12    274104.034982951 322.433701194959  349.22198459473   2.66575454333795
#> 13    505782.155194784 686.356304566642 776.966779603017  0.134841008809756
#> 14    182987.061878822 317.960275048208 367.856395289356   8.19966011603812
#> 15     272001.94792356 306.773010175799 333.706540317766   2.30146504171501
#> 16    527720.195011773 724.843274848236 829.870954228295 -0.166666898678583
#> 17    183303.830547641 311.099843640501 356.817859421225   6.97721442493896
#> 18    277308.086058789 338.902378066958  362.53296432948     4.314719919874
#> 19    521909.739675744  740.93841906927 837.308488839587  0.131406502791825
#> 20    185263.642735831 318.805244018641 378.525722016765   8.84549084682955
#> 21    275550.441546552 326.645923514465   354.3978448306   4.40662547178688
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7950639222768 3345.62937391138 3468.49135024772    77382.5321829956
#> 2   35.8971018593616 1430.18283229818 1518.32016207617    41709.3072892201
#> 3     23.98745491268 2181.31355862075 2210.17659092188     61902.498711351
#> 4   15.6990630578164 7506.28266410059 7646.13203475072    146628.168005719
#> 5   27.4881010726336  2284.5704169533 2442.82871070758     54137.820587279
#> 6   12.9951126547919 4802.29861089592 4839.79749370606    127912.088185118
#> 7   7.77176610891669 16406.9703559354 17427.4139719422    328234.486820004
#> 8   14.3116435995679 5101.66557802623 5279.79218860542    114504.053726035
#> 9   5.34677356551714 7714.22597199799 7819.17602146757    213789.222080267
#> 10  1.69408405889926 19906.6621272016 20204.8381421652    461732.474512849
#> 11   11.596465738609 6592.18564112551 6965.05540760842    160925.241603623
#> 12  5.25027551346474 9168.68537049405 9215.01897794569    257978.555827942
#> 13  1.43742461479449  20176.315218444 20615.3306463258    465080.372229407
#> 14  10.7875490052108 6825.49247333558  7098.9795775199    169463.387287945
#> 15   4.5437565980193 8972.42886823039 9028.85355600602    255649.569524235
#> 16  1.31681897624322 20509.6788032415 20820.5558475589    496108.402256346
#> 17  9.20592574820296 6703.46722012015 7101.06371485717     167106.29094777
#> 18  5.75862085904581 9302.26057188994 9343.51265018243    262613.047620341
#> 19  1.48457326634148 19606.6090961043 19933.9198459629     476621.58674479
#> 20  10.3056338706292 6999.01136433127 7294.69172585226    174569.479366703
#> 21  6.25560490386691 9236.93428301006  9270.8097116378    258251.307452413
#>    ci95.integral.bt.up
#> 1     85177.2669673069
#> 2     43881.1023588073
#> 3     69473.1269538011
#> 4     153788.635546701
#> 5     59980.4050656157
#> 6     134264.961855459
#> 7      383986.56808168
#> 8     123314.092748846
#> 9     229164.048939512
#> 10    504217.189789602
#> 11    173785.189893437
#> 12    275513.057433389
#> 13    509338.621667487
#> 14    184168.742182879
#> 15    273430.796521559
#> 16    530482.390495257
#> 17    184719.149541805
#> 18    278592.118543702
#> 19    525866.956922138
#> 20     186198.08419527
#> 21    277062.016370312