Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  92.130021210824  15.5560587508401
#> 2                 FALSE          0.75 32.3809778993067   31.265808464409
#> 3                 FALSE          0.75 64.6583361159513   20.227313631379
#> 4                 FALSE          0.75 241.451730584773  14.9547860564372
#> 5                 FALSE          0.75 65.8684829813076   21.657979699665
#> 6                 FALSE          0.75 204.049885510265  12.2021966467906
#> 7                 FALSE          0.75 508.583332357459  7.09078215443886
#> 8                 FALSE          0.75 201.238468032627    13.34886912427
#> 9                 FALSE          0.75 289.750396866723   4.6811787509396
#> 10                FALSE          0.75 656.547592772416 0.530758257269013
#> 11                FALSE          0.75  319.72957523147  9.68626620792523
#> 12                FALSE          0.75 333.579293208917  3.74285371862528
#> 13                FALSE          0.75 728.361027547843 0.784037328556102
#> 14                FALSE          0.75 344.507826927097  9.81779624422939
#> 15                FALSE          0.75 317.288353977005  3.36971693282976
#> 16                FALSE          0.75 774.857879784214 0.548935910907531
#> 17                FALSE          0.75 334.873650150577  8.13036110107368
#> 18                FALSE          0.75 349.062084588651  4.93647345618348
#> 19                FALSE          0.75 771.600714368039 0.569174765965002
#> 20                FALSE          0.75 349.775731773778   9.5214300291355
#> 21                FALSE          0.75 341.492657746879  5.40080388356303
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3408.45022322997 2114.21096511256  81271.192343031 2.12079728873264
#> 2  1456.49403719394 263.024077306644 42588.9925221689 1.71899263809294
#> 3   2196.9091605457 685.082501159119 66089.4856267802 1.46876380712309
#> 4  7567.49431426271 6047.46584233362 149490.712033443 2.99340783743846
#> 5  2352.57355672199 1145.57608561773 57108.8574248513 3.08341438107468
#> 6  4819.43919866924 3259.52318044644 131159.694731256 2.61688724171357
#> 7  17006.7636475043 15781.7402795549 360068.981276918 13.8013034181732
#> 8  5175.79800818707 3773.97600198148 118686.093359017 7.39973875820126
#> 9   7767.4946000843 6275.17850861079 222445.064660392 6.63825200728558
#> 10 20067.1792485065  18656.826914251 485556.259693977 19.4849415579773
#> 11 6808.13905680496 5620.25830077515 168061.608694395 18.0158902968117
#> 12 9190.77113677231 7672.82576638208 267294.428510167 3.53182031993604
#> 13 20468.2408997329 19429.1913082827 493807.019963602 27.0723783120891
#> 14 6982.22064682398 5652.65812272904 177331.352599736  15.815776032283
#> 15 9007.11144243064 7374.18060292836 263221.134465187 5.92441314232654
#> 16 20664.3523698725 19319.3408834327 511208.420029072 23.3097746514936
#> 17 6929.53826184644 6026.04133675506 176896.989926136 13.5164441621801
#> 18  9322.0817649444 7658.32222218805 272877.081881312 5.25669166029214
#> 19 19805.6089772016 18659.2677256766 502161.883052453  27.436094984768
#> 20 7171.05538780584 5827.90596100551 181839.732485552 12.8906622825699
#> 21  9255.3430063807 7445.31283257353 269472.550272545 8.27017086363643
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.639634191405854 21.6319920008514 68.5921081248263 1727.46595116915
#> 2   2.36964279761597 20.5370514694439 93.0143808364453 892.155629674998
#> 3   1.57388772754309 4.63943397872938 97.1874463911834 1147.16804789889
#> 4  0.323343017963329 54.3904145403189 89.6455715494057 2927.37147686252
#> 5   2.73803716839759 44.0511426534883 151.159120490189 1848.41453161556
#> 6  0.572718109688139 8.97821898841558 113.150796062097 1454.93942803142
#> 7   0.52594282371622 198.887559605981 206.440091321941 10987.8895868463
#> 8   0.70162587765858 62.1493005825099 124.561364073871 2799.34210267133
#> 9  0.425164560591061 13.7149998816592 68.7154325873727 2376.85172341215
#> 10 0.302244726344253 77.3171965596772 238.717042703052 8184.39501899504
#> 11 0.848958397914783  53.869885047575 217.954543040189 2103.34325922902
#> 12 0.343231292193193 12.1752261063722 81.9397894875738 3212.77571408169
#> 13 0.225652193947613 63.0196881331232 217.791992754502 5149.73802184862
#> 14 0.679443850654513 42.4021702191617 205.322338833944 2987.17267469035
#> 15 0.493660045578939 16.3237225462368 134.104449122564 9255.25627999208
#> 16 0.331879474120612 91.6251749086949 214.365835519927 10766.8818171121
#> 17 0.551708754745188 45.6194209415301 164.973770013787 1823.38752337431
#> 18 0.279346015474829 7.70906851450893 76.6123477255936 1787.12877181611
#> 19 0.258483255286509 77.6060356633253 207.217586557611 8242.91413498899
#> 20 0.435698532194185 61.9179455245569 111.299988106738 2001.09660915363
#> 21 0.712583718775949 10.1672881857581 174.654005108491 2770.47623002582
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.6413096708588 95.6187327507892    14.5038605059775
#> 2             TRUE 29.5532350096438 35.2087207889696    27.3677460623307
#> 3             TRUE 62.2422196532339 67.0744525786688    17.6382683195706
#> 4             TRUE 236.527574692187 246.375886477359    14.4228867918875
#> 5             TRUE 60.7962663244397 70.9406996381754     17.153908557651
#> 6             TRUE 199.745105997646 208.354665022884    11.2600753563536
#> 7             TRUE 485.880188234564 531.286476480354    6.22560620942568
#> 8             TRUE 189.065897775386 213.411038289868    12.1946945555216
#> 9             TRUE 278.830472314738 300.670321418708    3.98178304876731
#> 10            TRUE 624.494863909544 688.600321635289  0.0335656824327166
#> 11            TRUE 290.093435693215 349.365714769725    8.28972964335541
#> 12            TRUE 327.769448782622 339.389137635212    3.17823824296747
#> 13            TRUE 683.826965224456  772.89508987123   0.412839469512279
#> 14            TRUE 318.490875353991 370.524778500202    8.70011110990272
#> 15            TRUE 307.542694357878 327.034013596132    2.55764615785241
#> 16            TRUE 736.513300482507 813.202459085921 0.00299417597912499
#> 17            TRUE 312.639099503791 357.108200797364    7.22280019951785
#> 18            TRUE 340.414826807471 357.709342369832    4.47694926072739
#> 19            TRUE 726.468338118096 816.733090617982   0.143969811018695
#> 20            TRUE 328.570592318951 370.980871228606    8.80470594367607
#> 21            TRUE 327.888226676197 355.097088817561    4.22860366617659
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6082569957027 3372.86559638857 3444.03485007137    78429.5108533577
#> 2   35.1638708664872  1422.7105875267 1490.27748686117    41121.3965113535
#> 3   22.8163589431873 2189.27729165069 2204.54102944071    64202.3941879866
#> 4   15.4866853209869 7478.02208234388 7656.96654618153    144675.185954004
#> 5   26.1620508416791   2280.109427057 2425.03768638698    54068.2155203437
#> 6   13.1443179372276  4804.6700284333 4834.20836890518    128766.319372144
#> 7   7.95595809945205 16679.5936119524 17333.9336830561    341993.902906556
#> 8   14.5030436930184 5073.56240872884  5278.0336076453    114081.175600123
#> 9    5.3805744531119 7744.93342527897 7790.05577488963    218535.143575379
#> 10  1.02795083210531 19939.9924601658 20194.3660368472     472092.92988773
#> 11   11.082802772495  6719.5230959017 6896.75501770822    164601.609032963
#> 12  4.30746919428308 9170.74288982733  9210.7993837173    262009.412460502
#> 13  1.15523518759993 20364.5735127539 20571.9082867118    485335.700917661
#> 14  10.9354813785561 6912.46907681346  7051.9722168345    172417.453549871
#> 15  4.18178770780712 8980.25891884208  9033.9639660192      247996.2378846
#> 16  1.09487764583594 20513.6289571477 20815.0757825973    493496.899439923
#> 17  9.03792200262952 6854.49431439762 7004.58220929526    173897.517450185
#> 18  5.39599765163957 9309.40034723803 9334.76318265076    269937.255051674
#> 19  0.99437972091131 19677.9470485355 19933.2709058678    488602.289300396
#> 20  10.2381541145949 7069.20036741795 7272.91040819374    178547.928563495
#> 21  6.57300410094946 9238.61781731513 9272.06819544627    264915.116874153
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84112.8738327042  87.973258524908   96.28678389674    14.3023757356846
#> 2     44056.5885329843 29.0117523286445 35.7502034699689    26.6213085810817
#> 3     67976.5770655739 61.7795590539901 67.5371131779126    17.1424936853945
#> 4     154306.238112882 235.584651223394 247.318809946152    14.3210337412291
#> 5     60149.4993293589 59.8249907944012 71.9119751682139    16.2914268496058
#> 6     133553.070090367 198.920786516507 209.178984504024    11.0796691518019
#> 7      378144.05964728  481.53277765784 535.633887057078    6.05993421995507
#> 8     123291.011117911 186.734980066553 215.741955998702    11.9736824040592
#> 9     226354.985745405 276.739422932443 302.761370801003    3.84785621218112
#> 10    499019.589500224 618.357107318781 694.738078226052 -0.0616414063657231
#> 11    171521.608355826 284.418430249719 355.040720213221    8.02230774801225
#> 12    272579.444559831 326.656925381842 340.501661035992    3.07012038592662
#> 13    502278.339009543 675.299166056148 781.422889039538   0.341759028418781
#> 14    182245.251649602 313.508905903822 375.506747950372    8.48608629694655
#> 15    278446.031045774 305.676504218045 328.900203735965    2.40214324349504
#> 16    528919.940618222 729.170721467286 820.545038101141  -0.101547858368868
#> 17    179896.462402087 308.381419592704  361.36588070845    7.04901194177311
#> 18    275816.908710949 338.758968934479 359.365200242824    4.38895526585282
#> 19     515721.47680451 717.825968197894 825.375460538184  0.0625475856034449
#> 20     185131.53640761 324.510033699941 375.041429847615     8.6674609060349
#> 21    274029.983670938 325.283122854152 357.702192639606    4.00413979476217
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8097417659956  3366.0515189083 3450.84892755163    77885.3590787394
#> 2   35.9103083477363 1416.24141631383 1496.74665807405    40840.3674880059
#> 3   23.3121335773634 2187.81586994739 2206.00245114401    63841.0362528984
#> 4   15.5885383716453 7460.88910176368 7674.09952676173    143753.063938792
#> 5   27.0245325497243 2266.23331712115 2438.91379632283    53485.9649428848
#> 6   13.3247241417794 4801.84188945195 4837.03650788653    128308.013452314
#> 7   8.12163008892265 16616.9440306765  17396.583264332      338532.7176867
#> 8   14.7240558444808 5053.98537904535 5297.61063732879    113199.382837781
#> 9   5.51450128969809 7740.61320031625 7794.37599985235    217786.435282505
#> 10  1.12315792090375 19915.6375432495 20218.7209537635    469514.845456747
#> 11  11.3502246678382 6702.55408211171 6913.72403149821    163939.055906306
#> 12  4.41558705132394 9166.90769360382  9214.6345799408    260997.388110567
#> 13  1.22631562869342 20344.7223109919 20591.7594884738    483713.533440778
#> 14  11.1495061915122 6899.11239319442 7065.32890045354    171476.494157343
#> 15  4.33729062216448 8975.11694624001 9039.10593862126    245080.832156403
#> 16  1.19941968018393 20484.7670270515 20843.9377126936    490105.331667533
#> 17  9.21171026037425 6840.12419680104 7018.95232689184    173323.150380322
#> 18  5.48399164651414 9306.97199065596 9337.19153923283    269374.309488552
#> 19  1.07580194632656 19653.5011473015 19957.7168071018    486005.771347874
#> 20  10.3753991522361 7049.69621457771 7292.41456103398    177917.583131611
#> 21  6.79746797236389 9235.41512153661 9275.27089122479    264042.416861695
#>    ci95.integral.bt.up
#> 1     84657.0256073225
#> 2     44337.6175563319
#> 3     68337.9350006621
#> 4     155228.360128093
#> 5     60731.7499068178
#> 6     134011.376010197
#> 7     381605.244867137
#> 8     124172.803880253
#> 9      227103.69403828
#> 10    501597.673931208
#> 11    172184.161482484
#> 12    273591.468909767
#> 13    503900.506486425
#> 14    183186.211042129
#> 15    281361.436773972
#> 16    532311.508390612
#> 17     180470.82947195
#> 18    276379.854274071
#> 19    518317.994757031
#> 20    185761.881839494
#> 21    274902.683683396