Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.3420575002194  15.4818014470451
#> 2                 FALSE          0.75 32.3230281095557  30.5403525929139
#> 3                 FALSE          0.75 65.2286745842334  20.4583144358096
#> 4                 FALSE          0.75  242.47165432115  14.9254648737917
#> 5                 FALSE          0.75 66.0718165444403   22.516463420523
#> 6                 FALSE          0.75 205.323555182748   12.410368420147
#> 7                 FALSE          0.75 512.037714914447  7.21209242050676
#> 8                 FALSE          0.75 204.336752334003  13.5429952786158
#> 9                 FALSE          0.75 289.571939275543  4.66293376679909
#> 10                FALSE          0.75 678.987485281938 0.553438911024879
#> 11                FALSE          0.75 322.229500672065  9.84344536426629
#> 12                FALSE          0.75 333.289771098273  3.80400323656574
#> 13                FALSE          0.75 730.351347004657     0.94920594974
#> 14                FALSE          0.75 338.682260805059  9.81396803100674
#> 15                FALSE          0.75 319.855662427388  3.46486514808543
#> 16                FALSE          0.75 778.202161581766 0.617925938845204
#> 17                FALSE          0.75 334.906959136868  8.11085251591333
#> 18                FALSE          0.75 348.188210419273  4.88698287660396
#> 19                FALSE          0.75 773.126773442524 0.632605825453363
#> 20                FALSE          0.75 355.271007764842   9.5853125685489
#> 21                FALSE          0.75 338.751531510797  5.19507920615123
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3410.47518884853 2127.87284460327 81359.9402798675 2.20353727041012
#> 2  1458.41555036609 299.222423201949 42424.9283605333 2.09582402132147
#> 3  2196.67564275469 663.887186911335 66154.8114249112 1.20436600658023
#> 4  7556.06116699764 6052.92384433603 148978.805353008 2.43030897538161
#> 5  2370.97825019404 1138.46737035244 56974.9774933612 1.80736919865963
#> 6  4822.50388206302 3229.36423094005 131712.367274203 2.49982527250572
#> 7  16982.3461296063  15744.052251925 358009.171955939 15.2250852750082
#> 8  5188.99270243206 3756.06709620139 118658.660910888 7.74727066319966
#> 9  7758.91161307751 6288.11620028224 221942.834375741 5.47291200747338
#> 10 20046.8048074676 18759.3793491458 483382.602912138 19.1454438165001
#> 11 6810.36676699014 5606.88411421607 168347.409253251 9.89440330328763
#> 12 9188.31015169469 7622.13626559532 267615.141958174 5.67571518571892
#> 13 20438.9431678797 19293.0564090071 489910.088014149 28.2501216685637
#> 14 6993.51974754848 5661.23903680701  178034.52281146 18.4839130770822
#> 15 9008.33932230652 7368.69460415468 263367.797156938 5.13345462454083
#> 16 20640.5543914922 19207.0352200287 511076.082440369 31.2228599310479
#> 17 6883.80924215649  5992.9563079243 175026.618955682 14.0512912082236
#> 18 9322.52236511504 7651.96744066503 270573.663409898 7.99814334958439
#> 19 19757.7168209226  18521.475915989 504466.139861829 22.9535810476391
#> 20 7131.61045222929 5820.04973127449  179498.23326459 7.59177818043419
#> 21 9250.94281167651 7484.94009098368  267864.31751764  7.5672888874151
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.685019954959914 35.3380877957416  57.036412536469 1911.02348653693
#> 2   2.55213947353463 19.2917203723725 80.1513759455943 1055.86943925339
#> 3   1.34315454034056 3.76032249118099 82.8644875615596 1380.89549155122
#> 4  0.374580964420823 71.1813383228129 115.035439917704 3560.48625242425
#> 5   2.51724109316241 32.5471871031677 129.138936996792 1663.66987218571
#> 6   0.50927945745983 10.8581692974005 97.4063021787008 1238.40363658246
#> 7  0.577787698808857 171.989961844522 228.068013762047 9501.69845255498
#> 8   0.63556525697234 27.8269030241837 111.723461900532 1910.16267805499
#> 9  0.442605430079908 23.0653096857459 93.4063797661162 2013.86320852018
#> 10  0.24793415690858 74.1531051624824 197.339401972129 10900.4021939378
#> 11 0.517894983861377 52.5322288469464 159.842577779916 1959.54341670824
#> 12 0.736543511183541 14.3548481158187 192.154746064884 3424.09490013617
#> 13 0.344638258501078 92.7114023003192 211.582114121597 9976.04327834585
#> 14 0.732402286609224 42.7766311194914 217.539971038629 1813.04163544019
#> 15 0.503113142965863 12.6265654571621 122.522979694064 4473.22199817388
#> 16 0.554275748480291 92.4687536975963  341.11810978858 11118.9231029365
#> 17 0.564283472354339 103.154784061747 190.036671167363 3991.46927872765
#> 18  0.52274678878047 9.58714816134579 112.100357972239  3810.9421316027
#> 19 0.325923255556457 65.5857798263153 206.702635520697 7478.60589732472
#> 20 0.359241735896573 93.2474776660393 145.856302399231 3739.35310630581
#> 21 0.453512134094948 13.0380007344903 112.066945634176 5169.65872513654
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.7172386903948 95.9668763100441    14.354943621136
#> 2             TRUE 28.8753975944819 35.7706586246296   26.3420831589494
#> 3             TRUE 63.2474925034089 67.2098566650579   18.2488252169494
#> 4             TRUE 238.473796056648 246.469512585653   14.3092791873195
#> 5             TRUE 63.0986942126452 69.0449388762354   18.3756018222708
#> 6             TRUE 201.211342609477  209.43576775602   11.5726037126256
#> 7             TRUE 486.992449637059 537.082980191836   6.26163165596619
#> 8             TRUE  191.59249209304 217.081012574967   12.4974904308963
#> 9             TRUE 280.568999023249 298.574879527837   3.93484783431764
#> 10            TRUE 647.493230203795  710.48174036008  0.145587222910266
#> 11            TRUE 305.953207238157 338.505794105973   8.99150811581433
#> 12            TRUE 323.953219617766 342.626322578781   2.59238916066881
#> 13            TRUE  683.87989685987 776.822797149445  0.382276014505727
#> 14            TRUE 308.276223793259 369.088297816859   8.60916626953456
#> 15            TRUE 311.411129570018 328.300195284757   2.63724402790658
#> 16            TRUE 726.840556995192  829.56376616834 -0.293857667404875
#> 17            TRUE  311.79258509934 358.021333174396   7.18260620389044
#> 18            TRUE 335.031264609207 361.345156229339   4.02706440906009
#> 19            TRUE 735.368132619158 810.885414265891 0.0964620700629917
#> 20            TRUE 342.782532658028 367.759482871656   8.99435991299904
#> 21            TRUE 326.303341290999 351.199721730595   4.44905174556504
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6086592729541 3352.34403442453 3468.60634327252    78216.3066445142
#> 2   34.7386220268784 1426.68067035354 1490.15043037864    40688.0231329614
#> 3   22.6678036546699  2190.4899122567 2202.86137325268    63883.2383413094
#> 4    15.541650560264 7438.96786545661 7673.15446853867     143121.80546777
#> 5   26.6573250187752 2317.43812740933 2424.51837297875    54238.2405536157
#> 6   13.2481331276684  4804.6421935688 4840.36557055725    129675.193292025
#> 7   8.16255318504733 16699.4226423721 17265.2696168406    342378.878001486
#> 8   14.5885001263353 5143.21744695727 5234.76795790684    115516.443305488
#> 9   5.39101969928053 7720.96917864446 7796.85404751056    218630.029397725
#> 10 0.961290599139493 19924.8229494753 20168.7866654599    465451.441303111
#> 11  10.6953826127183 6723.95125053691 6896.78228344337    165123.960332766
#> 12  5.01561731246266 9164.69642654417 9211.92387684522     261982.50584745
#> 13  1.51613588497427 20286.4329110957 20591.4534246637     473499.49682127
#> 14  11.0187697924789 6923.15218935692 7063.88730574004    175052.069321161
#> 15  4.29248626826427 8987.56862212949 9029.11002248356    256009.346969942
#> 16  1.52970954509528 20488.4432916596 20792.6654913247    492785.453936038
#> 17  9.03909882793621 6714.11962237491 7053.49886193806    168460.651992175
#> 18  5.74690134414784 9306.75150638962 9338.29322384045    264304.663603412
#> 19  1.16874958084373 19649.8282131083 19865.6054287369     492163.83316073
#> 20  10.1762652240988 6978.21835146865 7285.00255298992    173346.997404717
#> 21  5.94110666673742 9229.49530046827 9272.39032288475    259360.228914791
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84503.5739152207 88.0231244502156 96.6609905502233     14.1391623353237
#> 2     44161.8335881051 28.2152130277657 36.4308431913458      25.538159224786
#> 3     68426.3845085129 62.8681172113362 67.5892319571307     17.8257315367421
#> 4     154835.805238246 237.708248729402 247.235059912898     14.1912861835269
#> 5     59711.7144331067 62.5293729150675 69.6142601738132     17.5826708779247
#> 6     133749.541256381 200.423897648637  210.22321271686     11.4121806835257
#> 7     373639.465910392 482.196547775431 541.878882053463      6.0796285308414
#> 8     121800.878516289 189.152101834132 219.521402833874       12.29728737495
#> 9     225255.639353757 278.845031740895 300.298846810191     3.79542712384247
#> 10    501313.764521166 641.462415401597 716.512555162278   0.0674879634840631
#> 11    171570.858173736 302.836470197621 341.622531146509     8.82837119589799
#> 12    273247.778068898 322.165369334264 344.414172862282       2.360377954646
#> 13    506320.679207028 674.981108534272 785.721585475042    0.273714963077888
#> 14    181016.976301759 302.453791173978  374.91073043614     8.37845954925266
#> 15    270726.247343934 309.794091363288 329.917233491488     2.47876338787234
#> 16    529366.710944699 717.005356116912  839.39896704662   -0.468454528176167
#> 17    181592.585919189 307.366428368749 362.447489904986     7.00485691009882
#> 18    276842.663216384 332.511849454087 363.864571384458     3.86239917059424
#> 19    516768.446562928 728.137754589152 818.115792295897 -0.00620375543729212
#> 20    185649.469124463 340.391122531191 370.150892998493     8.88119876619162
#> 21     276368.40612049 323.919645291464 353.583417730131     4.30619542332513
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8244405587665 3341.21253676887 3479.73784092818    77614.3342462551
#> 2   35.5425459610418 1420.60377843624 1496.22732229594    40355.4242595966
#> 3   23.0908973348771 2189.30541067198  2204.0458748374    63448.2562614708
#> 4   15.6596435640565 7416.54574388492 7695.57659011035    142000.252298257
#> 5   27.4502559631213 2307.18576347184 2434.77073691625    53714.1845438772
#> 6   13.4085561567683 4801.22187024012 4843.78589388593    129285.096146501
#> 7   8.34455631017212 16645.2458043911 17319.4464548216    339385.842988931
#> 8   14.7887031822816 5134.45197250465 5243.53343235946      114914.7420619
#> 9   5.53044040975571 7713.70360609345 7804.11962006157    217995.662487041
#> 10   1.0393898585657 19901.4647213492 20192.1448935861     462017.81461202
#> 11  10.8585195326346 6707.40359845012 6913.32993553015    164506.704156503
#> 12  5.24762851848548 9160.17464938769  9216.4456540017    260903.915953907
#> 13  1.62469693640211 20257.2288193711 20620.6575163883    470357.043188591
#> 14  11.2494765127608 6909.67755055428 7077.36194454268    174480.961205998
#> 15  4.45096690829852 8983.59125401049 9033.08739060256    254600.282040517
#> 16  1.70430640586657 20459.3156342449 20821.7931487395    489282.993158613
#> 17  9.21684812172783 6681.62586539546 7085.99261891751    167203.339169376
#> 18  5.91156658261368  9303.7315547188 9341.31317551128    263104.216831957
#> 19  1.27141540634402 19629.1686924631 19886.2649493822    489808.072303073
#> 20  10.2894263709062 6948.84539600385 7314.37550845472    172169.101176231
#> 21  6.08396298897733 9225.38833023691 9276.49729311611    257731.786416373
#>    ci95.integral.bt.up
#> 1     85105.5463134798
#> 2     44494.4324614699
#> 3     68861.3665883516
#> 4      155957.35840776
#> 5     60235.7704428451
#> 6     134139.638401905
#> 7     376632.500922946
#> 8     122402.579759876
#> 9      225890.00626444
#> 10    504747.391212256
#> 11    172188.114349999
#> 12    274326.367962441
#> 13    509463.132839707
#> 14    181588.084416923
#> 15    272135.312273359
#> 16    532869.171722124
#> 17    182849.898741989
#> 18    278043.109987839
#> 19    519124.207420586
#> 20     186827.36535295
#> 21    277996.848618908