Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.3609104161732  15.6104109168641
#> 2                 FALSE          0.75  33.277034136169  30.5679844173489
#> 3                 FALSE          0.75 64.7655067383413  19.8095929147516
#> 4                 FALSE          0.75 243.463313700444  15.0428156848095
#> 5                 FALSE          0.75 65.6498486732437  21.6957393719081
#> 6                 FALSE          0.75 204.932385646623   12.342926524121
#> 7                 FALSE          0.75  507.81596523191  7.01383866134456
#> 8                 FALSE          0.75 200.056957256341  13.5885071611119
#> 9                 FALSE          0.75 289.510047324609  4.56741387602007
#> 10                FALSE          0.75 669.657423917788 0.629764501098033
#> 11                FALSE          0.75 324.338709085063   9.7969808081992
#> 12                FALSE          0.75 334.501058342806  3.78170072683708
#> 13                FALSE          0.75 727.954171238893 0.727773829043308
#> 14                FALSE          0.75 346.431147845294  9.86242174220282
#> 15                FALSE          0.75 320.112710122753   3.6097755161796
#> 16                FALSE          0.75 780.172756239534 0.680393479837546
#> 17                FALSE          0.75 334.849496398909  8.19331456630972
#> 18                FALSE          0.75 350.197813563196  5.07583784292017
#> 19                FALSE          0.75 784.907684067577 0.668072369853073
#> 20                FALSE          0.75 349.118695095604  9.48534991553809
#> 21                FALSE          0.75 338.965473301977  5.05477977538896
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3416.79099589522 2118.33579219988 81596.6295178998 1.91298401936839
#> 2  1468.36802044191 295.532528121954 42542.2866092286 2.10477911910529
#> 3  2197.42473524526 705.215225963942 65747.0963730733 1.22428334330724
#> 4  7526.91795015554 6003.88894419107 147392.361004035 2.92469459988387
#> 5  2350.07646676021 1126.54350611525 56654.2284344987 3.30835031041056
#> 6  4822.61708631242 3239.67229540798  130778.85371929 2.48095068842247
#> 7   17060.377602737 15871.2049476184 363534.651085941 12.6044141925647
#> 8  5176.94146420541 3736.89968153062 118744.344694115 7.42932427381918
#> 9   7765.2904272206 6309.17866771115 221666.563706984 7.28819456436366
#> 10 20064.3222636056 18675.2648760694  482073.11152064 24.4782355716299
#> 11 6766.52602348185 5558.21333175281 166492.877734796  15.023431441899
#> 12 9188.97570192238 7662.41751893272 267897.441971094 7.60212797859825
#> 13 20431.6106472747 19417.2637741451 488896.361786856 18.0348032785466
#> 14 6980.47243056065 5642.26026747771 178461.599185978 13.6799611620405
#> 15 9004.52924983522 7339.08597034162 265488.726639634 7.12408380105206
#> 16  20677.094161151 19246.0086072521 513537.982358089 32.8457473093883
#> 17 6896.77529829173 5976.19443448879 174960.858920743 19.8831348653057
#> 18  9320.3151984436 7613.18631121951 269441.881882839 5.24272918005374
#> 19 19796.3643821932 18642.3000512277 500385.692752975 28.0738127365259
#> 20  7123.9750739391 5796.08789150192 180384.478734687 13.3815973003332
#> 21 9259.46804313953 7525.09306874173 267645.301242879 8.94103032489831
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.891797374129239 20.8375302036042 77.8182282995641 1488.21132753072
#> 2   2.91872886883934 19.4183755500808 101.131242663151 766.308813331798
#> 3   1.98046114777368 3.83743550903172 129.628296741595 1089.91994774617
#> 4  0.290337701918838 82.8448629876228 121.932508559153 4272.51101522578
#> 5   1.82726541431315 58.5855135278694 155.425329728545 2044.18280125541
#> 6  0.442814958335475 9.11033101495219 89.7348844705887 3231.46237115366
#> 7  0.460076092041279 144.829922978403 158.561083384653 7184.63839095589
#> 8  0.570163109426984 57.3563943650585 141.038098407295 2933.50250509634
#> 9  0.434891147752397 21.7278413909752 64.4923205893896 3599.56320516367
#> 10 0.266074254970017 105.249333532755   264.6200271581 12532.2749317596
#> 11 0.811077633197232 118.430884542869 256.716704015086 5280.48623009807
#> 12 0.331250388612468 13.8813890113348 88.3847894209059 3348.75483112712
#> 13 0.194789417461779 83.8748334285776 187.335389653025 9325.66791436144
#> 14  0.60854483554829 38.7902230845807 177.400304750507 2028.84450351997
#> 15 0.757552940495176 13.5262424749525 191.196271965508 3569.61238771782
#> 16 0.490891748110373 85.9662898954574 324.341179880173 7504.95348076178
#> 17 0.614315580230988 103.812887902498 162.029304580108 4337.09367572632
#> 18  0.63466356891123 7.20196674619059 181.913058165756 5012.76214981886
#> 19 0.492701260737387 83.0037116435353 318.448457661937 10007.1474756355
#> 20 0.324419116156651 76.4859781374196 79.0174567512711 2546.65951128348
#> 21 0.471641543856644 10.4947692658594 95.1466775263658 4362.67269198754
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.2140517043122 95.5077691280342   14.1434042364215
#> 2             TRUE 29.8146724852408 36.7393957870972   25.7666754281082
#> 3             TRUE 62.7515606386009 66.7794528380818   16.5517343266638
#> 4             TRUE 238.652191083635 248.274436317253    14.565210165153
#> 5             TRUE 60.2076124126183  71.092084933869    18.689887765363
#> 6             TRUE 200.851221764168 209.013549529078   11.6144959176591
#> 7             TRUE 487.081703885141 528.550226578679   6.25701348993666
#> 8             TRUE 187.835718825909 212.278195686774   12.6505888461045
#> 9             TRUE 277.520967266231 301.499127382987   3.85201793796738
#> 10            TRUE 629.390726402457 709.924121433119  0.192072351672355
#> 11            TRUE 299.625164363139 349.052253806987   8.46275810158976
#> 12            TRUE 321.995557818012   347.0065588676   3.23679383756957
#> 13            TRUE 698.286919845684 757.621422632103  0.407345237318682
#> 14            TRUE 323.927611733737 368.934683956851   8.86136548772588
#> 15            TRUE 308.393592270022 331.831827975483   2.36360092906504
#> 16            TRUE  726.14150191559 834.204010563477 -0.127123445804017
#> 17            TRUE 302.141739545482 367.557253252337   7.18276543682974
#> 18            TRUE 341.573524062008 358.822103064385   4.03181627206119
#> 19            TRUE 738.726262115992 831.089106019162 -0.142421204059928
#> 20            TRUE 327.105967536556 371.131422654652    8.9516804694604
#> 21            TRUE  324.25747841752 353.673468186435   4.27892943574478
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   17.0774175973067  3382.5132587103 3451.06873308015    79148.5218841117
#> 2   35.3692934065896 1436.42479266202 1500.31124822179    41281.7086112978
#> 3   23.0674515028393 2191.11215383291 2203.73731665762    63954.1780590308
#> 4    15.520421204466  7390.6381505409 7663.19774977018    140364.080383988
#> 5   24.7015909784533 2253.70329700687 2446.44963651356    53291.5477264336
#> 6   13.0713571305828 4807.63059179282 4837.60358083202    125463.098118743
#> 7   7.77066383275246 16822.1323794375 17298.6228260365    351715.920932819
#> 8   14.5264254761193 5082.59019547489 5271.29273293593    113918.733073232
#> 9   5.28280981407276 7729.54812813245 7801.03272630875     215745.28223449
#> 10  1.06745665052371 19891.1871099442  20237.457417267    461457.519257895
#> 11  11.1312035148087 6571.70721840883 6961.34482855487    157806.477886285
#> 12  4.32660761610459 9166.14081699874 9211.81058684603     262388.74027389
#> 13  1.04820242076793 20293.6365462847 20569.5847482647    473555.638067731
#> 14  10.8634779966798 6916.66251358651 7044.28234753478    175124.149977688
#> 15  4.85595010329417 8982.27858096392 9026.77991870652    259616.714261838
#> 16  1.48791040547911  20535.679614273  20818.508708029    501192.333882236
#> 17  9.20386369578969 6726.00309769212 7067.54749889134    167826.339824173
#> 18  6.11985941377914 9308.46796314611 9332.16243374108    261195.888146387
#> 19  1.47856594376607 19659.8232765396 19932.9054878468    483923.935155554
#> 20  10.0190193616158 6998.15563990305 7249.79450797516    176195.223838625
#> 21  5.83063011503314 9242.20414769719 9276.73193858187     260468.70466456
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84044.7371516878 88.6114617382112 96.1103590941353   13.8624880635708
#> 2     43802.8646071594 29.1516670627226 37.4024012096153   24.8472758344238
#> 3     67540.0146871157 62.3659113854591 67.1651020912235   15.9278890651151
#> 4     154420.641624081 237.730912284671 249.195715116216   14.4737537890486
#> 5     60016.9091425639  59.165482064839 72.1342152816484   18.1142991598544
#> 6     136094.609319838 200.069722297315 209.795048995931   11.4750092057835
#> 7     375353.381239064 483.111313414483 532.520617049337   6.11208952094365
#> 8     123569.956314999 185.495481679656 214.618432833027    12.470987466635
#> 9     227587.845179478 275.225185978456 303.794908670762   3.71502722642537
#> 10    502688.703783384 621.680082197394 717.634765638183    0.1082589613568
#> 11    175179.277583307 294.892783458941 353.784634711185   8.20726864713263
#> 12    273406.143668298 319.600887504754 349.401229180859   3.13244996515664
#> 13     504237.08550598 692.605956812942 763.302385664845  0.345986570818222
#> 14    181799.048394268 319.618423967695 373.243871722893   8.66967386452817
#> 15     271360.73901743 306.149505872691 334.075914372815   2.12497175280906
#> 16    525883.630833942 715.795091513133 844.550420965935 -0.281754346458784
#> 17    182095.378017312  295.87855206291 373.820440734909   6.98925602905698
#> 18    277687.875619291 339.922064370291 360.473562756102   3.83189724785415
#> 19    516847.450350395 729.883011103986 839.932357031167 -0.297622101192205
#> 20    184573.733630748 322.890764386951 375.346625804257   8.84948844787105
#> 21    274821.897821199 321.441053865177 356.489892738778   4.13036234942994
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.3583337701574 3375.94943669616 3457.63255509429    78679.7353159396
#> 2    36.288693000274 1430.30800436375 1506.42803652006    41040.3213350983
#> 3    23.691296764388 2189.90336164756 2204.94610884297    63610.8532754908
#> 4   15.6118775805705  7364.5420186998 7689.29388161128    139018.239414192
#> 5   25.2771795839619 2235.24886024559 2464.90407327484    52647.6301440381
#> 6   13.2108438424585 4804.76083752311 4840.47333510173    124445.187471829
#> 7   7.91558780174547 16776.5109536993 17344.2442517747    349452.759839668
#> 8   14.7060268555888 5064.52293124989 5289.35999716092    112994.679784126
#> 9   5.41980052561477 7722.70385809429 7807.87699634691    214611.419824863
#> 10  1.15127004083927 19858.0335698814 20270.6109573298    457509.852654391
#> 11  11.3866929692658 6534.40148977783 6998.65055718587    156143.124723804
#> 12  4.43095148851751 9161.76817946017  9216.1832243846    261333.882502085
#> 13  1.10956108726839 20267.2159737547 20596.0053207947    470618.052674707
#> 14  11.0551696198775 6904.44359331487 7056.50126780643    174485.063959079
#> 15  5.09457927955015 8978.01781458431 9031.04068508613    258492.286359707
#> 16  1.64254130613388 20508.6002329559 20845.5880893461    498828.273535796
#> 17  9.39737310356245 6693.30203800283 7100.24855858062    166460.155316319
#> 18  6.31977843798618 9306.19934362106 9334.43105326613    259616.868069194
#> 19  1.63376684089835 19633.6771073719 19959.0516570145    480771.683700729
#> 20  10.1212113832051 6974.06255678976 7273.88759108845    175393.026092571
#> 21  5.97919720134798 9238.89829537844 9280.03779090061    259094.462766584
#>    ci95.integral.bt.up
#> 1       84513.52371986
#> 2     44044.2518833589
#> 3     67883.3394706558
#> 4     155766.482593877
#> 5     60660.8267249593
#> 6     137112.519966752
#> 7     377616.542332215
#> 8     124494.009604104
#> 9     228721.707589105
#> 10    506636.370386888
#> 11    176842.630745788
#> 12    274461.001440103
#> 13    507174.670899004
#> 14    182438.134412877
#> 15    272485.166919561
#> 16    528247.691180382
#> 17    183461.562525166
#> 18    279266.895696484
#> 19    519999.701805221
#> 20    185375.931376802
#> 21    276196.139719175