Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.0455910371266  15.3051613770695
#> 2                 FALSE          0.75 32.1630774693964  30.0692819300847
#> 3                 FALSE          0.75 65.2631892935848  20.5866020395795
#> 4                 FALSE          0.75 242.369009803703  14.9939358664685
#> 5                 FALSE          0.75 66.2070279570766  22.9211687750307
#> 6                 FALSE          0.75 205.496951820609  12.3646870338862
#> 7                 FALSE          0.75 506.164711873368  6.99230363719535
#> 8                 FALSE          0.75 200.312380211696  13.4706743405378
#> 9                 FALSE          0.75 291.186365246273  4.79472939406499
#> 10                FALSE          0.75 676.940905888192 0.464095925187942
#> 11                FALSE          0.75 326.693493151043  10.0485409903157
#> 12                FALSE          0.75 335.280222598333  3.86676666893436
#> 13                FALSE          0.75 728.491313012987 0.770769176649501
#> 14                FALSE          0.75 348.174809013245  9.72206947041798
#> 15                FALSE          0.75 321.643877556162  3.43227541718406
#> 16                FALSE          0.75 774.800202978973 0.519506539112866
#> 17                FALSE          0.75 337.647854586826  8.40066950731783
#> 18                FALSE          0.75 348.977099067315  4.88022680705815
#> 19                FALSE          0.75 783.583565043204 0.659026808266885
#> 20                FALSE          0.75 358.650296114866  9.76209516763521
#> 21                FALSE          0.75 337.032119806036  5.07452884657735
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.44442592989 2139.38238776169 81791.0612700924 2.51160280872141
#> 2  1457.92328223093 314.019989673587 42053.5599219936 2.26444357297714
#> 3  2194.85319134026 654.137176869777 66033.8587920779  1.7016617419344
#> 4  7551.20304972575 6023.12445168427 148933.521148375 2.31426235592174
#> 5  2375.18166621499  1080.1384978649 57456.0049721598 2.59637387199066
#> 6  4818.83970708127 3233.61704263099 131200.904022595 2.65158139078062
#> 7  17065.7642678047 15860.2512908741 363147.077954839 10.9877886615217
#> 8  5192.73501502543 3757.12130254697 119365.456841894  8.8511615014424
#> 9   7760.2228906629 6251.25461687432  220899.14328845  6.7933716367807
#> 10 20103.8135223462 18843.9144894664 486476.233942891 19.7403103451501
#> 11 6797.96215656739 5544.32438662795  167409.04612498 12.4733557203162
#> 12 9189.07253193446 7646.75691021461 266708.244594419 7.46169957221757
#> 13 20470.3452423045  19421.835936963 494086.459016481 22.4432793419154
#> 14 6969.52917652871 5722.10179559369 176963.047762678 14.7504393689833
#> 15 9001.24266310582 7403.10023307236 265745.397757474 6.87449697251957
#> 16 20663.0396797222 19323.0482394002 512645.709005507 19.7173156708929
#> 17 6928.92278110033  5978.9257292857 176497.005350678 12.1325844543787
#> 18 9321.12952703023 7658.62346197533 270051.776809775 8.77385949795666
#> 19 19790.3503919943  18628.020935487  500914.69579109 22.0282605997905
#> 20 7155.39721638197 5791.61503588016 181113.988962873 12.7386746293519
#> 21 9257.58071378889 7505.23760252965 268894.918097724 5.49276240826525
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.639243659298299   21.64103078694 50.5265332690327 1398.98078385719
#> 2   1.65059157859201 28.0324190826669 59.9947374521842 1208.09067650909
#> 3   1.51911032606862 5.27376869899074 102.282039470066  1101.0764176982
#> 4  0.328822332927831 77.1527674694329 109.488050258213 4040.48597026129
#> 5   2.11209189460818 24.7331488838063 121.549104967101 887.501791560766
#> 6  0.492446796407766 11.3895158472882 98.4443613812096 1969.89887745218
#> 7  0.428300083160949 125.196639784258  165.37394877013 6368.41082408822
#> 8  0.880249035195701 29.6498441454314 158.793573358618 1926.54309646753
#> 9  0.484009593737432 14.8839882687485 83.6715174634198 3829.82111451186
#> 10 0.236010037413216 86.6755060452611 187.346442597933 8871.24873539303
#> 11 0.608801033487221 87.1223941405918 245.400480497999 3478.04841323567
#> 12 0.553832377907544 15.5437142452054 137.525409279989 3673.12634793026
#> 13 0.410783406788398 46.7434621567126 261.590064330046  5835.6553640596
#> 14 0.674973804312221  47.800858201192 187.708962899057 2520.45949461077
#> 15  0.56477572125404  13.383955639665 148.900007392382 3921.49040378132
#> 16 0.319854561938822 116.314845836528 220.058729889787 11598.4567557327
#> 17 0.370299310552204 75.4191630591824 134.393954326387  3120.0520062488
#> 18 0.432157820173835 9.30471702202488 64.5438700246255 4892.16409744979
#> 19 0.335861382896054 80.3327073337749 231.747667031983 9904.87041392434
#> 20 0.327857973799399 106.834378251936 129.027597149226 4580.17761077145
#> 21  0.40125247219485  12.708022484514 98.8736350486356 3021.18432952971
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE 86.9140044167799 95.1771776574733     14.2536055575238
#> 2             TRUE  28.438067791849 35.8880871469438     27.3540587833009
#> 3             TRUE 62.4639557281027 68.0624228590669     18.0876655531967
#> 4             TRUE 238.562048228212 246.175971379194     14.4530231288022
#> 5             TRUE  61.935992937652 70.4780629765012     19.4467776084002
#> 6             TRUE 201.135100432774 209.858803208443     11.5546120537954
#> 7             TRUE 488.089799525164 524.239624221571     6.28775000039559
#> 8             TRUE 185.752219541823 214.872540881569     12.0226646776409
#> 9             TRUE 280.011268903768 302.361461588777     3.99853361236691
#> 10            TRUE  644.46809537042 709.413716405963   0.0758594136432022
#> 11            TRUE 306.174822991122 347.212163310963     9.04706329022926
#> 12            TRUE 323.005726802035 347.554718394631     2.95571240727645
#> 13            TRUE 691.572118495537 765.410507530438    0.095030472482587
#> 14            TRUE 323.910336251267 372.439281775222     8.61173756232438
#> 15            TRUE 310.335330036367 332.952425075956     2.50321935572117
#> 16            TRUE 742.365218700354 807.235187257592 -0.00665421527649535
#> 17            TRUE 317.689753159373 357.605956014279     7.79152714145946
#> 18            TRUE 334.544100193176 363.410097941453     4.16932719287219
#> 19            TRUE 747.347076356549 819.820053729859    0.106534833402876
#> 20            TRUE 337.695176349582  379.60541588015      9.2227688007352
#> 21            TRUE  327.99652564444 346.067713967633     4.41446852981682
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3567171966152 3378.84493028538 3450.04392157441    79489.7378806474
#> 2   32.7845050768686 1411.80995283994 1504.03661162191    40066.2507591362
#> 3   23.0855385259624 2186.17784183042  2203.5285408501    64222.5880849644
#> 4   15.5348486041348 7424.28674723853 7678.11935221296    142286.921727295
#> 5   26.3955599416611 2334.49563630112 2415.86769612885    55996.0645250423
#> 6   13.1747620139769 4800.10395351249 4837.57546065006    127960.420369186
#> 7   7.69685727399511 16859.8157953596 17271.7127402499    352671.042149213
#> 8   14.9186840034347  5143.9610214062 5241.50900864467    116196.293448205
#> 9   5.59092517576306 7735.73872996081   7784.707051365    214599.087555078
#> 10 0.852332436732683 19961.2323149018 20246.3947297907     471883.02977317
#> 11  11.0500186904022 6654.64581820612 6941.27849492866    161687.656485207
#> 12  4.77782093059227  9163.5031220011 9214.64194186783    260665.951752074
#> 13  1.44650788081642 20393.4522470567 20547.2382375523    484486.805942603
#> 14  10.8324013785116 6890.89676478775 7048.16158826967    172816.891894043
#> 15  4.36133147864696 8979.22605607857 9023.25927013307    259294.546043254
#> 16  1.04566729350223 20471.7017583212 20854.3776011233    493566.247642327
#> 17  9.00981187317621 6804.85825786797 7052.98730433268    171364.519800398
#> 18  5.59112642124411   9305.823267529 9336.43578653146     262004.16686947
#> 19  1.21151878313089 19658.2030884302 19922.4976955583    484621.183960184
#> 20  10.3014215345352 6979.65466415754  7331.1397686064    173579.596793154
#> 21  5.73458916333787 9236.67601680187 9278.48541077592    263925.069875648
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84092.3846595375 86.1228495320326 95.9683325422206     14.0522438048448
#> 2     44040.8690848511 27.7247680663612 36.6013868724316     26.8341224360444
#> 3     67845.1294991915 61.9279322793934 68.5984463077762      17.609145800485
#> 4     155580.120569455 237.833055586096 246.904964021309       14.34944409393
#> 5     58915.9454192772 61.1181351679749 71.2959207461783     18.7814686615986
#> 6     134441.387676004 200.299852294679 210.694051346539     11.3994913129269
#> 7     373623.113760464 484.628646096785  527.70077764995     6.15283547419989
#> 8     122534.620235583 182.964103668869 217.660656754523     11.7453862315542
#> 9     227199.199021822 277.871356838183 304.501373654363     3.84607059033962
#> 10    501069.438112613 638.249897611697 715.631914164686  0.00151625185803916
#> 11    173130.435764753 302.245715939223 351.141270362862     8.85529096468079
#> 12    272750.537436764 320.655291436787  349.90515375988     2.78125520823558
#> 13    503686.112090359 684.502485502833 772.480140523141  -0.0343663006557583
#> 14    181109.203631313 319.263947850037 377.085670176452     8.39912081396603
#> 15    272196.249471694 308.169863490023   335.1178916223     2.32531500352614
#> 16    531725.170368687 736.154264264022 813.446141693923   -0.107408402287224
#> 17    181629.490900957 313.867989056244 361.427720117408     7.67488285863551
#> 18    278099.386750079  331.78033445132  366.17386368331     4.03319747951743
#> 19    517208.207621995 740.408174267615 826.758955818793 0.000738497790618964
#> 20    188648.381132592 333.682493841337 383.618098388396     9.11949353898839
#> 21    273864.766319801 326.266305485837 347.797934126236     4.28807400107544
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5580789492942 3372.02800558749 3456.86084627229    79049.0589337323
#> 2   33.3044414241251  1402.9797408289 1512.86682363295    39685.7021960358
#> 3   23.5640582786741 2184.51660469023 2205.18977799028    63875.7490133895
#> 4   15.6384276390071 7399.98362548566 7702.42247396583    141014.168646663
#> 5   27.0608688884627 2326.70469440273 2423.65863802725    55716.5014607007
#> 6   13.3298827548454 4796.51625602059 4841.16315814196    127339.902222789
#> 7   7.83177180019081 16820.3788538276 17311.1496817819    350664.992739626
#> 8   15.1959624495214 5134.62132050039 5250.84870955048    115589.432372818
#> 9   5.74338819779035 7731.05027365616 7789.39550766965    213392.693904007
#> 10 0.926675598517846 19933.9295304975 20273.6975141949    469088.586421521
#> 11  11.2417910159507 6627.20226405183 6968.72204908295    160592.071235038
#> 12  4.95227812963315 9158.60685201386 9219.53821185506    259508.916952476
#> 13  1.57590465395476 20378.7280564773 20561.9624281316    482648.574502924
#> 14  11.0450181268699 6875.83949445438 7063.21885860305    172022.947153241
#> 15  4.53923583084198 8975.01011005208 9027.47521615956    258059.276566063
#> 16  1.14642148051296 20435.0625818826 20891.0167775618    489912.733764271
#> 17  9.12645615600015 6781.10122150433 7076.74434069632     170381.70341843
#> 18  5.72725613459887 9302.89228166706  9339.3667723934    260463.135178773
#> 19  1.31731511874315 19632.8982856201 19947.8024983685    481501.149779798
#> 20   10.404696796282 6946.00183500818 7364.79259775576    172136.840845761
#> 21  5.86098369207925 9232.67298971924 9282.48843785854    262973.396811846
#>    ci95.integral.bt.up
#> 1     84533.0636064525
#> 2     44421.4176479515
#> 3     68191.9685707664
#> 4     156852.873650087
#> 5     59195.5084836189
#> 6     135061.905822401
#> 7     375629.163170052
#> 8     123141.481310971
#> 9     228405.592672894
#> 10    503863.881464261
#> 11    174226.021014922
#> 12    273907.572236362
#> 13    505524.343530038
#> 14    181903.148372115
#> 15    273431.518948885
#> 16    535378.684246743
#> 17    182612.307282925
#> 18    279640.418440776
#> 19    520328.241802381
#> 20    190091.137079985
#> 21    274816.439383602