Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.5362473283375  15.5927061751356
#> 2                 FALSE          0.75 31.8000041939138  31.8938089761736
#> 3                 FALSE          0.75 64.8541256419349  20.6020094583357
#> 4                 FALSE          0.75  242.75718958607   14.975332486873
#> 5                 FALSE          0.75 65.8944482101323  20.7453043688616
#> 6                 FALSE          0.75 204.766963450061  12.2303881533899
#> 7                 FALSE          0.75 511.868393038619  7.17133699530292
#> 8                 FALSE          0.75 203.915596862762  13.3311781513818
#> 9                 FALSE          0.75 287.711489609714   4.7182022502812
#> 10                FALSE          0.75  670.71992681955 0.526884634407709
#> 11                FALSE          0.75 330.581022521089  10.0894612768275
#> 12                FALSE          0.75 337.516243698012  3.95580561781824
#> 13                FALSE          0.75 718.133089883385 0.597030885093768
#> 14                FALSE          0.75 344.383464194415  9.69858035541854
#> 15                FALSE          0.75 317.532645434158  3.23895978463792
#> 16                FALSE          0.75 770.948246997165 0.354223240303791
#> 17                FALSE          0.75 337.582945631131  8.14495327288001
#> 18                FALSE          0.75 347.945116483698  4.91037514920997
#> 19                FALSE          0.75 774.804364759231 0.679606152028546
#> 20                FALSE          0.75 347.027978451235  9.49664329165493
#> 21                FALSE          0.75 338.202043798571  5.21881336261642
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.28466163247 2116.44290005399 81283.1887558262 2.55031488777401
#> 2  1459.18220341114 268.476767122745 42068.4563955573 2.08414487349022
#> 3  2194.04632100224 652.810618409876 66148.6379331223  1.7043542673267
#> 4  7554.81765148783 6038.30149234348  148727.32100991 2.39933727089034
#> 5  2358.45226264703 1145.44409645958 56853.5557059146  2.5876284437815
#> 6  4818.23869416765  3260.5242062097 130841.480971004 1.60039689599464
#> 7  17042.5898396387 15837.2306938793 361722.424092024 12.3005823914231
#> 8  5178.84866034054 3799.21604900931 118844.323229823 8.46491253948659
#> 9  7766.27809432345 6266.15902327938 220920.571509531 7.20039374972037
#> 10 20084.2262786117 18745.8986406554 483556.414938891 25.1776474193443
#> 11 6769.04720967294 5535.68518170364 167001.126992732 12.9455412788213
#> 12 9191.74489677397 7651.27667514498 264964.932183724  8.3408452824273
#> 13 20439.5636008861 19420.4789509102 492012.178787213 17.4833267063044
#> 14   6983.915012649 5700.23078397777 177301.817481693 17.3450858626693
#> 15 9005.41668635484 7416.63393723298 264999.792105159 6.62480802277881
#> 16 20643.9431204006 19409.2113802648  509577.97405637 15.1485841855736
#> 17 6897.11084690868 6016.29314088566 175925.142543904 10.3618697055214
#> 18 9324.07652086807 7651.26873921407 271301.707775924 5.46719622482003
#> 19 19807.9076218422 18586.6845577643 504342.996370307 36.3944212737005
#> 20 7164.72641441502  5826.7476729102 181325.020469943  12.855260580472
#> 21 9257.32785180155 7462.82578412692 268964.970442222 8.04419900587483
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.599308941634987 38.0889869991804 54.3834075765464 2649.82999779049
#> 2   1.88766308835904 28.8299816814277 87.3510047960465 1222.54767457678
#> 3   1.79786932771263 6.70105541425432  122.69948781678 1101.99877473945
#> 4  0.269340400297217 55.2814164374781 62.6467016550646 3038.87620148569
#> 5   5.30167850223921 40.2590626220345 160.714804246926 1604.31301207853
#> 6  0.488916648265678  7.5001357188769 98.4703217484705 2402.68780150445
#> 7  0.463411964704637 161.848215031811 168.237304378447 9289.85496472988
#> 8  0.637942838079034 43.4919241261165 133.896283151284 1969.06242492896
#> 9  0.393872281338396 27.9044138280806 88.4098482358021 3408.64144424237
#> 10 0.205434828935451 71.3333345736152 221.571308632216 9497.33799541562
#> 11 0.700150114891122 105.296190269687 262.195471937937 3884.12636549564
#> 12  0.61167874230014 11.0720121448466 145.144314656688 6650.55375042582
#> 13 0.244230434887246 82.9357357187055 173.363892098476 8186.47857188507
#> 14 0.639450136719306 42.1709411336969  187.89646285393 2401.22814755428
#> 15 0.346966091627565 15.9659834790284 48.9197432932046 2831.28761552128
#> 16 0.244939228914575 102.035725420588  198.62995512029 11405.5127571807
#> 17 0.512283201032183  75.972317589364 156.118035364242  2643.0022481883
#> 18 0.270502042747829 11.7922276917047 85.9562412006666  3103.0417090218
#> 19  0.34500231583378  99.138423646826 193.702141633388 6470.05830677934
#> 20 0.347276741537027 61.3524685829254 133.223449810746 1886.36615405554
#> 21 0.565893574823027 10.0341933494632 152.963310616868 2631.22187250152
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.3409793379492 96.7315153187257     14.606842966146
#> 2             TRUE 28.3715858770224 35.2284225108052     28.788603195823
#> 3             TRUE 62.0504628721825 67.6577884116873    17.6445144142484
#> 4             TRUE 238.810279775455 246.704099396684    14.5322675283841
#> 5             TRUE 61.6377994201118 70.1510970001529    12.0240432326781
#> 6             TRUE  202.13431055615 207.399616343972    11.4261202669929
#> 7             TRUE 491.633935004728  532.10285107251    6.40902431336379
#> 8             TRUE 189.990815735307 217.840377990217    12.2817621827418
#> 9             TRUE 275.866841891424 299.556137328004    4.07028234747954
#> 10            TRUE 629.302696814729 712.137156824372   0.188944340808892
#> 11            TRUE 309.285607117428  351.87643792475    8.93771433783163
#> 12            TRUE 323.795553208419 351.236934187605    2.94959408673451
#> 13            TRUE 689.373017451515 746.893162315256   0.195271819704248
#> 14            TRUE 315.850797950324 372.916130438506    8.64668488051528
#> 15            TRUE 306.634836236687 328.430454631629    2.66820056391058
#> 16            TRUE 746.028826011897 795.867667982434 -0.0487017912606857
#> 17            TRUE 320.537669965549 354.628221296714    7.30224740718207
#> 18            TRUE 338.951578693869 356.938654273527    4.46539928888979
#> 19            TRUE 714.935541763993 834.673187754468   0.112077342481978
#> 20            TRUE 325.881074796359 368.174882106112    8.92537305182652
#> 21            TRUE 324.969336433907 351.434751163235    4.28791843203254
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5785693841251 3348.62827801881 3473.94104524612    76924.2184094609
#> 2   34.9990147565243 1411.75688354519 1506.60752327708    40057.3654708785
#> 3    23.559504502423 2183.02308484579 2205.06955715868    64335.8499486759
#> 4   15.4183974453619 7463.87972144818 7645.75558152748    143728.369658466
#> 5   29.4665655050452 2292.22610463378 2424.67842066028    54214.4608010454
#> 6   13.0346560397869  4805.9009709101  4830.5764174252    126889.059537529
#> 7   7.93364967724205 16776.3495259114 17308.8301533661    346440.612675044
#> 8   14.3805941200218 5107.30444515308   5250.392875528    115605.215540815
#> 9   5.36612215308286 7720.37533357626 7812.18085507064    215313.356333752
#> 10 0.864824928006526 19966.8829432381 20201.5696139853    467933.293936433
#> 11  11.2412082158234  6595.8349766793 6942.25944266657    160611.739121492
#> 12  4.96201714890197  9173.5314367957 9209.95835675224    254024.771264273
#> 13 0.998789950483288 20303.1343156289 20575.9928861434    478545.421536462
#> 14  10.7504758303218 6914.54381448407 7053.28621081393    173351.797178966
#> 15  3.80971900536527 8979.15264353183 9031.68072917784    260342.323977627
#> 16 0.757148271868267 20476.0943520837 20811.7918887175    490815.905570808
#> 17  8.98765913857795 6772.13638447418 7022.08530934318    171577.403845634
#> 18  5.35535100953015 9304.67830631522 9343.47473542092    266197.204164583
#> 19  1.24713496157511 19644.8249149432 19970.9903287412    493699.750455655
#> 20  10.0679135314833  7063.8016035961 7265.65122523393    178221.948146522
#> 21   6.1497082932003 9240.82160374168 9273.83409986141    264636.610461957
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     85642.1591021916 87.5376301483004 97.5348645083745     14.418060649531
#> 2     44079.5473202361  27.715080241873 35.8849281459547    28.1939893229899
#> 3     67961.4259175687 61.5135912779746 68.1946600058952     17.078185576019
#> 4     153726.272361354 238.054488535125 247.459890637015    14.4474253022904
#> 5     59492.6506107838 60.8226964603206 70.9661999599441    10.3540145044728
#> 6     134793.902404479 201.630185533912 207.903741366211    11.2721115227892
#> 7     377004.235509005  487.75925155143 535.977534525809    6.26304954448183
#> 8     122083.430918831 187.324368285368 220.506825440156    12.0808101887469
#> 9     226527.786685309 273.598717860262 301.824261359166    3.94621257885795
#> 10     499179.53594135 621.371737877635 720.068115761465   0.124232369694225
#> 11    173390.514863973 305.207761614599 355.954283427579    8.71716705164092
#> 12    275905.093103174 321.168186944454 353.864300451569    2.75691528290997
#> 13    505478.936037964 683.865769539029 752.400410227742   0.118339232714766
#> 14     181251.83778442 310.387095903583 378.379832485247     8.4452580874487
#> 15    269657.260232692 304.548021709512 330.517269158805    2.55890624504789
#> 16    528340.042541932 741.257021993441  800.63947200089  -0.125857648368777
#> 17    180272.881242174 317.273681008309 357.892210253953    7.14087819885693
#> 18    276406.211387264 337.229411883051 358.660821084345    4.38019114542423
#> 19    514986.242284959 703.471299062778 846.137430455684 0.00340161299433761
#> 20    184428.092793364  321.83166771351  372.22428918896    8.81598087824236
#> 21    273293.330422487 322.435413747056 353.968673850086    4.10966195596328
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7673517007402 3336.63024711407 3485.93907615086    76089.5219601568
#> 2   35.5936286293574 1402.67543931554 1515.68896750673    39672.2629533868
#> 3   24.1258333406525  2180.9122523903 2207.18038961417     63988.720334633
#> 4   15.5032396714555 7446.46607527037 7663.16922770528    142771.123654999
#> 5   31.1365942332505 2279.54449990784 2437.36002538622    53709.1022022407
#> 6   13.1886647839906 4803.53842815865 4832.93896017665    126132.212880056
#> 7   8.07962444612401 16725.3673381764 17359.8123411011    343514.308361154
#> 8   14.5815461140167 5093.60448905335 5264.09283162773    114984.960876962
#> 9   5.49019192170446 7711.58544322041 7820.97074542649    214239.634278816
#> 10 0.929536899121193 19944.4129428475  20224.039614376    464941.632467877
#> 11  11.4617555020141 6562.66667674435 6975.42774260153    159388.239316361
#> 12  5.15469595272652 9170.04375297007 9213.44604057787    251929.846832889
#> 13  1.07572253747277 20277.0095588775 20602.1176428948    475966.680786318
#> 14  10.9519026233884 6901.25996802695 7066.57005727104    172595.410312486
#> 15  3.91901332422795 8974.12335873594 9036.71001397373    259450.468378738
#> 16 0.834304128976358 20443.9530985762 20843.9331422249    487223.169052296
#> 17  9.14902834690309 6748.20510443353 7046.01658938383    170744.858137455
#> 18  5.44055915299572 9300.96375459233 9347.18928714381    265219.746026241
#> 19  1.35581069106275 19613.5963114944   20002.21893219     491661.68208902
#> 20  10.1773057050675 7044.47557599248 7284.97725283755    177627.742807994
#> 21  6.32796476926955  9237.6608328366  9276.9948707665    263807.775572119
#>    ci95.integral.bt.up
#> 1     86476.8555514956
#> 2     44464.6498377277
#> 3     68308.5555316116
#> 4     154683.518364822
#> 5     59998.0092095886
#> 6     135550.749061953
#> 7     379930.539822895
#> 8     122703.685582684
#> 9     227601.508740246
#> 10    502171.197409906
#> 11    174614.014669104
#> 12    278000.017534558
#> 13    508057.676788107
#> 14    182008.224650899
#> 15    270549.115831581
#> 16    531932.779060444
#> 17    181105.426950353
#> 18    277383.669525606
#> 19    517024.310651595
#> 20    185022.298131892
#> 21    274122.165312325