Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0736642883379  15.5940803109497
#> 2                 FALSE          0.75 32.7791037771887   31.367602598818
#> 3                 FALSE          0.75  64.546834851444  20.6893306864588
#> 4                 FALSE          0.75 242.975571579373    15.02595654844
#> 5                 FALSE          0.75 64.8899091152305  22.4209313714309
#> 6                 FALSE          0.75   205.6596305753  12.3987585585721
#> 7                 FALSE          0.75 509.148554948449  7.19185854539174
#> 8                 FALSE          0.75 203.834537282786  13.4286714832099
#> 9                 FALSE          0.75  286.74629676028  4.51131975568596
#> 10                FALSE          0.75 669.481797374673 0.494193559873044
#> 11                FALSE          0.75 326.091988859404  9.63606408680469
#> 12                FALSE          0.75 334.463704395893  3.84853986611805
#> 13                FALSE          0.75 725.984403698412 0.842751148881478
#> 14                FALSE          0.75  346.62526427495  9.87090129342383
#> 15                FALSE          0.75 320.262569992047  3.41335222743319
#> 16                FALSE          0.75 769.174458991796 0.456557829683332
#> 17                FALSE          0.75 338.909496534049  8.19981226473024
#> 18                FALSE          0.75  351.92382764383  5.01184441256931
#> 19                FALSE          0.75 770.580472152824 0.640547291879385
#> 20                FALSE          0.75  357.56780864263  9.52319184324763
#> 21                FALSE          0.75 334.343414291552  5.02030316204657
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3409.4410831535 2109.64994078395 81514.4220410217 1.92547662088019
#> 2  1459.83639532671 279.407604619972 42499.7490018879 2.15295918595576
#> 3  2197.54454913912  649.04932784903 65659.4681222223 1.60662167605083
#> 4  7535.85427873683 6005.25112679362 148062.942428071 1.97432065101021
#> 5  2370.02213953939 1149.04150028098 57596.9496188408  3.0047163602037
#> 6  4820.92042273616 3232.89365673752  130628.35728719 1.99350477628096
#> 7  17003.3135631407 15744.3528184749 360941.488843862 16.3072288442258
#> 8  5193.07399983136 3789.39672328562 119236.290412398 8.78273701648094
#> 9  7761.56668652425 6293.02944932428 221699.190880106 7.26242071941275
#> 10 20116.1472472756 18806.2107743407  484703.44000667 21.0239950327153
#> 11 6797.54120083805 5657.78984619595 168154.777659252 12.6070749812844
#> 12 9183.09093176551 7625.48097964797 267456.238963077 6.05831626011807
#> 13 20413.4046180117 19306.5954998324 488226.338675366 18.7910673329851
#> 14 6988.28707704977 5663.58719709596 177293.996475211 10.6080605851047
#> 15 8998.76727897914 7382.00757303762 265663.051553742 4.84106227641053
#> 16 20671.8661163385 19344.8122924645 515038.570614263 17.4965806264134
#> 17 6894.60480068227 6012.39663027874 175223.734976223 9.57389106807436
#> 18 9322.82488088672 7666.20047757342 269874.477502837 7.05417884785424
#> 19 19806.1627887038 18591.9320044006 503030.039190701 18.6880202464402
#> 20 7158.22956343449 5863.53662302332  181031.67443099 16.0612996693755
#> 21 9254.58900666281 7496.72213592244 268891.562388359 6.23359358883963
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.488884109816981  23.830704447446 53.5982975286798 1397.64799810469
#> 2   2.06160019691808 23.4918254498326 72.3265405212924  835.63451910088
#> 3   1.16092449811577 8.50808821954604 81.1221215038648  2259.5870260393
#> 4  0.261918134991327 78.3018594039515 101.257646401686  4311.3309036693
#> 5   2.21724709456998 28.2338262186982 157.424002568707 1027.95902394592
#> 6  0.496747010614173 10.0711955342136 102.464441113545 3060.74137064504
#> 7  0.558715907602506 205.685950233059 227.875270988821 10869.3085199111
#> 8  0.671947104436794  25.385901700258 137.761196433352 1620.19716654747
#> 9  0.447778547656827 23.2042854856371 66.3683880005922 4708.95711202362
#> 10 0.291492469662218  81.140989116022 247.012993783929 8818.44396000711
#> 11 0.505740639757451 64.1355529009886 164.059505878667 2507.62832193288
#> 12 0.494288689774376 12.2274028317135 113.601957489336 3188.72200289621
#> 13 0.326350572300556 95.3012515443498 248.506766120328 10148.8091401784
#> 14 0.681201124052747 52.0922585274194 191.729320631621 2393.51399387442
#> 15 0.489183655646508 15.4745999172635 116.732272983667 2049.98796195087
#> 16 0.330485335756333 82.8658952634854 247.010027107675 9368.80446639238
#> 17  0.54106433401866 71.4702886423942 156.475038269805 3181.19249691277
#> 18 0.306986573306333 7.32805474747089 68.5320750170243 3764.44750793932
#> 19 0.294791220801457 71.0856508817262 205.337678042814 8412.05327926093
#> 20 0.498878985191405  64.613582279591 109.177265105076 1843.83616767268
#> 21 0.463550678690274 8.52288232411985 136.612397595471 3227.57229587824
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE   88.90625524699 95.2410733296858    14.7898659503008
#> 2             TRUE 29.2374859162915 36.3207216380859    27.9762702748877
#> 3             TRUE 61.9039421943403 67.1897275085476    18.7796098870583
#> 4             TRUE 239.727814108461 246.223329050284    14.5951012163793
#> 5             TRUE 59.9471507026954 69.8326675277656    18.7735599008632
#> 6             TRUE 202.380315218318 208.938945932282    11.5816097261118
#> 7             TRUE 482.323163499697   535.9739463972    6.27277087738562
#> 8             TRUE 189.386934890675 218.282139674898    12.3233184964114
#> 9             TRUE 274.799614676846 298.692978843714    3.77472404479048
#> 10            TRUE 634.897325545857  704.06626920349  0.0146884472786958
#> 11            TRUE 305.353350515191 346.830627203617    8.80412073440368
#> 12            TRUE 324.497774147999 344.429634643788     3.0354349714392
#> 13            TRUE 695.073097935652 756.895709461173   0.305904457447064
#> 14            TRUE 329.175004612453 364.075523937448    8.75032544435706
#> 15            TRUE 312.299022547352 328.226117436742    2.60864511389468
#> 16            TRUE 740.392583861346 797.956334122246 -0.0870905476358352
#> 17            TRUE 323.160445727066 354.658547341031    7.30976143526955
#> 18            TRUE 340.319703439109  363.52795184855    4.50685149948039
#> 19            TRUE  739.83867884743 801.322265458218   0.155615733660989
#> 20            TRUE 331.146970686507 383.988646598752    8.70253591260776
#> 21            TRUE  324.08915283791 344.597675745193    4.25776229560107
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3982946715987 3370.23957433745 3448.64259196955    79215.2910841395
#> 2   34.7589349227482 1421.19234246173 1498.48044819168     41125.130217967
#> 3   22.5990514858592 2183.54874401796 2211.54035426027    61942.4474643876
#> 4   15.4568118805008 7407.04772001733 7664.66083745633    140970.803091535
#> 5   26.0683028419985 2323.57749540963 2416.46678366915    55905.9570244498
#> 6   13.2159073910324 4804.35330608238 4837.48753938994    125593.437732479
#> 7   8.11094621339786 16664.9601750073  17341.666951274    343061.476328608
#> 8   14.5340244700085 5151.31419153443 5234.83380812828    116571.066073428
#> 9   5.24791546658144 7723.39563690038 7799.73773614812    213952.956430827
#> 10 0.973698672467393 19982.6703201797 20249.6241743714    470197.099692458
#> 11  10.4680074392057 6692.03821631592 6903.04418536017    164029.729069673
#> 12   4.6616447607969 9162.97685410734 9203.20500942368    262210.791268313
#> 13  1.37959784031589 20256.6340592213 20570.1751768022    471531.547639772
#> 14  10.9914771424906 6902.59531177216 7073.97884232737    173356.665955287
#> 15  4.21805934097169 8973.31156211524 9024.22299584304    262290.821356332
#> 16   1.0002062070025 20535.5517186301 20808.1805140469    499626.887267047
#> 17  9.08986309419094 6777.03617586553 7012.17342549901    169990.673318802
#> 18  5.51683732565822 9310.77023082713 9334.87953094631    263681.961352276
#> 19  1.12547885009778 19689.2268930033 19923.0986844042    489192.211546316
#> 20  10.3438477738875 7051.94022058456 7264.51890628442    177998.563935168
#> 21  5.78284402849207 9240.56886523963 9268.60914808599     263582.20596164
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83813.5529979039 88.2997301114127 95.8475984652631    14.6358674557085
#> 2     43874.3677858088 28.5593037727154  36.998903781662    27.3268662128585
#> 3     69376.4887800569 61.3978563663843 67.6958133365036    18.4139186701519
#> 4     155155.081764607 239.105903103393 246.845240055353     14.512597003857
#> 5     59287.9422132319 59.0006650492312 70.7791531812297    18.0751270660737
#> 6     135663.276841901 201.752361213789 209.566899936811    11.4251344177683
#> 7     378821.501359116 477.186386413766 541.110723483131    6.09677536649083
#> 8     121901.514751369 186.620372730484 221.048701835089    12.1116551585138
#> 9     229445.425329385 272.511952150231 300.980641370329    3.63367380227858
#> 10    499209.780320881 628.274767110551 710.688827638795 -0.0771316806649028
#> 11    172279.826248832 301.382121896087 350.801855822721    8.64481243288009
#> 12    272701.686657841 322.589404526062 346.338004265725    2.87973403416027
#> 13    504921.129710959 689.153911725761 762.814895671063   0.203104027172389
#> 14    181231.326995134 325.833465528145 367.417063021756    8.53574709028045
#> 15    269035.281751151 310.774087930282 329.751052053812    2.45455226236603
#> 16    530450.253961478 734.881160964025 803.467757019566   -0.19119342839908
#> 17    180456.796633645 320.144670040623 357.674323027474    7.13932617005367
#> 18    276066.993653397 338.097637102035 365.750018185624    4.41015072888889
#> 19    516867.866835085 733.951952469802 807.208991835847  0.0627564991085301
#> 20    184064.784926811 326.087661290654 389.047955994606    8.54538903227247
#> 21    274200.918815079 322.125570857426 346.561257725677    4.11174383181363
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    16.552293166191  3362.7329024365 3456.14926387049    78775.0319647365
#> 2   35.4083389847774 1413.79241744503 1505.88037320838    40861.9053444502
#> 3   22.9647427027657 2180.86869622881 2214.22040204943    61230.6775511852
#> 4    15.539316093023 7382.38263430509 7689.32592316857    139612.733856879
#> 5    26.766735676788 2314.68384015074 2425.36043892804    55582.1499319068
#> 6   13.3723826993759  4801.1808794891 4840.65996598322    124629.304200726
#> 7   8.28694172429265 16600.1691006839 17406.4580255975    339637.644144836
#> 8   14.7456878079061 5143.31763249885 5242.83036716386    116060.703965965
#> 9   5.38896570909334  7716.0862869724  7807.0470860761     212469.63494054
#> 10  1.06551880041099 19957.1109086082  20275.183585943    467419.289845056
#> 11  10.6273157407293 6671.83551715211 6923.24688452398    163239.826148264
#> 12  4.81734569807583 9159.12522221535 9207.05664131567      261206.3438374
#> 13  1.48239827059057 20226.6141649848 20600.1950710387    468334.672760616
#> 14  11.2060554965672 6886.18625033603 7090.38790376351    172602.709047217
#> 15  4.37215219250034  8968.4370631413 9029.09749481697    261645.075148318
#> 16  1.10430908776574 20509.4489616221 20834.2832710549    496675.713860133
#> 17  9.26029835940682 6754.52303494318 7034.68656642136    168988.597682274
#> 18  5.61353809624972 9308.46189358168 9337.18786819177    262496.160387276
#> 19  1.21833808465024 19666.8349129756 19945.4906644319    486542.414763349
#> 20  10.5009946542228 7031.58694216649 7284.87218470249    177417.755542351
#> 21  5.92886249227951 9237.88415730754 9271.29385601809    262565.520688438
#>    ci95.integral.bt.up
#> 1     84253.8121173069
#> 2     44137.5926593256
#> 3     70088.2586932593
#> 4     156513.150999262
#> 5     59611.7493057748
#> 6     136627.410373654
#> 7     382245.333542888
#> 8     122411.876858831
#> 9     230928.746819673
#> 10    501987.590168284
#> 11    173069.729170241
#> 12    273706.134088754
#> 13    508118.004590115
#> 14    181985.283903204
#> 15    269681.027959165
#> 16    533401.427368392
#> 17    181458.872270172
#> 18    277252.794618398
#> 19    519517.663618052
#> 20    184645.593319628
#> 21    275217.604088281