Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9819656877686  15.3254849084202
#> 2                 FALSE          0.75 33.1809432434336  32.0081366346347
#> 3                 FALSE          0.75 64.5016228161022  19.9397548187663
#> 4                 FALSE          0.75 243.006635003984  15.0451775196681
#> 5                 FALSE          0.75 65.5045713772095  23.7171537363808
#> 6                 FALSE          0.75 204.773714912938  12.2942364930323
#> 7                 FALSE          0.75 512.764747429043  7.16531847343229
#> 8                 FALSE          0.75 202.176630160688  13.1042747837513
#> 9                 FALSE          0.75 291.923716945271  4.71737997828045
#> 10                FALSE          0.75 666.579718643954 0.642341447971397
#> 11                FALSE          0.75 329.553321154493  9.95683492659158
#> 12                FALSE          0.75 334.998502973337  3.78304798948195
#> 13                FALSE          0.75 723.621206179771 0.613102569932525
#> 14                FALSE          0.75 340.279950766676  10.0124292844192
#> 15                FALSE          0.75 319.453201507254  3.22578235790549
#> 16                FALSE          0.75 774.036275060354 0.475876803316684
#> 17                FALSE          0.75 331.544094975417  8.04848080163651
#> 18                FALSE          0.75 349.518412812867  4.91351456991381
#> 19                FALSE          0.75 783.000833582944 0.642128099185709
#> 20                FALSE          0.75 349.762335637313  9.48117150102778
#> 21                FALSE          0.75 336.765487652095  4.99490934196234
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3405.72772521095  2133.0941893621 81558.3665492373  2.3611845566828
#> 2  1473.23862073052 282.198348417665 42900.6307138467 2.43853685359464
#> 3  2197.20744653333 702.704381532865 66079.6731131945  1.0414624778754
#> 4  7537.27790650002 6007.30652926975 147907.363260308 2.14457474591968
#> 5   2366.5422652167 1068.99456103341 57286.3483723826 2.54493289477606
#> 6  4817.80390408459 3244.77735425203 131202.763055401  3.0428638868913
#> 7  17036.0477188468 15837.5440394611 361348.244698123 12.5502331307958
#> 8  5178.94381361894 3833.04231077937  117849.34654097  7.1072590186407
#> 9  7766.23288795848 6293.22316689832  221559.43211524 5.50241796095951
#> 10 20051.0918336753 18619.8958921012   481302.6997765 28.9632304709694
#> 11    6742.82941244 5554.40273064981  165835.69584964  12.723188072654
#> 12 9187.69065046852 7674.53461120076 267139.855346204 4.94642280678975
#> 13 20386.4971978832 19403.3846224518 484863.317651224 30.5874964454323
#> 14 6964.48914955147 5582.17758651497 176579.831686685 15.6674768518602
#> 15 9000.84705206105 7433.68495358522  264536.52616443 6.82536108857874
#> 16 20656.3914015447  19326.957082154 512216.995730475 20.7657563515617
#> 17 6884.02105351963 5976.80645772384 175521.490035516  12.766696701528
#> 18  9317.2858516291 7659.14081457161 268970.103247148 6.29272086104899
#> 19 19817.1210731825 18706.2442959542 499513.527738143  18.942186949101
#> 20  7140.4649264534 5813.10170921461 180992.483488959 19.2308902014562
#> 21 9257.82255933448 7529.42321784818 268097.931176806 5.57957606570255
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.640828041006794 41.5363033995562 56.6673252224596 2619.41380567555
#> 2   2.23385981801677 21.0083855189775 69.2573528570643 787.217029766265
#> 3   1.89087860474924 7.70542659391008  126.14803320086 885.286543203879
#> 4  0.335376810538222 56.7316095263158 114.811184044992 2886.44952840426
#> 5   2.27700441539556 35.7987051479586 92.2518727340231 1732.05276583146
#> 6  0.382230342288055 11.0721661179584 83.5291367614976 1362.55673269333
#> 7  0.410432897536898  90.309272724319 91.5130869967619 6359.55604702745
#> 8  0.732386872252768 50.0034721142653 108.394338557233 3020.12417822443
#> 9  0.404874690900955 20.0504815526412 84.8897834382038  3422.9953632211
#> 10 0.447159753424364 49.3713315017065 266.275016388474 9465.27647958152
#> 11 0.731491806885694 113.161440276488 220.908764977438 4431.48527861033
#> 12 0.339376902123207 8.88681465520523 93.8381624395607 4234.17313810985
#> 13 0.273505067186854 104.478546224461 247.426793279962 10566.6809093974
#> 14 0.610901326459692 51.4339131997584 195.608976343761  3301.1852771991
#> 15 0.403566118105963 14.4828838822965 68.5112219071577 4188.91854155191
#> 16 0.411931796424813 116.622741997696 279.439233793926  12448.622400195
#> 17 0.333607881838611 104.918730313061 140.322504655493 3940.79679108866
#> 18  0.29904728594184  13.657715430044 72.4083748573112 6039.54495890238
#> 19  0.25326867961312   72.81300614049 232.636349142333 9529.24229482789
#> 20 0.399627057047922 90.3996717430547 150.640633326175 2564.00723882037
#> 21 0.285925471979419 8.31882948833166 70.1938482722234 4499.34217459079
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.0978170920254 95.8661142835118     14.271322780964
#> 2             TRUE 29.1695501192704 37.1923363675968    28.3334372339971
#> 3             TRUE 62.7884170399972 66.2148285922073    16.8292595139538
#> 4             TRUE 239.478809546946 246.534460461022    14.4934826663327
#> 5             TRUE 61.3181567653028 69.6909859891161    19.9714814730551
#> 6             TRUE 199.768203819002 209.779226006875    11.6654675799684
#> 7             TRUE 492.119613928884 533.409880929202    6.49015635698409
#> 8             TRUE 190.485189075024 213.868071246352    11.8994983788955
#> 9             TRUE 282.872239399493  300.97519449105    4.05136111174838
#> 10            TRUE 618.935204519209 714.224232768699 -0.0932363464116825
#> 11            TRUE 308.623676774977 350.482965534009    8.75353090426462
#> 12            TRUE 326.861637456168 343.135368490506    3.22477298548928
#> 13            TRUE 673.304774527034 773.937637832507    0.16318673441015
#> 14            TRUE 314.506951345366 366.052950187986    9.00749660239299
#> 15            TRUE 308.225482516542 330.680920497966    2.56191609362118
#> 16            TRUE 739.876605862035 808.195944258673  -0.201751001802133
#> 17            TRUE 310.542878901403  352.54531104943      7.499695836012
#> 18            TRUE 339.166886996441 359.869938629292    4.42158178453948
#> 19            TRUE 751.840936051673 814.160731114215   0.225501121222126
#> 20            TRUE 318.127521255918 381.397150018709    8.82378499218395
#> 21            TRUE 327.587085024015 345.943890280176     4.5245619405562
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3796470358764 3337.40050611868 3474.05494430322     77249.430838901
#> 2   35.6828360352723  1438.6798265518 1507.79741490923    41605.6586998812
#> 3   23.0502501235788 2184.53201978634 2209.88287328031    64623.3767496241
#> 4   15.5968723730035 7443.95440882924 7630.60140417081    143159.153786083
#> 5   27.4628259997065 2307.65339524831 2425.43113518509    54437.1215725899
#> 6   12.9230054060961 4799.59019082055 4836.01761734863     128961.35723012
#> 7   7.84048058988049 16887.4889652153 17184.6064724783    350886.775000762
#> 8   14.3090511886071 5096.68810199098 5261.19952524691    112881.242267791
#> 9   5.38339884481252 7733.24984580439 7799.21593011258    215928.604742741
#> 10  1.37791924235448  19969.875993355 20132.3076739956    465732.319967589
#> 11  11.1601389489185 6556.67884318518 6928.97998169483    158545.902566326
#> 12  4.34132299347463  9173.0718403607 9202.30946057633    260174.640534014
#> 13   1.0630184054549 20214.6299893439 20558.3644064224    467481.127555265
#> 14  11.0173619664454 6879.88036233787 7049.09793676508    171149.381905692
#> 15   3.8896486221898 8977.02270807468 9024.67139604743    257645.755163577
#> 16   1.1535046084355 20464.5469909585 20848.2358121309    491739.011882154
#> 17  8.59726576726102 6711.42974215464 7056.61236488461    169038.879314176
#> 18  5.40544735528814 9294.81890974668 9339.75279351153    259035.051789754
#> 19  1.05875507714929 19697.3436780813 19936.8984682836    483837.924163151
#> 20  10.1385580098716 6991.75746643608 7289.17238647073    176774.691581099
#> 21  5.46525674336849 9244.13808482618 9271.50703384279    260696.513299604
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     85867.3022595735 87.3540439566703 96.6098874188669   14.0694619480469
#> 2     44195.6027278122 28.4014110103881 37.9604754764791   27.6297713913218
#> 3     67535.9694767649 62.4603563594665  66.542889272738   16.2336327534578
#> 4     152655.572734533 238.803268501982 247.210001505987   14.3878389710132
#> 5     60135.5751721753 60.5165029034484 70.4926398509705   19.2542250822055
#> 6     133444.168880681 198.809701694631 210.737728131245   11.5450650221477
#> 7     371809.714395483 488.166290492683 537.363204365403   6.36086999425997
#> 8     122817.450814149 188.246402484152 216.106857837224   11.6687965141359
#> 9     227190.259487738 281.138977741791 302.708456148752   3.92382558411458
#> 10    496873.079585412 609.811786920854 723.347650367054 -0.234091668740357
#> 11    173125.489132954 304.615872532091 354.490769776895   8.52311098509562
#> 12    274105.070158395 325.303514272029 344.693491674645   3.11786926132047
#> 13    502245.507747183 663.669713146723 783.572699212818 0.0770326382462908
#> 14    182010.281467677  309.57169613703 370.988205396322   8.81506268455819
#> 15    271427.297165283  306.07549377364 332.830909240868    2.4347927664178
#> 16    532694.979578796 733.335392611293 814.737157509415 -0.331509517675949
#> 17    182004.100756857 306.521369440422 356.566820510412   7.39460935323283
#> 18    278905.154704542 337.184679925211 361.852145700523    4.3273818894678
#> 19    515189.131313135 745.874147162706 820.127520003182  0.145721487143994
#> 20    185210.275396818 312.069790842459 387.454880432168   8.69790246921386
#> 21    275499.349054008 325.829518563318 347.701456740872   4.43449541688268
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5815078687935 3324.31657054781 3487.13887987408    76424.3154901132
#> 2   36.3865018779476 1432.06218511332 1514.41505634771    41357.6853355048
#> 3   23.6458768840748 2182.10481040926 2212.31008265739    64344.5114885149
#> 4    15.702516068323 7426.08395182845  7648.4718611716    142249.922184635
#> 5   28.1800823905561  2296.3768031267  2436.7077273067    53891.5249513529
#> 6   13.0434079639169 4796.10245849339 4839.50534967579    128532.151859322
#> 7   7.96976695260461 16859.0415443071 17213.0538933864    348883.514845949
#> 8   14.5397530533668 5080.93700827498  5276.9506189629     111929.90315165
#> 9   5.51093437244633 7726.93394411531 7805.53183180166    214850.361203326
#> 10  1.51877456468315 19954.3240239319 20147.8596434186     462750.75787652
#> 11  11.3905588680875 6521.03298949809 6964.62583538192    157149.984703564
#> 12  4.44822671764344 9170.27249374431 9205.10880719272    258840.875995509
#> 13  1.14917250161876 20181.7192472832 20591.2751484831    464152.623068805
#> 14  11.2097958842802 6863.67867967995   7065.299619423    170109.508543374
#> 15  4.01677194939317 8972.46059965175 9029.23350447035    256326.245822989
#> 16  1.28326312430932 20427.8108272292 20884.9719758602    487817.695826093
#> 17  8.70235225004019 6678.38034210603 7089.66176493323    167797.528324983
#> 18  5.49964725035982 9290.51672938622 9344.05497387199    257132.595127699
#> 19  1.13853471122742 19674.4075811471 19959.8345652178     480836.21284028
#> 20  10.2644405328417 6963.28156983702 7317.64828306979    175967.029300871
#> 21    5.555323267042 9241.51765353735 9274.12746513161    259279.220514608
#>    ci95.integral.bt.up
#> 1     86692.4176083613
#> 2     44443.5760921886
#> 3     67814.8347378741
#> 4      153564.80433598
#> 5     60681.1717934123
#> 6      133873.37425148
#> 7     373812.974550296
#> 8      123768.78993029
#> 9     228268.503027153
#> 10     499854.64167648
#> 11    174521.406995716
#> 12      275438.8346969
#> 13    505574.012233643
#> 14    183050.154829995
#> 15    272746.806505872
#> 16    536616.295634857
#> 17     183245.45174605
#> 18    280807.611366597
#> 19    518190.842636006
#> 20    186017.937677046
#> 21    276916.641839004