Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2415972850222  15.4398981911939
#> 2                 FALSE          0.75 32.7860425947657  32.5570418893933
#> 3                 FALSE          0.75 64.7362914160533  19.9294954804982
#> 4                 FALSE          0.75  243.00153289054  15.0520428964344
#> 5                 FALSE          0.75 64.4806378052197   21.728926064735
#> 6                 FALSE          0.75 205.823989285245  12.3617794689082
#> 7                 FALSE          0.75 507.865356705347  7.05639707569233
#> 8                 FALSE          0.75 203.730414841531  13.3882575285944
#> 9                 FALSE          0.75 289.035372249961  4.64873905684932
#> 10                FALSE          0.75 669.082510231771 0.701945954970446
#> 11                FALSE          0.75 327.467289466909  10.0519981380243
#> 12                FALSE          0.75 335.534057548946   3.7480775369088
#> 13                FALSE          0.75 728.327015789649 0.721882991904905
#> 14                FALSE          0.75 342.667054088877  9.90816768473561
#> 15                FALSE          0.75 319.545142927984  3.40464697286867
#> 16                FALSE          0.75 781.803316110423 0.632236068097203
#> 17                FALSE          0.75 329.090485377965  7.85663660122504
#> 18                FALSE          0.75 347.895241441493  4.95120766263515
#> 19                FALSE          0.75 777.915619099256 0.598354212716673
#> 20                FALSE          0.75 344.316687465861  9.46373842762258
#> 21                FALSE          0.75 337.842521878203  5.25533277882003
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3415.46630879877 2133.24935463525 81997.8072210766 2.03893821095579
#> 2  1471.58382883469 267.831196113295  42846.027142362  2.2410337793127
#> 3  2194.38774080521 696.205658094334 66057.9121099879 1.30255925086218
#> 4  7582.41318471728 6048.66020452602 150466.332540511 3.39581594235862
#> 5  2360.20774466431 1130.64664728339   57474.51299922 2.09737258793376
#> 6  4817.77181943839 3236.99131801191 129558.715801596 1.85387439208948
#> 7  17013.0220620837  15800.891886333 360783.987014881  15.501323217548
#> 8  5179.18126469086 3791.85527768683 118913.944847962  8.2139940153034
#> 9   7761.5158052642 6292.56159839847 221968.805153231 9.02237347744027
#> 10 20069.1258273699 18625.8986728777 481498.280678513 22.9020360109252
#> 11 6785.60673210257 5535.92701590341  167945.50379615 9.38510844394943
#> 12 9187.42181613906 7692.72138191224 267711.569146146 7.24101824979489
#> 13 20427.5674590139 19403.8573136631 489606.641690163 19.2906672379964
#> 14 6983.09142145824 5636.64010701224 177555.004961621 13.4041388822974
#> 15 9005.59413894127 7401.83958463572  264290.18129475 5.12847400044302
#> 16 20697.0032858815 19303.2609503608 513625.464721285 19.8982356593529
#> 17 6905.50841350116 6045.26928466257 175983.752491213 16.6061781698619
#> 18 9317.47535272288 7615.01597625724 271715.698707167 7.12224568672173
#> 19 19770.3618846428 18619.6842975698 501903.342064661 22.6457916396831
#> 20  7130.2070083773 5783.98385612994 180116.198257849  14.711373736354
#> 21 9254.84236539552 7457.59982556695 266981.744183971 5.55320685107193
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.652364858621944  25.120911482382 57.2084204263328 1585.54341222346
#> 2   2.40063931467998 23.2081709504024 67.0567793826483 750.438246788552
#> 3   1.58559623053408 6.75987443667484 109.402541996965 1201.30081652358
#> 4   0.30779842147651 38.2363741746059 62.9638100841341  2070.7467792446
#> 5   2.09698394549329 30.0007430977039  134.47781121182  1121.4921030259
#> 6   0.45173580665497 16.9039277205461  91.553194370152 3069.15443875349
#> 7  0.557387712836435 217.166995246493 257.083547455901 11300.9645071948
#> 8  0.505520681282733 44.3047141950684  102.41682229457 2021.47873835274
#> 9  0.373314103502217 17.6234394092375 73.2731405743411 2139.85873748286
#> 10  0.45725338251916 56.1659219353678 213.670400194072 8792.33685786939
#> 11 0.475127236529518 104.234470425073 191.380012810389 4038.47668007449
#> 12 0.390868305812957 14.6958345852477 74.6745237104753 2885.98834722663
#> 13 0.262413990054626 99.8882066527728 162.600499543281 9616.15261762442
#> 14 0.444976288808938  52.220254314003 166.439926676184 3077.08718645992
#> 15 0.412216309118206 18.9584662580279 102.965164016211 3420.47239554502
#> 16 0.567684168449183 92.7211435556721 388.230255803587 9916.87950956462
#> 17 0.669252047203657 89.6494983212713 164.800049316494 2792.24538102386
#> 18 0.477332453091116 7.25557042401991 111.272151795546 3196.45863052617
#> 19  0.26264484059737 82.5046775924958 138.861714525647 15972.4918114741
#> 20  0.46926227719975 88.3934303951956 109.354132690368 3170.40974591607
#> 21 0.360723361069037 6.56574005697065  135.56675705544 4692.32758240903
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.8875439279999 95.5956506420445    14.3667579987608
#> 2             TRUE 29.0995420277963 36.4725431617351    28.6079902167447
#> 3             TRUE  62.593581448385 66.8790013837216    17.3211896812696
#> 4             TRUE  237.41541566536  248.58765011572    14.5457144931056
#> 5             TRUE 61.0304598980686 67.9308157123707    18.2793874743986
#> 6             TRUE 202.774365910258 208.873612660232    11.6186740669608
#> 7             TRUE 482.365680012481 533.365033398214    6.13949428807639
#> 8             TRUE 190.218394686357 217.242434996705    12.5566760078843
#> 9             TRUE 274.193567879572 303.877176620351    4.03463735658817
#> 10            TRUE 631.408660993799 706.756359469743 -0.0502358592735717
#> 11            TRUE 312.028786076612 342.905792857206    9.27041383393325
#> 12            TRUE 323.622582528033 347.445532569858    3.10509917384648
#> 13            TRUE 696.593868183145 760.060163396153   0.290211978265045
#> 14            TRUE 320.617245627498 364.716862550256    9.17618168964491
#> 15            TRUE 311.108803197256 327.981482658713    2.72655114436922
#> 16            TRUE 749.070718450787 814.535913770058  -0.301604389001704
#> 17            TRUE 301.773322288542 356.407648467388    6.75571698357502
#> 18            TRUE 336.179147286835  359.61133559615    4.16599577730027
#> 19            TRUE 740.663291851978 815.167946346535   0.166303449933999
#> 20            TRUE 320.116477669559 368.516897262164    8.69180198162899
#> 21            TRUE  328.70749660819 346.977547148216    4.66194284986146
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.513038383627 3374.14240941025 3456.79020818729     79389.588307969
#> 2   36.5060935620418 1433.40638762127  1509.7612700481    41611.5562263948
#> 3   22.5378012797268 2183.26774735688 2205.50773425354    64081.7722668066
#> 4   15.5583712997633 7519.51434920005 7645.31202023451    147059.954088654
#> 5   25.1784646550715 2310.85652226858 2409.55896706003    55629.6584897423
#> 6   13.1048848708556 4789.96485833809 4845.57878053869    124509.956749846
#> 7   7.97329986330826 16655.7823549032 17370.2617692642    342193.900400545
#> 8   14.2198390493045 5106.30000983997 5252.06251954175    115588.612323372
#> 9   5.26284075711047   7732.525247436 7790.50636309239    218448.737530072
#> 10  1.45412776921446 19976.7328857863 20161.5187689536    467034.886547318
#> 11  10.8335824421154 6614.14102825332 6957.07243595181    161302.209657428
#> 12  4.39105589997111 9163.24716824633 9211.59646403179    262964.118314959
#> 13  1.15355400554477   20263.25135907 20591.8835589577    473788.070634171
#> 14  10.6401536798263 6897.18910311171 7068.99373980478    172493.196539895
#> 15  4.08274280136812 8974.40746194682 9036.78081593573    258663.504204079
#> 16  1.56607652519611 20544.4770047324 20849.5295670305    497312.197928052
#> 17  8.95755621887506 6758.03498876267 7052.98183823965    171390.508839429
#> 18  5.73641954797004 9305.53993937537 9329.41076607039    266457.524259951
#> 19  1.03040497549935 19634.6416900031 19906.0820792824    475628.593034786
#> 20  10.2356748736162  6984.7998153772 7275.61420137739    174900.874225817
#> 21  5.84872270777859  9244.0417230018 9265.64300778923    259262.865310908
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84606.0261341842 88.2452783915489 96.2379161784956   14.1612630682949
#> 2     44080.4980583292 28.3936163873128 37.1784688022186   27.8517888326205
#> 3     68034.0519531692 62.1832752843634 67.2893075477432   16.8217268686514
#> 4     153872.710992369 236.345733643517 249.657332137563   14.4487579903405
#> 5     59319.3675086976 60.3697875328695 68.5914880775698   17.6188375315682
#> 6     134607.474853345 202.190395476749  209.45758309374   11.4763772878645
#> 7     379374.073629216 477.482763198953 538.247950211741   5.96391715853291
#> 8     122239.277372552 187.630986571537 219.829843111526   12.3974369932802
#> 9      225488.87277639 271.351520234179 306.719224265744   3.91704341398497
#> 10    495961.674809708 624.194519650358 713.970500813184 -0.194270674767107
#> 11    174588.797934873 309.072476916768  345.86210201705   9.12074875442645
#> 12    272459.019977334 321.341661779348 349.726453318544    2.9819756575154
#> 13    505425.212746155 690.517308003176 766.136723576122  0.207551571397838
#> 14    182616.813383348 316.394941879574  368.93916629818   9.03601415867009
#> 15    269916.858385422 309.493333887116 329.596951968853   2.59670300699698
#> 16    529938.731514519 742.802774218091 820.803858002755 -0.480424902063196
#> 17    180576.996142997 296.542376165036 361.638594590894   6.54490258870587
#> 18    276973.873154382 333.935639895518 361.854842987467   4.01563605457657
#> 19    528178.091094536 733.529867485477 822.301370713035 0.0835703251458276
#> 20    185331.522289881 315.482394942608 373.150979989115   8.54398436431107
#> 21    274700.623057033 326.958236450102 348.726807306304   4.54831499112471
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7185333140929  3366.2293222933 3464.70329530424    78890.1421331186
#> 2    37.262294946166  1426.0958137719 1517.07184389747    41375.1681786564
#> 3    23.037264092345 2181.13838690933  2207.6370947011    63703.3625096017
#> 4   15.6553278025284 7507.46989133505 7657.35647809951    146407.668853192
#> 5   25.8390145979019 2301.40628819281 2419.00920113581    55276.3884772892
#> 6   13.2471816499519 4784.64012110612 4850.90351777066    123543.173101639
#> 7   8.14887699285174 16587.3747514006 17438.6693727668    338634.096580779
#> 8   14.3790780639085 5092.34402486852 5266.01850451319    114951.846520791
#> 9   5.38043469971366 7726.97386402209  7796.0577465063    217774.682027765
#> 10    1.598162584708 19959.0406203766 20179.2110343633    464265.300437089
#> 11  10.9832475216222 6581.30717006942 6989.90629413571    160030.089503204
#> 12  4.51417941630219 9158.61798035197 9216.22565192615    262055.031985582
#> 13  1.23621441241197 20231.7865739744 20623.3483440533    470758.982559619
#> 14  10.7803212108011  6880.7397230028 7085.44311991369     171523.91407616
#> 15  4.21259093874035 8968.43554507554 9042.75273280701    257586.055399482
#> 16   1.7448970382576 20515.2698445123 20878.7367272506    494188.380882539
#> 17  9.16837061374421 6729.79539679147 7081.22143021085    170510.951544406
#> 18  5.88677927069374  9303.2544346918 9331.69627075396    265450.639791335
#> 19  1.11313810028752 19608.6527165615 19932.0710527241    470597.258114172
#> 20  10.3834924909341 6956.95588480271 7303.45813195188    173902.195155854
#> 21  5.96235056651534 9241.97351488385 9267.71121590718    257784.782122449
#>    ci95.integral.bt.up
#> 1     85105.4723090346
#> 2     44316.8861060675
#> 3     68412.4617103741
#> 4     154524.996227831
#> 5     59672.6375211507
#> 6     135574.258501552
#> 7     382933.877448982
#> 8     122876.043175133
#> 9     226162.928278697
#> 10    498731.260919937
#> 11    175860.918089096
#> 12    273368.106306711
#> 13    508454.300820707
#> 14    183586.095847083
#> 15    270994.307190018
#> 16    533062.548560032
#> 17     181456.55343802
#> 18    277980.757622998
#> 19     533209.42601515
#> 20    186330.201359845
#> 21    276178.706245492