Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0423370334751  15.5357516547039
#> 2                 FALSE          0.75 33.0357637513421  31.0759051503371
#> 3                 FALSE          0.75 65.1660906395102  20.6031907709961
#> 4                 FALSE          0.75 243.381103031829  15.0543663636685
#> 5                 FALSE          0.75 64.3824528387296  22.9282060861871
#> 6                 FALSE          0.75 205.178887001702  12.4179619364328
#> 7                 FALSE          0.75 507.922017280603   7.1063626887467
#> 8                 FALSE          0.75 203.243389790424  13.4140662356312
#> 9                 FALSE          0.75 290.784668279298  4.72682019395766
#> 10                FALSE          0.75  676.73761413337 0.453762956177184
#> 11                FALSE          0.75 324.361762777792  9.77464869194524
#> 12                FALSE          0.75 333.951751148534    3.903037914967
#> 13                FALSE          0.75 721.516956213705 0.655289773029504
#> 14                FALSE          0.75 342.487822050084  9.73002292367231
#> 15                FALSE          0.75 319.105198328442   3.3415836166641
#> 16                FALSE          0.75 780.735261983634 0.427132174377498
#> 17                FALSE          0.75 332.880136675789  8.04869972146169
#> 18                FALSE          0.75 349.461030290344  5.08059496576879
#> 19                FALSE          0.75 781.880722077429 0.626046643008518
#> 20                FALSE          0.75 340.771329195809  9.25392455578418
#> 21                FALSE          0.75 335.606098516079  5.11266394845026
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3421.0988328764 2125.93237107988 82102.6021909774 3.11415871839765
#> 2  1460.37877411738 277.452299884111 42300.0181549786 2.04641991083578
#> 3  2196.12069376296  657.87467123475 66182.3024227003 1.72621411955209
#> 4   7565.4266132781  6034.8049341911 149163.183314458 2.44658976504011
#> 5  2361.10715468896 1089.00002839719 57121.5440334694 2.44787430141435
#> 6  4817.68912204065 3222.49837410124 130719.924256846 3.44358277346295
#> 7  17058.3575022624 15828.2445410861 362659.454854932 12.2256484803219
#> 8  5185.13222762529 3777.81079512759 118771.358506204 10.0499522285164
#> 9  7764.23536555087 6272.00221346343  221737.53042429 8.23537067533626
#> 10 20062.9635577125 18835.0600035138 484016.872413234 23.6806588986358
#> 11 6758.02581432492 5577.02601003071  167063.09425353 9.10234241328711
#> 12 9190.15556440824 7627.03127444736 266916.970627923 5.29779765142278
#> 13 20444.5101744598 19437.5600956628 489311.024867972 16.7179688959814
#> 14 6971.41461184378 5687.37077735664 176990.308396609 10.3133784234289
#> 15 8993.62931292918 7393.45516905188 264875.288994894 6.60535174621019
#> 16 20650.0782967264 19408.9963567883 510473.284722125 28.5648906631273
#> 17 6880.54891509674 5984.32539183836 175375.514908067 10.8959022623407
#> 18 9319.91881635854 7618.33216321625 271455.434259126 5.65109681017135
#> 19 19800.2928902212 18645.3881177486 507312.977324066 23.9867964535892
#> 20 7125.74719266937 5843.66715631784 180098.809299452 11.1078932356621
#> 21 9256.10377573187 7478.92333155708 269316.539677319 7.29485257788881
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.607089625224331 10.5794264121723 56.5867140386514   953.1937381826
#> 2     2.320944514177  19.279986618588 65.0589109073378 914.573594927335
#> 3   1.46480974657935 4.04517553345442  106.68475655628 1097.33857262289
#> 4  0.309425621015856  47.551973302939 84.3556205152996 2478.97541493506
#> 5    1.9906119332794 28.4301706019396 111.198211797425 1649.00541875933
#> 6  0.521713029465799 11.4156037000977 110.022498274881 1754.12293664793
#> 7  0.556214989463721 145.745138322291 186.296162996277 7813.20047532303
#> 8  0.822671002675104 37.8904041330416 120.983197860931 2220.07269823404
#> 9  0.391469949613867 25.2622830847188 76.9258146560432 2548.06800268378
#> 10 0.150272527885559 108.575440073536 239.685314586105 11730.4846765761
#> 11 0.558767917159631 82.6097728225612 177.327534006961  3222.6008552136
#> 12 0.454187405273907 8.95951929244505 120.785857532976 4256.13372695425
#> 13 0.288381041781055 104.049895276968 186.445492075761 8900.56774431523
#> 14 0.561920742442067 58.0680823306431 166.822403389189 3484.24553350351
#> 15 0.502051348688944 19.9957646701047  106.13068158377 2824.10724661394
#> 16 0.395275793608417 92.7958833063249 293.460306217457 12647.5514577477
#> 17 0.485404058305138 99.9392734689483 135.575553506994 3407.12232369846
#> 18 0.494014665484238 9.48876166817425 141.681811217424  2933.8377996951
#> 19 0.335688395619463 83.8718887236558  210.90607387585  4334.8189758897
#> 20 0.369822594406161  96.901302438543 111.287943663271 3420.02865539345
#> 21 0.410723948494356 11.4999911168184 129.746748305443  2355.6513192137
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE  86.919545941711 97.1651281252393   14.5370892212099
#> 2             TRUE 29.6694029980173  36.402124504667    27.257951424516
#> 3             TRUE 62.3264684128471 68.0057128661734    18.193578737873
#> 4             TRUE 239.356462868338  247.40574319532   14.5453612170974
#> 5             TRUE  60.355699612903 68.4092060645562   19.6536494559425
#> 6             TRUE 199.514193339356 210.843580664049   11.5597440029616
#> 7             TRUE 487.810825530474 528.033209030733   6.19138903107888
#> 8             TRUE 186.711218374514 219.775561206333   12.0607724362307
#> 9             TRUE  277.23748351837 304.331853040226   4.08285212684284
#> 10            TRUE 637.782930245115 715.692298021626   0.20656464780544
#> 11            TRUE 309.388409507934 339.335116047649   8.85547546821765
#> 12            TRUE 325.236874011943 342.666628285124   3.15589963329142
#> 13            TRUE 694.015897379815 749.018015047594  0.180902959299668
#> 14            TRUE 325.522314543544 359.453329556625   8.80566330235511
#> 15            TRUE 308.239394705926 329.971001950958   2.51570914807079
#> 16            TRUE  733.74601684279 827.724507124479 -0.223096506108347
#> 17            TRUE 314.956377454238 350.803895897339   7.25021004554974
#> 18            TRUE 340.164976037612 358.757084543075   4.26794084104722
#> 19            TRUE 742.422441911274 821.339002243583 0.0738392322145017
#> 20            TRUE 322.498844823145 359.043813568473   8.64556638798604
#> 21            TRUE 323.606066025452 347.606131006706   4.43702305317704
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.534414088198 3403.69567642838 3438.50198932442     80534.598491667
#> 2   34.8938588761583  1428.6631961298 1492.09435210495    40795.5445913231
#> 3   23.0128028041191 2189.46638001042 2202.77500751549    64377.1804707356
#> 4   15.5633715102396 7487.20361719477 7643.64960936144     145085.26875689
#> 5   26.2027627164317 2314.33952404877 2407.87478532915    54408.9301196103
#> 6   13.2761798699041 4798.91045395398 4836.46779012731    127834.392026061
#> 7   8.02133634641452 16818.6067497222 17298.1082548025    349806.740073026
#> 8   14.7673600350317 5122.80251282644 5247.46194242415    115119.338917609
#> 9   5.37078826107247 7722.67890987651 7805.79182122524    217545.958559875
#> 10 0.700961264548929 19884.3569587916 20241.5701566335    464720.225120267
#> 11  10.6938219156728  6622.1327380318 6893.91889061803    161761.915846703
#> 12  4.65017619664258 9175.41715517216 9204.89397364431    259915.630647083
#> 13  1.12967658675934 20273.3480967292 20615.6722521904    474669.590928574
#> 14  10.6543825449895 6875.89261640987 7066.93660727768    171258.724493996
#> 15  4.16745808525741 8960.73628004686  9026.5223458115    260229.632574214
#> 16  1.07736085486334 20497.4290686875 20802.7275247653     489668.06257413
#> 17  8.84718939737365 6716.14881024032 7044.94901995316    169770.798685583
#> 18  5.89324909049036 9304.30980341439 9335.52782930269    266629.271078628
#> 19  1.17825405380253 19662.3236332708 19938.2621471716    500182.200108727
#> 20  9.86228272358231 6966.34455015797 7285.14983518077     174472.86216133
#> 21  5.78830484372347 9237.18629034471 9275.02126111904    265441.493257212
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83670.6058902878 85.9385859454158 98.1460881215345    14.3458559892643
#> 2      43804.491718634  29.024780726104 37.0467467765803    26.5268539025502
#> 3     67987.4243746649 61.7827109651881 68.5494703138323    17.7321636677005
#> 4     153241.097872027  238.58578709235 248.176418971308    14.4478921464775
#> 5     59834.1579473285 59.5846192079574 69.1802864695017    19.0266066969594
#> 6     133605.456487632 198.429464765715 211.928309237689    11.3954043986799
#> 7     375512.169636839 483.959746259172 531.884288302034    6.01618130939781
#> 8     122423.378094799 183.545483422532 222.941296158316     11.801631070388
#> 9     225929.102288705 274.643341755639 306.925994802957    3.95953909271448
#> 10    503313.519706202 630.323522692044 723.151705574697   0.159228801521489
#> 11    172364.272660356 306.521171647749 342.202353907834    8.67946357431236
#> 12    273918.310608763 323.568067751745 344.335434545322    3.01283060063014
#> 13    503952.458807371 688.749737177581 754.284175249829  0.0900629311386361
#> 14    182721.892299223 322.273600340164 362.702043760005    8.62865826848586
#> 15    269520.945415574  306.15870890587 332.051687751014    2.35756297323377
#> 16     531278.50687012 724.748076283905 836.722447683364  -0.347608381094999
#> 17    180980.231130551 311.524168241601 354.236105109977    7.09730776718362
#> 18    276281.597439625 338.384880542408 360.537180038279    4.11232622141968
#> 19    514443.754539404 734.866601028394 828.894843126463 -0.0319026124056291
#> 20    185724.756437574 318.999858453911 362.542799937707     8.5290722707481
#> 21    273191.586097425 321.308187463417 349.904009568741    4.30764500940132
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7256473201436 3400.36315710854 3441.83450864426    80234.3424641395
#> 2   35.6249563981241 1422.59000034494 1498.16754788981     40507.453908921
#> 3   23.4742178742916 2188.19214971739 2204.04923780853    64031.5188203594
#> 4   15.6608405808596 7472.22474560434 7658.62848095186    144304.391501186
#> 5   26.8298054754147 2305.38402030916 2416.83028906876    53889.4934127011
#> 6   13.4405194741858 4795.31453878845 4840.06370529284    127281.843301016
#> 7   8.19654406809559 16772.6970311507 17344.0179733741    347345.581923299
#> 8   15.0265014008744 5110.86703552453 5259.39741972605    114420.016017666
#> 9   5.49410129520084 7714.72129070483 7813.74944039692     216743.31713903
#> 10  0.74829711083288 19850.1556951684 20275.7714202567    461025.122447145
#> 11  10.8698338095781  6596.1106595927 6919.94096905714    160746.796577311
#> 12  4.79324522930386 9172.59490659504 9207.71622222143    258574.948523093
#> 13  1.22051661492037  20240.572379717 20648.4479692027    471865.912089115
#> 14  10.8313875788588 6857.60117047571 7085.22805321184    170161.187150942
#> 15  4.32560426009443 8954.43761417577 9032.82101168259     259340.03879153
#> 16  1.20187272984999  20468.198365446 20831.9582280068     485684.08386494
#> 17  9.00009167573976  6684.6679390976 7076.42989109588    168697.555153618
#> 18   6.0488637101179 9301.32084348892 9338.51678922816    265705.112171724
#> 19  1.28399589842267 19635.9039883228 19964.6817921196    498816.732131322
#> 20  9.97877684082025 6935.82063988982 7315.67374544891    173395.553134881
#> 21   5.9176828874992 9233.56379314291 9278.64375832084     264699.46309166
#>    ci95.integral.bt.up
#> 1     83970.8619178153
#> 2     44092.5824010362
#> 3     68333.0860250411
#> 4     154021.975127731
#> 5     60353.5946542377
#> 6     134158.005212676
#> 7     377973.327786566
#> 8     123122.700994743
#> 9      226731.74370955
#> 10    507008.622379323
#> 11    173379.391929748
#> 12    275258.992732753
#> 13     506756.13764683
#> 14    183819.429642276
#> 15    270410.539198257
#> 16    535262.485579311
#> 17    182053.474662516
#> 18    277205.756346528
#> 19     515809.22251681
#> 20    186802.065464023
#> 21    273933.616262978