Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.6840224743289  15.3726749533089
#> 2                 FALSE          0.75 32.6286529799026  31.4666910967888
#> 3                 FALSE          0.75 65.0659552032031  20.5103705712155
#> 4                 FALSE          0.75 240.836413881215   14.922565353794
#> 5                 FALSE          0.75 65.2966534359275  22.7173569689297
#> 6                 FALSE          0.75 205.196884199885  12.2892420941833
#> 7                 FALSE          0.75 505.966244071449  7.00180356473784
#> 8                 FALSE          0.75  204.52799724903  13.5511803721628
#> 9                 FALSE          0.75 289.691792429433  4.69074750641626
#> 10                FALSE          0.75 670.780221949921 0.585253993620744
#> 11                FALSE          0.75 326.636469047342  10.2372755759329
#> 12                FALSE          0.75 334.608596217514  3.80008522629256
#> 13                FALSE          0.75 732.133288269961 0.876208349452725
#> 14                FALSE          0.75 339.332671215638  9.57174991633273
#> 15                FALSE          0.75  320.11533638677  3.35006151159489
#> 16                FALSE          0.75 785.087691985519  0.46915461847699
#> 17                FALSE          0.75  332.07518643888  8.09686392741249
#> 18                FALSE          0.75 349.384012030782  4.94439521343959
#> 19                FALSE          0.75  780.74029628874 0.662893205847199
#> 20                FALSE          0.75 350.869888483343  9.68083016216954
#> 21                FALSE          0.75 337.053545531426   5.1324829857062
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3413.18681890145 2132.42794280851 81694.4483971293 2.64428362052419
#> 2  1464.20196230526 281.989506880736  42366.718894436 1.32982870776684
#> 3  2196.12641507942 672.689275426723 65381.2708617646 1.69306188339772
#> 4  7563.58788434224 6043.14200761325 149385.088376506 2.74324727544536
#> 5  2367.88737764371 1139.89746915608 57130.4773726885 2.18165595899957
#> 6  4813.91727803432 3241.30339337673 130028.217203506 2.51635841223842
#> 7  17040.0198136103 15836.0472928138 362205.259959776  12.313809210345
#> 8  5193.09823674545 3774.66169578362  119082.71191876 7.03585336826622
#> 9  7768.63895963462 6275.61304290311 221720.448468117 8.95579943657502
#> 10 20061.3983587664 18705.4342578213 485449.831785711 22.9583607102031
#> 11 6769.69716995617  5464.2754444957  167406.83291703 15.9697122830548
#> 12 9189.91661670721 7649.15792022915 265330.338919087 5.49554723443897
#> 13 20475.4653921419 19357.7206799403  492483.64257985 22.7813099872722
#> 14 6973.26619407267 5714.40864934865 176603.590141186 15.9549269456165
#> 15 9009.17815428604 7414.20712462138 265522.890943206 6.61675379386906
#> 16 20664.6157831873 19377.8197987425 515467.637793314 26.0905208238702
#> 17 6906.01086413563 5996.03527565131 176353.496878604 13.6324786758761
#> 18  9320.5974150433 7648.78479448668  270616.86715757 6.92153959795177
#> 19 19832.3331407425  18646.313021223 504133.141682227 22.1831404606173
#> 20  7161.4976755405 5764.81525423492 181348.040359699 13.7493675226437
#> 21 9255.20834865372 7483.24420971752 269564.774101294 9.16910314934283
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.544327161907911 16.5559288260958 42.3270110037312 1171.36383998605
#> 2   2.35450864069056 20.2608552528636 81.0596276287641 1073.10092997659
#> 3   2.07316177271956 4.85290828294211 142.610896184854 1325.66557648458
#> 4  0.343086750774061 63.5044062191043  77.016049529447 3289.91546908798
#> 5   2.58611815651807 30.5678980923223 148.083380588175 914.502940165962
#> 6  0.622150759810289 9.96453825711851 126.882947439875 2262.14422737298
#> 7  0.378560573833128 108.781676947214 157.944556179758 5664.17420828818
#> 8  0.810363096447365 42.0641361199802 161.056888984624  1949.6567961672
#> 9  0.408628034762542 17.8928013104991 45.7098778564317 4518.74832291381
#> 10  0.28145426690023 83.8568114344634 217.569232577812 10980.4071003236
#> 11 0.536728252848675 94.3904863963236 152.393110096499 3885.32053161945
#> 12 0.464386253028556  11.072907487976 122.162867406294 4438.77053229731
#> 13 0.485825893779799 76.8822918925626 326.346707425541  7665.3182525306
#> 14  0.67093737552542 57.2480053785007 186.996244434539 3397.71712842442
#> 15 0.357586302286018 13.6434917107591 86.0453036181727 2116.75558763431
#> 16 0.473836035126636  79.801321423524 290.484099005414 8571.66265954265
#> 17 0.721891631283692 63.5990318075058 181.334464649779 2013.52595460131
#> 18 0.505236602105646 9.35236471219852 142.038197162773 5848.51952458422
#> 19 0.335093380080318 92.2378106894948 328.189076091284 7998.67109204342
#> 20 0.474497814509187 68.2842864817558 142.351937697466 3208.70010424611
#> 21 0.463250910813659 12.6504503023092 110.507367971719 3047.67020899226
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.3341759185667 96.0338690300912   14.4772567719704
#> 2             TRUE 30.4410847556262 34.8162212041791   27.5935243828528
#> 3             TRUE 62.2808684050139 67.8510420013924   17.1000194550918
#> 4             TRUE 236.323772113107 245.349055649322   14.3581876487707
#> 5             TRUE 61.7078293833732 68.8854774884818   18.4631926014575
#> 6             TRUE 201.057474611753 209.336293788017   11.2658040942954
#> 7             TRUE 485.710027920432 526.222460222467   6.37907142078234
#> 8             TRUE 192.954018458232 216.101976039828   12.2181330785068
#> 9             TRUE 274.959502356267 304.424082502599   4.01855438923188
#> 10            TRUE 633.013718581636 708.546725318205  0.122261724569866
#> 11            TRUE 300.366292341717 352.906645752967   9.35435759999681
#> 12            TRUE 325.568421016862 343.648771418166   3.03616984006059
#> 13            TRUE 694.658033340898 769.608543199024 0.0770247541849549
#> 14            TRUE 313.086816390099 365.578526041178   8.46805793359342
#> 15            TRUE 309.230776395856 330.999896377685   2.76183204433439
#> 16            TRUE 742.168785230252 828.006598740785 -0.310305659306327
#> 17            TRUE 309.649759017064 354.500613860696   6.90935219395082
#> 18            TRUE 337.998079392151 360.769944669413    4.1132810029758
#> 19            TRUE 744.249030231024 817.231562346455  0.111664595615076
#> 20            TRUE 328.252178908594 373.487598058092   8.90028125730193
#> 21            TRUE 321.970370850758 352.136720212095   4.37043523741773
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2680931346474 3385.95231598253 3440.42132182038    79767.5548803522
#> 2   35.3398578107247  1430.8728554143 1497.53106919622    40601.4678646245
#> 3   23.9207216873392 2188.14338095398 2204.10944920486    63200.5509884475
#> 4   15.4869430588173 7459.12313611181 7668.05263257267    143973.177429856
#> 5   26.9715213364019 2317.60318528184 2418.17157000558    55626.1200361155
#> 6   13.3126800940713 4797.52561260136 4830.30894346728    126306.989949478
#> 7   7.62453570869333 16861.0739550321 17218.9656721884    352887.693387142
#> 8   14.8842276658187 5123.90273282808 5262.29374066281    115875.526489065
#> 9   5.36294062360064 7739.20530147885 7798.07261779039    214287.107476924
#> 10  1.04824626267162 19923.4539039567 20199.3428135761    467387.062105679
#> 11   11.120193551869 6614.42481983422 6924.96952007812    161015.480642516
#> 12  4.56400061252454 9171.70168388949 9208.13154952493    258028.561393458
#> 13  1.67539194472049 20348.9940219786 20601.9367623051    479874.194054437
#> 14  10.6754418990721 6879.09322522504 7067.43916292031    171014.345464928
#> 15  3.93829097885539 8986.73461042184 9031.62169815024    262040.828001548
#> 16  1.24861489626031 20533.3426094456  20795.888956929    501367.252718366
#> 17  9.28437566087417 6801.39045681228 7010.63127145898    173041.246683285
#> 18  5.77550942390338 9305.21277509173 9335.98205499487    260996.052539629
#> 19  1.21412181607932 19680.6019421583 19984.0643393268    490975.327735815
#> 20  10.4613790670372 7049.17002427801 7273.82532680299    176069.728688214
#> 21  5.89453073399466 9234.39835790642 9276.01833940101    264551.356607502
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83621.3419139063 86.5012265781015 96.8668183705563    14.3057937159694
#> 2     44131.9699242475 30.0221887126796 35.2351172471257    26.8518541610353
#> 3     67561.9907350818 61.7475539117436 68.3843564946627    16.4469734966852
#> 4     154796.999323155 235.459649221342 246.213178541088    14.2501153222768
#> 5     58634.8347092615 61.0206077562883 69.5726991155666    17.6485653821543
#> 6     133749.444457535 200.264821711898 210.128946687873    11.0698266049552
#> 7      371522.82653241 481.831178019173 530.101310123725    6.25982484002491
#> 8     122289.897348455 190.737724647228 218.318269850832    11.9628687031259
#> 9     229153.789459311 272.138425533746  307.24515932512    3.88983655828168
#> 10    503512.601465744 625.781834957922 715.778608941919  0.0336036304962932
#> 11    173798.185191544 295.335832972555  357.93710512213    9.18528820034948
#> 12    272632.116444716 323.837323638013 345.379868797014    2.88988817035659
#> 13    505093.091105263 687.481920694908 776.784655845014 -0.0760104023556818
#> 14    182192.834817444  308.06101440223 370.604328029047    8.25671266030291
#> 15    269004.953884865 307.146498950787 333.084173822754    2.64919235911429
#> 16    529568.022868262 733.950271170733 836.225112800304  -0.459564010371217
#> 17    179665.747073923 305.355528234163 358.794844643597    6.68195633009646
#> 18    280237.681775511 335.817794418797 362.950229642768    3.95413147331252
#> 19    517290.955628638  737.26134098593  824.21925159155 0.00611018088977544
#> 20    186626.352031184 323.921128138961 377.818648827724    8.75081444573154
#> 21    274578.191595087 319.082103358715 355.024987704138    4.22451120051142
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4395561906484 3380.73719840231  3445.6364394006    79398.5752707566
#> 2   36.0815280325423 1424.49068600965 1503.91323860087    40263.4410716818
#> 3   24.5737676457458 2186.61471484486 2205.63811531399    62782.9663318549
#> 4   15.5950153853111  7439.1192481528 7688.05652053168    142936.854057093
#> 5   27.7861485557051 2307.97429738276 2427.80045790466    55338.0516099632
#> 6   13.5086575834115 4794.38678305036 4833.44777301827    125594.414517855
#> 7   7.74378228945077 16826.8077267937 17253.2319004268    351103.478511531
#> 8   15.1394920411996 5110.65252995028 5275.54394354061    115261.384598272
#> 9   5.49165845455085 7733.56906906604  7803.7088502032    212863.701755206
#> 10   1.1369043567452 19897.0390083549  20225.757709178    463928.233869077
#> 11  11.2892629515163 6584.69181661938 6954.70252329297    159791.604675056
#> 12  4.71028228222853 9168.21371803077 9211.61951538364    256630.348675784
#> 13  1.82842710126113 20324.7761000324 20626.1546842513     477459.61880489
#> 14  10.8867871723626 6861.06010353081 7085.47228461453    169944.064569474
#> 15  4.05093066407548 8982.43691053296 9035.91939803913    261374.049991443
#> 16   1.3978732473252 20508.2051931972 20821.0263731774     498667.17898061
#> 17  9.51177152472853 6781.35676179292 7030.66496647834    172406.986007585
#> 18  5.93465895356666 9302.26678020739 9338.92804987921    259153.768889385
#> 19  1.31967623080462 19651.5470317911 20013.1192496939    488455.746341822
#> 20  10.6108458786075 7027.66047403626 7295.33487704474    175058.988155377
#> 21  6.04045477090097 9230.41346606119 9280.00323124624     263591.34049167
#>    ci95.integral.bt.up
#> 1     83990.3215235019
#> 2     44469.9967171901
#> 3     67979.5753916744
#> 4     155833.322695918
#> 5     58922.9031354138
#> 6     134462.019889157
#> 7     373307.041408021
#> 8     122904.039239248
#> 9     230577.195181028
#> 10    506971.429702346
#> 11    175022.061159004
#> 12    274030.329162389
#> 13     507507.66635481
#> 14    183263.115712898
#> 15     269671.73189497
#> 16    532268.096606018
#> 17    180300.007749622
#> 18    282079.965425755
#> 19    519810.537022632
#> 20    187637.092564021
#> 21    275538.207710919