Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4431095695384  15.4609643094412
#> 2                 FALSE          0.75 32.4926544312958  30.8418286918682
#> 3                 FALSE          0.75 64.7207941365595  20.2764534173275
#> 4                 FALSE          0.75 243.516130176647  15.1076769413421
#> 5                 FALSE          0.75 65.3880167171431  22.7105593228782
#> 6                 FALSE          0.75 205.531062798562  12.3981621265956
#> 7                 FALSE          0.75 508.315195179094  7.09136143585361
#> 8                 FALSE          0.75 203.853551142838  13.2688082136496
#> 9                 FALSE          0.75 289.431768444016  4.63718220991695
#> 10                FALSE          0.75 664.931074372359 0.521262461689131
#> 11                FALSE          0.75 324.682929306028  9.89391322476481
#> 12                FALSE          0.75  333.65755191744  3.78228593808345
#> 13                FALSE          0.75 728.166281434169 0.753797845386839
#> 14                FALSE          0.75 343.040857693835  9.81436050856029
#> 15                FALSE          0.75 319.402484526004  3.28884462326281
#> 16                FALSE          0.75 774.297008098799 0.428291519704245
#> 17                FALSE          0.75  337.41374509832  8.23338490953524
#> 18                FALSE          0.75 351.334898498821  5.06182365822352
#> 19                FALSE          0.75 780.153806126357 0.695923171209813
#> 20                FALSE          0.75 349.348613266743  9.50541169881165
#> 21                FALSE          0.75 338.167896082958  5.06084405534053
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.97299184224 2130.58591235607 81712.1949386765 2.52524865485279
#> 2    1462.255925713 305.340846278075 42366.1294104543 1.55097607444273
#> 3  2195.63858239442 679.116216256083 66091.7730491893  2.1073359153024
#> 4  7556.79410997903 6014.36611625353  148824.52684526 2.56948863929986
#> 5  2357.65980846635 1102.82893969011  56828.872504865  2.4178118794423
#> 6  4821.30148444083 3230.35418389387 130866.098940481 2.97596500378122
#> 7  17094.0647691671 15888.8843100946 365105.848323187 13.3106626389598
#> 8  5173.12801414103 3808.11323100384  117992.05002104 7.32318624920559
#> 9  7771.38432613929 6301.01922360004 222080.130718176 4.59342673516535
#> 10 20080.2398763762 18723.5066621082 485512.763466698 24.4160875591752
#> 11 6772.62326731747 5545.59785340325 166865.921776143 18.3736619008336
#> 12 9185.05210126537 7657.35417415755 265998.105058085 6.29540806395011
#> 13  20451.337138056 19410.9540879168 491685.346792198 22.9963910393881
#> 14   6966.632154423 5651.70179740095 176667.739510345 15.3015744160661
#> 15 9007.83935655491 7433.95537606763 265262.326931616 5.13814333820786
#> 16 20666.3814153103 19375.5142816007 516315.864386999  28.583859699994
#> 17  6907.8332755921 5981.47002265123 176621.122361442 13.8113604750913
#> 18 9324.07258997371 7640.35198804212  270416.24920406 4.92762426148683
#> 19 19804.1991096501 18637.6079631934 504934.939403838 24.9213768094239
#> 20 7109.18177065029 5775.15355573494  179537.58525653 15.6521476147263
#> 21 9257.14351147563  7515.7572737819 269586.042857018 8.82892566696112
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.532513694160453 19.1710839886557 48.8141198899851 1383.93457633127
#> 2   2.12421782387602 22.9667421094103 72.2446891865379 999.572224986072
#> 3   1.30867112484577 6.74867013936419 81.8238163878814 1000.47068794203
#> 4   0.29307908456888 38.6798791093194 81.1656652073945 2314.65960935034
#> 5   2.66564747789778 34.9865065004074 169.003878284044   1074.355869182
#> 6  0.427628616333336 7.89624771462593 89.9635391039662 1536.61352377716
#> 7  0.388841299965698 96.6466070383738 111.753440140555 4655.77277658606
#> 8  0.628796391469911  51.503789783553 122.893877467798 2532.55250051785
#> 9  0.295134920769674  19.253327368942 64.0819664230198 2729.33868347981
#> 10 0.182973241139959  61.996817948617 136.545577039429 7280.25593999767
#> 11 0.818469252417579 91.1355268642671 224.675479659713 3357.73705077602
#> 12 0.422990897055525 19.9096320170302 79.7744239497979 4910.17763789492
#> 13 0.252867519706121 82.3668690756674 174.195155851288 8029.67807712676
#> 14  0.57468172762104 60.0254582029165 180.287987365756 3044.57403256741
#> 15 0.309246241212477 17.3494711640565  70.190259570032 2951.85786377978
#> 16  0.34750041428033 87.9726547549951 238.059391647869 7747.12364941151
#> 17 0.577790042684663 85.3579883502979 169.406580972863  3220.0153333156
#> 18 0.368483798894212 8.76537502011671 123.237983233634 5465.87653835725
#> 19  0.25339954929091 86.6134434084545 180.771158330911  7538.9185101163
#> 20 0.414145462892115 68.3154513466896 130.720899317329 2595.50482084862
#> 21 0.501731910999469 7.07690190893112  124.27409561444 2388.82496475315
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.2890755323056 96.5971436067713   14.5849792825472
#> 2             TRUE 29.9412987888375 35.0440100737541   27.3474903715921
#> 3             TRUE 61.2542265558871  68.187361717232   18.1236894169562
#> 4             TRUE 239.289321364999 247.742938988296   14.6255618472263
#> 5             TRUE 61.4107161754605 69.3653172588257   18.3255692217364
#> 6             TRUE 200.635600367342 210.426525229782   11.6947130527273
#> 7             TRUE 486.419155138006 530.211235220183   6.45171749741004
#> 8             TRUE 191.806909762895 215.900192522781   12.2344381496816
#> 9             TRUE 281.875581464669 296.987955423363   4.15168526525083
#> 10            TRUE 624.766610337516 705.095538407202  0.220271480013899
#> 11            TRUE 294.458255479157   354.9076031329   8.54753130453789
#> 12            TRUE 323.301605652242 344.013498182638   3.08646591242711
#> 13            TRUE 690.337218174376 765.995344693963   0.33783077547027
#> 14            TRUE 317.869767779406 368.211947608263   8.86900906662368
#> 15            TRUE 310.950238734652 327.854730317356   2.78013455646828
#> 16            TRUE 727.276558892309 821.317457305289 -0.143346661786899
#> 17            TRUE 314.694057116795 360.133433079845   7.28292028931897
#> 18            TRUE 343.228956588675 359.440840408967   4.45566780904254
#> 19            TRUE 739.158141274855  821.14947097786  0.279080912626266
#> 20            TRUE 323.600830440518 375.096396092968   8.82414241235412
#> 21            TRUE 323.644313360807 352.691478805109   4.23549506174641
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3369493363351  3380.4365586809 3443.50942500358    79435.6225606116
#> 2   34.3361670121443 1424.47563494302 1500.03621648298    40721.8331003522
#> 3   22.4292174176988 2184.53702001517 2206.74014477368    64445.9987675247
#> 4   15.5897920354579  7493.1657088442 7620.42251111386    145016.911787878
#> 5   27.0955494240201 2300.10700527318 2415.21261165952    55061.5571000606
#> 6    13.101611200464 4808.31215695027 4834.29081193139    128338.369693867
#> 7   7.73100537429718  16935.081100589 17253.0484377452    357447.102105703
#> 8   14.3031782776176 5088.40427994709 5257.85174833498    113826.001157688
#> 9   5.12267915458306 7739.71260261738  7803.0560496612    217590.368583852
#> 10 0.822253443364363 19978.2551108508 20182.2246419017    473536.742445402
#> 11  11.2402951449917 6622.70532562575 6922.54120900918    161342.444327617
#> 12  4.47810596373979 9152.30075659735 9217.80344593338    257920.862843748
#> 13  1.16976491530341 20315.8436384265 20586.8306376855    478476.526355324
#> 14  10.7597119504969 6867.89027567921  7065.3740331668    171659.415226771
#> 15  3.79755469005733 8979.29947649004 9036.37923661978    260406.520745698
#> 16 0.999929701195388 20521.6663982383 20811.0964323823    503571.845983718
#> 17  9.18384952975151 6767.41938475586 7048.24716642834    171324.197138138
#> 18  5.66797950740449 9309.65354806562  9338.4916318818    261424.882298462
#> 19  1.11276542979336 19661.7199952432  19946.678224057    492533.418454697
#> 20  10.1866809852692 6996.80285318498 7221.56068811559    175267.979826234
#> 21  5.88619304893466 9245.50200783544 9268.78501511582    265656.425789999
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83988.7673167415  87.493622206027 97.3925969330499   14.4172374688867
#> 2     44010.4257205563  29.452741325388 35.5325675372035   26.6783617570712
#> 3      67737.547330854 60.5904157425668 68.8511725305522   17.7114580126298
#> 4     152632.141902641  238.47993244362 248.552327909675   14.5332419355871
#> 5     58596.1879096694 60.6491054334362   70.12692800085   17.4858902661986
#> 6     133393.828187094 199.698171391151 211.363954205974   11.5600100385823
#> 7     372764.594540671 482.226296406733 534.404093951456   6.32923248792084
#> 8     122158.098884391 189.500106094395 218.206996191281   12.0363672863686
#> 9     226569.892852501 280.428652043092  298.43488484494   4.05871776520839
#> 10    497488.784487994 617.075542756376 712.786605988343  0.162634909054812
#> 11     172389.39922467 288.670551980395 360.695306631662   8.28971349002635
#> 12    274075.347272422 321.318552112098 345.996551722782   2.95322377985462
#> 13    504894.167229072 683.093354996969  773.23920787137  0.258177506762842
#> 14    181676.063793918 313.049771838345 373.031943549324   8.68798432242305
#> 15    270118.133117534 309.331723583116 329.473245468891   2.68272199048635
#> 16    529059.882790281  718.27264308681 830.321373110787 -0.252809292285203
#> 17    181918.047584747 310.343478567141 364.484011629499    7.1009164258733
#> 18    279407.616109657 341.676754946307 360.993042051335   4.33959541239086
#> 19     517336.46035298 731.307907579887 828.999704672828  0.199260054599629
#> 20    183807.190686826 318.670403941879 380.026822591607   8.69368659154311
#> 21    273515.659924037 320.863201775714 355.472590390202   4.07744950978157
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5046911499957 3374.39766722448 3449.54831646001    78999.6831690672
#> 2   35.0052956266652 1417.24111117856 1507.27074024745    40406.9678494816
#> 3   22.8414488220252 2182.41118892127 2208.86597586758    64130.8505008229
#> 4   15.6821119470971 7480.98154692476 7632.60667303329    144287.794010933
#> 5   27.9352283795579 2289.08625572555 2426.23336120715    54723.1350012682
#> 6    13.236314214609 4805.82483892016  4836.7781299615    127854.336433878
#> 7   7.85349038378638 16904.6374193719 17283.4921189623    355980.533681078
#> 8   14.5012491409306 5072.18058616527  5274.0754421168    113028.247120025
#> 9   5.21564665462551 7733.64780449616 7809.12084778241    216730.626898556
#> 10  0.87989001432345  19958.726113197 20201.7536395555    471243.461824303
#> 11  11.4981129595033  6593.9976346635 6951.24889997143    160284.757156622
#> 12  4.61134809631228 9146.02922251199 9224.07498001875    256374.156887811
#> 13  1.24941818401084 20289.8980746677 20612.7762014443     475947.17776103
#> 14  10.9407366946975 6848.98225634529 7084.28205250072    170700.374406512
#> 15  3.89496725603926 8973.83439307336 9041.84432003646    259476.685518607
#> 16  1.10939233169369 20493.9550119905 20838.8078186301    501131.502034153
#> 17  9.36585339319718 6740.53161842552 7075.13493275869    170309.892308144
#> 18  5.78405190405617 9306.89245493428 9341.25272501314    259703.131188879
#> 19     1.19258628782 19634.4367605696 19973.9614587307     490158.65912401
#> 20  10.3171368060802 6975.28348601078  7243.0800552898    174450.395807667
#> 21  6.04423860089949 9243.27278373413 9271.01423921714    264903.945926102
#>    ci95.integral.bt.up
#> 1     84424.7067082858
#> 2      44325.290971427
#> 3     68052.6955975557
#> 4     153361.259679586
#> 5     58934.6100084617
#> 6     133877.861447084
#> 7     374231.162965295
#> 8     122955.852922055
#> 9     227429.634537797
#> 10    499782.065109094
#> 11    173447.086395664
#> 12    275622.053228359
#> 13    507423.515823366
#> 14    182635.104614177
#> 15    271047.968344624
#> 16    531500.226739846
#> 17    182932.352414741
#> 18     281129.36721924
#> 19    519711.219683666
#> 20    184624.774705393
#> 21    274268.139787934