Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for class 'flFit'
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP     1             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP     1             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP     1             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP     1          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP     1          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP     1          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP     1          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP     1          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP     1          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP     1           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP     1           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP     1           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP     1           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP     1           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP     1           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP     1           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP     1           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP     1           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP     1             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP     1             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP     1             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.5213872435023  15.5987447817493
#> 2                 FALSE          0.75 33.1725790416079  31.6182224041606
#> 3                 FALSE          0.75 65.2895982907073  20.2892193561726
#> 4                 FALSE          0.75 242.817699528216  15.0176258568039
#> 5                 FALSE          0.75 65.2072839611544  22.7861495627657
#> 6                 FALSE          0.75 204.813961489213  12.3901828260063
#> 7                 FALSE          0.75 507.319374361024  7.23603860439045
#> 8                 FALSE          0.75 202.510212692418  13.1444895258791
#> 9                 FALSE          0.75 291.038594046276  4.68330437636535
#> 10                FALSE          0.75 659.551724837857 0.604227464753907
#> 11                FALSE          0.75 322.680944372277  9.65930495390285
#> 12                FALSE          0.75 334.034882313756  3.70869078160293
#> 13                FALSE          0.75 714.473856911692 0.645319040827579
#> 14                FALSE          0.75 344.164037013385  9.88055061876359
#> 15                FALSE          0.75 321.316046888237  3.35316711289008
#> 16                FALSE          0.75 778.854400400082 0.576757903267561
#> 17                FALSE          0.75 339.608959097071  8.37668171051924
#> 18                FALSE          0.75 351.716929191204  5.11377153414047
#> 19                FALSE          0.75 779.723265140738 0.627030709284753
#> 20                FALSE          0.75 350.752932418366  9.62039607672537
#> 21                FALSE          0.75 339.559233848715  5.38939205373965
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3415.75348256876 2119.78985673881 81571.0053951323 2.06568399104687
#> 2   1463.2837471301  269.35647199142 42372.8518233673  1.8557012612374
#> 3  2195.77617900399 674.982966388483 65523.4043382328 1.59987337821221
#> 4  7576.09841026955  6053.0079905398 149876.258354591 2.64015329182118
#> 5  2365.37045378971 1094.16441149522 57188.9019862296 2.99003983695897
#> 6  4817.10754674216 3227.43627757527 130183.083461398 2.74285521390135
#> 7  17022.9617366286 15717.2479266137   360825.9398968 13.1567788717553
#> 8  5185.60047752503 3828.68823375731  118457.08246533  8.9073140411367
#> 9  7766.45178983998 6289.85125368692 222434.496550874 10.0242318845652
#> 10 20110.1844578772 18657.8383477193 487112.425706482 29.5392781039055
#> 11 6792.67517473254 5636.82058703451 167741.185318876 12.2376002461142
#> 12 9192.12419124151 7685.51726132473 268617.278672522 6.65965677322641
#> 13 20413.7465002633 19328.8927951398 489470.926572003 24.1772926813561
#> 14 6983.75952676431 5644.21673207673 177575.608718189 10.8124799239313
#> 15 8999.41458142694  7419.4209320978 265065.295664661  6.7421277292022
#> 16 20676.2680358757 19310.0586705199 512857.290944285  30.131481124556
#> 17 6901.08654547344 5951.98591804959 175894.314535927 15.9549207492416
#> 18 9320.19562230127 7624.45365915058  270074.06751817 5.10238766818822
#> 19 19781.4809908501 18619.9501364044 503077.841417369 23.6621344415991
#> 20 7170.33818742448 5809.73567546035 181437.467818927 15.6102210952044
#> 21 9256.19435970487 7421.12954584325  268157.02252679 6.41992292705003
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.748835222258036 25.5794188332312 71.5174383679278 1490.17010249661
#> 2   2.14301792975676  21.484186036436   87.28127325831 939.242143008907
#> 3   1.75822756446127   4.774011212333  122.05231045477 1621.08123541931
#> 4  0.278344496943128  43.065370166325 93.8894626983834 2005.88605812086
#> 5   1.83001411729906  44.568074959621 155.325254098316 1581.34086812138
#> 6  0.566184435928185 9.99242070543473 102.497245263124 3032.41651434508
#> 7  0.613242238846449 134.899768735939 288.918942621333  7519.3071464242
#> 8  0.589752303708225 41.0538207462721 124.644081873828 2269.09467982763
#> 9  0.467412277364675 19.1826685715265 63.6483452211557 2533.93956472708
#> 10 0.368821169101467 89.5051638677971  257.19062868731 6465.40020569086
#> 11  0.53493943598979 81.8048984516694 117.064255570476 2585.64808423081
#> 12 0.355541134369938 11.6148497008301 82.5951905498349 2588.83741940813
#> 13 0.279207632420053 92.1499798617774 163.254433523801 8743.77731703591
#> 14  0.64811370212081 36.0772270014388 222.944666729322 2329.46905386717
#> 15 0.357240923579373 18.5396393231368 71.8822488420928 4286.50649146139
#> 16 0.575889473016909 83.6641818518513 423.210624593402 11387.5128471402
#> 17 0.619847771895343 91.8241209660982 202.634767521117 3712.50427961253
#> 18 0.398673424354048  7.7280522495693 156.813256805643 5841.98155499323
#> 19 0.258332392347734 56.7634600570247 205.160264485196 6683.74162424722
#> 20   0.3715715671201 49.6009614462971 127.077469748071 1475.14655659653
#> 21 0.806386356268581 7.87894226160996 247.967294941194 5141.93448892975
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE 89.1233370782302 95.9194374087745     14.3669108411348
#> 2             TRUE 30.1199504668724 36.2252076163434     28.0929579097108
#> 3             TRUE 62.6578065835482 67.9213899978664     17.3969350126338
#> 4             TRUE  238.47464736317 247.160751693262     14.5597491593325
#> 5             TRUE 60.2886684293569 70.1258994929519     19.7757763398088
#> 6             TRUE 200.301964662346 209.325958316081     11.4588094289044
#> 7             TRUE 485.676473116987 528.962275605062     6.22725512148804
#> 8             TRUE 187.857681094748 217.162744290088      12.174346986279
#> 9             TRUE 274.548732596166 307.528455496385     3.91441118010046
#> 10            TRUE 610.959612356932 708.143837318781 -0.00248335841800584
#> 11            TRUE 302.550091967419 342.811796777135     8.77932958169964
#> 12            TRUE 323.079746921799 344.990017705714     3.12382561556438
#> 13            TRUE 674.702210450861 754.245503372523    0.186022485496591
#> 14            TRUE 326.377507538518 361.950566488252     8.81440357877485
#> 15            TRUE 310.225246773699 332.406847002774     2.76550579360201
#> 16            TRUE 729.288113950188 828.420686849977   -0.370580279845255
#> 17            TRUE 313.363114464568 365.854803729573     7.35703212575141
#> 18            TRUE 343.323501477034 360.110356905373     4.45795375107806
#> 19            TRUE 740.799053984308 818.647476297169    0.202073923872731
#> 20            TRUE 325.074118716755 376.431746119977      9.0091608488128
#> 21            TRUE 328.998460633718 350.120007063713     4.06288649767783
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.8305787223638 3373.67533858809 3457.83162654942    79119.6755765254
#> 2   35.1434868986105 1427.94226110016 1498.62523316003    40827.7984981176
#> 3   23.1815036997114 2187.92293055971 2203.62942744828    62856.7257059681
#> 4   15.4755025542754 7505.25587634594 7646.94094419315    146576.575788982
#> 5   25.7965227857227 2292.05597048113 2438.68493709828    54587.5962581699
#> 6   13.3215562231081 4800.67001468172  4833.5450788026    125194.758295301
#> 7   8.24482208729286  16801.051617058 17244.8718561992    348456.679640933
#> 8   14.1146320654791 5118.06694239742 5253.13401265265    114724.421717014
#> 9   5.45219757263024 7734.89630003982 7798.00727964014    218266.165966898
#> 10  1.21093828792582 19962.9484633147 20257.4204524397     476476.84236812
#> 11  10.5392803261061 6658.10611677955 6927.24423268554    163487.794220317
#> 12  4.29355594764148 9173.01776348365 9211.23061899938    264358.641117596
#> 13  1.10461559615857 20262.1597833906 20565.3332171359    475087.412885479
#> 14  10.9466976587523 6924.41248834694 7043.10656518167    173743.632124577
#> 15  3.94082843217815 8968.91687474038  9029.9122881135    258013.992486207
#> 16  1.52409608638038 20538.6404567294  20813.895615022     494124.83231074
#> 17  9.39633129528708 6750.03586648421 7052.13722446267    169787.244995965
#> 18  5.76958931720288 9307.48297635073 9332.90826825181    260464.007860206
#> 19  1.05198749469678 19688.1050990562 19874.8568826439    492083.086445483
#> 20  10.2316313046379 7088.74460584532 7251.93176900364    179010.851733326
#> 21  6.71589760980146 9243.23349968452 9269.15521972522      259698.5402925
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84022.3352137393 88.4726466210505 96.5701278659542   14.1310277461235
#> 2     43917.9051486169 29.5354045695826 36.8097535136332   27.4179072618374
#> 3     68190.0829704976 62.1538464694113 68.4253501120032   16.8430933298285
#> 4       153175.9409202 237.642999076247 247.992399980186   14.4720706427954
#> 5     59790.2077142893 59.3468058807148 71.0677620415939   19.1993218928596
#> 6     135171.408627496 199.437965269967  210.18995770846    11.280461331587
#> 7     373195.200152668 481.532087772384 533.106660949665   6.03408381625141
#> 8     122189.743213647  185.05187717179 219.968548213046   11.9885750106109
#> 9      226602.82713485 271.391099552528 310.686088540023   3.76717631273059
#> 10    497748.009044843 601.654739754202 717.448709921512 -0.118662026684968
#> 11    171994.576417436 298.695247889893 346.666640854661   8.61082365936286
#> 12    272875.916227448 320.981955038232  347.08780958928   3.01183015823785
#> 13    503854.440258527 667.086363256234  761.86135056715 0.0980720812842747
#> 14      181407.5853118  322.97157636248  365.35649766429    8.6102477626068
#> 15    272116.598843115    308.101476539 334.530617237473   2.65297490267451
#> 16    531589.749577831 719.796697395953 837.912103404212 -0.551985463845581
#> 17     182001.38407589 308.337314428557 370.880603765584   7.16178007760437
#> 18    279684.127176133 341.716249361555 361.717609020852   4.33237162240654
#> 19    514072.596389256 733.345481635204 826.101048646273  0.120699220283195
#> 20    183864.083904529 320.156899071765 381.348965764967   8.89211580516997
#> 21    276615.504761079 326.976184911697 352.142282785733   3.80887479545323
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    17.066461817375 3365.61782165563 3465.88914348189     78650.271994239
#> 2   35.8185375464839 1421.17474249868 1505.39275176151    40531.9372230698
#> 3   23.7353453825167 2186.41911702782 2205.13324098017     62346.085116811
#> 4   15.5631810708125 7491.69028474355 7660.50653579554    145944.721680674
#> 5   26.3729772326719 2278.01702686885 2452.72388071057    54089.4738847117
#> 6   13.4999043204255  4797.5224021595 4836.69269132481    124239.547093282
#> 7   8.43799339252949 16758.5581899061  17287.365283351    346088.097889809
#> 8   14.3004040411472 5105.13498886234 5266.06596618773    114009.656892868
#> 9   5.59943244000012 7728.85375943979 7804.04982024017    217467.975004009
#> 10  1.32711695619278 19934.7543366963 20285.6145790581    474440.241303328
#> 11  10.7077862484428 6632.33757376727 6953.01277569782    162673.315073784
#> 12  4.40555140496801 9169.35908582788 9214.88929665514    263543.157330482
#> 13  1.19256600037088 20233.1325397342 20594.3604607924    472333.123030612
#> 14  11.1508534749204 6913.04816184149 7054.47089168713    173009.849372609
#> 15  4.05335932310565 8963.07688835359 9035.75227450029    256663.742941396
#> 16   1.7055012703807 20512.2862394461 20840.2498323053     490537.76576389
#> 17  9.59158334343412 6721.11126837989 7081.06182256699    168617.806147887
#> 18  5.89517144587441 9305.04863989211 9335.34260471042    258623.783670383
#> 19  1.13336219828631 19670.2246091383 19892.7373725618    489977.707833845
#> 20  10.3486763482808 7073.12030298974 7267.55607185922    178546.180567998
#> 21  6.96990931202607 9240.75163287212 9271.63708653763    258078.830928487
#>    ci95.integral.bt.up
#> 1     84491.7387960257
#> 2     44213.7664236648
#> 3     68700.7235596547
#> 4     153807.795028508
#> 5     60288.3300877475
#> 6     136126.619829515
#> 7     375563.781903792
#> 8     122904.508037793
#> 9     227401.018097739
#> 10    499784.610109636
#> 11    172809.055563969
#> 12    273691.400014562
#> 13    506608.730113393
#> 14    182141.368063769
#> 15    273466.848387925
#> 16     535176.81612468
#> 17    183170.822923968
#> 18    281524.351365956
#> 19    516177.975000894
#> 20    184328.755069856
#> 21    278235.214125092