Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2721635950595  15.4690760721507
#> 2                 FALSE          0.75 33.2343322024787  31.5428228308711
#> 3                 FALSE          0.75 64.6571083055199  20.7735241575404
#> 4                 FALSE          0.75 242.431501700014  14.9887584556641
#> 5                 FALSE          0.75 66.1215898610865  21.8054992509318
#> 6                 FALSE          0.75 206.535537965005  12.4809828493349
#> 7                 FALSE          0.75 508.317345341535  7.04588179584248
#> 8                 FALSE          0.75 203.434049988055  13.3727711057108
#> 9                 FALSE          0.75 287.907072115433   4.6103940970366
#> 10                FALSE          0.75 672.376031319947 0.715312225652348
#> 11                FALSE          0.75 326.149882043578  10.0193042576082
#> 12                FALSE          0.75 334.360894075675  3.81564348184207
#> 13                FALSE          0.75 727.685070203138 0.665119178758216
#> 14                FALSE          0.75 334.530616270349  9.64102322044051
#> 15                FALSE          0.75 321.866327982199  3.45608400131568
#> 16                FALSE          0.75 779.759071202201 0.420870063920438
#> 17                FALSE          0.75 334.245042199249  7.91395750997648
#> 18                FALSE          0.75  347.90959597219  5.05593916877121
#> 19                FALSE          0.75 784.657812444864  0.83938341239176
#> 20                FALSE          0.75 353.000281477963  9.70525302227239
#> 21                FALSE          0.75 333.118635517931  4.91780950645029
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.67558016104 2130.66399366202 81540.1378781503 2.37003592093654
#> 2  1459.95415697435   269.3398224064 42291.5161597607 2.65264533151868
#> 3  2195.82518768948 649.434897632755 66113.8285086445 1.04829111814554
#> 4  7556.19202051411 6032.53697316692 149236.141159573  2.9464447928655
#> 5  2369.19379534949 1159.94276058368 57068.8166385456 2.92338405430569
#> 6  4817.22428396208 3212.57899490944 130839.332628028 2.68762102415772
#> 7  17065.5247686765 15885.2782108222 363634.970253237 12.9907806647571
#> 8  5185.92337256017 3793.71449881959 119004.080205322 7.51979809996593
#> 9  7774.77690027177  6296.8401725699 222673.337337752 8.88229302691797
#> 10 20067.4279756671 18609.2932781258 485746.280488115 34.5203314482648
#> 11  6786.9786209993 5518.94122713703 168383.820765516 12.9811929155623
#> 12 9186.79711617186 7650.30139248368 267818.428611336 6.25074153017591
#> 13  20460.310951193 19486.0413757493 492015.120215971 20.9358902335344
#> 14 6972.91349952575 5660.32486834855 176892.366499288 9.71881823224716
#> 15 8997.62206282077 7387.09800009199 265720.527728666 7.23496857081944
#> 16 20653.3972368283 19387.7437794372 512132.226928931 22.3544886319775
#> 17 6888.27110197305 6041.50561115611 174974.489100082 15.3849247604615
#> 18 9321.98114782326 7588.29278954732 270894.411295095 8.36218431009501
#> 19  19789.371383829 18497.5730428445 503460.513414716 29.2504876889305
#> 20  7100.0309035851 5718.66714655503 179716.624498733 13.0638772130237
#> 21  9254.2454223663 7501.67896148023 269212.516392825 6.52484120717499
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.80053217291132 25.4222126093473 77.6777728007969 1494.08282221234
#> 2   2.30610556405128 21.6838917980368 67.6404966356979 917.109161824419
#> 3   1.63556713328071 3.78022543083349 99.4223436268305 1125.35191615713
#> 4  0.280193496990586 57.6978747368451  72.525273626083 3117.07389457814
#> 5   2.38253546650896 26.2654265749724 137.026477469418  1273.7480445763
#> 6  0.385309678191529 8.15013738774917 72.4481823890525 1998.98428300533
#> 7   0.41967882354434 103.897614118253 118.488339003986 5848.13316931761
#> 8  0.548235932508189  34.640709843839 100.358281418445 2199.54919914162
#> 9  0.429088494673257 17.4107956807421  63.829614389657 1915.09876363847
#> 10 0.550602098659986 90.6223607397299 324.178818921962 5889.50242990275
#> 11 0.641913682613309 84.8623005497325 197.770276487165 3660.48874022729
#> 12 0.414890021763328  12.873552037718 110.173574545166 2739.43258572609
#> 13 0.211555105318919 76.9359273415915  140.37340235539 7902.51152779039
#> 14 0.770354460335558 54.1723739057209 249.331755513589  2700.5820888453
#> 15 0.620218029233842 14.6254242545784 136.168079953975 2275.73040421055
#> 16 0.347074115764602 109.197890207295 178.540184313272 9633.57124189749
#> 17 0.581694023282588 87.4496633489957 171.931224859744 2778.29573379888
#> 18 0.887033799062289 9.50836249608348 267.066653504103 3584.22821937613
#> 19 0.627295892911751 79.7678569528863 385.239907129987 8032.41945930573
#> 20 0.370494514403002  92.010366409888 146.448199985822 3663.36297034487
#> 21  0.49498334500025  12.598112356413 140.868284809489 3803.03831194162
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.3734545051188 96.1708726850001   14.1522006477116
#> 2             TRUE 28.8707306321305  37.597933772827   27.7492791780067
#> 3             TRUE 62.9326694161704 66.3815471948693   18.0830162232937
#> 4             TRUE 237.584600015751 247.278403384278   14.5278401531146
#> 5             TRUE 61.3126230917536 70.9305566304194   17.8862284085246
#> 6             TRUE 202.114401380265 210.956674549744   11.8471484287099
#> 7             TRUE 486.947511148009  529.68717953506   6.35551013111205
#> 8             TRUE 191.063982113611 215.804117862499   12.4709229967348
#> 9             TRUE 273.295700086153 302.518444144713   3.90454352329909
#> 10            TRUE 615.590086087551 729.161976552343 -0.190428226643328
#> 11            TRUE 304.795819697478 347.503944389678   8.96335624970929
#> 12            TRUE 324.078424258535 344.643363892814   3.13314939604139
#> 13            TRUE 693.245530768974 762.124609637302  0.317111030508595
#> 14            TRUE 318.543160278302 350.518072262395   8.37379013318852
#> 15            TRUE 309.964804683201 333.767851281197   2.43582534322601
#> 16            TRUE 742.985937402598 816.532205001804 -0.150066856512332
#> 17            TRUE  308.93684096829 359.553243430209   6.95707084167662
#> 18            TRUE 334.153802782084 361.665389162296   3.59676856931374
#> 19            TRUE 736.540760196574 832.774864693155  -0.19251833144807
#> 20            TRUE  331.51020346254 374.490359493387   9.09578954607945
#> 21            TRUE 322.385271732128 343.851999303734   4.10356190392488
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7859514965898 3370.85604041866 3454.49511990341     79082.371635611
#> 2   35.3363664837355 1424.28415496658 1495.62415898212    40782.8715885596
#> 3   23.4640320917872 2189.60671685576  2202.0436585232     64262.624606566
#> 4   15.4496767582137   7461.279016572 7651.10502445622    144108.554602992
#> 5   25.7247700933391 2325.98716863366 2412.40042206532    54973.5011052176
#> 6     13.11481726996 4803.81730795923 4830.63125996493    127551.003482485
#> 7   7.73625346057292  16894.613193452 17236.4363439011     354014.79118971
#> 8   14.2746192146868 5128.93940486705 5242.90734025328    115385.821772734
#> 9    5.3162446707741 7746.13614137694 7803.41765916659    219522.999871567
#> 10  1.62105267794802 19918.3541922503  20216.501759084    476058.048990924
#> 11  11.0752522655071 6647.38013659499 6926.57710540361    162362.316787842
#> 12  4.49813756764274 9165.62012306981  9207.9741092739    263312.062007816
#> 13  1.01312732700784 20333.7513507161 20586.8705516699    479015.488752756
#> 14  10.9082563076925 6883.79994445084 7062.02705460066    172449.908963138
#> 15  4.47634265940535 8973.56323992199 9021.68088571955     261976.95121374
#> 16 0.991806984353207 20473.7667074373 20833.0277662193    496285.002236009
#> 17  8.87084417827633 6744.41640576395 7032.12579818214    170404.192617982
#> 18  6.51510976822867  9306.3398915172 9337.62240412931    264998.355874221
#> 19  1.87128515623159 19658.1532591415 19920.5895085165    490247.183404158
#> 20  10.3147164984653 6948.67385084084 7251.38795632937    173690.392412515
#> 21   5.7320571089757    9233.52152754  9274.9693171926    262956.518369681
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83997.9041206896 87.6268931900238 96.9174340000951   13.9000330132445
#> 2     43800.1607309619 28.0351473527021 38.4335170522553   27.0228559253306
#> 3      67965.032410723 62.6024577139546 66.7117588970851   17.5678125763102
#> 4     154363.727716154 236.656469905998 248.206533494031   14.4395792015626
#> 5     59164.1321718737 60.3917571146473 71.8514226075256   17.1357297365742
#> 6     134127.661773572 201.267800757655 211.803275172354   11.7257758800795
#> 7     373255.149316765 482.855415238611 533.779275444459   6.22331130169558
#> 8      122622.33863791 188.695245712122 218.172854263988   12.2982286779948
#> 9     225823.674803938 270.497777782674 305.316366448192   3.76938064747701
#> 10    495434.511985305 604.716181681348 740.035880958546 -0.363867887721224
#> 11     174405.32474319 300.706743929076  351.59302015808    8.7611534396861
#> 12    272324.795214855  322.10944067653 346.612347474819   3.00245903918595
#> 13    505014.751679187 686.650725345411 768.719415060866  0.250471172333136
#> 14    181334.824035439 315.481732535144 353.579500005553   8.13112847818281
#> 15    269464.104243593 307.685789583393 336.046866381005   2.24045666401735
#> 16    527979.451621852 735.944273483525 823.573868920877 -0.259395202978182
#> 17    179544.785582181 304.090589668745 364.399494729754    6.7738372243426
#> 18    276790.466715968 331.519714724404 364.299477219976   3.31735292260912
#> 19    516673.843425274  727.32685657456 841.988768315168 -0.390116537715271
#> 20     185742.85658495 327.395082140437  378.60548081549   8.97908377404251
#> 21    275468.514415969 320.329946751868 345.907324283994    3.9476421502498
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0381191310569 3362.84804344672 3462.50311687536    78611.7355466141
#> 2   36.0627897364116  1417.4537290502  1502.4545848985    40493.9822025849
#> 3   23.9792357387706 2188.41594584505 2203.23442953391    63908.1387529765
#> 4   15.5379377097657  7443.1041860299 7669.27985499833      143126.6763262
#> 5   26.4752687652894 2317.71355926254 2420.67403143644    54572.2704711761
#> 6   13.2361898185903 4801.25001468209 4833.19855324207    126921.323433338
#> 7   7.86845228998939 16861.8854450048 17269.1640923483    352172.629241375
#> 8   14.4473135334269 5118.02758126624 5253.81916385409    114692.963775005
#> 9   5.45140754659618 7740.65174073751 7808.90205980602    218919.743761021
#> 10  1.79449233902592 19889.8081486173  20245.047802717    474202.855725505
#> 11  11.2774550755303 6620.64851192182 6953.30873007677     161209.26283467
#> 12  4.62882792449819 9161.56495417793 9212.02927816578    262449.140743312
#> 13   1.0797671851833 20309.5165336035 20611.1053687825    476526.197621502
#> 14  11.1509179626982 6866.73564667054 7079.09135238096    171599.225605151
#> 15  4.67171133861401  8968.9562312818 9026.28789435975    261260.096136414
#> 16  1.10113533081906  20439.369372022 20867.4251016346    493250.427294812
#> 17  9.05407779561035 6716.86976180901 7059.67244213708    169529.029461836
#> 18  6.79452541493329 9303.34475733093 9340.61753831558    263869.323985117
#> 19  2.06888336249879 19633.0263842014 19945.7163834567    487716.971274477
#> 20  10.4314222705023 6919.69058542172 7280.37122174848    172536.433076857
#> 21  5.88797686265078 9229.55312214773 9278.93772258487    261758.561301419
#>    ci95.integral.bt.up
#> 1     84468.5402096865
#> 2     44089.0501169366
#> 3     68319.5182643124
#> 4     155345.605992946
#> 5     59565.3628059152
#> 6     134757.341822719
#> 7       375097.3112651
#> 8      123315.19663564
#> 9     226426.930914484
#> 10    497289.705250724
#> 11    175558.378696361
#> 12    273187.716479359
#> 13    507504.042810441
#> 14    182185.507393425
#> 15    270180.959320919
#> 16     531014.02656305
#> 17    180419.948738327
#> 18    277919.498605072
#> 19    519204.055554955
#> 20    186896.815920609
#> 21     276666.47148423