Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xls', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xls', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.7908101175902  15.7332009018935
#> 2                 FALSE          0.75 32.4023681587463  31.4471749914891
#> 3                 FALSE          0.75 64.6202486996163  20.9170278447465
#> 4                 FALSE          0.75 242.696500622968  14.8567346166699
#> 5                 FALSE          0.75 65.5645773443677  22.4586556436748
#> 6                 FALSE          0.75 205.952168093612  12.4489833189391
#> 7                 FALSE          0.75 511.674656344177  7.12101722504811
#> 8                 FALSE          0.75 199.599842818091  13.1428563936935
#> 9                 FALSE          0.75 290.572644730562  4.63099096772431
#> 10                FALSE          0.75 674.263432585639 0.543896850754324
#> 11                FALSE          0.75 324.385148607015  9.87861053620838
#> 12                FALSE          0.75  332.39952079975  3.74748453213162
#> 13                FALSE          0.75 730.862324560915 0.796041554548969
#> 14                FALSE          0.75 343.338842280596   9.8332889863791
#> 15                FALSE          0.75 317.424502667522  3.32706480124453
#> 16                FALSE          0.75 774.391460712391 0.437186148292789
#> 17                FALSE          0.75 336.897614871721  8.17958836185464
#> 18                FALSE          0.75 351.436114564713  4.93522679651427
#> 19                FALSE          0.75 781.159570884114 0.650316210556982
#> 20                FALSE          0.75 346.188613414026  9.44303176044143
#> 21                FALSE          0.75 336.628799466102  5.20150512622107
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3416.69405553951 2113.74260276315 81909.9904525567 2.39987527712994
#> 2  1464.09165630346 287.187085380696 42530.6930324425 2.32830109354138
#> 3  2197.26813665318 642.364453149552 66421.1884633376 1.52365323524975
#> 4  7563.87702139862 6070.78342839403 149289.226910713 2.07400108182099
#> 5  2371.48527071978  1146.6408888849 57639.5837320367 2.85359156618332
#> 6  4820.26936156103 3219.78282473283  131002.16874708 2.06075071515294
#> 7  17053.0349288279 15871.5509775675 363228.676743315 13.8261070262712
#> 8  5174.00544221906 3780.41056045836 118418.165112931 9.35872819858045
#> 9  7766.26252349955 6299.37124987929 221120.427077055 5.04453247898023
#> 10 20081.5595855279 18754.8193571964  486633.31346739 18.8406029080541
#> 11 6771.23636131561 5553.57489619918 167279.311593135 12.9299975213389
#> 12 9191.23455577629 7650.95628004185 265426.951919634 5.80843230727906
#> 13 20439.2090942993 19388.8457901973 488513.616421184 22.9722101178155
#> 14  6926.0538175245 5600.28715260776 174432.749701174 17.8020655497009
#> 15 8989.75895198277 7368.21617543394 264691.900082606 6.91399110348793
#> 16 20666.2517268566 19404.9839994612 514229.749979519 19.5476173541828
#> 17 6908.07559768189 6004.33298142552 176249.275796315  14.288687063142
#> 18 9318.78012381238 7670.01632477534 269535.909830915 8.89284667577651
#> 19  19794.539214497 18658.7373664364 502587.307871317 22.4905359134447
#> 20 7134.52708800731 5810.61692915107 179573.701504772 16.8551649545349
#> 21 9257.69646185185 7464.08842363903 268991.481059609 10.5004070345492
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.795669988346446 15.7518920155484 75.2301080603738 1157.86902253291
#> 2   1.95010309109246 22.3299795660374 56.7462908185357  1200.1651740832
#> 3   1.72149188754795 6.31607096407545 106.037613240942 1399.02556187689
#> 4  0.279971559388728 44.1923365759009 73.6849744210807 2327.94000860858
#> 5   2.65796115322501  25.490472060435 122.454171538622 831.628173378405
#> 6   0.36724523419213 6.61684613218294 74.1577242806191 2120.21416873599
#> 7  0.501318372610054 154.583521676323 206.270934245115 8376.45482607779
#> 8  0.950333579291329 57.6780407588479 186.198304364186  3547.5175013342
#> 9  0.386841536705513 18.6365007395926 86.4418022830728 2886.77454592704
#> 10 0.311184070794688 86.3470035928303 207.134290656938 5205.61592693676
#> 11 0.588928089739578 80.8420024120054  202.02710582637 3422.33092688736
#> 12 0.502160447158059  12.780770311546 147.974440625219 4972.18483141392
#> 13  0.44923091471864 87.7485564054276 265.380318737124 8382.80154868812
#> 14 0.676443461597784 97.2786260775535 187.686606434027 5524.45548074464
#> 15 0.601117290224115 25.7072263337733 137.807477836084 4539.19730200422
#> 16 0.278925389071487 101.387103318656 251.457043404572 10870.6021180229
#> 17 0.698938186154406 74.1297348867696 242.859113194751 2581.00372893218
#> 18 0.572864396535414 10.3757852778442   138.2049909819 4549.98781732131
#> 19 0.201224082235314 90.4960885083379  153.83060234551 9999.26750077612
#> 20  0.36858658515023 109.010348730549 148.515334401633 3877.93362917277
#> 21 0.695122332728443 9.92646398421627 155.575190692639  2933.7854326541
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.8430152867114 96.7386049484689    14.4243237710636
#> 2             TRUE 28.5723128598707 36.2324234576219     28.239255406642
#> 3             TRUE 62.1138391276304 67.1266582716021    18.0851736897301
#> 4             TRUE 239.284768843372 246.108232402563    14.3961814014755
#> 5             TRUE 60.8704192179961 70.2587354707392    18.0863095466197
#> 6             TRUE 202.562233167186 209.342103020039    11.8448649086931
#> 7             TRUE 488.930710285961 534.418602402393    6.29634850210457
#> 8             TRUE 184.204734931426 214.994950704756    11.5795576557592
#> 9             TRUE  282.27438880264 298.870900658485    3.99463663984374
#> 10            TRUE  643.27064080189 705.256224369388  0.0319990542970624
#> 11            TRUE 303.115302684412 345.654994529617    8.90982382858678
#> 12            TRUE 322.844649654276 341.954391945224    2.92143059655662
#> 13            TRUE 693.073038917108 768.651610204721  0.0570566998368059
#> 14            TRUE 314.054444451338 372.623240109854    8.72053949205075
#> 15            TRUE 306.050987302284  328.79801803276    2.33822685882586
#> 16            TRUE  742.23563016476 806.547291260021 -0.0216461167298073
#> 17            TRUE 313.392724652852 360.402505090589    7.02983504563065
#> 18            TRUE  336.80738178306 366.064847346365    3.99286486421351
#> 19            TRUE 744.162639306498 818.156502461731   0.319302595279889
#> 20            TRUE 318.461867063816 373.915359764236     8.8367068278693
#> 21            TRUE 319.355629894269 353.901969037935    4.05802888888278
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   17.0420780327234 3390.78219317393 3442.60591790508      80005.29591049
#> 2   34.6550945763362 1427.35883991733 1500.82447268959    40556.4213210757
#> 3   23.7488819997629 2186.87819991727 2207.65807338908    64119.7914140501
#> 4   15.3172878318644 7491.18062773126 7636.57341506598    145459.765596552
#> 5   26.8310017407299 2329.55344418037  2413.4170972592    56271.5553868293
#> 6   13.0531017291852 4809.38464967359 4831.15407344847     127514.41643951
#> 7   7.94568594799165 16798.7450356704 17307.3248219855    349449.408554417
#> 8   14.7061551316277 5079.12506517076 5268.88581926737    112582.498823236
#> 9   5.26734529560488 7735.60547978292 7796.91956721618    216371.682949005
#> 10  1.05579464721159 19939.5187646177 20223.6004064381    478070.075267579
#> 11    10.84739724383 6638.25126734786 6904.22145528336    161649.577218405
#> 12  4.57353846770663  9170.2101886138 9212.25892293878    257247.707871958
#> 13  1.53502640926113 20294.8627190124 20583.5554695863    474723.907873593
#> 14  10.9460384807075 6766.03047762693 7086.07715742208    165345.020435349
#> 15   4.3159027436632 8947.47056466371 9032.04733930182    257224.920520809
#> 16 0.896018413315386 20499.4699418974 20833.0335118158    496347.609495371
#> 17  9.32934167807864 6786.13218379315 7030.01901157062    172003.524662222
#> 18  5.87758872881502 9301.71195703032 9335.84829059443    262051.179871422
#> 19 0.981329825834074 19645.6731489008 19943.4052800932     486138.51283254
#> 20  10.0493566930136 6955.20506434556 7313.84911166907    173194.500684783
#> 21  6.34498136355936 9241.36742859781 9274.02549510588    264165.404022893
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83814.6849946233 88.0870545744155 97.4945656607649    14.1736877247344
#> 2     44504.9647438094 27.8388980154052 36.9658383020874    27.6249729329478
#> 3     68722.5855126251 61.6338883585268 67.6066090407058    17.5429037451525
#> 4     153118.688224874 238.631458502598 246.761542743337     14.307990360268
#> 5     59007.6120772442 59.9715378746483  71.157616814087    17.2490517833538
#> 6     134489.921054651 201.913096691913 209.991239495312    11.7291826599225
#> 7     377007.944932213 484.575486572685 538.773826115668     6.1384332147324
#> 8     124253.831402626 181.256735548873 217.942950087309    11.2802025782825
#> 9     225869.171205105 280.685361071761 300.459928389363    3.87278155578151
#> 10    495196.551667201 637.335850885853 711.191014285425 -0.0660239280032644
#> 11    172909.045967865  299.04235346519 349.727943748839    8.72431148031881
#> 12     273606.19596731 321.014993477483 343.784048122017    2.76325005570183
#> 13    502303.324968776 685.836792729996 775.887856391833 -0.0844510382995657
#> 14    183520.478966999 308.446793803182  378.23089075801    8.50745980164745
#> 15    272158.879644403 303.873080104686 330.975925230358    2.14887491240527
#> 16    532111.890463666 736.078130698192 812.704790726589  -0.109507614287326
#> 17    180495.026930409 308.891788227962 364.903441515479    6.80966951699201
#> 18    277020.639790409 334.006135080191 368.866094049235    3.81241257930486
#> 19    519036.102910094 737.078120493763 825.241021274466   0.255917009375765
#> 20    185952.902324761 313.152490103138 379.224736724915    8.72060205354698
#> 21    273817.558096325 316.048001678386 357.209597253818    3.83906535407332
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.2927140790525 3385.82034718903 3447.56776388998    79640.5671683922
#> 2   35.2693770500303 1420.32489635402 1507.85841625289    40178.3692912395
#> 3   24.2911519443405 2184.88863756359 2209.64763574277    63679.0983620589
#> 4   15.4054788730718 7477.26004170985 7650.49400108739    144726.464493841
#> 5   27.6682595039958 2321.52394548133 2421.44659595823    56009.5925122151
#> 6   13.1687839779557 4807.30034314195 4833.23837998011    126846.548976358
#> 7   8.10360123536381 16750.0512263423 17356.0186313135    346810.825284203
#> 8   15.0055102091045 5060.95648233172  5287.0544021064    111465.030810316
#> 9   5.38920037966712 7729.73498204995 7802.79006494916    215462.348967038
#> 10  1.15381762951191  19912.319458486 20250.7997125698    476430.306250594
#> 11   11.032909592098 6612.78603658808 6929.68668604314    160571.542976436
#> 12  4.73171900856142 9166.18424596566 9216.28486558692    255681.469650062
#> 13   1.6765341473975 20267.2219237447  20611.196264854    472083.325385756
#> 14  11.1591181711108  6735.3877104125 7116.71992463651    163604.816958914
#> 15   4.5052546900838 8939.37278836857 9040.14511559696    255795.073370677
#> 16 0.983879910872904  20467.533004352 20864.9704493612    492923.369828194
#> 17  9.54950720671728 6762.78131730382 7053.36987805995    171190.508487608
#> 18  6.05804101372368  9298.4435846678 9339.11666295695    260617.933708965
#> 19   1.0447154117382 19617.1668810207 19971.9115479733    482988.743569796
#> 20  10.1654614673359 6920.86680449544 7348.18737151919    171972.951591593
#> 21  6.56394489836882 9238.24059244278 9277.15233126091    263241.261611606
#>    ci95.integral.bt.up
#> 1     84179.4137367212
#> 2     44883.0167736456
#> 3     69163.2785646163
#> 4     153851.989327586
#> 5     59269.5749518584
#> 6     135157.788517803
#> 7     379646.528202428
#> 8     125371.299415546
#> 9     226778.505187072
#> 10    496836.320684186
#> 11    173987.080209834
#> 12    275172.434189205
#> 13    504943.907456613
#> 14    185260.682443433
#> 15    273588.726794534
#> 16    535536.130130844
#> 17    181308.043105022
#> 18    278453.885952865
#> 19    522185.872172838
#> 20     187174.45141795
#> 21    274741.700507611