Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8810181165145  15.4900781185611
#> 2                 FALSE          0.75 33.0142900666779  30.7755149690121
#> 3                 FALSE          0.75 65.0428286334576  20.4055781836549
#> 4                 FALSE          0.75 242.424724758198  14.9848518077673
#> 5                 FALSE          0.75 65.0613107012946  21.4341847723798
#> 6                 FALSE          0.75 206.208572390969  12.5229169862106
#> 7                 FALSE          0.75 509.618138451409  7.18941936043906
#> 8                 FALSE          0.75 200.515629921205  13.4771670153594
#> 9                 FALSE          0.75 289.376854283208   4.5902656850418
#> 10                FALSE          0.75 670.381032411219 0.516773495289976
#> 11                FALSE          0.75 325.362639950699  9.83860085940953
#> 12                FALSE          0.75 332.874185416819    3.611630535765
#> 13                FALSE          0.75 722.243256812362 0.750306992669745
#> 14                FALSE          0.75 342.480309708144  9.59903319996631
#> 15                FALSE          0.75 318.627517298318  3.46034133998785
#> 16                FALSE          0.75  772.19970447655 0.411813937789752
#> 17                FALSE          0.75  334.98371321026  8.06521801666775
#> 18                FALSE          0.75 351.756409544426  5.03091187398735
#> 19                FALSE          0.75 773.164719491762 0.668255419721168
#> 20                FALSE          0.75 347.903847195854  9.47752807167911
#> 21                FALSE          0.75 338.064728737589   5.1995397428779
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3418.11214647172 2138.38106224467 81669.0413392506  3.0182761516698
#> 2  1454.11969976142 276.357938629784 42294.2806023777  2.7699831148066
#> 3  2194.76203600654 662.239746864561 65985.4399407627 1.58394909754266
#> 4  7572.96304679935  6049.7200433695 150062.122980514 2.58040222217247
#> 5  2348.99710158214 1142.97239329388   56704.95943872 1.64311446702045
#> 6  4818.14912153881 3202.95969971951 130268.152336429 2.78573026780678
#> 7  17033.6408179951 15817.5525635798 361295.069825431 11.7699693149609
#> 8  5180.20090976115 3753.74958522042 119311.099761687 7.64423038004514
#> 9  7763.78725416434 6294.32150924792 221704.097797394 7.83279340996374
#> 10 20056.1053717756 18769.1196994065 486389.074189412 20.5640242950454
#> 11 6776.69612338501 5603.83538716945  167046.33288081 11.6243157635972
#> 12 9192.30702596943  7693.2817307622 266055.911027846 6.09318617843735
#> 13 20440.5405346592 19379.1501652637 488578.810615271 20.2063525487768
#> 14 6971.27261072962 5721.84409566675 176629.892803389 14.7551544007463
#> 15 9012.67297455358   7366.417174202   264251.8642482 6.71955901716152
#> 16 20696.1724288146 19428.5854765727 516515.139710702 22.1652822949523
#> 17 6929.36582651446 6045.05461271289 176647.391110417 11.1537139752463
#> 18 9324.10869012128 7651.79616649442 271530.885603696 6.46143732309459
#> 19 19785.8448451545 18543.2640501188 502892.403458601 27.0409536132915
#> 20 7145.14175704808 5820.04798581194 180601.928351256 14.1189520539073
#> 21 9256.85848690041  7481.3598339513 268153.757253582 6.27354169512366
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.58548940901223 20.0560546432344 47.8998623201657 1141.80212124974
#> 2   2.75178644413221 20.8996975602798  85.716956744691 888.374388730273
#> 3    1.5957312320542  4.9029711584183 97.0448691220552 1162.74038639237
#> 4  0.287874364062838 31.3298271048074 62.5393017910504 1816.12107254091
#> 5   2.56823884197928 45.0411681931999 177.836008849792 1407.54770577938
#> 6  0.471228061289239 11.3076385145851  88.954737207172 2025.33979712294
#> 7  0.380651486894158 145.695041664135 195.016117667462 8341.18268104955
#> 8   0.72805555128505 41.2315485383886 106.501036597265 1820.99137519907
#> 9  0.416996777909562 23.6251003499673 55.6495267548376 3562.29467627287
#> 10  0.20342184010391 75.3235210914007 152.443964895559 9777.54845918358
#> 11 0.772440802409909 102.854317135955 264.759862359578 3237.03181584027
#> 12 0.360677261776413 9.61334806444856 84.7264072504759 4812.06967658458
#> 13 0.304035757391457 110.380124829607 219.952181586735 10048.7791078586
#> 14 0.576582574643486 39.1316340631767 178.440790296871 2969.39260178936
#> 15 0.439324937356909 15.3116727329721 126.852824861811 3791.38070622528
#> 16 0.284393858454978 102.731884043279 239.794947034235 9750.05478219761
#> 17 0.411407404696006 65.3473145208782 132.382410956458 2485.67762790191
#> 18 0.460197704202806 8.92682858484104 136.683062188699 3320.36157710457
#> 19 0.382811397825506 73.4007837893212 233.062756486221 12073.6245905446
#> 20  0.41549892755836 82.8675549174435 132.812172827708 2903.76836541271
#> 21 0.398649548654521 10.3923205081438 116.627873812513 3576.24052500821
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.9159538470177 97.8460823860113     14.526948040736
#> 2             TRUE  28.457667842821 37.5709122905347    26.2488262684146
#> 3             TRUE 62.4372323679999 67.6484248989153    17.7806003069257
#> 4             TRUE 238.179963102725 246.669486413672    14.5112984788839
#> 5             TRUE  62.358387403046 67.7642339995433    17.2094318773239
#> 6             TRUE 201.626046100427 210.791098681511    11.7477468253898
#> 7             TRUE 490.256538928298 528.979737974519    6.56324766449817
#> 8             TRUE 187.940870946031 213.090388896379    12.2795156334955
#> 9             TRUE 276.491909123817 302.261799442598    3.90430598538057
#> 10            TRUE 636.553212445869 704.208852376569   0.182144568319044
#> 11            TRUE 306.240640519582 344.484639381817    8.56793573944523
#> 12            TRUE  322.85089415329 342.897476680349    3.01831644014281
#> 13            TRUE 689.003806869624   755.4827067551   0.250168171760799
#> 14            TRUE 318.208080718916 366.752538697371    8.65055486467777
#> 15            TRUE 307.573842715087 329.681191881548    2.73765181803573
#> 16            TRUE 735.737815101353 808.661593851746 -0.0560139593686862
#> 17            TRUE  316.63585372098  353.33157269954    7.38845283594282
#> 18            TRUE 341.127345147935 362.385473940917    4.27388665057373
#> 19            TRUE 728.682350797897 817.647088185626  0.0385306702982107
#> 20            TRUE 324.678171067177 371.129523324532    8.79403233584561
#> 21            TRUE  327.74475264911 348.384704826067    4.54376123534121
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4532081963862 3385.11993658359 3451.10435635984    79790.7768497948
#> 2   35.3022036696096 1419.73969727476 1488.49970224808    40832.9047329164
#> 3   23.0305560603841 2186.69664845094 2202.82742356214    64072.7320051473
#> 4   15.4584051366507 7521.42548121194 7624.50061238676    147074.603816184
#> 5   25.6589376674357 2274.90437990432 2423.08982325995    54389.5434627129
#> 6   13.2980871470314 4799.54805618232  4836.7501868953    126936.468370162
#> 7   7.81559105637995 16793.9724744576 17273.3091615326    347573.824315104
#> 8   14.6748183972233  5112.3750124155  5248.0268071068    116315.568949485
#> 9   5.27622538470303 7724.92396408864 7802.65054424003    215844.123054925
#> 10 0.851402422260907 19932.1981795803  20180.012563971    470305.006974055
#> 11  11.1092659793738 6607.50077169636 6945.89147507365    161721.415543753
#> 12   4.2049446313872 9176.49306840342 9208.12098353545    258140.056409864
#> 13  1.25044581357869 20258.9652293145 20622.1158400039    472048.568982844
#> 14  10.5475115352548 6906.90107269569 7035.64414876354    171745.241973445
#> 15  4.18303086193996 8987.48527290784 9037.86067619932     258015.04298646
#> 16  0.87964183494819 20527.1784795634 20865.1663780658    500476.299593987
#> 17  8.74198319739268 6821.86949412762 7036.86215890131    172558.451412518
#> 18  5.78793709740096 9309.42405709922 9338.79332314335    266068.890809359
#> 19  1.29798016914413  19665.100555821 19906.5891344879    483031.291007155
#> 20  10.1610238075126 7008.82462920888 7281.45888488727    175825.229390152
#> 21  5.85531825041459 9239.76311966451 9273.95385413631    262270.841589943
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83547.3058287064 86.9651968592417 98.7968393737873    14.3425188768971
#> 2      43755.656471839 27.5851231616569 38.4434569716988     25.382013538513
#> 3     67898.1478763782  61.938288402274 68.1473688646412    17.2779449688287
#> 4     153049.642144844  237.36713640274 247.482313113657    14.4206180542041
#> 5     59020.3754147271 61.8408063459345 68.2818150566547    16.4004366421004
#> 6     133599.836302696 200.748541066068  211.66860371587    11.5993099860837
#> 7     375016.315335757 486.548998594085 532.687278308732    6.44334244612651
#> 8      122306.63057389 185.532938376316 215.498321466093    12.0501781348407
#> 9     227564.072539863 274.024579199679 304.729129366737    3.77295200033906
#> 10    502473.141404769  630.07554479293 710.686520029508   0.118066688686313
#> 11    172371.250217867 302.578981054049  348.14629884735    8.32461688668611
#> 12    273971.765645827 320.931540507082 344.816830326556    2.90470310268324
#> 13    505109.052247698 682.638805816759 761.847707807964    0.15439690818249
#> 14    181514.543633332 313.560207082681 371.400412333606    8.46893135366508
#> 15    270488.685509941 305.457181624681 331.797852971954     2.5992644627683
#> 16    532553.979827417 728.755751178443 815.643657774656  -0.145598024782004
#> 17    180736.330808315 313.122433818777 356.844992601743    7.25885950346358
#> 18    276992.880398033 339.091992391161 364.420826697691    4.12892437374985
#> 19    522753.515910047  720.16445040971 826.164988573813 -0.0820549200168237
#> 20    185378.627312359 320.230701170196 375.576993221513    8.66315017366472
#> 21     274036.67291722 325.768587015146 350.360870460031    4.41818662751504
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    16.637637360225 3378.80227937098 3457.42201357245    79431.1091816011
#> 2   36.1690163995112 1413.15629254327 1495.08310697957    40553.0668004664
#> 3   23.5332113984811 2185.15221253604 2204.37185947704    63706.4687834337
#> 4   15.5490855613305 7511.55658567393 7634.36950792477    146502.525678334
#> 5   26.4679329026592 2260.71641192346 2437.27779124081    53946.1659353924
#> 6   13.4465239863375 4795.98615005022 4840.31209302739    126298.486334068
#> 7   7.93549627475161 16748.0785363334 17319.2030996568    344946.351770574
#> 8   14.9041558958781 5099.38707462591  5261.0147448964    115741.956666297
#> 9   5.40757936974454  7717.4820574784 7810.09245085027    214722.000231899
#> 10 0.915480301893638 19908.4712704365 20203.7394731148    467225.079209413
#> 11   11.352584832133 6575.10166179853 6978.29058497148    160701.750521763
#> 12  4.31855796884677 9173.46486376311 9211.14918817575     256624.25446174
#> 13    1.346217077157 20224.1954899931 20656.8855793252    468883.203563868
#> 14  10.7291350462675 6894.57460796579 7047.97061349344    170809.883303882
#> 15  4.32141821720739 8982.66209599695 9042.68385311021    256820.758063999
#> 16 0.969225900361508 20494.8179360898 20897.5269215394    497405.032337595
#> 17  8.87157652987192 6801.28509005354 7057.44656297538    171775.462959729
#> 18  5.93289937422485   9306.612106095 9341.60527414757    265022.976912571
#> 19  1.41856575945916 19641.9793089274 19929.7103813815    479228.099261134
#> 20  10.2919059696935 6982.72134940989 7307.56216468627    174910.542355047
#> 21  5.98089285824076 9236.48953870445 9277.22743509637    261144.325824566
#>    ci95.integral.bt.up
#> 1     83906.9734969001
#> 2      44035.494404289
#> 3     68264.4110980918
#> 4     153621.720282694
#> 5     59463.7529420476
#> 6      134237.81833879
#> 7     377643.787880288
#> 8     122880.242857078
#> 9     228686.195362889
#> 10    505553.069169412
#> 11    173390.915239857
#> 12    275487.567593951
#> 13    508274.417666674
#> 14    182449.902302896
#> 15    271682.970432402
#> 16    535625.247083809
#> 17    181519.319261104
#> 18    278038.794294821
#> 19    526556.707656068
#> 20    186293.314347464
#> 21    275163.188682598