Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  91.541263952423  15.5674111459657
#> 2                 FALSE          0.75    33.1689918256  30.9936439369247
#> 3                 FALSE          0.75 64.8999991514221  20.0603373416786
#> 4                 FALSE          0.75 243.490836841663  15.0197356402084
#> 5                 FALSE          0.75 66.0310759893646  22.8894185216761
#> 6                 FALSE          0.75 205.594192493031  12.4628896561011
#> 7                 FALSE          0.75 508.967861628192  7.07375258950668
#> 8                 FALSE          0.75 200.524150891106  13.4197422847444
#> 9                 FALSE          0.75 288.810060304082  4.82169077015851
#> 10                FALSE          0.75 664.304067333901 0.546792738837704
#> 11                FALSE          0.75 323.682307517857  9.72075772240953
#> 12                FALSE          0.75 334.006060851474  3.88876666971633
#> 13                FALSE          0.75  730.87751181698 0.858762690503597
#> 14                FALSE          0.75 344.945494269112  9.71231397852846
#> 15                FALSE          0.75 317.581796609523  3.23669313526902
#> 16                FALSE          0.75 773.235340043217 0.515176210490319
#> 17                FALSE          0.75 332.604956861108  7.94843963288934
#> 18                FALSE          0.75 349.945624638372  4.98023068042339
#> 19                FALSE          0.75 782.437260018572 0.790054754666233
#> 20                FALSE          0.75 353.049070027427  9.52089613867713
#> 21                FALSE          0.75 336.027328382625  5.22603796080182
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3409.76796557255 2112.89332155815 81221.2317454969 1.87187905888292
#> 2  1454.89893858899 270.339158572969 42279.3387736143 1.72139968739926
#> 3  2196.42089928696 691.366025900554 66082.1279511637 1.52578067561292
#> 4  7563.28907785722 6044.12734193786  149362.41788725   1.522998179971
#> 5  2358.65611464878 1080.78101567582 56711.6808991554 3.05873308760815
#> 6  4820.30434784953 3220.67375431978 131292.618842822 2.69143040737383
#> 7   17066.273389488 15862.2756313773 363612.066963307 16.4416471683495
#> 8  5172.21727849375 3759.66425539105 117961.463494273 8.75322533090102
#> 9  7771.09364692974 6242.14131902493 222734.458318298 6.32568824159299
#> 10 20102.8319583364 18736.4830751718 483981.511550667  35.999772499351
#> 11 6809.86012969836 5639.94079414002 168495.257376975 13.6808783342085
#> 12 9187.04887795279 7624.14091770848 266018.061063159 5.34642478002468
#> 13 20445.2623543967 19335.2221793008 490128.929109617  21.989417587426
#> 14 6917.95228607797 5639.92872388702 173879.597101031 12.4344664560373
#> 15 9002.63589482786 7413.90532371594 265575.686172763 4.84074499976457
#> 16 20667.5831325527 19324.5269134308 515191.678920585 17.4031656521841
#> 17 6875.72105172195 6028.64392469759 174590.720703502 12.1065038874987
#> 18  9319.1336450335 7640.56373373742 270942.800118516  6.3280278482923
#> 19 19781.5168297225 18500.4211641294 500444.640273844 30.8935045486798
#> 20 7117.30334327843 5790.45503700225 179950.474259939 10.7715271043451
#> 21 9254.26293479315 7425.50831480295 269800.699914803 8.87739916714583
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.535763452444629 25.7887929591308 50.9610218980879 1671.24252097784
#> 2   2.02310449064484 27.2451249147959 74.9380821572922 1339.93250032312
#> 3   1.50716229233471 5.21584751583383 89.1736870432514 942.183444999887
#> 4  0.317149535263015 49.8439819905381 89.0318963897669 2709.42697329417
#> 5   2.78186428225188 42.9812873078213 148.983602968896 1250.14229780268
#> 6  0.330444537631172 8.79706757522109  69.884871047363 1534.07917587923
#> 7  0.609372564154061 78.0282850901579 131.694837530339 4620.21912452164
#> 8  0.764148407727231 45.4216074938548 120.780926279903 2458.15637941352
#> 9  0.369205722161617 16.1575097847155 69.5309994889742 3029.69567705984
#> 10 0.435350644001485 68.2141369989155 275.813278091192 7395.52781836502
#> 11 0.570007585056379 63.6977288202815 146.273883583306 2388.57197220579
#> 12 0.552151924782095 8.33069021157518 170.996553244621 5123.56368480144
#> 13 0.427304367072001 108.950286509648 304.483948059853 11311.5557339987
#> 14 0.418572110431711 123.907372415641  137.41136484186 7766.38339865912
#> 15 0.280190762931487  19.116149318519 44.8715354583054  3201.4608304411
#> 16 0.344569216728892  90.249358071166  284.79045626739 7428.54477332572
#> 17 0.516961418619659 83.8659260635732 165.399989428163 3191.54550395815
#> 18 0.438676250512374  10.102823676413 104.274597267332 4451.61503781881
#> 19  0.57143179581523 85.1539766637147 346.408347771265 12381.6629484407
#> 20 0.363530065103577 86.3223980228467 104.440656593796 4496.53985856603
#> 21   0.7553659559841 9.95664151019205 225.687313341036 2324.49633273511
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.4620229005606 94.6205050042854    14.6860802666943
#> 2             TRUE 30.3372893398282 36.0006943113717     27.665637049814
#> 3             TRUE 62.3900899400388 67.4099083628053     17.581055370788
#> 4             TRUE 240.985504835611 245.996168847716    14.4980246547007
#> 5             TRUE 60.9994600602491   71.06269191848    18.3132517773718
#> 6             TRUE 201.166789472901 210.021595513161    11.9193083916978
#> 7             TRUE 481.921352036257 536.014371220127    6.07133472147325
#> 8             TRUE 186.125095221774 214.923206560439    12.1627181540331
#> 9             TRUE 278.404303146661 299.215817461502    4.21434735720265
#> 10            TRUE 605.084441572468 723.523693095333  -0.169359070544739
#> 11            TRUE 301.177262658084  346.18735237763    8.78309524499179
#> 12            TRUE 325.211192088333 342.800929614614    2.98047675344978
#> 13            TRUE 694.704919885664 767.050103748295   0.155847006670154
#> 14            TRUE  324.49079694893 365.400191589293     9.0237628568683
#> 15            TRUE  309.61877108491 325.544822134136    2.77577933024672
#> 16            TRUE 744.607132545374  801.86354754106 -0.0516401510287089
#> 17            TRUE 312.689757966173 352.520155756044       7.09803809926
#> 18            TRUE 339.536018827931 360.355230448813    4.25860824833053
#> 19            TRUE 731.617445035993  833.25707500115   -0.14995054944982
#> 20            TRUE 335.329907940779 370.768232114074    8.92288918158174
#> 21            TRUE  321.42400675267  350.63065001258    3.98346096320797
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4487420252372 3367.34540115478 3452.19052999032    78472.0377984884
#> 2   34.3216508240355 1410.08070810415 1499.71716907383    40075.1498105828
#> 3   22.5396193125692 2187.84083012342 2205.00096845051    64532.2361841389
#> 4    15.541446625716 7481.29572748279 7645.28242823166    144905.410516181
#> 5   27.4655852659804 2287.95189702741 2429.36033227015      54655.19681927
#> 6   13.0064709205043 4805.83317168829 4834.77552401077    128769.058598501
#> 7   8.07617045754012 16937.9168605146 17194.6299184613    356011.806503469
#> 8   14.6767664154557 5097.49873416636 5246.93582282114    113917.796250138
#> 9   5.42903418311437 7744.51454333389  7797.6727505256    217750.608929535
#> 10  1.26294454822015 19990.6197029732 20215.0442136996    471815.868289457
#> 11  10.6584201998273 6705.07736578899 6914.64289360772    164566.056482696
#> 12  4.79705658598288 9173.34489255475 9200.75286335083     257589.79880166
#> 13  1.56167837433704 20266.0391330883 20624.4855757051    471521.419927189
#> 14  10.4008651001886 6714.12465845424  7121.7799137017    161103.896410237
#> 15  3.69760694029131  8971.1898291989 9034.08196045682    260309.283106687
#> 16  1.08199257200935 20519.1229385256 20816.0433265797    502971.722768465
#> 17  8.79884116651868 6737.76160334737 7013.68050009653    169340.628349491
#> 18  5.70185311251624  9302.5145000858  9335.7527899812    263619.893381304
#> 19  1.73006005878229 19641.4385381107 19921.5951213343    480076.804723659
#> 20  10.1189030957725 6975.30299853084 7259.30368802601    172553.666192598
#> 21  6.46861495839566 9237.88425950889 9270.64161007742    265976.903447454
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83970.4256925055 87.8723809970125 95.2101469078335   14.5173147791743
#> 2     44483.5277366458 29.7950484382974 36.5429352129025   27.0283591352609
#> 3     67632.0197181885 61.9094690272207 67.8905292756234   17.1062992487026
#> 4     153819.425258319  240.50576040892 246.475913274407   14.3981225510929
#> 5     58768.1649790408 60.0359591376526 72.0261928410765   17.4369645284624
#> 6     133816.179087143 200.318988894579 210.869396091484    11.815218362344
#> 7     371212.327423145 476.742233178227 541.193490078157   5.87938236376472
#> 8     122005.130738409  183.36782924254 217.680472539672    11.922011405599
#> 9     227718.307707062  276.41171135056 301.208409257604   4.09804755472174
#> 10    496147.154811878 593.744513235172 734.863621432629 -0.306494523405206
#> 11    172424.458271253 296.867785982808 350.496829052906   8.60354285569903
#> 12    274446.323324657 323.527068282625 344.485053420322   2.80654889714342
#> 13    508736.438292045 687.778253345625 773.976770288335 0.0212461310424741
#> 14    186655.297791826 320.573940015279 369.317048522945   8.89191264208231
#> 15    270842.089238838 308.093936409984 327.069656809061    2.6875192399233
#> 16    527411.635072706 739.125135364937 807.345544721498  -0.16017945429831
#> 17    179840.813057513 308.876209241611 356.333704480606   6.93519525239481
#> 18    278265.706855728 337.542690055719 362.348559221025   4.12042522941914
#> 19    520812.475824029 721.885991103159 842.988528933984 -0.329951565131617
#> 20     187347.28232728  331.93687690291 374.161263151943   8.80837721107412
#> 21    273624.496382153 318.627626015019 353.427030750231   3.74552068707298
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6175075127572 3359.22193137266 3460.31399977245    77945.5964043804
#> 2   34.9589287385886 1401.49849375599 1508.29938342199     39653.071072981
#> 3   23.0143754346547 2186.19783815593   2206.643960418    64235.4483989639
#> 4   15.6413487293239 7465.59487315577 7660.98328255868    144051.941019593
#> 5   28.3418725148898 2274.41279152545 2442.89943777211    54261.4019954621
#> 6   13.1105609498582  4803.0620954021 4837.54660029696    128285.823658099
#> 7   8.26812281524864 16913.3379507112 17219.2088282647    354556.437479245
#> 8   14.9174731638898  5083.1909278058 5261.24362918171    113143.476990623
#> 9   5.54533398559528  7739.4249277517 7802.76236610779    216796.254791261
#> 10  1.40008000108062 19969.1322498186 20236.5316668543    469486.277026672
#> 11    10.83797258912  6685.0125812106 6934.70767818611    163813.656311452
#> 12  4.97098444228924 9170.72072513811 9203.37703076748    255975.876240948
#> 13  1.69627924996472 20231.7197928378 20658.8049159556    467958.279870979
#> 14  10.5327153149746 6675.09383614331 7160.81073601262     158657.48563966
#> 15  3.78586703061473 8965.16824216356 9040.10354749216    259300.822945098
#> 16  1.19053187527895 20490.6943907332 20844.4718743721    500631.731164867
#> 17  8.96168401338387 6711.34383663735 7040.09826680655    168335.291515744
#> 18  5.84003613142764 9299.33211062773 9338.93517943927    262217.634644391
#> 19  1.91006107446408 19614.6150354616 19948.4186239834      476176.5808949
#> 20  10.2334150662801 6948.11144315365 7286.49524340321     171137.25613715
#> 21  6.70655523453065 9234.74791743318 9273.77795215313    265244.687102643
#>    ci95.integral.bt.up
#> 1     84496.8670866135
#> 2     44905.6064742476
#> 3     67928.8075033635
#> 4     154672.894754907
#> 5     59161.9598028487
#> 6     134299.414027545
#> 7      372667.69644737
#> 8     122779.449997924
#> 9     228672.661845336
#> 10    498476.746074663
#> 11    173176.858442498
#> 12    276060.245885369
#> 13    512299.578348254
#> 14    189101.708562403
#> 15    271850.549400427
#> 16    529751.626676304
#> 17     180846.14989126
#> 18    279667.965592641
#> 19    524712.699652788
#> 20    188763.692382729
#> 21    274356.712726964