Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1017819760521  15.5990962974326
#> 2                 FALSE          0.75 33.1105668608948  31.8609757859369
#> 3                 FALSE          0.75 64.6481596683896  20.0905637881118
#> 4                 FALSE          0.75 242.340924853542  14.9130543516192
#> 5                 FALSE          0.75  66.534497147189  22.1752100834661
#> 6                 FALSE          0.75 206.155155657261  12.3580449031657
#> 7                 FALSE          0.75 505.362097183793  7.03803514077985
#> 8                 FALSE          0.75 201.868251569075  13.3580373229021
#> 9                 FALSE          0.75 288.403811996294  4.70794316006336
#> 10                FALSE          0.75 674.733152297293 0.749428256660441
#> 11                FALSE          0.75 323.599136630805  9.82603050388896
#> 12                FALSE          0.75 334.235274918028  3.93930196802206
#> 13                FALSE          0.75  723.07531340662 0.794731541101144
#> 14                FALSE          0.75  339.27031628781  9.76758386846417
#> 15                FALSE          0.75 317.504561044118  3.31779702466285
#> 16                FALSE          0.75 785.275298475167 0.549967255839216
#> 17                FALSE          0.75 336.217858923878  8.06031805385671
#> 18                FALSE          0.75 348.731810514424   4.8266297139649
#> 19                FALSE          0.75 775.035246963735 0.638409964839689
#> 20                FALSE          0.75 348.725411165244  9.53062436413473
#> 21                FALSE          0.75 340.241226898771  5.22207265897331
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3413.61170601106 2116.26432594895 81623.3565552331  2.2381699844035
#> 2  1458.04858726823 259.932190273456 42209.5019752656 2.67885353348599
#> 3  2193.45853335004 687.711321063766 65763.0897426357  1.8122101126911
#> 4  7579.17313807451 6077.53365446842   149776.0671117 1.99294734675059
#> 5  2364.11270478247 1129.67105946689 57399.4254726459 2.95546131303812
#> 6  4819.80752192925 3240.42453347655 130482.041686215 3.24230714305418
#> 7  16974.7438472301 15740.6073484506  358704.98774488 14.8039872460694
#> 8  5164.82119726859 3760.51431425993 118340.906855754 9.63880959334337
#> 9   7757.8713396438 6247.25995441089 220693.813727058 7.32372134534747
#> 10 20085.6797634778 18621.2915863365 485151.832778018 24.2756997302787
#> 11 6780.85773206551 5581.60433582949 167900.981334217 14.0579028489902
#> 12  9187.2235398431 7605.31392272425 265568.752714164  6.8202042599351
#> 13 20461.4494282201 19356.9241389933 491449.527355001  25.271353556434
#> 14 6981.21615135114 5664.67005961208 177561.763327382 14.5216275021273
#> 15 9000.88312525787 7367.98440168637 266342.974446725 5.70108911105239
#> 16 20678.6035428131 19371.8289049343 513377.525363605 27.7342661352007
#> 17 6866.67942179621 5990.87168568735  174327.85372191  9.8748971603196
#> 18  9321.9021747284 7677.75012874885 271436.978840725  8.3878593671763
#> 19 19790.9613081105 18612.0897847967 502760.729956874 18.2143420896789
#> 20  7129.2415650316 5796.87740545136 179368.828771685 15.5767935995951
#> 21  9256.1253435743 7491.79343625606 269251.813860486 7.02201156272332
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.781472663662533 24.4317401034569 68.5702610840613 1759.72760439251
#> 2   2.22658706090483 25.9901150153099 64.8905536464841 1087.52862311607
#> 3   2.44798973837111 5.15236828215786 168.809711000115 1373.07141624681
#> 4  0.302470707522994 43.4003783041208 66.9224846593315 2218.18108446316
#> 5   2.59754750114245 25.0500182259345  126.11487789359 1121.22218639352
#> 6  0.546490902810522 10.3171739195273  108.81249018701 2228.98920746256
#> 7  0.503917362320058 191.719212927377 253.654383781453   9967.339192883
#> 8  0.580433183316304 58.8043190185581 99.5320180628312 3192.06885917916
#> 9  0.473407606895068  24.365799384056 78.9524464280874 4652.79789310995
#> 10 0.436443077322128 98.4138698558396 260.242612687832 8064.31811462801
#> 11 0.727842010793956 64.1358128266037 222.417548507647 2416.64230835115
#> 12 0.577266963412294 12.7461578542138 175.822584920147 5554.34635870346
#> 13 0.363367459773767 94.3173179378875 217.296897847934 9375.73665008223
#> 14 0.762417340193517  34.810551967395 212.671884356399 2886.21670630052
#> 15 0.663951616761367  15.875949695374 153.553122228917 2844.63379604962
#> 16 0.377441425177635 69.7986929798509 238.565588671602 9446.53718966017
#> 17 0.511835459235883 100.014586512984 151.296528079954 4298.51470109593
#> 18  0.45090736797891 9.12503581401463 74.5091312197036 3315.70465282011
#> 19 0.259090494356101 77.7986305788612  189.83657102747 8396.67581267655
#> 20  0.48646359714255 92.4581184687522  152.80665213199 3957.42087187627
#> 21 0.516331004023978 7.94784281073107 154.493070864307 3496.56427384562
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.4199923517083 95.7835716003959    14.3135737657077
#> 2             TRUE 28.7038527983103 37.5172809234792    28.1982400707485
#> 3             TRUE 61.6670740330128 67.6292453037665    16.0636206684913
#> 4             TRUE 239.062526468137 245.619323238947    14.4154900377438
#> 5             TRUE 61.6727632872413 71.3962310071367    17.9022444440868
#> 6             TRUE 200.821560406937 211.488750907585    11.4590673680424
#> 7             TRUE 481.009538164008 529.714656203577    6.20909107976336
#> 8             TRUE 186.012409788025 217.724093350124    12.4032247363468
#> 9             TRUE 276.356290383197 300.451333609391    3.92918764672097
#> 10            TRUE 634.799626240985 714.666678353602  0.0314793944655414
#> 11            TRUE 300.473886444216 346.724386817394    8.62873039613291
#> 12            TRUE 323.016038910435 345.454510925621    2.98969781320883
#> 13            TRUE 681.503936806286 764.646690006954   0.196992069773297
#> 14            TRUE  315.38223904681 363.158393528809    8.51340734384584
#> 15            TRUE 308.126269456437   326.8828526318     2.2255966150904
#> 16            TRUE 739.652430682762 830.898166267572 -0.0709238885779939
#> 17            TRUE 319.973653095152 352.462064752604    7.21834872341368
#> 18            TRUE 334.933781855419 362.529839173429    4.08488709363959
#> 19            TRUE 745.072654226214 804.997839701257   0.212206101623903
#> 20            TRUE  323.10158569391 374.349236636578    8.73039174683524
#> 21            TRUE 328.690017878091  351.79243591945    4.37270815735387
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.8846188291574 3373.42149354087 3453.80191848125    78728.6046460074
#> 2   35.5237115011254 1415.29484806804 1500.80232646841    40420.5173902397
#> 3   24.1175069077323 2184.98288752589 2201.93417917419    63504.3872629097
#> 4   15.4106186654945 7507.77951576423 7650.56676038479    146127.159227758
#> 5   26.4481757228454 2322.90542480081 2405.31998476413    55555.0149760285
#> 6    13.257022438289 4802.83577083163 4836.77927302687    126815.354439939
#> 7   7.86697920179635 16659.3657419645 17290.1219524956    342308.714772587
#> 8   14.3128499094574 5068.08809248307 5261.55430205412    113089.953582405
#> 9   5.48669867340575 7717.78959965703 7797.95307963057    213039.961192892
#> 10  1.46737711885534  19923.788947565 20247.5705793907    471886.029479455
#> 11   11.023330611645 6675.35431996575 6886.36114416528    163925.604736979
#> 12  4.88890612283528 9166.25611017292 9208.19096951328    256431.852954097
#> 13  1.39247101242899 20306.2974402123 20616.6014162279    476026.440565616
#> 14  11.0217603930825 6923.95279336477  7038.4795093375    172813.936845517
#> 15   4.4099974342353 8974.76718800898 9026.99906250677    261663.551852224
#> 16  1.17085840025643 20563.7846928612 20793.4223927649    497837.971686614
#> 17  8.90228738429974 6702.15542698236 7031.20341661007    167256.797038608
#> 18  5.56837233429021 9306.89149081434 9336.91285864245    265982.644686836
#> 19  1.06461382805548 19662.9825608083 19918.9400554128    488948.198245021
#> 20  10.3308569814342  6977.1479601505 7281.33516991269    172858.871437448
#> 21  6.07143716059276 9243.05114215065 9269.19954499795     263499.96563001
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84518.1084644588 87.7149688066212  96.488595145483    14.067409876654
#> 2     43998.4865602916 27.8600139352622 38.3611197865273   27.4968651465635
#> 3     68021.7922223617 61.0962278475151 68.2000914892642   15.2925039009044
#> 4     153424.974995642 238.434748053911 246.247101653173   14.3202117648741
#> 5     59243.8359692632 60.7417929736343 72.3272013207437   17.0840169812269
#> 6     134148.728932491 199.800233656875 212.510077657647    11.286922733657
#> 7     375101.260717172 476.346282181496 534.377912186089   6.05035711063254
#> 8     123591.860129104 182.976184766122 220.760318372028   12.2203882836021
#> 9     228347.666261224 274.049318159413 302.758305833175   3.78006425054903
#> 10    498417.636076581 627.152780825947  722.31352376864 -0.106000174890929
#> 11    171876.357931454 296.045647046784 351.152626214826   8.39946016273281
#> 12    274705.652474231 320.867674568555 347.602875267501   2.80785871973396
#> 13    506872.614144387 673.543460436009 772.607166377231 0.0825313199445603
#> 14    182309.589809246  310.80792638364  367.73270619198   8.27324588168488
#> 15    271022.397041227 306.330426386456 328.678695701781   2.01645185581057
#> 16    528917.079040596 730.916136850173  839.63446010016 -0.189817937508949
#> 17    181398.910405213 316.863060489652 355.572657358104   7.05712055375438
#> 18    276891.312994615 332.291606154758 365.172014874089   3.94285127272624
#> 19    516573.261668727 739.335136467965 810.735357459506  0.130592595901731
#> 20    185878.786105921 318.194895710037  379.25592662045   8.57715571373533
#> 21    275003.662090962 326.478084235833 354.004369561708   4.21006389108632
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.1307827182111 3365.72549540829 3461.49791661384    78174.2904506237
#> 2   36.2250864253104 1407.10796183822 1508.98921269824    40077.9458739581
#> 3   24.8886236753192 2183.35989151701 2203.55717518307    63071.8697667919
#> 4   15.5058969383642 7494.10839659844 7664.23787955059    145428.432186152
#> 5   27.2664031857053 2315.01466905964  2413.2107405053    55201.8299873146
#> 6   13.4291670726743 4799.58586104698 4840.02918281152    126113.222839588
#> 7   8.02571317092717 16598.9741898924 17350.5135045677    339169.002926829
#> 8   14.4956863622021 5049.56473199222 5280.07766254497    112084.451891763
#> 9   5.63582206957769 7710.11437285105 7805.62830643655    211574.329856563
#> 10  1.60485668821181 19892.7885785604 20278.5709483953    469345.769273347
#> 11  11.2526008450451 6655.15153892537 6906.56392520566    163164.362409848
#> 12  5.07074521631015 9162.24107044884 9212.20600923736    254682.233851105
#> 13  1.50693176225773 20276.5874850619 20646.3113713784     473073.08352084
#> 14  11.2619218552435 6912.98746949504 7049.44483320723    171904.778583033
#> 15  4.61914219351513 8969.76626385494 9031.99998666081    260767.492206468
#> 16  1.28975244918738 20541.7981045726 20815.4089810536    494862.312471871
#> 17  9.06351555395904 6670.65083223077 7062.70801136166    165902.764907762
#> 18  5.71040815520356 9304.01710453293 9339.78724492386    264938.197721198
#> 19  1.14622733377765  19638.475992176 19943.4466240451    486303.245364028
#> 20  10.4840930145341 6948.02365283284 7310.45947723035    171612.283862807
#> 21  6.23408142686031 9240.54757166527 9271.70311548333    262398.547883748
#>    ci95.integral.bt.up
#> 1     85072.4226598424
#> 2     44341.0580765732
#> 3     68454.3097184794
#> 4     154123.702037248
#> 5     59597.0209579772
#> 6     134850.860532841
#> 7      378240.97256293
#> 8     124597.361819745
#> 9     229813.297597554
#> 10    500957.896282689
#> 11    172637.600258585
#> 12    276455.271577223
#> 13    509825.971189163
#> 14    183218.748071731
#> 15    271918.456686983
#> 16    531892.738255339
#> 17    182752.942536058
#> 18    277935.759960253
#> 19     519218.21454972
#> 20    187125.373680562
#> 21    276105.079837223