Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.3912911496101  15.5259667957667
#> 2                 FALSE          0.75 32.6183875489353   31.811927086335
#> 3                 FALSE          0.75 64.3168839304158  20.0924359690592
#> 4                 FALSE          0.75 242.188637168296  15.0864978608263
#> 5                 FALSE          0.75 65.6138390140064  22.6361935379202
#> 6                 FALSE          0.75 205.132978308476  12.5585598528028
#> 7                 FALSE          0.75 508.137338698955  7.13934020994497
#> 8                 FALSE          0.75 198.637349587391  12.9809228913761
#> 9                 FALSE          0.75 287.442940351745   4.6240720653514
#> 10                FALSE          0.75 671.613786156952 0.675674054645376
#> 11                FALSE          0.75 318.914184627881  9.79908993059875
#> 12                FALSE          0.75 333.567923832877  3.74413653070331
#> 13                FALSE          0.75 733.861953223437 0.659443626245528
#> 14                FALSE          0.75 347.266265554602  10.0278007659403
#> 15                FALSE          0.75 318.450010466757  3.50409372147679
#> 16                FALSE          0.75  780.86538846818 0.492923041939212
#> 17                FALSE          0.75 336.924846419111  7.95278252265891
#> 18                FALSE          0.75  351.60106547101  5.05240866299083
#> 19                FALSE          0.75 778.866549967339 0.733808731418825
#> 20                FALSE          0.75 350.212028824889  9.44276776747575
#> 21                FALSE          0.75 340.525107064846  5.17424293024327
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.81886150228 2117.54457601477 81922.3446973786 2.54487695101891
#> 2  1471.96570897373 286.955515575649 42600.0793572945  2.7676092385599
#> 3  2197.60186417771 703.115080138931 65884.4908753623 1.28907297505272
#> 4  7571.64247775882 6024.54342480862 149396.973493835  2.6347038382125
#> 5  2370.10245440949 1127.36952335563 57561.2587817961  2.2665694494263
#> 6  4817.02118414746 3196.17927919259 129958.659000367 3.00273634146253
#> 7  17052.5792282964 15806.9093443458 363155.172504538 14.3775460241961
#> 8  5175.64545695804 3826.53421988446 118581.683218722 7.63399200176667
#> 9  7758.46814443904  6274.0003674183 221370.824886601 6.10594297518738
#> 10 20075.8482990287 18630.1503406168 484708.617377617 33.8291573780505
#> 11 6806.68022195421  5566.4708405512 168520.378427088 12.6305218943649
#> 12 9188.83475182884 7663.05202166297 266059.278913162 4.93735656097107
#> 13 20489.7630619234 19551.0694821411 495162.089828563 19.9432020325914
#> 14 6983.72035748768 5613.33957343321 177209.139871229 15.8033179721888
#> 15 9001.42176235371 7346.92813290859 264243.004624196  4.9162994516159
#> 16 20672.2651134896  19375.320008908 513931.375259253 15.9891864741982
#> 17 6898.77493673054 6061.71842453183  175600.05025027 11.8188841498176
#> 18  9320.8978825243 7637.38841158148 270049.683968171 7.51858602059325
#> 19 19790.0975970479 18550.4861472085 504950.867965887 18.0801217180428
#> 20 7144.12841333983 5836.68817592319 180994.407250625 13.6611559093792
#> 21  9257.8174699935 7524.93773381928 268365.702165263 6.56298167459445
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.435344709677606 20.3613855643857 41.6011023283214 1383.10366120469
#> 2   2.34234271655773  22.932497773096 77.1041933836634 820.961851475274
#> 3   2.60914078889411 4.73646345341094 163.521577332348 1364.20415593667
#> 4  0.302086414992432 38.1217506352497 80.5879472193889 2162.32325720148
#> 5   2.54251257241061 25.7707872505329 149.356044684168 1004.00730594952
#> 6   0.32340981967968 11.8564334491932 61.4709158067428  2669.6930392913
#> 7  0.639442962382684 131.393474829914 162.212656097246 6758.70138457761
#> 8  0.709274738467981 40.0056313419555 112.028841896088 2432.11513309366
#> 9  0.240228018317745 18.3088846496676  52.126082707463 5147.85561312321
#> 10 0.643189340012181 67.8276779271593 368.961573850451  7207.0168866736
#> 11 0.740710876983409 72.6302541028079 204.990701864553 2354.49022633342
#> 12 0.441541959499305 14.6824285220007 116.755616424569 3574.41051682251
#> 13 0.272499742784395 41.6642020674686 155.657558147772 4086.94738550112
#> 14 0.837443386048397 43.2681482897536 230.027914363487 2015.71265752693
#> 15  0.65628527532055 16.4987953867346 176.953979267151 3747.32045793626
#> 16 0.257559411878141 60.5614273818171 202.360163231122  10154.177668538
#> 17 0.569542454618371 78.6604832228853 131.538436853877 2574.94693031639
#> 18 0.442274618327198 7.66413836423679 111.523736794982 3572.09779657706
#> 19 0.331045216789644 78.8048469022618 254.455203926773 8499.65755761518
#> 20 0.443294713134239  77.328920392856 106.189558560288 2601.50909971627
#> 21 0.423850593929163 7.63326177378976 98.7024105335732 2599.41827224367
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.2049685651839 95.5776137340362   14.8098247483471
#> 2             TRUE 28.0656703515042 37.1711047463663   27.9587733175975
#> 3             TRUE  62.196358886454 66.4374089743775   15.8003993713283
#> 4             TRUE 237.854549354437 246.522724982156   14.5895657081638
#> 5             TRUE 61.8853322697001 69.3423457583127   18.4537603563048
#> 6             TRUE  200.19347702677 210.072479590182   12.0265506994298
#> 7             TRUE 484.486275489153 531.788401908758   6.08745653682546
#> 8             TRUE 186.079432744485 211.195266430298   11.8141659465963
#> 9             TRUE 277.398664157562 297.487216545929   4.22889697521871
#> 10            TRUE 615.964822270059 727.262750043845 -0.382372409674661
#> 11            TRUE  298.13697611165 339.691393144111   8.58062053796104
#> 12            TRUE  325.44597229008 341.689875375675   3.01780000732695
#> 13            TRUE 701.055385879824  766.66852056705  0.211181549365199
#> 14            TRUE 321.269807490351 373.262723618852   8.65020639589067
#> 15            TRUE 310.362697868849 326.537323064666   2.42450444357448
#> 16            TRUE 754.563176718124 807.167600218236 0.0692378093996699
#> 17            TRUE 317.482781992661 356.366910845561   7.01588518481169
#> 18            TRUE 339.232991467134 363.969139474886   4.32486691584259
#> 19            TRUE 749.124749741159  808.60835019352   0.18923934979986
#> 20            TRUE  327.73942735396 372.684630295818   8.71354796436992
#> 21            TRUE 329.729002210139 351.321211919554   4.47700870322979
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2421088431864 3379.32438224886 3446.31334075569    79647.1391746969
#> 2   35.6650808550724 1434.24175013699 1509.68966781047    41249.5971116177
#> 3     24.38447256679 2189.81038179685 2205.39334655858    63640.3750388465
#> 4   15.5834300134889 7508.93219796383  7634.3527575538    145839.951735739
#> 5   26.8186267195357 2327.70950938237 2412.49539943662    55909.6667635092
#> 6   13.0905690061759 4797.51735112354 4836.52501717139    125567.013950732
#> 7   8.19122388306449 16836.4369622012 17268.7214943916    352037.108726908
#> 8   14.1476798361559 5109.83619340052 5241.45472051555    114580.853824783
#> 9   5.01924715548409 7728.35002919033 7788.58625968774    212902.602403013
#> 10  1.73372051896541 19964.2717688385 20187.4248292189    472853.074599039
#> 11  11.0175593232365 6687.20345395509 6926.15698995333     164647.24200477
#> 12  4.47047305407967 9164.68215691015 9212.98734674753    260179.373612989
#> 13  1.10770570312586 20421.2254495225 20558.3006743244    488439.061379413
#> 14  11.4053951359899 6912.54425355103 7054.89646142432    173893.292549597
#> 15  4.58368299937909 8974.28124394253 9028.56228076489    258078.662470891
#> 16 0.916608274478753 20572.6415654465 20771.8886615327    497227.752994508
#> 17  8.88967986050613 6769.37844182889 7028.17143163219      171364.2625499
#> 18  5.77995041013907 9308.29037491513 9333.50539013347    264173.583092802
#> 19  1.27837811303779 19660.4636238936 19919.7315702021     490968.93128361
#> 20  10.1719875705816 7016.92233929359 7271.33448738608    176714.924781592
#> 21  5.87147715725674 9245.26075437562 9270.37418561139    264089.659107422
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84197.5502200603  86.403332325613 96.3792499736071    14.6726911647986
#> 2     43950.5616029713 27.1938734413579 38.0429016565127    27.2209353618818
#> 3     68128.6067118782 61.7903008993124 66.8434669615191    14.9785200228267
#> 4     152953.995251932   237.0246176454 247.352656691193    14.4944084874412
#> 5     59212.8508000831 61.1713628931309  70.056315134882    17.6528688959954
#> 6     134350.304050001 199.247615079209 211.018341537743    11.9246766062307
#> 7     374273.236282168 479.957348491531  536.31732890638    5.88603200367491
#> 8     122582.512612661 183.674725263929 213.599973910854    11.5907444039789
#> 9     229839.047370188 275.475292120378 299.410588583113    4.15322514944862
#> 10    496564.160156195 605.308637695973 737.918934617931  -0.584977051778498
#> 11    172393.514849407 294.158361714925 343.670007540836    8.34729661171126
#> 12    271939.184213335 323.890704973374  343.24514269238    2.87871429008467
#> 13    501885.118277712 694.773277239558 772.950629207316   0.125344130388114
#> 14    180524.987192861 316.291762329112 378.240768780092    8.38641172928543
#> 15    270407.346777501  308.81406354159 328.085957391925    2.21777458184851
#> 16    530634.997523998 749.526582978752 812.204193957608 -0.0118934053419444
#> 17    179835.837950641 313.759833485468 360.089859352753     6.8364793116069
#> 18     275925.78484354 336.864636870647 366.337494071373    4.18555041106952
#> 19    518932.804648164 743.429511399975 814.303588534703  0.0849601065111224
#> 20    185273.889719658 323.436163242506 376.987894407272    8.57391012973264
#> 21    272641.745223104 327.661662982641 353.388551147052    4.34349576614211
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.3792424267348 3372.91054579608 3452.72717720847    79211.4615214174
#> 2   36.4029188107881 1427.01801333846   1516.913404609     40990.994128403
#> 3   25.2063519152916 2188.31839580903  2206.8853325464    63210.6507297265
#> 4   15.6785872342115 7496.92384651373 7646.36110900391     145158.81990972
#> 5    27.619518179845 2319.59171139845 2420.61319742054    55593.4044621351
#> 6    13.192443099375 4793.78257458705 4840.25979370788    124726.060643356
#> 7   8.39264841621503 16795.0480176298 17310.1104389631    349908.117790766
#> 8   14.3711013787733  5097.2344195278 5254.05649438827    113814.737557858
#> 9   5.09491898125418 7722.58273052569 7794.35355835238    211281.027884879
#> 10  1.93632516106925 19942.9060502915 20208.7905477659    470582.864279737
#> 11  11.2508832494862 6664.32492391271 6949.03551999571    163905.577583475
#> 12  4.60955877132195 9160.05719192572 9217.61231173196    259053.434300189
#> 13  1.19354312210294 20408.1012258712 20571.4248979757    487151.672952981
#> 14  11.6691898025951 6898.91478683976 7068.52592813559    173258.343062476
#> 15  4.79041286110506 8969.08412339571 9033.75940131171    256898.256526641
#> 16 0.997739489220367 20553.5647158213  20790.965511158    494029.187028918
#> 17  9.06908573371092 6744.60038961368 7052.94948384739     170553.15426685
#> 18  5.91926691491214  9305.8761713304 9335.91959371821     263048.37228688
#> 19  1.38265735632653 19635.6400971194 19944.5550969763    488291.539152961
#> 20  10.3116254052189 6992.56372936984 7295.69309730983    175895.449415181
#> 21  6.00499009434443 9242.85627691688 9272.77866307013    263270.842351665
#>    ci95.integral.bt.up
#> 1     84633.2278733398
#> 2     44209.1645861861
#> 3     68558.3310209982
#> 4      153635.12707795
#> 5     59529.1131014572
#> 6     135191.257357378
#> 7      376402.22721831
#> 8     123348.628879586
#> 9     231460.621888322
#> 10    498834.370475497
#> 11    173135.179270702
#> 12    273065.123526134
#> 13    503172.506704145
#> 14    181159.936679982
#> 15    271587.752721751
#> 16    533833.563489587
#> 17     180646.94623369
#> 18    277050.995649462
#> 19    521610.196778813
#> 20    186093.365086069
#> 21     273460.56197886