Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9442226239269  15.5586206952682
#> 2                 FALSE          0.75 33.5736112784189   31.061130264158
#> 3                 FALSE          0.75 65.1285563835813  19.7644964782219
#> 4                 FALSE          0.75 242.896269857631  14.9813098959264
#> 5                 FALSE          0.75 65.6945464092259  22.9503727503556
#> 6                 FALSE          0.75 205.886394407636  12.4766021440908
#> 7                 FALSE          0.75 511.036937302586  7.11553601315027
#> 8                 FALSE          0.75 204.064748884271   13.293083714619
#> 9                 FALSE          0.75 291.456340185224   4.6993233808633
#> 10                FALSE          0.75 670.095636676117 0.599296054964926
#> 11                FALSE          0.75 318.954657064053  9.96396451297147
#> 12                FALSE          0.75 335.465242362114  3.68901816362429
#> 13                FALSE          0.75 739.461845335168 0.901416971520294
#> 14                FALSE          0.75 343.177372757823  9.72405260819597
#> 15                FALSE          0.75 317.595371293946  3.30758125310782
#> 16                FALSE          0.75 775.509672436996 0.627725761970775
#> 17                FALSE          0.75 336.561304443916  7.91622529469045
#> 18                FALSE          0.75 352.144299456254   5.2292275453073
#> 19                FALSE          0.75 783.001082089361 0.627214728765436
#> 20                FALSE          0.75 353.127600577756  9.64804485780322
#> 21                FALSE          0.75 337.634951413811  5.22853616480576
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.45067994722 2113.92542655021 81275.8435708734 2.38047388742934
#> 2   1459.8778730075 279.762874940409 42324.7460068044 1.89033417511123
#> 3  2195.05218631103 698.039496442532 65951.3751560476 1.62322278853395
#> 4   7543.8964246174 6033.75668833172 148265.208533322 2.65653898640785
#> 5  2363.21514605773 1086.33086226271 56916.5635479873 2.73429193431717
#> 6  4822.23498844052 3218.60167465841 131141.828217901 2.66039764126141
#> 7  17004.0591805615 15815.1509261998 359649.516507298 11.5511949112586
#> 8  5159.60054271166 3772.07573075463 117771.787691803 7.12650411748305
#> 9  7775.87467284348 6300.70104621416 222036.939238499 6.99241198744474
#> 10 20049.3629522294 18670.5631626602  486069.30959488 32.1527409788141
#> 11 6801.80241616069 5535.74075721347 168413.811545418 11.5045329126972
#> 12 9187.06923808045 7697.93230667704 265759.620129119 6.04600717898677
#> 13  20425.198238011 19329.7002371048 489176.585296758  21.313129624367
#> 14 6972.19492876485 5665.08180615925  177167.46339439 13.4997862210928
#> 15 9002.45046603196 7391.28556253793 262946.635067517 4.59951396794297
#> 16 20686.3634245721  19282.711280415 515141.643427168 23.9652687296691
#> 17  6910.5930956845   6076.133935867 176577.841467763 13.5473554067377
#> 18 9324.54580300308 7582.42115411157 269643.855008143 6.81802739538517
#> 19 19789.4755048079 18664.8564640032 499885.732951358 20.1964655119056
#> 20 7096.18000668467 5761.26847238811 179487.199166005 11.3992646454291
#> 21  9259.7507388346 7479.53626622933 269225.672330117 6.63702507846873
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.606251947178335 34.2340690888071 66.8968520735826 2063.80911826134
#> 2   2.14483384983312   19.62572939033 63.8775050619328 851.365451345778
#> 3   2.26262337366725 3.77770138770296 152.744583226229 1203.34332901159
#> 4  0.326048063856906 96.0593701780428 118.249432052767 5281.10168378973
#> 5   2.78790342572328 47.1963978785153 155.617215940144 1713.28088856941
#> 6  0.462763298247951 13.0658768642262 97.7152100985358 1806.71032569198
#> 7  0.463728910267396 151.104640615252 223.487754502058 8125.44352447666
#> 8  0.697406484888045 81.2221044841606 144.417598480321 3983.74220756616
#> 9  0.202677222980708 19.9937015725818  49.845302222191 3139.52861223261
#> 10 0.513403525367416 95.2793276090714 311.423119008696 6052.34033852751
#> 11  0.63893593302424 64.9846364389886 228.972040312472 2270.90585825569
#> 12 0.313950351728343 11.1293757744026 65.9012642098476 5169.68860606855
#> 13 0.380040503247962 91.4175375050353 259.189221180342  8847.7850345589
#> 14  0.58652367593217  52.030359832525 161.889307706724 3187.38113982872
#> 15 0.434290773451278 16.4813305187429 106.546566161964 4090.16218461454
#> 16 0.430185190157157 69.6744223283215 252.876522455797 6332.02587133518
#> 17  0.56114395012022 81.8192858835562 198.782737194055 2687.10054661525
#> 18 0.742855690678612 7.22984498611256 212.769396076082 4780.90504701465
#> 19 0.245847074712927  72.351114539351 205.202166054571 12785.6928716762
#> 20 0.344558513552614 94.5363985330742    158.608628096  4046.4432529265
#> 21 0.440216629043421 10.6751183252021 147.297654838957 3944.98157227175
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.0283430791056 95.8601021687481    14.5613362421598
#> 2             TRUE 30.4640115603609 36.6832109964769    27.5328785811826
#> 3             TRUE  62.458354896443 67.7987578707197    16.0424810285392
#> 4             TRUE  238.52626322499 247.266276490272    14.4449608308818
#> 5             TRUE 61.1966361772742 70.1924566411777    18.3642716150408
#> 6             TRUE 201.510040287761 210.262748527511    11.7153565184729
#> 7             TRUE 492.035221673566 530.038652931606     6.3527019557604
#> 8             TRUE 192.341649611012 215.787848157531    12.1458500469781
#> 9             TRUE 279.953822465877  302.95885790457    4.36591934906003
#> 10            TRUE 617.204377765968 722.986895586266  -0.245252744264473
#> 11            TRUE 300.029700422667  337.87961370544     8.9129149031466
#> 12            TRUE 325.519560552681 345.410924171547    3.17256983503117
#> 13            TRUE 704.401747103085 774.521943567252   0.276250343677397
#> 14            TRUE 320.970224424126 365.384521091521    8.75922116128755
#> 15            TRUE  310.02917081668 325.161571771212    2.59317293078047
#> 16            TRUE 736.086805376691 814.932539497302 -0.0799288758377491
#> 17            TRUE 314.275904799832    358.846704088    6.99314349674269
#> 18            TRUE 340.928644390846 363.359954521663    4.00722993414099
#> 19            TRUE 749.777896322277 816.224267856446   0.222796290862671
#> 20            TRUE 334.375810236025 371.879390919487    9.08124610300917
#> 21            TRUE  326.71704515973 348.552857667892    4.50437981002933
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5559051483765 3351.13563629613  3463.7657235983    77880.8775713335
#> 2   34.5893819471335 1427.59354816041 1492.16219785459    40924.2498393406
#> 3   23.4865119279045 2188.83786752826  2201.2665050938    63971.8753798235
#> 4    15.517658960971 7385.87876067452 7701.91408856028    139577.796263488
#> 5   27.5364738856704 2285.57707154758 2440.85322056789    54098.2164862906
#> 6   13.2378477697087 4800.74162099886 4843.72835588217    128169.789732138
#> 7   7.87837007054014 16755.4920467494 17252.6263143736    346283.161909534
#> 8   14.4403173822598 5025.99018083522  5293.2109045881    111218.531760357
#> 9   5.03272741266657 7742.98503375658 7808.76431193037    216872.414671376
#> 10  1.44384485419432 19892.6284583124 20206.0974461463    476113.209738002
#> 11  11.0150141227963 6694.90268921855 6908.70214310283    164678.171408587
#> 12  4.20546649221741 9168.76141493156 9205.37706122934    257255.482372136
#> 13  1.52658359936319 20274.8163888153 20575.5800872068    474621.978914909
#> 14  10.6888840551044 6886.60498684034 7057.78487068935    171924.221419372
#> 15  4.02198957543517 8975.33867732863 9029.56225473529    256218.318273826
#> 16   1.3353803997793  20571.748999842 20800.9778493022    504725.460868822
#> 17  8.83930709263822 6776.00037040605 7045.18582096295    172157.561068581
#> 18  6.45122515647362 9312.65270800093 9336.43889800524    261779.266205804
#> 19   1.0316331666682 19670.4579213907 19908.4930882251    478853.268177451
#> 20  10.2148436125973 6940.66763109776 7251.69238227158    172830.800014941
#> 21  5.95269251958219 9242.19016918964 9277.31130847956     262736.17764373
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84670.8095704133 87.2784938045654 96.6099514432884   14.3703668787986
#> 2     43725.2421742682 29.8685562952009 37.2786662616369   26.8572559184851
#> 3     67930.8749322716 61.9470397180548 68.3100730491079   15.3297546658341
#> 4     156952.620803157 237.689453444272 248.103086270991   14.3422556907669
#> 5      59734.910609684 60.3353342179642 71.0537586004876    17.486082035938
#> 6     134113.866703664 200.672015030763 211.100773784508   11.5695860795248
#> 7     373015.871105062 488.396595276519 533.677279328653   6.20662734902617
#> 8      124325.04362325 190.096800814005 218.032696954538   11.9261670042384
#> 9     227201.463805622 277.751212689832 305.161467680615   4.30207602382111
#> 10    496025.409451757 607.076264357642 733.115008994593 -0.406974854755208
#> 11    172149.451682249 296.405772555167  341.50354157294   8.71165008424396
#> 12    274263.757886101   323.6150682913 347.315416432928   3.07367547423674
#> 13    503731.191678608 697.688111271409 781.235579398928  0.156537585154289
#> 14    182410.705369408 316.717791764481 369.636953751165   8.57446620336892
#> 15    269674.951861208 308.580323916778 326.610418671114   2.45637133714332
#> 16    525557.825985515 728.537745726845 822.481599147148 -0.215437210737254
#> 17    180998.121866945  310.00848784671 363.114121041122   6.81638315245482
#> 18    277508.443810482 338.780965761299 365.507633151209   3.77323039157722
#> 19    520918.197725265 743.416009686026 822.586154492696  0.145354462328099
#> 20    186143.598317069 330.785041872715 375.470159282797    8.9727101712401
#> 21    275715.167016504 324.626382260012 350.643520567609   4.36571157188066
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7468745117377 3340.35190453315 3474.54945536128    77230.7776990812
#> 2   35.2650046098309 1421.41144340245 1498.34430261255    40656.0697221667
#> 3   24.1992382906097 2187.64789159113 2202.45648103093    63592.8222311849
#> 4    15.620364101086 7355.62005906843 7732.17279016636    137914.249233095
#> 5   28.4146634647733 2270.71020621584 2455.72008589962    53558.5330063913
#> 6   13.3836182086568 4796.62586978663  4847.8441070944    127600.675979545
#> 7   8.02444467727437 16707.8940849556 17300.2242761674    343723.647199324
#> 8   14.6600004249995  5000.4052179227 5318.79586750061    109963.652964974
#> 9   5.09657073790549 7736.68701776122 7815.06232792574    215883.463158523
#> 10  1.60556696468506 19862.6154701156 20236.1104343431    474206.722531366
#> 11   11.216278941699 6674.43252874027 6929.17230358111    163962.836063237
#> 12  4.30436085301184 9165.25566156262 9208.88281459828    255627.030461224
#> 13   1.6462963578863 20246.0198645012 20604.3766115209    471834.926629023
#> 14   10.873639013023  6870.2154234931  7074.1744340366    170920.196360326
#> 15  4.15879116907233 8970.14705821522  9034.7538738487    254929.917185673
#> 16   1.4708887346788 20549.8015568086 20822.9252923356    502730.872719351
#> 17  9.01606743692609 6750.22729535273 7070.95889601627    171311.124396397
#> 18  6.68522469903738  9310.3753068303 9338.71629917587    260273.281115995
#> 19  1.10907499520277 19647.6673203108  19931.283689305    474825.774922873
#> 20  10.3233795443663 6910.88866555985  7281.4713478095    171556.170390269
#> 21  6.09136075773087  9238.8275069172 9280.67397075199    261493.508448464
#>    ci95.integral.bt.up
#> 1     85320.9094426656
#> 2     43993.4222914422
#> 3     68309.9280809103
#> 4      158616.16783355
#> 5     60274.5940895833
#> 6     134682.980456257
#> 7     375575.385815272
#> 8     125579.922418633
#> 9     228190.415318475
#> 10    497931.896658393
#> 11    172864.787027599
#> 12    275892.209797013
#> 13    506518.243964494
#> 14    183414.730428455
#> 15    270963.352949362
#> 16    527552.414134985
#> 17    181844.558539129
#> 18    279014.428900292
#> 19    524945.690979843
#> 20    187418.227941741
#> 21     276957.83621177