Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75   91.63730522703  15.6844072757192
#> 2                 FALSE          0.75 32.3483452765817   31.502556693397
#> 3                 FALSE          0.75 64.8243382337562  20.3594434644698
#> 4                 FALSE          0.75 243.350409808157  15.0360386612908
#> 5                 FALSE          0.75 66.4195727185255  22.7429234474323
#> 6                 FALSE          0.75 206.158486532196  12.3481795795361
#> 7                 FALSE          0.75 506.364989907731  6.98913412851779
#> 8                 FALSE          0.75 200.344001265293  13.3966015596243
#> 9                 FALSE          0.75  291.39423242257  4.72512376886502
#> 10                FALSE          0.75 676.857152584556 0.634646028459412
#> 11                FALSE          0.75 326.712892363866  9.83371355844878
#> 12                FALSE          0.75 337.627427273859  4.04811779589204
#> 13                FALSE          0.75 736.152693638133 0.728706288140433
#> 14                FALSE          0.75 341.416250248025  9.72005505520092
#> 15                FALSE          0.75 317.427893103127  3.30603382137641
#> 16                FALSE          0.75 778.361198862239 0.510657134235448
#> 17                FALSE          0.75 328.163050154257  7.87492706430775
#> 18                FALSE          0.75 350.486554527311    4.988621039478
#> 19                FALSE          0.75 784.891100128349 0.691816725713093
#> 20                FALSE          0.75 346.907985584335  9.37765190890644
#> 21                FALSE          0.75  339.93478129653  5.28837751614553
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3414.2410236747 2105.87897191653 81646.2702018246 3.27436374970471
#> 2  1461.01149334192 275.618067286522 42504.1507476728 1.79678436665162
#> 3  2194.80925358769  675.17765955949  66200.973772406 1.68790694903741
#> 4  7555.38700022071 6029.40712292744 148886.461818271 3.03778883790073
#> 5  2364.15821928073  1088.6965928959 57097.8480281665 2.72334555376721
#> 6  4822.09434019312 3245.21552755109 130356.154751317  3.0989383140738
#> 7    17022.18394882 15833.4350208257 361648.301744007 13.5969536453369
#> 8  5178.18800407132 3758.48461179683 118398.737150545 9.57532438295886
#> 9  7760.62744760782  6279.6580684126 221166.474235225 5.39169516917868
#> 10 20066.7584065915 18716.7024250607 482294.271053452 22.9582102801874
#> 11 6786.82386785051 5601.31201039781 167626.576089265 14.4605901681177
#> 12 9191.22347434375 7626.02351110492 266679.631308624 6.72386266374501
#> 13 20447.0741283426 19457.2555753076 490377.234308036  19.043691524106
#> 14 6960.66262453187 5655.92927516187 176842.260347813 12.9919337605841
#> 15 9003.57303263335 7393.89969576342 264765.135247029  7.3129538876986
#> 16 20602.4463262664 19287.9657062838 505877.394850895 20.2441032269747
#> 17 6897.94444547736 6021.07467329768 176129.562481361 11.4061536237888
#> 18 9319.99559554911 7645.64638740541 270392.860750901 8.46433441350445
#> 19 19743.6870833149 18592.0133297369 502562.450416991 24.0111789178228
#> 20 7151.69215880043 5830.35928641811 180782.621222345 9.03130389752945
#> 21 9256.25775808531 7470.61441742043 269481.635137861 8.16981879318416
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.558071974275886 24.2705459456985 49.2133953100554 1627.88582576676
#> 2   2.14087031008086 16.6097819513945 62.5543229173302 813.374070935816
#> 3   2.01063843173435 5.97977739284151 115.641627439721 1055.36497991791
#> 4  0.349067506911216 55.5929098664487 97.9789671166993 2928.28701467574
#> 5   2.07929150821213 37.7259824157886 125.704539991721  1500.1536045449
#> 6  0.602418065518105 9.97369567138319 114.284631034197 2863.83640289988
#> 7  0.520334048078629 153.397726219056 199.080507100205 8562.92543093824
#> 8  0.746625138729586 33.5565314819346 119.971787582096 2101.53658249653
#> 9  0.444966188559042 26.5876413532283 95.1452875030837 2454.39987145909
#> 10 0.426287187012588 88.4255339756615 281.841481469975 10651.9128550111
#> 11  0.73103428964494 75.0679628994291 225.867921372763 2671.12912516853
#> 12 0.650959708440612 13.2471782194767 195.274267386678 4767.63780178505
#> 13 0.345330374099815  96.294978578598 231.170433944202 10636.4200373292
#> 14 0.386439757610064 67.3609105447063 118.410557581775 3584.44892425529
#> 15 0.492323809098557 14.5989176400033 100.932148926828 3989.95792322126
#> 16  0.32199503987991 149.948427587363 322.993516655084 16985.7251046937
#> 17 0.496081654402899 89.4610511663017 150.622446125328 3466.87023332088
#> 18 0.408693365898784 8.17651481925653 96.3766081342131  3754.7585088056
#> 19   0.3430413565794 70.3443447780752 248.131389936613 8546.27687193944
#> 20 0.324115749953716 92.4836369469517 121.793134902799 3173.69709317009
#> 21  0.55986015983781 8.37031989507095 149.937523932113 2942.30301376136
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 86.2509768587658 97.0236335952943    14.7663788780354
#> 2             TRUE 29.3926349934398 35.3040555597236     27.980825033314
#> 3             TRUE 62.0477313025897 67.6009451649228    17.0519432442668
#> 4             TRUE  238.35324716981 248.347572446504    14.4618226124218
#> 5             TRUE 61.9396692825784 70.8994761544726    19.3224889164233
#> 6             TRUE 201.060733005545 211.256240058847    11.3572018617589
#> 7             TRUE 483.998001161151  528.73197865431    6.13318461942845
#> 8             TRUE 184.592592655326 216.095409875261    12.1684032064141
#> 9             TRUE 282.524893869271 300.263570975869     3.9931543886854
#> 10            TRUE 639.090896673648 714.623408495464 -0.0665963941762955
#> 11            TRUE 302.925221537312 350.500563190419    8.63116215198285
#> 12            TRUE 326.566673191999  348.68818135572    2.97728907550723
#> 13            TRUE 704.825821080979 767.479566195287   0.160637822746237
#> 14            TRUE 320.044519211864 362.787981284186    9.08436165393236
#> 15            TRUE 305.398083957863 329.457702248391    2.49616115540928
#> 16            TRUE 745.059649053866 811.662748670613 -0.0190247063670037
#> 17            TRUE 309.399927443124  346.92617286539    7.05887274281498
#> 18            TRUE 336.562724417096 364.410384637526     4.3163204525745
#> 19            TRUE  745.39271080853 824.389489448168    0.12751369413998
#> 20            TRUE 332.051490672899 361.764480495771    8.84448150023257
#> 21            TRUE 326.495429381742 353.374133211318    4.36740755321234
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.602435673403 3374.31597559402 3454.16607175537    78968.3980184383
#> 2     35.02428835348 1433.68840203188 1488.33458465197    41166.1504009834
#> 3   23.6669436846729 2184.97251977646 2204.64598739891     64464.898380441
#> 4   15.6102547101597  7463.9366634904 7646.83733695102    144069.429679129
#> 5   26.1633579784412 2302.09897820676  2426.2174603547    54630.0953486901
#> 6   13.3391572973134 4805.68761081369 4838.50106957254    125645.143868546
#> 7   7.84508363760714 16769.8446891896 17274.5232084503    347562.289410113
#> 8   14.6247999128344 5122.98750978354  5233.3884983591    114941.709472339
#> 9   5.45709314904465 7716.89077758176 7804.36411763388    217128.986446675
#> 10  1.33588845109512 19921.2984032015 20212.2184099815    464771.874406959
#> 11  11.0362649649147 6663.33706888095 6910.31066682008    163232.568678362
#> 12  5.11894651627684 9169.43186617271 9213.01508251479    258836.867124687
#> 13  1.29677475353463 20288.6688885808 20605.4793681044     472880.32334663
#> 14  10.3557484564695 6849.85392668583 7071.47132237791    170945.841867413
#> 15  4.11590648734353 8979.55781311554 9027.58825215115     258201.65446333
#> 16   1.0403389748379 20355.7811628852 20849.1114896476    477935.877053674
#> 17  8.69098138580052 6750.78101630879 7045.10787464592    170426.560947548
#> 18   5.6609216263815 9306.54522867143 9333.44596242678    264216.283003916
#> 19  1.25611975728621  19627.970636155 19859.4035304749    488503.824962651
#> 20   9.9108223175803  6999.5565760227 7303.82774157817     175561.88950408
#> 21  6.20934747907873 9242.48858185792  9270.0269343127    264641.546680223
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84324.1423852109 85.2195522776088 98.0550581764513   14.5905862061385
#> 2     43842.1510943622 28.8266479179446 35.8700426352189   27.3064508856385
#> 3     67937.0491643709 61.5160406136429 68.1326358538695   16.4185921382705
#> 4     153703.493957412 237.396343685872 249.304475930443   14.3518663477448
#> 5     59565.6007076429 61.0818154331417 71.7573300039092   18.6675120913365
#> 6     135067.165634087 200.084567436611 212.232405627781   11.1674401711207
#> 7       375734.3140779  479.71496076287 533.015019052591   5.96927939428368
#> 8     121855.764828752 181.576365474694 219.111637055893   11.9332162877143
#> 9     225203.962023775  280.82650989098  301.96195495416    3.8529900392893
#> 10    499816.667699946 631.859060435389 721.855244733723 -0.200876858085261
#> 11    172020.583500167 298.370135634355 355.055649093376    8.4008863507447
#> 12     274522.39549256 324.448656452919 350.806198094799   2.77223676734844
#> 13    507874.145269443 698.827058250886 773.478329025381 0.0518587549047956
#> 14    182738.678828213  315.95206007728  366.88044041877   8.96263313028519
#> 15    271328.616030728 303.094503483238 331.761282723016   2.34107915554324
#> 16    533818.912648116 738.682756537369  818.03964118711 -0.120453143929175
#> 17    181832.564015174 305.806989051631 350.519111256883   6.90260702167807
#> 18    276569.438497886 333.896459076842  367.07664997778   4.18758204231639
#> 19    516621.075871332 737.829189449416 831.953010807282 0.0194556668174691
#> 20     186003.35294061 329.206629945177 364.609341223493   8.74238503899715
#> 21    274321.723595498 323.921936461889 355.947626131171   4.19105160286343
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7782283452999 3366.67075362113 3461.81129372826    78455.6139833218
#> 2   35.6986625011555 1428.45632071719 1493.56666596665    40909.9375686386
#> 3   24.3002947906692 2183.08888989772 2206.52961727766    64132.4584117668
#> 4   15.7202109748368 7446.42489688247 7664.34910355895    143147.019269506
#> 5   26.8183348035281 2290.21529374579 2438.10114481568    54157.5469632585
#> 6   13.5289189879516 4802.54589667721 4841.64278370903    124743.035401633
#> 7   8.00898886275191 16721.5244054306 17322.8434922093    344864.967899368
#> 8   14.8599868315342 5112.41720236673 5243.95880577591    114279.725448852
#> 9   5.59725749844075 7708.51567055549 7812.73922466015    216355.850487165
#> 10  1.47016891500408 19893.4443599992 20240.0724531838    461416.521857631
#> 11  11.2665407661529 6639.69066056763 6933.95707513339    162391.163003934
#> 12  5.32399882443564 9165.25900503358 9217.18794365393    257335.061217125
#> 13  1.40555382137607 20258.3359703286 20635.8122863567    469529.851034871
#> 14  10.4774769801166 6828.63523986424 7092.69000919949    169816.740456273
#> 15  4.27098848720958 8974.95915405894 9032.18691120775    256944.817717516
#> 16  1.14176741240007 20308.5474081952 20896.3452443376    472585.373645695
#> 17  8.84724710693743 6722.60078519141 7073.28810576331    169334.496824052
#> 18  5.78966003663962 9303.96962650336 9336.02156459485    263033.534073642
#> 19  1.36417778460872 19605.8121675499   19881.56199908     485811.74774799
#> 20  10.0129187788157 6970.42423038441 7332.96008721646    174562.174919732
#> 21  6.38570342942764 9239.85193109097 9272.66358507965    263714.721230889
#>    ci95.integral.bt.up
#> 1     84836.9264203275
#> 2      44098.363926707
#> 3     68269.4891330451
#> 4     154625.904367035
#> 5     60038.1490930745
#> 6        135969.274101
#> 7     378431.635588645
#> 8     122517.748852239
#> 9     225977.097983285
#> 10    503172.020249274
#> 11    172861.989174595
#> 12    276024.201400122
#> 13    511224.617581202
#> 14    183867.780239354
#> 15    272585.452776543
#> 16    539169.416056095
#> 17     182924.62813867
#> 18     277752.18742816
#> 19    519313.153085992
#> 20    187003.067524959
#> 21    275248.549044833