Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9048008906872  15.3293002921799
#> 2                 FALSE          0.75 33.1129020357537  30.9602838225788
#> 3                 FALSE          0.75 64.7640373322732  20.5078447226822
#> 4                 FALSE          0.75  243.06771749935  15.0534433337388
#> 5                 FALSE          0.75 66.3605050931127  23.0208045338911
#> 6                 FALSE          0.75 205.844555651964  12.4511919883958
#> 7                 FALSE          0.75 507.782045890233  7.02136159469988
#> 8                 FALSE          0.75 205.103321179004  13.6103951961136
#> 9                 FALSE          0.75 288.879697395885  4.60813153665333
#> 10                FALSE          0.75 668.869556069009 0.442466359465097
#> 11                FALSE          0.75 320.712782472166  9.63705251567027
#> 12                FALSE          0.75 332.082511082405  3.72265264400567
#> 13                FALSE          0.75 730.915267926396 0.853017322797684
#> 14                FALSE          0.75 342.382501971186  9.93574019619607
#> 15                FALSE          0.75 319.254420891676  3.29474352887238
#> 16                FALSE          0.75 769.054543522436 0.412703106525307
#> 17                FALSE          0.75 336.305419890159  8.05950833625938
#> 18                FALSE          0.75 351.627958919648  5.16571271350806
#> 19                FALSE          0.75 775.245335821704 0.725784702230261
#> 20                FALSE          0.75 353.812264366168  9.61270729118099
#> 21                FALSE          0.75 336.639822238105  5.05395484304405
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3420.14511434928 2144.61349106703 82224.6045936637 2.36948781917483
#> 2  1464.71156882869 277.722549093022 42254.2939321884 3.20664194179926
#> 3  2195.38384470937 657.960573414777 66389.9503227211  1.2373952210436
#> 4  7532.51765866756 6002.64417752598 147722.531952406 2.55375530158918
#> 5  2365.92283591275 1080.05298338786 57054.0170231349 2.57732586823819
#> 6  4822.26156166578 3219.34762639437 130435.940746067 2.35071101941598
#> 7  17027.3907298002 15822.2169314699 361060.192003421 14.3187972844458
#> 8  5194.92536164519 3763.61613616088 119426.361153197 9.03842418126939
#> 9  7764.56546505477 6289.93995299343 220507.019991654 8.83163923963542
#> 10 20048.2791282209 18770.4222530409 479599.517885798 24.6172807680691
#> 11 6765.28487623588 5575.32037021793 167245.017181252 11.4981399095893
#> 12 9193.38095427039 7651.83131819311 267188.651721105 4.94150149782224
#> 13 20466.0061602256 19355.7072343665 490484.833432327 28.3984988444638
#> 14 6971.40574583461 5612.25308626158 176741.035322755 17.1211437276931
#> 15 9001.84607081439 7411.32614161518 264633.685730346 5.36823179538832
#> 16 20677.0213560233 19373.7170979246 514606.566513585 17.5955773903857
#> 17 6908.61875168335 6036.55521847144  176150.23595851 12.3353039981267
#> 18 9323.31132530983 7609.98574546901 270684.762191366 5.94546456782485
#> 19 19775.4554726437 18517.5051611933 503686.991773331 23.6717314937769
#> 20 7131.31233081638 5810.48362639589 179527.736841993 14.2053356391428
#> 21 9253.03782094712 7504.98870540776 270076.726747636 7.96079122715013
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.715962473585455 16.9424332086934 56.9568141120081 988.982693187428
#> 2   2.06306529448509 17.9292938149145 48.7395375336322 846.358599993553
#> 3   1.62342499432438 5.83637675794228 105.912853325153 1138.02151405469
#> 4  0.348287205458835 75.5107306170826 100.837654285242 3953.14652286725
#> 5   2.25707214145991 28.4615026460272 111.631715192367 992.903051975799
#> 6  0.602075298249283 6.94822045525716 124.814636234167 2392.91001726713
#> 7   0.52904633453571 167.175514099274 270.017360571421  9681.9333633792
#> 8  0.638498940311663 30.9055491856257 131.884973784786 1718.33582981645
#> 9  0.477994382477355 18.8877344837214 76.7631014119635 4554.98488399328
#> 10  0.25883417422422 62.7163771507416 171.930425339893 11812.5040529908
#> 11 0.737249232807393 87.4196182953812 198.330699401531 3512.27284428855
#> 12 0.548409478449792 9.97657461139764 163.647793970408 3894.20939138863
#> 13 0.432219965411062 111.037569592621  292.94701837548 9436.94398159746
#> 14 0.738420012028469 46.3654032406396 198.035331904939 2582.33699169589
#> 15 0.389263689435613 24.6978779420881 75.7184287632646 5519.57159831706
#> 16 0.286267357475729 87.2994651729472 196.570272310435 9255.04427241805
#> 17 0.638947538930333 67.6066347631881 142.637187417314  2640.5265830883
#> 18 0.517084148601891 10.4367024463417 144.010613144752 3649.24349950317
#> 19 0.353176932445361 71.4220687817854 215.087647455669 7606.74004110275
#> 20 0.282803583108128 95.1973349671171 124.842985907282 3544.69981842771
#> 21 0.420237358520469 11.1840649923447  74.554255464111 2421.54118509517
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.0069934281446 95.8026083532298    14.1515420231318
#> 2             TRUE 27.8379760414939 38.3878280300135    27.5665414131508
#> 3             TRUE 62.7285221936565 66.7995524708899    17.8373106070186
#> 4             TRUE 238.866790028235 247.268644970464     14.480510880759
#> 5             TRUE 62.1208040398609 70.6002061463646    19.3079208611896
#> 6             TRUE 201.977636025025 209.711475278904    11.4607781227757
#> 7             TRUE  484.22762435732 531.336467423146    6.15108037438864
#> 8             TRUE 190.235113400816 219.971528957192    12.5600644393009
#> 9             TRUE 274.351650846685 303.407743945085    3.82183077747808
#> 10            TRUE 628.374129205536 709.364982932483  0.0166841428662557
#> 11            TRUE 301.798342320892 339.627222623441    8.42427752770211
#> 12            TRUE 323.953741118488 340.211281046323    2.82051905195576
#> 13            TRUE 684.199737327254 777.630798525539   0.142015479696488
#> 14            TRUE 314.218220539131 370.546783403241    8.72103927640924
#> 15            TRUE 310.423679588262 328.085162195089     2.6544047597508
#> 16            TRUE 740.109818715251  797.99926832962 -0.0582066965222667
#> 17            TRUE 316.013844813241 356.596994967078    7.00843963471898
#> 18            TRUE 341.847669705576  361.40824813372    4.31510928905794
#> 19            TRUE 736.305337514441 814.185334128967   0.144808648357643
#> 20            TRUE 330.444487239778 377.180041492558    9.14749539696811
#> 21            TRUE 323.544320669443 349.735323806767    4.36266438827788
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.507058561228 3392.27481172098 3448.01541697758    80597.7280633704
#> 2   34.3540262320067 1435.21788050316 1494.20525715423     40862.034035199
#> 3   23.1783788383458 2185.78300494255 2204.98468447618    64517.9049321011
#> 4   15.6263757867186 7408.30250680246 7656.73281053266    141219.605922289
#> 5   26.7336882065927 2319.10366406004 2412.74200776547    55420.6915026347
#> 6   13.4416058540158 4810.83173901688 4833.69138431468    126499.603767662
#> 7   7.89164281501112 16752.3870091068 17302.3944504935    345133.411620663
#> 8   14.6607259529263 5144.08573323483 5245.76499005554    116599.698713149
#> 9   5.39443229582858 7733.49514182905 7795.63578828049    213014.069857485
#> 10 0.868248576063939 19945.1106878079 20151.4475686339    460167.948718628
#> 11  10.8498275036384 6621.47960413997 6909.09014833178    161467.328352397
#> 12  4.62478623605558 9176.96948903464 9209.79241950614     260782.67727227
#> 13  1.56401916589888 20283.3493582457 20648.6629622054    474961.060582599
#> 14  11.1504411159829 6895.13465750376 7047.67683416546    172493.090971416
#> 15  3.93508229799397 8961.21806159965 9042.47408002912    255553.990451114
#> 16  0.88361290957288 20533.4137358138 20820.6289762328    499382.018685457
#> 17  9.11057703779978 6797.40583749791  7019.8316658688     171806.56972933
#> 18  6.01631613795817  9306.1429497856 9340.47970083406    264681.756634683
#> 19  1.30676075610288 19657.9661694976 19892.9447757897    491173.904405717
#> 20  10.0779191853939 6974.71271479548 7287.91194683729    173696.705640679
#> 21  5.74524529781022 9234.64003403472 9271.43560785953    266093.291498154
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1      83851.481123957 87.2606047651045 96.5489970162699    13.9260138439524
#> 2     43646.5538291778 26.8278838298272 39.3979202416803     26.916675845388
#> 3      68261.995713341 62.3387426990278 67.1893319655187    17.3259317338064
#> 4     154225.457982522 238.062357108235 248.073077890465    14.3708004110395
#> 5     58687.3425436351 61.3089463913659 71.4120637948596    18.5969431366297
#> 6     134372.277724471 201.237162053909  210.45194925002    11.2711244038272
#> 7      376986.97238618 479.717203212719 535.846888567747    5.98443077900989
#> 8     122253.023593245 187.388009783716 222.818632574292    12.3589372731027
#> 9     227999.970125823   271.5696844862 306.189710305571    3.67126254699771
#> 10    499031.087052968 620.619685763594 717.119426374425 -0.0648486220143734
#> 11    173022.706010107 298.176428249371 343.249136694961    8.19204401936778
#> 12    273594.626169939 322.397168146674 341.767854018137    2.64777006624408
#> 13    506008.606282055 675.254210191247 786.576325661546 0.00586619059200333
#> 14    180988.979674095 308.825060264908 375.939943677465    8.48843697262027
#> 15    273713.381009577 308.732686572714 329.776155210637    2.53178669757858
#> 16    529831.114341712  734.56721183728 803.541875207592  -0.148380914127121
#> 17     180493.90218769 312.128224053831 360.482615726488    6.80717115995593
#> 18    276687.767748048 339.974848366711 363.281069472584    4.15222778224835
#> 19    516200.079140945 728.848742093901 821.641929549507   0.033557914637354
#> 20    185358.768043306 325.969806513448 381.654722218888    9.05841226828905
#> 21    274060.161997117 321.036671432891 352.242973043319    4.23028962034393
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7325867404074 3386.93794526024 3453.35228343832    80286.1985150164
#> 2   35.0038917997695 1429.57015295146 1499.85298470592     40595.431076201
#> 3    23.689757711558  2183.9445462638 2206.82314315494    64159.4281551739
#> 4   15.7360862564381 7384.51662665808 7680.51869067704    139974.364767586
#> 5   27.4446659311525 2310.13829072654 2421.70738109897    55107.9270412624
#> 6   13.6312595729644 4808.64304957348 4835.88007375809    125745.837112223
#> 7   8.05829241038987 16699.7267221656 17355.0547374347    342083.602611198
#> 8   14.8618531191244 5134.35048524136 5255.50023804902    116058.422926756
#> 9   5.54500052630894 7727.54550546667 7801.58542464286    211579.249619027
#> 10 0.949781340944568 19925.3550290055 20171.2032274364    456447.009941936
#> 11  11.0820610119728 6593.94242437693 6936.62732809482    160360.962406446
#> 12  4.79753522176726 9173.82686803205 9212.93504050873    259556.001313983
#> 13  1.70016845500336  20248.372523824 20683.6397966271    471988.423228396
#> 14  11.3830434197719 6880.52955548296 7062.28193618626    171679.654819031
#> 15  4.05770036016618  8953.4382300479 9050.25391158088    253815.325397644
#> 16 0.973787127177735 20505.9144042843 20848.1283077622    496466.679739645
#> 17  9.31184551256283  6776.1097475475  7041.1277558192    170974.803855657
#> 18  6.17919764476776   9302.855388515 9343.76726210466    263532.244932339
#> 19  1.41801148982317 19635.4682178314  19915.442727456    488777.781292769
#> 20  10.1670023140729 6944.72555428083 7317.89910735193    172580.125197874
#> 21  5.87762006574417 9231.11705356213 9274.95858833212    265330.506024849
#>    ci95.integral.bt.up
#> 1     84163.0106723111
#> 2     43913.1567881757
#> 3     68620.4724902683
#> 4     155470.699137225
#> 5     59000.1070050075
#> 6      135126.04437991
#> 7     380036.781395645
#> 8     122794.299379637
#> 9      229434.79036428
#> 10     502752.02582966
#> 11    174129.071956058
#> 12    274821.302128226
#> 13    508981.243636258
#> 14    181802.415826479
#> 15    275452.046063047
#> 16    532746.453287524
#> 17    181325.668061363
#> 18    277837.279450392
#> 19    518596.202253892
#> 20    186475.348486111
#> 21    274822.947470422