Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  92.026403827577  15.4095233820838
#> 2                 FALSE          0.75 33.2356009101549  31.6127676794221
#> 3                 FALSE          0.75 64.8863221475228  20.0254762092608
#> 4                 FALSE          0.75 242.712045496836  14.9553688285213
#> 5                 FALSE          0.75 65.7906914305242  21.9877570755524
#> 6                 FALSE          0.75 206.194421811619  12.5007127307483
#> 7                 FALSE          0.75 507.863641435599  7.07912202410988
#> 8                 FALSE          0.75 201.587694173347  13.3595901532345
#> 9                 FALSE          0.75 290.874547529016  4.72953363574638
#> 10                FALSE          0.75 675.062512625664 0.562272861526407
#> 11                FALSE          0.75  319.90119219143  9.76496079554976
#> 12                FALSE          0.75 334.872137600555   3.7792426369811
#> 13                FALSE          0.75 735.467100227552 0.863505941699347
#> 14                FALSE          0.75 346.800785122545  9.97568860409255
#> 15                FALSE          0.75 319.395500849826  3.49462227502807
#> 16                FALSE          0.75 776.775421567224 0.482749346158321
#> 17                FALSE          0.75 335.616050025169   8.1895874198866
#> 18                FALSE          0.75 349.612966174778  4.86232013236301
#> 19                FALSE          0.75 779.677389537186 0.648655374921612
#> 20                FALSE          0.75 352.253557259476  9.55660306881462
#> 21                FALSE          0.75 338.093456191127  5.25005450931052
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3420.46223512679 2139.09784881812 82139.3127256234 1.93401991055349
#> 2  1458.80808194143 256.363860478799 42258.1755147646  2.7063464853223
#> 3  2195.51639813157 685.345424738116 65848.6110116353  1.3174887600261
#> 4  7574.97034701995 6065.73912229523 149551.254963105 2.87504642150415
#> 5  2356.71641616215 1121.13276696877 57002.5749578498 2.68822245468069
#> 6  4818.87752623113 3210.63496981106 130897.624509874 2.26377804735899
#> 7  17048.6881034955  15830.774851098 363115.432051174 8.68432803650928
#> 8  5175.23614263214 3763.61324352906 118441.410079695 8.44822400201699
#> 9  7772.34761596469 6287.23844576256 220879.531030852 7.86344738707524
#> 10 20062.8807175767 18730.9103957178 484554.750188049 26.7421479742048
#> 11 6744.52906061554 5560.63983713918 166652.375531278 7.61349394650728
#> 12 9181.27964058805 7665.62463711299  268347.57326141 4.98240916747728
#> 13 20442.8391809205 19333.5746857617 491078.508424312 32.9591112320461
#> 14 6974.47657787978 5622.71390460717 177081.191579305 11.5314572475029
#> 15  8999.0370060544 7358.09370399618 265253.747428989  6.6170014583794
#> 16 20672.4713051406 19369.7894117639 513515.899956329 21.7977442109555
#> 17  6881.0890786927 5965.60821187806 174667.887270857 16.0734120177611
#> 18 9312.96193906291 7666.46197565085 270092.330284939 5.13392992781902
#> 19 19783.5031276184 18620.2453805182 503913.690466872 27.6655020451429
#> 20  7144.2647410753 5818.22155606212 180109.462077534  10.240995598037
#> 21 9255.08344826815 7467.68178825309 268524.369805714 5.63670400093989
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.620808643910937 16.8065196430895 50.3877465843027 1395.76588928319
#> 2   2.32919822917531 20.1432455688035 59.8718399056327 932.367453048624
#> 3   1.93495725394233 5.22761154261091 122.555034128748  1086.7907628519
#> 4  0.321708328012671  42.111195359602 91.9552121552244  2349.0204052852
#> 5   2.68897933080443 38.2382946181822 176.280400935799 1418.64230398643
#> 6  0.347970837715392  10.930173625687 68.2329145342636 1724.20674414593
#> 7  0.368466621307349 119.934445250193  136.33612148703 6748.62450270984
#> 8  0.798588492026803 36.4638877979836 144.623548914205 2213.07912980701
#> 9   0.38289361460363 14.3983395416142 60.6631385047107 3424.76778706314
#> 10  0.32354401025446 76.0804437408543  196.78255197239 10335.1526675934
#> 11 0.396456810372863 93.5516072407324  141.57160277158 3694.87172959209
#> 12 0.334535608741691 24.3932402691006 109.754752494804 3104.41675888565
#> 13 0.465098112041074 70.4740927343792 260.792423204446 8105.61041971212
#> 14 0.667106014269741 54.9142085031061 232.902262657501 2684.49823111822
#> 15 0.547088593445419 15.0461521362494 154.234257134286 4035.61457848268
#> 16 0.397512354299007 112.476288117005 285.454302034266 10492.5589160241
#> 17 0.686108654284922 92.4578393733782 178.200146061589 3409.82358001516
#> 18 0.252717729937627 9.94387285655359 70.0573496573362 5040.45288671062
#> 19 0.320354025380033 84.3041316009122 243.157827006131 7453.62535889597
#> 20 0.344272711450606 84.1614892015442 113.970055148203 3022.03444873323
#> 21 0.366693470032128 5.78115999783254 113.332670864647 4258.39462486256
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.8449410747165 95.2078665804375   14.3882931628503
#> 2             TRUE 28.7836609417997 37.6875408785101   27.7812365924287
#> 3             TRUE 62.7190531372798 67.0535911577657   16.8424715265257
#> 4             TRUE 237.982594133462 247.441496860211   14.4261586289405
#> 5             TRUE 61.3685654925744 70.2128173684739   17.5643860763792
#> 6             TRUE 202.470506923713 209.918336699524   11.9283007027065
#> 7             TRUE 493.577921815541 522.149361055657   6.47299443205929
#> 8             TRUE 187.690365690029 215.485022656665   12.0459120838504
#> 9             TRUE 277.939176577277 303.809918480755   4.09967363972341
#> 10            TRUE 631.071679208097  719.05334604323 0.0300429646578201
#> 11            TRUE 307.376994649425 332.425389733434    9.1127893424864
#> 12            TRUE 326.676074520055 343.068200681055   3.22893156060102
#> 13            TRUE 681.249362250836 789.684838204268 0.0984195473917792
#> 14            TRUE 327.831537950403 365.770032294688   8.87829921061883
#> 15            TRUE 308.510533450792  330.28046824886   2.59466153881036
#> 16            TRUE 740.918132340202 812.632710794246 -0.171158476663545
#> 17            TRUE 309.175287255952 362.056812794386   7.06093868358791
#> 18            TRUE 341.167651443515  358.05828090604   4.44659946661561
#> 19            TRUE 734.167638672926 825.187140401447  0.121673003171458
#> 20            TRUE 335.407119500706 369.099995018247   8.99027445847838
#> 21            TRUE 328.821078109581 347.365834272673   4.64684375110767
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4307536013173 3392.81551031391 3448.10895993968    79843.2778377525
#> 2   35.4442987664155 1425.67244298075 1491.94372090211    40724.4310544996
#> 3   23.2084808919959 2186.91697714397 2204.11581911916    64060.8402067439
#> 4   15.4845790281022  7505.6974306534  7644.2432633865    145687.116396411
#> 5   26.4111280747257 2293.81442151524 2419.61841080906    54668.9083677921
#> 6   13.0731247587901 4800.89739061687 4836.85766184538    128061.304415754
#> 7   7.68524961616047  16851.395941059 17245.9802659321    352013.944744216
#> 8   14.6732682226186 5115.25304720446 5235.21923805983    114800.894911162
#> 9   5.35939363176935 7748.66234741873 7796.03288451065    215245.788021133
#> 10  1.09450275839499  19937.728387623 20188.0330475304    467553.424049858
#> 11  10.4171322486131 6590.63666670453 6898.42145452654    160574.311536099
#> 12  4.32955371336118 9141.15276034538 9221.40652083072    263240.807693043
#> 13  1.62859233600691 20326.9092983724 20558.7690634685    477744.779283886
#> 14  11.0730779975663 6884.14270489217 7064.81045086739    172665.191989115
#> 15  4.39458301124579 8974.28608579027 9023.78792631853    258615.161447385
#> 16  1.13665716898019 20487.4478111882 20857.4947990931     496255.64053947
#> 17   9.3182361561853 6728.99593292349  7033.1822244619    169058.727481732
#> 18  5.27804079811041 9296.60426821388 9329.31960991194      261800.7852863
#> 19  1.17563774667177 19644.8228311349 19922.1834241019    491652.476751488
#> 20  10.1229316791509 7005.81909133876 7282.71039081184    175138.215409368
#> 21  5.85326526751337 9245.57344007171 9264.59345646458    261519.310647815
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84435.3476134942 88.2357248028922 95.8170828522619    14.1927384400184
#> 2     43791.9199750296 27.9311617989232 38.5400400213866    27.0475391502385
#> 3     67636.3818165267 62.3040441778716 67.4686001171739    16.2329599915338
#> 4     153415.393529799 237.076954510688 248.347136482985    14.3248205056165
#> 5     59336.2415479075   60.52177541935 71.0596074416983    16.7173575871758
#> 6     133733.944603994 201.757416838795 210.631426784443    11.8186898888261
#> 7     374216.919358132 490.842358484041 524.884924387157    6.35692744634747
#> 8     122081.925248228 185.029175129394   218.1462132173     11.794356708862
#> 9      226513.27404057 275.462190650348 306.286904407683    3.97906215112326
#> 10    501556.076326241 622.647902596222 727.477122655105 -0.0718733985723349
#> 11    172730.439526457 304.978744056276 334.823640326584    8.98790544721895
#> 12    273454.338829777 325.106615632299  344.63765956881    3.12355284384738
#> 13    504412.237564739 670.867242212742 800.066958242363 -0.0480863579011593
#> 14    181497.191169494  324.19912891744 369.402441327651    8.66816081612386
#> 15    271892.333410593 306.426177991402  332.36482370825    2.42232863187505
#> 16    530776.159373189 734.051842913751 819.499000220697  -0.296374868267733
#> 17    180277.047059981 304.112162470358 367.119937579981    6.84481445748816
#> 18    278383.875283577 339.550463516252 359.675468833303    4.36699338168526
#> 19    516174.904182256 725.453005528706 833.901773545667  0.0207614851767471
#> 20    185080.708745701 332.181205887324 372.325908631629    8.88182855437144
#> 21    275529.428963613 327.045516349285 349.141396032969    4.53133530804755
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6263083241493 3387.52145662634 3453.40301362725    79403.6115826283
#> 2   36.1779962086057 1419.32732062657 1498.28884325628    40430.7353067893
#> 3   23.8179924269878 2185.27027950805 2205.76251675508    63718.5011164456
#> 4   15.5859171514262 7492.43240411513 7657.50828992477    144947.174968746
#> 5   27.2581565639291 2281.76935871051 2431.66347361379    54222.0360420364
#> 6   13.1827355726705 4797.45438592478 4840.30066653748    127518.179291348
#> 7   7.80131660187228 16813.6165908052 17283.7596161859    349888.128025863
#> 8    14.924823597607 5103.76692254809 5246.70536271619    114103.774985273
#> 9   5.48000512036949 7744.12687046313 7800.56836146625    214166.986168208
#> 10  1.19641912162515 19913.7630478446 20211.9983873088    464297.850959566
#> 11  10.5420161438806  6561.1679104237 6927.89021080737    159410.426941278
#> 12  4.43493243011481 9133.46888966061 9229.09039151549    262262.916413995
#> 13  1.77509824129985 20304.7099591611 20580.9684026798    475191.512001676
#> 14  11.2832163920612 6866.84472921369 7082.10842654587    171819.575046313
#> 15  4.56691591818109 8969.54654786735 9028.52746424144    257343.942855163
#> 16  1.26187356058437 20452.0177804313   20892.92482985    492950.484480922
#> 17  9.53436038228505 6699.87171352088 7062.30644386452    167984.633054027
#> 18  5.35764688304076 9293.47194826406 9332.45192986175    260213.042626986
#> 19  1.27654926466648 19618.2670296806 19948.7392255562    489304.584763436
#> 20  10.2313775832578 6979.30822224027 7309.22125991033    174186.274558017
#> 21  5.96877371057349 9243.75237467239  9266.4145218639    260177.916340984
#>    ci95.integral.bt.up
#> 1     84875.0138686184
#> 2     44085.6157227399
#> 3      67978.720906825
#> 4     154155.334957464
#> 5     59783.1138736632
#> 6       134277.0697284
#> 7     376342.736076485
#> 8     122779.045174117
#> 9     227592.075893495
#> 10    504811.649416533
#> 11    173894.324121278
#> 12    274432.230108826
#> 13    506965.504846948
#> 14    182342.808112297
#> 15    273163.552002815
#> 16    534081.315431737
#> 17    181351.141487686
#> 18    279971.617942891
#> 19    518522.796170308
#> 20    186032.649597051
#> 21    276870.823270445