Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2492959402625  15.3827108048907
#> 2                 FALSE          0.75 32.6270536308262    32.00023420256
#> 3                 FALSE          0.75 64.7508553400194  19.9365640512067
#> 4                 FALSE          0.75 242.105682301437  14.9663039572265
#> 5                 FALSE          0.75 65.0802200072715  21.9482880398184
#> 6                 FALSE          0.75 205.106691931076  12.3459774968432
#> 7                 FALSE          0.75 505.483770130124  7.02266717853062
#> 8                 FALSE          0.75 200.754442943335  13.1820404954911
#> 9                 FALSE          0.75 289.926883066613  4.82350963453962
#> 10                FALSE          0.75 673.215206793912 0.603061260220768
#> 11                FALSE          0.75 322.908214613241  9.90671400863138
#> 12                FALSE          0.75 334.784713596267  3.84913648489079
#> 13                FALSE          0.75 732.940196012464 0.839543373255565
#> 14                FALSE          0.75 341.866951319086  9.72004386462703
#> 15                FALSE          0.75 321.136401858921  3.38980158886283
#> 16                FALSE          0.75 778.853801498857 0.452174918723667
#> 17                FALSE          0.75 343.238797196516  8.21878422283926
#> 18                FALSE          0.75 352.571481133462  5.05390057047736
#> 19                FALSE          0.75  762.29307024464 0.464259812712253
#> 20                FALSE          0.75 349.308378673063  9.49593899938548
#> 21                FALSE          0.75 336.246716522253  5.04723115462518
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.87646763593 2139.76698385842 81960.1712668387 2.45270335835928
#> 2  1465.91496636413 266.943673840679  42538.928469948 1.58995249489054
#> 3  2196.58834069369 703.770221374684  65564.434805502 1.09978188351085
#> 4  7541.59177159365 6017.75523366051 148115.040800527 2.27888064530426
#> 5  2371.28191782537 1117.69906315286 57538.4495304886 2.85245347297297
#> 6  4813.11377512905 3228.33912157873 131330.678785908 2.91205162501968
#> 7  17029.4340195341 15795.3860896467 362161.278831905 12.5787029239212
#> 8   5187.1502888015  3810.5055263727 119559.856218406 8.07742850003214
#> 9  7759.46784931829 6234.68035877051  221492.12108911 6.34714065556022
#> 10 20081.8528957849 18702.5147261465 484743.095922813 26.6859410175812
#> 11 6813.56666371745 5562.95706754613 168027.824465709 13.5931455564401
#> 12 9194.19141835677 7660.01627276942 267505.588309725 6.53657976842649
#> 13 20438.0507705614 19359.9136105449 489486.543560016 24.8439114262434
#> 14 6985.21573433741 5679.33840040919 177784.229499713 17.3998305907145
#> 15 9004.72957161971 7413.69594444104 265117.159645345 7.55314475450763
#> 16  20693.582551544 19401.4037934351 515172.021509108 21.2464886590876
#> 17 6901.96275471368 6028.03600374237 176376.996314331 13.1861019554587
#> 18 9317.89886882119 7661.49202870732 271067.763565489 5.36824315687446
#> 19 19774.5921066696  18660.935346061 498559.937988991 24.3332365730743
#> 20 7142.41729937786 5824.18120769567 180648.724892034 11.0886085270676
#> 21 9261.28817909935 7494.08780387586 270059.199069391 8.91693354062921
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.491887702979614 25.8799007014979 56.9730624276391 1363.81380283105
#> 2   2.80854050661866 19.1779335555071 84.9981377058771 947.521327201009
#> 3   1.98180779501868 4.74137059282082 129.345939200892 1428.88996161191
#> 4  0.307556097379432 98.8337614870309 88.6061110181686  5252.8580039776
#> 5   5.56616139325527 25.3329609964468 152.517136097237 942.340719653127
#> 6  0.369269247367029 10.7002347124708  80.686579083516 1701.77652726101
#> 7  0.519389606316602 125.612788540138 218.911377681454 6292.55215148998
#> 8   0.67884448631221 32.1835632714423 114.811116357443 2286.17335126329
#> 9  0.438963896440642 28.4336668269122 89.3047192792715 3261.28841084331
#> 10 0.495800122932799 57.1688617790684 286.616083923291 10186.3584147797
#> 11  0.85707177186308  79.218731436645 214.674216946676 2517.80841794303
#> 12 0.338947528660221 13.7131053416341 100.688480172202 3516.32108993093
#> 13 0.437908851595478 118.851898178123  247.97633562972 10410.5230120532
#> 14 0.479164167516696 38.9843765397684 149.107756260201 2207.23639765064
#> 15 0.366809823432049 13.7104160520029  67.075789497457 2634.12313081973
#> 16 0.322400239526911 107.860761658896 245.981964308441 9410.91876373178
#> 17 0.532357105146975 73.3826952236913 134.085960066515 3189.80238298479
#> 18 0.284287657613647  11.144615386902 74.8756113680429 4625.13658059115
#> 19 0.277405404249026 70.3819377764025  194.81545246371 16298.4413812916
#> 20 0.362581660073423   63.73568798711 110.565399225064 2404.25690687975
#> 21 0.713404360699584 8.87564311756139 144.115904891375 1753.37306321089
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.2145989157614 96.2839929647635    14.5735555334892
#> 2             TRUE 30.0115817767313 35.2425254849211    27.3801850691723
#> 3             TRUE  62.941714141644 66.5599965383948    16.6764902284009
#> 4             TRUE 238.356923639912 245.854440962963    14.4603741770374
#> 5             TRUE  60.387934044231 69.7725059703121    12.7919525479134
#> 6             TRUE 200.316367007919 209.897016854233    11.7385295849244
#> 7             TRUE 484.791803820273 526.175736439974    6.16827127613981
#> 8             TRUE 187.467073060783 214.041812825888    12.0653413155075
#> 9             TRUE 279.485836688217  300.36792944501    4.10141402489477
#> 10            TRUE 629.316833819991 717.113579767833  -0.212529942003686
#> 11            TRUE 300.547490172897 345.268939053585    8.49683094391661
#> 12            TRUE 324.032039877205 345.537387315328    3.29156780024473
#> 13            TRUE 692.071961716294 773.808430308635   0.119183312381005
#> 14            TRUE  313.24422999736 370.489672640811    8.93181880906207
#> 15            TRUE 308.711478737756 333.561324980086    2.78639942931711
#> 16            TRUE 743.903327654658 813.804275343056 -0.0781734752981005
#> 17            TRUE 321.547659479787 364.929934913246    7.34305678487249
#> 18            TRUE 343.740721140403  361.40224112652    4.58624737370291
#> 19            TRUE 722.264896081933 802.321244407347 0.00792792272260529
#> 20            TRUE 331.067617646037  367.54913970009     8.8994921685647
#> 21            TRUE 321.578360847918 350.915072196588    3.87368098127436
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.1918660762921 3372.30403098197  3457.4489042899    79716.6975611816
#> 2   36.6202833359477 1434.36726566532 1497.46266706294    40980.2558867023
#> 3   23.1966378740124  2188.7887860685 2204.38789531888    63213.9108186504
#> 4   15.4722337374157 7379.01023394748 7704.17330923981    139474.089383984
#> 5   31.1046235317233 2329.60919698622 2412.95463866453    55988.2990466592
#> 6    12.953425408762 4795.51188902703 4830.71566123106    128531.256398564
#> 7   7.87706308092143 16822.8009823855 17236.0670566826    351810.030542704
#> 8   14.2987396754747 5134.20832721998 5240.09225038303    115799.101055578
#> 9   5.54560524418448 7712.69446738802 7806.24123124856    216127.301653273
#> 10  1.41865246244522 19987.8101181583 20175.8956734114      467986.5363305
#> 11  11.3165970733461 6683.25185050417 6943.88147693074    163886.029618193
#> 12  4.40670516953686 9171.63336006979 9216.74947664376    261721.240116788
#> 13  1.55990343413013 20242.5393980584 20633.5621430644    472361.233205188
#> 14   10.508268920192 6921.08643492949 7049.34503374533    174153.325625578
#> 15  3.99320374840855 8982.17593721416 9027.28320602525    260784.027095147
#> 16 0.982523312745435 20516.1515986151 20871.0135044728     499691.06014277
#> 17  9.09451166080603 6781.24822107071 7022.67728835666    171129.771394321
#> 18  5.52155376725181 9299.56597650974 9336.23176113265    263459.413890416
#> 19 0.920591702701901 19658.8138190274 19890.3703943118    471749.001916767
#> 20  10.0923858302063 7037.57209263906 7247.26250611665    176693.722280217
#> 21  6.22078132797599 9246.68774617096 9275.88861202774    267174.900380409
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84203.6449724958 87.4419973578783 97.0565945226466    14.4186109070506
#> 2     44097.6010531936 29.5107467408407 35.7433605208116    26.4954948095875
#> 3     67914.9587923536 62.5952828483381 66.9064278317007    16.0522207729701
#> 4      156755.99221707 237.639076236641 246.572288366233    14.3634940063628
#> 5      59088.600014318 59.4894112002445 70.6710288142985     11.038611709038
#> 6     134130.101173252 199.399070746038 210.814313116115    11.6222097720038
#> 7     372512.527121106 480.829512399238 530.138027861009    6.00466355015008
#> 8     123320.611381234 184.922683083272 216.586202803398    11.8515053023192
#> 9     226856.940524947 277.486487381715 302.367278751511    3.96314039751596
#> 10    501499.655515125 620.910762399453 725.519651188371  -0.368706980727518
#> 11    172169.619313225 296.265649322618 349.550779903864    8.22685333577974
#> 12    273289.936502661 321.973017250151 347.596409942382    3.18479932871676
#> 13    506611.853914843 684.246129617027 781.634262407901 -0.0187579758715707
#> 14    181415.133373849 307.763283361285 375.970619276886    8.78088209629431
#> 15    269450.292195544 306.332238140086 335.940565577756    2.67085433493601
#> 16    530652.982875447 737.210683727045 820.496919270668  -0.179729550749077
#> 17    181624.221234341 317.394037363817 369.083557029215    7.17536429675119
#> 18    278676.113240561 342.049724545988 363.093237720936    4.49669676155461
#> 19    525370.874061216 714.599926561414 809.986213927865  -0.079454779615838
#> 20    184603.727503851 327.574705960011 371.042051386116    8.78527894564157
#> 21    272943.497758373  318.76952678262 353.723906261886    3.64895860765399
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.3468107027307   3364.151862261 3465.60107301087    79287.0962132898
#> 2   37.5049735955326 1428.32621659534 1503.50371613293     40681.786668634
#> 3   23.8209073294433 2187.29525433176 2205.88142705562    62763.8104807426
#> 4   15.5691139080902 7347.87759907907 7735.30594410823    137819.439112731
#> 5   32.8579643705987 2321.62931427234 2420.93452137841    55691.4617199685
#> 6   13.0697452216826 4792.14131509261 4834.08623516549    127995.196792476
#> 7   8.04067080691116 16783.2329539954 17275.6350850727    349827.876614985
#> 8    14.512575688663 5124.07050478948 5250.23007281353     115078.95644993
#> 9   5.68387887156328 7703.73786233754 7815.19783629904    215099.995803857
#> 10  1.57482950116905 19969.8019266979 20193.9038648718    464777.833429844
#> 11   11.586574681483 6658.29795010163 6968.83537733328    163092.919966541
#> 12  4.51347364106483 9167.31373188717 9221.06910482638     260613.59897346
#> 13   1.6978447223827 20205.1010501323 20671.0004909905    469081.918456391
#> 14  10.6592056329598 6908.80635631946 7061.62511235535    173458.046160318
#> 15  4.10874884278964 8977.85715615778 9031.60198708163    259954.278308939
#> 16  1.08407938819641 20482.1754586925 20904.9896443954    496726.620732194
#> 17  9.26220414892733 6758.13267207525 7045.79283735212    170124.983643681
#> 18  5.61110437940011 9296.05542266286 9339.74231497952     262002.49586753
#> 19  1.00797440504034 19636.6435086279 19912.5407047114     466614.99288166
#> 20  10.2065990531294 7017.49535092312 7267.33924783259     175936.38135455
#> 21  6.44550370159636 9243.89191858893 9278.68443960977    266622.587865497
#>    ci95.integral.bt.up
#> 1     84633.2463203876
#> 2     44396.0702712619
#> 3     68365.0591302613
#> 4     158410.642488323
#> 5     59385.4373410087
#> 6     134666.160779339
#> 7     374494.681048825
#> 8     124040.755986882
#> 9     227884.246374363
#> 10    504708.358415781
#> 11    172962.728964877
#> 12    274397.577645989
#> 13     509891.16866364
#> 14    182110.412839109
#> 15    270280.040981752
#> 16    533617.422286023
#> 17    182629.008984981
#> 18    280133.031263448
#> 19    530504.883096323
#> 20    185361.068429518
#> 21    273495.810273284