Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.3330151197039  15.4991565235164
#> 2                 FALSE          0.75 33.6887623244629   31.211217351845
#> 3                 FALSE          0.75  65.070300211199  20.5787554912976
#> 4                 FALSE          0.75  241.20412581162  14.9386882647236
#> 5                 FALSE          0.75 65.4963411851528  21.8263894413835
#> 6                 FALSE          0.75 205.367088159208  12.3118169292003
#> 7                 FALSE          0.75 503.485753318542   6.9893632969568
#> 8                 FALSE          0.75 201.257585100596  13.0950888232807
#> 9                 FALSE          0.75   291.2205793118  4.77837975231581
#> 10                FALSE          0.75 667.806152377113 0.537133475952834
#> 11                FALSE          0.75 326.489858830524  10.1534159655289
#> 12                FALSE          0.75 336.297604492394  3.94166497117744
#> 13                FALSE          0.75 732.973618534451 0.697779712999479
#> 14                FALSE          0.75 344.223669847546  10.0563619064627
#> 15                FALSE          0.75 317.975621758818  3.25895658725243
#> 16                FALSE          0.75 785.314705124916 0.594596483327733
#> 17                FALSE          0.75 336.124082075481  7.92790505429349
#> 18                FALSE          0.75 351.122602942745  5.02520323340103
#> 19                FALSE          0.75 783.606674910632 0.666335688631106
#> 20                FALSE          0.75 350.031333670088  9.54589676802421
#> 21                FALSE          0.75 340.158729348988  5.25945542232393
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.08561036634 2125.34011358003 81485.3695644872 2.29161145910268
#> 2  1462.73518053059  274.31145031534 42149.8897650162 2.87142270717503
#> 3  2198.19960664932 658.965069892007 66331.5090310424 1.64059892193997
#> 4  7563.38162711303 6040.88134584694 149196.756447227 3.00789751020381
#> 5  2348.13318362205 1129.05017535452 56434.4630790055 2.81256607256425
#> 6  4818.53886110869 3242.27858846601 130687.604217305 2.57308737520008
#> 7  17054.4405348365  15821.444899669 363298.266768195 12.7087534363955
#> 8  5181.28432015015  3825.7675628534 118546.796062345  8.2939679686891
#> 9  7758.20150966424 6270.63189795457 222488.231182498 6.91604345413409
#> 10 20043.5757833276 18691.1171672669 481369.091315417 20.7563938098887
#> 11 6801.97394188933  5530.5525850318 168143.741297529 9.48166911196195
#> 12 9181.90501890965 7627.77575351994 267240.820153318 7.72433862721109
#> 13  20474.674018104 19500.2271398463 492384.897933087 17.1311679547896
#> 14 6980.25154712441 5590.02581303577 177286.784169262 14.5078497907843
#> 15 9003.66253066652 7411.44228951072 265917.404762467 6.31015726580953
#> 16 20665.0209609952 19321.9442555525 511703.046613476 28.7957723138354
#> 17 6879.98105210932 6048.67155845674 174897.514184836 13.4668162150495
#> 18 9316.16706364287 7643.34624713733  270968.67412202 5.59767324104468
#> 19  19771.495659476 18630.2655549064 499077.768413944 19.0314288860811
#> 20 7126.53821534109 5793.50806823274 180080.429801455 16.3073600110147
#> 21 9254.97883847151 7472.50589974717 267113.732185054 7.49119770825382
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.750859409720075 24.2130179172927 74.8773254627064 1635.02959040772
#> 2   3.06730843543105 22.1545427221899 93.9880486227475   1466.967889085
#> 3   1.69071961797469 6.42984029642434 107.680707326136 1024.78057314866
#> 4  0.329023650058477 59.7673265816815 80.9497932501377 3105.46706843216
#> 5   2.43293892155653 28.0553239845785 132.794270617631 1250.65700172261
#> 6  0.657166825496663 8.85429068900767 131.299457747178 2910.30608023372
#> 7  0.491958982833063 145.576248429794 197.672848586143 8059.72307263797
#> 8  0.604326651797684 53.2236341414733 116.254005051756 2533.16711168027
#> 9  0.291386177727555  19.169286265595 52.9896347746292 2121.80615916392
#> 10 0.325821655605852 190.446697009785 268.608456717092 18091.2925143472
#> 11 0.512083448337903 83.6685337162998 184.553432324252 3068.06630808609
#> 12 0.491390232020146 13.3155294089087 158.710182830279 3769.98032516259
#> 13 0.321799357077902 94.8016695377909 258.343195016336 6955.68864150085
#> 14 0.594932335397349 56.2663938635664 202.862521981471 3202.13693398127
#> 15 0.391253478832441 16.8677544111955 94.4797911962448 3853.56565148562
#> 16 0.416368706984543 112.120621645507 284.508552330103 10909.6887627878
#> 17 0.604078386386705 120.655459152479   156.6120807581 4421.98703676362
#> 18 0.401643877539532 12.7327622738598  87.343684893547 3128.52130470187
#> 19 0.252021445346326 64.9704302088836 222.511847776752 8938.76133716717
#> 20 0.372092954654932 67.8283535602932 71.1646536181753 2531.57499216077
#> 21 0.612669403062447 8.57233797533918 150.001478960536 5037.83358125383
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE   88.56331426948 96.1027159699278    14.2639927945268
#> 2             TRUE   28.96527197116 38.4122526777658    26.1654949755609
#> 3             TRUE 62.3715149846077 67.7690854377902    17.7975217197293
#> 4             TRUE 236.256134407334 246.152117215905    14.3974443603774
#> 5             TRUE 60.8696699957846  70.123012374521     17.824204915423
#> 6             TRUE 201.134359427004 209.599816891412    11.2307775012583
#> 7             TRUE 482.579853915672 524.391652721413    6.18009077019641
#> 8             TRUE 187.614007792102 214.901162409089    12.1009714810735
#> 9             TRUE 279.843687829749  302.59747079385    4.29904948995398
#> 10            TRUE 633.661884559846  701.95042019438 0.00115685248120856
#> 11            TRUE 310.892513141347 342.087204519701    9.31103869301308
#> 12            TRUE 323.591067450632 349.004141534156     3.1333280395043
#> 13            TRUE 704.792847248822  761.15438982008   0.168419770606331
#> 14            TRUE 320.358256941706 368.089082753386    9.07769821473402
#> 15            TRUE 307.595413056561 328.355830461075    2.61534461457306
#> 16            TRUE 737.945659668657 832.683750581175 -0.0903300396618406
#> 17            TRUE 313.971169401725 358.276994749238    6.93419610868736
#> 18            TRUE 341.914430461227 360.330775424264     4.3644990548485
#> 19            TRUE 752.299974393029 814.913375428236     0.2517604110364
#> 20            TRUE 323.205726451968 376.856940888207    8.93380385761685
#> 21            TRUE 327.835709118911 352.481749579066    4.25161425428621
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7343202525059  3374.2551958924 3453.91602484029    78795.7458882666
#> 2   36.2569397281291 1426.29095775259 1499.17940330859    39736.7275874714
#> 3    23.359989262866 2187.62251936171 2208.77669393694    64645.7449882129
#> 4   15.4799321690698 7465.06437488616  7661.6988793399    144088.263119656
#> 5    25.828573967344 2301.98217566742 2394.28419157668    54377.1323111718
#> 6   13.3928563571423 4803.97355292527 4833.10416929211    125900.150715321
#> 7   7.79863582371719 16814.9676061695 17293.9134635035    350040.022313705
#> 8   14.0892061654879 5093.73144198743 5268.83719831287    114379.736163631
#> 9   5.25771001467763 7726.66803375734 7789.73498557114    218997.860050673
#> 10  1.07311009942446 19730.2909667466 20356.8605999087    451608.915129316
#> 11  10.9957932380448 6664.33920392602 6939.60867985264    163096.772220728
#> 12  4.75000190285058 9160.00097303199  9203.8090647873    261039.202518426
#> 13  1.22713965539263 20318.7252717144 20630.6227644937    480942.790117818
#> 14  11.0350255981913 6887.69332921884 7072.80976502998    172019.268912863
#> 15  3.90256855993179  8975.9150746601 9031.40998667294    259578.289265773
#> 16  1.27952300631731 20480.5825383883  20849.459383602     493756.60859869
#> 17  8.92161399989962 6681.50282180349 7078.45928241514    167623.345509359
#> 18  5.68590741195356 9295.22166970237 9337.11245758337    265822.256575785
#> 19  1.08091096622581 19664.6193017824 19878.3720171696    484373.506014304
#> 20  10.1579896784316 7014.96057373441 7238.11585694777    175915.988939351
#> 21  6.26729659036166 9240.87734250208 9269.08033444094    258826.495943891
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84174.9932407079 87.8414566598627 96.8245735795452    14.027472080465
#> 2      44563.051942561 28.0607738183998 39.3167508305259   25.1992928184001
#> 3      68017.273073872 61.8547263241966 68.2858740982013   17.2649450400672
#> 4     154305.249774797  235.30864669162 247.099604931619    14.293801910609
#> 5     58491.7938468392 59.9837116829269 71.0089706873787   17.0578291551327
#> 6      135475.05771929 200.323836903816 210.410339414601   11.0237699512268
#> 7     376556.511222684 478.576596583207 528.394910053878     6.025123690604
#> 8     122713.855961059 185.001407881965 217.513762319226   11.9106085857573
#> 9     225978.602314322 277.665134141697 304.776024481902    4.2072628439698
#> 10    511129.267501519 627.123620509731 708.488684244495 -0.101476969034635
#> 11    173190.710374331 307.905787371079 345.073930289969   9.14973240678664
#> 12    273442.437788211  321.15790078306 351.437308201728   2.97854011641795
#> 13    503827.005748356 699.396529343063 766.550707725838 0.0670529731267915
#> 14    182554.299425661 315.788284257608 372.659055437483   8.89029452908385
#> 15    272256.520259161 305.607713517831 330.343529999805   2.49209976874084
#> 16    529649.484628262 728.874991389799 841.754418860034 -0.221486182361972
#> 17    182171.682860312 309.729122293984 362.519041856978   6.74391141697555
#> 18    276115.091668254 340.151163390298 362.094042495193   4.23798123342355
#> 19    513782.030813584 746.305074293914 820.908275527351  0.172373655752308
#> 20     184244.87066356 318.068908048499 381.993759291676   8.81659457690055
#> 21    275400.968426217 325.475981840811 354.841476857166   4.05862339232154
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9708409665677 3366.62809524845 3461.54312548424    78280.7115672881
#> 2   37.2231418852899  1419.3122767951 1506.15808426608    39274.6327024096
#> 3    23.892565942528 2185.59711966833 2210.80209363032    64322.9391076711
#> 4   15.5835746188382 7446.23766701294 7680.52558721313      143110.0409931
#> 5   26.5949497276343 2293.14474861227 2403.12161863182    53983.1753556292
#> 6   13.5998639071737 4801.18445135824 4835.89327085915    124983.404300047
#> 7    7.9536029033096 16769.1110879141 17339.7699817589    347501.209545824
#> 8   14.2795690608042 5076.96599723286 5285.60264306744    113581.788523451
#> 9   5.34949666066181 7720.62970858367 7795.77331074481    218329.491110536
#> 10   1.1757439209403 19670.3002571885 20416.8513094668    445910.157987297
#> 11  11.1570995242712 6637.98361580538 6965.96426797328     162130.33133368
#> 12  4.90478982593692 9155.80658126819 9208.00345655111       259851.658716
#> 13  1.32850645287217   20288.86274581 20660.4852903981    478751.748195745
#> 14  11.2224292838415 6869.96941515182   7090.533679097    171010.595778659
#> 15  4.02581340576401 8970.60173202058 9036.72332931246    258364.416085555
#> 16  1.41067914901744   20445.26454257 20884.7773794204    490320.056638412
#> 17  9.11189869161144 6643.49635217046 7116.46575204818    166230.419592779
#> 18  5.81242523337851 9291.21084958611 9341.12327769964    264836.772364804
#> 19   1.1602977215099 19644.1536162666 19898.8377026854    481557.796193097
#> 20  10.2751989591479 6993.59464236291 7259.48178831926     175118.54281682
#> 21  6.46028745232633 9238.17705603984 9271.78062090317    257239.578365797
#>    ci95.integral.bt.up
#> 1     84690.0275616864
#> 2     45025.1468276228
#> 3     68340.0789544138
#> 4     155283.471901354
#> 5     58885.7508023818
#> 6     136391.804134563
#> 7     379095.323990565
#> 8     123511.803601238
#> 9     226646.971254459
#> 10    516828.024643538
#> 11    174157.151261378
#> 12    274629.981590637
#> 13    506018.047670429
#> 14    183562.972559866
#> 15    273470.393439379
#> 16     533086.03658854
#> 17    183564.608776892
#> 18    277100.575879235
#> 19    516597.740634792
#> 20     185042.31678609
#> 21    276987.886004312