Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.2707947897573  15.4547407650277
#> 2                 FALSE          0.75 32.4741825828648  31.6797285027831
#> 3                 FALSE          0.75 65.5251111152462  20.6929136027249
#> 4                 FALSE          0.75 242.450816515411  15.0158050189341
#> 5                 FALSE          0.75 65.0877398313435  22.4012962300589
#> 6                 FALSE          0.75 204.644054819502  12.4811541929722
#> 7                 FALSE          0.75 506.166726079974  7.01683489190215
#> 8                 FALSE          0.75 203.329259757074  13.2556987922197
#> 9                 FALSE          0.75  287.86361051588  4.56813269812229
#> 10                FALSE          0.75 682.496198437206 0.859391308925949
#> 11                FALSE          0.75 324.755985894098  9.98997024598248
#> 12                FALSE          0.75 334.524723782608  3.80516327153407
#> 13                FALSE          0.75 724.996364994571 0.731442621624983
#> 14                FALSE          0.75 340.685665901559  9.79958910597842
#> 15                FALSE          0.75 321.007689012141  3.47888358419838
#> 16                FALSE          0.75 773.899264673795 0.329745692707853
#> 17                FALSE          0.75 334.187975497669   7.9916735387071
#> 18                FALSE          0.75 350.805916564066  5.01897171071947
#> 19                FALSE          0.75 777.619906688953 0.670147571099361
#> 20                FALSE          0.75 352.526026672048  9.66908737933685
#> 21                FALSE          0.75 338.144037933817  5.21838908063639
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1    3415.259478859 2125.26927808098 81736.8060834833 2.49487570003499
#> 2   1471.1876649764 290.633929486737 42944.2113609474 2.09252625905742
#> 3   2196.0857511556 651.396875161874 66686.5685161454 1.23027043412363
#> 4  7532.13338121434 6004.07528076924 147504.834922221 2.75823243793514
#> 5  2369.72065599841  1126.1565324038 57433.6999986488 2.87977329662621
#> 6  4817.58386210329 3210.29354005599 130802.184045015 2.40610545497331
#> 7   17031.711255096 15820.9898485709 362040.047642849 11.5128620254466
#> 8  5189.62471191293 3813.66337076892 119060.755396764 7.21825251409138
#> 9   7759.2672962221 6282.83206543488 221085.491727586 7.48234253257763
#> 10 20066.1336448881 18600.4667326601 481166.484688637 26.3352713244339
#> 11 6800.36444157663 5564.76156368631  168368.40043749 11.0746423592272
#> 12 9188.80174165495 7660.90493238701 265671.328224566 5.67695517128012
#> 13 20463.3532261977 19418.9705798886 491135.178994883 22.3386690684727
#> 14 6965.36450458551 5616.20564413506 176714.595484519 9.43386312584078
#> 15 9007.85414170643 7393.86134696395 263737.011180468 6.45804240520773
#> 16 20670.0395958159 19475.3291978467 514073.105259365 20.8157761285605
#> 17 6926.85692253666 6061.89251974338 176874.712144915 13.3576241355214
#> 18 9314.18472755713 7634.59331264741 270707.850730623 7.91389224682865
#> 19 19813.7818390901 18599.6225301127 504535.892249249 27.8685587334316
#> 20  7119.9741702645 5764.35408413056 179610.407534841 14.3958945866759
#> 21 9255.73866786446 7466.35248277097 269667.500983879 7.00810567173876
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.667252608668275 16.5796232918684 55.4328581196434 1258.38059075303
#> 2   2.78790669157256 18.9397970644807 76.1752187544356 631.241134244876
#> 3   1.38198975278239   4.991109318066 80.2989685522541 777.581143762497
#> 4  0.282654627518491 88.8901276045776 97.9564829049495 4437.83785084683
#> 5   2.56848266536408 27.5623383295989 158.995327541169 895.371389995873
#> 6  0.298916521928997 10.6189265453865 59.0367439691232 2425.49726434685
#> 7  0.383185601578241  153.02793094209 177.418045559786 8279.64410703583
#> 8  0.607528759231514 43.6219712040406 128.046502242736 2240.19483308994
#> 9  0.341025956143583  22.828826621087 74.5774543801747 3515.44361918309
#> 10 0.565443699179838 110.364695500389 333.922544168622 13369.7063785034
#> 11 0.645887234484757  76.756249647147 179.621606294149 2339.95762898377
#> 12 0.354203160840614 13.8308062814783 112.242067794855 4441.22935971812
#> 13 0.263899855243448 73.3088161241634 170.603056932962 6834.45775178767
#> 14  0.79563658017195 45.2116260167908 253.364692129807 3182.16532555168
#> 15 0.472484232674802 16.8355459061628 125.639240416821  4041.9831314422
#> 16 0.221496960293321 107.657528726165 188.076947484012 11875.2299267775
#> 17 0.565120213764974 61.1438585401946 135.899876023849 2945.12928516972
#> 18 0.451530113148667 14.7754662246537 93.3319079699025 3470.39359949195
#> 19 0.418865385532655 91.7239312008025  345.28944691533 7920.53554415079
#> 20 0.433467133212513 100.952449335004 167.520478659117  3499.9660938102
#> 21 0.627593607273694 11.4868520389214 175.688013800461 2445.33686498655
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.1667242631998 95.3748653163149    14.3571102237684
#> 2             TRUE 29.0319768867153 35.9163882790142    27.0936219951463
#> 3             TRUE 63.5013162511129 67.5489059793796    18.4195404593979
#> 4             TRUE 237.913524155008 246.988108875814    14.5508381566662
#> 5             TRUE 60.3505127583934 69.8249669042937     18.176142245535
#> 6             TRUE 200.686011346071 208.602098292933     11.989436514399
#> 7             TRUE 487.228068048114 525.105384111834    6.38649457730595
#> 8             TRUE 191.455234371394 215.203285142754    12.2563139832839
#> 9             TRUE 275.555157049789  300.17206398197    4.00714500026609
#> 10            TRUE 639.174677108512   725.8177197659 -0.0707635762248853
#> 11            TRUE  306.53819921317 342.973772575027    8.92748574525506
#> 12            TRUE 325.186132525852 343.863315039364    3.22249907195126
#> 13            TRUE 688.249254376933 761.743475612208    0.29732735974951
#> 14            TRUE 325.166961059551 356.204370743567    8.49076693159556
#> 15            TRUE 310.384209255574 331.631168768708    2.70164702144834
#> 16            TRUE 739.657312942313 808.141216405277 -0.0346168069746601
#> 17            TRUE 312.214683794736 356.161267200602    7.06205078706372
#> 18            TRUE 337.787563818033 363.824269310099    4.27620467458991
#> 19            TRUE 731.776127572458 823.463685805448  -0.018885988101857
#> 20            TRUE 328.844780076966  376.20727326713    8.95603394520227
#> 21            TRUE 326.615704103807 349.672371763827    4.18599759667116
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.552371306287 3387.98599854388 3442.53295917412    79666.7700116946
#> 2     36.26583501042 1440.03169880533 1502.34363114747    41905.8196951146
#> 3   22.9662867460519 2187.87537632738 2204.29612598382    65407.4475346561
#> 4    15.480771881202 7385.90912130481 7678.35764112387    140204.591657578
#> 5   26.6264502145828 2324.38060944622  2415.0607025506    55960.8140621056
#> 6   12.9728718715454 4800.11572793613 4835.05199627045    126812.241045164
#> 7   7.64717520649836 16779.9803086962 17283.4422014957    348420.033086775
#> 8   14.2550836011555 5117.86656928228 5261.38285454358    115375.634896331
#> 9   5.12912039597848 7721.71387643041 7796.82071601379     215302.58697403
#> 10  1.78954619407678   19884.58372079 20247.6835689863    459173.317695999
#> 11  11.0524547467099 6674.10041090707 6926.62847224619    164519.170137811
#> 12  4.38782747111688 9166.05006532192 9211.55341798798     258365.50592783
#> 13  1.16555788350046 20342.7602236734 20583.9462287219    479892.495993192
#> 14  11.1084112803613 6890.99137978789 7039.73762938313    171479.933523987
#> 15  4.25612014694843 8980.15966869079 9035.54861472207    257087.948929245
#> 16 0.694108192390366 20492.9429610614 20847.1362305705    494538.352029816
#> 17  8.92129629035049 6826.27527523804 7027.43856983528    172029.974470811
#> 18  5.76173874684902 9289.87908561757 9338.49036949668    264999.053259458
#> 19  1.35918113030058 19662.8959722648 19964.6677059155    491506.611279121
#> 20  10.3821408134714 6953.90739110841 7286.04094942058    173852.963310523
#> 21  6.25078056460162 9236.84279626044 9274.63453946849    265644.921840976
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83806.8421552721 86.3808384176888 96.1607511618259   14.1469256520379
#> 2     43982.6030267802 28.3728311151122 36.5755340506173   26.2154313873009
#> 3     67965.6894976347 63.1137810643639 67.9364411661285   17.9842136872714
#> 4     154805.078186864 237.044680937058 247.856952093764   14.4618019489979
#> 5      58906.585935192 59.4433841699562 70.7320954927309   17.3670702059453
#> 6     134792.127044865 199.928088127754 209.360021511249   11.8952778099913
#> 7     375660.062198923 483.601516510099 528.731935649849    6.2657911128088
#> 8     122745.875897197 189.181484829455 217.477034684693   12.0649424241259
#> 9     226868.396481143 273.198219152027 302.529001879732   3.89972182408086
#> 10    503159.651681275 630.879066641315 734.113330233096 -0.248878341466534
#> 11    172217.630737168 303.049686870013 346.462284918184   8.72403126639236
#> 12    272977.150521303 323.397891646899 345.651555918317   3.11092507628646
#> 13    502377.861996573 681.212573620364 768.780156368777  0.214198905347824
#> 14    181949.257445052 322.195294174911 359.176037628207    8.2401414088414
#> 15     270386.07343169 308.349925897934 333.665452126348   2.55281448815577
#> 16    533607.858488914 733.100343461817 814.698185885774 -0.104388349467056
#> 17    181719.449819019 308.007032192047 360.368918803291   6.88403791972775
#> 18    276416.648201787 335.294687760282  366.31714536785   4.13397268894808
#> 19    517565.173219377 722.997531571427 832.242281806479 -0.150828584544644
#> 20    185367.851759159 324.310073282164 380.741980061933   8.81949179824032
#> 21    273690.080126782 324.408150817209 351.879925050425   3.98830561037995
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7625558780175 3382.76341720694 3447.75554051106    79270.3801256074
#> 2   37.1440256182653 1434.06566273001 1508.30966722278    41706.9787378274
#> 3   23.4016135181784 2186.30317689219 2205.86832541901    65162.5094743709
#> 4   15.5698080888704 7357.90873110937 7706.35803131931    138806.672734561
#> 5   27.4355222541725  2315.6984728724 2423.74283912443    55678.7720742569
#> 6    13.067030575953 4796.77076607433 4838.39695813225    126048.209406895
#> 7   7.76787867099551 16731.7765104495 17331.6459997425    345811.945193059
#> 8   14.4464551603135 5104.12564835301 5275.12377547285    114669.973523908
#> 9   5.23654357216371 7714.52279604477 7804.01179639943    214195.222233988
#> 10  1.96766095931843 19849.8188417074 20282.4484480689     454961.86018677
#> 11  11.2559092255726 6649.92219226822 6950.80669088504    163782.083484681
#> 12  4.49940146678167 9161.69336134325 9215.91012196665    256966.518679519
#> 13  1.24868633790214 20319.6679465943 20607.0385058011    477739.641801379
#> 14  11.3590368031154  6876.7497175926 7053.97929157842    170477.551446438
#> 15    4.404952680241 8974.85647173035 9040.85181168251    255814.724242841
#> 16 0.763879734882762 20459.0308395126 20881.0483521192    490797.654602881
#> 17  9.09930915768645 6807.01495979788 7046.69888527544    171102.258745982
#> 18  5.90397073249085 9285.22481375681 9343.14464135745    263905.879275618
#> 19  1.49112372674337 19634.0029339366 19993.5607442437    489011.642582713
#> 20  10.5186829604334 6922.10736956789  7317.8409709611    172750.473990973
#> 21  6.44847255089283 9233.22443786818 9278.25289786075    264874.640728506
#>    ci95.integral.bt.up
#> 1     84203.2320413593
#> 2     44181.4439840673
#> 3     68210.6275579199
#> 4     156202.997109881
#> 5     59188.6279230407
#> 6     135556.158683135
#> 7     378268.150092639
#> 8     123451.537269621
#> 9     227975.761221185
#> 10    507371.109190504
#> 11    172954.717390298
#> 12    274376.137769614
#> 13    504530.716188387
#> 14    182951.639522601
#> 15    271659.298118094
#> 16    537348.555915849
#> 17    182647.165543847
#> 18    277509.822185627
#> 19    520060.141915784
#> 20    186470.341078709
#> 21    274460.361239253