Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.5658177167295  15.6161971754015
#> 2                 FALSE          0.75 33.5249819029871  30.2061422818196
#> 3                 FALSE          0.75 64.8887445333687  20.2716349824232
#> 4                 FALSE          0.75 242.983575261632  14.9912360985286
#> 5                 FALSE          0.75 65.4530382918005   23.379457555201
#> 6                 FALSE          0.75 204.521353185011  12.2718235614659
#> 7                 FALSE          0.75 512.466597195983  7.18524771438502
#> 8                 FALSE          0.75 201.570403214162  13.0364977909669
#> 9                 FALSE          0.75 288.988750624302  4.64086164355454
#> 10                FALSE          0.75 673.379451164428 0.623577813697254
#> 11                FALSE          0.75 322.388797254602  9.83424540914113
#> 12                FALSE          0.75 334.934581778953   3.8356279804922
#> 13                FALSE          0.75 733.697363851329 0.916968945732117
#> 14                FALSE          0.75 346.046597545853    9.879895147172
#> 15                FALSE          0.75 320.226958940594  3.44266854312786
#> 16                FALSE          0.75  773.10187635822 0.489072704318116
#> 17                FALSE          0.75 336.669557695897  8.02663126585174
#> 18                FALSE          0.75 350.085342244717  4.93043614604398
#> 19                FALSE          0.75 770.273938102233 0.634938504498021
#> 20                FALSE          0.75 353.728201203825  9.56187109628747
#> 21                FALSE          0.75 334.237693328936  5.22383365651536
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3420.45042891579  2118.0779289982 82189.6856703796 2.42465509047731
#> 2  1470.68022745956 318.938411679877 42627.3890178795 2.44007018171543
#> 3   2195.0634509842 679.418747014818 66009.7396288248 1.52398545114555
#> 4  7572.93587600403 6054.84777030304 149066.368897046  2.8614934468036
#> 5  2362.96183670638 1090.24752871321 57324.6725917538 3.21579057689085
#> 6  4820.24412262583  3250.1028658761 131103.004249096 2.20677365028477
#> 7  17075.0476868959 15862.7505514468 364161.492202994 16.4840235919126
#> 8  5171.63761397652 3841.28857895248 117675.505566601 7.87112026526936
#> 9  7761.46187930019 6280.26787644807 221432.615263371 7.01158793120482
#> 10 20090.7981279304 18721.1530844368 485837.179760254 26.2308309280409
#> 11 6760.60512445024 5566.94775068861 166769.243275932  14.721051920969
#> 12 9188.47833764911 7650.07735535961 266308.488950006 4.90599021511316
#> 13 20422.7134499856 19298.6265396779 488724.491308771  26.804319151889
#> 14 6968.73602172968 5648.96869489986 176620.261071413 15.6768366837454
#> 15 9010.88446038174 7388.55280927314 264614.033388986 5.51085253555769
#> 16 20699.7437110606 19368.4202940414 516847.628068869 23.9228550489695
#> 17 6882.83456116899 6012.32652104797 175485.292549609 12.1607086533157
#> 18 9319.90656497607 7663.42531865008   270984.5925345 6.89837088247683
#> 19 19799.1774393903  18570.408631365 504120.976897768 28.8314066595047
#> 20  7148.8668172515 5810.81977890398 180687.138958622 12.6620039733262
#> 21  9257.5322513541 7432.29602434242 267698.223815356 9.05728499894355
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.567163903964987 15.0836199842998 51.8982832269152 830.988140842292
#> 2   2.24582804919643 21.5258354715368 75.7947126950206 992.263202410888
#> 3   1.27828834147965 6.95220828539375    84.3877851388 1254.26348146899
#> 4  0.293618744320448 47.3999215228853 77.4793835064146 2445.82563598925
#> 5   1.84710292942568 34.4261128460976 130.422324206426 1320.19855446605
#> 6  0.625602562530388 10.7458788036275 123.773945304491 1915.84990700647
#> 7   0.48705642655304 140.209472255723 141.748783439493 6928.63621094268
#> 8  0.728477595361311 47.8619897271285 155.143664265896 2660.97807219814
#> 9   0.31762699530394 18.5384620084821   65.03184026081 3502.27169831693
#> 10  0.27296731004554 88.6553490498357 198.730439372506 6534.60161376599
#> 11 0.472428270317614 72.8243818539036 156.068306300791  2861.0733533105
#> 12 0.412699242384584 15.9185978550947 104.074293250448 4268.36603085491
#> 13 0.382262040892129 98.7653232765934 195.781281507729 11482.9940384469
#> 14 0.598886498025459 60.0746725623997 200.080675298637 3057.29137512871
#> 15 0.565819426070428 14.8038655102211 151.303677619952 3083.84277134161
#> 16 0.346346651240413 76.6310834478147 198.755500869992 5540.11392273626
#> 17 0.564471055812419 98.6206140107504 163.113912048625 3640.61689795818
#> 18  0.40993074136976 10.3970740280478 87.1209429984488 3774.34959505047
#> 19 0.453096661844727  80.131518520293 271.957714738961 9287.79851187981
#> 20 0.345308822481914 101.281477867315 136.546615067388  3557.7546643302
#> 21 0.726825767846637 9.14835844406152  203.50674151165 4734.35809781475
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.5772600928944 95.5543753405647    14.6832125533791
#> 2             TRUE 29.5110664540652  37.538897351909    26.5117551408914
#> 3             TRUE 62.3817884662342 67.3957006005031    18.1688506606892
#> 4             TRUE  238.27641854164 247.690731981623    14.5082332641215
#> 5             TRUE  60.163062792815 70.7430137907859    20.3409732362957
#> 6             TRUE 200.891210530293 208.151495839729    11.2427073461034
#> 7             TRUE 485.350378387287 539.582816004679    6.38403989270527
#> 8             TRUE 188.622410377794  214.51839605053    11.8381521465976
#> 9             TRUE  277.45468847747 300.522812771134    4.11836523627956
#> 10            TRUE 630.229734287801 716.529168041055    0.17454658867234
#> 11            TRUE 298.172666844608 346.604927664596    9.05710090446865
#> 12            TRUE 326.864227875092 343.004935682814    3.15673772676956
#> 13            TRUE 689.604258846471 777.790468856186   0.288147888464564
#> 14            TRUE 320.258201201091 371.834993890614    8.89472685792012
#> 15            TRUE 311.161606519602 329.292311361586    2.51189558724201
#> 16            TRUE 733.748779802665 812.454972913775 -0.0806675369723633
#> 17            TRUE 316.665191961192 356.673923430601    7.09807637904031
#> 18            TRUE 338.737522143043 361.433162346392    4.25610007649072
#> 19            TRUE 722.846274147348 817.701602057118  -0.110405504236555
#> 20            TRUE 332.899204667703 374.557197739946    8.99383808330472
#> 21            TRUE 319.338459505674 349.136927152198    4.02820526840764
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5491817974239 3395.63787404162 3445.26298378997     80822.710178694
#> 2   33.9005294227477 1435.27022810888 1506.09022681024    40995.1160499136
#> 3   22.3744193041573 2183.62706835473 2206.49983361368    63946.4762018083
#> 4   15.4742389329357 7494.96300509889 7650.90874690918    145042.985725843
#> 5   26.4179418741062 2306.33088107454 2419.59279233821    55152.9459696571
#> 6   13.3009397768284 4802.56715199386 4837.92109325779    127951.431152071
#> 7   7.98645553606477 16844.4031050352 17305.6922687566    352763.885635993
#> 8   14.2348434353363  5092.9046408754 5250.37058707765    113298.196637835
#> 9   5.16335805082953 7730.96610929624 7791.95764930414     215671.37831964
#> 10  1.07260903872217 19944.9600787435 20236.6361771174    475087.760105609
#> 11  10.6113899138136 6640.80901630057 6880.40123259991    162062.777609736
#> 12  4.51451823421484 9162.29224417748 9214.66443112074     259287.02682925
#> 13  1.54579000299967 20260.2444931956 20585.1824067756    469834.966115526
#> 14  10.8650634364239 6869.91318536454 7067.55885809483    171591.016759327
#> 15  4.37344149901372 8986.53210161742 9035.23681914605    259541.112030129
#> 16  1.05881294560859 20573.6855787889 20825.8018433322    507734.140665967
#> 17  8.95518615266317 6720.60365112131 7045.06547121668    169496.477752468
#> 18  5.60477221559723 9302.80337819993 9337.00975175221    264775.787450642
#> 19   1.3802825132326 19667.3610914244 19930.9937873562    488842.548345725
#> 20  10.1299041092702 6982.25878615977 7315.47484834323    174834.632535799
#> 21  6.41946204462307 9242.48320171362 9272.58130099458     259910.20474445
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83556.6611620652  86.813493739394 96.3181416940651   14.5045559236301
#> 2     44259.6619858454 28.7424443468249 38.3075194591493   25.8043193053946
#> 3     68073.0030558413 61.9017330491234 67.8757560176139   17.7661898331231
#> 4     153089.752068248 237.375048105896 248.592102417367   14.4157433596605
#> 5     59496.3992138504 59.1500887610944 71.7559878225065   19.7591358135267
#> 6     134254.577346122 200.196076830453 208.846629539569   11.0456425389063
#> 7     375559.098769995 480.157910955834 544.775283436131   6.23061711834106
#> 8     122052.814495367 186.143007494234  216.99779893409   11.6086817040588
#> 9     227193.852207102 275.246038279141 302.731462969464   4.01831273275882
#> 10    496586.599414899 621.967022545468 724.791879783388 0.0885618860079948
#> 11    171475.708942127 293.535535489503 351.242059019701    8.9082859993186
#> 12    273329.951070763 325.318840957332 344.550322600575   3.02673746541842
#> 13    507614.016502017 681.160898313626 786.233829389031  0.167735345583543
#> 14      181649.5053835 315.319997645712 376.773197445993    8.7060776110421
#> 15    269686.954747843 309.425687970901 331.028229910287   2.33366246802982
#> 16     525961.11547177  726.21308046224   819.9906722542 -0.189766732113093
#> 17     181474.10734675 312.834568735398 360.504546656395    6.9202679964594
#> 18    277193.397618358 336.564535315063 363.606149174372   4.12697189295925
#> 19     519399.40544981 713.764381049604 826.783495154862 -0.253130952717645
#> 20    186539.645381445 328.910673416106 378.545728991544   8.88506580422291
#> 21    275486.242886261 316.485414731007 351.989971926865   3.79925515153595
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7278384271728 3390.88653374657 3450.01432408502    80560.9489143287
#> 2   34.6079652582446 1428.48958993535 1512.87086498377    40682.5531411542
#> 3   22.7770801317234 2181.43712274483 2208.68977922358    63551.3832051456
#> 4   15.5667288373967 7480.03202981918 7665.83972218889    144272.550650507
#> 5   26.9997792968753 2295.48665552802 2430.43701788473    54737.0834250003
#> 6   13.4980045840254 4799.18220017072 4841.30604508094    127347.938431364
#> 7   8.13987831042898 16800.2371212747 17349.8582525171    350581.365229546
#> 8   14.4643138778751 5077.82811411135 5265.44711384169    112459.988545093
#> 9   5.26341055435027 7725.12649376356 7797.79726483681     214568.16273467
#> 10  1.15859374138651 19917.0336437928 20264.5626120681    473029.360597273
#> 11  10.7602048189636 6617.86933601659 6903.34091288389    161161.539503443
#> 12  4.64451849556599 9157.27788585313  9219.6787894451    257942.491529531
#> 13  1.66620254588069 20229.1334163635 20616.2934836077    466217.822993415
#> 14  11.0537126833019 6850.98966350738 7086.48237995199    170627.969976161
#> 15   4.5516746182259  8981.8688839817 9039.90003678177    258569.701557156
#> 16  1.16791214074932 20549.5467875029 20849.9406346183    505989.004780306
#> 17  9.13299453524408 6689.53815770792 7076.13096463006    168349.683429611
#> 18  5.73390039912871  9299.5282998811 9340.28483007105    263586.867328201
#> 19  1.52300796171369 19642.1196630905   19956.23521569    485916.891814483
#> 20   10.238676388352 6950.35512063156 7347.37851387144    173713.939816534
#> 21  6.64841216149477 9239.60146880374 9275.46303390446    258418.881943639
#>    ci95.integral.bt.up
#> 1     83818.4224264305
#> 2     44572.2248946048
#> 3      68468.096052504
#> 4     153860.187143585
#> 5     59912.2617585072
#> 6     134858.070066829
#> 7     377741.619176442
#> 8     122891.022588109
#> 9     228297.067792072
#> 10    498644.998923236
#> 11     172376.94704842
#> 12    274674.486370482
#> 13    511231.159624127
#> 14    182612.552166666
#> 15    270658.365220815
#> 16    527706.251357432
#> 17    182620.901669607
#> 18    278382.317740799
#> 19    522325.061981052
#> 20    187660.338100709
#> 21    276977.565687072