Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4844938363185  15.6543603488896
#> 2                 FALSE          0.75 31.7076938885484  31.6800718538508
#> 3                 FALSE          0.75 65.3122907744685  20.8277299961184
#> 4                 FALSE          0.75 242.780967417896  15.1419433242986
#> 5                 FALSE          0.75 66.2730025429974  22.6067517138089
#> 6                 FALSE          0.75 205.056868250043  12.3315261991328
#> 7                 FALSE          0.75 509.272182552284   7.0935252058103
#> 8                 FALSE          0.75 202.748097070933  13.2661663901411
#> 9                 FALSE          0.75 286.435854596202  4.53628263556726
#> 10                FALSE          0.75  666.22825279687 0.527970384930579
#> 11                FALSE          0.75 326.889941316287  9.79302406792295
#> 12                FALSE          0.75  335.92894003029   4.0168575733966
#> 13                FALSE          0.75 738.682561842288 0.887328944842808
#> 14                FALSE          0.75 347.044506094271  9.58458461799686
#> 15                FALSE          0.75 319.127392206358  3.29147389017642
#> 16                FALSE          0.75 775.522232750357 0.370273056960966
#> 17                FALSE          0.75  334.58782348793  8.20106912365054
#> 18                FALSE          0.75 350.019539945649   4.9261504668008
#> 19                FALSE          0.75 779.612442011912 0.613649130332815
#> 20                FALSE          0.75 355.413868855564  9.74458868907018
#> 21                FALSE          0.75 337.356997260909  5.14761954555742
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3408.34166976092 2111.13986499648 81036.5375539441 2.21838391067456
#> 2  1453.10686627129 255.296743198083 42432.3383072885 2.19138923677203
#> 3  2194.33981939771 643.451459861996 66644.2149873263 1.56951567627034
#> 4   7546.6254327753 5989.48987969706 148627.244218919 2.36182730108927
#> 5  2369.19989195291 1084.18793745503 56987.9563588191 2.43935252930171
#> 6  4822.61574111864 3244.92004673312 129671.064930374 3.18971208615407
#> 7  17029.4381418949 15829.3504682458  361766.27832967 15.1822378593818
#> 8    5173.825945047 3785.68312987127 118403.193614446 8.57898137160202
#> 9  7767.85596552293 6288.65050479861  220899.94654647 6.99338384771466
#> 10 20100.6164982491 18735.1544185497 485012.422082521  25.348982436846
#> 11 6785.20471203567 5584.47475639734 167721.872862808 9.99109061482769
#> 12  9188.7696520946 7600.34518097089 265778.633507672 8.46056434939766
#> 13 20472.3742460725 19379.0412596338  492317.23770415 23.7034706341484
#> 14 6964.89033982183  5747.0189805301 175969.536485239 12.6729092749235
#> 15 9000.09669236175 7422.02914058015 265683.868830722 5.98183944275346
#> 16 20628.3622967666 19403.1584132237 510484.714977116 15.1915900547635
#> 17 6854.29980741917 5917.95619173232 174513.622002784 11.7641738205958
#> 18 9319.40418839267 7663.78691522784 270073.084211807 5.48773152286997
#> 19 19787.1816315658 18669.0715947171 502571.356146799 29.8580380347753
#> 20 7144.84694658093  5770.3700910986 179923.314174418 13.8364730029499
#> 21 9255.16411899596 7485.50861256838 268738.587882838 6.60470039439244
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.69938232016967 28.0064540334754 68.3292345095629  1647.8546492982
#> 2   2.27098319565043 23.1270722171404 76.4368247092966 1242.29502261379
#> 3   1.94804091782455 5.04999502931117 127.399625570135 910.656780615961
#> 4  0.278135309661836 54.1060030057676 68.9938166467227 3139.94666788994
#> 5   2.14791402767311 27.4276976119229 152.616522128332 1014.83560108489
#> 6  0.427808688916575  10.031221852781 88.5024759102749 2654.98944249017
#> 7  0.571180322516084 152.170886317773  168.67859215094 8521.04435291632
#> 8  0.748605135450299 37.4668914578733 137.255335854399 2262.57005741324
#> 9  0.413442281230526 18.1325139499973 65.6146897570376 4158.21121273362
#> 10 0.279373908840387 89.0742799893333 270.355754325287 6683.16570764033
#> 11 0.796495246439449 92.8337431543738  228.45621152387 3936.02286084944
#> 12 0.682539971716956 12.0835628077485 178.507984456828 3134.24100593146
#> 13 0.406258908241624 75.2862703681445 271.336948398939 6270.06116733908
#> 14 0.509860841090344 48.9855094058971 155.075854275241 2739.90330011884
#> 15 0.329557896389662 16.1554679467594 68.1174142032063 2843.00973643921
#> 16  0.21438981776522 105.030182842004 165.052108449719 12431.2563520026
#> 17 0.579984050193243  116.88772847621 207.048853418903 5345.03870426829
#> 18 0.324652055443586 10.5876370207001 60.0258819657099 4209.75844660572
#> 19  0.26275512742078 82.9414876715721 206.159199894178 7466.62529721102
#> 20 0.436282051137041  92.489229760561 137.682814128686 3669.62263875744
#> 21 0.536454179040823 12.0542631105133 154.784122256971 4375.17280779161
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.8352523032589 96.1337353693782   14.5038764322105
#> 2             TRUE 28.1028585940584 35.3125291830384   27.9443044970059
#> 3             TRUE 62.7304374870038 67.8941440619332    17.623202686297
#> 4             TRUE 238.895761507604 246.666173328187   14.6844107399049
#> 5             TRUE 62.2602676322961 70.2857374536987   19.0734331382866
#> 6             TRUE 199.809791868319 210.303944631766    11.627780905865
#> 7             TRUE 484.297401273601 534.246963830967   6.15393357527134
#> 8             TRUE 188.635672714648 216.860521427218   12.0347109423254
#> 9             TRUE 274.931738166711 297.939971025692   3.85617008294305
#> 10            TRUE 624.529176688259 707.927328905482  0.068400304888142
#> 11            TRUE 310.454597254896 343.325285377679   8.48278938753006
#> 12            TRUE 322.011311675531  349.84656838505   2.89407931992221
#> 13            TRUE 699.690352649114 777.674771035462  0.219033040785337
#> 14            TRUE 326.197570337022  367.89144185152   8.74586353440325
#> 15            TRUE 309.287266323029 328.967518089688   2.74935115061542
#> 16            TRUE 750.532067110272 800.512398390443 0.0176018067371793
#> 17            TRUE  315.23575755305  353.93988942281   7.24699536108265
#> 18            TRUE 340.992221590528  359.04685830077    4.3920978355961
#> 19            TRUE 730.495969444707 828.728914579118  0.181416945725632
#> 20            TRUE 332.652870765712 378.174866945417   9.02690471494975
#> 21            TRUE 326.492265112133 348.221729409684   4.26515242103526
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.8048442655687 3362.27105287585 3454.41228664598    78325.8166558486
#> 2   35.4158392106958  1415.0628324741 1491.15090006849    40388.7629950889
#> 3   24.0322573059398 2186.03257757449 2202.64706122092     65146.184583213
#> 4   15.5994759086924 7457.62105783081 7635.62980771979     143462.03195024
#> 5   26.1400702893312 2324.08132938129 2414.31845452452    55318.5517950344
#> 6   13.0352714924005 4806.11438117081 4839.11710106646    125303.607297478
#> 7   8.03311683634925 16779.1170339022 17279.7592498876    347749.160369122
#> 8   14.4976218379569  5112.1929085988  5235.4589814952    114681.265870001
#> 9   5.21639518819148 7738.02798007519 7797.68395097068    214059.689101524
#> 10 0.987540464973016 19954.0893076667 20247.1436888316    474018.614493453
#> 11  11.1032587483158 6632.49320454673 6937.91621952462     161247.11525671
#> 12    5.139635826871 9168.89219127586 9208.64711291335    260622.807052915
#> 13  1.55562484890028 20348.5283313169 20596.2201608281    482002.987083877
#> 14  10.4233057015905 6884.30917684913 7045.47150279453    171462.395556544
#> 15  3.83359662973741 8973.52094758933 9026.67243713417     261007.11781428
#> 16 0.722944307184753 20455.5876459915 20801.1369475417    490035.298278071
#> 17  9.15514288621842  6662.0194940758 7046.58012076253    165721.033334262
#> 18   5.4602030980055 9301.98752549362 9336.82085129172    263148.031567141
#> 19     1.04588131494 19650.7428843461 19923.6203787856    490288.757532886
#> 20  10.4622726631906 6992.70216362481 7296.99172953706    173886.784933662
#> 21  6.03008667007957 9235.33485617916 9274.99338181275    261541.428614021
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83747.2584520397 88.1364613713964 96.8325263012407     14.283571001357
#> 2     44475.9136194882 27.4125709844752 36.0028167926216     27.228944790376
#> 3     68142.2453914395 62.2360400489787 68.3885414999584    17.0095697971823
#> 4     153792.456487598 238.151785907761 247.410148928031    14.5967981173614
#> 5     58657.3609226037 61.4918715855661 71.0541335004287    18.3968402195696
#> 6     134038.522563271 198.805032561181 211.308703938905    11.4930211688563
#> 7     375783.396290217 479.514996347896 539.029368756672    5.97401177367877
#> 8      122125.12135889 185.933293582593 219.562900559273    11.7989003246586
#> 9     227740.203991417 272.728822254681 300.142886937723    3.72593576435543
#> 10    496006.229671589 616.544247220652 715.912258373089 -0.0196024763965799
#> 11    174196.630468905 307.307403711225  346.47247892135    8.23189338490163
#> 12    270934.459962429 319.346233905471  352.51164615511    2.67907922883137
#> 13    502631.488324422 692.223759399357 785.141364285219  0.0910614846892257
#> 14    180476.677413935 322.205603915421 371.883408273121    8.58525736945979
#> 15    270360.619847165 307.402986898562 330.851797514155    2.64554041325268
#> 16     530934.13167616 745.746716243021 805.297749257694  -0.049930985858865
#> 17    183306.210671305 311.530042799562 357.645604176297    7.06430038527178
#> 18    276998.136856474 339.263586160824 360.775493730474    4.28983243813137
#> 19    514853.954760711 721.090687463753 838.134196560072  0.0986490805880859
#> 20    185959.843415174 328.294381769782 382.533355941346    8.88947586884158
#> 21    275935.747151655   324.4117844879 350.302210033918     4.0961693546374
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0251496964222 3353.44901985531 3463.23431966653    77806.7424413197
#> 2   36.1311989173257  1407.7778047257 1498.43592781689    39997.4400629655
#> 3   24.6458901950545 2184.44182914026 2204.23780965516     64859.327697319
#> 4   15.6870885312358 7440.57766688399  7652.6731986666    142472.948749855
#> 5   26.8166632080482 2315.44160463354 2422.95817927228    54998.8785806927
#> 6   13.1700312294093 4802.95454628718 4842.27693595009    124467.285623094
#> 7   8.21303863794182 16731.1832047121 17327.6930790777    345065.031397954
#> 8   14.7334324556237 5100.39083778957 5247.26105230443    113968.556301916
#> 9   5.34662950677909 7732.31623818094 7803.39569286493    212749.852569513
#> 10  1.07554324625774   19926.03090947 20275.2020870282    471913.417295546
#> 11  11.3541547509443  6603.2505754531 6967.15884861825    160007.268055543
#> 12  5.35463591796184 9165.08586899142 9212.45343519779    259635.521136047
#> 13  1.68359640499639 20324.8131561509 20619.9353359941    480027.917816165
#> 14  10.5839118665339 6868.87874138627 7060.90193825739    170599.326017006
#> 15  3.93740736710015  8968.4319751861  9031.7614095374    260111.569747301
#> 16 0.790477099780797 20422.5031383963 20834.2214551369    486119.452527191
#> 17  9.33783786202929 6625.19985960579 7083.39975523254    164037.346142418
#> 18  5.56246849547022  9298.6524198321 9340.15595695325     261821.95765646
#> 19  1.12864918007754 19624.6163157295 19949.7469474021    487936.770564265
#> 20  10.5997015092988 6963.56805625023 7326.12583691163    172730.853802454
#> 21  6.19906973647743 9231.53776329935 9278.79047469256    260163.249179566
#>    ci95.integral.bt.up
#> 1     84266.3326665686
#> 2     44867.2365516116
#> 3     68429.1022773335
#> 4     154781.539687983
#> 5     58977.0341369455
#> 6     134874.844237655
#> 7     378467.525261386
#> 8     122837.830926976
#> 9     229050.040523428
#> 10    498111.426869496
#> 11    175436.477670073
#> 12    271921.745879298
#> 13    504606.557592134
#> 14    181339.746953472
#> 15    271256.167914143
#> 16     534849.97742704
#> 17    184989.897863149
#> 18    278324.210767155
#> 19    517205.941729332
#> 20    187115.774546383
#> 21    277313.926586109