Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.7113698855767  15.6567959025837
#> 2                 FALSE          0.75 33.0237637207245  31.2426256350954
#> 3                 FALSE          0.75 64.6448874906704  20.8722094502498
#> 4                 FALSE          0.75 243.039434294457  15.0138474968982
#> 5                 FALSE          0.75 65.3800227330826  22.2894162472991
#> 6                 FALSE          0.75 205.355167806404  12.4773493168925
#> 7                 FALSE          0.75 504.686641391715  6.93771874339948
#> 8                 FALSE          0.75 203.116370858382  13.3967518185521
#> 9                 FALSE          0.75 290.926565271363  4.89275561345617
#> 10                FALSE          0.75 674.651802630847  0.69577616169344
#> 11                FALSE          0.75 325.967082067139   9.9350388460172
#> 12                FALSE          0.75 334.346223573799  3.85631182251196
#> 13                FALSE          0.75 722.174464985743  0.70417136347944
#> 14                FALSE          0.75 343.706775069879  9.68422801275717
#> 15                FALSE          0.75 321.066753258835  3.38278094937656
#> 16                FALSE          0.75 779.323279392898 0.591633342439984
#> 17                FALSE          0.75 335.061198899067  8.06969167387561
#> 18                FALSE          0.75 351.733906001469  5.12905873218054
#> 19                FALSE          0.75 772.292837928185 0.558579420359934
#> 20                FALSE          0.75 350.882879615549  9.59229333217424
#> 21                FALSE          0.75 337.907210456033  5.12351284343579
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3417.21985197898 2118.62901203943 81750.2512859459 2.00908427108629
#> 2  1458.80282220448 259.339148309782 42134.9930673494 2.83422241969815
#> 3  2199.15669332778 643.182718727493 65780.6204603153 1.34756476039162
#> 4  7564.94141063742   6039.054829607 149124.448150867 2.04167699769201
#> 5  2353.67981070768 1128.08521431467 56676.0470537578 2.92806172816846
#> 6  4820.35461996424 3211.94240189024 130223.140699199 2.35802476151393
#> 7  17010.7841883724 15815.5892895179 360622.623801822  12.799397326151
#> 8  5179.51800762435 3762.32406682214 118842.847616848 6.83213791529264
#> 9  7768.24661727547 6235.09636440954 222759.494662244 5.74117727971647
#> 10  20085.692889606 18651.9316866953 485848.548770855 43.8264531818405
#> 11 6758.75893057615 5537.18256616163 166354.182543797 8.70378147761941
#> 12 9190.03975857132  7640.4282464644 265706.305929215 6.49738260215059
#> 13 20473.4372575496 19432.1925417189 492557.070658295 19.4709797480512
#> 14 6971.00134553028 5673.00927999113 176630.642929887 15.3715039372045
#> 15 8997.99593375365 7408.83819653186 265141.441921213 5.22802311062295
#> 16 20704.3298873806 19277.6001836703 516544.792527814  23.998660351044
#> 17 6893.15885931508 6011.89544772396 175623.291069122 17.3486481503416
#> 18 9321.69835096807 7619.73795499516 272170.164390457 7.06309972686647
#> 19 19759.6599537647 18612.2237195149 503550.981557492 24.5581444786474
#> 20 7123.56934666395 5784.87895829357 179814.384602171 16.0431278124404
#> 21 9259.30614417442 7501.91994244868 269177.078330636 7.05750264245794
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.564439905594154 13.9081317887855 52.5091259777114 1039.31214933385
#> 2   2.54538950299341 22.9528094906641 98.2735141843923 1399.90839890965
#> 3     1.343345809341 5.21737157929601 92.5129277259937 1284.72973044841
#> 4  0.296302556765492 44.2743549433245 67.9224113161059 2462.93171469912
#> 5   2.12229387889231 48.9544759039716 141.849988826362 1844.74145948091
#> 6  0.384038968275355 12.1890954805547  82.158149010177 3197.25062510553
#> 7  0.475503837890693 131.417677115042 159.745806541347 8332.96103431918
#> 8  0.661147408669942 57.7820819977437 116.618603281982  3350.5454745177
#> 9  0.341439703615668 20.8494225889513 81.8526628740527 2166.55448105118
#> 10 0.597074954494463 53.8109734708514 273.200314425239 6930.14210760473
#> 11 0.809878333676073 120.321851649241 262.264150249324  4873.3339057776
#> 12 0.607561384940043 12.0113519276822 175.518527242759 5722.47284081525
#> 13 0.208571694209437 77.0690085690778 162.362796277898 7844.30827721418
#> 14 0.921017223310909 42.9401003561491 256.211278858309  3467.6392892083
#> 15 0.357040582220196 18.1798107613595 82.0151115350117 3113.53839845916
#> 16 0.729107985327217 60.6318967146153 576.653354846994 5973.34656922254
#> 17 0.613271822245267  104.16788618237 179.915689348741 4626.94822961171
#> 18  0.43253554812077 8.22677292556438 129.710334050164 2693.29915127506
#> 19 0.274944350495397 76.7299288695009 154.487360360821 9081.64573423707
#> 20 0.395388067561226 119.779063804924 141.326372843003  4545.9215172086
#> 21 0.442437290040198 8.92028995967355 118.067608970881  3449.3764569034
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.4064262596397 96.0163135115136   14.7282922578814
#> 2             TRUE  28.361467840321  37.686059601128   27.0554599026713
#> 3             TRUE 62.4281434598262 66.8616315215146   18.6624055938839
#> 4             TRUE 239.680875633254  246.39799295566   14.5264297910189
#> 5             TRUE 60.5633611902454 70.1966842759197   18.7982428165213
#> 6             TRUE 201.476217073714 209.234118539094   11.8456052140795
#> 7             TRUE 483.631632790197 525.741649993234   6.15551493006929
#> 8             TRUE 191.877503987726 214.355237729039     12.30916433129
#> 9             TRUE 281.482328646229 300.370801896497    4.3310873010084
#> 10            TRUE  602.55728714672 746.746318114975 -0.286412138449951
#> 11            TRUE 311.649361536456 340.284802597823   8.60278898712006
#> 12            TRUE 323.658029193262 345.034417954337   2.85687334428559
#> 13            TRUE 690.144703300199 754.204226671287  0.361070926504916
#> 14            TRUE 318.420651093178 368.992899046581   8.16915468041072
#> 15            TRUE 312.466655241861  329.66685127581   2.79544919162434
#> 16            TRUE  739.84548311543 818.801075670365 -0.607749293423287
#> 17            TRUE 306.522672691755 363.599725106379   7.06085952628215
#> 18            TRUE 340.115106950773 363.352705052164   4.41753775552187
#> 19            TRUE 731.894690260811  812.69098559556  0.106295963795006
#> 20            TRUE 324.491934364084 377.273824867013   8.94187996103602
#> 21            TRUE  326.29761860919 349.516802302877   4.39570350131966
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5852995472861 3394.34097518643 3440.09872877153    80040.5828002917
#> 2   35.4297913675196 1421.04545059234 1496.56019381662    39832.1437511431
#> 3   23.0820133066158 2190.57411707984 2207.73926957572    63667.2400537277
#> 4   15.5012652027774 7492.11009675565 7637.77272451919    145072.925480187
#> 5    25.780589678077 2273.14969784565 2434.20992356972    53641.4473529117
#> 6   13.1090934197054 4800.30355789873 4840.40568202976    124963.663420901
#> 7   7.71992255672967 16794.6021095182 17226.9662672267    346914.902900367
#> 8   14.4843393058141 5084.46648273806 5274.56953251064    113331.200311266
#> 9   5.45442392590394 7733.94931711664 7802.54391743429    219195.512540915
#> 10  1.67796446183683 19997.1738382464 20174.2119409655    474448.465003846
#> 11  11.2672887049143 6560.82948461315 6956.68837653915    158337.548268792
#> 12  4.85575030073833 9170.28108465028 9209.79843249235    256292.838106074
#> 13  1.04727180045396 20346.6587384535 20600.2157766457    479653.183542278
#> 14  11.1993013451036 6900.36488044441 7041.63781061614    170926.376299139
#> 15  3.97011270712879 8968.09014505121 9027.90172245608    260019.671255748
#> 16  1.79101597830326  20604.590417285 20804.0693574761    506718.637421443
#> 17  9.07852382146908 6721.80268654508 7064.51503208508    168011.961231411
#> 18  5.84057970883921 9308.16530950552 9335.23139243063    267739.687286609
#> 19  1.01086287692486 19633.4392207744 19885.8806867551    488611.674324672
#> 20  10.2427067033125 6926.53278670485 7320.60590662305    172336.343706363
#> 21  5.85132218555192 9244.63226719076 9273.98002115808     263502.85405903
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1        83459.9197716 88.7735647142476 96.6491750569058   14.5504936876192
#> 2     44437.8423835558 27.4686877781161 38.5788396633329   26.2536622092283
#> 3      67894.000866903 62.0036605603028 67.2861144210379   18.2392516639415
#> 4     153175.970821547 239.037747378981 247.041121209933   14.4330944856378
#> 5     59710.6467546039 59.6410217458724 71.1190237202928   18.1297202446702
#> 6     135482.617977498 200.733439273837 209.976896338971   11.7246329390728
#> 7     374330.344703277 479.599822632459 529.773460150971   6.00573122113372
#> 8     124354.494922429 189.725380544409 216.507361172356    12.100902897559
#> 9     226323.476783573 279.673857803119 302.179272739607   4.22353379436946
#> 10    497248.632537865  588.75195439444 760.551650867255 -0.474490749115707
#> 11    174370.816818801 308.907670371005 343.026493763273   8.34767731201209
#> 12    275119.773752356 321.611353673584 347.081093474015   2.66549150802948
#> 13    505460.957774313 684.011344679563 760.337585291923  0.295370842828944
#> 14    182334.909560634 313.578627352958   373.8349227868   7.87903425506779
#> 15    270263.212586679 310.819827962014 331.313678555656   2.68298140822498
#> 16    526370.947634185 732.285905104851 826.360653680944 -0.837418308801361
#> 17    183234.620906833 301.057848524397 369.064549273736   6.86767890227489
#> 18    276600.641494304  337.89023053681 365.577581466127   4.28128905786383
#> 19    518490.288790312 724.158874750037 820.426801106334  0.019688493388956
#> 20    187292.425497979 319.438349103166 382.327410127932   8.81733271975424
#> 21    274851.302602242 324.074505276816 351.739915635251     4.256335754957
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7630981175483 3389.95991367296   3444.479790285    79713.1994732515
#> 2   36.2315890609625 1413.81531560278 1503.79032880618    39391.1726054865
#> 3   23.5051672365582 2188.93064503236  2209.3827416232    63262.5501886365
#> 4   15.5946005081585  7478.1636749485 7651.71914632633    144297.101990057
#> 5    26.449112249928  2257.7290379359 2449.63058347947    53060.3537931752
#> 6   13.2300656947122 4796.46399282236 4844.24524710613    123956.529473992
#> 7   7.86970626566524  16753.205541227 17268.3628355179    344290.020174556
#> 8   14.6926007395451 5066.26512690877 5292.77088833992    112275.778486793
#> 9   5.56197743254288 7727.38174900112 7809.11148554981    218513.047879384
#> 10  1.86604307250259 19980.2233816031 20191.1623976088     472265.47023995
#> 11  11.5224003800223 6522.92810134364 6994.58975980866    156802.448088473
#> 12  5.04713213699445 9166.49750879306 9213.58200834957    254490.259161217
#> 13  1.11297188412994 20322.3820007542  20624.492514345    477182.226434956
#> 14  11.4894217704465 6886.83874883223 7055.16394222833    169834.069923039
#> 15  4.08258049052815 8962.36350466138 9033.62836284591    259038.906660234
#> 16  2.02068499368133 20585.4913698199 20823.1684049412    504837.033252138
#> 17  9.27170444547633 6688.98980239764 7097.32791623253    166554.472539083
#> 18  5.97682840649725 9305.57387603397 9337.82282590218    266891.298053958
#> 19  1.09747034733091 19609.2692931805  19910.050614349    485750.955918387
#> 20  10.3672539445942  6888.8023816063  7358.3363117216    170904.378428442
#> 21  5.99068993191458 9241.82237585346 9276.78991249538    262416.300475106
#>    ci95.integral.bt.up
#> 1     83787.3030986402
#> 2     44878.8135292123
#> 3     68298.6907319942
#> 4     153951.794311677
#> 5     60291.7403143403
#> 6     136489.751924406
#> 7     376955.227429087
#> 8     125409.916746902
#> 9     227005.941445104
#> 10    499431.627301761
#> 11    175905.916999121
#> 12    276922.352697213
#> 13    507931.914881635
#> 14    183427.215936735
#> 15    271243.977182193
#> 16     528252.55180349
#> 17    184692.109599161
#> 18    277449.030726956
#> 19    521351.007196596
#> 20      188724.3907759
#> 21    275937.856186167