Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4766204433267  15.5883553360382
#> 2                 FALSE          0.75 32.2853927978573  30.5350866820648
#> 3                 FALSE          0.75 64.8791947890199  20.4583654029071
#> 4                 FALSE          0.75 243.305079740433  15.0260826409056
#> 5                 FALSE          0.75 65.5656691036314   22.661765715911
#> 6                 FALSE          0.75 204.826698753539  12.4514866606825
#> 7                 FALSE          0.75 507.659161994458  7.07789586780629
#> 8                 FALSE          0.75 203.084220315289  13.6970773586882
#> 9                 FALSE          0.75 289.775955607263   4.6921378435415
#> 10                FALSE          0.75 666.232695271873 0.508742866627282
#> 11                FALSE          0.75 318.048221004921  9.71565510942697
#> 12                FALSE          0.75 335.599933784809  3.97781746048343
#> 13                FALSE          0.75 729.601113094478 0.841541832000623
#> 14                FALSE          0.75 346.363443787634  9.50658580308931
#> 15                FALSE          0.75 318.252454388246  3.52581205427913
#> 16                FALSE          0.75 773.443191627811 0.543804988471044
#> 17                FALSE          0.75 335.648681993088  8.05143033473319
#> 18                FALSE          0.75 350.133259841167  4.98580451633853
#> 19                FALSE          0.75 768.246006565799 0.614649571129635
#> 20                FALSE          0.75 351.793577815023  9.51212742562823
#> 21                FALSE          0.75 338.845648752407  5.34651514810302
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.49270002533 2120.29353692129 81699.1391991976 2.19915395167246
#> 2  1456.98179085763  287.81437466831 42576.7900939633 2.27397820718577
#> 3    2195.853961529 672.651184342038 66340.5159559531 1.55013344882234
#> 4  7561.63859436027 6038.98638845289 149187.851443697 2.83754750773352
#> 5   2366.3059070951 1114.32026827561 57500.9591031853 2.13053385083052
#> 6  4817.75276783145 3215.38324451294 130560.602430431 3.31253132783253
#> 7  17009.2940013228 15771.1211263883 361164.940627488 12.0268828356682
#> 8  5176.82568165197 3719.16280791698 118456.022453695 7.70969253770018
#> 9  7772.67782417208 6287.99153811736 221695.680814936 5.98587758224212
#> 10 20072.8859154939 18757.2879495102 481710.337350796 15.5199181883069
#> 11 6793.75403055356 5586.89538336484  168121.84054238 14.2357340214869
#> 12 9186.98578415018 7614.47957408072 267092.715370791 5.83212111198114
#> 13 20472.5751153676 19378.4931435399 492872.595946499 23.1318620275398
#> 14 6957.32173217399 5735.24842538044 175559.066411506 14.8038988028282
#> 15 9010.11824728707 7345.96045940086 264780.343232885 4.79383742958957
#> 16 20694.4640900135 19331.9686333271 514852.379278802 27.7228388856381
#> 17  6871.8758545092 5981.26776759192 175228.435925782 13.1403494962423
#> 18 9320.14049854997 7640.97891601294 272223.414033313 7.60451697279643
#> 19 19790.5605974699 18595.8861530864 502322.599131891 32.5345142333985
#> 20 7138.69956342365 5807.14319753601 180525.071121783  13.376188792013
#> 21 9256.41569499616 7442.12734849943 267864.954237538 6.64004534422022
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.647481956390713 22.6763092539711 52.8098691883529 1647.24437963491
#> 2   1.72740023809283 15.2108344342015 60.3870589783845 895.858664885676
#> 3   2.03876327211815 6.58670323384574 139.978953805752 1166.25477822571
#> 4  0.273435260975753 61.4592804553022    77.6377300573 3351.88118695948
#> 5   2.85889542858189 26.0804853916107 123.665111791549  1124.4478349247
#> 6  0.481822011305395 7.54471618180402 98.7239789077799 2313.75735667919
#> 7  0.526726406854272 120.128773214787 210.172528300311 6887.85981387798
#> 8  0.531214607242246 52.2515291115695 140.980408663821 2599.86227973143
#> 9  0.398029243580012 20.5883551627223 79.3075967699928 2512.42738429697
#> 10 0.207065011880906 94.4741176243221 195.668847167172  10588.981687357
#> 11 0.592094119567834  61.129021413846 134.463137896166 2116.31967143598
#> 12 0.575260929515027 11.3371765453985   149.4706828009  2515.4788369252
#> 13 0.385896812878265 58.0340776596757 230.321368292601 6833.73951986921
#> 14  0.56244930051653 58.0377012630179 160.294827900523 3724.59899505133
#> 15 0.529281297607731 16.4624550687115 135.818073516197 4178.86547772775
#> 16 0.547923066030559  92.678934145845 351.887871617345 9009.44262038471
#> 17 0.541128505747904 97.8894903448123 132.164977931679 4198.40414373388
#> 18 0.353085501081427 8.15455151082258 97.4004925155172 2586.39459763316
#> 19 0.323070997407473 72.1954567523372 152.940516844372 9241.39970179299
#> 20 0.380996913714752 81.5756722257606 96.4604650667415 2209.96655223747
#> 21 0.561487937857948 10.3632158372932 145.969572071941 4738.41999077941
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.8590121928255 96.0942286938279   14.5232475177755
#> 2             TRUE 28.5446986470367 36.0260869486779    27.693513290402
#> 3             TRUE 62.3292252657071 67.4291643123326   17.1045998202727
#> 4             TRUE 238.637314090211 247.972845390654   14.5762816366005
#> 5             TRUE 62.0609409190152 69.0703972882476   17.9588827358938
#> 6             TRUE 199.377584719254 210.275812787823   11.6588894520851
#> 7             TRUE 487.874939729783 527.443384259132   6.21143092853101
#> 8             TRUE 190.401776090772 215.766664539806   12.8232293297747
#> 9             TRUE 279.929186984474 299.622724230051   4.03737973785238
#> 10            TRUE 640.702429852108 691.762960691638  0.168120922083192
#> 11            TRUE 294.630438539575 341.466003470267   8.74166028273789
#> 12            TRUE   326.0060945556 345.193773014018   3.03151323143121
#> 13            TRUE 691.549200059175 767.653026129781  0.206741574815877
#> 14            TRUE 322.011030256982 370.715857318287   8.58135670373961
#> 15            TRUE 310.366591816571 326.138316959921   2.65514431971441
#> 16            TRUE 727.839121660937 819.047261594686 -0.357528455149225
#> 17            TRUE 314.032807071769 357.264556914407   7.16127394277788
#> 18            TRUE 337.623829420917 362.642690261417   4.40497886705958
#> 19            TRUE 714.726730651858  821.76528247974  0.083197780394343
#> 20            TRUE 329.789747252162 373.797408377885   8.88538750256747
#> 21            TRUE 327.922774161165  349.76852334365    4.4228674903267
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6534631543009 3377.19017130255 3451.79522874811    78989.4221946982
#> 2   33.3766600737275 1431.95996821337  1482.0036135019    41103.1025902263
#> 3   23.8121309855414 2185.01883470932 2206.68908834867    64422.0268457718
#> 4   15.4758836452107  7460.5380780113 7662.73911070925    143674.006891148
#> 5   27.3646486959282  2323.4035086259  2409.2083055643    55651.2424147342
#> 6   13.2440838692798 4805.34170971238 4830.16382595052    126754.471578694
#> 7   7.94436080708156 16811.6821693845 17206.9058332612    349834.411233659
#> 8   14.5709253876017 5090.87191626343  5262.7794470405    114179.249003537
#> 9   5.34689594923062  7738.8099799294 7806.54566841476    217562.737767768
#> 10 0.849364811171373 19917.4759920019 20228.2958389859    464291.462475094
#> 11  10.6896499361161 6693.19679032779 6894.31127077934    164640.494682868
#> 12  4.92412168953565   9168.336128733 9205.63543956736    262954.752684049
#> 13  1.47634208918537 20377.1090576174 20568.0411731177    481631.094436314
#> 14   10.431814902439 6861.84971359633 7052.79375075165    169432.101064647
#> 15  4.39647978884385 8983.03750869904  9037.1989858751    257906.109522023
#> 16  1.44513843209131 20542.0072433436 20846.9209366834     500031.84616827
#> 17  8.94158672668849 6710.84764289199 7032.90406612642    168322.061109339
#> 18  5.56663016561747 9306.72626131467 9333.55473578528    267968.794920206
#> 19  1.14610136186493 19671.7990711123 19909.3221238275    487120.496622441
#> 20   10.138867348689 7004.50758261227 7272.89154423502    176889.676143352
#> 21  6.27016280587935 9239.36820494382 9273.46318504851    260070.253352706
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84408.8562036971 88.1662786980487 96.7869621886048    14.3192907015124
#> 2     44050.4775977002 27.8283955117732 36.7423900839414    27.1493822154028
#> 3     68259.0050661344 61.8409332293281 67.9174563487117    16.4623893895555
#> 4     154701.695996245 237.743486625275  248.86667285559    14.4901495293931
#> 5     59350.6757916364 61.3898227560036 69.7415154512592    17.0583306758905
#> 6     134366.733282168 198.334137350987 211.319260156091    11.5071155185239
#> 7     372495.470021318 484.086471636548 531.231852352367    6.04551211037191
#> 8     122732.795903853 187.973222941397 218.195217689182    12.6558967284934
#> 9     225828.623862105 278.043635546068 301.508275668457    3.91200052612467
#> 10    499129.212226499 635.813655622791 696.651734920954   0.102895443340706
#> 11    171603.186401892 290.146182322807 345.950259687035    8.55515063507402
#> 12    271230.678057533 324.168976405326 347.030891164292    2.85030603863398
#> 13    504114.097456684   684.2626635205 774.939562668456  0.0851840787592232
#> 14    181686.031758366 317.347802134091 375.379085441177    8.40418517407691
#> 15    271654.576943747 308.856533026251 327.648375750242    2.48842071096797
#> 16    529672.912389335 719.106427411961 827.779955843662  -0.530124220948851
#> 17    182134.810742224 309.893596980453 361.403767005723    6.99081846346729
#> 18    276478.033146419 335.228406574486 365.038113107848    4.29375693421893
#> 19     517524.70164134 704.478358668338  832.01365446326 -0.0185695837890107
#> 20    184160.466100213 325.576247782678 378.010907847369    8.76537347474732
#> 21     275659.65512237 325.831159877736 351.860137627079    4.24599878990144
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    16.857419970564 3370.04713388755 3458.93826616312    78470.5402151132
#> 2   33.9207911487267  1427.1685553666 1486.79502634867    40820.9071107873
#> 3   24.4543414162586 2182.94402319066 2208.76389986734    64054.6565906308
#> 4   15.5620157524181 7441.17840466788 7682.09878405267    142618.164317256
#> 5   28.2652007559315 2315.18815572754 2417.42365846265    55297.0413467329
#> 6    13.395857802841 4802.96512411512 4832.54041154779     126025.63801134
#> 7   8.11027962524066 16773.8416058219 17244.7463968238    347664.735392288
#> 8    14.738257988883 5074.41268459329 5279.23867871064    113360.292385421
#> 9   5.47227516095832 7732.32464805314 7813.03100029101    216771.323141714
#> 10 0.914590289913858 19887.7166449503 20258.0551860376    460955.933243576
#> 11  10.8761595837799 6673.94114858242  6913.5669125247    163973.853986365
#> 12  5.10532888233289 9164.76491812119 9209.20665017916    262162.376850418
#> 13  1.59789958524202 20358.8283231546 20586.3219075805    479478.466487555
#> 14  10.6089864321017 6843.56783769848 7071.07562664951    168258.852381206
#> 15  4.56320339759028 8977.85183535239 9042.38465922174    256589.766896538
#> 16  1.61773419789094 20512.8133790876 20876.1148009394    497193.871742848
#> 17  9.11204220599908 6680.01245343337 7063.73925558504    166999.563804063
#> 18  5.67785209845812 9304.15757758876 9336.12341951118    267154.080621952
#> 19  1.24786872604828 19649.0575022353 19932.0636927045    484209.455716376
#> 20  10.2588813765091 6978.81124586116 7298.58788098614    176193.536679397
#> 21   6.4470315063046 9236.10379195507 9276.72759803726     258577.65105561
#>    ci95.integral.bt.up
#> 1     84927.7381832821
#> 2     44332.6730771392
#> 3     68626.3753212755
#> 4     155757.538570137
#> 5     59704.8768596377
#> 6     135095.566849522
#> 7     374665.145862689
#> 8     123551.752521969
#> 9     226620.038488158
#> 10    502464.741458016
#> 11    172269.827098394
#> 12    272023.053891165
#> 13    506266.725405442
#> 14    182859.280441807
#> 15    272970.919569231
#> 16    532510.886814757
#> 17      183457.3080475
#> 18    277292.747444674
#> 19    520435.742547405
#> 20    184856.605564168
#> 21    277152.257419465