Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0853024013121  15.2720135021091
#> 2                 FALSE          0.75 32.1806466251388  30.4885193364194
#> 3                 FALSE          0.75 64.8012348265449  20.2664523424842
#> 4                 FALSE          0.75  243.37973180183   15.018124693838
#> 5                 FALSE          0.75  65.927381498101  24.0639956510308
#> 6                 FALSE          0.75 205.402608533811  12.1817694123729
#> 7                 FALSE          0.75 506.327207355519  6.92908870718016
#> 8                 FALSE          0.75  203.07590364471   13.392057926756
#> 9                 FALSE          0.75 290.184655111522   4.6690171918222
#> 10                FALSE          0.75 675.112554652385 0.636709423612121
#> 11                FALSE          0.75  324.64051544442  9.91371213532602
#> 12                FALSE          0.75 334.746447365835  3.83852736169667
#> 13                FALSE          0.75 729.595928597818 0.694961201041401
#> 14                FALSE          0.75 342.071628804663  9.73705636668392
#> 15                FALSE          0.75 322.073665183936  3.39481930753919
#> 16                FALSE          0.75 771.947148719348 0.394941753635118
#> 17                FALSE          0.75 331.725292666922  8.09405672948615
#> 18                FALSE          0.75 351.196825255955  5.04325650236757
#> 19                FALSE          0.75 773.084720561344  0.65013784316977
#> 20                FALSE          0.75 350.573991497908  9.54895070695135
#> 21                FALSE          0.75 339.512775912444  5.38236560815143
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3421.0993379482 2151.52161345763 81772.2049576473 3.00331009549678
#> 2  1455.69728793912 297.969154108362 42432.1271263443 1.62349144361957
#> 3  2196.74139266442 683.925747601605 66031.5130977874 1.45815730521105
#> 4  7554.50853154138 6035.71547956164 148383.085240824 3.11452541257976
#> 5  2371.28916368426 1059.77879829706 57574.2397064171 3.40933684904507
#> 6  4820.62387201959 3273.78809087526  129305.65437853 2.32064142805125
#> 7  17045.4583475455 15872.3249182995 362491.530012133 15.6148207547454
#> 8  5195.54811134892 3783.91223320408 119317.625389666 8.95826373947456
#> 9  7765.34290152818 6290.45387781013  220720.64818902 6.15792158149296
#> 10 20067.9343350142 18668.6364993454 486903.624264295 22.5518020261737
#> 11 6784.78978506812 5553.93875963609 168143.659477605 11.6419648061594
#> 12 9193.16110728839 7657.39483369636 265176.611762499 5.51598590469253
#> 13 20449.2900956054 19451.0576129014  489763.17389191 16.0129436554493
#> 14 6961.56063343808 5665.07099299107 176472.408874379 13.1329505659479
#> 15  9002.2447107004 7417.62305038821 263010.568043572 4.80299421745619
#> 16 20689.9691440746 19429.8223642754 514817.794858738 27.9818554398688
#> 17  6906.4418879494 5993.42598137607 176622.900874345 13.7278433728675
#> 18 9318.66285451908 7642.36711390104 271725.036621757 5.32684383286836
#> 19 19776.1790393768 18519.7878495216 500234.000397982 30.6988413664033
#> 20 7126.99435579022 5787.68096470379 179776.866619097 11.7457606606065
#> 21 9256.08012780153 7425.10314875678   267095.0677503 6.84228180251804
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1    0.6264842228626 18.3338823339622 37.0843628803176 1436.14564297055
#> 2   2.23612698361141 14.8444126387324 86.2147329696148 982.779607401342
#> 3   2.04474349088771 5.83992371066842 137.617064698638 1372.54269565835
#> 4  0.213126598028351 57.8075388697462 80.0545963407473 3702.02599786643
#> 5   1.49553564468886 29.5413964060526 117.307281089173  1156.9185502391
#> 6  0.646658828698702 9.24715291406102 130.838139461148 2806.14403057594
#> 7  0.557578039075502 111.888916866534 124.681812797074 6215.02434163415
#> 8  0.745450814485457 26.5898647272458 113.871136467646 1749.63112385007
#> 9  0.325011613763642 24.1880133572432 77.9027068757407 3346.42120803324
#> 10 0.367994511686868 82.4543903763109 282.380983991831 7308.29909638226
#> 11 0.670895065388335 89.4978674772504  197.79012790492 3622.64123200936
#> 12 0.383984769150264 13.5112396738718 100.646756871359 4430.29163466862
#> 13 0.263923991426379 103.529572925982 188.491390433963 9200.19374577205
#> 14 0.633418948262013 75.6915693654106 192.656949749368 4108.13673292009
#> 15 0.271256505887017 17.7636844172889 40.6218691076209 6951.86633399838
#> 16  0.37931074490925 89.4621623229321  271.12128858301 6858.78537755446
#> 17 0.518707404004454 84.3854248810548 135.153268861324 3439.09986305236
#> 18 0.407661469461807 8.44131674102497 108.792053910999 3219.89677615995
#> 19 0.428704635787987 74.8437281264083 281.720396114228 14347.4845749363
#> 20 0.350496231935811 110.971525005249 157.554205547883 4555.82611698024
#> 21 0.559117759568631 10.5897202235738 156.200598234007 6334.08471736959
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.1448572942199 97.0257475084043    14.2414469555001
#> 2             TRUE 29.5100032003846  34.851290049893    26.8100904483786
#> 3             TRUE 62.4025660594728 67.1999035936171    16.9028492999739
#> 4             TRUE 238.256337498136 248.503126105524    14.6675314400814
#> 5             TRUE 60.3190223814218 71.5357406147801    21.6038395155177
#> 6             TRUE 201.585153384667 209.220063682956    11.1180156391635
#> 7             TRUE 480.640827213963 532.013587497075    6.01187283290096
#> 8             TRUE 188.339559793275 217.812247496146    12.1657913369274
#> 9             TRUE 280.054874109966 300.314436113078      4.134373087181
#> 10            TRUE 638.014840319329 712.210268985441  0.0313584518872241
#> 11            TRUE 305.489483338288 343.791547550553     8.8100897527622
#> 12            TRUE 325.672650552616 343.820244179054    3.20687241644449
#> 13            TRUE 703.254636284604 755.937220911032   0.260806235145008
#> 14            TRUE 320.467925123679 363.675332485648    8.69508219679291
#> 15            TRUE 314.172739696221 329.974590671652    2.94860235535504
#> 16            TRUE 725.916996520764 817.977300917932  -0.229024421740598
#> 17            TRUE 309.142990318555 354.307595015289    7.24078304989883
#> 18            TRUE 342.434167150886 359.959483361023    4.37265338510289
#> 19            TRUE 722.585126513611 823.584314609078 -0.0550812827014678
#> 20            TRUE  331.25221521121 369.895767784605    8.97238440541694
#> 21            TRUE 328.257222347301 350.768329477586    4.46261689366103
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3025800487181 3390.94010150883 3451.25857438756    79409.7453749607
#> 2   34.1669482244601 1431.27822914841 1480.11634672984    40815.4546721691
#> 3   23.6300553849944 2187.13471816037 2206.34806716847    63773.6803634294
#> 4   15.3687179475946 7459.41513010065 7649.60193298211    142293.252474333
#> 5    26.524151786544  2322.6935665963 2419.88476077222    55671.1086912738
#> 6   13.2455231855823 4805.41230547596 4835.83543856322    124689.547448233
#> 7   7.84630458145936 16861.4010793001  17229.515615791    352267.814970145
#> 8   14.6183245165846  5151.8077838726 5239.28843882524    116439.482190932
#> 9   5.20366129646339 7725.55361955551 7805.13218350084    215215.785301805
#> 10  1.24206039533702 19932.2968628452 20203.5718071832    474881.472250746
#> 11  11.0173345178898 6637.56579306805  6932.0137770682     162184.41465095
#> 12  4.47018230694886 9170.93511802487 9215.38709655191    257888.782023469
#> 13  1.12911616693779 20278.9839481422 20619.5962430687    474628.855180115
#> 14  10.7790305365749 6837.04800183198 7086.07326504419    169714.523948725
#> 15  3.84103625972333 8973.02344983396 9031.46597156683    251574.747924144
#> 16  1.01890792901083 20542.8038870533 20837.1344010958    503535.092912661
#> 17  8.94733040907348 6767.62786402007 7045.25591187874    170965.581599624
#> 18  5.71385961963224 9304.77688848009 9332.54882055806    266428.306424974
#> 19  1.35535696904101 19653.0611066089 19899.2969721448    476632.388272211
#> 20  10.1255170084858 6944.44619715658 7309.54251442385    172282.532656664
#> 21  6.30211432264183 9238.66003803375 9273.50021756931    256675.498390227
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84134.6645403338 86.1988146141384 97.9717901884858    14.0441044252984
#> 2     44048.7995805195 28.9986033956445 35.3626898546332     26.105710448541
#> 3     68289.3458321454 61.9432465083313 67.6592231447586    16.2587551003442
#> 4     154472.918007314 237.275261993174 249.484201610486    14.6003965617024
#> 5     59477.3707215604 59.2450812739726 72.6096817222293    21.1327457874407
#> 6     133921.761308828 200.854151334831 209.951065732792    10.9143181081235
#> 7     372715.245054121 475.722158676218  536.93225603482    5.83623575059217
#> 8     122195.768588399  185.51770671534 220.634100574081    11.9309743303645
#> 9     226225.511076234 278.115128811796 302.254181411248    4.03199442884546
#> 10    498925.776277843 630.911022681085 719.314086623685 -0.0845598192941392
#> 11    174102.904304261 301.822264424348 347.458766464493    8.59875780716488
#> 12    272464.441501529 323.935114992638 345.557779739033    3.08591721416216
#> 13    504897.492603705 698.210559033138 760.981298162499   0.177670177845699
#> 14    183230.293800032 316.331045695406 367.812211913921    8.49555522809038
#> 15    274446.388162999 312.659796517722  331.48753385015    2.86315655600063
#> 16    526100.496804815 717.102712057205 826.791585381491  -0.348507306387011
#> 17    182280.220149066 304.818719656101 358.631865677742    7.07739021763742
#> 18     277021.76681854 340.756211343533 361.637439168377    4.24424002222242
#> 19    523835.612523752 712.914991483194 833.254449639495  -0.190123242974684
#> 20    187271.200581529 327.552300603119 373.595682392696    8.86197809235716
#> 21    277514.637110373 326.101903579508 352.923648245379    4.28649479939691
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4999225789198 3385.16492857363 3457.03374732276     78957.359497425
#> 2   34.8713282242977 1426.60223916721 1484.79233671104    40505.8790958377
#> 3   24.2741495846241 2185.29514219151 2208.18764313733    63341.3294142971
#> 4   15.4358528259736 7441.20575535667 7667.81130772608    141127.114285005
#> 5    26.995245514621  2313.3880267284 2429.19030064012    55306.6793479485
#> 6   13.4492207166224 4802.49945230803 4838.74829173115    123805.612078601
#> 7   8.02194166376814 16826.1560704871 17264.7606246039     350310.08230253
#> 8   14.8531415231475 5143.43197648352 5247.66424621432     115888.34838692
#> 9   5.30603995479893 7717.93439534798 7812.75140770837    214161.662621275
#> 10  1.35797866651838 19906.3237298767 20229.5449401518    472579.358035385
#> 11  11.2286664634872 6609.37396481271 6960.20560532353    161043.282662867
#> 12  4.59113750923119  9166.6790775276 9219.64313704918    256493.240158548
#> 13   1.2122522242371 20246.3721326705 20652.2080585404    471730.794150196
#> 14  10.9785575052775 6813.20515748188 7109.91610939429    168420.460877856
#> 15  3.92648205907774 8967.42788924251 9037.06153215828    249384.910028935
#> 16  1.13839081365725 20514.6233059216 20865.3149822275    501374.575518731
#> 17  9.11072324133488 6741.04645518254 7071.83732071627    169882.265142763
#> 18  5.84227298251271 9302.11787370667 9335.20783533148    265414.038940484
#> 19  1.49039892931422 19629.4853322491 19922.8727465046    472112.930631106
#> 20  10.2359233215455 6909.49016677993 7344.49854480051    170847.447429815
#> 21  6.47823641690594 9235.32427616332 9276.83597943973    254680.261704256
#>    ci95.integral.bt.up
#> 1     84587.0504178696
#> 2     44358.3751568509
#> 3     68721.6967812778
#> 4     155639.056196642
#> 5     59841.8000648858
#> 6     134805.696678459
#> 7     374672.977721736
#> 8     122746.902392412
#> 9     227279.633756765
#> 10    501227.890493204
#> 11    175244.036292344
#> 12    273859.983366449
#> 13    507795.553633623
#> 14    184524.356870902
#> 15    276636.226058209
#> 16    528261.014198745
#> 17    183363.536605928
#> 18    278036.034303031
#> 19    528355.070164857
#> 20    188706.285808378
#> 21    279509.873796345