Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4739380711881   15.633992287068
#> 2                 FALSE          0.75 32.8092224565786  31.3835592401036
#> 3                 FALSE          0.75 64.5515071707432  20.3174982791946
#> 4                 FALSE          0.75 243.832601181182  15.0935766153077
#> 5                 FALSE          0.75 65.3645631940446  22.7340114000361
#> 6                 FALSE          0.75 206.085263918579  12.3721769332325
#> 7                 FALSE          0.75 507.328482138298  6.89331476972163
#> 8                 FALSE          0.75 204.312592823668  13.2753024075725
#> 9                 FALSE          0.75 290.888294944004  4.77295940455427
#> 10                FALSE          0.75 672.657900255964 0.578906134297003
#> 11                FALSE          0.75 325.613631962596  9.87290526589808
#> 12                FALSE          0.75 332.832242869503   3.7913356115914
#> 13                FALSE          0.75 730.598373403677 0.769959526359372
#> 14                FALSE          0.75 340.644342768607  9.54130520177716
#> 15                FALSE          0.75 319.521597486447  3.39980338260013
#> 16                FALSE          0.75  769.78470904678 0.404718412014467
#> 17                FALSE          0.75 334.119517261934  8.09752909913547
#> 18                FALSE          0.75 349.561978260783   4.9749177808601
#> 19                FALSE          0.75 779.949690101176 0.646783617850254
#> 20                FALSE          0.75 352.209310600726   9.5521052567534
#> 21                FALSE          0.75 341.282360094807  5.46989673898201
#>                A.bt            dY.bt      integral.bt   stdmax_slope.bt
#> 1  3416.39185403118 2118.84279083776 81741.1015629213    2.435919026031
#> 2  1465.80009255604 273.829980271934 42400.1429372906   2.2634079606512
#> 3  2196.62791560166 684.361323282299 66330.1488986703 0.961170644732891
#> 4   7557.6200696206 6025.67497896628 148828.864140173  2.36922622930248
#> 5  2359.32768418033 1135.37994116761 56872.2939448285   2.3818725905699
#> 6  4818.38423853557 3239.31601164737 130727.873693146  3.30821652734502
#> 7  17037.7891736752 15900.5328173483 361889.756240167  11.1303559714417
#> 8  5171.11569340563 3798.92965628301 118563.242559644  9.96852003910239
#> 9  7769.55148287773 6262.29659826819 220452.878742552  5.71441280160228
#> 10 20066.3964532167 18734.0729055157  486361.48615878  28.2919227908147
#> 11 6801.48053351344 5606.96622660111  168249.28012051  11.3461860775063
#> 12 9188.90723257052 7641.09076284369 267300.720203856  6.54779824095848
#> 13  20459.421886717 19411.2952282324 491986.474988444  29.2121267607509
#> 14 6970.45035563105 5706.50493893254  177206.13797067  15.8396283215382
#> 15 8998.05007364355 7381.04626101354 264777.506908192  6.13630260917141
#> 16 20673.4301909771 19399.1429354515  514738.39978439  18.4558662588899
#> 17 6902.93559270526 6027.54274070102 176250.906601665  12.2011438738017
#> 18  9322.3174324682 7642.51219256471  271771.55605212  7.26909932473223
#> 19 19810.8159508521 18655.5673097479 505197.694131053  26.4701919058415
#> 20 7155.00283100151 5824.50440779623 180808.645514841  13.1842039123661
#> 21 9255.72080591869 7420.91137173713 267219.825017841  3.89821943117622
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.706060230501369 19.2541770243736 66.6745063132376 1258.88975229513
#> 2   2.26170476315709  20.524770666871 72.3668256226841 928.565367586634
#> 3   1.94615972875845 4.86096433558174 130.312929835006 888.626188492516
#> 4   0.35458166312642 71.9097635999472 111.263853364082 3507.00182055822
#> 5   2.35033639178414   38.62172345972 127.383397258834 1310.78810308765
#> 6  0.551184709756251 8.43461989268947 106.764803318844 1827.71466793748
#> 7  0.398983498945252 127.803475885322 163.676805520874 7218.28374474395
#> 8  0.813557108510104 70.8654995727236 149.342367524489 3232.61326738399
#> 9  0.334411648643813 20.6461699254994 76.8253722150442  3255.3368809345
#> 10 0.380294605941104 64.0050103697211 232.138897134677 7990.35514570811
#> 11 0.494555960604608 56.8249370685611 150.665582574801 2404.54744990474
#> 12 0.465031503483829 13.1609636718087 106.591678242445 4186.44748343473
#> 13 0.318559418542827 70.8020174273141 191.352208938636 7121.66273295689
#> 14 0.753795866927296  51.787757492851 183.110806604996 3101.93250463872
#> 15 0.395231587971077 16.5341539463439  92.324529768338  3484.1445367223
#> 16 0.309918341881055 83.2751509987538 198.253430559181 8499.69496658662
#> 17 0.457921696172014 76.8254086756433 136.046758641188 2464.35652480384
#> 18 0.390802362354234 8.07839517050637  95.137266086838   2714.677605763
#> 19 0.307959067738003  83.016049717902  210.54946341917 10592.8527314992
#> 20 0.317498875850692 79.7551784852893 124.698605755043 2212.03110163968
#> 21 0.451931891227015 10.5878739574457 141.037083709033 4100.25811156554
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.4668512733671 96.4810248690091    14.4725232078933
#> 2             TRUE 29.0859163613074 36.5325285518498    27.6630549047101
#> 3             TRUE 62.9703814601576 66.1326328813289     17.116065525387
#> 4             TRUE 239.935224033979 247.729978328384    14.5102897794647
#> 5             TRUE 61.4463827825571 69.2827436055321    18.8677080355512
#> 6             TRUE 200.643247731097 211.527280106062    11.4654780856835
#> 7             TRUE 489.019046565277  525.63791771132    6.23698691395669
#> 8             TRUE 187.914377359345 220.710808287992    11.9370009640734
#> 9             TRUE 281.488085885368  300.28850400264     4.2228522425352
#> 10            TRUE 626.117687265074 719.198113246854 -0.0466784924761127
#> 11            TRUE 306.949155865099 344.278108060094     9.0593607107035
#> 12            TRUE 322.061114763127  343.60337097588     3.0263587883605
#> 13            TRUE 682.544424882241 778.652321925112   0.245929282856422
#> 14            TRUE 314.588154179676 366.700531357537    8.30131100068176
#> 15            TRUE  309.42737969436 329.615815278534     2.7496474203877
#> 16            TRUE 739.424809050906 800.144609042653  -0.105097260379868
#> 17            TRUE  314.04863558953 354.190398934338    7.34424790893251
#> 18            TRUE 337.604309871598 361.519646649967    4.33204789478738
#> 19            TRUE 736.406224416067 823.493155786286   0.140190951421238
#> 20            TRUE 330.521295164884 373.897326036569    9.02981960597901
#> 21            TRUE 334.869789130522 347.694931059092    4.72646877791357
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7954613662428 3384.71873282608 3448.06497523627    79670.2279203958
#> 2    35.104063575497 1432.03684480904 1499.56334030305    40872.6529076106
#> 3   23.5189310330023 2188.63162926963 2204.62420193369    64868.3588186001
#> 4   15.6768634511506 7439.32850849869 7675.91163074252    143059.846145355
#> 5    26.600314764521 2295.79494908909 2422.86041927157    54716.0475152493
#> 6   13.2788757807816 4804.50928881209 4832.25918825904    127721.283064389
#> 7   7.54964262548657 16827.5524558439 17248.0258915066    350015.679480063
#> 8   14.6136038510717  5054.5419466085 5287.68944020276    113245.593734797
#> 9   5.32306656657335 7735.58853335028 7803.51443240517    215097.849573415
#> 10  1.20449076107012 19961.1082111586 20171.6846952749     473217.35194409
#> 11  10.6864498210927 6708.00351203566 6894.95755499123    164293.799565417
#> 12  4.55631243482229 9167.25744733039 9210.55701781064    260414.014093606
#> 13  1.29398976986232 20342.9525680491 20575.8912053849     480271.33979273
#> 14  10.7812994028726 6885.25949455531 7055.64121670679     172103.45900054
#> 15  4.04995934481255 8970.85139040181 9025.24875688528    259046.089145284
#> 16 0.914534084408803 20536.4425675842 20810.4178143701    500756.401564355
#> 17  8.85081028933844 6776.55779543383 7029.31338997669    172197.040118363
#> 18  5.61778766693281 9309.02847241271 9335.60639252368     267305.91139064
#> 19  1.15337628427927 19674.2545490661  19947.377352638    487772.451387737
#> 20  10.0743909075278 7023.80556239321 7286.20009960981    177169.854352644
#> 21  6.21332470005045 9238.30375325869 9273.13785857869    260474.900424316
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83811.9752054468 87.6995367801673 97.2483393622088   14.2501142352854
#> 2     43927.6329669706 28.3729428537023  37.245502059455   26.9506179043157
#> 3     67791.9389787405 62.6676127070668 66.4354016344197   16.5030252108281
#> 4     154597.882134992 239.188917771749 248.476284590614   14.3985965555799
#> 5     59028.5403744077 60.6960929165276 70.0330334715616   18.1273520721392
#> 6     133734.464321903 199.601159524983 212.569368312176   11.2918549021103
#> 7     373763.833000271 485.512984434273 529.143979842324   6.11130711178894
#> 8      123880.89138449 184.774293547027 223.850892100309   11.6807304748927
#> 9      225807.90791169 279.688045852863 302.088544035144    4.1175125732124
#> 10     499505.62037347 617.205731585967 728.110068925961  -0.16647129334756
#> 11    172204.760675603 303.375107250684 347.852156674509   8.90357558311305
#> 12    274187.426314107 319.998558317225 345.665927421782   2.87987386476309
#> 13    503701.610184158 673.342604952605 787.854141854748  0.145583066015431
#> 14    182308.816940801 309.598671258392 371.690014278822   8.06386530259966
#> 15    270508.924671101 307.494444372471 331.548750600423   2.62514947017681
#> 16    528720.398004425 733.611211179355 805.958206914204 -0.202721538072401
#> 17    180304.773084968 310.205275269283 358.033759254586   7.20000257463832
#> 18      276237.2007136 335.314543584307 363.809412937258    4.2089451506458
#> 19    522622.936874369 728.068113965727 831.831266236626 0.0431838450837669
#> 20    184447.436677038 326.368270932489 378.050350268964   8.92980746008604
#> 21    273964.749611367 333.641850009702 348.922870179913   4.58411023217706
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0178703388507 3378.65366706341 3454.13004099895    79273.6776484228
#> 2   35.8165005758915 1425.57154204898 1506.02864306311    40580.1548168208
#> 3   24.1319713475612 2187.10042550392  2206.1554056994     64588.441569225
#> 4   15.7885566750354 7416.67693296471  7698.5632062765    141955.140571879
#> 5   27.3406707279331 2283.62910619928 2435.02626216138    54303.1492627767
#> 6   13.4524989643548 4801.85238354589 4834.91609352524    127145.552943989
#> 7   7.67532242765432   16787.29436094 17288.2839864104    347741.920100469
#> 8   14.8698743402524 5032.21931424309 5310.01207256817    112227.320555571
#> 9   5.42840623589615 7729.08498982375 7810.01797593171    214072.418455921
#> 10  1.32428356194157 19940.9466328921 20191.8462735414    470700.390073192
#> 11  10.8422349486831 6690.10365685906 6912.85741016782    163536.367118697
#> 12   4.7027973584197 9163.11174377377 9214.70272136726    259095.283136324
#> 13  1.39433598670331 20320.6499325595 20598.1938408745    478028.016031848
#> 14  11.0187451009547 6868.94635094507 7071.95436031704    171126.350261578
#> 15  4.17445729502344 8965.64313190872 9030.45701537838    257948.583616217
#> 16  1.01215836210134 20510.2108950196 20836.6494869347    498078.997649881
#> 17  8.99505562363262   6752.357791701 7053.51339370952     171420.76781305
#> 18   5.7408904110744 9306.48377793401 9338.15108700239    266450.787944825
#> 19  1.25038339061674  19648.104493405 19973.5274082992    484435.702777315
#> 20  10.1744030534208 6998.68268117034 7311.32298083268    176473.064555627
#> 21  6.35568324578696  9234.9685729621 9276.47303887529    259183.319119173
#>    ci95.integral.bt.up
#> 1     84208.5254774197
#> 2     44220.1310577604
#> 3     68071.8562281157
#> 4     155702.587708467
#> 5     59441.4386268803
#> 6     134310.194442304
#> 7     376037.592379865
#> 8     124899.164563716
#> 9     226833.339029184
#> 10    502022.582244368
#> 11    172962.193122323
#> 12    275506.157271389
#> 13    505944.933945039
#> 14    183285.925679762
#> 15    271606.430200168
#> 16      531397.8019189
#> 17    181081.045390281
#> 18    277092.324159416
#> 19    525959.685484792
#> 20    185144.226474055
#> 21     275256.33091651