Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9847396084785  15.4028735135695
#> 2                 FALSE          0.75 31.9116280727235  30.9242935493401
#> 3                 FALSE          0.75 64.7353416049868  20.5988448847981
#> 4                 FALSE          0.75 243.060296027202  14.9462135575211
#> 5                 FALSE          0.75 66.0120345458013  20.9762701072249
#> 6                 FALSE          0.75 205.686936827387   12.198386964902
#> 7                 FALSE          0.75 509.564516672429  7.12617492524293
#> 8                 FALSE          0.75 202.093505646135  13.2510009636451
#> 9                 FALSE          0.75 289.479630575918  4.71190473270897
#> 10                FALSE          0.75 669.995244486281 0.648390502524811
#> 11                FALSE          0.75 324.849590213907  9.90638498176219
#> 12                FALSE          0.75   333.8318022554  3.73492497796903
#> 13                FALSE          0.75  729.21319099289 0.667610452697099
#> 14                FALSE          0.75 344.572450220117  9.56392303355361
#> 15                FALSE          0.75 318.554062085658  3.34932798880783
#> 16                FALSE          0.75 782.789713341766 0.541931793537221
#> 17                FALSE          0.75 333.844013969973  8.03621846822582
#> 18                FALSE          0.75 350.254128400577    4.879087077856
#> 19                FALSE          0.75 765.592742803219 0.596983742070676
#> 20                FALSE          0.75 349.161970321928  9.54784950864743
#> 21                FALSE          0.75 335.903978592808   5.0894913749117
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3413.54722005564 2133.87775599064 81276.9724149024 2.34922736515823
#> 2  1464.24247824754 300.290871997481 42514.7513972085 1.67657790924948
#> 3  2196.34127579597 654.026934645951 66337.4935263552 1.42228710064167
#> 4  7563.17248658671 6059.28753950949 149449.570383118 2.43960732850802
#> 5   2373.9349224867 1178.75584955987 57619.7168179584 2.16817041402187
#> 6  4818.54932891019  3270.3140166527 130204.745308063 2.02041438768116
#> 7  17014.4654726135 15792.3111149016 360841.000885631 10.9386248784728
#> 8  5188.28418831615 3812.63748283723 118909.971165765 6.05789460434922
#> 9  7769.57140188432 6274.42201708037 223274.323513731 5.11876533264732
#> 10 20118.7044965794 18691.1297768164 485510.129404334 19.9036193165471
#> 11 6799.95031457499  5591.5769584043 168021.822982046 10.6220874183989
#> 12 9185.67628245786 7662.52479795538 267700.649339289 6.63338286645051
#> 13 20425.1204865519 19423.3948350119 489225.238585307 24.1050283279745
#> 14 6966.84384934091 5704.61204599343  176780.20525863 15.3499883944468
#> 15 9005.95626959705 7394.51499341451  266128.12938683 5.83667655271188
#> 16 20703.5455140766 19355.4198309457 514142.578183432 26.5485270429508
#> 17 6898.40936577634 6012.76853171412 175899.126408411 13.9078645888431
#> 18 9322.15629621686 7693.58943318104 272215.273202785 5.99756217064345
#> 19 19795.9345632655 18606.7231022084 503901.039310867 24.6916760577765
#> 20 7163.92347435998 5836.25052784422 181248.317568546 14.6955421556789
#> 21  9258.7408376378 7494.87446111457 268914.088617072 6.80646053147984
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.624433528790086 28.9538450421274 59.7018919861655 1991.61218102767
#> 2   2.01267819278319 22.1629571773314 56.5334418546512 932.373930215163
#> 3   1.81236687496764 5.39315942152888 117.746284029136 931.071851255244
#> 4   0.32183505763517 42.8964376097909 83.7844373880625 2269.95751258733
#> 5   5.41600341277183 23.1662664900432 159.811717930352 1030.61328712827
#> 6  0.521832196710161 11.6949533503422 103.459449669921 2309.30796916189
#> 7  0.380394999312114 160.995823415647 203.090426376379 8482.89584789579
#> 8  0.432860645739267 54.1731063574409 112.628136051412 2746.30906779817
#> 9  0.321203038320881 16.8859331700856 63.5773461901347 1129.64691241388
#> 10 0.286352448116245 78.6784861974415 147.976172440999 9585.49803210908
#> 11 0.696458995429265 70.1658331524183 237.658901316648 2311.03628823581
#> 12 0.395954284089254 14.4556449293504 82.8161379825908  4566.5700399348
#> 13  0.35386064969515 110.222555255546 208.618394418146    11902.9184864
#> 14 0.743212996925575 46.7076508658296 194.490896808546 3238.00402212546
#> 15 0.534930618561087 16.1747565431402 122.920176783245 2336.79100950714
#> 16 0.368900196039344 96.5532333373838 226.866720903891 9227.31598131232
#> 17 0.480087049862503 97.6265740848617 135.841777942479 3271.53154124486
#> 18 0.312359238901585 13.7106637244988 54.7271057769968 2246.67492154699
#> 19 0.308930575970496 99.7838488313039 214.133272328693 7841.81381062888
#> 20 0.323470662325715 86.3707072917077 95.9353590751731 2460.22221465916
#> 21  0.40610850545117 9.08444670728527 119.874369863894 2949.55425268299
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.1202605927932 95.8492186241638    14.3756803587098
#> 2             TRUE 29.1536574120081 34.6695987334389    27.6134379222118
#> 3             TRUE 62.3956793244313 67.0750038855424    17.6175013754763
#> 4             TRUE 239.047141971806 247.073450082598    14.4167948877112
#> 5             TRUE 62.4453942147353 69.5786748768672    12.0669444932153
#> 6             TRUE 202.363355159651 209.010518495123    11.3399730013138
#> 7             TRUE 491.570478747341 527.558554597516    6.50042515137451
#> 8             TRUE 192.128269021981  212.05874227029     12.538945201404
#> 9             TRUE 281.059261603713 297.899999548123    4.18352573467112
#> 10            TRUE 637.253790710561 702.736698262001   0.177340725373587
#> 11            TRUE  307.37625641064 342.322924017173    8.76070993428104
#> 12            TRUE 322.919887440089 344.743717070711    3.08358018064221
#> 13            TRUE 689.560419393372 768.865962592408  0.0855096839485764
#> 14            TRUE 319.321719311252 369.823181128982    8.34133765361103
#> 15            TRUE 308.952729156447 328.155395014869    2.46936712127485
#> 16            TRUE 739.117386356112  826.46204032742 -0.0649090289474991
#> 17            TRUE 310.965576721326  356.72245121862    7.24647527120201
#> 18            TRUE 340.388138629869 360.120118171286    4.36525612986289
#> 19            TRUE 724.974935688177 806.210549918261  0.0887929445992098
#> 20            TRUE 324.987803475837  373.33613716802    9.01574026912162
#> 21            TRUE 324.707351018523 347.100606167092    4.42144288344452
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4300666684292 3365.91814496134 3461.17629514994    78000.7703771118
#> 2   34.2351491764685 1427.78441369083 1500.70054280425    40980.9962820046
#> 3   23.5801883941199 2187.46952854755 2205.21302304438    64805.8803310404
#> 4    15.475632227331  7492.6078467186 7633.73712645481    145715.490274912
#> 5   29.8855957212346 2335.82641411058 2412.04343086282    55924.3579606324
#> 6   13.0568009284902 4799.31113064887  4837.7875271715    126405.933698791
#> 7   7.75192469911136 16749.6273430948 17279.3036021323    346886.637215842
#> 8   13.9630567258862 5099.16942835816 5277.39894827414    114392.292749237
#> 9   5.24028373074682 7741.79404181953 7797.34876194911     221416.05434281
#> 10  1.11944027967603 19989.2783867846 20248.1306063742    469741.985141515
#> 11  11.0520600292433 6684.52751903926 6915.37311011072    164220.168287898
#> 12  4.38626977529586 9161.89674654908 9209.45581836664    260188.641623596
#> 13  1.24971122144562 20243.8043831565 20606.4365899473    469644.937675179
#> 14  10.7865084134962 6890.00976366662  7043.6779350152    171453.688642233
#> 15  4.22928885634082 8979.34879508358 9032.56374411051     262284.10817619
#> 16  1.14877261602194 20544.7154452366 20862.3755829166    498963.643394173
#> 17  8.82596166524964 6737.81365140674 7059.00508014594    170517.457023063
#> 18  5.39291802584911 9299.60225439006 9344.71033804366     268519.49295684
#> 19  1.10517453954214  19631.790131938  19960.078994593    491001.255592382
#> 20  10.0799587481732 7021.84366086512 7306.00328785484    177201.252025431
#> 21  5.75753986637887 9243.79692280432 9273.68475247128    264062.071871409
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84553.1744526929 87.3802539727684 96.5892252441886     14.1789837971409
#> 2     44048.5065124125 28.6255353705945 35.1977207748525     26.9794442914851
#> 3     67869.1067216701 61.9476588877292 67.5230243222445     17.0466058098615
#> 4     153183.650491324 238.278665663326 247.841926391078     14.3154168445562
#> 5     59315.0756752844 61.7624205343184 70.2616485572841     10.3609034181921
#> 6     134003.556917334 201.726924627532 209.646949027242     11.1755958593501
#> 7      374795.36455542 488.124811910622 531.004221434235     6.38060072659119
#> 8     123427.649582293 190.220032221611  213.96697907066     12.4025940979961
#> 9     225132.592684652 279.446850523929 299.512410627907     4.08234677760004
#> 10    501278.273667154 630.984150625849 709.006338346713     0.08713970421697
#> 11    171823.477676194 304.030298873845 345.668881553969     8.54132535072083
#> 12    275212.657054981 320.830371837157 346.833232673643      2.9588545811541
#> 13    508805.539495435  681.96733547006  776.45904651572  -0.0259564207053959
#> 14    182106.721875026 314.486472967001 374.658427473233     8.10722555957948
#> 15    269972.150597469 307.114176042343 329.993948128974      2.3008639764281
#> 16    529321.512972691 730.754600337583  834.82482634595   -0.181112590699892
#> 17    181280.795793758  306.58459937584 361.103428564105     7.09524785049532
#> 18     275911.05344873 338.498906546116 362.009350255039      4.2668629696089
#> 19    516800.823029351 717.197057729977 813.988427876461 -0.00852018683149636
#> 20     185295.38311166 320.358707696798 377.965232947059     8.91384701048902
#> 21    273766.105362736 322.563315951107 349.244641234508     4.29351870422741
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6267632299981 3356.79768377307 3470.29675633821    77373.4125400881
#> 2   34.8691428071952 1420.80308217997 1507.68187431511    40687.2984939868
#> 3   24.1510839597347 2185.77068332977 2206.91186826217     64512.592697895
#> 4    15.577010270486 7479.09546887152  7647.2495043019    145000.453658447
#> 5   31.5916367962577 2328.52904016622 2419.34080480719     55599.714775187
#> 6   13.2211780704539 4795.62722034352 4841.47143747686    125678.501688505
#> 7   7.87174912389468 16698.9136587188 17330.0172865082    344214.525023755
#> 8   14.0994078292941 5082.10489985557 5294.46347677673     113527.20539288
#> 9    5.3414626878179 7736.47497287096 7802.66783089769      221060.2155654
#> 10  1.20964130083265 19964.4946636324 20272.9143295264    466722.553261401
#> 11  11.2714446128035 6662.42528159625 6937.47534755373    163492.191857104
#> 12  4.51099537478397 9157.34321839633 9214.00934651939    258750.172061016
#> 13  1.36117732609959  20209.084278251 20641.1566948528    465895.518351963
#> 14  11.0206205075277 6875.29685364388 7058.39084503794    170433.717375264
#> 15  4.39779200118757 8974.25374677249  9037.6587924216    261548.019008196
#> 16  1.26497617777433 20514.3011767353 20892.7898514178     496057.03886006
#> 17  8.97718908595633 6707.06128057001 7089.75745098267    169486.924587571
#> 18  5.49131118610311 9295.28339531684 9349.02919711688    267811.790356553
#> 19  1.20248767097285 19600.3582195561 19991.5109069748    488531.084242034
#> 20  10.1818520068058 6994.63688806823 7333.21006065172    176426.282027814
#> 21  5.88546404559599 9240.93532209152 9276.54635318408    263132.962281814
#>    ci95.integral.bt.up
#> 1     85180.5322897166
#> 2     44342.2043004303
#> 3     68162.3943548155
#> 4     153898.687107789
#> 5     59639.7188607298
#> 6      134730.98892762
#> 7     377467.476747507
#> 8     124292.736938649
#> 9     225488.431462062
#> 10    504297.705547268
#> 11    172551.454106988
#> 12    276651.126617561
#> 13    512554.958818651
#> 14    183126.693141996
#> 15    270708.239765464
#> 16    532228.117506804
#> 17    182311.328229251
#> 18    276618.756049017
#> 19    519270.994379699
#> 20    186070.353109278
#> 21    274695.214952331