Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  91.851496873442  15.5975274732329
#> 2                 FALSE          0.75 32.4202326855298  31.1432565781901
#> 3                 FALSE          0.75 65.3640952465674  19.8634961476677
#> 4                 FALSE          0.75 243.004988980223  15.0766585115689
#> 5                 FALSE          0.75 66.1885712486276  22.4630760256347
#> 6                 FALSE          0.75 204.558640946139  12.3851698781514
#> 7                 FALSE          0.75 509.396989698221  7.14925593378405
#> 8                 FALSE          0.75 199.676424068462   13.338789365213
#> 9                 FALSE          0.75  290.18486102932  4.65622664142553
#> 10                FALSE          0.75 668.706066818699 0.432495497343966
#> 11                FALSE          0.75 324.630743941544  9.99812407263169
#> 12                FALSE          0.75 333.802120892202  3.74303915087931
#> 13                FALSE          0.75 723.496558157816  0.75738912206119
#> 14                FALSE          0.75 348.707584859931  10.0008076067796
#> 15                FALSE          0.75 319.034413540785  3.32825584019286
#> 16                FALSE          0.75 778.060408619532 0.465147107124701
#> 17                FALSE          0.75 332.314271279698  8.20455749626973
#> 18                FALSE          0.75 350.997810610678  4.98238966169389
#> 19                FALSE          0.75  777.01163692489 0.531695534222105
#> 20                FALSE          0.75 347.129493159884  9.55192434155334
#> 21                FALSE          0.75 335.637213576894  5.16833105540271
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3400.6937430869 2100.08295910604 81047.5984505028 2.39061695932005
#> 2  1463.23675884398 290.654382831323 42796.4974564614 2.53730475371684
#> 3  2195.21474504597  689.04253833488 65556.1352649144 1.38847908899858
#> 4  7528.90953929585  5989.8609594154 147655.266819109 3.24134957604164
#> 5  2363.98114781632 1135.00864585504 57005.5875598119  2.5982913856659
#> 6  4819.96070839239 3228.01329910914 130294.522472549 2.72034847977658
#> 7  17043.4710049531 15816.2278572482  362668.30323952 11.4815264954533
#> 8  5176.17646577233 3778.40786549304 118704.985410906 8.19941858389188
#> 9    7769.271305653 6300.24038243187 221853.317366682 7.37149384027882
#> 10 20062.5165744483 18783.2956564023 484602.442003929  30.140053408301
#> 11   6798.607223212 5554.15772761943 168459.143178568 14.3394555791448
#> 12 9186.74788122376 7668.07218217523 266835.019205316 6.33159123630918
#> 13   20472.60790324 19383.9110674501 495603.087938908 23.8775897764754
#> 14  6970.2024936754 5596.88943947966 176880.358763791 10.2219260264349
#> 15 9000.38784781189 7403.26691735152 265404.048929901 6.49448549059382
#> 16 20674.4632918853 19368.7129774769 516924.278814471 20.9942252720496
#> 17 6914.41107548024 5973.83398376491 176515.497405471 13.9539954165667
#> 18 9319.88799336405   7652.716701115   270648.8905775 6.74283999924212
#> 19 19782.1501759621 18681.3056136122 503733.006381412 25.5402007416903
#> 20 7142.26042065401 5799.18003546225 180640.979707524 17.5635048578811
#> 21 9254.85673418835   7450.862146567 268249.390795863 6.80345764761961
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.48729906201602 31.9419500244886 49.6972118596689 2134.03406746183
#> 2   2.05194515520507 18.6545919770092 73.3911594240063 1064.66901740173
#> 3   1.51118528514509 5.06442593651636 96.1806272789425 1380.74382181954
#> 4  0.303842871496628 98.6648583289709 115.109868841608 4830.96002513048
#> 5   2.42571998876683 41.1174941535949 149.381209523427 1178.58884432525
#> 6   0.47449341587356 11.9219424887748 91.5656942699807 2236.13260199885
#> 7  0.394313368303829 112.959215671164 166.562877831122 6077.92236137648
#> 8  0.773598065385921 55.5433883548808 142.066950130556 3018.22411478641
#> 9  0.287291338838263  18.957793530162 32.2663464131516 2630.02894141577
#> 10 0.185014876366932  87.053468308418 156.911158487468 10426.3861998073
#> 11 0.767139457098482 67.6976681557686 231.877958927485 2541.03915818015
#> 12  0.33731273459286 9.86982116431692 97.6482630912054 2870.48429811317
#> 13 0.273245038891267 51.6769772099725 185.688454597797 4244.05358024379
#> 14 0.713570420271079 53.8251033474714 216.007419838195 3412.44005160582
#> 15 0.438053772830653 23.9883225736931 103.845058943732 3256.80152966028
#> 16 0.224897531595813  105.88955465963 166.988884092666 6053.06156560109
#> 17 0.640479120120525 67.0856248294757 173.101267340253 2287.10481619015
#> 18 0.342971624137746 7.77578993398622 46.5597788667447  4162.9730116878
#> 19 0.208024863663637 74.6584864229726 141.098341641779 6947.12960265911
#> 20 0.395700653314455 74.5909424418063 110.831391608282 2898.73992422663
#> 21 0.700384348526948 9.98180663949878 200.657434692297  3544.9543142605
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.9189319753606 95.7840617715235   14.7959205162165
#> 2             TRUE 28.2463663656656  36.594099005394   27.7678067978777
#> 3             TRUE 63.0800471451647   67.64814334797    17.377596353604
#> 4             TRUE 237.672968927634 248.337009032811    14.576836987957
#> 5             TRUE 61.9143819192072  70.462760578048   18.4727666441133
#> 6             TRUE 200.083667696907 209.033614195372   11.6046282090394
#> 7             TRUE   490.5098786132 528.284100783241   6.50061044292425
#> 8             TRUE  186.18838049796 213.164467638965   12.0662205476532
#> 9             TRUE 278.058753662061 302.310968396579   4.18363238903658
#> 10            TRUE 619.125678962044 718.286454675355  0.128146025720362
#> 11            TRUE 301.042339513851 348.219148369237   8.73617966570469
#> 12            TRUE 323.386653308473  344.21758847593   3.18815970247406
#> 13            TRUE 684.217922975514 762.775193340119  0.307901033085057
#> 14            TRUE 331.892516546446 365.522653173417   8.82698426543367
#> 15            TRUE 308.350984908758 329.717842172812   2.60765738388644
#> 16            TRUE  743.52490804701 812.595909192053 0.0951906676495876
#> 17            TRUE 309.359948819446  355.26859373995   7.15096934367147
#> 18            TRUE 339.905838811925 362.089782409432    4.4182013399873
#> 19            TRUE  734.99800670481 819.025267144971  0.189494633495423
#> 20            TRUE  318.23752766867 376.021458651098   8.90099676685106
#> 21            TRUE  324.44552574656 346.828901407228   4.01619880207588
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3991344302492 3348.14923529662 3453.23825087719    77537.1124095281
#> 2   34.5187063585024  1432.5499550418 1493.92356264616    41045.1169228355
#> 3   22.3493959417314  2186.8837643804 2203.54572571154    63284.8116780213
#> 4   15.5764800351809 7366.60584734469 7691.21323124701     139708.33757777
#> 5   26.4533854071561 2296.34286993366 2431.61942569898    55066.8089108969
#> 6   13.1657115472634 4800.34911299835 4839.57230378642    126616.084342261
#> 7   7.79790142464385  16857.653095174 17229.2889147322    352670.120955055
#> 8   14.6113581827728 5084.80759192855  5267.5453396161    113740.006742082
#> 9   5.12882089381447 7738.08573529589 7800.45687601012    217526.919758053
#> 10 0.736844968967569 19919.3136190809 20205.7195298156    467451.036705246
#> 11  11.2600684795587 6687.24455909576 6909.96988732824    164279.133763361
#> 12  4.29791859928456 9170.51202540845 9202.98373703906     262113.07253492
#> 13  1.20687721103732 20387.5992757296 20557.6165307505    488621.619799407
#> 14  11.1746309481255 6881.66019866881 7058.74478868199      171266.8948789
#> 15  4.04885429649929 8960.92705717817 9039.84863844562     260046.61041361
#> 16 0.835103546599814 20500.2749744702 20848.6516093004    506966.992539057
#> 17    9.258145648868 6804.05522263575 7024.76692832473    172753.209982838
#> 18  5.54657798340049 9307.09681892264 9332.67916780545    263800.799973273
#> 19 0.873896434948787 19659.3369657963 19904.9633861279    492304.978185038
#> 20  10.2028519162556 7019.55832033724 7264.96252097078    175872.552532171
#> 21  6.32046330872954 9238.43666226638 9271.27680611033    262417.940948904
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84558.0844914775 87.1658876331747 96.5371061137094   14.6424213116815
#> 2     44547.8779900872 27.4471153682448 37.3933500028148   27.1214440739881
#> 3     67827.4588518076 62.6426762321301 68.0855142610046   16.9015729887833
#> 4     155602.196060449 236.651943811181 249.358034149264   14.4811264834355
#> 5      58944.366208727 61.0959201327224 71.2812223645328   17.7086648476517
#> 6     133972.960602837 199.226757925777 209.890523966501   11.4551627830392
#> 7     372666.485523984 486.893197767132 531.900781629309   6.37640173190855
#> 8     123669.964079729 183.605563644034  215.74728449289   11.8225371570566
#> 9     226179.714975311 275.736733102374 304.632988956267   4.09313561730253
#> 10    501753.847302612 609.631562138429 727.780571498969 0.0698663396647782
#> 11    172639.152593774  296.52541100642 352.736076876668   8.49453073671867
#> 12    271556.965875712 321.392202069036 346.212039715368   3.08190619107731
#> 13    502584.556078409 676.696482195925 770.296634119708  0.221828845834308
#> 14    182493.822648683 328.672609848119 368.742559871743   8.60220958304828
#> 15    270761.487446192 306.305221979221 331.763605102349   2.46967044544478
#> 16    526881.565089884 736.911727086314 819.209090152749 0.0243479451969064
#> 17    180277.784828104 304.964440263228 359.664102296169   6.94921842083351
#> 18    277496.981181726 337.781844212164 364.213777009193   4.31016527838391
#> 19    515161.034577786 726.952843471177 827.070430378603  0.123966801441377
#> 20    185409.406882876 312.705023638437 381.553962681331   8.77635106105701
#> 21    274080.840642821  322.30243658756 348.971990566228   3.79557773228989
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5526336347843  3338.0875210389  3463.2999651349    76864.8916782776
#> 2    35.165069082392 1426.67375856904 1499.79975911892     40709.746182354
#> 3   22.8254193065521 2185.28847021039 2205.14101988154    62849.8773741481
#> 4   15.6721905397023 7335.52641697107 7722.29266162063    138186.585169854
#> 5   27.2174872036177 2283.39085927527 2444.57143635737    54695.5534249344
#> 6   13.3151769732635 4796.59370111439 4843.32771567038    125911.702572632
#> 7   7.92211013565955 16822.0709422376 17264.8710676686    350755.575411222
#> 8   14.8550415733694 5067.31142459676 5285.04150694789    112789.266145924
#> 9   5.21931766554852 7732.11403033389 7806.42858097212    216698.460641507
#> 10 0.795124655023153 19891.8917765638 20233.1413723328    464166.725052307
#> 11  11.5017174085447 6665.91979362669  6931.2946527973    163478.706428535
#> 12  4.40417211068132 9167.40303174169 9206.09273070582    261208.869981014
#> 13  1.29294939828807 20371.3210279085 20573.8947785716     487284.74292163
#> 14  11.3994056305109 6864.70529111436 7075.69969623644    170191.976262644
#> 15  4.18684123494094 8953.37073556745 9047.40496005633    259020.717931767
#> 16 0.905946269052495 20466.9197647524 20882.0068190182    505060.278145892
#> 17  9.45989657170596 6782.92325081447 7045.89890014601    172032.771965738
#> 18  5.65461404500387 9304.64744509343 9335.12854163466    262489.463474592
#> 19 0.939424267002833 19635.8195425731 19928.4808093511      490116.6323602
#> 20  10.3274976220497 6996.06217346807 7288.45866783995    174959.449456039
#> 21  6.54108437851552 9235.29239317493 9274.42107520177    261301.280339912
#>    ci95.integral.bt.up
#> 1     85230.3052227279
#> 2     44883.2487305688
#> 3     68262.3931556807
#> 4     157123.948468365
#> 5     59315.6216946894
#> 6     134677.342372467
#> 7     374581.031067818
#> 8     124620.704675887
#> 9     227008.174091857
#> 10    505038.158955551
#> 11    173439.579928601
#> 12    272461.168429618
#> 13    503921.432956186
#> 14    183568.741264939
#> 15    271787.379928035
#> 16    528788.279483049
#> 17    180998.222845204
#> 18    278808.317680408
#> 19    517349.380402624
#> 20    186322.509959008
#> 21    275197.501251813