Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1641133562059  15.6428316939959
#> 2                 FALSE          0.75 32.1571957907379  32.0782972810881
#> 3                 FALSE          0.75 65.1074909756038  20.3173279548951
#> 4                 FALSE          0.75 242.667274502497  14.9311720561609
#> 5                 FALSE          0.75 66.1075989463557  22.5161749053965
#> 6                 FALSE          0.75 204.695406639352  12.5981118016559
#> 7                 FALSE          0.75 509.406476447072  7.04453910097496
#> 8                 FALSE          0.75 202.087653646315  13.6309822195996
#> 9                 FALSE          0.75  289.11859589273  4.77390922970573
#> 10                FALSE          0.75 665.656839217619 0.561046266462658
#> 11                FALSE          0.75 323.827739963688  9.92125265898992
#> 12                FALSE          0.75 332.851172768507  3.70883821664595
#> 13                FALSE          0.75 724.810987118822 0.729614812195231
#> 14                FALSE          0.75 346.447443820792   9.5612446509734
#> 15                FALSE          0.75 319.527600206404  3.35504226167471
#> 16                FALSE          0.75 773.709822862686 0.458750014580135
#> 17                FALSE          0.75 333.560032787822  7.91285339147876
#> 18                FALSE          0.75  351.03237910853  5.10225245892574
#> 19                FALSE          0.75 785.323759839419  0.60508486650787
#> 20                FALSE          0.75 352.502857623222  9.51161355160592
#> 21                FALSE          0.75 337.398895665726  5.34185145430479
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3406.99007889592 2105.82340815711  81303.444028924 3.18460351198517
#> 2  1460.78956412267 255.479347568095 42505.1163680938 2.04535370713849
#> 3  2195.84020714003 672.993300742719  66332.727160771 1.13916985717811
#> 4  7570.65796933301 6067.15384012466 149281.776589619 2.59383353473181
#> 5  2363.52473890452 1095.63665754536 57458.6049968059 2.83286588033182
#> 6  4819.15798416532 3189.24604185286 131489.221614018 2.08296557983923
#> 7   17069.890803796 15892.5646529475 363411.462058033 15.7694528887888
#> 8  5186.12431930348 3736.24486495658 118771.462291003 7.92246385196975
#> 9  7757.39596866318  6261.3145696882 222346.416694842 6.47187789339563
#> 10  20084.574296901 18699.2705016085 489099.939845258 29.6744256063443
#> 11 6803.98969896309 5596.77277047392 167750.525634288 9.10787770400324
#> 12 9187.18071157526 7670.22742726973 267145.345625219 7.36599286406357
#> 13 20434.9041947948 19391.1094906397 489211.827346722 25.2605984924281
#> 14 6939.58799205652 5714.78803871358 175225.295862394 9.25230699411141
#> 15 9005.32964006564  7398.9768129943 266753.310064128 4.16488998476442
#> 16 20687.7135439918 19384.1946847469 517448.149606909 19.2748188097886
#> 17 6896.16612972633 6052.60624102678   175144.5170506 7.91020173457659
#> 18 9320.16136504948 7613.77200844609 269217.603190093 5.67063555688225
#> 19 19823.0552879262 18694.7190695919 503658.660065043 21.6085432408239
#> 20 7111.65060093581 5798.06288182609 179482.660830493 10.4787093256568
#> 21 9256.22086516219 7428.17248131841   267355.9524654 7.22536663845986
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.69388225141405 26.5352582388491 62.5506287031145  1955.6239262578
#> 2   2.49792184058405 23.1766027982963  77.882062173566 974.568077484732
#> 3   1.87749055067752 3.84747417767748 126.244004953942 1001.66096913687
#> 4  0.257350720278652 47.7127966285714 56.7191584148984 2600.43475654013
#> 5   2.86678522864611  33.149199342964 134.475627897636 1256.40796533208
#> 6  0.390866714424301 8.17552658135834 75.8373082565728 1574.64670589359
#> 7  0.501871330460534 107.146957770656 119.678979184044 5140.43715580633
#> 8  0.625807871844163 49.7225019462686 139.216918442452 2335.98191174498
#> 9  0.280098800634146 23.2896087093056 65.7128567207543  2389.8568395021
#> 10 0.427059729131864 79.0050654041224 235.584675239891 3506.62834334388
#> 11 0.702550459559599  62.039535215797  214.58083679528 2370.22361385571
#> 12 0.482937995266916 13.1704552871665 94.4250170258692 4365.46413171683
#> 13 0.315297828956876 99.1329513613397 222.771596881306  9622.4557631189
#> 14 0.619992878488387  79.657066012129 221.114046976327  4422.3756109966
#> 15 0.272134186932751 15.5896447116437 71.8264294888483 1841.75585919831
#> 16 0.243447053911957 81.7944315491109 185.191885296157 4906.87304076033
#> 17 0.433325103197829 106.637856902442 129.729697678443 4766.48741756318
#> 18 0.537248927924722 8.85235895464816 158.263206904469 5917.55535732113
#> 19 0.323466762477752 67.5219271947165 225.633558082554 8847.98460309996
#> 20 0.334635941959803 110.088482518739 137.473822651797 3398.71351128429
#> 21 0.499475466053684 13.9155463273464 159.434274243504 3702.24969340496
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 86.9254405789903 97.4027861334215   14.5013953904198
#> 2             TRUE 28.7925889424951 35.5218026389807   27.9692158533273
#> 3             TRUE 63.2335565605458 66.9814253906618   17.2288559990306
#> 4             TRUE 238.400418337863  246.93413066713   14.5078301213025
#> 5             TRUE 61.4475345732099 70.7676633195015   17.8003132042737
#> 6             TRUE 201.268928260517 208.121885018188   11.9551360564279
#> 7             TRUE 483.465726445014  535.34722644913   6.21896076236739
#> 8             TRUE 189.055200609825 215.120106682805    12.601528270416
#> 9             TRUE 278.472356758095 299.764835027366   4.31314670266256
#> 10            TRUE 616.842409095183 714.471269340056 -0.141466987959259
#> 11            TRUE 308.845281140603 338.810198786774   8.76555715301438
#> 12            TRUE 320.734114507122 344.968231029891   2.91440521443187
#> 13            TRUE 683.257302598778 766.364671638867  0.210949883561171
#> 14            TRUE 331.227398815479 361.667488826106   8.54135636586001
#> 15            TRUE 312.676356181466 326.378844231341   2.90738152417033
#> 16            TRUE 742.002745920584 805.416899804788 0.0582796108949657
#> 17            TRUE 320.547750934444 346.572314641201   7.20003359671833
#> 18            TRUE 341.704183617458 360.360574599601   4.21847797248957
#> 19            TRUE 749.777706208264 820.869813470575 0.0729820422319679
#> 20            TRUE 335.265380782517 369.740334463928   8.96113742708205
#> 21            TRUE 325.513167545459 349.284623785992   4.52021431264648
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7842679975721 3363.33957909301 3450.64057869883    78086.4426702299
#> 2   36.1873787088488 1422.66405251948 1498.91507572587    40901.9518806314
#> 3   23.4057999107597 2189.51111211775 2202.16930216231    64684.9948665409
#> 4   15.3545139910193 7492.17041887902 7649.14551978701     145004.06141511
#> 5   27.2320366065194 2308.99430598535  2418.0551718237    55391.8138938346
#> 6   13.2410875468838 4805.70924293898 4832.60672539165    128898.927782823
#> 7   7.87011743958254 16893.6340582632 17246.1475493287    354955.442936732
#> 8   14.6604361687833 5104.33080360187 5267.91783500509    114928.772046183
#> 9    5.2346717567489 7719.08456233638 7795.70737498999    218415.102193861
#> 10  1.26355952088457 19954.6109643113 20214.5376294908    483331.536220457
#> 11  11.0769481649655  6701.9346635331 6906.04473439307    163851.507789495
#> 12  4.50327121886002 9165.51531262787 9208.84611052265    259964.157128545
#> 13  1.24827974082929 20271.8304898054 20597.9778997842    473382.887616391
#> 14  10.5811329360868 6808.55211846657 7070.62386564647    167950.487982305
#> 15  3.80270299917908 8979.68467451499 9030.97460561629    263723.621675747
#> 16 0.859220418265303 20553.1617040935   20822.26538389    509376.343454858
#> 17  8.62567318623919 6720.74685512182 7071.58540433085    167303.645248709
#> 18  5.98602694536191 9305.59923456909 9334.72349552988      259483.2246273
#> 19  1.13718769078377 19711.9817176909 19934.1288581615    489103.725392944
#> 20  10.0620896761298 6930.55504719248 7292.74615467913    173891.777104431
#> 21   6.1634885959631  9233.3297914537 9279.11193887067    261265.751719749
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1      84520.445387618  85.922290472715 98.4059362396968    14.2828224812244
#> 2     44108.2808555561 28.1483025247465 36.1660890567294    27.1823704735433
#> 3     67980.4594550012 62.8747180555347 67.3402638956729    16.6374464755672
#> 4     153559.491764127 237.583360774422 247.751188230571    14.4267646444147
#> 5     59525.3960997772 60.5551818209053 71.6600160718061    16.8972758572502
#> 6     134079.515445213 200.612794102867 208.778019175837    11.8320130413842
#> 7     371867.481179335 478.498348785046 540.314604109098    6.06087129327232
#> 8     122614.152535824 186.559624496454 217.615682796176     12.404398790785
#> 9     226277.731195823 276.433715221675 301.803476563786    4.22491558046281
#> 10    494868.343470058 607.494965029184 723.818713406054  -0.275990802635796
#> 11     171649.54347908 305.976299663842 341.679180263535    8.54425375825311
#> 12    274326.534121893 318.413826754942 347.288518782071    2.76227974592279
#> 13    505040.767077052 675.300214073663 774.321760163982   0.111631067439755
#> 14    182500.103742484 328.312922112334 364.581965529251    8.34605860913616
#> 15    269782.998452509 311.364415836265 327.690784576542    2.82165925528652
#> 16     525519.95575896   735.9311779955 811.488467729872 -0.0184062110873006
#> 17    182985.388852492 318.056037388052 349.064028187592    7.06353618921102
#> 18    278951.981752886  339.91793341704 362.146824800019    4.04924456019329
#> 19    518213.594737143 742.971015087404 827.676504591434 -0.0289099879485238
#> 20    185073.544556556 331.964587344935  373.04112790151    8.85572710536471
#> 21    273446.153211052 323.237177054344 351.560614277107    4.36287954083957
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0028409067675 3354.98097274778 3458.99918504406    77470.4211334587
#> 2   36.9742240886328 1415.36342263801 1506.21570560734    40594.9629362237
#> 3   23.9972094342231 2188.29915775178 2203.38125652828    64369.4716612628
#> 4    15.435579467907 7477.14088794101 7664.17505072502      144184.9244668
#> 5   28.1350739535429 2298.55230819231 2428.49716961673     54996.045384755
#> 6   13.3642105619275 4803.13395206585 4835.18201626478    128402.914070467
#> 7   8.02820690867761 16859.8827665655 17279.8988410265    353336.205232653
#> 8   14.8575656484142  5088.6682154888 5283.58042311817    114192.937743983
#> 9   5.32290287894866 7711.74833559295 7803.04360173342    217662.297289418
#> 10  1.39808333556111  19929.724368709 20239.4242250931    482226.948292304
#> 11  11.2982515597267 6682.39220994013 6925.58718798605     163104.88735113
#> 12   4.6553966873691 9161.36661921241 9212.99480393811    258589.035927054
#> 13  1.34759855695071 20240.6036101266 20629.2047794631    470351.814051009
#> 14  10.7764306928106 6783.46014267274 7095.71584144029    166557.439664841
#> 15   3.8884252680629 8974.77393643082 9035.88534370046      263143.4685801
#> 16  0.93590624024757 20527.3964581555  20848.030629828    507830.678447019
#> 17   8.7621705937465 6687.15593019755 7105.17632925512    165802.201712176
#> 18   6.1552603576582 9302.81074149837 9337.51198860059    257619.194689744
#> 19  1.23907972096426 19690.7123106245 19955.3982652278    486316.610242967
#> 20  10.1674999978471 6895.87717519908 7327.42402667254    172821.182348376
#> 21  6.32082336777001 9228.94639436059 9283.49533596379    260099.543066327
#>    ci95.integral.bt.up
#> 1     85136.4669243892
#> 2     44415.2697999638
#> 3     68295.9826602793
#> 4     154378.628712438
#> 5     59921.1646088568
#> 6      134575.52915757
#> 7     373486.718883414
#> 8     123349.986838023
#> 9     227030.536100266
#> 10    495972.931398212
#> 11    172396.163917445
#> 12    275701.655323384
#> 13    508071.840642435
#> 14    183893.152059948
#> 15    270363.151548157
#> 16    527065.620766799
#> 17    184486.832389024
#> 18    280816.011690443
#> 19    521000.709887119
#> 20    186144.139312611
#> 21    274612.361864474