Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9952099517254  15.5778030859291
#> 2                 FALSE          0.75 33.0906520146095  32.3590119742361
#> 3                 FALSE          0.75 64.6647033221032  20.3872980851063
#> 4                 FALSE          0.75 241.962407517675  15.0013590639755
#> 5                 FALSE          0.75 65.9523501211116  23.1143431694657
#> 6                 FALSE          0.75 206.399535543369  12.3326804606649
#> 7                 FALSE          0.75 508.581283841252   7.1144219520313
#> 8                 FALSE          0.75 201.991787196377  13.3832662288703
#> 9                 FALSE          0.75  291.63438817662   4.7447382407372
#> 10                FALSE          0.75 669.624477564338 0.613192445498658
#> 11                FALSE          0.75 328.204015187048  9.82913893350642
#> 12                FALSE          0.75 334.354471467357  3.82194636367018
#> 13                FALSE          0.75  735.56276781336 0.891831585463879
#> 14                FALSE          0.75 347.143027856251  9.67744271763977
#> 15                FALSE          0.75  321.53138348915  3.60664946447049
#> 16                FALSE          0.75 772.396675612484 0.465812919894401
#> 17                FALSE          0.75 327.974626206032  7.89470880293074
#> 18                FALSE          0.75 351.588374251472  5.07002607144154
#> 19                FALSE          0.75 796.666780036566 0.981778373487728
#> 20                FALSE          0.75 348.084542073749  9.59477652719288
#> 21                FALSE          0.75 334.566907997235  5.00399083921291
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3395.61705098808 2101.60510192078 80469.7439046917 2.70994264777639
#> 2  1466.65014335581 253.753452125746 42482.8755064257 1.59842972308487
#> 3  2197.14862780077 672.825019157517 65832.1178182617 1.49983811304787
#> 4  7547.18326638506 6018.95891590734  148479.46641616 1.90621528982804
#> 5  2373.47282113415  1084.6189312443 57417.4811383709 2.17650324062339
#> 6   4819.8982089737 3244.99649588462 130036.262202992 2.81694023579628
#> 7  17038.1119315775 15809.3763874084 361965.262386877 11.7313747314195
#> 8  5186.65414158624  3765.2060299692 119248.275769798  9.4407027896237
#> 9  7762.03628142558 6274.03311406768  221503.64845684  7.2398799222671
#> 10 20092.1460749099 18715.6313303046  483340.13607941 27.1704987564737
#> 11 6786.34532510109 5611.94780476207 167696.620186593 11.6518699381872
#> 12 9191.05734992831 7657.93659708387  267007.20527368 5.14288322050658
#> 13 20441.4618297425 19333.9907768344 492231.577085694 23.4258659834483
#> 14 6954.94537239545 5706.85126315996 176624.316104049 12.1310188544812
#> 15 8999.72849944229 7346.59963258917 265261.252332834 7.10806947819815
#> 16 20639.3976365184 19334.8658129183 510368.208759929 25.0059257505736
#> 17 6913.36367523478 6025.37611557033 176271.307185543  14.880412656406
#> 18 9321.82620747309 7635.97165705775  270233.26931594 4.89763194922577
#> 19 19790.6325472417 18449.2881756691 502029.991672761 14.2439653960991
#> 20 7105.35727470103 5751.20824829976 179024.949987201  13.881466101389
#> 21 9259.28470725358 7501.70005915109 268546.748759312 9.36007991498578
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.673170430304176 40.5625951158743 53.5994791569005 2342.76560530532
#> 2   2.38184607979145 20.6584456335416 60.8389105961013 743.997639165871
#> 3    2.1952723978031 5.36208763438974 146.934672652693 1550.75019996752
#> 4  0.343320634632273 84.3194107934953 128.004130826621  4239.9771746057
#> 5   2.90010406486698 28.7963490849507 160.680926652147 897.199285728691
#> 6  0.575661041141324  10.364595169454 113.790939976021 2399.52030739745
#> 7  0.538830520947315 171.925012769074 217.743807164861 8495.93387914458
#> 8  0.750525545497634 48.2858131718136 136.612472014158 2052.83889163283
#> 9  0.338971999926571 26.4609700149744 57.3945436492282 3750.05953503564
#> 10 0.347269149838679 90.5617304926069 215.253609713677 9951.84166940787
#> 11 0.563334185776093 74.9486019654142 122.927455364733 3274.36782557195
#> 12 0.404389814883535 13.0134322076298 99.5068473099256 4185.68203897472
#> 13 0.518877957540204 69.2562830939589 314.209630689545 7283.91856118705
#> 14 0.606686118536526 61.6249658621961  160.90818541618  3686.4059183465
#> 15 0.712909181316958 13.0315368564314 167.386882838369 2908.84254956595
#> 16 0.324828759857311 149.544216363613 221.547787668114 14110.6766296516
#> 17 0.570567385533256 75.9447434397065  116.01657491131 2481.92810487969
#> 18 0.490364585515015  7.6940850870028 147.133179250142 4910.92994497509
#> 19 0.450228090465128  80.122480183059 376.262003524263 9297.40549380043
#> 20 0.244964854084648 94.2702713607903 122.889970720063 3285.95735617298
#> 21 0.539561326557382 8.19783611878229 103.083233488912 4310.93410376111
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.5373542961332 96.4530656073176    14.4704377280787
#> 2             TRUE 30.4612351201349 35.7200689090841    28.4408751729791
#> 3             TRUE 62.1974696261395  67.131937018067    16.7760749907202
#> 4             TRUE 238.826683365908 245.098131669442    14.4365966200054
#> 5             TRUE 62.3720022902862 69.5326979519371    18.3436719827596
#> 6             TRUE 201.765668855484 211.033402231254    11.3857180479875
#> 7             TRUE 489.283172408067 527.879395274437    6.22804574507297
#> 8             TRUE 186.461831107446 217.521743285308    12.1486517065266
#> 9             TRUE 279.724785704491  303.54399064875    4.18712930085799
#> 10            TRUE 624.929007109939 714.319948018737  0.0419346940140306
#> 11            TRUE  309.03668913873 347.371341235365    8.90245419790475
#> 12            TRUE 325.894428569624  342.81451436509    3.15672511818677
#> 13            TRUE 697.027218270587 774.098317356132  0.0382773453102433
#> 14            TRUE  327.18750184063 367.098553871873    8.67944405264718
#> 15            TRUE 309.838609197514 333.224157780786     2.4339138612041
#> 16            TRUE  731.26192775279 813.531423472177 -0.0685303900708759
#> 17            TRUE 303.496347386244  352.45290502582    6.95612545372854
#> 18            TRUE 343.531769694996 359.644978807949    4.26337632826934
#> 19            TRUE 773.235456959983 820.098103113149   0.241153164672593
#> 20            TRUE 325.249530336964 370.919553810534    9.19180934222364
#> 21            TRUE 319.169576537083 349.964239457386    4.11641245702602
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6851684437794 3328.89158202247 3462.34251995369    76615.8944839644
#> 2    36.277148775493 1432.66700028863 1500.63328642299    41258.9993899979
#> 3   23.9985211794924  2188.3279936422 2205.96926195934    63281.1337393151
#> 4   15.5661215079456 7408.47783562976 7685.88869714036    141504.703963933
#> 5   27.8850143561719  2326.1028268894 2420.84281537889    55941.5883133472
#> 6   13.2796428733424 4802.84844991995 4836.94796802745    126089.051297323
#> 7   8.00079815898964 16755.2952855724 17320.9285775827    347989.451155684
#> 8   14.6178807512139 5107.22397891861 5266.08430425388    115871.355793062
#> 9   5.30234718061641 7718.50798575095 7805.56457710021    215334.800521706
#> 10  1.18445019698329 19943.1720282496 20241.1201215703    466969.356533234
#> 11  10.7558236691081 6663.05487486798 6909.63577533419    162310.285113528
#> 12   4.4871676091536 9169.65025394676 9212.46444590986    260121.758319567
#> 13  1.74538582561751  20327.535244053 20555.3884154321    480249.531052541
#> 14  10.6754413826324 6853.57230355214 7056.31844123877    170560.178368369
#> 15  4.77938506773689 8978.29162131346 9021.16537757112    260476.206338798
#> 16  1.00015622985968 20393.3974006002 20885.3978724365    487156.145704152
#> 17  8.83329215213295 6788.43457227647  7038.2927781931    172188.535453016
#> 18  5.87667581461374 9309.16943750497 9334.48297744121    262154.789556456
#> 19  1.72240358230286 19658.8310673406 19922.4340271429     486735.75963546
#> 20  9.99774371216213 6950.28267831253 7260.43187108953    173619.550136296
#> 21  5.89156922139981 9245.79926683818 9272.77014766898    261455.262158625
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84323.5933254189 86.6837223620837 97.3066975413671    14.2583890425329
#> 2     43706.7516228536 29.9577297573631 36.2235742718558    27.6905936578448
#> 3     68383.1018972082 61.7250206205294 67.6043860236771    16.0845641854122
#> 4     155454.228868386 238.226225549612 245.698589485738    14.3284506200962
#> 5     58893.3739633946 61.6864037694898 70.2182964727335    17.4301392023265
#> 6      133983.47310866 200.878332681208  211.92073840553    11.2043848200279
#> 7      375941.07361807  485.58778936767 531.574778314834    6.05831413097457
#> 8     122625.195746534 183.488009728714 220.495564664039    11.9122361596949
#> 9     227672.496391974 277.444223528977 305.824552824264    4.08035312088112
#> 10    499710.915625586 616.370300001649 722.878655127026 -0.0674550881851533
#> 11    173082.955259659 305.366350108201 351.041680265894    8.72500392938528
#> 12    273892.652227794 324.274420355164  344.43452257955    3.02934232649845
#> 13    504213.623118847 689.648070485801 781.477465140918  -0.125169211314921
#> 14    182688.453839729 323.366230901468 370.919824811034    8.48833792530818
#> 15     270046.29832687 307.599567311882 335.463199666418    2.20934746908925
#> 16    533580.271815706  723.38506114136 821.408290083608  -0.170851449425929
#> 17     180354.07891807 298.809017399476 357.140235012588    6.77639672728556
#> 18    278311.749075424  341.98901563099 361.187732871955    4.10891148383211
#> 19    517324.223710063 768.748607860212 824.584952212921  0.0993313161760777
#> 20    184430.349838105 320.876868515027 375.292215632472    9.11464541318697
#> 21    275638.235359999 316.221151363863 352.912664630607    3.94645063916044
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8972171293253 3316.11436456097 3475.11973741519    75877.9233182932
#> 2   37.0274302906273 1426.15958991407 1507.14069679755    41024.6401336606
#> 3   24.6900319848004 2186.63893603736 2207.65831956417    62792.6474263253
#> 4   15.6742675078547  7381.9172212298 7712.44931154031    140169.111153932
#> 5    28.798547136605 2317.03197692764 2429.91366534065    55658.9705383427
#> 6   13.4609761013019 4799.58360244157 4840.21281550583    125333.202400493
#> 7   8.17052977308804 16701.1389065502 17375.0849566049    345313.231983754
#> 8   14.8542962980456 5092.01394776949   5281.294335403    115224.711542198
#> 9   5.40912336059328 7710.17278019623 7813.89978265493     214153.53176817
#> 10  1.29383997918247 19914.6450831444 20269.6470666754     463834.52640737
#> 11  10.9332739376276 6639.44606524888  6933.2445849533    161278.859248472
#> 12  4.61455040084191 9165.55102280135 9216.56367705526     258803.26847729
#> 13  1.90883238224268 20305.7195148784 20577.2041446067    477955.096705767
#> 14  10.8665475099714 6834.16043930555 7075.73030548536     169398.96050409
#> 15  5.00395145985173 8974.18668720368  9025.2703116809    259559.920935685
#> 16  1.10247728921473 20346.2909724457  20932.504300591    482711.282565812
#> 17  9.01302087857592 6764.51197809296 7062.21537237661    171406.728099979
#> 18  6.03114065905097 9306.74580070256 9336.90661424361    260607.846623788
#> 19  1.86422543079938 19633.5924860829 19947.6726084005    483807.076904913
#> 20  10.0749076411988 6920.58754283389 7290.12700656818    172584.473569102
#> 21  6.06153103926538 9243.21694846077 9275.35246604639     260097.31791594
#>    ci95.integral.bt.up
#> 1     85061.5644910901
#> 2     43941.1108791908
#> 3      68871.588210198
#> 4     156789.821678387
#> 5     59175.9917383992
#> 6     134739.322005491
#> 7         378617.29279
#> 8     123271.839997399
#> 9      228853.76514551
#> 10    502845.745751449
#> 11    174114.381124714
#> 12    275211.142070071
#> 13     506508.05746562
#> 14    183849.671704008
#> 15    270962.583729983
#> 16    538025.134954046
#> 17    181135.886271107
#> 18    279858.692008091
#> 19     520252.90644061
#> 20      185465.4264053
#> 21    276996.179602684