Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0671840929814  15.3799020583716
#> 2                 FALSE          0.75 32.2131837212547  31.0002383717417
#> 3                 FALSE          0.75 64.9799330109722  20.5528054894806
#> 4                 FALSE          0.75 242.187627286035  14.9171284617503
#> 5                 FALSE          0.75 65.7850211292554  22.4006264796701
#> 6                 FALSE          0.75 204.202167684627   12.269954602921
#> 7                 FALSE          0.75 506.058627170412  7.06130215612537
#> 8                 FALSE          0.75 201.600080924731   13.445354628044
#> 9                 FALSE          0.75 289.072630439928  4.69573845602086
#> 10                FALSE          0.75 675.986389114928 0.701899263717344
#> 11                FALSE          0.75 328.791467546393  10.0840912456903
#> 12                FALSE          0.75 335.227065210034  3.91888309134603
#> 13                FALSE          0.75 726.547591906358 0.831100451729851
#> 14                FALSE          0.75  349.81404697498  9.99590170506672
#> 15                FALSE          0.75 320.663797909539  3.63769183893278
#> 16                FALSE          0.75 769.571274727896 0.432788701062811
#> 17                FALSE          0.75 337.194103452975  8.15274332606284
#> 18                FALSE          0.75 350.166990894793  4.88884745931672
#> 19                FALSE          0.75 778.584067429087 0.642970968405296
#> 20                FALSE          0.75 353.827673646073  9.70150097053248
#> 21                FALSE          0.75 339.635485655712  5.31802041927723
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3414.7795497698 2137.30152381816 81615.3525218124  1.2594903801206
#> 2  1461.96770360128 292.290581913988 42429.0894967302 2.25304742001807
#> 3  2197.14429410629 655.529171082877 66012.9221313914 1.19740460528511
#> 4  7555.90836319688 6044.76069173659   149106.4896728 2.68779397085818
#> 5  2361.44397172807 1101.00580689347 57362.6012195583 2.95147151678104
#> 6  4820.09712863051 3250.29623064966 129776.479654463 2.30430712814861
#> 7  16971.8198562976 15726.2187574382 358230.670735582 13.3113721778922
#> 8  5169.60896066113 3756.55946647036 118669.461012708 7.12848061407772
#> 9  7762.12891744105 6270.52864628992 221384.609586163 6.95401398866697
#> 10  20086.436940347 18672.3379550851 484801.249074357 29.4040139612236
#> 11 6798.73496551402 5569.07548859565 168140.756215831 13.1249265437472
#> 12 9189.53136976651 7623.22217110631 266287.058851761 6.56317247758625
#> 13 20462.1692981398 19377.8664839461 490395.282709359 23.4400927696268
#> 14 6964.65674396769 5635.98678609093 176292.668252034 15.0035145260508
#> 15 9005.20393569714 7330.50538470185 264608.812128839  4.4801775346432
#> 16 20686.5837066311 19401.4812794699  513978.10596206 17.0466384529175
#> 17 6916.52484159943 6018.30074347726 176480.735485503 15.3844514114922
#> 18 9317.23730137439 7673.11831935337 271872.914113908 6.86980802567811
#> 19 19778.5075427894 18613.5055364402  503081.52273929 26.3617629253416
#> 20 7142.73550960718 5776.48589915492 181022.106782911 12.5379006662627
#> 21 9258.17341532369 7456.46186423401 268565.523897185 7.33341246188388
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.547767931423112 22.3273875194647  60.387570258307 2072.96488234965
#> 2   2.78095361072536 16.5663876340682 90.3360423485133 1138.99173879037
#> 3   1.67789537632899 4.65639814400365  98.412020623323 1384.84304398896
#> 4  0.331480365841443 66.6430790199104 98.2698094529009 3316.16241806412
#> 5     2.171364521206 37.2616366175324 138.025811963722 1210.48162916875
#> 6  0.590312514433816 10.3861020831248 116.551341547419 3008.21807305233
#> 7  0.418300767848305 198.176128635761 236.727367590753 10014.9545494064
#> 8   0.53666724778723 69.4967022801869 123.270739012682 3737.54242950105
#> 9  0.296512063672823 15.7179410933052 69.3469997305824 2737.83498153002
#> 10 0.583165956160105 74.1775450272048 373.328067782299 8074.80632573245
#> 11 0.544417630289219 88.0518149686244 175.961013613978 3203.59582568599
#> 12 0.695835445052897 10.0005764225138 209.883385781089 3704.86469669028
#> 13 0.292848790995863 57.8874998386787  179.78298474662 7084.11172790181
#> 14 0.403988304872111 43.5231215686496 110.818670257449  2978.4092120988
#> 15 0.502751279641276 14.6150707286883 145.818850127695 4376.74880777715
#> 16 0.217967266672528 102.662032469173 157.328258828239 9418.67077914482
#> 17 0.541459933437407 60.1558691950601 167.505324273395 1859.10845206524
#> 18 0.301627880444884 11.9575140212072  58.910067349272 3429.03228344838
#> 19 0.331262703835688 72.5788431124028  201.20851137331  8717.4221038302
#> 20  0.28114280918266 74.8340601353412 120.128346334683 2149.45797709182
#> 21 0.585754249212927 8.76884620369931 162.287135394933 5707.45257802889
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE  89.995322417683 94.1390457682798   14.4788238111806
#> 2             TRUE  28.506920715325 35.9194467271845   26.4255696820984
#> 3             TRUE 63.0102024352782 66.9496635866662   17.7926675954194
#> 4             TRUE 237.766206203973 246.609048368097   14.3718432599412
#> 5             TRUE 60.9298504841505 70.6401917743602   18.8287318422862
#> 6             TRUE 200.411582458823 207.992752910432   11.2988905166774
#> 7             TRUE 484.161419937779 527.955834403045   6.37319739301491
#> 8             TRUE 189.873730314573 213.326431534889    12.562537005434
#> 9             TRUE 277.633277428571 300.511983451285   4.20797611127907
#> 10            TRUE 627.616786148716 724.355992081141 -0.257408734166029
#> 11            TRUE 307.200963381929 350.381971710857   9.18852424386456
#> 12            TRUE 324.430646484404 346.023483935663   2.77423378423401
#> 13            TRUE 687.988639300322 765.106544512394  0.349364190541657
#> 14            TRUE 325.133265579626 374.494828370333    9.3313409435521
#> 15            TRUE  313.29390586505 328.033689954027   2.81066598392288
#> 16            TRUE 741.529554472847 797.612994982945 0.0742325473865035
#> 17            TRUE  311.88668088107 362.501526024879    7.2620417355583
#> 18            TRUE 338.866156692552 361.467825097033   4.39266959598489
#> 19            TRUE   735.2189674169 821.949167441274 0.0980438205955881
#> 20            TRUE  333.20282705007 374.452520242075   9.23902104942701
#> 21            TRUE 327.572022155913 351.698949155511   4.35445467932197
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2809803055626 3378.05099730028 3451.50810223932    78205.3252903472
#> 2   35.5749070613849 1434.71599594324 1489.21941125932    40555.4480864201
#> 3   23.3129433835418 2189.48451915941 2204.80406905318    63734.8553240295
#> 4   15.4624136635595 7446.28049820913 7665.53622818463    143651.402495085
#> 5   25.9725211170539 2300.14857949222 2422.73936396391    55371.3589395758
#> 6   13.2410186891646 4803.01199070377 4837.18226655725    124827.960924292
#> 7   7.74940691923583 16645.8201246917 17297.8195879034    341756.070501809
#> 8    14.328172250654 5055.28688541023 5283.93103591204    112521.203716179
#> 9   5.18350080076266 7736.27290434256 7787.98493053954    216880.871041546
#> 10  1.66120726160072 19964.4148787773 20208.4590019168    471518.192668528
#> 11  10.9796582475161 6653.88972989064 6943.58020113741    162870.841082578
#> 12  5.06353239845805 9173.08042155148 9205.98231798155    260192.556425706
#> 13  1.31283671291805 20366.9443609052 20557.3942353744    478741.918916961
#> 14  10.6604624665813 6893.06120898726 7036.25227894812    171393.185098131
#> 15  4.46471769394268 8981.16214434845 9029.24572704583    257409.060340045
#> 16 0.791344854739119 20517.7046632193 20855.4627500429    498484.392530367
#> 17  9.04344491656737 6817.56843677356  7015.4812464253    173422.502081856
#> 18  5.38502532264856 9297.56719080951 9336.90741193928    266232.156007636
#> 19    1.187898116215 19659.1153458695 19897.8997397093     488741.36337849
#> 20   10.163980891638 7019.63348068455 7265.83753852982    177486.248410595
#> 21   6.2815861592325  9243.7486633186 9272.59816732877    259176.764406328
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     85025.3797532775  89.598582947945 94.5357852380178    14.3062769127823
#> 2     44302.7309070404 27.7972107780193 36.6291566644902      25.54956929472
#> 3     68290.9889387532 62.6330199846134  67.326846037331    17.2641305518758
#> 4     154561.576850516 236.919551103153 247.455703468917    14.2674269447011
#> 5     59353.8434995409 60.0001369563645 71.5699053021462    18.1447520181063
#> 6     134724.998384634 199.685725713456 208.718609655799    11.1129420746307
#> 7     374705.270969356 479.968337701743 532.148916639081    6.24143265114269
#> 8     124817.718309238 187.628258921138 215.571902928323     12.393486822381
#> 9      225888.34813078 275.442763022141 302.702497857715    4.11457481122213
#> 10    498084.305480187  618.35452175093 733.618256478927  -0.441106010356463
#> 11    173410.671349085 303.066611520649 354.516323572138    9.01703269032346
#> 12    272381.561277817 322.363247153964 348.090883266103    2.55504561904235
#> 13    502048.646501758 680.605010077889 772.490173734826   0.257116821377961
#> 14    181192.151405936  320.40715850392 379.220935446039    9.20408462751739
#> 15    271808.563917632 311.882649941638 329.444945877439    2.65229933083588
#> 16    529471.819393753 736.159863360178 802.982686095614 0.00557285838465726
#> 17    179538.968889151  307.04057868645 367.347628219499    7.09148185652552
#> 18    277513.672220181 336.702167164463 363.631814625122    4.29765681364475
#> 19    517421.682100091 726.915012095418 830.253122762756 -0.0063039311126537
#> 20    184557.965155228 329.253388340198 378.401958951948    9.15046106453447
#> 21    277954.283388043 325.261997230419 354.008974081004     4.1699420908199
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4535272039609 3371.01787023165 3458.54122930795    77552.3413524071
#> 2   36.4509074487634 1429.49758383851 1494.43782336406    40196.6656887011
#> 3   23.8414804270854 2188.01775374405 2206.27083446854     63298.629765173
#> 4   15.5668299787996 7425.28792831786  7686.5287980759    142606.811333395
#> 5   26.6565009412338  2288.4111639577 2434.47677949843    54990.0572263876
#> 6   13.4269671312113 4799.74036854758 4840.45388871343    123880.372231281
#> 7   7.88117166110805 16583.3946441715 17360.2450684237    338601.359818746
#> 8   14.4972224337069 5033.39542419197  5305.8224971303    111343.877850886
#> 9    5.2769021008196 7731.32175289817 7792.93608198393    216018.453022364
#> 10  1.84490453779115 19941.0489520937 20231.8249286003    468974.628675922
#> 11  11.1511498010572 6626.15340817552 6971.31652285253    161861.708397487
#> 12  5.28272056364971 9169.93023997838 9209.13249955464    259025.524046248
#> 13  1.40508408208174  20348.709798456 20575.6287978236    476510.423722672
#> 14  10.7877187826161 6879.35142569313 7049.96206224224     170454.98619632
#> 15  4.62308434702968 8976.55839706891 9033.84947432537    256030.384465596
#> 16 0.860004543740966 20485.3661229915 20887.8012902707    495517.511234936
#> 17  9.21400479560015 6798.61933797711 7034.43034522175    172836.882919455
#> 18   5.4800381049887 9293.80057389283 9340.67402885596    265152.010838349
#> 19  1.29224586792325 19636.2530102891 19920.7620752897    485995.375415783
#> 20  10.2525408765305 6996.06075174192 7289.41026747245    176809.169147811
#> 21  6.46609874773457 9240.98647676444 9275.36035388294    257378.916844249
#>    ci95.integral.bt.up
#> 1     85678.3636912177
#> 2     44661.5133047593
#> 3     68727.2144976097
#> 4     155606.168012206
#> 5     59735.1452127291
#> 6     135672.587077646
#> 7     377859.981652419
#> 8     125995.044174531
#> 9     226750.766149962
#> 10    500627.869472793
#> 11    174419.804034176
#> 12    273548.593657274
#> 13    504280.141696047
#> 14    182130.350307747
#> 15    273187.239792082
#> 16    532438.700689184
#> 17    180124.588051551
#> 18    278593.817389467
#> 19    520167.670062797
#> 20    185235.044418011
#> 21    279752.130950122