Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.6170631028232  15.4606801535949
#> 2                 FALSE          0.75 32.9308397844563  31.5028054275702
#> 3                 FALSE          0.75 65.1117103340992   20.312467609959
#> 4                 FALSE          0.75 242.763662356108  14.9950754375098
#> 5                 FALSE          0.75 66.2952463043789  22.6769958468588
#> 6                 FALSE          0.75 206.049958749923   12.429333779687
#> 7                 FALSE          0.75  504.95073364711  6.98352512254965
#> 8                 FALSE          0.75 207.341679193762  13.5288835539165
#> 9                 FALSE          0.75 291.644818758161  4.65556442522215
#> 10                FALSE          0.75 663.150613425174 0.531320394835316
#> 11                FALSE          0.75 324.758482462361  10.0064022946832
#> 12                FALSE          0.75 334.741754394618  3.82081572219249
#> 13                FALSE          0.75 721.459545069061 0.720741200191882
#> 14                FALSE          0.75 342.630143132016  9.78203532119863
#> 15                FALSE          0.75 320.001435838518  3.42731187089584
#> 16                FALSE          0.75 775.776069783035 0.594780964065961
#> 17                FALSE          0.75 336.266856344491   8.1388308021743
#> 18                FALSE          0.75 349.374177414096  4.96661222058685
#> 19                FALSE          0.75 784.865738633292 0.764774310690306
#> 20                FALSE          0.75 350.812871384631  9.45775460239349
#> 21                FALSE          0.75 339.090232477366  5.56022181731262
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3413.19322388463 2134.19457212783 81661.1377296532 2.42967683954306
#> 2  1466.19186115363  279.09710090311 42650.5997081648 2.17269153486662
#> 3  2195.36581781465 659.354540029217 66248.5642474058 1.49502603653341
#> 4  7553.01032566567  6035.2811206431 148604.002939741  3.2900573443998
#> 5  2370.67357069786 1151.20551240755 57239.5785385525  2.5553067815662
#> 6  4817.19746530309 3224.30464446224 130161.782679806 3.00050075850206
#> 7  17066.9459278793 15858.6252111715 363940.147290235 12.2491002567901
#> 8  5183.29886279124 3774.36923978671 118493.392831235 9.41856638505881
#> 9   7763.2220456618 6305.39949447743 221766.103392852 4.38198182235657
#> 10 20117.5906660804 18721.2123009513 484024.848633965 32.9133960552599
#> 11 6749.37400255624 5499.50330864788 165923.339911328 15.3359554828017
#> 12 9189.15366469252 7647.12148934643 267743.120877135 6.45765083642592
#> 13 20473.5465870949 19412.9596308175  493487.52986888 23.6830009775754
#> 14 6930.62192506687 5613.60174099813 175224.042892579 16.7230518074489
#> 15 9002.66236819871 7386.92103770432 264144.365349572 5.66366755709459
#> 16 20619.8678424346 19225.5573826262 507373.751321214 23.4403647189799
#> 17 6875.54335340122 5972.67382783046 174892.772506191 8.37211493983876
#> 18 9320.22378013718 7632.92076270202 270740.597810224 8.98156504302304
#> 19 19803.0411520959 18580.8954661518 504056.196923617  30.041278463286
#> 20 7110.92367346327 5805.51393946719 178693.198272476 12.5210123802329
#> 21 9256.88664069993 7366.02964947664 268345.799616717 7.86669695317459
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.530792266192247 44.5447353920064 59.2240083798596 2538.22733191477
#> 2   2.70973440253852 25.9558366489761 103.428047586158 900.682331847474
#> 3   1.81337944181136 4.38311726409364 116.480044465687 1274.94675381332
#> 4  0.377619130536976 68.4634899684572 118.716040803667 3366.31368050937
#> 5    2.6481637023054 34.4740418656723 142.205705664431 1350.89141504879
#> 6  0.482121197467598 10.8543267249052 96.5385794835203 2603.81260766256
#> 7  0.560108171369724 87.1422633381099 129.925300003774 5071.62001582433
#> 8  0.625406427278523 33.9143683185808 124.347492585908 2168.14200574617
#> 9  0.292892701906295 20.0962804167376 71.2684445036365 2716.33244227402
#> 10 0.428267688304682 81.0978882562837 267.408207544641 8167.63314375261
#> 11 0.715097105404583 105.829115581538 204.909446667617 4332.00602718419
#> 12 0.442913620968664 11.6231508406465 92.2163213642196 2220.64159580732
#> 13 0.319726630227444 61.6341106279942  226.94512219739 4573.25702499331
#> 14 0.518845797923623 97.5342693991277 164.032760805204 5492.53058005366
#> 15 0.372075714522998 18.1195189646134 106.079161464857 4483.63579419069
#> 16 0.478718502312303 120.398090068608 308.090122799882 14629.4152855797
#> 17 0.522166599897025 107.100559566741  172.07761221176  4561.0813427353
#> 18 0.519668132025865 10.2087443749468 96.5901631156597 3291.94748241459
#> 19 0.343710337397994  84.246585716574 222.956044065208 7481.81637403179
#> 20 0.376870150308114 125.150006358124 160.867266882319  4994.6110773682
#> 21 0.921453187594243 11.2363823272137 257.071400029138 4062.90928678435
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.6202447017749 96.6138815038716   14.5875268757086
#> 2             TRUE 29.3567622096007 36.5049173593119   27.0452923353943
#> 3             TRUE 62.6523925040018 67.5710281641967   17.3294584281793
#> 4             TRUE  237.35151802457 248.175806687646   14.3738919677765
#> 5             TRUE 62.0917666487025 70.4987259600553   18.3207665565664
#> 6             TRUE 201.114135002187 210.985782497659   11.6362444098528
#> 7             TRUE  484.80096372469 525.100503569529   6.06214718064646
#> 8             TRUE  191.84813749034 222.835220897184   12.5000899810433
#> 9             TRUE 284.436458660385 298.853178855938   4.17375593058629
#> 10            TRUE 609.008076914272 717.293149936077 -0.173179952425885
#> 11            TRUE 299.530835693152  349.98612923157   8.83006755629264
#> 12            TRUE 324.118918768697 345.364590020538   3.09222281569904
#> 13            TRUE 682.501008460949 760.418081677172  0.194790893467736
#> 14            TRUE 315.120722908763 370.139563355269   8.92853398361427
#> 15            TRUE 310.684702707097 329.318168969938   2.81524732050551
#> 16            TRUE 737.216669820313 814.335469745757 -0.192710972237778
#> 17            TRUE 322.494727268456 350.038985420526   7.27986674534369
#> 18            TRUE 334.599502918323 364.148851909868    4.1117581434043
#> 19            TRUE 735.447835561187 834.283641705398  0.199370805670606
#> 20            TRUE 330.215806019148 371.409936750114   8.83780320513664
#> 21            TRUE 326.149515989394 352.030948965338   4.04443132372009
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3338334314811 3339.91713416478 3486.46931360448    77485.7537686534
#> 2    35.960318519746 1423.49450986607  1508.8892124412    41168.9772722757
#> 3   23.2954767917387 2188.15558991522 2202.57604571409    64151.2768373829
#> 4   15.6162589072432 7440.38788466756 7665.63276666378    143066.416935303
#> 5   27.0332251371512 2313.96377182883 2427.38336956689    55017.3621607972
#> 6   13.2224231495212 4799.34209784062 4835.05283276556    125878.510940201
#> 7   7.90490306445285 16923.5969046881 17210.2949510705    355597.332364204
#> 8   14.5576771267897 5127.50972690718 5239.08799867531    114926.799231782
#> 9     5.137372919858 7730.16366437627 7796.28042694734    217297.736525311
#> 10  1.23582074209652 19984.1846398988  20250.996692262    470589.092112492
#> 11  11.1827370330737 6575.28510742461 6923.46289768787     158797.18999661
#> 12  4.54940862868594 9170.03358155965 9208.27374782538    264090.165452032
#> 13  1.24669150691603 20372.1584751119  20574.934699078    485964.522062766
#> 14   10.635536658783 6770.17805190531 7091.06579822844     166188.83008839
#> 15  4.03937642128617 8972.85575950192  9032.4689768955    256768.784468128
#> 16   1.3822729003697 20421.8129842717 20817.9227005974    483308.363176435
#> 17   8.9977948590049 6699.36293291393 7051.72377388851    167389.793697392
#> 18  5.82146629776939 9303.43039564039 9337.01716463397    265325.344201652
#> 19  1.33017781571001 19664.4555185921 19941.6267855996    491748.608988334
#> 20  10.0777059996503 6905.05191300415 7316.79543392238    170477.063050206
#> 21  7.07601231090515 9238.40279177166  9275.3704896282    261662.313839956
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1      85836.521690653 87.8548964973188 97.3792297083276   14.4203273118581
#> 2     44132.2221440539 28.6723643761177 37.1893151927949   26.1917259985947
#> 3     68345.8516574287 62.1814593024937 68.0419613657047   16.7582439040087
#> 4     154141.588944178 236.315149961084 249.212174751132   14.2549419416574
#> 5     59461.7949163077 61.2868450125091 71.3036475962487   17.4865949903402
#> 6     134445.054419411 200.168977263259 211.930940236587   11.4843762326505
#> 7     372282.962216266 480.942497143801 528.958970150418   5.88571310666499
#> 8     122059.986430687 188.881289079047 225.802069308477   12.3030869564506
#> 9     226234.470260393 283.056134386342  300.23350312998   4.08149472948581
#> 10    497460.605155438 598.640357156865 727.660869693484  -0.30808427424186
#> 11    173049.489826046  294.70000971607 354.816955208652    8.6048119680902
#> 12    271396.076302238 322.084758755223 347.398750034013   2.95270502509391
#> 13    501010.537674994 675.040863153013 767.878226985109 0.0940770049460913
#> 14    184259.255696767 309.852961589416 375.407324674616   8.76509755726833
#> 15    271519.946231016 308.900647426612 331.102224250423   2.69804347043076
#> 16    531439.139465992 729.832954933835 821.719184632236 -0.343507300466154
#> 17    182395.751314991 319.857511062407 352.676201626575   7.11538426637613
#> 18    276155.851418796  331.77030992977 366.978044898421   3.94806268181615
#> 19    516363.784858899 725.984832845252 843.746644421333 0.0911020493902377
#> 20    186909.333494747 326.271687119375 375.354055649888   8.71908910778958
#> 21    275029.285393477 323.671506449144 354.508958505588    3.7541735696279
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5010329953317  3325.8855425163 3500.50090525297    76686.2121591003
#> 2   36.8138848565456 1415.31842132164 1517.06530098563    40885.2623377438
#> 3   23.8666913159093 2186.77490797703 2203.95672765228    63749.6686099317
#> 4   15.7352089333623 7418.82188532749 7687.19876600384    142006.028125942
#> 5   27.8673967033774 2303.10444864114 2438.24269275458    54591.8313650568
#> 6   13.3742913267235 4795.92298492228 4838.47194568391    125058.309968788
#> 7   8.08133713843431 16896.1470917366  17237.744764022    353999.772059219
#> 8   14.7546801513824 5116.82670088682 5249.77102469566    114243.834499972
#> 9   5.22963412095848   7723.833336045 7802.61075527861    216442.091805995
#> 10  1.37072506391249 19958.6388050981 20276.5425270627     468016.28767221
#> 11  11.4079926212762 6541.94893601643 6956.79906909605    157432.608098047
#> 12  4.68892641929107 9166.37228904485 9211.93504034018    263390.663349353
#> 13  1.34740539543767 20352.7437302641 20594.3494439258    484523.946099893
#> 14  10.7989730851289 6739.45475704458 7121.78909308916    164458.682955673
#> 15  4.15658027136092 8967.14811102807 9038.17662536935    255356.439192958
#> 16  1.53306922859808 20383.8875859001  20855.848098969    478700.097361478
#> 17  9.16227733797247 6665.62625665041 7085.46045015204     165953.05307443
#> 18  5.98516175935754 9300.21464116229 9340.23291911208    264288.380744691
#> 19  1.43844657199037 19637.9178440914 19968.1644601003    489391.836830514
#> 20  10.1964200969974 6865.62966100135 7356.21768592519    168903.760560835
#> 21  7.36627006499733 9234.86333133859 9278.90995006127    260382.497414619
#>    ci95.integral.bt.up
#> 1     86636.0633002061
#> 2     44415.9370785859
#> 3     68747.4598848799
#> 4     155201.977753539
#> 5     59887.3257120481
#> 6     135265.255390825
#> 7     373880.522521251
#> 8     122742.951162497
#> 9     227090.114979709
#> 10     500033.40959572
#> 11    174414.071724609
#> 12    272095.578404918
#> 13    502451.113637867
#> 14    185989.402829484
#> 15    272932.291506186
#> 16     536047.40528095
#> 17    183832.491937952
#> 18    277192.814875757
#> 19    518720.557016719
#> 20    188482.635984118
#> 21    276309.101818814