Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8762241643169  15.3672784560059
#> 2                 FALSE          0.75 32.7275044723585  31.1608612348947
#> 3                 FALSE          0.75 64.9170264249556  20.6055220616854
#> 4                 FALSE          0.75 242.574351060953  15.0006254542463
#> 5                 FALSE          0.75 65.2408344934224  21.5853899153771
#> 6                 FALSE          0.75 206.437743795225  12.2682273973659
#> 7                 FALSE          0.75 510.566151053762  7.12104915525614
#> 8                 FALSE          0.75 201.916063653555  13.4284148209472
#> 9                 FALSE          0.75 288.906637955599   4.6840345947154
#> 10                FALSE          0.75 669.003048313122 0.566058323438562
#> 11                FALSE          0.75 329.517211415682  10.1241063498274
#> 12                FALSE          0.75 335.185273778197  3.86653650776144
#> 13                FALSE          0.75 732.795088880059 0.765723918103858
#> 14                FALSE          0.75 342.203864651301  9.65593454355756
#> 15                FALSE          0.75 319.610010989417  3.46105794896159
#> 16                FALSE          0.75 775.318516430293 0.523504783806373
#> 17                FALSE          0.75 331.859290265041   7.9618230081151
#> 18                FALSE          0.75 353.009882066469  5.12187381754258
#> 19                FALSE          0.75 772.855407757769 0.582138493178468
#> 20                FALSE          0.75 351.498595005605  9.48606405041844
#> 21                FALSE          0.75 337.805776650821  5.14427202727589
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.90833095573 2141.39937007132 81763.8289770792  2.7823515358927
#> 2  1466.09835239093 300.267226765091 42390.9952093032 2.02381664980362
#> 3  2196.07832093271 653.153954597602 66339.1508986466 1.19205959300324
#> 4  7526.80271080058 6006.24239742879 147420.497450319 2.64811929335007
#> 5  2360.19468045523 1178.52591093525 57409.6235123753 2.81358436682908
#> 6  4814.24537047166 3253.51115688863 130570.395259877 2.44228220742879
#> 7  17025.8334656403 15808.3545265322 361654.587990045 12.1257923226685
#> 8  5183.84546408602 3769.41681191272 119327.407457715 8.70249728244811
#> 9  7763.72836195055  6279.4424420176 221564.779062314 4.76958198374535
#> 10 20097.0288246017  18730.674261954 484667.922610675 25.7484839480733
#> 11 6766.11951041903 5483.48528521017 167129.184234239 12.1273979862419
#> 12 9194.13797867882 7651.40129825842 267244.747911814 6.56850960038452
#> 13 20472.0374820498  19449.266304756 492032.637786866 17.8845047722595
#> 14 6972.43336483015  5712.4446664577 175872.133201651 15.5713559413046
#> 15 9011.21425187649 7386.46178775009 265072.328518962 6.08084484578907
#> 16  20691.938985023 19345.1452196531 514470.852775564 27.2832512983514
#> 17 6895.76707386599 6034.87264519511 175263.539807705   16.06289571689
#> 18 9320.98137870539 7629.67981409541 271805.818891812 6.52390416381076
#> 19 19798.2804712891 18626.2723668701 500734.081392896 18.0494316791883
#> 20 7128.16613620513 5820.04614791172 180097.082056447 16.5984979998798
#> 21 9256.03329176854 7486.34699396506  267598.04508346 10.4511221422799
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.592013878929685 19.4662128928186 55.4534597388804 1208.89897247559
#> 2   2.47315031986496 23.6039274435714 62.4201808048839 1343.80474775279
#> 3   1.97440600138469 4.96715038816627 129.280653386592 1133.44226114279
#> 4  0.359201052407859 76.0593255866335 131.108110884307  3829.0378955647
#> 5   2.37865954529539 29.1698704199145 140.644809387751 1007.33721717569
#> 6  0.496540768782063 12.3616348535128 93.6402838646025 1841.31595876333
#> 7  0.392773738810024 117.590641393158 162.811592102992 6875.85431307664
#> 8  0.631100299819971 55.6946094070733 130.111872297248 2862.37864908942
#> 9  0.359348486775538   24.02017771766 82.6223781705186 3822.35348480416
#> 10 0.475586664655595 84.0186233902084 335.014035843468 9885.51313973117
#> 11   0.7582936558605 118.871774505888 213.012381333947 4472.19021592412
#> 12 0.503681687250255   12.58588538224 141.443876834955 3028.14864071375
#> 13 0.560317039933295 73.1864894019602 378.460214703557 8188.08763811084
#> 14  0.56558032076169 69.0265729586135 193.427133290619 3591.86394128453
#> 15 0.453294953200454  10.288271364467 117.832771979231 4445.44619419808
#> 16  0.38241033942769 105.924674007001 198.970923825878 7081.80064560354
#> 17 0.660534033643524 128.123912024713   204.8762821281 4897.99649672783
#> 18 0.334625443556376 12.0381517003166  96.637541060692 3049.54017905579
#> 19 0.269394084959816 66.5077021976983 243.315783966338 10295.4022516516
#> 20 0.419668385324585 70.2836844901701 116.728793072349 3320.01253608812
#> 21 0.610586699522601 8.32392467663991 129.151531842603  4224.7415951324
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.2992558877734 97.4531924408604   14.3934156251666
#> 2             TRUE 29.3983260834315 36.0566828612854   27.0925289587168
#> 3             TRUE 62.9560883944653 66.8779644554459   17.3576241894076
#> 4             TRUE 238.218194823392 246.930507298513   14.4097397230354
#> 5             TRUE 60.6124882099886 69.8691807768562   17.6724949633662
#> 6             TRUE 202.420189564005 210.455298026445   11.4514178327194
#> 7             TRUE 490.619222682972 530.513079424552   6.47493635491365
#> 8             TRUE 187.600455623927 216.231671683182   12.3902548277434
#> 9             TRUE 281.060675592338  296.75260031886   4.09290633396964
#> 10            TRUE 626.646792218542 711.359304407703 -0.216281739919892
#> 11            TRUE 309.567641728314  349.46678110305   8.87671328593688
#> 12            TRUE 324.380075485565  345.99047207083   3.03798013223477
#> 13            TRUE 703.375078529692 762.215099230425 -0.155997612586413
#> 14            TRUE 316.588984127855 367.818745174747   8.72555491590458
#> 15            TRUE 309.607021218094  329.61300076074   2.71538775094684
#> 16            TRUE 730.437568044505 820.199464816081 -0.105560224552178
#> 17            TRUE 305.435826810757 358.282753719325    6.8752445227715
#> 18            TRUE    342.278059717 363.741704415938   4.57141496289234
#> 19            TRUE 743.164092645504 802.546722870033   0.13898522341957
#> 20            TRUE 324.194065795803 378.803124215407    8.7957095565595
#> 21            TRUE 320.613680726771 354.997872574871   4.13985690656121
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3411412868452 3380.88641074704 3444.93025116441    79775.1901673569
#> 2   35.2291935110725 1427.26989174625  1504.9268130356    40180.4363992498
#> 3   23.8534199339632 2187.90735854418 2204.24928332124    64474.6383790667
#> 4   15.5915111854572 7401.68512021057 7651.92030139059    141121.730112116
#> 5    25.498284867388 2312.21024361447 2408.17911729599    55752.5537901213
#> 6   13.0850369620124 4793.91048113763 4834.58025980569    127541.430507711
#> 7   7.76716195559863 16832.3968605486 17219.2700707321    350343.807645034
#> 8   14.4665748141511 5092.22783161138 5275.46309656066    114618.794579963
#> 9   5.27516285546116   7724.215169605  7803.2415542961    215277.007579811
#> 10  1.34839838679702 19958.8181891248 20235.2394600786    468406.253495817
#> 11  11.3714994137179 6570.57544135684 6961.66357948121    159772.431329043
#> 12  4.69509288328811 9173.43419722504 9214.84176013261     262263.44339784
#> 13  1.68744544879413 20351.6457069836 20592.4292571161    478563.233622173
#> 14  10.5863141712105 6858.88465231323 7085.98207734707    169963.517018237
#> 15  4.20672814697633 8994.29004548194 9028.13845827104    257759.569529506
#> 16  1.15256979216492 20517.6928962815 20866.1850737645    502821.290713546
#> 17   9.0484014934587 6685.00323858533 7106.53090914664    167206.335570587
#> 18  5.67233267219282 9301.17861915837 9340.78413825241    266789.325297265
#> 19  1.02529176293737 19688.8753011739 19907.6856414044    483798.144688929
#> 20  10.1764185442774  7012.5494752188 7243.78279719146    174635.661434582
#> 21  6.14868714799057 9242.34043567546 9269.72614786161    260648.345159467
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83752.4677868016 87.4228151539672 98.3296331746666   14.2069312533037
#> 2     44601.5540193565 28.7608238387434 36.6941851059736   26.3134866079594
#> 3     68203.6634182265 62.5805896226693  67.253463227242   16.7356862989714
#> 4     153719.264788523 237.384037245986 247.764664875919   14.2965913915269
#> 5     59066.6932346293 59.7262091344374 70.7554598524074   16.9232172065981
#> 6     133599.360012043 201.650870668665 211.224616921786   11.2950074905531
#> 7     372965.368335056 486.799598101332 534.332704006192   6.35121262718849
#> 8     124036.020335468 184.859168979956 218.972958327153   12.1914582333001
#> 9     227852.550544817 279.558257267458  298.25501864374   3.97971156063535
#> 10    500929.591725533 618.536019774899 719.470076851346 -0.366091539286404
#> 11    174485.937139434 305.747511362648 353.286911468716   8.63785078434082
#> 12    272226.052425789 322.310994961443 348.059552594951   2.87932040075094
#> 13    505502.041951558  697.74145952643 767.848718233687 -0.332497480165401
#> 14    181780.749385064 311.684007006344 372.723722296258   8.54739711486464
#> 15    272385.087508418  307.69155509167 331.528466887163    2.5725998406887
#> 16    526120.414837582 721.843343885525 828.793688975062   -0.2260194814719
#> 17    183320.744044822 300.376014659937 363.342565870145   6.66717630217379
#> 18    276822.312486359   340.2230299054 365.796734227538   4.46600794817208
#> 19    517670.018096863  737.47852166656 808.232293848978 0.0541260866572283
#> 20    185558.502678312 318.965538925841  384.03165108537   8.66351401518226
#> 21    274547.745007453 317.321577251952  358.28997604969   3.94752209621159
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5276256587081  3374.7545536858 3451.06210822565    79394.3869910271
#> 2     36.00823586183 1419.83465460153 1512.36205018033    39757.1379037077
#> 3   24.4753578243994  2186.3427061719 2205.81393569352    64117.6040668067
#> 4   15.7046595169657 7377.72643265078 7675.87898895038    139915.583175013
#> 5    26.247562624156 2303.02173443219 2417.36762647826    55435.2425667109
#> 6   13.2414473041788 4790.01656615877 4838.47417478454    126961.415980701
#> 7   7.89088568332378 16795.3558085097 17256.3111227709    348177.913536415
#> 8   14.6653714085944 5074.68402964816 5293.00689852388      113717.1453055
#> 9   5.38835762879545 7716.64881362394 7810.80791027716    214072.966232098
#> 10  1.49820818616353 19932.3523227569 20261.7053264465    465292.316856802
#> 11   11.610361915314 6533.13083238749 6999.10818845057    158363.691411027
#> 12  4.85375261477194 9169.46964332963 9218.80631402801    261309.576576015
#> 13  1.86394531637312  20328.591962822 20615.4830012777    475983.986016168
#> 14  10.7644719722505 6837.14128183127 7107.72544782903    168832.079876733
#> 15  4.34951605723448 8991.04924000214 9031.37926375084    256359.253978334
#> 16  1.27302904908465 20484.3266239693 20899.5513460767    500590.523510181
#> 17  9.25646971405641 6644.64420629755 7146.88994143443    165663.466674118
#> 18  5.77773968691308 9297.38660137277 9344.57615603801    265828.720140863
#> 19  1.11015089969971 19667.9253749816 19928.6355675966    480555.092979659
#> 20  10.3086140856546  6990.4101146044 7265.92215780586    173589.857485714
#> 21  6.34102195834019 9239.71839940232 9272.34818413475       259317.551557
#>    ci95.integral.bt.up
#> 1     84133.2709631314
#> 2     45024.8525148986
#> 3     68560.6977304865
#> 4     154925.411725626
#> 5     59384.0044580396
#> 6     134179.374539053
#> 7     375131.262443675
#> 8     124937.669609931
#> 9      229056.59189253
#> 10    504043.528364548
#> 11     175894.67705745
#> 12    273179.919247613
#> 13    508081.289557563
#> 14    182912.186526568
#> 15     273785.40305959
#> 16    528351.182040947
#> 17    184863.612941291
#> 18    277782.917642762
#> 19    520913.069806134
#> 20     186604.30662718
#> 21    275878.538609919