Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8093827721733   15.295250833016
#> 2                 FALSE          0.75 32.3972094234851  32.3285154334713
#> 3                 FALSE          0.75 64.8448253445116  20.6125544944584
#> 4                 FALSE          0.75 242.440511671682  14.9615997142653
#> 5                 FALSE          0.75   64.95526622944  23.1018724113592
#> 6                 FALSE          0.75 205.620365208611  12.5710046259788
#> 7                 FALSE          0.75 507.151391101821  7.06876023540435
#> 8                 FALSE          0.75 201.811024649383  13.2967325384555
#> 9                 FALSE          0.75 290.585756034615  4.71300413785535
#> 10                FALSE          0.75 665.511850046541 0.470268988975198
#> 11                FALSE          0.75 323.226325535441   9.8221161950837
#> 12                FALSE          0.75 332.572840611874  3.69000005336621
#> 13                FALSE          0.75 718.349957311491 0.649634975283135
#> 14                FALSE          0.75 342.172196251852  9.89594647976264
#> 15                FALSE          0.75 318.566151782307  3.18008038548727
#> 16                FALSE          0.75 774.389613912744 0.455237878242207
#> 17                FALSE          0.75 332.784480385142  8.00592452889293
#> 18                FALSE          0.75  348.60717198338  5.03606809906177
#> 19                FALSE          0.75 781.761425161131 0.800230204123869
#> 20                FALSE          0.75 355.779959043562  9.48027414065057
#> 21                FALSE          0.75 338.865369448568  5.17526350096898
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3409.2509867004 2138.55048504386 81349.3870155309 2.91661482430288
#> 2  1464.12619796377 269.644228845216 42704.3759344302 2.40953583989183
#> 3  2195.68158842553 654.007813322175 66424.5931347228 1.23645054492251
#> 4  7553.34186949873 6038.43370243985 148319.843115345 2.56020082118043
#> 5  2370.98875255232   1088.399576087 57348.1532190172  3.0140508233047
#> 6  4821.56520426751 3200.82907714584 130988.072618444 3.33794063466378
#> 7     17017.1578493 15787.1351380332 361075.799755709 13.8267445397896
#> 8  5180.02405741194  3786.0729702664 118463.736636933 7.70721032429043
#> 9  7761.57326991202 6271.83746460624 222352.951987585 7.45290264828525
#> 10 20117.5235494217 18809.7734800959 482224.597283001 28.9735052272446
#> 11 6782.59707083577 5560.50527546347 168068.554138125 10.7203615586709
#> 12 9190.51290332363 7673.28721682222 267096.373446339 5.14585339252187
#> 13 20419.1572333778 19379.3858616213 490094.101631844 25.2303994871866
#> 14 6973.22905097651  5587.0042831674  177780.51771988 15.4869865007395
#> 15  9002.8403558057 7440.28956088955 265597.217626949 4.24492896869356
#> 16 20627.2993949081 19319.6415468758 509746.012855541 22.5081651758037
#> 17 6887.36014438896 6015.76677450166 175350.558511498 12.7293084497903
#> 18   9323.310396519 7626.68552768803 272269.104210111  4.8323528445304
#> 19 19822.8093706447 18526.9424431623 502066.555498631 29.3619414363483
#> 20 7121.06574457612 5831.95215708687 179942.687367352 13.5819760658479
#> 21 9256.66090303114 7486.50284138599 267063.011148381  8.4734121313532
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.548332089184036 24.0787547112366 59.2679801456772 1619.46466396537
#> 2   2.53112539148618 25.3268508451779 60.4907593660246 680.458708640731
#> 3   1.77407643548389 4.14708164867773 102.614060688357 887.279623495073
#> 4  0.283697215304517 50.8203075075932 87.3543252898509 2707.15319161333
#> 5   2.61725430580152 30.7524925025458 151.115708063314 1023.83557173738
#> 6  0.325731571351171 10.6965558362946 63.3847544675435 2511.93037484556
#> 7  0.532589727842909 169.071137025349 212.162192949849  8871.9888554213
#> 8   0.93926009704167 44.7539229588369 175.596941867117 2558.19757674185
#> 9  0.320549240058369 23.8842920724315 77.6110225249995 2542.12108771097
#> 10 0.227109221369308 91.9165059211864 200.822052175193 10262.9912060343
#> 11 0.529445915038851 85.9889874232119 162.193739322848 3138.58131758453
#> 12 0.345819984217366 12.3532980069962 87.1276262099017 4414.68992146157
#> 13 0.238611759641958 66.0795444098249 164.521119403395 7804.37399702292
#> 14 0.452368095423672 40.3048131585096 160.125428858567 2605.15275899115
#> 15 0.259575740843577 19.3449232596953 56.7859627618021 3173.33818308557
#> 16  0.28246219390905 156.556467186086 271.264392521276 16902.7355603947
#> 17 0.561368560862745 81.1210516057026 157.600918302791 3028.03033336668
#> 18 0.320088362183701 9.06119783145404  95.776198277078 2339.89453702581
#> 19   0.5011031481271 88.4668412897031 303.693216349992 9276.23909459936
#> 20  0.43824148712428 96.9888009935234 131.560044683237 4030.05780791907
#> 21 0.584654008745835 7.13031663227312 160.420567458936 5583.73198085884
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE  87.011551386195 96.6072141581515     14.3932445463082
#> 2             TRUE  28.433522966863 36.3608958801071     28.1648141644765
#> 3             TRUE  62.810864198114 66.8787864909091     17.6941987580874
#> 4             TRUE 238.228981320841 246.652042022524     14.4949177950894
#> 5             TRUE 59.9971526251037 69.9133798337762     18.7964890783158
#> 6             TRUE 200.129452864589 211.111277552633     12.0351761911061
#> 7             TRUE 484.406396333867 529.896385869774     6.19265013310277
#> 8             TRUE 189.132663665925  214.48938563284     11.7516496788219
#> 9             TRUE 278.325731178186 302.845780891044     4.18570063795934
#> 10            TRUE 617.850433947724 713.173266145358   0.0966743198226854
#> 11            TRUE 305.591330771427 340.861320299455     8.95117766484479
#> 12            TRUE 324.107911781176 341.037769442573     3.12112617932864
#> 13            TRUE 676.845950155069 759.853964467913    0.257118630672114
#> 14            TRUE 316.696103458135 367.648289045568     9.15180096279069
#> 15            TRUE 311.583243628806 325.549059935808     2.75307829179959
#> 16            TRUE 737.363682198547 811.415545626941 -0.00941243073818099
#> 17            TRUE 311.844767985237 353.724192785047     7.08247324627371
#> 18            TRUE 340.657951554127 356.556392412632     4.50952274326958
#> 19            TRUE 733.461031498338 830.061818823924  -0.0240844745452108
#> 20            TRUE 333.437608415242 378.122309671882     8.75936689433113
#> 21            TRUE 324.926606492492 352.804132404644     4.21350765658208
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.1972571197237 3369.64143520041 3448.86053820038    78685.3676433079
#> 2    36.492216702466 1422.46352832345 1505.78886760409    41585.0213587162
#> 3   23.5309102308294 2188.85963911346 2202.50353773761    64965.0181540734
#> 4   15.4282816334413 7469.74246364874 7636.94127534872    143866.576115141
#> 5   27.4072557444027 2320.40090238563 2421.57660271901    55663.9437035092
#> 6   13.1068330608515 4803.96936991681 4839.16103861821    126855.947151823
#> 7   7.94487033770594 16739.0358288933 17295.2798697067    346481.378088541
#> 8    14.841815398089 5106.40385414465 5253.64426067922    114255.501623193
#> 9   5.24030763775137 7722.28360945287 7800.86293037117    218171.162798301
#> 10  0.84386365812771 19966.3208971813  20268.726201662    465341.976749074
#> 11  10.6930547253226 6641.14518652458 6924.04895514695    162905.587870699
#> 12  4.25887392740377 9170.19172810212 9210.83407854513    259834.208525535
#> 13  1.04215131989416 20310.4563828236  20527.858083932    477255.906406742
#> 14  10.6400919967346 6906.92763333076 7039.53046862226    173495.041431339
#> 15  3.60708247917496  8971.0179570435  9034.6627545679    260377.076315774
#> 16 0.919888187222595  20369.764006387 20884.8347834292    481941.012858692
#> 17  8.92937581151214 6753.91601449758 7020.80427428034     170369.44861311
#> 18  5.56261345485396 9308.40472608626 9338.21606695175    268419.977696704
#> 19  1.62454488279295 19677.2814167231 19968.3373245662    486807.142188016
#> 20    10.20118138697 6961.51916694177 7280.61232221046    173313.242273325
#> 21  6.13701934535588 9244.93153217105 9268.39027389123    257877.772039868
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1      84013.406387754 86.0928177165396 97.5259478278069    14.2205199382153
#> 2     43823.7305101442 27.6745191772971  37.119899669673    27.3675096661584
#> 3     67884.1681153722 62.4213822764635 67.2682684125597      17.13536468091
#> 4     152773.110115549 237.422518062169 247.458505281196    14.4055531722685
#> 5     59032.3627345252 59.0477266157628 70.8628058431172    17.9720539719883
#> 6     135120.198085065  199.07800156467 212.162728852552    11.9325707461305
#> 7     375670.221422877 480.050971803833 534.251810399808    6.02488436883225
#> 8     122671.971650673 186.704892413773 216.917156884992    11.4557827482538
#> 9      226534.74117687 275.978066843976 305.193445225254    4.08472762734095
#> 10    499107.217816927 608.723779801142  722.29992029194  0.0251349150913532
#> 11    173231.520405552 302.214416880446 344.238234190436    8.78440220160755
#> 12    274358.538367143 322.486967962531 342.658713261217    3.01219288430017
#> 13    502932.296856947 668.898374316606 767.801540306377   0.181955926384897
#> 14     182065.99400842 311.817702710402 372.526689793301    9.00930501273224
#> 15    270817.358938125 310.246091003667 326.886212560946    2.67131193343386
#> 16     537551.01285239 730.273610168169  818.50561765732 -0.0983880218195318
#> 17    180331.668409886 307.835035823553 357.733924946731    6.90564214960195
#> 18    276118.230723519   339.1357604081 358.078583558659    4.40869490918172
#> 19    517325.968809247 724.212019945888 839.310830376374  -0.181931966205247
#> 20    186572.132461379   329.1592859545 382.400632132624    8.62132082588698
#> 21    276248.250256893 322.257481671116 355.473257226021    4.02934164382714
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.3699817278167 3362.05662746637 3456.44534593442    78175.2362741588
#> 2   37.2895212007842 1414.48557030722 1513.76682562032    41370.6768654943
#> 3   24.0897443080068 2187.55330839412 2203.80986845694    64685.5250726724
#> 4   15.5176462562622 7453.73406678385 7652.94967221361    143013.822859783
#> 5   28.2316908507302 2310.71386724733 2431.26363785731     55341.435498412
#> 6   13.2094385058271 4800.59995482837 4842.53045370665    126064.689083747
#> 7   8.11263610197646 16685.7784207303 17348.5372778697    343686.701599083
#> 8   15.1376823286571 5092.30636841262 5267.74174641126    113449.669386519
#> 9   5.34128064836975 7714.76005745006 7808.38648237399    217370.394655672
#> 10 0.915403062859042 19937.3671978162 20297.6799010272    462109.134519174
#> 11  10.8598301885598 6614.05865548627 6951.13548618526     161916.93475566
#> 12  4.36780722243225 9166.30043922991 9214.72536741734    258443.581200274
#> 13  1.11731402418137 20289.6413263345 20548.6731404211    474797.528597679
#> 14   10.782587946793 6894.23161718583 7052.22648476719    172674.418312257
#> 15  3.68884883754068  8964.9243062167  9040.7564053947    259377.474788102
#> 16  1.00886377830395 20320.4487192234 20934.1500705928    476616.651157168
#> 17  9.10620690818391 6728.36288324179 7046.35740553614      169415.6190581
#> 18  5.66344128894182 9305.55044876935 9341.07034426865    267682.910917541
#> 19  1.78239237445298 19649.4143617169 19996.2043795725    483885.126873217
#> 20  10.3392274554142 6930.96769462881 7311.16379452342     172043.77406383
#> 21  6.32118535811082 9242.68548243188 9270.63632363039    256118.896465897
#>    ci95.integral.bt.up
#> 1     84523.5377569031
#> 2      44038.075003366
#> 3     68163.6611967731
#> 4     153625.863370908
#> 5     59354.8709396225
#> 6     135911.456153142
#> 7     378464.897912335
#> 8     123477.803887347
#> 9     227335.509319499
#> 10    502340.060046828
#> 11    174220.173520591
#> 12    275749.165692404
#> 13    505390.674666009
#> 14    182886.617127502
#> 15    271816.960465797
#> 16    542875.374553915
#> 17    181285.497964897
#> 18    276855.297502682
#> 19    520247.984124046
#> 20    187841.600670873
#> 21    278007.125830864