Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.0441345727525   15.594479820668
#> 2                 FALSE          0.75  32.661606681345  31.4722645211441
#> 3                 FALSE          0.75 64.7273188382502  20.3512617944584
#> 4                 FALSE          0.75  242.87782389365  15.0611851014519
#> 5                 FALSE          0.75 66.0243440078427  22.9594593683799
#> 6                 FALSE          0.75  205.02414049487  12.3345698882059
#> 7                 FALSE          0.75 507.755425923568  7.00600963921171
#> 8                 FALSE          0.75 203.119188237943  13.2851695217286
#> 9                 FALSE          0.75 289.463101695324  4.68854068379787
#> 10                FALSE          0.75 673.409137460804  0.63513857439728
#> 11                FALSE          0.75 327.774231344276  9.95549313166676
#> 12                FALSE          0.75 336.512612779605  3.82547556616791
#> 13                FALSE          0.75 725.429884810564 0.680262312769594
#> 14                FALSE          0.75 348.032428029812  9.86420200184535
#> 15                FALSE          0.75 319.964239393248  3.65331701557066
#> 16                FALSE          0.75 773.536561512044 0.379411582398788
#> 17                FALSE          0.75 337.594412334755  8.13349228421786
#> 18                FALSE          0.75 349.262928332818  4.79850282662746
#> 19                FALSE          0.75  779.50122885751 0.597031284260358
#> 20                FALSE          0.75 356.741022940181  9.60264783532843
#> 21                FALSE          0.75 338.500100737609  5.15435335463203
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3420.73898194669 2124.97465450692 82172.5580453571 2.46419286472986
#> 2  1456.36152523272 283.111596912824 42444.4883162729 2.60804402164974
#> 3  2196.03112760994 684.808548309425  66099.671186449 1.18642650585673
#> 4  7561.68980122974 6027.83863634789 148998.918221544 2.38384419027274
#> 5  2368.93698531745 1076.19885932925 57331.3650967715 2.31792484540801
#> 6   4817.3997141997 3236.16077992461 130623.133092841 2.92160349051633
#> 7  17032.4297763017 15839.7006572466 362078.060698442 15.3553097663928
#> 8  5182.84727333308 3790.76719390705 118793.269458466 7.55356698339678
#> 9  7769.28711526788 6274.93263206364 222119.423223931 6.07133117202718
#> 10 20059.2805946151 18685.1740159163   486179.5568056 25.8636511820567
#> 11 6777.95760984135 5547.31274021555  167148.91711561 14.3969495388389
#> 12 9185.58419951303 7666.97338395032 266107.207050877 6.13156566587293
#> 13 20443.4118869889  19443.437402101 490226.976097437  23.509281796284
#> 14 6959.16677949912 5633.92056129101 176636.012892555 13.5843551957145
#> 15 9004.32479796223 7317.11526805713 265335.496112644 7.26123433809978
#> 16 20710.3231504049 19449.7910008652 515269.757617504 23.1307925483886
#> 17 6895.34112988927 5977.19876652841 175756.038059378 15.8252242487688
#> 18  9320.6846810281 7696.77511154514  272033.02037525 7.26645414317199
#> 19 19767.8736919207 18645.1047239301  498945.21715619 23.2286632006148
#> 20 7111.62486019609 5779.22412309465  180089.36770954 13.5863797123418
#> 21 9257.03252088135 7496.84307752972 269029.713380049 7.05841578421038
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.617103693099538 16.2470302270933 55.5526302234623 1107.60252403601
#> 2    2.0540110818741  23.544845750253  83.740836275243 720.594720910857
#> 3   1.66865041485714  5.4171048977983 109.792336277233 919.842278395448
#> 4   0.34128656447334 53.0902687960509 91.5299871501633 2720.06856683816
#> 5   2.72806139230433 28.1640973201632 141.497634619396 977.682470443022
#> 6  0.439745570678843 9.20389420058138 86.0046889588301   1823.845495499
#> 7  0.562094842777655 133.040509648128 153.238488220732 6752.42615300996
#> 8  0.585370716716793 30.2358289843404 90.0081291262965  2655.5352522331
#> 9   0.31914454352669 19.3582414358199 67.9541192916776 3822.83684976057
#> 10 0.348776570839949 82.2715830060721 243.003564757577 6641.98521725803
#> 11 0.687265479726785 75.5796376442967 226.204912315874 2476.53577870228
#> 12 0.447562626009234 14.3113698524126 111.853913449584 4637.43247589839
#> 13 0.283417636343518 104.007968871855 218.912765592999 11779.7787326646
#> 14 0.670621973970765 56.0601968708562  205.94938820942 2982.69722647723
#> 15 0.709315764974156 15.7535144392955 172.969697161392 3318.47405810723
#> 16 0.261057867862662  108.52927470309 230.534425582274 7140.07560404067
#> 17 0.814242000467083 72.7724425147666 206.854387246061 2798.66563673042
#> 18 0.413785278528304 10.6120583352211 54.6156467934206 3552.43396237132
#> 19 0.152478328131338  64.580841833874  124.25267251836 13420.8333056897
#> 20 0.317408798135413 91.3642207082261 145.230403419782 3491.82904903175
#> 21 0.532069539621927 8.94128562700228 124.408957000503 3203.61616060343
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.9905373102719 96.0977318352331    14.5793442455193
#> 2             TRUE 28.3713742657311 36.9518390969588    28.0934162914612
#> 3             TRUE 62.7756472361159 66.6789904403845    17.6063318620184
#> 4             TRUE 238.956400200651 246.799247586648    14.4997687028932
#> 5             TRUE 62.2113576371465 69.8373303785389    18.4717983780392
#> 6             TRUE 200.218102752971 209.830178236769    11.6111884244392
#> 7             TRUE 482.495941357852 533.014910489284    6.08136362284247
#> 8             TRUE 190.693570550256 215.544805925631    12.3222346927295
#> 9             TRUE 279.475761917339 299.450441473309    4.16354790969647
#> 10            TRUE 630.863431266321 715.954843655288  0.0614011153655644
#> 11            TRUE 304.091249352886 351.457213335666     8.8249414175162
#> 12            TRUE 326.426187259244 346.599038299966    3.08923504638272
#> 13            TRUE 686.757116255677 764.102653365451   0.214040300984507
#> 14            TRUE 325.686163732861 370.378692326762    8.76102885466344
#> 15            TRUE 308.019508907074 331.908969879423    2.48649258218817
#> 16            TRUE 735.486407769945 811.586715254143 -0.0500286102352898
#> 17            TRUE 311.561918445531  363.62690622398    6.79406419344951
#> 18            TRUE   337.3096112673 361.216245398336     4.1178260434484
#> 19            TRUE 741.290077892498 817.712379822521   0.346204434484308
#> 20            TRUE 334.391428313379 379.090617566983    9.08051036239567
#> 21            TRUE 326.889006772583 350.111194702635    4.27909896195396
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6096153958167 3394.01261722312 3447.46534667025    80350.5518933178
#> 2    34.851112750827 1417.63025397356 1495.09279649189    41259.1100003745
#> 3   23.0961917268984 2187.11999005306 2204.94226516682    64586.5306384885
#> 4   15.6226015000105 7474.35630906024 7649.02329339925    144524.405429095
#> 5   27.4471203587205 2322.60704522578 2415.26692540912    55723.0774328927
#> 6   13.0579513519726 4802.25930823974 4832.54012015966    127622.907252745
#> 7   7.93065565558095 16813.5781379305 17251.2814146729     350970.31967674
#> 8   14.2481043507277 5133.10933465384 5232.58521201232    114424.913968542
#> 9   5.21353345789928 7737.44280810595  7801.1314224298    215830.856606075
#> 10    1.208876033429 19923.9438405701 20194.6173486601     475253.49112321
#> 11  11.0860448458173 6653.62910591649 6902.28611376622    163075.015759645
#> 12   4.5617160859531 9162.04199610581 9209.12640292024    258478.630628024
#> 13  1.14648432455468 20272.3187781947 20614.5049957831    470849.240082204
#> 14  10.9673751490273 6866.94775564656 7051.38580335168       171729.475955
#> 15  4.82014144895315 8978.41026670958 9030.23932921487    259876.606287057
#> 16 0.808851775032867 20531.7924935183 20888.8538072914    503524.333248857
#> 17  9.47292037498621 6775.63046195248 7015.05179782607    171152.233086957
#> 18  5.47917960980652 9303.22784506666 9338.14151698954     266189.26650715
#> 19 0.847858134036408  19661.638207104 19874.1091767374     476867.94636833
#> 20  10.1247853082612 6961.33071713106 7261.91900326113    174345.308923883
#> 21   6.0296077473101 9242.32410602493 9271.74093573777    263759.764795857
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83994.5641973963  87.214316557882 96.8739525876231    14.3849565821929
#> 2     43629.8666321713 27.5498403989115 37.7733729637785    27.4464028006708
#> 3     67612.8117344096  62.401922886771 67.0527147897294    17.0807069813384
#> 4     153473.431013993 238.205489280715 247.550158506584    14.3922634350841
#> 5     58939.6527606502  61.481211310843 70.5674767048424    17.6124590394634
#> 6     133623.358932937 199.297797653458 210.750483336282    11.4726685696753
#> 7     373185.801720143 477.659018781438 537.851833065698    5.90430374736751
#> 8     123161.624948389 188.314196950486 217.924179525401    12.1378429169637
#> 9     228407.989841787 277.563292598151 301.362910792497    4.06301737848556
#> 10    497105.622487989 622.716381143973 724.101893777635 -0.0484635044490195
#> 11    171222.818471576 299.556210248152   355.9922524404    8.60845279140226
#> 12     273735.78347373 324.494744074494 348.530481484716    2.94825281918981
#> 13    509604.712112671 679.351692489847 771.508077131281   0.124763745536299
#> 14     181542.54983011 321.407091846211 374.657764213412    8.54978293286265
#> 15     270794.38593823 305.732220090573 334.196258695924    2.26305811622131
#> 16    527015.181986151 728.200208117202 818.872914906886  -0.132261838612028
#> 17      180359.8430318 306.576972807169 368.611851862342    6.53757796330238
#> 18    277876.774243351 335.020678212201 363.505178453435    3.98748368071199
#> 19    521022.487944049 733.973048984304 825.029408730715   0.298173761122936
#> 20    185833.426495198 330.111718703991 383.370327176371    8.98052659098302
#> 21    274299.661964242 324.665605800557 352.334595674662    4.11149705697305
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8040030591431 3388.89480270158 3452.58316119179    80001.6570982465
#> 2   35.4981262416173 1410.21362756223 1502.50942290322    41032.1226632876
#> 3   23.6218166075784 2185.41360201025 2206.64865320962     64296.780320794
#> 4   15.7301067678196 7457.63287438948    7665.74672807    143667.583830541
#> 5   28.3064596972964 2313.73535456993 2424.13861606497    55415.1074547031
#> 6   13.1964712067364 4799.36008156656 4835.43934683284    127048.395921663
#> 7   8.10771553105591 16771.6703773914  17293.189175212    348843.305438542
#> 8   14.4324961264935 5123.58504852378 5242.10949814239    113588.420364089
#> 9   5.31406398911019 7731.34496205367 7807.22926848208      214626.6629984
#> 10  1.31874065324358 19898.0282919232  20220.532897307    473161.265779774
#> 11  11.3025334719313 6629.82152005853 6926.09369962417    162294.906989354
#> 12  4.70269831314601  9157.5339146023 9213.63448442376    257017.839398116
#> 13  1.23576088000289     20239.556268 20647.2675059777    467138.609781415
#> 14   11.178621070828 6849.28879363224   7069.044765366     170789.92632866
#> 15  5.04357591492001 8973.44790966121 9035.20168626325    258831.286958754
#> 16 0.891085003409605 20497.6057719868 20923.0405288229    501275.209433584
#> 17  9.72940660513335 6752.70714256033 7037.97511721822    170270.653411387
#> 18  5.60952197254294 9299.88504669106 9341.48431536513    265070.249809003
#> 19  0.89588880739778 19641.2952419263 19894.4521419151    472640.383877038
#> 20  10.2247690796738 6932.55098760797 7290.69873278422    173245.382773438
#> 21  6.19720965229101 9239.50760105242 9274.55744071027    262750.625705267
#>    ci95.integral.bt.up
#> 1     84343.4589924676
#> 2     43856.8539692582
#> 3     67902.5620521041
#> 4     154330.252612547
#> 5     59247.6227388398
#> 6     134197.870264019
#> 7     375312.815958341
#> 8     123998.118552842
#> 9     229612.183449461
#> 10    499197.847831425
#> 11    172002.927241867
#> 12    275196.574703638
#> 13     513315.34241346
#> 14    182482.099456451
#> 15    271839.705266534
#> 16    529264.305801424
#> 17     181241.42270737
#> 18    278995.790941498
#> 19    525250.050435341
#> 20    186933.352645643
#> 21    275308.801054832