Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 90.9905170028456  15.5206506152889
#> 2                 FALSE          0.75 33.4059797055901  30.9126973030873
#> 3                 FALSE          0.75 64.1509688347428  19.7484370447585
#> 4                 FALSE          0.75 242.513078637151  15.0281602619717
#> 5                 FALSE          0.75 65.8448734469251  22.0897125284634
#> 6                 FALSE          0.75 204.458240539434  12.4299578926755
#> 7                 FALSE          0.75 508.892433178152  7.08951672788614
#> 8                 FALSE          0.75 203.880774995925  13.2707873800052
#> 9                 FALSE          0.75 290.740416937162  4.70803079427144
#> 10                FALSE          0.75 671.777843342457 0.572932896332284
#> 11                FALSE          0.75 324.822341518943  10.1007194272427
#> 12                FALSE          0.75  334.80558482132  3.76772644068201
#> 13                FALSE          0.75 731.809264358929 0.746101604514925
#> 14                FALSE          0.75 343.638342099949  9.65869311236059
#> 15                FALSE          0.75 319.433738880829  3.52754087991862
#> 16                FALSE          0.75 776.996133349435  0.48643886032234
#> 17                FALSE          0.75 333.453499491923  8.24049892615253
#> 18                FALSE          0.75 351.571581250235  5.10304742793517
#> 19                FALSE          0.75 780.398100522647 0.734007213124811
#> 20                FALSE          0.75 356.725068316976  9.69164264200321
#> 21                FALSE          0.75 336.126934570737  5.00767115760032
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.58243820296 2113.61896896155 81674.1761722183 2.60951942963007
#> 2  1471.47923111114 317.409852149246  42692.362937714 2.65536600624906
#> 3  2194.71244636187 720.808675080534 65909.7034720813 1.66764429306705
#> 4  7566.46676298501 6037.14495444923 148624.879164988  2.0107512869546
#> 5  2366.74450609626 1150.86280560874 57197.8712805024 2.75941835450413
#> 6  4821.00092071659 3223.52305981833 129694.050331558 1.97122162751446
#> 7  17004.7547455334 15790.7390821163 360792.819021783 15.8131468738085
#> 8  5184.16123833306 3822.25743612557 118519.552714931 7.08614892152363
#> 9  7761.72493393821 6278.19593968834 222090.680542968 6.46752304079339
#> 10 20098.8804301879 18735.3363377333 485893.297186822  23.808282769753
#> 11 6756.50079888842 5484.20359345308 166665.887716982 11.9977649094136
#> 12 9187.48768558247  7674.4752245559 267372.943079538 6.65964270608013
#> 13 20409.5636862706 19367.8684241136 489140.982937318 28.1031499103731
#> 14 6945.78732498676 5651.31615113759 175860.214269277 14.2248690414445
#> 15 9007.52737930841 7348.73573760454 265448.879108446 5.44034793900751
#> 16 20657.6393803462 19335.6150590472 514239.382283059 17.4502460049156
#> 17 6902.28730101761 5967.19032440369 175595.230242703 14.0901696442823
#> 18 9317.96591686614 7628.76036248931 270293.556956971 7.53701903317187
#> 19 19778.2717330324 18552.2948503146 504148.293152473 23.5587579866277
#> 20  7124.5534554146 5788.95666466444 179869.017434543  12.842262580423
#> 21 9260.33126456603 7513.28134581352 267420.736817347 6.77947350343818
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.585515155187468 16.6748128611708 51.4608919679394 1177.68115159091
#> 2   2.96677106833102 24.1144741085963 81.4428419138508 830.508728618781
#> 3   2.51169567595666 4.39842441787458 167.041437369404 1221.24114011127
#> 4  0.310415176163563 37.1145903661229 76.6522206875246 3062.04639557109
#> 5   2.38725637817941 23.8139444472342 117.851652362536 1180.07870157639
#> 6  0.417830440808574 9.36160943061246 90.4705366403166 4086.58753470763
#> 7  0.491718412653341 268.383338816419 266.951206013031 15115.9190722271
#> 8  0.766201581604561 36.4764597927995 160.605270280173 2073.21937333675
#> 9  0.489399535336048 22.6873075761185 98.5321804003542 3574.63391075709
#> 10 0.355727445649728  72.221751027959 181.455711860845 8555.89430991159
#> 11 0.645347776063842 109.482941828745 199.182190066127 4731.61089676473
#> 12 0.389579987459896 15.7296674331617 102.434861365829 4058.41488373066
#> 13 0.440723730525297  127.66206274462 245.493444925013 12741.2747507211
#> 14  0.48964483271909 63.6838871072756 189.871569499833 3917.40184901692
#> 15 0.678621368500604 14.1507203456317 171.962612900661 4463.41990244457
#> 16 0.378968486092087 101.052734216888 272.132687853126 6554.46821938214
#> 17 0.681623454922626 85.7362023643711 170.585854644573 3404.80282278482
#> 18 0.423469213686298 13.3168751035124 90.8326162176428  4776.4944221698
#> 19 0.357322256420793 68.9296512264257 234.989319198161 9328.19838504832
#> 20 0.351733266457987  89.520088880626 142.878632715749  2750.9534872506
#> 21 0.401460893568644   10.28854526982 93.3702266607776  5302.9430824632
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 86.6978575411041 95.2831764645871    14.5574781850055
#> 2             TRUE 29.0379026253104 37.7740567858698    26.0323588956828
#> 3             TRUE 61.4076939726475 66.8942436968381    15.6166976578098
#> 4             TRUE 239.205392770111 245.820764504191    14.5175272971826
#> 5             TRUE 61.3056302537658 70.3841166400844    18.1626757863583
#> 6             TRUE 201.215580962172 207.700900116695    11.7426268175454
#> 7             TRUE 482.879806570737 534.905059785567    6.28063993907139
#> 8             TRUE 192.224060020019 215.537489971832    12.0103857782657
#> 9             TRUE 280.101341535057 301.379492339267    3.90296855864365
#> 10            TRUE 632.613218186214 710.942468498701 -0.0122387517615181
#> 11            TRUE 305.086018242957 344.558664794928    9.03912233561769
#> 12            TRUE 323.850472569818 345.760697072822    3.12686736131048
#> 13            TRUE 685.579582756366 778.038945961493  0.0211110678008114
#> 14            TRUE 320.238432526773 367.038251673125    8.85322736253769
#> 15            TRUE 310.484366521162 328.383111240496    2.41120872873513
#> 16            TRUE 748.290478671348 805.701788027521  -0.136964299299143
#> 17            TRUE 310.275170427079 356.631828556768    7.11922834280481
#> 18            TRUE 339.173184940668 363.969977559803    4.40644057142121
#> 19            TRUE 741.643943634644  819.15225741065   0.146212101312605
#> 20            TRUE  335.59954637218 377.850590261771    9.11304141867982
#> 21            TRUE 324.974700657581 347.279168483893     4.3472679876799
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4838230455723 3385.15237104634 3440.01250535959    79736.8906778513
#> 2   35.7930357104919  1431.8109212025 1511.14754101979    41326.1760791361
#> 3   23.8801764317072 2187.47703819446 2201.94785452927    63900.7617965982
#> 4   15.5387932267607 7505.41326183274 7627.52026413729    143587.812844273
#> 5   26.0167492705686 2327.57056748056 2405.91844471196    55256.6418164092
#> 6   13.1172889678056 4805.60107320324 4836.40076822995    122971.613836964
#> 7   7.89839351670089 16563.2641531804 17446.2453378864    335927.132147969
#> 8   14.5311889817447 5124.15746197391 5244.16501469222    115109.106845792
#> 9   5.51309302989924  7724.4043129755 7799.04555490093    216210.407759773
#> 10  1.15810454442609 19980.0756497469 20217.6852106289    471818.851047017
#> 11  11.1623165188677 6576.40135958014 6936.60023819671    158882.387791804
#> 12  4.40858552005354 9161.61238265492 9213.36298851002    260696.850595801
#> 13  1.47109214122904 20199.5595930557 20619.5677794854    468181.585972381
#> 14  10.4641588621835 6841.02733069529 7050.54731927822    169416.088227644
#> 15  4.64387303110211 8984.24944433984 9030.80531427697    258106.553368925
#> 16  1.10984201994382 20491.4076325595  20823.871128133    503457.282062176
#> 17  9.36176950950025 6761.25124812822 7043.32335390701    169994.329599222
#> 18  5.79965428444913 9296.05965732086 9339.87217641141    262436.223632502
#> 19  1.32180232493702  19664.882456765 19891.6610092999    488803.406809069
#> 20  10.2702438653266 6977.29290920597 7271.81400162323    175343.698948016
#> 21  5.66807432752074 9243.40660759718 9277.25592153489    258697.395446695
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83611.4616665854 85.8758589207707 96.1051750849205   14.3730409111215
#> 2     44058.5497962919 28.2014623333419 38.6104970778382   25.0978260091585
#> 3     67918.6451475643 60.8823860203314 67.4195516491542   14.8255135198835
#> 4     153661.945485702  238.57200611472 246.454151159582   14.4197465166911
#> 5     59139.1007445955  60.436413472097 71.2533334217532   17.4106900272318
#> 6     136416.486826152 200.594646149505 208.321834929362   11.6110102286907
#> 7     385658.505895596 477.898665305488 539.886201050817   6.12574863908559
#> 8      121929.99858407 189.991923109739 217.769626882112   11.7690322800602
#> 9     227970.953326164 278.064071777207 303.416762097117   3.74880770501279
#> 10    499967.743326627 625.113609113741 718.442077571173 -0.124292897141182
#> 11     174449.38764216 301.306722296492 348.337960741393   8.83583778615758
#> 12    274049.035563275 321.752685117403 347.858484525237   3.00414966526061
#> 13    510100.379902254 676.727090534598 786.891438183261 -0.117716907314657
#> 14     182304.34031091 315.757598778718  371.51908542118   8.69898924023118
#> 15    272791.204847968 308.770656920374 330.096820841284   2.19744299765744
#> 16    525021.482503943   742.7936511798 811.198615519069  -0.25633937241815
#> 17    181196.130886184  305.83676698913 361.070231994716   6.90451695450418
#> 18    278150.890281441 336.799023945218 366.344138555252   4.27304776911002
#> 19    519493.179495877 734.222934868857 826.573266176437 0.0336555905400556
#> 20     184394.33592107 331.554233659347 381.895902974605   9.00224543974556
#> 21    276144.078187999 322.839166503998 349.414702637476   4.22080780620578
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6682603194564 3379.89980499507 3445.26507141086    79365.9211151001
#> 2   36.7275685970161  1424.2148618583 1518.74360036399    41064.5658296212
#> 3   24.6713605696336 2186.09153450283  2203.3333582209    63516.0708374632
#> 4   15.6365740072522 7493.72216586741 7639.21136010262    142623.268229668
#> 5   26.7687350296951 2320.06917497968 2413.41983721284    54884.9170254126
#> 6   13.2489055566603 4802.65216623259 4839.34967520059    121684.338763531
#> 7   8.05328481668669 16478.7234014532 17530.7860896136    331165.617640218
#> 8   14.7725424799501 5112.66737713918 5255.65509952695    114456.042743191
#> 9    5.6672538835301 7717.25781108902 7806.19205678741    215084.398077884
#> 10  1.27015868980575 19957.3257981731 20240.4350622027    469123.744339395
#> 11  11.3656010683278 6541.91423290408 6971.08736487276    157391.930359323
#> 12  4.53130321610341 9156.65753741347 9218.31783375146    259418.449907426
#> 13  1.60992011634451 20159.3460432911   20659.78132925    464168.084425904
#> 14    10.61839698449  6820.9669062565 7070.60774371702    168182.106645204
#> 15   4.8576387621798 8979.79196743097 9035.26279118585    256700.576099655
#> 16  1.22921709306283 20459.5760212811 20855.7027394113     501392.62457307
#> 17  9.57648089780088 6734.24434438345 7070.33025765178    168921.816710045
#> 18  5.93304708676031 9291.86484166325 9344.06699206902    260931.627889519
#> 19  1.43435883570957 19643.1696166286 19913.3738494362    485865.024317778
#> 20  10.3810398442609 6949.09408120858 7300.01282962063    174477.148599532
#> 21  5.79453450899486 9240.16571583718 9280.49681329488    257026.968375719
#>    ci95.integral.bt.up
#> 1     83982.4312293365
#> 2     44320.1600458068
#> 3     68303.3361066994
#> 4     154626.490100307
#> 5     59510.8255355921
#> 6     137703.761899585
#> 7     390420.020403348
#> 8     122583.062686671
#> 9     229096.963008052
#> 10    502662.850034249
#> 11    175939.845074641
#> 12     275327.43625165
#> 13    514113.881448731
#> 14     183538.32189335
#> 15    274197.182117238
#> 16    527086.139993048
#> 17    182268.643775361
#> 18    279655.486024424
#> 19    522431.561987168
#> 20    185260.886269554
#> 21    277814.505258975