Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 93.0894682252451  15.5340042572768
#> 2                 FALSE          0.75 32.0495772563672  30.9046551988378
#> 3                 FALSE          0.75 65.3829737210963  21.2099708097334
#> 4                 FALSE          0.75 242.280029606264  14.8763364059805
#> 5                 FALSE          0.75 65.4366251162898  22.5325726246374
#> 6                 FALSE          0.75 205.619539344714  12.4569007053025
#> 7                 FALSE          0.75 514.873585550747  7.30228751327026
#> 8                 FALSE          0.75 200.774428918002  13.1595173301919
#> 9                 FALSE          0.75 293.578466017421  4.71377706572094
#> 10                FALSE          0.75 659.733503211847 0.478993058091008
#> 11                FALSE          0.75 319.410931924262  9.68855444156247
#> 12                FALSE          0.75 331.685446264404  3.70405255861831
#> 13                FALSE          0.75 740.737530016558 0.788607873392764
#> 14                FALSE          0.75 342.623821046306  9.81730696456535
#> 15                FALSE          0.75 319.207591535787  3.38382977895614
#> 16                FALSE          0.75 769.227968371615 0.430692729422154
#> 17                FALSE          0.75 340.118127637363  8.07924761334738
#> 18                FALSE          0.75  350.90762347745  5.00577637582048
#> 19                FALSE          0.75 769.740912654494 0.638361542024679
#> 20                FALSE          0.75 347.808039789154  9.51304677874319
#> 21                FALSE          0.75 336.919743941324  5.25530826731021
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3419.94805253881 2131.27344077127 82084.6223234082 3.17932438497924
#> 2  1449.49143553829 277.422333389824  42357.325746103  2.1386467436452
#> 3  2194.95220897468 633.535011903945 66550.4725782615  1.6496335811459
#> 4  7550.96013962929 6055.09388792434 148698.256883707 2.58979139964009
#> 5  2334.76054346746 1096.10447899394 56372.7760128771 3.15773572464059
#> 6  4821.97161730101 3224.23481973511 130242.194344898 2.25240070918384
#> 7  17060.8893772987  15832.020670818 362377.894785149 7.90618670354183
#> 8  5177.05458308315 3807.07193728419 118370.727269745 10.8493982988775
#> 9  7767.29835738529 6293.78033986627 219320.347077658 6.61139504682504
#> 10 20110.8448515342 18759.4240180852 486571.143833443 16.0353036632381
#> 11 6806.42769834596 5605.76135796975 169149.112094225 14.2461060188919
#> 12 9186.56945557938   7644.436854738 266904.638507535 5.22511357044321
#> 13 20467.5229481773 19458.2189384116 492004.421886316 25.2662359041115
#> 14 6960.66050184381 5666.05209789365 175425.172739733 13.4851769186802
#> 15 9006.35261244683 7395.53183097686 263681.097485525  6.0896183561833
#> 16 20665.0634861881 19344.6401606472 514436.607245381 20.2619636000566
#> 17 6895.79587607707 6057.11475950198 175088.592475529 12.4611105751518
#> 18 9324.63295275177 7662.01114526548 272113.820109863 6.54308764334506
#> 19 19761.6437628137 18556.3994420251 502448.363036101 21.8257515655317
#> 20  7140.5394069093 5819.20701386701 179836.034883019 10.7079854433333
#> 21 9256.83274977645 7439.75196057704 268550.713519288 8.12948228464165
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.523206128902734 13.4372884973209 47.4904383628492 1095.67071673997
#> 2   2.63315620314439  17.914205636501 80.6574531184791 802.528966448015
#> 3   1.42188644385834 4.53167009107663 81.8331879406589 576.351104426407
#> 4  0.330207967597156 93.8175103915491 107.262782977699  4843.1685260869
#> 5   2.35676397112564 81.6597878293663 157.455117624082 3045.45546115412
#> 6  0.604952385813699 11.3468970321411 127.718726454585 3458.72288224086
#> 7  0.358163045570587 150.755757574382 199.051640253041 8120.56798791026
#> 8  0.609989386285446 56.8711547710857 109.692452329857 3092.06563319137
#> 9  0.293171400138908 24.1923321753381 91.6460752189616 4526.67012435691
#> 10 0.300785578585808 68.5220342304489  195.30041255999 7945.37039402138
#> 11 0.740707399346723 43.8465367897691 140.910277570378 1561.19732060619
#> 12 0.637230960815745  11.891996147046 170.193814157159 4624.94960082992
#> 13 0.401373293910512  96.792711998917 279.058533956808 9573.00627163506
#> 14 0.885763111177539 54.0762962190114 252.160644954572 3546.25851974049
#> 15 0.553934758148926 16.3223864237079 158.264502581002 6273.12752607298
#> 16  0.27979972869282 87.9325150354407  202.95376373603 8642.66525426495
#> 17 0.569207295552142 135.010238882566  200.14781090401 5541.18818194796
#> 18  0.33718504978619 7.35935270254445 97.6707140939635 3422.79748220547
#> 19 0.304363719367686 76.0811410166699 197.661137891165 7447.82755729607
#> 20 0.343201217889403 87.2489993642025  106.34821611455 3534.79992230853
#> 21 0.549306206380933 11.9671417691591 142.550156161417 3920.79139895679
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.8594796119543  98.319456838536    14.6733301752318
#> 2             TRUE 28.5315033630708 35.5676511496635    26.5731132446653
#> 3             TRUE 62.6693264801113 68.0966209620813    18.8709676095864
#> 4             TRUE 238.019822753856 246.540236458672    14.3331442992832
#> 5             TRUE 60.2421498492561 70.6311003833236    18.6556958921357
#> 6             TRUE 201.914340178107 209.324738511322     11.461754030639
#> 7             TRUE 501.867908423421 527.879262678073    6.71310930330665
#> 8             TRUE 182.927168716349 218.621689119656    12.1560847897524
#> 9             TRUE 282.702721165393 304.454210869448    4.23151011249244
#> 10            TRUE  633.35542868582 686.111577737874 -0.0157992186826471
#> 11            TRUE 295.976087523185 342.845776325339    8.47009076963711
#> 12            TRUE 323.090134441025 340.280758087783    2.65580762807641
#> 13            TRUE 699.174571954295 782.300488078822   0.128348804909971
#> 14            TRUE 320.440705015077 364.806937077535     8.3602266466783
#> 15            TRUE 309.190169339866 329.225013731709    2.47260710180116
#> 16            TRUE 735.897038249522 802.558898493708 -0.0295778242775353
#> 17            TRUE 319.619600741238 360.616654533488    7.14290161216411
#> 18            TRUE 340.144244304147 361.671002650752     4.4511069689222
#> 19            TRUE 733.837551329194 805.644273979794   0.137683223664836
#> 20            TRUE  330.19340373487 365.422675843437    8.94848077531512
#> 21            TRUE 323.546745583088 350.292742299559    4.35169955781357
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3946783393218 3397.84371296071  3442.0523921169     80282.243994371
#> 2   35.2361971530103 1420.02256726625 1478.96030381034     41037.165596296
#> 3   23.5489740098804 2187.49761167486  2202.4068062745    65602.3750114801
#> 4   15.4195285126779  7396.6303350352 7705.28994422339    140731.244658294
#> 5   26.4094493571391 2200.43019248815 2469.09089444677    51363.0017792786
#> 6    13.452047379966 4803.30597168313 4840.63726291888    124552.595203611
#> 7   7.89146572323388 16812.8961560889 17308.8825985086    349019.560445036
#> 8   14.1629498706315 5083.50153348471 5270.60763268158    113284.279303145
#> 9   5.19604401894944 7727.50197095686 7807.09474381372    211873.974723091
#> 10 0.973785334864663 19998.1261052251 20223.5635978433    473501.009535278
#> 11  10.9070181134878  6734.3001453268 6878.55525136513    166580.942501828
#> 12  4.75229748916021 9167.00712191749 9206.13178924127     259296.59641417
#> 13  1.44886694187556 20308.2989369391 20626.7469594156    476256.826569477
#> 14  11.2743872824524 6871.70499456353 7049.61600912408    169591.577474759
#> 15  4.29505245611112 8979.50228677983 9033.20293811383    253361.802705135
#> 16 0.890963283121842 20520.4144989548 20809.7124734214    500219.422902115
#> 17  9.01559361453066 6673.70403311525 7117.88771903889    165973.337916225
#> 18  5.56044578271876 9312.52681755609 9336.73908794746    266483.318251635
#> 19  1.13903986038452 19636.4902858413 19886.7972397861    490196.686704348
#> 20  10.0776127821713 6997.01480295519 7284.06401086342    174021.289010822
#> 21  6.15891697680684 9237.14680156618 9276.51869798672    262101.011668004
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83887.0006524455 86.8579924306858 99.3209440198044    14.5085202446274
#> 2       43677.48589591 27.8578296388226 36.2413248739118    25.7436690406748
#> 3     67498.5701450429 62.1496919020503 68.6162555401422    18.4230733797711
#> 4      156665.26910912 237.204038462969 247.356020749558    14.2291287894901
#> 5     61382.5502464757 59.2474630959943 71.6257871365854    17.9133152412312
#> 6     135931.793486184 201.204833954714 210.034244734714    11.2711940291077
#> 7     375736.229125261 499.377459611805 530.369711489689    6.60028794395191
#> 8     123457.175236345 179.509608252202 222.039249583802    11.9639381330725
#> 9     226766.719432225 280.620131725643 306.536800309198    4.13916112144868
#> 10    499641.278131609   628.3043080319 691.162698391793  -0.110546675937177
#> 11    171717.281686622 291.488564127234  347.33329972129    8.23676793884289
#> 12      274512.6806009 321.444223666335 341.926668862472    2.45507987541945
#> 13    507752.017203156   691.2157076445 790.259352388617 0.00191621732816005
#> 14    181258.768004706 316.192874285692 369.054767806919    8.08121126665738
#> 15    274000.392265915 307.271939557668 331.143243513907    2.29811765298425
#> 16    528653.791588647 729.514519715504 808.941417027726  -0.117714738815774
#> 17    184203.847034834 315.694350910065 364.541904364661    6.96360131406518
#> 18    277744.321968091 338.083171696494 363.732075258406    4.34489367823955
#> 19    514700.039367853 726.962439586052 812.519385722936  0.0418086520640153
#> 20    185650.780755217  326.82038832022 368.795691258087    8.84037239167996
#> 21    275000.415370572 320.985958663426 352.853529219222    4.17866810280358
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5594882699261 3393.61096708406 3446.28513799355    79937.1077185979
#> 2   36.0656413570008 1414.37959249075 1484.60327858583    40784.3689718649
#> 3   23.9968682396958 2186.07013559617 2203.83428235319    65420.8244135858
#> 4    15.523544022471 7367.07781926186 7734.84245999673    139205.646572577
#> 5   27.1518300080437  2174.7073593219 2494.81372761302    50403.6833090151
#> 6   13.6426073814974 4799.73169911801   4844.211535484    123463.097495706
#> 7   8.00428708258861 16765.4080924529 17356.3706621445    346461.581528845
#> 8   14.3550965273114 5065.58711973182 5288.52204643448     112310.27862869
#> 9    5.2883930099932 7719.88138632163 7814.71532844895    210448.073633918
#> 10  1.06853279211919 19976.5416644425 20245.1480386259    470998.217861161
#> 11   11.140340944282 6720.48848623802 6892.36691045391    166089.165345837
#> 12  4.95302524181717 9163.26114313117 9209.87776802759    257839.737289909
#> 13  1.57529952945737 20277.8092326595 20657.2366636952    473241.329593912
#> 14  11.5534026624733 6854.67096125454 7066.65004243307    168474.506041041
#> 15  4.46954190492804 8974.36073505636  9038.3444898373    251385.767534422
#> 16 0.979100197660081 20492.7157567186 20837.4112156575    497496.983347022
#> 17  9.19489391262958 6631.17580786724  7160.4159442869    164227.863638911
#> 18  5.66665907340141 9310.20862145478 9339.05728404876     265405.13704474
#> 19  1.23491443198534  19612.524726421 19910.7627992064      487850.6210238
#> 20  10.1857211658064 6969.53136815547 7311.54744566314    172907.827035295
#> 21  6.33194843181683  9233.3771519089   9280.288347644    260865.962377333
#>    ci95.integral.bt.up
#> 1     84232.1369282186
#> 2     43930.2825203411
#> 3     67680.1207429373
#> 4     158190.867194837
#> 5     62341.8687167392
#> 6      137021.29119409
#> 7     378294.208041453
#> 8       124431.1759108
#> 9     228192.620521397
#> 10    502144.069805725
#> 11    172209.058842613
#> 12    275969.539725162
#> 13    510767.514178721
#> 14    182375.839438424
#> 15    275976.427436628
#> 16     531376.23114374
#> 17    185949.321312148
#> 18    278822.503174985
#> 19    517046.105048401
#> 20    186764.242730744
#> 21    276235.464661244