Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4753647328568   15.617305982068
#> 2                 FALSE          0.75 32.4689159590392  31.4265471779149
#> 3                 FALSE          0.75 64.9811189353131   20.030490438859
#> 4                 FALSE          0.75 242.884240525427  14.9770121072743
#> 5                 FALSE          0.75 65.8777035916386  23.8990369154238
#> 6                 FALSE          0.75 205.749078918619  12.5217335311376
#> 7                 FALSE          0.75 506.809077661488  6.95581093241921
#> 8                 FALSE          0.75 204.576627476901  13.5599750001487
#> 9                 FALSE          0.75 289.728983248297  4.59299264082676
#> 10                FALSE          0.75 675.882986016521 0.677225689130804
#> 11                FALSE          0.75 326.708116615925  9.94981827457408
#> 12                FALSE          0.75 334.189051253656  3.89116011995442
#> 13                FALSE          0.75 725.931766420521 0.694041913601797
#> 14                FALSE          0.75 344.863928080854  9.74872087459159
#> 15                FALSE          0.75 319.302013230781   3.4579925305502
#> 16                FALSE          0.75 771.182347763141 0.326393909688059
#> 17                FALSE          0.75 335.793839281179  8.16659488335018
#> 18                FALSE          0.75 350.516918924979  5.11818363410899
#> 19                FALSE          0.75 777.611147103259 0.601116091334244
#> 20                FALSE          0.75 352.734322568916  9.55984213137715
#> 21                FALSE          0.75 340.645264107991   5.2959850404696
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3408.13367577157 2113.08510389535 81468.4654061565 2.12585042366049
#> 2  1467.67868495656 294.482953071241 42772.4493511063  2.4201259553951
#> 3  2197.19467074766 698.992979771117 65789.8373636781  1.2504576446757
#> 4  7559.57697814512 6051.25936692604 148679.486334688 2.85553802360732
#> 5  2362.65633287653 1092.41393567189 56963.5879181332 1.86386302245915
#> 6  4823.80499298679  3211.6233373555 131036.607657972  2.9147165971847
#> 7   17045.430234717  15865.379007714 362657.223219073 14.4878616521801
#> 8  5180.94118653656 3748.48410553968 118888.663915861 6.76203069416189
#> 9  7762.22967375276 6293.93706986227 220953.503629732 8.07044160616124
#> 10   20111.81708366 18732.4930501664 483657.528147639 21.1899688001364
#> 11 6765.43524475487 5541.25384634246 167085.150268163 13.6497813447171
#> 12 9192.04305136264  7631.7722317537 265621.963630279  6.0687134789977
#> 13 20435.9126108302 19430.2029685858  487730.26693691 24.9426234234656
#> 14 6965.72953949462 5658.82621565205 176892.767085537 15.4403318224037
#> 15 9003.60259260814 7376.46356258487 266719.917292751 5.16533358755637
#> 16 20669.2303272413 19441.1387609613 511410.685067416 23.3986955264785
#> 17 6871.42513207051 5976.11496979649 174853.955722377 14.9001362134203
#> 18 9320.34032751475 7609.22522076983 270158.751477895 5.56029217238584
#> 19 19790.2718104496 18675.4244563554 504643.906176041 22.0180661961878
#> 20 7142.34248869355 5816.33562750155 179984.035037796 12.6386832366201
#> 21 9256.48703720925 7469.93809362369 268037.954607991 7.80323451055535
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.419273485938499 31.9019773367304 49.7612676689721 1774.38108765917
#> 2   1.86284393683765 27.1268403389395 64.2705339300715  747.05537342633
#> 3   1.79849245620071 6.18677073021239 124.711091478905 1004.52418454647
#> 4  0.347415555562914   48.68278510635 71.1623911515817 3086.64206436302
#> 5   1.93667586474833 49.2294725438782 144.303593397524 1661.42449876442
#> 6  0.376838154168833 8.76552483743163 74.0962518812483 2435.50189183766
#> 7  0.596427285980335 140.341531687121 189.004566417121 6640.46324736353
#> 8  0.762296238230581 60.9339293233978 163.678550384347 2786.71553499191
#> 9  0.330945263135766 21.6746723606297  62.732211798232 4378.64476622237
#> 10 0.360687527059105  83.662301830903 254.331245275299 7329.24081739947
#> 11 0.776206039003489 96.7707526875153 221.767719133187 3892.02955723438
#> 12 0.430828168573635 10.4304440258033 110.233260038154  5139.1851052654
#> 13  0.36068592552617  114.99450740911 272.715540263069 10553.3289980055
#> 14 0.749207221087069 62.6664878517687 161.406416738488 3184.18805169762
#> 15 0.380339144953958 9.76874395009983 116.202781835139 1662.17085223228
#> 16 0.263513526631299 110.435454018704 192.961834395987 10581.3907898933
#> 17 0.444672798458084 91.8283554774098 128.904332911511 3508.63838567118
#> 18 0.474376923256338 10.9828353622223  146.25802407158 5248.35770320744
#> 19 0.241740488647206 76.2339582071444 161.859788407485  7429.1971338624
#> 20 0.389944971818146 79.5518593416099 138.065345823274 2654.60306632346
#> 21 0.513274672611748 9.37448600739286 106.058851806268 4578.61454656747
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.9783407859353 95.9723886797784   14.9276010976992
#> 2             TRUE 28.4878087624142 36.4500231556641    28.362168901817
#> 3             TRUE 62.9241161098216 67.0381217608046   17.0719703484088
#> 4             TRUE 238.186880476593 247.581600574261   14.4055135183733
#> 5             TRUE 62.8116489196933 68.9437582635839   20.7132051179128
#> 6             TRUE  200.95437011625 210.543787720987   11.9018347675299
#> 7             TRUE 482.976545243652 530.641610079324   5.97468804698156
#> 8             TRUE 193.453086985004 215.700167968797   12.3059976882594
#> 9             TRUE 276.453106806162 303.004859690433   4.04858768296843
#> 10            TRUE 641.025487340297 710.740484692745 0.0838947071185766
#> 11            TRUE 304.254226303866 349.162006927985   8.67295934041334
#> 12            TRUE 324.206017580705 344.172084926607   3.18244778265079
#> 13            TRUE  684.90115088892 766.962381952122  0.100713566111246
#> 14            TRUE    319.464582233 370.263273928708   8.51627499590336
#> 15            TRUE 310.805039479251 327.798986982311   2.83233463710094
#> 16            TRUE 732.691493622084 809.673201904198 -0.107085841620428
#> 17            TRUE 311.283115210103 360.304563352256   7.43510812988663
#> 18            TRUE 341.370238301405 359.663599548554   4.33783359535232
#> 19            TRUE  741.39142821053 813.830865995988  0.203452987509591
#> 20            TRUE 331.943688644675 373.524956493156    8.9183826527363
#> 21            TRUE 327.808943338127 353.481584877854   4.45164820402327
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3070108664368 3355.65492305265 3460.61242849049    78549.6085169572
#> 2   34.4909254540128 1423.05503259901 1512.30233731412      41543.54326182
#> 3   22.9890105293092 2187.01743289646 2207.37190859886    64137.3950800991
#> 4   15.5485106961753 7479.49379664517 7639.66015964506    143601.960138811
#> 5   27.0848687129348 2281.67385054185 2443.63881521121    54230.5446176658
#> 6   13.1416322947453 4809.38570462921 4838.22428134436    127030.207045899
#> 7   7.93693381785686 16814.5684150916 17276.2920543423     351733.66117716
#> 8    14.813952312038 5080.70487279957 5281.17750027355    114304.516860799
#> 9    5.1373975986851 7726.57483771953   7797.884509786    213750.632989296
#> 10  1.27055667114303 19974.1925971482 20249.4415701718    471600.927003017
#> 11  11.2266772087348  6606.2473565839 6924.62313292583    160682.761646513
#> 12  4.59987245725805 9174.88497094019 9209.20113178509    257168.004132118
#> 13  1.28737026109235 20246.7466461422 20625.0785755182    470370.040735191
#> 14  10.9811667532798 6862.64316697846 7068.81591201078    171654.777740494
#> 15  4.08365042399946 8987.53300881022 9019.67217640605    263985.646240829
#> 16 0.759873660996546 20487.5640053805 20850.8966491021    494004.297218042
#> 17  8.89808163681372 6720.36748731017 7022.48277683085    169082.245577948
#> 18  5.89853367286567  9302.2735633439 9338.40709168561    261525.203056119
#> 19 0.998779195158898 19664.8669491988 19915.6766717003    492422.876890838
#> 20   10.201301610018 7011.47968007661  7273.2052973105    175617.212993694
#> 21  6.14032187691592 9241.06600772709 9271.90806669141    260506.133678887
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84387.3222953558 88.3086979024823 96.6420315632314    14.7955299496286
#> 2     44001.3554403926 27.7254690864648 37.2123628316136    27.7753730617131
#> 3      67442.279647257 62.5302219517487 67.4320159188775    16.5054452247056
#> 4     153757.012530565 237.287385999157 248.481095051698     14.296077618371
#> 5     59696.6312186007 62.2245320676187 69.5308751156586    20.1031522205171
#> 6     135043.008270045 200.036234388137 211.461923449101    11.7831307489667
#> 7     373580.785260986 478.412868823215 535.205286499761    5.78681345189775
#> 8     123472.810970923 191.323047316343 217.830207637458    12.0658743732168
#> 9     228156.374270168 273.910917700221 305.547048796373    3.94433992508066
#> 10    495714.129292261 634.350647168254 717.415324864788 -0.0297218639050415
#> 11    173487.538889814  299.95454518028 353.461688051571    8.42845443812724
#> 12    274075.923128441  322.29437283482 346.083729672491    3.04673690955009
#> 13    505090.493138629 677.044224510529 774.819308330514 -0.0129025004294971
#> 14     182130.75643058 314.600877708943 375.126978452765    8.28027472126093
#> 15    269454.188344673  309.17795939917 329.426067062391    2.71252780644044
#> 16    528817.072916791 725.320904531243 817.043790995039  -0.190092602509287
#> 17    180625.665866806 306.589572302875 364.998106259483    7.29503619837233
#> 18    278792.299899672 339.618746267103 361.415091582855    4.18840486452657
#> 19    516864.935461245 734.455737358731 820.766556847787   0.127304733585721
#> 20    184350.857081898  327.96250342514 377.506141712691    8.79554998661359
#> 21    275569.775537094 325.350924467302 355.939603748679    4.28996668215057
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4390820145075 3345.60580019158 3470.66155135156    77990.6784743445
#> 2   35.0777212941167 1414.51007789224 1520.84729202088    41308.2208191907
#> 3   23.5555356530124 2185.06860011644 2209.32074137887     63820.969961967
#> 4   15.6579465961776 7464.15871933667 7654.99523695356    142629.667888536
#> 5   27.6949216103305 2266.16656669053 2459.14609906254     53707.195900555
#> 6   13.2603363133085 4806.62456430542 4840.98542166815     126263.02394997
#> 7   8.12480841294067 16770.3608326102 17320.4996368237     349641.91525424
#> 8   15.0540756270807  5061.5106850627 5300.37168801042    113426.701467277
#> 9   5.24164535657286 7719.74731592593  7804.7120315796    212371.359887936
#> 10  1.38417324216665 19947.8389720714 20275.7951952486    469292.216145536
#> 11  11.4711821110209 6575.76456948734  6955.1059200224    159456.772335984
#> 12  4.73558333035874 9171.59938107206 9212.48672165321    255549.160823959
#> 13  1.40098632763309 20210.5233763083  20661.301845352    467045.742100819
#> 14  11.2171670279222 6842.90322330516 7088.55585568409     170651.75850421
#> 15  4.20345725465995 8984.45585446594 9022.74933075033    263462.062422376
#> 16 0.842880421885405 20452.7768373647  20885.683817118    490671.159119226
#> 17  9.03815356832802 6691.44155533478 7051.40870880623    167977.024486461
#> 18  6.04796240369141  9298.8139702048 9341.86668482471    259871.970379609
#> 19  1.07492744908277 19640.8532523636 19939.6903685356    490082.679793671
#> 20  10.3241342761407   6986.420844384 7298.26413300311    174781.013027802
#> 21  6.30200339878862 9238.11304463476 9274.86102978374    259063.870096719
#>    ci95.integral.bt.up
#> 1     84946.2523379685
#> 2     44236.6778830219
#> 3     67758.7047653891
#> 4     154729.304780839
#> 5     60219.9799357115
#> 6     135810.191365974
#> 7     375672.531183905
#> 8     124350.626364445
#> 9     229535.647371528
#> 10    498022.840149742
#> 11    174713.528200342
#> 12    275694.766436599
#> 13       508414.791773
#> 14    183133.775666864
#> 15    269977.772163126
#> 16    532150.211015607
#> 17    181730.886958292
#> 18    280445.532576182
#> 19    519205.132558412
#> 20     185187.05704779
#> 21    277012.039119263