Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8016790483609  15.3743445897743
#> 2                 FALSE          0.75 33.6067847528454  30.5610366946447
#> 3                 FALSE          0.75 64.6934323359896  19.9285764061143
#> 4                 FALSE          0.75 242.902074591418  15.0663674424839
#> 5                 FALSE          0.75 64.8858352427538  22.1483779674199
#> 6                 FALSE          0.75 205.641805409754  12.3519003360088
#> 7                 FALSE          0.75 504.593763578409  6.90456787680785
#> 8                 FALSE          0.75 201.355263742959  13.1920623450886
#> 9                 FALSE          0.75 287.119138779027  4.60350918996493
#> 10                FALSE          0.75 672.990212616612 0.486917778941171
#> 11                FALSE          0.75 323.132075077608   9.8945084185801
#> 12                FALSE          0.75 334.797215406846  3.74134915955064
#> 13                FALSE          0.75 719.877477644529 0.655339278590234
#> 14                FALSE          0.75  347.21130955233  9.83379821850092
#> 15                FALSE          0.75 321.855440424114  3.54515722682421
#> 16                FALSE          0.75 780.131233914478 0.561408836658005
#> 17                FALSE          0.75 332.435352276811  7.84756013059551
#> 18                FALSE          0.75 350.077698580343  4.94372922800779
#> 19                FALSE          0.75 770.022375129222 0.606182384595955
#> 20                FALSE          0.75  346.62740936778  9.34834225896178
#> 21                FALSE          0.75 334.607607163078  5.19666057407527
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3423.72883377517 2141.62115891001 82217.3776803047 2.77038482684004
#> 2  1467.35763110072 311.057608369173 42530.5882082273 2.48116356671139
#> 3  2195.70909827986 704.062032665986 65658.7822413388 1.57938929781223
#> 4  7564.70264115928 6029.34702078301 149308.509490548 2.64642155093123
#> 5  2375.08175794264  1123.0469719683 57667.3532277868 2.12684884414131
#> 6  4820.87426417191 3243.03347644139 130485.579569219 2.28682846363692
#> 7  16981.3873086841 15797.9467205636 359983.410120065 14.0627151793229
#> 8  5166.56559231569 3783.65672804274 117867.798956235 8.09905570907588
#> 9  7768.00951384308 6301.11494632837 221130.057595885 5.68404149653996
#> 10 20065.2546590943 18763.7090176489 484807.033033412 23.4441015338291
#> 11 6762.54528751209 5537.65768382142 167268.218003078 15.2036176645192
#> 12 9187.49835144166 7689.73711026634 265410.197313014 5.08026446356654
#> 13 20461.8575807647 19438.7285318338 491226.793293548 18.3642001481398
#> 14 6954.03715636166 5641.59049789658 176301.215219107 12.0800195771271
#> 15 9000.48741597947 7368.25901057768 266230.461816547 4.79584773593585
#> 16 20690.4187691255 19339.8625069955 516005.244953945  26.826338026321
#> 17 6913.23439832382 6092.81904577816 176338.805595256 12.5330182841362
#> 18 9320.86516701764 7649.99709158175 270928.699218871  8.2161343373212
#> 19 19787.7832819375 18606.1880986719 505541.854215466 27.1365662886218
#> 20 7125.93783949742 5832.86881020564  180294.92615794 14.2457058541962
#> 21 9256.98904176949 7430.53466413766 268576.619593319 8.17040893344327
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.675765067386024  18.462287146765 56.4411702362051 1109.74206533164
#> 2   2.03505815100165 23.2234890249588 74.8766430855499 921.032505906492
#> 3    1.9422327725946 5.25049000554233 119.560831924677 1487.57022830089
#> 4  0.304856334175372 41.2866034045001 60.9133979514345 2593.66686594617
#> 5   1.94464624490334 18.8490735454777 151.289328834728  786.72253880256
#> 6  0.429575394001039 9.22047303255429  86.506762176651 4610.34892108357
#> 7   0.50560323440667  188.17445622576 230.400017261281  9732.1736871506
#> 8  0.699545654243266  52.041397419578 103.357759118594  3114.1388067691
#> 9  0.428563898798063 13.6108004920455 85.0275226794977 2497.08913362886
#> 10   0.3727477812823  96.597081592773 269.377401945361 12356.8297437121
#> 11 0.767441300158076 96.8714284713749  200.78343943119 4057.42952043284
#> 12 0.289211255110619 12.4542572035387 78.0752486814471 4238.45838395172
#> 13 0.219905430628182 90.6187280256526 111.996401228549 8512.27595286636
#> 14 0.553388325285872 76.4509662589777 205.962119952979 4105.29056674821
#> 15 0.531654485640901 17.1778077497551 151.764496825638 3198.40340280846
#> 16 0.438333863781766 86.1093640580286 249.373031680231 8401.31559340081
#> 17  0.43524533176257 65.2223468183261 149.298940416067 2225.11525953347
#> 18 0.363371833559079 7.72292996621707 95.8504532802987 3044.94464814887
#> 19 0.202926387605643 60.4014111422471 162.002980037671 8223.99842202151
#> 20 0.320846196196688 81.1898108231775 111.887701474705  2438.9479116667
#> 21 0.842755055782401 10.6249775578582 231.582007788832 3465.96344543747
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE  87.244396008209 96.3589620885128   14.2627110539243
#> 2             TRUE 29.5252706856052 37.6882988200857    27.213366036247
#> 3             TRUE 62.0953369410885 67.2915277308908   16.7336034951961
#> 4             TRUE 238.548711140136   247.2554380427   14.5648787727654
#> 5             TRUE 61.3871688941413 68.3845015913663   18.9494348945539
#> 6             TRUE 201.879972587071 209.403638232436   11.6452488128771
#> 7             TRUE 481.460597108423 527.726930048395   6.07285055620887
#> 8             TRUE 188.032317101529 214.678210384389   12.0413097438584
#> 9             TRUE 277.768890517219 296.469387040836   3.89852157644212
#> 10            TRUE 634.424665593463 711.555759639761 -0.126252321268212
#> 11            TRUE 298.122124019474 348.142026135742   8.63206747982006
#> 12            TRUE  326.44018036428 343.154250449413   3.26559664489367
#> 13            TRUE 689.668368400839 750.086586888219  0.293594845206875
#> 14            TRUE 327.339677347956 367.082941756704   8.92347442340566
#> 15            TRUE 313.966270898499 329.744609949728   2.67058559794493
#> 16            TRUE  736.00190786118 824.260559967776    -0.159650369263
#> 17            TRUE 311.818537199407 353.052167354215   7.13158155984608
#> 18            TRUE  336.56215759545 363.593239565237   4.34598256180311
#> 19            TRUE 725.382723584439 814.662026674005  0.272368476984673
#> 20            TRUE 323.193223237627 370.061595497933   8.82055026621823
#> 21            TRUE 321.167284467564 348.047929858593   3.81032850731322
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4859781256243 3393.35837141874  3454.0992961316    80391.8519828341
#> 2   33.9087073530424 1429.15499165467 1505.56027054678    41015.4897360112
#> 3   23.1235493170324 2187.07204222075 2204.34615433898    63211.7292157838
#> 4   15.5678561122024 7496.78617855888 7632.61910375968    145041.927496067
#> 5   25.3473210402859 2344.07503196033 2406.08848392495    56373.1946514566
#> 6   13.0585518591405 4805.70658603336 4836.04194231046    122901.555594036
#> 7   7.73628519740682 16671.8403281928 17290.9342891755    343973.984404702
#> 8   14.3428149463188 5080.95749356049  5252.1736910709      112745.0406191
#> 9   5.30849680348774 7745.61974703367  7790.3992806525    217022.345971065
#> 10  1.10008787915056 19906.3524598742 20224.1568583144    464480.048105006
#> 11  11.1569493573401 6603.19178767668  6921.8987873475    160593.746441966
#> 12  4.21710167420761 9167.01109834183 9207.98560454148    258437.933271413
#> 13  1.01708371197359 20312.7897731625 20610.9253883669    477224.099351082
#> 14  10.7441220135962 6828.27531686564 7079.79899585767    169548.012236806
#> 15  4.41972885570349 8972.22992223112 9028.74490972781    260969.088218927
#> 16  1.28246804257901   20548.76886525 20832.0686730009      502185.0808028
#> 17  8.56353870134493 6805.94363780768 7020.52515883997    172678.490993323
#> 18  5.54147589421248 9308.16094722321 9333.56938681207    265919.765272666
#> 19 0.939996292207238 19688.4229606085 19887.1436032665    492013.376811241
#> 20  9.87613425170533 6992.38060069329 7259.49507830155    176282.856843248
#> 21  6.58299264083732 9239.51095368681 9274.46712985217    262875.109725574
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84042.9033777752 86.3717247877544 97.2316333089674   14.0498450576977
#> 2     44045.6866804435 28.7437041620911 38.4698653435997   26.5723227186814
#> 3     68105.8352668937 61.5978293122777 67.7890353597016   16.1218001718288
#> 4      153575.09148503 237.715088351593 248.089060831243   14.4688490275002
#> 5     58961.5118041171 60.7172115082368 69.0544589772708   18.3368713274094
#> 6     138069.603544401 201.159621621025 210.123989198482   11.5099325637667
#> 7     375992.835835427 477.030841826936 532.156685329882   5.91358553737077
#> 8      122990.55729337  185.48111455317 217.229412932748   11.8209528627718
#> 9     225237.769220704 275.978417445809 298.259860112246   3.76352394832073
#> 10    505134.017961819 627.039773610307 718.940651622917 -0.243667872372137
#> 11     173942.68956419  293.33298445515 352.931165700065   8.39032347027027
#> 12    272382.461354614 324.839897058256 344.754533755437   3.17449509953382
#> 13    505229.487236013 683.883645354175 755.871309934883  0.224324634558998
#> 14    183054.418201408 323.534471181161 370.888147923499   8.74915710094061
#> 15    271491.835414167 312.455578861679 331.255301986548   2.50311443496804
#> 16    529825.409105089 727.551611382889 832.710856446067 -0.297725536354257
#> 17    179999.120197189 307.870636439904 357.000068113718   6.99447928034087
#> 18    275937.633165076 333.974075279194 366.181321881493     4.231520434232
#> 19    519070.331619691 716.834705203523 823.210045054921  0.208446664888896
#> 20    184306.995472632 318.705825893556 374.548992842005   8.71948371441627
#> 21    274278.129461064  318.59360565353 350.621608672627   3.54486066474176
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6988441218509 3387.54275096751 3459.91491658283    80042.2832322546
#> 2   34.5497506706079  1421.8395926118 1512.87566958964    40725.3644966506
#> 3   23.7353526403997   2185.418137869 2206.00005869073     62743.144593869
#> 4   15.6638858574677 7483.78089848646  7645.6243838321    144224.922433294
#> 5   25.9598846074305  2338.1375737935 2412.02594209178    56125.3770517338
#> 6   13.1938681082508 4802.80213702811 4838.94639131572    121449.295683895
#> 7   7.89555021624492 16612.5653744817 17350.2092428866    340908.349693249
#> 8   14.5631718274054 5064.56445337332 5268.56673125807    111764.086894967
#> 9   5.44349443160913 7741.33234487867 7794.68668280749    216235.762893972
#> 10  1.21750343025448 19875.9243791725 20254.5849390162    460587.646735737
#> 11  11.3986933668899 6572.67728770819 6952.41328731598    159315.656143029
#> 12  4.30820321956745 9163.08800732272 9211.90869556059    257102.818880468
#> 13  1.08635392262147 20284.2448738344  20639.470287695     474542.73242593
#> 14  10.9184393360612 6804.19326249406 7103.88105022925    168254.845708281
#> 15  4.58720001868037 8966.81891278995 9034.15591916899    259961.591147042
#> 16  1.42054320967027 20521.6444155718 20859.1931226792    499538.666390879
#> 17  8.70064098085014  6785.3985985599 7041.07019808774     171977.57968657
#> 18  5.65593802178359 9305.72822428385 9336.00210975142    264960.607708499
#> 19  1.00391810430301 19669.3965160987 19906.1700477763    489422.817308304
#> 20  9.97720080350729 6966.80581028399 7285.06986871085    175514.588251073
#> 21  6.84846048340877 9236.16408575609 9277.81399778289    261783.331240262
#>    ci95.integral.bt.up
#> 1     84392.4721283547
#> 2     44335.8119198041
#> 3     68574.4198888085
#> 4     154392.096547803
#> 5     59209.3294038399
#> 6     139521.863454543
#> 7      379058.47054688
#> 8     123971.511017502
#> 9     226024.352297797
#> 10    509026.419331088
#> 11    175220.779863126
#> 12    273717.575745559
#> 13    507910.854161166
#> 14    184347.584729934
#> 15    272499.332486052
#> 16     532471.82351701
#> 17    180700.031503942
#> 18    276896.790729243
#> 19    521660.891122628
#> 20    185075.264064807
#> 21    275369.907946377