Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1480000228262  15.2997011773877
#> 2                 FALSE          0.75 31.4987470582848  31.8725068903176
#> 3                 FALSE          0.75 64.9236032898515  20.3159441930191
#> 4                 FALSE          0.75 243.122158176647  15.1491844865132
#> 5                 FALSE          0.75 66.1605116431693  22.0793467132709
#> 6                 FALSE          0.75 206.538244586154  12.3413298999378
#> 7                 FALSE          0.75 507.754882190561  7.07835472514792
#> 8                 FALSE          0.75 202.201020675643  13.6549286330007
#> 9                 FALSE          0.75  287.16708181813  4.80051736172521
#> 10                FALSE          0.75 688.805152275539 0.676575969619161
#> 11                FALSE          0.75 325.779611100594  9.67078442162658
#> 12                FALSE          0.75 335.508833717271  3.81656647752937
#> 13                FALSE          0.75 726.362909882305 0.795847561127147
#> 14                FALSE          0.75   348.1064799454  9.89098806244534
#> 15                FALSE          0.75  321.86207382622  3.42306617957514
#> 16                FALSE          0.75 774.335267726506 0.554531609800118
#> 17                FALSE          0.75 333.115184195391  8.00831476629021
#> 18                FALSE          0.75 350.577476026709  5.03242950965036
#> 19                FALSE          0.75  777.95498304294 0.808006180524069
#> 20                FALSE          0.75 349.858223858429  9.47405194315902
#> 21                FALSE          0.75 337.809296747214  5.23120592725325
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.28687370061 2138.36175204377 80990.5009873533 1.91039275857657
#> 2   1459.1690169138 269.427408710672 42538.9291942917 2.02863863979513
#> 3  2195.75133218731 671.952614878881 65833.9744009881 1.67903935321727
#> 4  7519.28717981337 5968.63870947862 147281.698006193 3.19239003011566
#> 5  2368.57237652218 1147.96751767578 57539.0845897394  2.4355714272711
#> 6  4816.07933455179 3242.73099575267 130336.231252899 2.63859728693238
#> 7  17034.7218567622 15823.6433773843 361454.863442683 13.6042466731104
#> 8  5183.36455182257 3735.66506227888 119420.111440752 10.8159348579319
#> 9  7767.48725838884 6240.80830598634 221432.992368345 6.46792290490963
#> 10 20074.0139846136 18730.0687047844 483936.205849091 30.4306504834682
#> 11 6761.13868056954 5612.90511932561 166427.462096084 12.5257386841849
#> 12 9187.50258750897 7656.76159513884 268318.401371761 6.26968623001247
#> 13 20398.0494519613 19325.2001969338 484589.028242089 26.1437950899896
#> 14 6972.89059233291 5659.22013234269 177192.530967717 12.9981798015647
#> 15 8997.96594503096 7407.25698658226 265313.598730917 6.69580902017483
#> 16 20707.3998837313 19357.9929037711 514953.828818172 27.4206415864526
#> 17 6886.94629469442 6002.22474207162 175987.987225171 14.7148649750912
#> 18 9325.55816573421 7646.34382171609 271537.964811395 7.80475603500122
#> 19 19806.3501742929 18514.4706605464 502856.244272865 24.8523465452432
#> 20 7169.46409756381 5847.73909801473 180695.089593475 10.6967762080954
#> 21 9252.53132356335 7452.75288454563 266972.151667424  8.0739785030903
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.542503008915575 44.8724944115391 62.6962680640127  2836.6160788853
#> 2   2.69275088602291 19.9588577556107 78.1574615818749 834.767497364944
#> 3   2.17450322491717 5.50661633691646 134.392311430595 1314.53526850745
#> 4  0.350817424939507 151.475842852754   175.6332874137 6624.60395789875
#> 5   2.97852806173051 36.6985935353922 133.189259660111 1394.62652706391
#> 6   0.38810761744532 9.27131375024646 83.5292437147722 2502.83825331855
#> 7  0.413342289263279 133.581148527538  157.47822131274 7682.52819865129
#> 8  0.850979529139698 35.8905513364324 152.179566715424 2323.65980909835
#> 9  0.426303752636454 20.9908239578398 93.4380718685558 3471.57106424784
#> 10 0.455025019253515 68.6594563312887 291.207667925206 10569.1841745643
#> 11 0.526229916815324 119.395387916455 185.397077491699 4410.25515873751
#> 12  0.46630148847304 11.5379918877887 135.631340210965 2778.03653389578
#> 13 0.362602188958081 127.806025257318 212.064237335754 13285.8932002647
#> 14 0.681098364987012 58.4439692335693  187.97100930689  2932.2411980426
#> 15 0.481304834336163 22.9764780172073 102.177017592202 2710.78931819836
#> 16 0.394595802631493 106.628557657817 268.644273162036 8434.63416071942
#> 17  0.64998535351673 119.342454702174 175.782934526107 4590.80637496765
#> 18 0.462312095065891 7.31845130859986 120.259647695395 2486.99187417155
#> 19 0.383362464116823 99.2582911992406 250.759237819795 7908.49483568946
#> 20 0.376698230436487 71.7681292105379 128.002910543323 3829.45905003792
#> 21 0.540056772855767 10.5720205922385 138.494058849246 3929.62190862241
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 89.0054039349677 95.2905961106846    14.4072837277216
#> 2             TRUE 28.1616364958218 34.8358576207477      27.44293168281
#> 3             TRUE 62.1615835538091  67.685623025894    16.7388863880303
#> 4             TRUE 237.870676577106 248.373639776187    14.5720898224877
#> 5             TRUE 62.1539966453083 70.1670266410302    17.1796680517242
#> 6             TRUE 202.197752049151 210.878737123158    11.7028928692402
#> 7             TRUE 485.375896413294 530.133867967827    6.39840665930983
#> 8             TRUE 184.408807834345 219.993233516941    12.2550673075659
#> 9             TRUE 276.527348639554 297.806814996706    4.09924768863824
#> 10            TRUE 638.746732230234 738.863572320844 -0.0719401870528712
#> 11            TRUE  305.17477096511 346.384451236078    8.80513620846537
#> 12            TRUE 325.195199868901 345.822467565642    3.04950052899122
#> 13            TRUE 683.356366959272 769.369452805337   0.199366960291104
#> 14            TRUE 326.724474171826 369.488485718974     8.7705812520417
#> 15            TRUE 310.847467988032 332.876679664408    2.63131972709215
#> 16            TRUE 729.228312316791  819.44222313622 -0.0945784855286882
#> 17            TRUE 308.909231311366 357.321137079416    6.93908885975519
#> 18            TRUE 337.738652349132 363.416299704286    4.27192611326697
#> 19            TRUE 737.072872976015 818.837093109865   0.177374927051894
#> 20            TRUE 332.262026996112 367.454420720746      8.854383354091
#> 21            TRUE 324.527602109631 351.090991384798    4.34281253590551
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.1921186270538 3333.47162039363 3481.10212700759    76324.2675375869
#> 2   36.3020820978253 1426.33669590582 1492.00133792178    41165.7366611264
#> 3   23.8930019980078 2186.69294831308 2204.80971606154    63671.5638842933
#> 4   15.7262791505387 7270.10941832059 7768.46494130615    136384.224495449
#> 5   26.9790253748176 2308.20319015646  2428.9415628879    55244.9239527192
#> 6   12.9797669306353 4800.82802343264 4831.33064567095     126219.06232619
#> 7   7.75830279098602 16814.9808674344   17254.46284609    348817.104555902
#> 8   15.0547899584355 5124.32459487414 5242.40450877101    115597.691054786
#> 9   5.50178703481218 7732.95735297819 7802.01716379949    215722.257967657
#> 10  1.42509212629119 19961.0691789487 20186.9587902786    466549.897881932
#> 11  10.5364326347878 6564.73326744697  6957.5440936921    159172.592359961
#> 12  4.58363242606752 9168.52259085356 9206.48258416439    263748.531273502
#> 13  1.39232816196319  20187.808540413 20608.2903635095    462733.733927654
#> 14   11.011394872849 6876.75026294369 7069.03092172213    172368.994196937
#> 15  4.21481263205812 8960.16963869265 9035.76225136926    260854.350302481
#> 16  1.20364170512892 20531.9959063842 20882.8038610784    501078.855623789
#> 17  9.07754067282523 6690.62795670934  7083.2646326795    168436.110738349
#> 18  5.79293290603375 9313.51931333156 9337.59701813685    267446.863178383
#> 19  1.43863743399624 19643.0702852702 19969.6300633157    489846.770268156
#> 20   10.093720532227 7051.40552501247 7287.52267011514    174395.629456163
#> 21  6.11959931860098 9235.14034968912 9269.92229743759     260507.92362774
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     85656.7344371196 88.4036302160161 95.8923698296363   14.2363952799132
#> 2     43912.1217274571 27.5226153242863 35.4748787922832   26.5947151537127
#> 3     67996.3849176828 61.6326861575457 68.2145204221574   16.0539178721814
#> 4     158179.171516936  236.86507371762 249.379242635673   14.4615823336318
#> 5     59833.2452267595  61.386791645718 70.9342316406207   16.2414317122791
#> 6     134453.400179608 201.366593903767 211.709895268542    11.580638969745
#> 7     374092.622329464 481.090558711264 534.419205669857    6.2682038381919
#> 8     123242.531826719 181.001788354096  223.40025299719   11.9870087558869
#> 9     227143.726769033 274.489952924507 299.844210711753   3.96496200655776
#> 10    501322.513816249 629.161077327941 748.449227223137 -0.215273068117728
#> 11    173682.331832207 301.229163279591 350.330058921596   8.63937378466854
#> 12    272888.271470019 323.220248706447 347.797418728096   2.90261556012222
#> 13    506444.322556525 675.121071505925 777.604748258684 0.0851472707693083
#> 14    182016.067738497 322.630047534333 373.582912356466   8.55603526707079
#> 15    269772.847159354 308.738288146677 334.985859505763   2.47970870427626
#> 16    528828.802012556 720.590810217059 828.079725235953 -0.218876163357609
#> 17    183539.863711993 304.274048844212 361.956319546569   6.73434347339742
#> 18    275629.066444408 335.280154198106 365.874797855311   4.12629780332121
#> 19    515865.718277574 729.244383814264 826.665582271617  0.056615750855095
#> 20    186994.549730787 328.892542490562 370.823905226296    8.7357234115035
#> 21    273436.379707108 321.984298881157 353.634294613271   4.17269465245594
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.3630070748623   3319.336784654 3495.23696274723    75430.7334727381
#> 2   37.1502986269225  1420.0496557128  1498.2883781148    40902.7848994565
#> 3   24.5779705138567 2184.95836416695 2206.54430020766    63257.4852747135
#> 4   15.8367866393946 7222.39452782197 7816.17983180476    134297.474248711
#> 5   27.9172617142627 2296.64313319282 2440.50161985155    54805.6165966941
#> 6   13.1020208301306 4797.90755960131 4834.25110950227    125430.668276395
#> 7   7.88850561210395 16772.9028056482 17296.5409078761    346397.108173326
#> 8   15.3228485101145 5113.01907120317 5253.71003244198     114865.73821492
#> 9   5.63607271689266 7726.34524343147 7808.62927334621    214628.713082419
#> 10  1.56842500735605 19939.4414502043  20208.586519023    463220.604866945
#> 11  10.7021950585846 6527.12372025329 6995.15364088579    157783.361984959
#> 12  4.73051739493653 9164.88812340891 9210.11705160904    262873.449765325
#> 13  1.50654785148499 20147.5496424569 20648.5492614656     458548.67756957
#> 14  11.2259408578199 6858.34041263511  7087.4407720307    171445.338219553
#> 15  4.36642365487402 8952.93204811723 9042.99984194468    260000.451667249
#> 16  1.32793938295784  20498.407910722 20916.3918567406    498421.945863162
#> 17    9.282286059183 6653.03508347816 7120.85750591068    166990.006730234
#> 18   5.9385612159795 9311.21400116935 9339.90233029906    266663.460738019
#> 19  1.55939661019304 19611.8039235424 20000.8964250434    487355.594394914
#> 20  10.2123804748145 7028.79856431115 7310.12963081646    173189.349855401
#> 21  6.28971720205055 9231.81016320257 9273.25248392414    259270.092726524
#>    ci95.integral.bt.up
#> 1     86550.2685019685
#> 2      44175.073489127
#> 3     68410.4635272626
#> 4     160265.921763674
#> 5     60272.5525827846
#> 6     135241.794229403
#> 7      376512.61871204
#> 8     123974.484666585
#> 9     228237.271654271
#> 10    504651.806831237
#> 11     175071.56220721
#> 12    273763.352978196
#> 13    510629.378914608
#> 14     182939.72371588
#> 15    270626.745794586
#> 16    531485.711773182
#> 17    184985.967720108
#> 18    276412.468884772
#> 19    518356.894150817
#> 20    188200.829331549
#> 21    274674.210608324