Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.3563882163188  15.2339432557087
#> 2                 FALSE          0.75 32.8511901701752  31.2800598488224
#> 3                 FALSE          0.75 64.9079876358958   20.328324562421
#> 4                 FALSE          0.75 241.674580593105  15.0004523410413
#> 5                 FALSE          0.75 65.8944381111761  22.1174559250932
#> 6                 FALSE          0.75 205.567637794528  12.3225047843447
#> 7                 FALSE          0.75  505.36778759869  6.92057596490398
#> 8                 FALSE          0.75 198.182502073168  13.0864992812418
#> 9                 FALSE          0.75 289.714516341521  4.69849192721864
#> 10                FALSE          0.75 664.009641340674 0.467457070413824
#> 11                FALSE          0.75 325.462059911396  9.96418669987222
#> 12                FALSE          0.75  334.80058638418  3.71256654977188
#> 13                FALSE          0.75 721.482244987257 0.722023149141278
#> 14                FALSE          0.75 345.751059326871   9.8906766825979
#> 15                FALSE          0.75 320.594005327478  3.40903727233812
#> 16                FALSE          0.75 785.922870811186 0.595913458765901
#> 17                FALSE          0.75 333.142730956248  8.12974805954302
#> 18                FALSE          0.75 350.525303790799  4.97671929014247
#> 19                FALSE          0.75 767.366589150338 0.537794228400094
#> 20                FALSE          0.75 359.308415116381  9.65640683688898
#> 21                FALSE          0.75 337.819562725643  5.20766902317625
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1   3413.2052778077 2140.46856427811 81600.3793328507 2.53268665790117
#> 2   1460.7890174716 285.542246575742 42336.0586389896  2.1669881128955
#> 3  2195.15130591354 669.962272785054 66112.7598213123 1.58677410317117
#> 4  7549.04123078006 6019.63834825272 148786.899215952 2.50688304788434
#> 5  2365.23662721315 1136.16993019846 57231.3725644957 2.96386797257477
#> 6   4823.5429301832 3250.78747160441 130709.025240961 2.66652633555223
#> 7  17093.4348558056 15912.9450910129 364603.649629227 15.5051575216393
#> 8  5190.47106860773 3823.98255184986 119524.802215028 10.3841661593018
#> 9  7772.17072666457 6291.38018961499 221386.338077918 7.94880392122444
#> 10 20064.9868763505  18751.026214843 485740.901201508 27.9710288811273
#> 11  6757.8818979672 5522.83163700304 166620.168175271 13.6984386809834
#> 12 9181.34671287298 7665.87331419063 266610.529921926 6.79363674758534
#> 13 20419.4548927131 19353.6951168461 487416.888163821 21.4738199388727
#> 14 6969.44457432499 5635.55948815765 176868.175529528 12.8070401729751
#> 15 9001.71347569709 7394.34744149645 265346.556051283 4.60442124043672
#> 16  20634.098195715 19252.4265554203 511381.791166781 21.9442847586181
#> 17  6890.0271443951 5956.44283225541 176435.257887401 14.2966325018174
#> 18 9318.42070263151 7648.38695852044 271764.660838781 5.63907082472158
#> 19  19818.165763841 18687.0189869851 505146.859346007  21.257455693735
#> 20 7130.49606700815 5787.63939050455 179912.448697129 12.7178784378716
#> 21 9254.17155905928 7460.15865853387 267420.058931959 7.53753135971641
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.691979149943407 27.3157958608933 63.8154796008622  1597.4876490876
#> 2   1.95140163741828 19.8766086061148 60.6211327127519 896.812868115905
#> 3   1.75647186655743 5.76233851924659 117.453073328398 1079.36237346046
#> 4  0.337752458949931 86.6240895412809 120.880342262609 4229.53027473237
#> 5   2.72414228430841 36.6043065237578  144.68941096782 1791.42555404023
#> 6  0.426889601279553 6.96989270410895 87.0397658865014 2023.12822446587
#> 7   0.52432586182248 105.910259730362 157.326610769253 5485.21387072466
#> 8  0.921963093472167 38.1552087254064 167.704419165366 1673.12346093194
#> 9  0.443532991907324 21.5819840574429 77.4851273531097 3907.20269606912
#> 10 0.327572805943697 58.4555694125927 155.857800082395  8643.0821353632
#> 11 0.718072706054863   106.5538382967 234.715212930069 3827.16790658243
#> 12 0.390499408880852 16.3990343163669 100.690538256115 4095.98459412809
#> 13 0.281376201174933 128.864754059413 233.245180593846 12064.4129310432
#> 14 0.665173655746957 59.6068130901956 201.628891286759 3139.52009664083
#> 15 0.409996278058658 25.1845928689434 124.515576695033 2966.84976190176
#> 16 0.419628958873485 85.8185053640305  266.81919635241 9543.32244461244
#> 17  0.73663113637542 72.7577359160913 177.131058271653 2413.79979123823
#> 18 0.421066227539002 11.4356681592841 126.504089235993  2055.0703597537
#> 19 0.156241482447051 97.3067046109046 152.687470496129  6416.9503011323
#> 20  0.31714696591495 128.739027451857 144.116004323974 4319.70069169244
#> 21 0.767518438781747 12.8486088596156 194.850969543147 4583.74455870217
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.1901186640714 95.5226577685662    14.0956375540518
#> 2             TRUE 29.2864947244621 36.4158856158883    28.0700041552694
#> 3             TRUE 62.2977442361792 67.5182310356123     17.438928341934
#> 4             TRUE 237.550757979335 245.798403206875    14.4448495460687
#> 5             TRUE 61.0188752962906 70.7700009260616    17.6362418674058
#> 6             TRUE 201.181201972545 209.954073616512    11.6202713902398
#> 7             TRUE 479.861803475593 530.873771721787      6.058059922206
#> 8             TRUE 181.100548741116 215.264455405219      11.56986999248
#> 9             TRUE 276.638733891107 302.790298791935    3.96888015553109
#> 10            TRUE  617.99729883122 710.021983850129 -0.0714001953635585
#> 11            TRUE 302.928128281178 347.995991541613    8.78295709841197
#> 12            TRUE 323.625053934402 345.976118833958    3.07019502216288
#> 13            TRUE 686.157811187811 756.806678786702   0.259159298208514
#> 14            TRUE 324.683478242327 366.818640411415    8.79646601889415
#> 15            TRUE  313.01973238696 328.168278267996    2.73459339493163
#> 16            TRUE  749.82452238326 822.021219239113 -0.0943761785809821
#> 17            TRUE 309.624770490759 356.660691421738    6.91798984020546
#> 18            TRUE 341.249032284132 359.801575297466    4.28406534584081
#> 19            TRUE 732.398074534143 802.335103766532   0.280776989774696
#> 20            TRUE 338.387505086082  380.22932514668    9.13470007795889
#> 21            TRUE  325.42032363891 350.218801812377    3.94510119138027
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.3722489573656 3368.27079361653 3458.13976199887    78972.5121501016
#> 2   34.4901155423755 1428.09199631454 1493.48603862866    40860.8014709389
#> 3    23.217720782908 2185.67225904938  2204.6303527777    64337.2087169698
#> 4    15.556055136014 7406.54460348465 7691.53785807547    141829.321914018
#> 5   26.5986699827805 2305.02254298157 2425.45071144474    54284.4775280996
#> 6   13.0247381784495 4812.07745668494 4835.00840368146    127380.979311715
#> 7   7.78309200760196 16919.2124785492 17267.6572330621    355580.472811885
#> 8   14.6031285700035 5127.70575025444 5253.23638696102    116772.514121795
#> 9   5.42810369890618 7736.66836289008 7807.67309043906    214958.989642884
#> 10  1.00631433619121 19968.8274646668 20161.1462880342    471523.031088835
#> 11  11.1454163013325 6582.60083396913 6933.16296196527    160324.476968943
#> 12  4.35493807738088 9154.37030142255  9208.3231243234    259872.635264585
#> 13  1.18488700007404 20207.4723722854 20631.4374131409    467570.928892255
#> 14  10.9848873463016 6871.39136679162 7067.49778185836    171703.664970554
#> 15  4.08348114974461 8960.28482042768  9043.1421309665    260466.088192955
#> 16  1.28620309611278 20492.9267543911 20775.2696370388    495683.025745393
#> 17  9.34150627888059 6770.34066881313 7009.71361997707    172464.557230814
#> 18  5.66937323444412 9299.60902850949 9337.23237675353    268384.070096986
#> 19 0.794811467025493  19658.096234756 19978.2352929259    494590.976100644
#> 20  10.1781135958191 6918.72036684984 7342.27176716645    172806.541059295
#> 21  6.47023685497222 9233.03559748522 9275.30752063335    259879.799132894
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84228.2465155998 86.3923223668325 96.3204540658051   13.8776641218196
#> 2     43811.3158070403    28.6038934689 37.0984868714503   27.4553126394826
#> 3     67888.3109256547 61.7979103936803 68.0180648781112   16.8856397039685
#> 4     155744.476517887 236.761089819252 246.588071366958   14.3384575214995
#> 5     60178.2676008919 60.0852568849295 71.7036193374227   16.7781370478487
#> 6     134037.071170207 200.341246176846  210.79402941221   11.4858011658367
#> 7      373626.82644657 474.977678856277 535.757896341103   5.89289727573192
#> 8     122277.090308261 177.829536400936 218.535467745399   11.2794516180363
#> 9     227813.686512951 274.134860655921 305.294172027121   3.82916726308028
#> 10     499958.77131418 609.186424733665 718.832857947684 -0.174585629235823
#> 11    172915.859381599 298.613120096668 352.310999726123   8.55676419600469
#> 12    273348.424579267 321.485058358912 348.116114409447   2.94718770836541
#> 13    507262.847435387 679.393557907066 763.570932067447   0.17052579483841
#> 14    182032.686088503 320.649260587839 370.852858065902   8.58693631733386
#> 15    270227.023909612 311.569339696222 329.618670958734   2.60544456734315
#> 16    527080.556588168 742.912072684295 828.933668938078  -0.22655930062613
#> 17    180405.958543988 305.121331252686  361.16413065981    6.6859510322472
#> 18    275145.251580576 339.472724974345 361.577882607253   4.15142948416602
#> 19     515702.74259137 725.701975990617 809.031202310058  0.231560922803875
#> 20    187018.356334963 334.381373378153 384.235456854609   9.03479878369568
#> 21    274960.318731024 323.046001260599 352.593124190687   3.70333288316402
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.5902223895978 3359.66631792035 3466.74423769505     78469.303540639
#> 2   35.1048070581623 1421.83086460361 1499.74717033958    40578.3054174824
#> 3   23.7710094208736 2183.85712241582 2206.44548941127    63997.2095693298
#> 4   15.6624471605832 7379.25801527915 7718.82444628097    140497.019877477
#> 5   27.4567748023376 2293.49218642659 2436.98106799972    53720.1784785769
#> 6   13.1592084028526 4809.88194048315 4837.20391988325    126743.693921008
#> 7   7.94825465407604 16885.8507467341 17301.0189648771    353852.630442607
#> 8   14.8935469444472 5115.68685950593 5265.25527770953    116245.480231601
#> 9   5.56781659135699 7729.87003791198 7814.47141541716    213728.220793622
#> 10  1.10949977006347 19950.4139603018 20179.5597923992    468800.460216196
#> 11  11.3716092037397 6549.03637490567 6966.72742102873    159118.919078369
#> 12  4.47794539117835  9149.2046056129 9213.48882013306    258582.400117435
#> 13  1.27352050344415 20166.8799747567 20672.0298106696    463770.638818977
#> 14  11.1944170478619 6852.61522066821 7086.27392798177    170714.716140112
#> 15  4.21262997733309 8952.35167367396 9051.07527772022    259531.530517956
#> 16  1.41838621815793 20465.8939252015 20802.3024662285     492676.87917534
#> 17  9.57354508683885 6747.42198199956 7032.63230679064    171704.210296574
#> 18  5.80200909611891 9296.00679303931 9340.83461222371    267736.722933664
#> 19 0.844027533996314 19627.4446228036 20008.8869048783    492569.636755788
#> 20  10.2780148900823 6878.16757320251 7382.82456081379    171445.835341412
#> 21  6.71200516318847 9228.98828569444 9279.35483242413    258435.919596903
#>    ci95.integral.bt.up
#> 1     84731.4551250624
#> 2     44093.8118604968
#> 3     68228.3100732948
#> 4     157076.778554428
#> 5     60742.5666504146
#> 6     134674.356560914
#> 7     375354.668815848
#> 8     122804.124198455
#> 9     229044.455362213
#> 10     502681.34218682
#> 11    174121.417272172
#> 12    274638.659726417
#> 13    511063.137508666
#> 14    183021.634918944
#> 15    271161.581584611
#> 16    530086.703158221
#> 17    181166.305478228
#> 18    275792.598743898
#> 19    517724.081936226
#> 20    188379.062052846
#> 21    276404.198267015