Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  91.928662397574   15.617673412894
#> 2                 FALSE          0.75 32.7239121600829  31.2787379319294
#> 3                 FALSE          0.75 64.6330519411418  20.8812397903349
#> 4                 FALSE          0.75 243.239540686929  15.0275156316926
#> 5                 FALSE          0.75 64.9142621098959  23.0225904072301
#> 6                 FALSE          0.75 205.545856954219  12.4742354354949
#> 7                 FALSE          0.75  508.25980060088  7.08443612727579
#> 8                 FALSE          0.75 201.663802535489  13.3801169606339
#> 9                 FALSE          0.75  289.49683813056  4.71424063695979
#> 10                FALSE          0.75 664.933694548276 0.636118951503195
#> 11                FALSE          0.75 327.243111153674   10.074883706131
#> 12                FALSE          0.75 332.038826221035  3.58759428175996
#> 13                FALSE          0.75 729.877388121957 0.762102005887756
#> 14                FALSE          0.75 340.537481760443   9.8415938871556
#> 15                FALSE          0.75 319.962841683185  3.51404072833879
#> 16                FALSE          0.75 771.462702027213 0.457283927838987
#> 17                FALSE          0.75 336.539037511752  8.15586144267998
#> 18                FALSE          0.75 348.027983403982  5.11255771526349
#> 19                FALSE          0.75 777.871882276928 0.688951580752098
#> 20                FALSE          0.75 348.371137326209  9.62251559571361
#> 21                FALSE          0.75 340.448655298577  5.40490701702058
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.00291088999 2105.61712783707 81572.3867978242 2.61763120591053
#> 2  1467.84408290936 283.590305852503 42661.5209199582 2.18964728284461
#> 3  2195.87255660208 642.063033846485 66155.7374445901 1.44953090109774
#> 4  7550.93863107372 6026.05325059231 148104.732188659 1.79507495626533
#> 5  2362.38591976271 1047.64954231254  57264.079632139 1.80205101242146
#> 6  4822.15506290271 3218.19563520121 130507.313178755 3.12855128568748
#> 7  17060.3446821336 15831.5201158028 363793.127468529 13.4953455632508
#> 8    5170.859502499  3762.0672917417  118617.75778079 7.91215757247584
#> 9  7763.99145223556 6277.32129307807 222997.930173883 4.81837482787872
#> 10 20083.1385779567 18647.5978951886   487664.0374224 23.9906266141637
#> 11 6800.80747426633 5555.42665162316 168364.833266198 11.6260430185307
#> 12 9189.90412257835 7698.66090141472  266909.74989909  7.0439659002967
#> 13 20438.0805723531 19393.7646321977 490381.930443772 19.1403494714216
#> 14 6972.24519139127 5631.86094143374 177419.967171082 12.4234289083863
#> 15 9004.24516980705 7362.90621037688   264378.0703206 5.76260790544522
#> 16 20643.8877000325 19298.9701524538 515715.934146354 22.0219955039657
#> 17 6885.21452865732 5993.56993553439 175121.870309987 10.0816637005031
#> 18 9318.24096132304 7566.28499615341 271788.656990842 7.22271680830837
#> 19 19798.1505173928 18566.9282390806 500557.478912104 22.9089311038809
#> 20 7145.83486361617 5773.51001039034 180432.197128329 13.9785959302773
#> 21 9255.70140052725 7441.34990669717 268339.526637449 7.12684163101403
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.611616314894755 33.8937744399411 67.0342735490062 1993.40208682525
#> 2    2.0316631975056 23.2384419549005 83.6151884866491  810.54379241554
#> 3   1.65650356475929 6.48869051977054 107.742052697126 1125.49639258985
#> 4  0.253620778862851 68.0847609969499 69.9521927752216 3879.68862695549
#> 5   2.08172693450912 38.3011985488292   107.1842096442 1402.38707833415
#> 6  0.343952097500378 8.13212201071559 65.6319070462943 2002.54662055338
#> 7  0.606480577546572 91.2873405377938 224.587422431471 6083.17826371781
#> 8  0.550143347182159 54.5921269238519 90.9618700786806 2801.09554623941
#> 9  0.346433936003955 25.4583614521094 90.6718754840593 1534.84863961503
#> 10  0.43913469212858  73.300342844434 230.775284419656 4787.99137231047
#> 11 0.563974339616574 73.8460264604716 165.265474945008 2129.15167036802
#> 12 0.342682117080887 9.90081565220473 112.817144007287 4690.30003795151
#> 13 0.309966748303293 74.4546377385537 194.054058754434 7148.23664185911
#> 14  0.49874863029421 40.3004135551663 148.490322379519 2772.48757601573
#> 15 0.502802742522489 12.0863523418187  145.91122410616 4969.14401232719
#> 16 0.412101070891541 76.1509733376412 333.555533733172 7078.39188591244
#> 17 0.497262782977991 90.5508026487685 162.010692395594 3271.90980384898
#> 18 0.706845547084734 7.54665204802612 194.110067446547  2498.9854436489
#> 19 0.304708950423619 72.5808727374912 169.638929376614  8830.2250459449
#> 20 0.493569098876977 83.0919471424923 139.401980783292 3265.17131395704
#> 21  0.57692904524446 14.2244545224662 169.026400537728 2703.56338315108
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.6226590638512 96.2346657312968   14.6115645748921
#> 2             TRUE 29.1219423798035 36.3258819403623   27.9366519720327
#> 3             TRUE  62.248573608836 67.0175302734476   18.1562914263058
#> 4             TRUE 240.286642383873 246.192438989986   14.6103094504632
#> 5             TRUE 61.9498881944626 67.8786360253292   19.5981495999626
#> 6             TRUE 200.399390089263 210.692323819175   11.9084342351068
#> 7             TRUE 486.059957149332 530.459644052428   6.08677557721168
#> 8             TRUE 188.648303328766 214.679301742212   12.4751311545193
#> 9             TRUE   281.5706115387 297.423064722421   4.14435681223328
#> 10            TRUE 625.469113767976 704.398275328575  -0.08625761704832
#> 11            TRUE 308.118270388191 346.367951919157   9.14714591746171
#> 12            TRUE 320.451502315047 343.626150127023   3.02388219916191
#> 13            TRUE 698.391513241469 761.363263002446  0.252206704928839
#> 14            TRUE 320.100941206148 360.974022314739   9.02115239032162
#> 15            TRUE 310.483351678728 329.442331687642    2.6869302168893
#> 16            TRUE  735.23651942319 807.688884631237 -0.220622333777598
#> 17            TRUE 319.954700724424 353.123374299079   7.33786416468119
#> 18            TRUE 336.146614254314 359.909352553649   3.94979679030911
#> 19            TRUE 740.186690611044 815.557073942812  0.187705357305244
#> 20            TRUE 325.376347020903 371.365927631515   8.81059442806098
#> 21            TRUE 328.725000815559 352.172309781595   4.45585873759344
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6237822508959 3351.24765193629 3462.75816984369    78293.2403649966
#> 2   34.6208238918261 1429.61684589355 1506.07131992517    41328.1763814347
#> 3   23.6061881543639 2185.19866069706  2206.5464525071    64304.2958787798
#> 4    15.444721812922 7438.93919923374  7662.9380629137    141722.644397317
#> 5   26.4470312144976 2299.38044814988 2425.39139137553    54957.1528882793
#> 6    13.040036635883 4808.77772219508 4835.53240361033    127213.123987945
#> 7    8.0820966773399 16910.1770069489 17210.5123573183    353786.299224713
#> 8   14.2851027667486 5081.05545370926 5260.66355128873    114009.955607226
#> 9   5.28412446168629 7722.11244764684 7805.87045682428    220473.104161716
#> 10  1.35849552005471 19962.5595139776 20203.7176419358     479787.79161495
#> 11  11.0026214948002 6679.33076073885  6922.2841877938    164862.378768442
#> 12  4.15130636435802 9173.61728083048 9206.19096432623     259194.20633666
#> 13  1.27199730684667 20315.6026932732  20560.558451433    478623.081167913
#> 14  10.6620353839896 6905.95101109302 7038.53937168952    172859.225108536
#> 15  4.34115123978829 8984.36312020476 9024.12721940934    256203.828420322
#> 16  1.13519018945557  20518.619348892 20769.1560511729    504071.979494028
#> 17  8.97385872067878  6736.2584583001 7034.17059901455    169739.578682656
#> 18  6.27531864021788 9305.82671870403 9330.65520394204     267677.82593604
#> 19  1.19019780419895 19678.7549817396 19917.5460530459    486031.758711524
#> 20  10.4344367633662 7009.14861056677 7282.52111666557    175060.990316869
#> 21  6.35395529644771 9232.30217283779 9279.10062821671    263892.164872166
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     84851.5332306517 86.7981052339894 97.0592195611586   14.4189054357003
#> 2     43994.8654584818 28.4322034857075 37.0156208344583   27.2966780648185
#> 3     68007.1790104004 61.7919713749903 67.4741325072934   17.6344928034067
#> 4         154486.81998 239.721193772649 246.757887601209   14.5304189051214
#> 5     59571.0063759986 61.3822421255498  68.446282094242   18.9424056155922
#> 6     133801.502369565 199.413896434272 211.677817474167   11.8000893243942
#> 7     373799.955712345 481.808923296908 534.710677904852   5.89573419528451
#> 8     123225.559954354 186.155973693436 217.171631377541   12.3018360001569
#> 9      225522.75618605 280.052823467918 298.940852793203   4.03523012239203
#> 10    495540.283229851 617.912066384515 711.955322712036 -0.224585045068823
#> 11    171867.287763953 304.456066837354 350.030155469994   8.96949400048249
#> 12    274625.293461521 318.232653056454 345.844999385617   2.91593733228143
#> 13     502140.77971963 692.362303157971 767.392473085944  0.154567179213302
#> 14    181980.709233628 316.187561100006  364.88740242088   8.86404657177895
#> 15    272552.312220878 308.668130188512 331.257553177858   2.52854735299472
#> 16     527359.88879868  728.29959083944 814.625813214986 -0.350434171108433
#> 17    180504.161937319 316.778976658765 356.299098364738   7.18122638804312
#> 18    275899.488045645 333.871458459697 362.184508348266   3.72714044297742
#> 19    515083.199112683 732.970377313322 822.773387240535 0.0917220379218039
#> 20    185803.403939788 320.973089302866 375.769185349553   8.65512016191473
#> 21    272786.888402733 326.480045701789 354.417264895364   4.27412608834144
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8164413900877  3340.5711129877 3473.43470879227    77665.3187076467
#> 2   35.2607977990404 1422.29673667775 1513.39142914096    41072.8550868238
#> 3   24.1279867772631 2183.15472318333 2208.59039002083     63949.764515114
#> 4   15.5246123582638  7417.4924995197 7684.38476262774    140500.542479826
#> 5   27.1027751988679   2287.315570607 2437.45626891841     54515.400958604
#> 6   13.1483815465957  4806.2161037617 4838.09402204371     126582.32180247
#> 7   8.27313805926707 16881.4214946795 17239.2678695877    351870.098071642
#> 8    14.458397921111 5063.85893372825 5277.86007126975    113127.610510161
#> 9   5.39325115152754 7714.09306378942 7813.88984068169    219989.626840237
#> 10  1.49682294807521 19939.4699059816 20226.8072499318    478279.574332672
#> 11  11.1802734117795  6656.0692624038 6945.54568612885    164191.695992276
#> 12   4.2592512312385 9170.49852390003 9209.30972125667    257716.761824705
#> 13  1.36963683256221 20292.1494823855 20584.0116623207    476371.386625728
#> 14  10.8191412025322 6893.25638082315  7051.2340019594    171985.891522091
#> 15  4.49953410368287 8980.55591921708 9027.93442039701    254638.548056439
#> 16  1.26500202678641 20494.6317922907 20793.1436077742    501842.286049965
#> 17  9.13049649731685 6707.73495546574 7062.69410184891    168708.927094443
#> 18  6.49797498754957  9303.4495233089 9333.03239933717     266890.64552129
#> 19  1.28618112358239 19655.8920068273 19940.4090279583    483250.237822052
#> 20  10.5899110295125 6982.97464721689 7308.69508001546    174032.461352973
#> 21  6.53568794569972 9227.82146966322 9283.58133139128    263040.542406473
#>    ci95.integral.bt.up
#> 1     85479.4548880017
#> 2     44250.1867530927
#> 3     68361.7103740662
#> 4     155708.921897491
#> 5     60012.7583056739
#> 6      134432.30455504
#> 7     375716.156865416
#> 8     124107.905051419
#> 9     226006.233507528
#> 10    497048.500512129
#> 11    172537.970540119
#> 12    276102.737973475
#> 13    504392.474261816
#> 14    182854.042820073
#> 15    274117.592584761
#> 16    529589.582242742
#> 17    181534.813525531
#> 18    276686.668460394
#> 19    517864.720002156
#> 20    186831.932903684
#> 21    273638.510868426