Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2910064576407  15.4814042381414
#> 2                 FALSE          0.75 32.6306423853985  31.4741925943155
#> 3                 FALSE          0.75 65.3554425323453  21.0654629898889
#> 4                 FALSE          0.75  241.62290039962  14.8928812612654
#> 5                 FALSE          0.75 65.8903171406681  23.3676739787547
#> 6                 FALSE          0.75 205.635184294491   12.365454353154
#> 7                 FALSE          0.75 504.994555985478  6.91879376347214
#> 8                 FALSE          0.75 202.425934946429  13.3337510058902
#> 9                 FALSE          0.75 290.968874808613  4.88086950551834
#> 10                FALSE          0.75 675.066934359897  0.61762436401615
#> 11                FALSE          0.75 324.749082399661  9.82242355144679
#> 12                FALSE          0.75 333.840724877451  3.89229742609489
#> 13                FALSE          0.75 734.505770138636 0.904764106581717
#> 14                FALSE          0.75 341.813777633018  9.81831462831966
#> 15                FALSE          0.75 320.290105901161  3.60889541525197
#> 16                FALSE          0.75  770.31848398135 0.397220504133315
#> 17                FALSE          0.75 335.347272757032   7.9830313234379
#> 18                FALSE          0.75 350.552600852264  5.04520565436502
#> 19                FALSE          0.75 773.528384241943 0.622505452013766
#> 20                FALSE          0.75 353.199258623594   9.5082051437057
#> 21                FALSE          0.75 337.892625903196  5.47258640263888
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3408.40268519061 2124.40294471181 81510.7643312887 2.03888376497802
#> 2   1471.1332447054  295.36993803114 42743.2820089998 1.91284507816303
#> 3  2198.02622580799 628.122862160932 66339.7368427261 1.95426719363437
#> 4  7566.11005624968 6059.00346965305 148956.464635858 2.09771605491649
#> 5   2374.4229317639 1088.13626905311  57168.137483905 2.08554769981191
#> 6  4821.35334350286 3236.79863704238 130851.540750378 3.12093118906315
#> 7   17006.918050125 15816.6567957026 361009.577277215  8.1253174644759
#> 8  5172.39645446108  3771.2324771122 118648.464257736 7.31304650126344
#> 9  7760.72188807277  6236.5590070481 220223.386139799 7.13039657415131
#> 10 20059.5255506886  18714.217292501 486350.445933913 19.2915536095232
#> 11 6782.98471877407 5596.38090993537 167015.875100223 11.4808727054364
#> 12  9191.0073902592 7628.53212637826 268469.971895678 5.08870710312994
#> 13 20433.4028590029 19336.9155519515 489524.577820569 30.1763633187993
#> 14 6962.11208862192    5621.55200211 176622.252140486 17.4982249494963
#> 15 9006.66453977016  7336.0131770026 265668.096939781 5.36592981657441
#> 16 20705.5961310786  19442.343246318 516632.460081698 19.5577665891297
#> 17 6911.61145019101 6040.51508651241 176380.377875249 10.8209015797723
#> 18 9320.56561752757 7640.25293498138 270200.557836242 7.37067385945023
#> 19 19794.6862791222 18592.3520736418 500072.448989315 21.7901919866237
#> 20 7154.33781874146 5846.10519261846 181096.364586643 13.0808601131834
#> 21 9258.04502606145  7384.1496486985 267402.749289572 5.97161811307048
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.617669538860365 24.7601532141327 62.7539047048849 1629.10027988338
#> 2   1.67755271617104 21.5032987498979 56.0854699840153  668.94630171885
#> 3   1.37783912341709 5.09529654061946 90.6161178136526  813.38482389882
#> 4  0.312170652931141 51.9847983377782 90.3828852572659  2539.0086383687
#> 5   2.90550438169526 27.5960575323795 123.226022897919 1112.56307893926
#> 6  0.543973596780986 8.46634188738119 111.746337051614 2332.52278057939
#> 7  0.336669919359873 172.368457490007 212.304629323294  9645.5090352209
#> 8   0.50158643401343  38.465258258905 110.852248219817 2349.34808766719
#> 9  0.433057406855743 39.2808653943499 83.3253974696555 4666.18319656981
#> 10 0.347117803750261 84.5398664866994 206.591074135103 6112.21844368423
#> 11 0.623521206627198 97.0808437377153  182.67595778325 3757.36756168675
#> 12 0.302859442497074 12.0982191741661 78.1659546423388 2332.91523170266
#> 13 0.373920189102384 81.9558662156915 229.361795367674  9040.8820317974
#> 14 0.937701731631848 59.6064003786265 267.989429416046 3331.42635757186
#> 15  0.45364663532781 21.9583518668154 109.659815135886 2736.57011339036
#> 16 0.203491973708649  92.751959563451 142.164991046924 9396.44200006793
#> 17 0.576577684673451 87.1757534695215 130.451831607534 2930.84473649441
#> 18 0.294413548723016 9.21105937032667 83.6836258433887 4073.94225898748
#> 19 0.288074914641697 93.3113188869236 204.097280242763 11624.7159648969
#> 20 0.368993454877757 68.5566885894793 134.204518013951 2193.12435729203
#> 21 0.846914900765841 10.2731802734819 234.083404284182 4012.27943600488
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.9370426642518 95.6449702510295   14.4653378467161
#> 2             TRUE 29.4840122318203 35.7772725389767   28.7146183762141
#> 3             TRUE 62.1406729988167 68.5702120658738   18.7989176318678
#> 4             TRUE 238.172157489283 245.073643309958   14.3793605371936
#> 5             TRUE 62.4595911744776 69.3210431068587    18.588119270866
#> 6             TRUE 200.501252488482   210.7691161005   11.4706177864492
#> 7             TRUE 491.628408756415 518.360703214541   6.36497174612515
#> 8             TRUE 190.395973451851 214.455896441008   12.5086413219381
#> 9             TRUE 279.239372444134 302.698377173092   4.16849007124064
#> 10            TRUE 643.332328672231 706.801540047563 0.0466155768469715
#> 11            TRUE 305.863046799218 343.635118000104   8.79673116654505
#> 12            TRUE 325.469801692803   342.2116480621    3.3940936431872
#> 13            TRUE 684.865652479211 784.145887798061  0.289665395508296
#> 14            TRUE 313.029197591097  370.59835767494   8.27579527978527
#> 15            TRUE 311.463151352897 329.117060449426   2.86264670013772
#> 16            TRUE 738.145957942232 802.491010020469 0.0624762073825879
#> 17            TRUE 317.546889658307 353.147655855757   7.03456103215007
#> 18            TRUE 338.427842353468  362.67735935106   4.56089536671565
#> 19            TRUE 737.683518423947 809.373250059939  0.148622217428175
#> 20            TRUE 331.681243737408 374.717273509781   8.90121091043179
#> 21            TRUE 328.069314107195 347.715937699197   4.07941139087907
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4974706295667 3367.67223315336 3449.13313722786    78830.8943708806
#> 2   34.2337668124168 1435.76031826182 1506.50617114898    41642.8653426723
#> 3     23.33200834791 2189.64446299867 2206.40798861731    65001.7188074126
#> 4   15.4064019853371 7480.59506298403 7651.62504951532    144779.795425741
#> 5   28.1472286866434 2329.02741712314 2419.81844640467    55337.9712190499
#> 6   13.2602909198587 4807.42621109812  4835.2804759076    127014.540776325
#> 7   7.47261578081913 16723.3719375539  17290.464162696    345142.714914276
#> 8   14.1588606898423 5109.12110462518 5235.67180429698    114783.786653523
#> 9   5.59324893979604 7696.10486449906 7825.33891164647    212547.514781441
#> 10  1.18863315118533  19920.457470318 20198.5936310592    476295.846594053
#> 11  10.8481159363485 6623.28673082553 6942.68270672261    160835.005461249
#> 12  4.39050120900258 9171.10581971769  9210.9089608007    264632.326339528
#> 13  1.51986281765514 20298.5854590781 20568.2202589278    474652.326878263
#> 14  11.3608339768541 6864.05955999907 7060.16461724476     171142.05578228
#> 15  4.35514413036622 8970.54305094925 9042.78602859107    261166.439103254
#> 16 0.731964800884042 20553.0191575968 20858.1731045605    501175.312991586
#> 17  8.93150161472572 6768.20733573365 7055.01556464837    171559.138283716
#> 18  5.52951594201438 9305.41342486339 9335.71781019176    263498.922820208
#> 19  1.09638868659936 19641.1891595532 19948.1833986911     480949.79122706
#> 20  10.1151993769796 7041.56206601177 7267.11357147116    177488.675018898
#> 21  6.86576141439869 9241.14564451157 9274.94440761133    260802.549617344
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up    ci95.lambda.bt.lo
#> 1     84190.6342916969 88.2947942782838 96.2872186369976      14.270771941975
#> 2     43843.6986753273  28.881466032199  36.379818738598     28.1861892706202
#> 3     67677.7548780397 61.5250788328219 69.1858062318686     18.3648983079914
#> 4     153133.133845975 237.511376931984 245.734423867257     14.2810267815203
#> 5     58998.3037487601 61.8026436490368 69.9779906322995      17.672885390632
#> 6     134688.540724431 199.518159163927 211.752209425055     11.2992661034632
#> 7     376876.439640153 489.068933755105 520.920178215851     6.25892072152678
#> 8     122513.141861948 188.092363803953 216.759506088906     12.3506415952239
#> 9     227899.257498156 276.993297523276  304.94445209395     4.03207698808108
#> 10    496405.045273774 637.255489285231 712.878379434562  -0.0627265313343607
#> 11    173196.744739198 302.246571897006 347.251592902316     8.60032198645748
#> 12    272307.617451829 323.866858955317 343.814590799586     3.29869291880063
#> 13    504396.828762876 675.360098033789 793.651442243483    0.171880535941045
#> 14    182102.448498692 307.517256732006 376.110298534031     7.98041923432124
#> 15    270169.754776308 309.772883460676 330.807328341647     2.71974801000946
#> 16     532089.60717181 731.985261466656 808.651706496044 -0.00162376433563649
#> 17    181201.617466783 314.138305660678 356.556239853386     6.85293906147793
#> 18    276902.192852277 336.106080087742 364.999121616787      4.4681550988679
#> 19    519195.106751571  730.81960794816 816.237160535725   0.0578786193160411
#> 20    184704.054154389 327.560772801755 378.837744445434     8.78497797214529
#> 21      274002.9489618 326.188254401578 349.596997404814     3.81263319713783
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6920365343077 3359.87278489091 3456.93258549031    78317.7277827173
#> 2   34.7621959180107  1428.9867791556  1513.2797102552    41432.1472576308
#> 3   23.7660276717864 2188.03944458837  2208.0130070276    64745.5025878844
#> 4   15.5047357410104 7464.21985150763 7668.00026099172    143980.007704655
#> 5   29.0624625668774 2320.33465900044 2428.51120452737    54987.5138491841
#> 6   13.4316426028447 4804.75931340359 4837.94737360212    126279.796100442
#> 7   7.57866680541749 16669.0758734445 17344.7602268054    342104.379568182
#> 8   14.3168604165565 5097.00454827362 5247.78836064853    114043.742005908
#> 9    5.7296620229556 7683.73139189984 7837.71238424569    211077.667074522
#> 10  1.29797525936666 19893.8274123747 20225.2236890026    474370.497784292
#> 11  11.0445251164361 6592.70626504815 6973.26317249999    159651.434679317
#> 12  4.48590193338915 9167.29488067783 9214.71989984056    263897.458041541
#> 13  1.63764767722239 20272.7693612202 20594.0363567857    471804.449038246
#> 14  11.6562100223181 6845.28354387981 7078.94063336402    170092.656479645
#> 15  4.49804282049448 8963.62617011121 9049.70290942912    260304.419517536
#> 16 0.796064772602266 20523.8022903343  20887.389971823    498215.433761565
#> 17  9.11312358539786 6740.74697339075 7082.47592699127     170635.92219172
#> 18  5.62225620986213 9302.51194116173 9338.61929389341    262215.631008627
#> 19  1.18713228471149 19611.7960941038 19977.5764641405    477288.005698117
#> 20  10.2314323152661 7019.96670910608 7288.70892837684    176797.840846351
#> 21  7.13253960813993 9237.90959272542 9278.18045939747    259538.681595003
#>    ci95.integral.bt.up
#> 1     84703.8008798601
#> 2     44054.4167603687
#> 3     67933.9710975678
#> 4     153932.921567061
#> 5      59348.761118626
#> 6     135423.285400313
#> 7     379914.774986248
#> 8     123253.186509563
#> 9     229369.105205076
#> 10    498330.394083534
#> 11    174380.315521129
#> 12    273042.485749816
#> 13    507244.706602892
#> 14    183151.847801327
#> 15    271031.774362026
#> 16    535049.486401831
#> 17    182124.833558778
#> 18    278185.484663858
#> 19    522856.892280513
#> 20    185394.888326936
#> 21    275266.816984142