Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 90.9757681944976  15.5198817120875
#> 2                 FALSE          0.75 33.1286406745465  30.8483336277911
#> 3                 FALSE          0.75 64.7858763082086  20.0264738480128
#> 4                 FALSE          0.75 242.537272654946  15.0038488379999
#> 5                 FALSE          0.75 65.2876865522601  23.3135308307861
#> 6                 FALSE          0.75 206.478044858852  12.4103542811484
#> 7                 FALSE          0.75 505.604227433824  6.88345240415273
#> 8                 FALSE          0.75 203.217283610652  13.6726778524348
#> 9                 FALSE          0.75 289.897134834546  4.72329266536535
#> 10                FALSE          0.75 659.411801151514 0.516557041935237
#> 11                FALSE          0.75 324.697054163745  9.84407588804053
#> 12                FALSE          0.75 336.146498153974  3.85983133870195
#> 13                FALSE          0.75 720.158755973946 0.757389888024887
#> 14                FALSE          0.75 347.984348468644  9.87150941561576
#> 15                FALSE          0.75 320.782181665082  3.38480116666028
#> 16                FALSE          0.75 772.623005158797 0.442116781390169
#> 17                FALSE          0.75 331.208103150747  8.03390341969915
#> 18                FALSE          0.75 351.224847777708  5.05594077982858
#> 19                FALSE          0.75 769.872429712236 0.638482426304935
#> 20                FALSE          0.75 350.091807512737  9.43458712195567
#> 21                FALSE          0.75 336.238100962108  5.10506247642776
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3417.08444458114 2118.94449532449  81784.711892648 2.18331865579481
#> 2  1465.64367191269 309.825730540907 42471.5082675949 2.03422372845325
#> 3  2195.29311996469 687.805555119239  66221.203452191  1.6486215929847
#> 4  7544.38974430523 6021.22765107366 148093.565843287 2.52691346619928
#> 5  2366.49792835148  1072.1150035276 57550.5126353953 2.53677183661265
#> 6  4816.88072999728 3226.03493209309 130710.732834234 2.61258226839173
#> 7  17075.6261103895 15901.1990630718 365024.429866635 14.3102211091615
#> 8  5178.58546088533 3734.89305924955 118456.192288845 8.00066974817117
#> 9   7768.3884818891 6272.42976736611 219809.640287042 7.20574544588711
#> 10 20087.4175888978 18714.5626848158 483922.695041359   13.11756077295
#> 11 6760.89896534114 5565.90831700192 166658.496732008 13.7853128915669
#> 12 9189.21423236953 7660.33349016668 265706.931561305 8.29548661108594
#> 13 20435.8731765774 19301.6633048089 490597.131165933 24.6060585613636
#> 14 6957.55782451447 5653.06333151876  176067.15367333 17.6750272971342
#> 15 8998.03185130571 7411.84153232675 264885.308407612 5.35980099571297
#> 16 20701.9991371535 19394.4125430611 514309.889810843 22.4167723453004
#> 17  6920.3251928497 6024.94444875635 176595.463271637  12.752941850527
#> 18 9317.18376843216 7626.03795991743 270825.714615691 6.16654994140115
#> 19 19785.5807788166 18571.3699162834 502976.981278343 36.4863770798162
#> 20 7160.46030660956 5863.56471335087 181012.056148734 12.0980609052483
#> 21 9255.97279896366 7480.92943825858 266938.516532017 8.59642089029588
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.564956702959024 24.4578611141073 58.1025033971806 1330.45698857852
#> 2    2.9264763932895 26.2601668907822 87.0479915407236 947.900623383904
#> 3   1.31101626549829 4.51342206837948 86.8788566059423 872.378144338747
#> 4  0.329089525010674 72.3146015907057 104.112993003418 3736.43456220547
#> 5   2.02815434245518  26.380938521642 111.267322218012 1141.09785966715
#> 6  0.521783951396585 11.1333519855724 106.833382769084 1548.54241194346
#> 7  0.545976537636622 108.622285882415 166.611332593602 5813.32504271652
#> 8  0.685206608551927  39.215757307992 156.759164696197 2602.39831581705
#> 9  0.407300318216648 25.1856440571488 71.1274762401788 4099.02048941989
#> 10 0.291037654895445 79.1381934850833 274.523078579441 7702.13175950943
#> 11 0.460465089952328 104.834096491009 160.127761178019 4504.62539653579
#> 12 0.392278705614545 9.86254100797107 85.5663250778388 6256.15301560314
#> 13 0.330504763724202 110.238909582162 253.020229251943 10045.7490549035
#> 14 0.616824656219076 65.1289173168553 230.321929719899 3570.23288137771
#> 15 0.366791689362688 19.2526196572166 85.8493508926973 3249.15494433559
#> 16 0.311314086930907 101.154188108127 267.893095615563 8810.40584197865
#> 17 0.441445416306274  67.024714742844 111.238664886453 2001.89194858871
#> 18   0.5521136048745  10.531566011629 124.994522687428 2832.76098551325
#> 19 0.432047139106309 82.4815481674618 258.487349235341 9248.63315392471
#> 20 0.350573167780315 99.2274899500793 129.087686099094 3524.69999667346
#> 21 0.457173612197859 8.49191317303373 120.626709493844  4232.1145868426
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.3842090057151   94.56732738328    14.5905279357199
#> 2             TRUE 29.7823426412409 36.4749387078521    26.0342799608299
#> 3             TRUE 62.0738937877488 67.4978588286685    17.8698520912681
#> 4             TRUE 238.380500003048 246.694045306844    14.4624965693574
#> 5             TRUE 61.1146968810323 69.4606762234879    19.9772169374473
#> 6             TRUE 202.180347027347 210.775742690356     11.552019681101
#> 7             TRUE 482.063913709253 529.144541158395    5.98532099974049
#> 8             TRUE  190.05618187491 216.378385346394    12.5455129813668
#> 9             TRUE 278.043683576062  301.75058609303    4.05328364189897
#> 10            TRUE 637.833413680011 680.990188623017    0.03780009963223
#> 11            TRUE 302.020214457117 347.373893870373    9.08661081506895
#> 12            TRUE 322.500422678737  349.79257362921    3.21453286796602
#> 13            TRUE 679.681789640503 760.635722307389   0.213709551698575
#> 14            TRUE 318.908928564858  377.05976837243    8.85683285613538
#> 15            TRUE 311.965309027134  329.59905430303    2.78142883765866
#> 16            TRUE 735.747414650778 809.498595666816 -0.0699948916111731
#> 17            TRUE  310.22951380663 352.186692494864    7.30772570987533
#> 18            TRUE 341.080873124103 361.368822431313    4.14771389981003
#> 19            TRUE 709.852339415939 829.892520008534 -0.0722351175249437
#> 20            TRUE 330.190497323603  369.99311770187    8.85789426095705
#> 21            TRUE 322.096988597572 350.379213326645    4.35301188436228
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4492354884551 3376.85126304843 3457.31762611385    79596.1101464363
#> 2   35.6623872947524 1422.44569737735 1508.84164644802    40912.2117421284
#> 3   22.1830956047575 2187.86854066221 2202.71769926718    64786.1414047538
#> 4   15.5452011066425 7425.43222468852 7663.34726392194    141947.130988459
#> 5   26.6498447241249 2323.10128448337 2409.89457221958    55673.4066562428
#> 6   13.2686888811958 4798.56636598102 4835.19509401355    128163.380566587
#> 7   7.78158380856498 16896.9424501129 17254.3097706661    355461.510171367
#> 8   14.7998427235027 5114.07554011368 5243.09538165698    114175.247059326
#> 9   5.39330168883174 7726.95809741509 7809.81886636311    213066.751581946
#> 10 0.995313984238244 19957.2352606149 20217.5999171808    471252.688296966
#> 11  10.6015409610121 6588.44687661343 6933.35105406885    159248.387954706
#> 12  4.50512980943788 9172.99035241142 9205.43811232764    255415.559850638
#> 13   1.3010702243512 20254.5301703148 20617.2161828401    474071.873970617
#> 14  10.8861859750961 6850.42075552824  7064.6948935007    170194.120583463
#> 15   3.9881734956619 8966.36129196958 9029.70241064183     259540.44852418
#> 16 0.954228454391511 20535.6004977156 20868.3977765913    499816.772200789
#> 17  8.76008112952297 6810.06953709772 7030.58084860168    173302.351016208
#> 18  5.96416765984714 9299.85934234303 9334.50819452129    266165.822794522
#> 19  1.34919997013481 19649.8986320811 19921.2629255521    487762.979740137
#> 20  10.0112799829543 6997.23108564168 7323.68952757744    175213.924654207
#> 21  5.85711306849324 9242.00360179402  9269.9419961333    259976.688036661
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83973.3136388596 86.6964636291397 95.2550727598554    14.4125665742878
#> 2     44030.8047930615 29.1415621667782 37.1157191823149    25.1124398969437
#> 3     67656.2654996283 61.5545779859586 68.0171746304587    17.4568819676361
#> 4     154240.000698115 237.584522261195 247.490023048697     14.358833368979
#> 5     59427.6186145478 60.3156137524993 70.2597593520209    19.3383483195739
#> 6     133258.085101881 201.357383612804 211.598706104899    11.3876577364111
#> 7     374587.349561904 477.556194059867 533.652260807781    5.81333839038496
#> 8     122737.137518364 187.535970904237 218.898596317068     12.329672899673
#> 9     226552.528992138 275.773873760607 304.020395908485    3.92498404166072
#> 10    496592.701785752 633.701382036532 685.122220266496 -0.0538767616598351
#> 11    174068.605509309 297.677840896274 351.716267431216    8.94156431173397
#> 12    275998.303271972 319.887344396245 352.405651911702    3.09096507569744
#> 13    507122.388361249 671.930881193673 768.386630754219   0.109600551125451
#> 14    181940.186763196 313.341294966261 382.627401971027    8.66253308942637
#> 15    270230.168291045 310.276971713485  331.28739161668    2.66588945550941
#> 16    528803.007420898 728.686131362008 816.559878955586  -0.168058828994409
#> 17    179888.575527065 306.212337123714  356.20386917778    7.16867040373885
#> 18     275485.60643686 339.138409892561 363.311285662854    3.97379811427456
#> 19    518190.982816549 698.359130635797 841.385728788676  -0.208329966343431
#> 20    186810.187643262  326.37960813845 373.804006887024    8.74746371310625
#> 21    273900.345027373 319.389116017128 353.087085907088    4.20900219651996
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6271968498872 3369.14703679749 3465.02185236479    79177.0161950341
#> 2   36.5842273586386 1414.17374480675 1517.11359901862    40613.6230457625
#> 3   22.5960657283894 2186.44681271067 2204.13942721872    64511.3422892871
#> 4   15.6488643070209 7402.65312518745 7686.12636342301    140770.154101365
#> 5   27.2887133419982 2314.79128884906 2418.20456785389    55313.9608304477
#> 6   13.4330508258857 4795.05936010556 4838.70209988901    127675.589706825
#> 7   7.95356641792051 16862.7264300599  17288.525790719    353630.312782911
#> 8   15.0156828051965 5101.72257656167 5255.44834520899    113355.491589844
#> 9   5.52160128906998 7719.02461953709 7817.75234424111    211775.560127779
#> 10  1.08699084553031 19932.3067296671 20242.5284481286     468826.51679272
#> 11  10.7465874643471 6555.42413621877 6966.37379446352    157829.430954797
#> 12  4.62869760170646 9169.88365199391 9208.54481274515    253444.871650723
#> 13  1.40517922492432 20219.8049137964 20651.9414393585    470907.463018322
#> 14  11.0804857418052 6829.90514657343 7085.21050245551    169069.497225829
#> 15  4.10371287781115 8960.29671677756 9035.76698583385    258516.964716715
#> 16  1.05229239177475 20503.7369284615 20900.2613458454    497041.494360565
#> 17  8.89913643565944 6788.95675195373 7051.69363374568    172671.755052403
#> 18   6.1380834453826 9296.54189904937 9337.82563781495    265273.503084085
#> 19   1.4852948189533 19623.9169444084 19947.2446132248    484849.660296651
#> 20  10.1217105308051 6965.97442630741 7354.94618691172    174103.644155254
#> 21  6.00112275633557 9239.32864914452 9272.61694878281    258643.571941806
#>    ci95.integral.bt.up
#> 1     84392.4075902618
#> 2     44329.3934894274
#> 3      67931.064615095
#> 4      155416.97758521
#> 5     59787.0644403429
#> 6     133745.875961643
#> 7      376418.54695036
#> 8     123556.892987846
#> 9     227843.720446305
#> 10    499018.873289997
#> 11    175487.562509218
#> 12    277968.991471887
#> 13    510286.799313544
#> 14     183064.81012083
#> 15     271253.65209851
#> 16    531578.285261122
#> 17    180519.171490871
#> 18    276377.926147297
#> 19    521104.302260036
#> 20    187920.468142214
#> 21    275233.461122229