Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8650805987712  15.4002228468992
#> 2                 FALSE          0.75 32.7546285965595  30.3051090346229
#> 3                 FALSE          0.75 64.2710148489488  20.1308403590214
#> 4                 FALSE          0.75 241.082298090455  14.9882573914826
#> 5                 FALSE          0.75 65.6270446698123  22.4332973851127
#> 6                 FALSE          0.75 205.942945856768  12.2907184226981
#> 7                 FALSE          0.75 509.111811121025  7.02630396623374
#> 8                 FALSE          0.75 200.407330773502  13.4796077174996
#> 9                 FALSE          0.75 289.767087518019  4.65358168043154
#> 10                FALSE          0.75 668.140902473875 0.498421067739532
#> 11                FALSE          0.75 327.659547283373  9.94613632250777
#> 12                FALSE          0.75 331.943325117883  3.74719348027477
#> 13                FALSE          0.75 733.594246538989 0.858188180445382
#> 14                FALSE          0.75 338.425785738329  9.71451420221371
#> 15                FALSE          0.75 319.989729551849  3.35193740600515
#> 16                FALSE          0.75 780.733521371086 0.446426697998718
#> 17                FALSE          0.75 332.603168281945  8.02688407768145
#> 18                FALSE          0.75 352.593453742862  5.01502338076252
#> 19                FALSE          0.75 779.333530306393 0.667098981466821
#> 20                FALSE          0.75 353.800098069599  9.61004657332146
#> 21                FALSE          0.75  337.59120473926  5.42283583301188
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3405.57744981178  2123.4785550433 81165.0364802617  2.1597714587126
#> 2  1455.93728582576 296.263542802696 42654.0704799213 1.79084483460293
#> 3   2195.5055396136 686.936599946082 66139.4480914087 1.89381793511821
#> 4  7552.00268643033 6019.02628318264 148407.647817358 3.32235594769591
#> 5  2357.73189099819 1085.53170777094 57276.8307545383 2.96735066318656
#> 6  4815.67118976176 3251.46997336903 129963.657130337 3.22094072803532
#> 7  17023.5487368537 15847.4700374006 360964.402128326  9.8697453634781
#> 8  5193.81797435853 3778.32811596666 119428.047020583 8.28774577464031
#> 9  7760.99833636841 6278.14760275604 221594.341240279 7.26520078163729
#> 10 20074.3103869185 18766.0852319497  483326.81903664 27.3588009867267
#> 11 6788.45520092894 5564.59953544122 167881.430224937 11.3186345726569
#> 12  9189.2655465641 7645.78478420253 267788.363388242 5.84021089126108
#> 13   20432.07104239   19342.73688549 491010.725112114 28.4933235226007
#> 14 6953.51803425407 5614.90046049637 176041.244389328 16.6904471673675
#> 15 8999.48535485187 7408.52563832903 264860.913695042 5.61882875204373
#> 16 20713.4770726663 19412.8027196253 514613.112484806 25.0124582696224
#> 17 6878.94367550251 5980.10827661779 174563.559001222  15.755871297994
#> 18    9318.91794637 7660.62893195234 271475.176185474 7.26262753998582
#> 19 19791.8287019072 18591.8825761635 506278.817021193 23.3130560164387
#> 20 7104.23532203484 5754.84494948332 179626.011296631  13.605976098738
#> 21 9258.75609596377 7396.77848319209  267488.67298371 6.87211514527282
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.633927458405965 42.8605067639494 79.3293339884838 2342.18943638769
#> 2   2.65509607727347 20.1951608727753 76.3906923452899 619.333059372775
#> 3   1.67647518746202  5.2527972866959 101.796989832152 995.028997732572
#> 4  0.255258181012053 75.1574387757175 92.4236280578589 3916.79289577137
#> 5   2.16444828301973 37.8862862311917 143.841304384983 1687.84531520377
#> 6  0.493463889667277  9.7251858687842 103.914242582323 2907.45892419088
#> 7  0.337047127486051 139.014965415753 148.019335470124 7609.54979779371
#> 8  0.695009366613334 36.3174806475479 102.449612538776 1897.59993362384
#> 9  0.349738628037608 18.8179140318882 80.9602602141232 3553.87165178235
#> 10 0.175291062395751 84.5819118519236 231.494865633338 10239.1255082251
#> 11 0.762061415526743 74.3432133924111 238.255895782806 3083.55109050052
#> 12 0.526520808283586 10.1833856430349 140.611133517142  4072.4870167814
#> 13 0.278951758380223 93.3788542543585 258.176266461633 9970.52385381904
#> 14 0.782216646274604 69.1927700660618 254.760119679669 3671.32665046813
#> 15 0.450742217087151 16.7236073482984 122.730776945233 3570.14616807233
#> 16 0.377912923051336 114.406808876842 282.833449492652 10467.2289234178
#> 17 0.595848111858591 128.367079423924 169.037785442253 6005.08912574633
#> 18 0.410267416462188  11.247783831713 95.7007126330348 3224.11179422792
#> 19 0.263890769745846 71.5954867469463 174.428272541952 5698.17725275336
#> 20 0.452033023218237 99.9815478896231 159.967834474808 4012.24239663676
#> 21 0.659894439278687 8.58006159610296 204.953250393015 4478.47648182339
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up ci90.lambda.bt.lo
#> 1             TRUE  88.312256549189 95.4179046483534  14.3574121778214
#> 2             TRUE 29.8086888436377 35.7005683494813   25.937475987508
#> 3             TRUE 61.1556843456793 67.3863453522183  17.3730386756463
#> 4             TRUE 235.617022556495 246.547573624415  14.5683576837178
#> 5             TRUE 60.7457528288704 70.5083365107542  18.8727799595453
#> 6             TRUE  200.64449835915 211.241393354386  11.4789703241955
#> 7             TRUE 492.876079998104 525.347542243947  6.47186144151918
#> 8             TRUE 186.773988974219 214.040672572785  12.3363173094206
#> 9             TRUE 277.815832232226 301.718342803813  4.07826163730968
#> 10            TRUE 623.135674850709  713.14613009704 0.210067270098521
#> 11            TRUE 309.040393411352 346.278701155393  8.69254529396627
#> 12            TRUE 322.336178201759 341.550472034008  2.88106675064827
#> 13            TRUE 686.722729344311 780.465763733667 0.399312537909916
#> 14            TRUE  310.97000014801 365.881571328649  8.42776781909198
#> 15            TRUE 310.746756254737 329.232702848961  2.61046645889679
#> 16            TRUE 739.588027517557 821.879015224615 -0.17524006042073
#> 17            TRUE 306.684759996745 358.521576567145  7.04671393367407
#> 18            TRUE 340.646431439585 364.540476046138  4.34013348068222
#> 19            TRUE 740.983553159352 817.683507453435 0.232998665234905
#> 20            TRUE 331.418267387175 376.181928752023  8.86645225012746
#> 21            TRUE 326.286575325287 348.895834153234  4.33730948039844
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.443033515977 3335.07191618508 3476.08298343847    77312.1348574039
#> 2   34.6727420817377 1422.71624619004 1489.15832546148    41635.2675972531
#> 3   22.8886420423964 2186.86468807699 2204.14639115022    64502.6253901387
#> 4   15.4081570992474 7428.36869964427 7675.63667321639    141964.523503815
#> 5   25.9938148106801 2295.40895014788  2420.0548318485    54500.3252110281
#> 6   13.1024665212008 4799.67325900761 4831.66912051591    125180.887200043
#> 7   7.58074649094829 16794.8691187448 17252.2283549626    348446.692710955
#> 8   14.6228981255785 5134.07571869332 5253.56023002375    116306.495129772
#> 9   5.22890172355341 7730.04286778595 7791.95380495086    215748.222373097
#> 10 0.786774865380542  19935.173141922 20213.4476319149     466483.45757561
#> 11  11.1997273510493 6666.16061489843 6910.74978695946    162808.988681063
#> 12  4.61332020990127  9172.5138771813 9206.01721594689    261089.122245637
#> 13  1.31706382298085 20278.4628271416 20585.6792576384    474609.213372581
#> 14  11.0012605853354  6839.6959274954 7067.34014101274    170001.912049308
#> 15  4.09340835311352 8971.97502076392 9026.99568893982    258988.023248563
#> 16  1.06809345641817 20525.2778720639 20901.6762732687    497394.520905784
#> 17  9.00705422168883 6667.77982985015 7090.10752115486     164685.18738937
#> 18  5.68991328084282 9300.41534196683 9337.42055077317    266171.512283969
#> 19  1.10119929769874 19674.0541262085  19909.603277606    496905.315440414
#> 20  10.3536408965155 6939.76567575641 7268.70496831327    173025.872554164
#> 21  6.50836218562532 9244.64189463819 9272.87029728936    260121.579171111
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up ci95.lambda.bt.lo
#> 1     85017.9381031194 87.6319285396945 96.0982326578479  14.1577250284235
#> 2     43672.8733625896 29.2445727207378 36.2646844723813  25.1011207231669
#> 3     67776.2707926788 60.5591316961171 67.9828980017805  16.8449489915958
#> 4     154850.772130902 234.570480432971 247.594115747939   14.487951356699
#> 5     60053.3362980485 59.8110373699666 71.4430519696579   18.190978750394
#> 6     134746.427060631 199.629902029819 212.255989683717  11.3235291989503
#> 7     373482.111545697 489.767110208608 528.456512033443  6.36569159636108
#> 8     122549.598911394 184.163349055207 216.651312491797  12.1173893589374
#> 9     227440.460107461  275.52729398601 304.006881050028  3.96809396947783
#> 10    500170.180497671  614.51765253989 721.764152407859  0.15485058544386
#> 11     172953.87176881 305.475023520965  349.84407104578  8.45249594807535
#> 12    274487.604530848 320.496511771012 343.390138464755  2.71521269603894
#> 13    507412.236851646 677.747332434692 789.441160643286 0.311442734020145
#> 14    182080.576729348 305.712509290289  371.13906218637  8.18136957551548
#> 15    270733.804141521 308.976825197843 331.002633905855  2.46848266051434
#> 16    531831.704063828 731.709103162626 829.757939579546  -0.2942826311819
#> 17    184441.930613075 301.721660537877 363.484676026014  6.85902177843861
#> 18    276778.840086979 338.358703764489 366.828203721234  4.21089924449663
#> 19    515652.318601973 733.639940514173 825.027120098613 0.149873072764964
#> 20    186226.150039099 327.132384916073 380.467811223126  8.72406184781372
#> 21     274855.76679631 324.121859054526 351.060550423995  4.12944273202566
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6427206653749 3321.57085655444 3489.58404306912    76574.3451849418
#> 2   35.5090973460789 1416.35477051512  1495.5198011364    41440.1776835507
#> 3   23.4167317264469 2185.21005693168 2205.80102229553    64189.1912558529
#> 4   15.4885634262662 7404.69410642992 7699.31126643074    140730.733741647
#> 5   26.6756160198314 2283.47476998505 2431.98901201133    53968.6539367389
#> 6    13.257907646446 4796.60982545894 4834.73255406458    124265.037638923
#> 7    7.6869163361064 16751.0794046388 17296.0180690686     346049.68452465
#> 8   14.8418260760617 5122.63571228934 5265.00023642773     115708.75115068
#> 9   5.33906939138526 7724.11522486591 7797.88144787091    214628.752802785
#> 10 0.841991550035203 19908.5298396887 20240.0909341482    463258.133040519
#> 11  11.4397766969402 6642.74250267982 6934.16789917807    161837.670087555
#> 12   4.7791742645106 9169.30611070375 9209.22498242444    259806.288835351
#> 13  1.40493362687062 20249.0484880515 20615.0935967286    471468.498358628
#> 14  11.2476588289119 6817.90020492459 7089.13586358355     168845.44415441
#> 15  4.23539215149597  8966.7070844492 9032.26362525453    257863.427205621
#> 16  1.18713602717934 20489.2397272677 20937.7144180649    494097.343794907
#> 17  9.19474637692429 6627.34419983162  7130.5431511734     162793.58431476
#> 18  5.81914751702841 9296.87229005984 9340.96360268016    265155.917068787
#> 19  1.18432489016868 19651.5015478832 19932.1558559313    495110.389605797
#> 20  10.4960312988292 6908.27148817118  7300.1991558985    171762.016199223
#> 21  6.71622893399811 9241.93917523541 9275.57301669214    258710.859079336
#>    ci95.integral.bt.up
#> 1     85755.7277755815
#> 2      43867.963276292
#> 3     68089.7049269646
#> 4      156084.56189307
#> 5     60585.0075723377
#> 6     135662.276621751
#> 7     375879.119732002
#> 8     123147.342890486
#> 9     228559.929677772
#> 10    503395.505032761
#> 11    173925.190362318
#> 12    275770.437941134
#> 13    510552.951865599
#> 14    183237.044624245
#> 15    271858.400184464
#> 16    535128.881174705
#> 17    186333.533687685
#> 18     277794.43530216
#> 19     517447.24443659
#> 20    187490.006394039
#> 21    276266.486888084