Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.6977191840187   15.522910481383
#> 2                 FALSE          0.75 32.9291935764032  31.5135758975565
#> 3                 FALSE          0.75   65.18336278984  20.1786583417858
#> 4                 FALSE          0.75 243.329416346266  15.0288739758756
#> 5                 FALSE          0.75  65.035483875467  23.2366175931248
#> 6                 FALSE          0.75 205.056033765439  12.4729490553264
#> 7                 FALSE          0.75 508.167335095352  7.17330946101166
#> 8                 FALSE          0.75 202.517473328268  13.1353450479819
#> 9                 FALSE          0.75 290.969002676215  4.67155750718639
#> 10                FALSE          0.75 661.071652324237 0.641744373229173
#> 11                FALSE          0.75  320.57123533317  9.64937932823561
#> 12                FALSE          0.75 333.840962236081  3.76681079958032
#> 13                FALSE          0.75 714.835324897641 0.627102453302298
#> 14                FALSE          0.75 344.777173565639  9.87830901636049
#> 15                FALSE          0.75 321.490597758951  3.35961875693891
#> 16                FALSE          0.75 778.658300812427 0.536773399218228
#> 17                FALSE          0.75 338.498298253312  8.34749160233813
#> 18                FALSE          0.75 350.860477431796  5.14963876492764
#> 19                FALSE          0.75 782.210556817061 0.709284121657007
#> 20                FALSE          0.75 356.099644687351  9.73682990877608
#> 21                FALSE          0.75 338.452001588414  5.13316344591916
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.24722550126 2124.03500276106 81453.6998366471 2.06281883165924
#> 2  1462.63797476839 272.127727704973 42097.4920719934 1.64730274160602
#> 3  2196.25294476574  679.69112303507 65604.3889768068 1.82080914455231
#> 4  7562.03209960683 6039.53742798563 149225.046367333 2.27317681755409
#> 5  2375.75520045354 1088.38492683158 57583.4214998694 2.77171861543848
#> 6  4817.52244642325 3212.09448206571 130404.377940465 3.12083461985348
#> 7  17013.2360760219 15748.6007282474 360726.477161037 15.5445159980726
#> 8  5189.24692833802 3834.01846846764  118452.77681643 8.76600157649482
#> 9  7767.42645623424  6294.4136829444 221916.983567777 8.00614936145091
#> 10 20097.6402145818 18616.5056326297 485081.137345154 36.0412003374441
#> 11 6794.51752125706 5622.41905088905 168132.509456838 11.6893534993676
#> 12 9192.18202360259 7665.30377639905 267591.133515902 6.54295974369938
#> 13 20412.5917689558 19340.6534035053 489579.258112319 21.2048080572222
#> 14 6985.23876784488 5640.78512575021 177563.018568531 13.1937497437545
#> 15 9000.58118067112 7421.39150038568 264451.031671892 5.61830453747018
#> 16 20696.2176085328 19361.8998693849 515069.193506164  24.239408740523
#> 17 6904.30306490611 5953.02910460705 176268.723280414 13.0372083306349
#> 18 9318.91976447102 7595.77818662543  271229.68243725 6.29204704228672
#> 19  19763.944398555 18552.1180272098 504023.259770422 26.4434711495887
#> 20 7163.94827495786 5795.71508537963 181079.139688948 17.5875242232661
#> 21 9255.32155800194 7494.39889137601 267114.628014676 6.57639007144767
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.613130519611383 28.6590984371144  65.948842592615 1757.21468487703
#> 2   3.26269583398767 20.1635283655219 98.9444315756067 1099.56417074099
#> 3   2.05183876609081 4.06320205935177 139.481152958023 1746.36679993008
#> 4  0.203589574726863 68.6392192168425 82.3584259703554 3693.05049247838
#> 5   2.23657282935833 25.7443771832307 110.494187286041 851.519258010784
#> 6  0.526475329536298 8.51061659525583  96.556167257469 2325.56642361695
#> 7  0.537670855739018 140.171949926712 184.487014515745  7563.7714595104
#> 8  0.732025607273997 39.9875598876654 136.292272022737 2330.03953532667
#> 9  0.415961983980409 19.7785812899566 62.7261172275163 2946.56362059856
#> 10 0.421430505734107 76.2582033707359 166.084302814731 7194.37846778988
#> 11 0.557867420763672  76.827148563782 127.234358937319 2372.87941694345
#> 12 0.412013917620082 10.8317949254776 105.567639340806 3253.81405589668
#> 13 0.291030642525586 73.9419398173773 153.949601557657 8547.51346140798
#> 14 0.516883673443851 32.3210802459151 133.718993786738 2205.57763681303
#> 15 0.322321711601597 14.0988562662507 49.1134538404333 5987.79093060926
#> 16 0.426729084788161 75.1403049347588 309.059429394427   8182.939846131
#> 17 0.751402438390542 77.0995417124826 210.094879582181 3270.60438944105
#> 18 0.625147660221281 9.73111602549859 186.373411031039 2901.33289112672
#> 19 0.372239019522698  58.177767255919 201.797331159556 6346.64524293788
#> 20 0.476834299671331 69.4366706684213 120.174185948335 1582.31904712515
#> 21 0.389652239004028 9.83235027480653 107.802877604568 6538.73105675001
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.3043822059392 96.0910561620981   14.5143107766223
#> 2             TRUE 30.2193805664613 35.6390065863451   26.1464412506468
#> 3             TRUE 62.1881317470514 68.1785938326285   16.8033835715665
#> 4             TRUE 239.590040481389 247.068792211142     14.69396912545
#> 5             TRUE 60.4760067530707 69.5949609978633   19.5574552888304
#> 6             TRUE  199.92226081578 210.189806715098   11.6068971382391
#> 7             TRUE 482.596606278522 533.738063912181   6.28884090332097
#> 8             TRUE 188.097400734934 216.937545921602   11.9311629240162
#> 9             TRUE 277.798886976628 304.139118375802   3.98730004353861
#> 10            TRUE 601.783877769142 720.359426879333 -0.051508808703433
#> 11            TRUE  301.34224882671  339.80022183963   8.73168742107937
#> 12            TRUE 323.077793457696 344.604131014467   3.08904790509528
#> 13            TRUE 679.953415643511 749.717234151772  0.148357046347709
#> 14            TRUE 323.073455237163 366.480891894115   9.02803537354536
#> 15            TRUE 312.248486794813  330.73270872309   2.82939954135428
#> 16            TRUE 738.784473434267 818.532128190588 -0.165195945258298
#> 17            TRUE 317.052090549418 359.944505957207   7.11143459118569
#> 18            TRUE 340.510060047235 361.210894816358   4.12127086386363
#> 19            TRUE 738.711046775988 825.710066858135 0.0969509345421692
#> 20            TRUE 327.168167340078 385.031122034623   8.95243748581674
#> 21            TRUE 327.633839920882 349.270163255945   4.49218551275753
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5315101861437 3364.10300857221 3458.39144243032    78563.0816800244
#> 2   36.8807105444663 1429.46897060711 1495.80697892968    40288.7090111245
#> 3   23.5539331120052 2189.56897737811 2202.93691215338    62731.6155909218
#> 4   15.3637788263013 7449.12058399513 7674.94361521854    143149.978307206
#> 5   26.9157798974193 2333.40569998713 2418.10470091996    56182.6723204416
#> 6   13.3390009724136 4803.52248212405 4831.52241072244    126578.821173615
#> 7   8.05777801870234 16782.6532183924 17243.8189336513    348284.073110142
#> 8   14.3395271719477 5123.46739232281 5255.02646435323    114619.861780818
#> 9   5.35581497083416 7734.89069001226 7799.96222245622    217069.886411893
#> 10  1.33499755516178  19972.195470037 20223.0849591267     473246.38476564
#> 11  10.5670712353919 6668.13686186964 6920.89818064448    164229.122815966
#> 12  4.44457369406535 9174.36372095018   9210.000326255    262238.609393952
#> 13  1.10584786025689 20290.9572779562 20534.2262599554    475518.598468303
#> 14  10.7285826591756 6932.07059084035 7038.40694484941    173934.843355974
#> 15  3.88983797252354 8977.38856211314  9023.7737992291    254601.115591039
#> 16  1.23874274369475 20572.6118069152 20819.8234101505    501608.257459279
#> 17  9.58354861349057 6777.47431878907 7031.13181102314    170888.579059783
#> 18  6.17800666599165 9302.91207860908 9334.92745033297    266456.989831347
#> 19  1.32161730877185  19668.241971419  19859.646825691     493583.02834579
#> 20  10.5212223317354 7049.72495170831 7278.17159820741    178476.224856427
#> 21  5.77414137908078 9239.14734179988   9271.495774204    256358.415426322
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84344.3179932698 88.6545942739666 96.7408440940708    14.3211746629447
#> 2     43906.2751328623 29.7004802028554  36.157906949951    25.1186920629407
#> 3     68477.1623626917 61.6145768665175 68.7521487131625    16.1570543602479
#> 4      155300.11442746  238.87398978386 247.784842908672     14.629838409411
#> 5     58984.1706792971 59.6029153892076 70.4680523617264    18.8529348475825
#> 6     134229.934707315 198.939197910526 211.172869620352    11.4410574094352
#> 7     373168.881211931 477.700083739129 538.634586451574    6.11947458376318
#> 8     122285.691852043 185.336110238338 219.698836418198    11.7005748577249
#> 9     226764.080723662 275.276949927771 306.661055424659    3.85627201858479
#> 10    496915.889924669 590.430899662847 731.712404985628  -0.184259418009677
#> 11     172035.89609771 297.660102474409  343.48236819193    8.55595918353882
#> 12    272943.657637852 321.016761138431 346.665163333732    2.95926352104496
#> 13    503639.917756335 673.273901105486 756.396748689797   0.056682393952149
#> 14    181191.193781089  318.91742406788 370.636923063398    8.86521701641055
#> 15    274300.947752744  310.47872086551 332.502474652393    2.72786820219978
#> 16     528530.12955305 731.149059681002 826.167541943852  -0.299615606966569
#> 17    181648.867501044 312.945369925268 364.051226581357    6.87474282309266
#> 18    276002.375043154 338.528065228914 363.192889634678    3.92434935089393
#> 19    514463.491195055 730.381353363867 834.039760270255 -0.0203043566074806
#> 20    183682.054521469 321.628097209749 390.571192164952    8.80223468142027
#> 21     277870.84060303 325.562277048376 351.341726128451    4.36944505747126
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7246462998213 3355.07539256452 3467.41905843801    78009.5590542881
#> 2   37.9084597321724 1423.11745917197 1502.15849036482    39942.3462973411
#> 3   24.2002623233238 2188.28906872941 2204.21682080207    62181.5100489438
#> 4   15.4279095423403 7427.49922994182 7696.56496927185    141986.667402075
#> 5   27.6203003386672 2325.29622117441 2426.21417973267    55914.4437541682
#> 6   13.5048407012175 4800.84163789654 4834.20325494995    125846.267750176
#> 7   8.22714433826013 16738.4990541655 17287.9730978782    345901.485100396
#> 8    14.570115238239  5110.8713109582 5267.62254571785     113885.89932719
#> 9   5.48684299578799 7728.66043690593 7806.19247556256    216141.718871404
#> 10  1.46774816446802 19948.1741359752 20247.1062931885    470980.155548286
#> 11  10.7427994729324 6643.93631007205 6945.09873244208    163481.665799629
#> 12  4.57435807811568 9170.95170554865 9213.41234165653    261213.657966344
#> 13  1.19752251265245 20267.6655669137 20557.5179709979    472826.131727959
#> 14  10.8914010163104 6921.88945056289 7048.58808512687    173240.086400378
#> 15  3.99136931167804 8972.94742238927 9028.21493895297    252714.961447897
#> 16  1.37316240540302 20548.9426108607  20843.492606205    499030.631407747
#> 17  9.82024038158359 6753.18796314964 7055.41816666257    169858.338677109
#> 18  6.37492817896135 9299.84677706104   9337.992751881    265543.069970642
#> 19   1.4388725999215 19649.9159747334 19877.9728223766    491583.835094264
#> 20  10.6714251361319 7027.85240044776 7300.04414946797    177977.794356583
#> 21  5.89688183436705 9236.05015146332 9274.59296454056    254298.715143446
#>    ci95.integral.bt.up
#> 1     84897.8406190061
#> 2     44252.6378466457
#> 3     69027.2679046697
#> 4      156463.42533259
#> 5     59252.3992455705
#> 6     134962.488130754
#> 7     375551.469221677
#> 8     123019.654305671
#> 9      227692.24826415
#> 10    499182.119142023
#> 11    172783.353114047
#> 12    273968.609065459
#> 13    506332.384496678
#> 14    181885.950736685
#> 15    276187.101895886
#> 16    531107.755604581
#> 17    182679.107883718
#> 18    276916.294903859
#> 19    516462.684446581
#> 20    184180.485021314
#> 21    279930.540885906