Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.7222037131013  15.5642949959269
#> 2                 FALSE          0.75 32.6450367436557  31.1619323569464
#> 3                 FALSE          0.75 64.7213956849209  20.3082469359277
#> 4                 FALSE          0.75 241.900714067339  15.0118823787915
#> 5                 FALSE          0.75 66.2239308487023  22.9253322404774
#> 6                 FALSE          0.75 204.249432089037  12.1887195433786
#> 7                 FALSE          0.75 509.308696818769  7.24272395853882
#> 8                 FALSE          0.75  200.76567416168  13.2351046405465
#> 9                 FALSE          0.75 291.153074141475   4.7889107363548
#> 10                FALSE          0.75 670.207657511446 0.518526928985196
#> 11                FALSE          0.75 325.201342163526  9.95857019861798
#> 12                FALSE          0.75 335.530572665047  3.82190517121198
#> 13                FALSE          0.75 727.664968134656 0.770707989813348
#> 14                FALSE          0.75  343.88821947765  9.71453822781107
#> 15                FALSE          0.75 319.678162300422  3.44308159699836
#> 16                FALSE          0.75 776.055087505371 0.489562165287326
#> 17                FALSE          0.75 338.109172296659  8.05083266184193
#> 18                FALSE          0.75 350.697198071681  5.05083965549416
#> 19                FALSE          0.75 782.193379974219 0.685815718354791
#> 20                FALSE          0.75 349.893904838923  9.61418758648909
#> 21                FALSE          0.75  338.76181328662  5.34115889479829
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3405.87322056126 2110.39883761285 81361.4080158932 2.34968704211559
#> 2  1464.85171751559 279.516777761907  42363.044298123 2.25323439852673
#> 3  2198.51132503653 682.984422854256 65916.6495002193  1.6367380104747
#> 4  7564.72398172237 6029.13772152126 149525.804614677 2.08751660407205
#> 5  2367.97597644961  1103.8402919566 57033.8745022227 1.93761046792704
#> 6   4818.9485865906 3265.95855256057   130659.3419114 2.87157072333549
#> 7  17078.5877852486 15792.0924434422 363805.685632421 13.1465266529358
#> 8  5178.70189878521 3799.82861227919 119533.682185291 9.52583601077301
#> 9   7763.3642156547  6257.8634452025 221509.338370178 7.18593037318387
#> 10  20104.707197995 18785.5669947185  483928.48519919 22.7171602427378
#> 11 6767.80203953764 5519.80908666403 167187.022387256 12.9117748253633
#> 12 9186.66764778308 7658.34879631227 267478.350855694 4.93442549911668
#> 13 20406.2011833751   19334.61722699 487121.568177012 29.4610371302371
#> 14 6959.84176151854  5677.8663934438 176040.001970418 13.3888543578296
#> 15 9006.83025901054  7385.4041399143 264624.497612853 5.88887286198767
#> 16 20655.3655051498 19333.1713264542 514536.833033462 13.7214953818048
#> 17 6908.58406592782 6019.45990843997 176729.645041555 12.8594501666506
#> 18 9318.95057171859 7628.16297167327 270567.882658485 6.55924127459451
#> 19 19785.0122588436  18589.180137983 499854.279089785 18.0676486856206
#> 20 7144.18658153998 5786.32091378367  179973.11771829 7.59924803630082
#> 21 9257.53329426863 7436.70136376436 269267.274601642 7.07671543429994
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.671071745408235 23.6642604841999 65.6016987967218 1523.39271743447
#> 2   2.43084936160568 24.8962921241455 72.0336999717592 1064.17874106325
#> 3   1.67243710792514 2.91097471553656 119.175590079251 1482.59941023268
#> 4  0.284217988740078 36.6696348636795 72.2864299361358 2700.98184151989
#> 5   2.39467772228839 30.4532553944271 138.820106025471 1223.24492249104
#> 6  0.579425417777449 10.0367711995544 122.318125123379 2375.93723390191
#> 7  0.662264534289749 101.004663852731 204.748230723978 5099.23044803631
#> 8  0.649499090419702 39.4641394518423 128.276043193052 1800.16222533539
#> 9  0.389916804649777 19.7038807161335 86.1962523468634 2886.55464185564
#> 10 0.276360430521041 86.4535330980418 161.241507734745 8289.83232422956
#> 11 0.814072028156896 105.927822326659 213.910439434941 4460.28752716597
#> 12 0.518237431720109 8.76599972275244  130.83099372261 3984.04066212747
#> 13  0.46270840634382 102.840058233097  265.03392195534 9311.51707157327
#> 14 0.792242991934561 88.9245850548818  213.64732070188 5097.61613278158
#> 15 0.426366506103218 15.6650676092157 118.292217173837 3791.60239509867
#> 16 0.266212446459064 92.6645349339396 186.484561682658 9582.01290664412
#> 17 0.612524941175894 71.7198978406837 170.141714754292 3234.28797923698
#> 18  0.59643535962157 9.83365125200334 159.829972206158 3904.60883640951
#> 19 0.309214848106268 71.9368987506901 246.249447984863 12327.5089522323
#> 20 0.234375444194144 86.2021678527277 115.623008568982 3207.07353786554
#> 21 0.703445139884317 9.17831797764377 209.835506115292 2477.44340174831
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.8569685288212 95.5874388973815    14.4603819747304
#> 2             TRUE 28.9384661580793 36.3516073292322    27.1631851571051
#> 3             TRUE   62.02896165769 67.4138297121518    17.5570878933909
#> 4             TRUE  238.46674925364 245.334678881037     14.544343787314
#> 5             TRUE 63.0365616289623 69.4113000684423     18.986087387313
#> 6             TRUE  199.52569824915 208.973165928924    11.2355647311347
#> 7             TRUE  487.68266047469 530.934733162849    6.15329879963218
#> 8             TRUE 185.095673923958 216.435674399401    12.1666786368061
#> 9             TRUE 279.332218677588 302.973929605363    4.14749759270592
#> 10            TRUE 632.837928912142  707.57738611075  0.0639140207780845
#> 11            TRUE 303.961472575804 346.441211751249    8.61942171229989
#> 12            TRUE    327.413442719 343.647702611094     2.9694045960324
#> 13            TRUE 679.201562055416 776.128374213896 0.00955266137776345
#> 14            TRUE  321.86355405902 365.912884896279    8.41129850607872
#> 15            TRUE 309.990966442452 329.365358158391    2.74170869445857
#> 16            TRUE 753.483227602302  798.62694740844  0.0516426908621662
#> 17            TRUE 316.955376772518 359.262967820799    7.04322913360758
#> 18            TRUE 339.907246174973 361.487149968389    4.06970348891667
#> 19            TRUE 752.472097886373 811.914662062065   0.177157293219981
#> 20            TRUE 337.393141819208 362.394667858638    9.22863998078972
#> 21            TRUE 327.120616397196 350.403010176043    4.18399163968859
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6682080171234 3366.94551206476 3444.80092905777    78855.4269957136
#> 2   35.1606795567878 1423.89731697137 1505.80611805981    40612.4702690739
#> 3   23.0594059784646 2193.72277162948 2203.29987844359    63477.7734703865
#> 4   15.4794209702689 7504.40243237162 7625.04553107312    145082.689485377
#> 5   26.8645770936418 2317.88037132578 2418.07158157345    55021.6366047249
#> 6   13.1418743556225 4802.43809796733 4835.45907521387    126750.925161632
#> 7   8.33214911744545 16912.4351132108 17244.7404572863    355417.451545401
#> 8    14.303530644287 5113.78338938693 5243.62040818349    116572.415324614
#> 9   5.43032388000368 7730.95133187666 7795.77709943274    216760.955984326
#> 10 0.973139837192308 19962.4911360488 20246.9232599413    470291.711025832
#> 11  11.2977186849361 6593.55077181028 6942.05330726499    159849.849405068
#> 12  4.67440574639155 9172.24757823916 9201.08771732701    260924.603966494
#> 13  1.53186331824893 20237.0292875816 20575.3730791685    471804.122594274
#> 14  11.0177779495434 6813.56081910326 7106.12270393382    167654.423431992
#> 15  4.14445449953816 8981.06122279338  9032.5992952277    258387.311672916
#> 16 0.927481639712486 20502.9323451834 20807.7986651161    498774.421802032
#> 17  9.05843619007627  6790.6048339799 7026.56329787575     171409.24131571
#> 18  6.03197582207164 9302.77421540905 9335.12692802814    264144.801122591
#> 19   1.1944741434896 19666.6760603987 19903.3484572885    479575.526863363
#> 20  9.99973519218846 7002.38401542225 7285.98914765772    174697.481748501
#> 21  6.49832614990799 9242.43496119541 9272.63162734185    265191.880205766
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83867.3890360729 87.1168171105547 96.3275903156479    14.2489943749268
#> 2      44113.618327172 28.2286973225433 37.0613761647681    26.3974676081993
#> 3      68355.525530052 61.5133891843905 67.9294021854513    17.0302702043945
#> 4     153968.919743977 237.809181523357  245.99224661132    14.4548151208609
#> 5     59046.1123997205 62.4262143315653 70.0216473658393    18.2317639047922
#> 6     134567.758661169 198.621153471299 209.877710706774    11.0530457245348
#> 7     372193.919719441 483.541504579015 535.075889058523    5.94468547133091
#> 8     122494.949045968 182.095035580565 219.436312742795    11.9620864233239
#> 9     226257.720756031 277.068650610035 305.237497672916    4.02467379924124
#> 10    497565.259372548  625.68202343568 714.733291587212 -0.0231395148360433
#> 11    174524.195369444 299.894263505814 350.508420821238    8.36298902343047
#> 12    274032.097744894 325.859098686779 345.202046643316    2.80615980504056
#> 13     502439.01375975 669.921335359391 785.408600909921   -0.13620048662054
#> 14    184425.580508844 317.646064936304 370.130374018996    8.16174196361933
#> 15     270861.68355279 308.135971490926 331.220353109918    2.60740324503606
#> 16    530299.244264891 749.160956557034 802.949218453708 -0.0322142297724389
#> 17    182050.048767399 312.904649970024 363.313694623294    6.85028377713718
#> 18    276990.964194378 337.841085173476 363.553310969886    3.88182635063588
#> 19    520133.031316208 746.780788550403 817.605971398036  0.0797546160665065
#> 20    185248.753688078 334.999378687773 364.788430990072    9.15481171586857
#> 21    273342.668997518 324.891451035392 352.632175537848    3.96240642062503
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    16.879595616927 3359.49127001223  3452.2551711103    78375.5582897217
#> 2   35.9263971056936 1416.05498495227 1513.64845007892     40277.253965639
#> 3    23.586223667461 2192.80581459408 2204.21683547899    63010.7546561632
#> 4    15.568949636722 7492.85149738956 7636.59646605518    144231.880205298
#> 5   27.6189005761626 2308.28759587654 2427.66435702269    54636.3144541403
#> 6   13.3243933622224 4799.27651503947 4838.62065814173    126002.504932953
#> 7   8.54076244574672 16880.6186440972 17276.5569263999     353811.19395427
#> 8   14.5081228577692  5101.3521854596 5256.05161211082    116005.364223634
#> 9   5.55314767346836 7724.74460945108 7801.98382185832    215851.691272141
#> 10  1.06019337280644 19935.2582731229 20274.1561228672      467680.4138437
#> 11  11.5541513738055 6560.18350777738 6975.42057129789    158444.858834011
#> 12  4.83765053738339 9169.48628832649 9203.84900723968    259669.631157924
#> 13  1.67761646624724 20204.6346692382  20607.767697512    468870.994716729
#> 14  11.2673344920028 6785.54957481097 7134.13394822611    166048.674350166
#> 15  4.27875994896067 8976.12672649648  9037.5337915246     257192.95691846
#> 16  1.01133856034709 20473.7430166792 20836.9879936203    495756.087736439
#> 17  9.25138154654668 6768.01306616008 7049.15506569556     170390.44060225
#> 18  6.21985296035243 9299.67661526467 9338.22452817252    262914.849339122
#> 19  1.29187682064308 19644.0159372922 19926.0085803949     475692.36154341
#> 20  10.0735634571096 6975.23033254864 7313.14283053133    173687.253584073
#> 21  6.71991136897155 9239.54379103245 9275.52279750481    264411.485534216
#>    ci95.integral.bt.up
#> 1     84347.2577420648
#> 2     44448.8346306069
#> 3     68822.5443442753
#> 4     154819.729024056
#> 5     59431.4345503051
#> 6     135316.178889848
#> 7     373800.177310572
#> 8     123062.000146948
#> 9     227166.985468215
#> 10     500176.55655468
#> 11    175929.185940501
#> 12    275287.070553464
#> 13    505372.141637296
#> 14     186031.32959067
#> 15    272056.038307247
#> 16    533317.578330484
#> 17    183068.849480859
#> 18    278220.915977847
#> 19    524016.196636161
#> 20    186258.981852506
#> 21    274123.063669069