Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.8939230818734  15.4923953284178
#> 2                 FALSE          0.75 32.2272439575906  31.5525000788677
#> 3                 FALSE          0.75 64.8329591625595  20.3105270777427
#> 4                 FALSE          0.75  242.42386630698  15.0371518483565
#> 5                 FALSE          0.75 64.6209449684656  22.7754577770151
#> 6                 FALSE          0.75 205.698173727597  12.3562521523141
#> 7                 FALSE          0.75  507.69853876856  6.98737078386404
#> 8                 FALSE          0.75 200.352377414324  13.3695097803639
#> 9                 FALSE          0.75 288.391869802119  4.58713563028583
#> 10                FALSE          0.75 674.910719445378 0.565170411697046
#> 11                FALSE          0.75 321.261639815129  9.72452391569747
#> 12                FALSE          0.75 333.434128118214   3.7055728644806
#> 13                FALSE          0.75 739.190289100122  0.76234696780517
#> 14                FALSE          0.75 344.369969585224  10.0859153365848
#> 15                FALSE          0.75 320.287565971515  3.59139389037079
#> 16                FALSE          0.75 786.575881035983 0.653045966103129
#> 17                FALSE          0.75 332.879205473308  7.98456947781133
#> 18                FALSE          0.75 350.916825015749  5.01116292985805
#> 19                FALSE          0.75 781.140656932682 0.608250957858685
#> 20                FALSE          0.75 348.325904638956  9.53817590031042
#> 21                FALSE          0.75 339.566158312275   5.3110228296927
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3420.94356870052 2132.78628605557 82254.3514776643 1.94547474647758
#> 2  1451.04882893675 257.378866812427 41862.5171317414 1.34925242334292
#> 3  2195.82726632585  680.36343419983 66223.5101445792 1.40342440491586
#> 4  7557.41551000067 6022.79844612012 149240.232252833 3.25108711720549
#> 5  2364.42182537778 1099.69934402715 57548.4523528384 2.85345962436803
#> 6  4818.01568030921 3234.69090202529 131434.869434174 2.29063241797205
#> 7  17086.4525973407 15913.0051431809 365192.507839433 15.9735578997743
#> 8  5200.52144414846 3783.75075739715 119959.693886952 12.1257375741282
#> 9  7765.88711249272 6294.76429459048 221474.077517464 7.50082492819531
#> 10 20097.8624096191 18744.5814778589 484283.362701157 27.7730802217493
#> 11 6788.44031445744 5599.07794886192 167382.296725302 12.8092793424012
#> 12 9195.76585985402 7684.21655757721 265227.484273074 5.27355225945346
#> 13 20431.6207588658 19444.1975784003 489082.688826551 25.4434721730232
#> 14 6983.92305843603 5595.57609072366 177676.020781049 16.1590764225371
#> 15 8996.25765064177 7331.19698182625  264185.82803829  7.3000264572249
#> 16 20677.0135528208 19291.4345349859 514726.591612512 26.5228824176766
#> 17 6896.86867579199  6037.3332133728 175759.207242346 10.5275944129813
#> 18 9320.00660075107  7664.8359678784 269323.481443668 5.70724645214066
#> 19 19779.3143929227 18647.1130985138 506806.948263746  22.207958851435
#> 20 7157.34411036296 5820.90366127447 180763.894824975 12.3674530374282
#> 21 9255.44885856022 7463.20109761869   268130.0590633 7.00205499392147
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.594192157145654 13.5718598387618 58.8771609525715 694.316003623425
#> 2   2.77532177436143 22.0863387119494  79.924983025421 972.381784375961
#> 3   1.95695289011732 4.92291067389186 122.318690117985 920.413406592474
#> 4  0.268860413885932 35.1656590420901 64.6472537981001 2118.96361322532
#> 5   2.67642690638403 26.4584258029363 134.832414199987  1115.3027332517
#> 6  0.467933501156178  11.906193581834  99.693653262556 1451.39650997359
#> 7  0.574602835265161  85.709419386375 126.649558170609 4722.90347856658
#> 8  0.811564285805302 27.2649884012565 132.157849065488 1666.96957442622
#> 9  0.413790186175156 19.4836972097409 54.8116153161219  3075.0619087762
#> 10 0.389859509532279 94.6869561971533 282.704154879104 10742.5734561029
#> 11 0.790832637878992  99.691004154831 258.161393171707 3660.17563165764
#> 12 0.350853497784569  8.7352537633512 103.789699411211 4559.58921760355
#> 13  0.41228619820056  104.59923846433 230.949314938769 10911.1485357019
#> 14 0.592023960182281  30.139142633372 153.865413123478 2105.57736412077
#> 15 0.735252982298011 16.7769800616245 204.661915202755 5227.99695990136
#> 16 0.557797591740089  78.750752448699 361.641507976445 7834.81260009694
#> 17  0.60518761978567 58.3490908168326  150.73112381669 2755.82790395226
#> 18 0.373594284691044 10.3943015314294 72.6827860145275  5254.9961115377
#> 19 0.287133799513844 82.0443584143115 299.255541841343  4008.0549953667
#> 20 0.311773793164657 75.5996475633427 113.627610425351 3282.57357791528
#> 21 0.699141997063344 10.8958289667977 214.131986858746 4062.28521043459
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.6936171239178  95.094229039829    14.5149492299132
#> 2             TRUE 30.0077237211915 34.4467641939897    26.9870957600431
#> 3             TRUE 62.5243260164729 67.1415923086461    17.0913395734997
#> 4             TRUE 237.075827999177 247.771904614783    14.5948764675141
#> 5             TRUE 59.9270038863802  69.314886050551    18.3727355160134
#> 6             TRUE 201.930083400033 209.466264055161    11.5865015429122
#> 7             TRUE 481.422036023431 533.975041513689    6.04214911985285
#> 8             TRUE 180.405539104884 220.299215723765    12.0344865302141
#> 9             TRUE 276.053012795238    300.730726809     3.9064507740277
#> 10            TRUE 629.224002480601 720.597436410156 -0.0761484814835529
#> 11            TRUE 300.190375296879 342.332904333379    8.42360422638653
#> 12            TRUE 324.759134651413 342.109121585015    3.12841886062498
#> 13            TRUE 697.335777375499 781.044800824745  0.0841361717652495
#> 14            TRUE  317.78828887015 370.951650300297    9.11203592208493
#> 15            TRUE  308.27902244938  332.29610949365    2.38190273449056
#> 16            TRUE 742.945739458905 830.206022613061  -0.264531072309318
#> 17            TRUE 315.561312663954 350.197098282662     6.9890358432639
#> 18            TRUE 341.528404601978  360.30524542952    4.39660033154129
#> 19            TRUE 744.608564622071 817.672749243293   0.135915857658412
#> 20            TRUE 327.981444392387 368.670364885525    9.02530801055456
#> 21            TRUE 328.047777847275 351.084538777276     4.1609342445235
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4698414269224 3398.61785926575 3443.26927813528    81112.2016517038
#> 2   36.1179043976922  1414.7168017556 1487.38085611791    40262.9490964429
#> 3   23.5297145819856 2187.72907826729  2203.9254543844    64709.4300907346
#> 4   15.4794272291988 7499.56800087643  7615.2630191249    145754.537109078
#> 5   27.1781800380168 2320.89771493195 2407.94593582361    55713.7793566393
#> 6    13.126002761716 4798.42999186709 4837.60136875133    129047.322175268
#> 7   7.93259244787523 16945.4606024501 17227.4445922313    357423.331617191
#> 8   14.7045330305136 5155.67053822839 5245.37235006852    117217.528937021
#> 9   5.26782048654396  7733.8364305827 7797.93779440275    216415.600677527
#> 10  1.20648930487764 19942.1023666748 20253.6224525635    466611.829365868
#> 11  11.0254436050084 6624.44861262274 6952.43201629213    161361.307811226
#> 12  4.28272686833622 9181.39636741331 9210.13535229473    257726.960010116
#> 13  1.44055776384509  20259.555011592 20603.6865061396    471133.849485321
#> 14  11.0597947510846 6934.34416880413 7033.50194806793     174212.34601707
#> 15  4.80088504625102  8968.6595184404 9023.85578284314    255585.773039252
#> 16  1.57062300451558 20547.4685650427 20806.5585405989    501838.324885353
#> 17  8.98010311235875  6800.8844213983 6992.85293018568    171225.870340344
#> 18  5.62572552817482 9302.90797473187 9337.10522677027    260679.012840189
#> 19  1.08058605805896 19644.3514233312 19914.2773625143    500213.697796368
#> 20  10.0510437900663 7032.98269012126 7281.70553060465    175364.061289304
#> 21   6.4611114148619 9237.52521990983  9273.3724972106    261447.599892135
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83396.5013036248 88.0807925787774 95.7070535849695    14.3277787004123
#> 2     43462.0851670398 29.5827092078385 34.8717787073427    26.1128694011193
#> 3     67737.5901984238 62.0822473289244 67.5836709961946    16.4748994131127
#> 4     152725.927396589 236.051735557257 248.795997056703      14.51018543714
#> 5     59383.1253490374 59.0281641047043 70.2137258322269    17.5296610405024
#> 6     133822.416693081 201.208534188372 210.187813266822     11.439102490048
#> 7     372961.684061675 476.390365285002 539.006712252117    5.86114922674433
#> 8     122701.858836883 176.585931769033 224.118823059616    11.7788437801855
#> 9     226532.554357401 273.690252942856 303.093486661382    3.77610686538253
#> 10    501954.896036446  620.47548221075 729.345956680007  -0.198954226986221
#> 11    173403.285639379 296.155452304022 346.367827326235    8.17449194545465
#> 12    272728.008536032 323.097965689686 343.770290546743    3.01790000882284
#> 13    507031.528167781 689.321083640997 789.059494559247 -0.0457339806679269
#> 14    181139.695545027 312.698179797051 376.041759373396    8.92554837462751
#> 15    272785.883037328 305.979514115354 334.595617827676    2.15029804506669
#> 16    527614.858339672 734.591031497337 838.560730574629  -0.440237313707446
#> 17    180292.544144347 312.245120423865 353.513290522751    6.79840174303141
#> 18    277967.950047148 339.730621969553 362.103028061945    4.27891813186361
#> 19    513400.198731125 737.613057583869 824.668256281495  0.0454687108115508
#> 20    186163.728360645 324.085696685597 372.566112592315    8.92709926570769
#> 21    274812.518234465 325.842130524189 353.290186100361    3.94070451544855
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6570119564233 3394.34272341654 3447.54441398449    80893.4921105624
#> 2   36.9921307566161 1407.75960506133 1494.33805281217    39956.6488343645
#> 3   24.1461547423726 2186.17836140502 2205.47617124667    64419.4998676579
#> 4   15.5641182595729 7488.49081827817 7626.34020172316    145087.063570912
#> 5   28.0212545135278 2312.56331080403 2416.28033995154    55362.4589956651
#> 6   13.2734018145802 4794.67954088882 4841.35181972961    128590.132274626
#> 7   8.11359234098376 16918.4621353434  17254.443059338    355935.617021443
#> 8   14.9601757805423   5147.082066882 5253.96082141492    116692.433521077
#> 9   5.39816439518914 7727.69906596163 7804.07515902381    215446.956176262
#> 10  1.32929505038031 19912.2759754727 20283.4488437656    463227.918727196
#> 11  11.2745558859403 6593.04594631397 6983.83468260091    160208.352487253
#> 12  4.39324572013836 9178.64476247785 9212.88695723019    256290.689406571
#> 13  1.57042791627827 20226.6062514757 20636.6352662559    467696.837696575
#> 14  11.2462822985421 6924.85033887462 7042.99577799744    173549.089147372
#> 15  5.03248973567489 8963.37476972099 9029.14053156255    253938.953996883
#> 16   1.7463292459137 20522.6620780213 20831.3650276202    499370.358916322
#> 17  9.17073721259124   6782.504457791 7011.23289379298    170357.784550599
#> 18   5.7434077278525 9299.63376974947 9340.37943175267    259023.689065054
#> 19  1.17103320490582 19618.5074504307 19940.1213354148    498951.160472828
#> 20  10.1492525349132 7009.16880113881 7305.51941958711    174330.050612261
#> 21  6.68134114393686 9234.09303378529 9276.80468333514    260167.980050848
#>    ci95.integral.bt.up
#> 1     83615.2108447662
#> 2     43768.3854291183
#> 3     68027.5204215004
#> 4     153393.400934755
#> 5     59734.4457100117
#> 6     134279.606593722
#> 7     374449.398657424
#> 8     123226.954252827
#> 9     227501.198858665
#> 10    505338.806675119
#> 11    174556.240963351
#> 12    274164.279139577
#> 13    510468.539956527
#> 14    181802.952414725
#> 15    274432.702079696
#> 16    530082.824308702
#> 17    181160.629934092
#> 18    279623.273822282
#> 19    514662.736054665
#> 20    187197.739037688
#> 21    276092.138075752