Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8303250934476  15.4685700963394
#> 2                 FALSE          0.75 32.3685764963975  31.2967628744644
#> 3                 FALSE          0.75 64.8298806623752  19.9110870427424
#> 4                 FALSE          0.75 242.782283074191  15.0787356277786
#> 5                 FALSE          0.75    65.3599549722  23.5754783289454
#> 6                 FALSE          0.75 205.344453211019  12.3388838564171
#> 7                 FALSE          0.75 507.672513620782  7.02525507005159
#> 8                 FALSE          0.75 201.249758664549   13.159435151171
#> 9                 FALSE          0.75 290.835169098333   4.7046244612353
#> 10                FALSE          0.75  660.90975088275 0.623665686869775
#> 11                FALSE          0.75 327.495717159235  10.2741302562244
#> 12                FALSE          0.75 334.483465066749  3.80300905218277
#> 13                FALSE          0.75 727.715640040042 0.692261460410552
#> 14                FALSE          0.75 346.630697091829  10.0650471326718
#> 15                FALSE          0.75 320.448845989642  3.45269979014513
#> 16                FALSE          0.75 777.285690372658 0.464134271320807
#> 17                FALSE          0.75 336.287239420093  8.03771994285374
#> 18                FALSE          0.75 350.667906878335  5.03620677237977
#> 19                FALSE          0.75 771.427441056199 0.586465661471232
#> 20                FALSE          0.75 346.281857802616  9.42006864035372
#> 21                FALSE          0.75 336.183571850488  5.33368277919971
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3417.39602022193 2139.60653279588 81841.0471341092 2.47963307732526
#> 2  1455.15825524341 276.756442754527 42324.4560585645 2.26234519916754
#> 3  2197.95786115839 704.191769857665 65677.4850991838 1.82202435908446
#> 4  7564.31339506046 6026.17346404665 149308.835442714 2.89135803655495
#> 5  2370.10380639049  1091.3513753523 57089.8723994367 2.17499016427613
#> 6  4816.48930119886 3237.79097398608  131070.37901403 2.26286938442885
#> 7  17032.1151572834 15824.7393163823 361584.594396032 13.9791942138003
#> 8  5178.88616099673 3809.91974853785 118662.465023556 8.54753988379841
#> 9  7753.27011507109 6269.45762639792 221828.698424448 4.69350983794212
#> 10 20074.8434307594 18631.6405098739 484450.234920338 23.4294730599267
#> 11 6791.19266179692 5485.56073524439 167832.014537016 13.5229775894607
#> 12 9187.18281599135 7663.05664860765 266728.144564539  5.3042284151631
#> 13 20432.6533571959 19432.0521491934 492747.385381094 22.6976787352958
#> 14  6998.4574226702 5617.19875450161 178173.547193042 10.9337405509916
#> 15 9006.16237734725 7389.88633599592 264714.720245045 5.34178893135051
#> 16 20680.1533152741 19386.3962321084 514071.420668407 24.1558125828118
#> 17 6896.20581877645 6032.13187963056 175723.485741663  14.984291063428
#> 18 9321.91613028864 7637.47986756027 271537.904119995 6.77567682122422
#> 19 19804.9115192956 18621.2554914138 500760.814251763 24.5032928479005
#> 20 7171.51870736686 5861.16058731433 181341.374353954 14.6851073178478
#> 21 9252.95638959088 7394.82826245994 268246.461948293 7.56171022652978
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.497425999249654 18.1328120737197 36.0475460686912 1190.36295740456
#> 2   1.96123466571639 23.1700521840791 59.5130553455388 1112.01810630094
#> 3   1.80554361351676 4.41001187571064 109.379895509822 1167.39469708064
#> 4  0.223262236386366  50.596725429038 75.5792500374219 2624.16294240961
#> 5   2.66076511440841 30.6680035393409 138.603749859963 1345.47818222069
#> 6  0.558834842511365  8.6541109461114 109.395014608509 2700.41726936118
#> 7  0.480780392590217 154.948062031737 202.029690906659 8424.43774830779
#> 8  0.874237758236605 49.1624366612237 158.870254577848 2683.93699765495
#> 9  0.253718406284633 24.3551423215998 68.6932980770141 2611.24214140567
#> 10 0.816713717261664 89.4166367210986 547.788243843805 7407.26841853513
#> 11  0.78729282584029 64.0610130513956 226.874599827688 1946.24175776848
#> 12 0.376925024197257 10.1573218325124 100.901947810843 3304.53620446872
#> 13  0.30243444361169 85.1665492153775 220.348301026489 7880.41820414491
#> 14 0.545785508464913 41.0741310013554 172.514787995738 2462.27291839808
#> 15 0.360018349420479 16.4112684309032 89.1668675019784 4230.59569629808
#> 16 0.224872775144958 125.181040455149 198.007624245766 12631.6855063173
#> 17 0.464512396741566 86.2848093161303 138.832654202313 3412.34967214878
#> 18 0.318710937794875 10.0680563792438 86.9645969884702 4160.63453661096
#> 19 0.263479576394878 73.0315544522796 154.094413172007 11923.0027185492
#> 20 0.388922473398064 40.0500640252562 96.0039790141115 2071.62813812935
#> 21 0.681721968106111 14.7835745945833 174.861155062841 4965.36382457628
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.7513286812476 96.9093215056477   14.6503043275737
#> 2             TRUE 28.6470186437669 36.0901343490281    28.070531849361
#> 3             TRUE 61.8326505916813 67.8271107330691   16.9409677985073
#> 4             TRUE 238.025999104058 247.538567044324    14.711469248923
#> 5             TRUE 61.7820961519658 68.9378137924343   19.1985197157435
#> 6             TRUE 201.622033073633 209.066873348404    11.419600540486
#> 7             TRUE 484.676739139081 530.668288102484   6.23437132424069
#> 8             TRUE 187.189055555701 215.310461773397   11.7213140388718
#> 9             TRUE 283.114345414918 298.555992781748   4.28725768289708
#> 10            TRUE  622.36826769917 699.451234066329 -0.719828378025662
#> 11            TRUE 305.250419024572 349.741015293898   8.97903355771716
#> 12            TRUE 325.758009323806 343.208920809693   3.18296738737829
#> 13            TRUE  690.37795852048 765.053321559603  0.194756800669322
#> 14            TRUE 328.644693885448  364.61670029821   9.16722997124706
#> 15            TRUE 311.661603197571 329.236088781714   2.86046960534844
#> 16            TRUE 737.549378673932 817.022002071383 0.0942185562073513
#> 17            TRUE 311.638080620754 360.936398219432   7.27359705021387
#> 18            TRUE 339.521918507421 361.813895249249    4.5119272797072
#> 19            TRUE 731.119524321403 811.735357790996  0.153041758301658
#> 20            TRUE 322.124856264756 370.438859340476    8.7802911716139
#> 21            TRUE 323.744558527846 348.622585173129   4.21225014166516
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2868358651051 3387.56754436066  3447.2244960832    79882.9000691786
#> 2   34.5229938995679  1417.0435194006 1493.27299108622    40495.1862736995
#> 3   22.8812062869775 2190.70339162285 2205.21233069394    63757.1208224862
#> 4   15.4460020066341 7481.08178172969 7647.54500839123     144992.08740245
#> 5   27.9524369421472 2319.65494056828 2420.55267221271    54876.5607896837
#> 6   13.2581671723483 4802.25328869251 4830.72531370521     126628.19260593
#> 7    7.8161388158625 16777.2255952412 17287.0047193256    347726.394300065
#> 8   14.5975562634702 5098.01395268902 5259.75836930445    114247.388662414
#> 9   5.12199123957352 7713.20590595205 7793.33432419012    217533.205101835
#> 10  1.96715975176521 19927.7530633532 20221.9337981656    472265.278371847
#> 11  11.5692269547317 6685.81229532737 6896.57302826646    164630.446845487
#> 12  4.42305071698726 9170.47402157686 9203.89161040583    261292.182508188
#> 13  1.18976612015178 20292.5543837366 20572.7523306552    479784.097435276
#> 14  10.9628642940966 6930.89047717297 7066.02436816742    174123.108242277
#> 15  4.04492997494181 8979.16584077841 9033.15891391608    257755.390324635
#> 16 0.834049986434262 20474.2305037253 20886.0761268228    493292.298010515
#> 17  8.80184283549362 6754.26730745142 7038.14433010149    170110.170530978
#> 18  5.56048626505234 9305.35417754479  9338.4780830325     264693.66030727
#> 19  1.01988956464081 19684.7746122216 19925.0484263696     481147.47477975
#> 20  10.0598461090935 7105.63635204531  7237.4010626884    177933.546066731
#> 21  6.45511541673427 9228.63740938279 9277.27536979897    260078.438456865
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83799.1941990397 87.9702442618901 97.6904059250051   14.4936151378101
#> 2     44153.7258434296 27.9343799060292 36.8027730867659   27.4527429296603
#> 3     67597.8493758815 61.2587129185697 68.4010484061807   16.3722215602496
#> 4     153625.583482977 237.115221322544 248.449344825839   14.6411416444613
#> 5     59303.1840091898 61.0969742502188 69.6229356941812   18.3603787047049
#> 6     135512.565422129 200.909229217538 209.779677204499   11.2435675650949
#> 7     375442.794491998 480.273292961734 535.071734279831   6.08292550057477
#> 8     123077.541384698 184.496580492304 218.002936836794   11.4459291450272
#> 9      226124.19174706 281.635889815966 300.034448380699   4.20733638491742
#> 10    496635.191468828 614.987983685294 706.831518080206 -0.977093198963086
#> 11    171033.582228545 300.990681083892 354.000753234578   8.73103631757747
#> 12    272164.106620891  324.08717737303 344.879752760469   3.06423600475615
#> 13    505710.673326913 683.228189718862 772.203090361222 0.0994899509316401
#> 14    182223.986143806 325.200565611885 368.060828571772   8.99530753608062
#> 15    271674.050165456 309.978939684195 330.918752295089   2.74706382528099
#> 16    534850.543326299 729.940297710347 824.631083034969 0.0233836320366896
#> 17    181336.800952348 306.918028935774 365.656449904412   7.12727564524028
#> 18     278382.14793272 337.387580308736 363.948233447935   4.41153333430182
#> 19    520374.153723777 723.400987074314 819.453895038085 0.0700456917372717
#> 20    184749.202641176 317.499047459634 375.064668145598   8.65778059249351
#> 21    276414.485439721 321.362619806489 351.004523894486   3.99750772171174
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4435250548687 3381.85570855744 3452.93633188642    79507.9357375962
#> 2   35.1407828192686 1409.74495296262 1500.57155752421    40144.9005702147
#> 3   23.4499525252352   2189.314237882 2206.60148443478    63389.3914929058
#> 4   15.5163296110959 7465.14381321955 7663.48297690138    144165.476075591
#> 5   28.7905779531859 2309.99451945338  2430.2130933276    54452.7351622842
#> 6   13.4342001477394 4799.52724374448 4833.45135865324    125777.561166082
#> 7   7.96758463952842 16728.4169557012 17335.8133588656    345072.696409348
#> 8   14.8729411573147 5082.52778514073 5275.24453685273    113401.948508152
#> 9   5.20191253755318 7705.53403612075 7801.00619402142    216710.663827292
#> 10  2.22442457270264 19899.5868227861 20250.1000387328    469931.988820009
#> 11  11.8172241948714 6665.63307621618 6916.75224737765     164017.38069179
#> 12   4.5417820996094 9167.27446519962 9207.09116678307    260251.253603781
#> 13  1.28503296988946 20265.7269207337  20599.579793658     477301.76570097
#> 14  11.1347867292631 6917.95212590754 7078.96271943285    173347.492272981
#> 15  4.15833575500927 8973.99629122268 9038.32846347182    256422.752680301
#> 16 0.904884910604924  20434.798475982 20925.5081545661    489313.317076025
#> 17  8.94816424046721 6727.08759251684 7065.32404503607    169035.280384251
#> 18  5.66088021045772 9302.18273978533 9341.64952079196    263383.060428238
#> 19  1.10288563120519 19661.7696725691 19948.0533660221    477391.728923407
#> 20  10.1823566882139 7093.02058187735 7250.01683285636     177280.98320322
#> 21  6.66985783668769  9223.9805833855 9281.93219579626    258514.348852124
#>    ci95.integral.bt.up
#> 1     84174.1585306221
#> 2     44504.0115469144
#> 3     67965.5787054619
#> 4     154452.194809836
#> 5     59727.0096365893
#> 6     136363.196861977
#> 7     378096.492382715
#> 8      123922.98153896
#> 9     226946.733021603
#> 10    498968.481020667
#> 11    171646.648382242
#> 12    273205.035525298
#> 13    508193.005061218
#> 14    182999.602113102
#> 15     273006.68780979
#> 16    538829.524260789
#> 17    182411.691099075
#> 18    279692.747811753
#> 19     524129.89958012
#> 20    185401.765504687
#> 21    277978.575044463