Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.1590694329444   15.602124372729
#> 2                 FALSE          0.75  31.940439518593    31.45215752039
#> 3                 FALSE          0.75 65.0375149818132  20.4539411788005
#> 4                 FALSE          0.75 241.737262072449  14.9554648521025
#> 5                 FALSE          0.75  65.492949671137  22.5179069143675
#> 6                 FALSE          0.75 204.974071624406  12.3480101192292
#> 7                 FALSE          0.75  506.00514432932  6.96827720272766
#> 8                 FALSE          0.75 206.104347964657  13.3995042003188
#> 9                 FALSE          0.75  290.02055370339  4.68926029020997
#> 10                FALSE          0.75 661.590357918583  0.40463383284764
#> 11                FALSE          0.75 327.678151865371  9.85899210258669
#> 12                FALSE          0.75 332.558414760669  3.70758905416476
#> 13                FALSE          0.75 733.007821053809 0.797897807818719
#> 14                FALSE          0.75  345.86393683167  9.71841877139993
#> 15                FALSE          0.75 318.936565881761  3.40200357479603
#> 16                FALSE          0.75 777.175397928912   0.4273329365524
#> 17                FALSE          0.75 338.801969040367  8.12630567336833
#> 18                FALSE          0.75 351.250789423081  4.99901128271502
#> 19                FALSE          0.75 773.486964141471 0.617852004994429
#> 20                FALSE          0.75 353.184518075692  9.58399076050124
#> 21                FALSE          0.75 333.866055195165  4.88502965869672
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3414.03890441717  2112.4319142056 81587.9481035389 2.18377568318609
#> 2  1458.40192303369 266.667677240839 42433.4180936984 2.06573531090038
#> 3  2194.71383768492 663.776698590418 65779.5864743899 1.41025876884048
#> 4  7556.61951907068 6038.83407896655 148806.454902929 2.80313820635927
#> 5  2351.99769011567 1084.35767660397  57068.711052577 3.00156066098334
#> 6  4820.38616419921 3239.69498239036  130417.36089572 2.93617675498369
#> 7  17095.8702530834 15900.3517710883 364887.014749454 14.1656442770909
#> 8  5161.01463688827 3769.58490762051 117887.944883343  9.9609721273716
#> 9  7771.25329513381 6295.03834284309 221396.002655543 6.40774398215116
#> 10 20063.9542359127 18781.4001574373 482718.648052271 26.0925364169778
#> 11 6781.53509920931 5603.47702707852 167239.805096903 10.2946756403084
#> 12 9185.95052066703  7668.8352110914 266200.093644228 5.83228339186168
#> 13   20431.70137203 19383.3686700544 490379.426685463 22.0390654363737
#> 14 6968.04501582103 5690.28915986441 177307.334797388 16.0573415558451
#> 15 9004.42645098725 7387.72401666629 263531.372700483 4.97459847222557
#> 16 20677.6664726639 19441.6602316349 514601.512461678 17.7717858187293
#> 17 6880.08371309136 6008.06412158997 175380.939731582 10.2990918501365
#> 18  9319.9792565101 7655.47814129155 270129.462274107 5.61980304579158
#> 19 19817.8797901057 18663.4320817639   504673.2573539 26.1927877471229
#> 20 7131.32913896488 5810.27947031844 180257.387407792 13.8850088608286
#> 21 9257.00412413785 7517.86322717169 269144.978868002  6.8912401958268
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.84917043140128  31.663278824461 78.3429995170675 1702.28070019234
#> 2   1.79686626071576 21.2138616905322 54.6340270336994 606.336259395984
#> 3   1.85207750750354 7.17427017172948 115.608667974169  1692.2766635478
#> 4  0.274545852069107 60.7583881034642 85.8887413550566 3143.32373960431
#> 5   2.48398813470865 32.3372564303302 151.650885474543 1132.00399296702
#> 6  0.457133340533587 11.6441901395599 88.9489470221299 1804.83833868451
#> 7  0.453424828638096 75.3906682133206 104.792773011143 4152.66989493516
#> 8  0.587520316068801 65.5153604216526 110.200754204716 3231.75668799999
#> 9  0.330053909287109 16.4341125725721 68.7593178156431 3940.18154761545
#> 10 0.203849530944873 92.5445011937554  211.89860331093 9881.76491956716
#> 11 0.455969183790387 93.8083172071517 164.618828566489 3695.66420599838
#> 12 0.302070661257922 16.0019565693402 77.7226483495127 4076.05469913968
#> 13 0.285059927804652 89.6851134911563 154.956819421094 9834.76479166973
#> 14 0.538937661203767 32.6442734858752 184.489559302743 2141.59331714644
#> 15 0.718288123024039 13.8595574774658  191.44900581288 4544.77071852575
#> 16 0.271218911007313 98.3641041318787 213.360704439791 9164.48752880198
#> 17 0.463596387400923 98.9172825130989 148.557863065187 4051.27995015947
#> 18 0.371482356063474 7.59467624782903 89.5843510014508  5599.7083894458
#> 19 0.365141340282838 93.8518512708775 268.741933467999  6346.0970873313
#> 20 0.371412226735881 87.9807691246282 105.284286900131  3195.1579611385
#> 21  0.43136568576762 8.41598350915802 99.5275402362731 3428.17437420182
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.5667584341032 95.7513804317855    14.2052390130739
#> 2             TRUE 28.5423049321619 35.3385741050241    28.4963125215126
#> 3             TRUE 62.7176393070706 67.3573906565558    17.4072736789571
#> 4             TRUE 237.126099722988  246.34842442191    14.5038369254488
#> 5             TRUE 60.5553823838194 70.4305169584546    18.4317464327718
#> 6             TRUE 200.144060862458 209.804082386354    11.5960257740515
#> 7             TRUE 482.702659493505 529.307629165134    6.22239335961799
#> 8             TRUE  189.71854881513 222.490147114183    12.4330332803856
#> 9             TRUE 279.479814852751 300.561292554028    4.14632160943267
#> 10            TRUE 618.668135512655 704.512580324511  0.0693013544433247
#> 11            TRUE 310.743410437063 344.612893293678     9.1089227952515
#> 12            TRUE 322.964308581057 342.152520940281    3.21068281639547
#> 13            TRUE 696.753558410974 769.262083696644   0.328974226580066
#> 14            TRUE 319.449609972305 372.278263691035    8.83186631871973
#> 15            TRUE  310.75335139495 327.119780368572    2.22041961242149
#> 16            TRUE 747.940810257102 806.409985600722 -0.0188221720546302
#> 17            TRUE 321.859962946892 355.743975133841    7.36368961609381
#> 18            TRUE 342.006213412754 360.495365433408     4.3879228069906
#> 19            TRUE 730.399828297454 816.574099985488  0.0171945002291601
#> 20            TRUE 330.343678499628 376.025357651755    8.97301764752072
#> 21            TRUE  322.52996507303   345.2021453173    4.17543310560899
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.9990097323841 3361.95281075093 3466.12499808341    78787.6963517226
#> 2   34.4080025192674 1423.50512055277 1493.29872551462     41435.994946992
#> 3   23.5006086786438 2182.91216325243 2206.51551211742    62995.7913628537
#> 4   15.4070927787562 7456.67197064048 7656.56706750088     143635.68735128
#> 5   26.6040673959632 2298.80290328777 2405.19247694356    55206.5644841463
#> 6    13.099994464407 4801.23147141963 4839.54085697879    127448.401828584
#> 7   7.71416104583733 16971.8526038725 17219.8879022943    358055.872772285
#> 8   14.3659751202519 5053.24186899465 5268.78740478189    112571.705131583
#> 9   5.23219897098726 7744.21917995193 7798.28741031569    214914.404009715
#> 10 0.739966311251956 19911.7185314489 20216.1899403764    466463.144759583
#> 11  10.6090614099219 6627.22041740355 6935.84978101508    161160.437478036
#> 12  4.20449529193404 9159.62730211047  9212.2737392236    259494.983664144
#> 13  1.26682138905737  20284.169360337 20579.2333837229    474201.238603166
#> 14  10.6049712240801 6914.34518593676 7021.74484570529    173784.413790683
#> 15  4.58358753717057 8981.62747893682 9027.22542303769    256055.224868509
#> 16  0.87348804515943 20515.8575213669 20839.4754239608    499525.930476799
#> 17  8.88892173064285 6717.36478335732 7042.80264282541    168716.584213569
#> 18  5.61009975843943 9307.48601408242 9332.47249893778    260917.941973469
#> 19   1.2185095097597 19663.4934947651 19972.2660854463     494233.92764524
#> 20  10.1949638734818 6986.60077375487  7276.0575041749     175001.35256172
#> 21  5.59462621178446 9243.15983126529 9270.84841701042     263505.63202244
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84388.1998553553 87.8788690938996 96.4392697719891    13.9377503271825
#> 2     43430.8412404048 27.8915983092283 35.9892807279577    27.9302996493871
#> 3      68563.381585926 62.2734077948859 67.8016221687406    16.8238692640935
#> 4     153977.222454578 236.243111187985 247.231412956913    14.4173549820471
#> 5     58930.8576210078 59.6098907756097 71.3760085666644    17.6492901703386
#> 6     133386.319962856 199.219165184638 210.728978064174    11.4520287717834
#> 7     371718.156726622 478.240481546221 533.769807112418    6.07956453859699
#> 8     123204.184635103 186.580842595008 225.627853334305    12.2479643808239
#> 9      227877.60130137 277.461375498373 302.579731908406    4.04235462800723
#> 10    498974.151344959 610.448986541307 712.731729295859 0.00508875219568988
#> 11    173319.172715771 307.500587610366 347.855716120375    8.96529250235753
#> 12    272905.203624313  321.12713931262 343.989690208718    3.11553055809923
#> 13    506557.614767759 689.811252798516 776.204389309101   0.239180349321601
#> 14    180830.255804094 314.391547382214 377.336326281126    8.66210095544054
#> 15    271007.520532458 309.186352876199 328.686778887323    1.99415885366891
#> 16    529677.094446557 742.342697724203 812.008098133622  -0.104256129021934
#> 17    182045.295249594 318.615749014099 358.988189066634    7.21765675406252
#> 18    279340.982574745 340.235975453329 362.265603392832    4.27090586483061
#> 19     515112.58706256 722.149100157111 824.824828125832  -0.097825021959934
#> 20    185513.422253865 325.969900708467 380.399135442916    8.85602279609891
#> 21    274784.325713564 320.359224411344 347.372885978985    4.03955291459219
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.2664984182755 3351.97887792122 3476.09893091311     78251.477931162
#> 2   34.9740153913929 1416.82275412025 1499.98109194713    41244.9990252823
#> 3   24.0840130935074 2180.65226814833 2208.77540722151    62462.7242138362
#> 4    15.493574722158 7437.53307838789 7675.70595975347    142645.540373304
#> 5   27.3865236583965 2288.61666751222 2415.37871271911    54849.9832263616
#> 6   13.2439914666751 4797.56355152567 4843.20877687275    126879.877751898
#> 7   7.85698986685833 16948.1045433853 17243.6359627815    356747.781755381
#> 8   14.5510440198136 5032.60453046183 5289.42474331471    111553.701774863
#> 9    5.3361659524127 7739.04243449157 7803.46415577605    213673.246822216
#> 10  0.80417891349959 19882.5670135729 20245.3414582524     463350.38880992
#> 11  10.7526917028158  6597.6707974833 6965.39940093533    159996.303253146
#> 12  4.29964755023028 9154.58668579113 9217.31435554294    258211.026433915
#> 13  1.35661526631584 20255.9185495873 20607.4841944726     471103.28769379
#> 14  10.7747365873593 6904.06223978871 7032.02779185334    173109.811895781
#> 15  4.80984829592315 8977.26171833142 9031.59118364309    254623.622092173
#> 16 0.958922002126734 20484.8728285654 20870.4601167623    496639.116905226
#> 17  9.03495459267414 6686.20583936569 7073.96158681704    167440.431029269
#> 18  5.72711670059943 9305.09369106435 9334.86482195584    259154.033830793
#> 19  1.33352903194879 19633.9301616147 20001.8294185966    492234.907062731
#> 20  10.3119587249036 6958.88683148061 7303.77144644915    173994.877803961
#> 21  5.73050640280126 9240.50879645991  9273.4994518158    262425.757094566
#>    ci95.integral.bt.up
#> 1     84924.4182759159
#> 2     43621.8371621146
#> 3     69096.4487349436
#> 4     154967.369432553
#> 5     59287.4388787924
#> 6     133954.844039541
#> 7     373026.247743527
#> 8     124222.187991823
#> 9     229118.758488869
#> 10    502086.907294623
#> 11     174483.30694066
#> 12    274189.160854542
#> 13    509655.565677135
#> 14    181504.857698996
#> 15    272439.123308794
#> 16     532563.90801813
#> 17    183321.448433894
#> 18    281104.890717421
#> 19    517111.607645069
#> 20    186519.897011624
#> 21    275864.200641438