Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.4616255295257  15.4493707193959
#> 2                 FALSE          0.75   33.24395193047  31.5382266906721
#> 3                 FALSE          0.75 64.3924147628775  20.1943405512404
#> 4                 FALSE          0.75 242.260138223854  14.9032091708579
#> 5                 FALSE          0.75 65.5747997374169  22.0067731143869
#> 6                 FALSE          0.75 204.250293486563  12.2638489040074
#> 7                 FALSE          0.75 510.507131236819  7.06085951103583
#> 8                 FALSE          0.75 206.112477279335  13.4513872251961
#> 9                 FALSE          0.75 292.032539034418  4.71512135024482
#> 10                FALSE          0.75  668.02779572601 0.672005529819509
#> 11                FALSE          0.75 326.423226000933  9.74821566977046
#> 12                FALSE          0.75  335.34155098246  3.94776153818528
#> 13                FALSE          0.75 726.700138690217  0.93481992001004
#> 14                FALSE          0.75 342.519778776778  9.79714335830787
#> 15                FALSE          0.75 319.406054611495  3.37235537027276
#> 16                FALSE          0.75 782.961565043889 0.607504679361842
#> 17                FALSE          0.75 336.903383960615  8.10963923645285
#> 18                FALSE          0.75 351.186000701585  5.06477524435959
#> 19                FALSE          0.75 769.720617332503 0.619123911106317
#> 20                FALSE          0.75 351.303722839746   9.5488513830505
#> 21                FALSE          0.75  338.96353503679  5.23193277138631
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.52733774615 2130.14219868837 81496.9224829084 2.62835136910395
#> 2  1462.13676107529 269.543373181933 42735.7351818599 2.53070856080918
#> 3  2197.14897892007 678.099755708464 66122.3005104061 1.69697831490353
#> 4  7568.11716866695 6066.29817916069 149257.130715691 2.33341389580112
#> 5  2364.21914107748 1135.74778738559  57080.738175459 2.79398339346283
#> 6  4816.56915635052 3249.00050049365 130452.472768827 2.11937371263123
#> 7   17056.918392187 15894.2707907165 362866.591578594 13.7546523427911
#> 8  5193.73117747511 3791.20003513094 119129.882601358 7.10005718909485
#> 9  7767.05751733884 6290.11545656591 221853.820271093 6.24272707328593
#> 10 20077.0144610364 18638.8355278372 484764.695042345 22.3878492891294
#> 11 6801.93071457707  5638.0540865368 168558.449285428 12.4714095921541
#> 12 9192.53816999058 7622.58238814572 267166.601766858 6.43985186754955
#> 13  20464.777688613 19291.6797060002 490222.965820442 30.0619876904627
#> 14 6966.81374608182 5654.63779198082 176052.915676317 14.1575952489274
#> 15 9006.79723502226 7405.19709520951 263333.987796928 5.72347752050934
#> 16 20668.1748978904 19293.6676390368 515572.475003712 20.9269886742593
#> 17  6907.3750117427  6027.8546682072  175381.94800111 12.0284924098007
#> 18 9319.60355142123 7628.17202569523 270357.352541949 7.11703065455518
#> 19 19792.8270066845 18600.2191631736 501246.359021842 26.5966422022464
#> 20 7125.97379569533 5801.19576458409 179128.502294334 14.4227417120289
#> 21 9256.56664480473 7468.62354015686 268591.805042401 7.65976627171772
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.684189512805081 29.2198170883293 60.2173868535872 1766.20406417624
#> 2   2.35855320053947 21.5275323790053 52.6003238602196 693.262763065158
#> 3   2.53422444719738 5.51165673714769 161.075759414776 1163.97822803853
#> 4  0.228255775855565 62.5823432973396 81.3170145304034 3329.91471342066
#> 5   2.51994477690163 32.2229782662955 128.799071793895 1207.15299169396
#> 6  0.643441332242392 9.32486676002419   124.1611776837 3548.40244219644
#> 7  0.494836354371471 155.908749695311 238.600877260478 8487.34859429725
#> 8  0.701565520885931 40.8435431310197  113.40852734644 1984.52576707153
#> 9  0.356499212324723 22.6353952029544 75.6499145285976 2798.49804165832
#> 10 0.365805179187839 72.4246163731868 215.661363503333 10245.5388610402
#> 11 0.611197927050266 68.4540550106985 129.398056065714  2006.4257545167
#> 12 0.410868910032005 11.6883137221057 95.5145609543788 4655.44726676367
#> 13 0.455319100638083 90.0351066103952 277.674684773069 9943.32768808302
#> 14 0.546786278490544 44.8857007320952 153.339075569126 2514.43370068633
#> 15 0.299634477811379 18.0230820184818 41.3729878775241  5220.9740206128
#> 16 0.419830397886265  77.474744909317 301.485542564504  6593.3445676846
#> 17 0.538141415598388 114.936227638841  213.88729001021 5461.68379695543
#> 18 0.505085855699128 10.9943255863998 138.741306990223 4771.89020512154
#> 19 0.281951286012309 65.4527307952084 182.785353659294 13804.6815723824
#> 20 0.428204671490438 141.886801133336  179.50869056141 6351.01987327674
#> 21 0.445729668159222 14.6875803243451 139.629077356485 3030.87981514129
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.1379875273497 96.7852635317017    14.3238789708315
#> 2             TRUE 29.0809363479389 37.4069675130011    27.6584066757847
#> 3             TRUE 61.6008854348612 67.1839440908938    16.0255413356007
#> 4             TRUE 238.421672365262 246.098604082447    14.5277284195755
#> 5             TRUE 60.9786970551706 70.1709024196633    17.8614639563838
#> 6             TRUE 200.763923729285 207.736663243842    11.2053879124687
#> 7             TRUE 487.880728132928 533.133534340711    6.24685370809476
#> 8             TRUE 194.432883203274 217.792071355396    12.2973119433387
#> 9             TRUE 281.763252998863 302.301825069974    4.12868014597065
#> 10            TRUE 631.199783645392 704.855807806628  0.0702560100555144
#> 11            TRUE 305.907757221839 346.938694780026    8.74279507977277
#> 12            TRUE 324.747994660341 345.935107304579    3.27188218118263
#> 13            TRUE 677.248168939405 776.152108441028   0.185819999460393
#> 14            TRUE 319.230534592292 365.809022961264    8.89767993019092
#> 15            TRUE 309.990934090257 328.821175132733    2.87945665427304
#> 16            TRUE 748.536668674733 817.386461413046 -0.0831163251610635
#> 17            TRUE 317.116513946493 356.690253974738     7.2243966077935
#> 18            TRUE 339.478485274842 362.893516128329    4.23390901173452
#> 19            TRUE 725.969140909808 813.472093755198   0.155314045616068
#> 20            TRUE 327.578312723458 375.029132956033    8.84445469844873
#> 21            TRUE 326.363219519814 351.563850553766    4.49870746726439
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5748624679602 3364.46073863585 3460.59393685646    78591.5167973385
#> 2   35.4180467055596 1426.72397031183 1497.54955183876    41595.3179366177
#> 3     24.36313976688 2188.08230358746 2206.21565425268    64207.5563252827
#> 4   15.2786899221403 7465.16921394283 7671.06512339108    143779.421012114
#> 5   26.1520822723901 2311.21234182943 2417.22594032554    55094.9715041225
#> 6   13.3223098955461 4801.22975053028 4831.90856217076    124615.350751414
#> 7    7.8748653139769 16800.4484989382 17313.3882854358    348904.903140975
#> 8   14.6054625070535 5126.54354902458 5260.91880592563    115865.337714525
#> 9   5.30156255451899 7729.82229222998  7804.2927424477    217250.290992565
#> 10   1.2737550495835 19957.8759671025 20196.1529549703    467910.783615934
#> 11  10.7536362597682 6689.32379408447 6914.53763506967    165257.878919248
#> 12  4.62364089518792 9173.31089391771 9211.76544606344    259508.391013032
#> 13  1.68381984055969 20316.6699382389 20612.8854389871    473866.191773546
#> 14  10.6966067864248 6892.97676837752 7040.65072378612    171916.672238688
#> 15  3.86525408627248 8977.14926510185 9036.44520494266     254745.48553302
#> 16  1.29812568388475 20540.7289425146 20795.6208532662    504726.423189871
#> 17   8.9948818651122 6718.30491727681 7096.44510620859    166397.478155118
#> 18  5.89564147698465  9301.5178858316 9337.68921701086    262507.593154524
#> 19  1.08293377659657 19685.1572645264 19900.4967488426    478537.657835273
#> 20  10.2532480676523 6892.57000783099 7359.37758355967    168681.074602794
#> 21  5.96515807550823 9232.40557517118 9280.72771443828    263606.007746493
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84402.3281684783 87.3100568460819 97.6131942129694    14.1083592742979
#> 2     43876.1524271021  28.283763151284  38.204140709656    26.9154624176148
#> 3     68037.0446955295 61.0663372656666 67.7184922600884    15.2272606347335
#> 4     154734.840419268 237.686646988084 246.833629459625     14.455827850181
#> 5     59066.5048467956 60.0985922862298 71.0510071886041    17.0676813516598
#> 6      136289.59478624 200.096321009806 208.404265963321    11.0027038928123
#> 7     376828.280016213 483.548012644949  537.46624982869    6.09098025646775
#> 8     122394.427488191 192.196365188709 220.028589369961    12.0763188042597
#> 9      226457.34954962 279.796793970778 304.268284098059    4.01638289408836
#> 10    501618.606468757 624.147611119316 711.907980332704 -0.0449726213886548
#> 11    171859.019651608  301.97926320031 350.867188801555    8.55026773275194
#> 12    274824.812520684 322.719441322063 347.963660642857    3.14245847452255
#> 13    506579.739867339  667.77864281691 785.621634563524  0.0423944827593968
#> 14    180189.159113946  314.77089208888 370.268665464676     8.7254422524664
#> 15    271922.490060836 308.188038671297 330.624070551693    2.78507179376246
#> 16    526418.526817554 741.944667242341 823.978462845437  -0.215362900495237
#> 17    184366.417847101 313.327538837406 360.479229083825    7.05488206188001
#> 18    278207.111929374 337.236620618657 365.135380784514     4.0748069671893
#> 19    523955.060208411   717.5911986161 821.850036048906  0.0664993905221909
#> 20    189575.929985874 323.035149084169 379.572296595322    8.70957022692924
#> 21    273577.602338308 323.950393144223 353.976676929357    4.35830262179424
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7903821644938 3355.25649625303 3469.79817923928     78035.162517123
#> 2   36.1609909637295 1419.94279761244 1504.33072453814    41376.9401662522
#> 3   25.1614204677472 2186.34613171526 2207.95182612488    63840.9031834506
#> 4   15.3505904915348 7445.45577580417 7690.77856152974    142730.497877386
#> 5   26.9458648771141 2301.06210367554 2427.37617847942    54714.7183117389
#> 6   13.5249939152025 4798.29241750087 4834.84589520017    123497.603982122
#> 7   8.03073876560391 16751.3372427842 17362.4995415898    346231.388333771
#> 8   14.8264556461325 5113.67783293831  5273.7845220119    115240.212097898
#> 9   5.41385980640127 7722.69214274105 7811.42289193663    216368.764109442
#> 10  1.38898368102767  19935.062212945 20218.9667091278    464683.438874707
#> 11   10.946163606789  6667.7607667561 6936.10066239804    164625.854806576
#> 12  4.75306460184801 9169.62907509525 9215.44726488591    258041.925124001
#> 13  1.82724535726068 20288.3088796566 20641.2464975694    470734.043551799
#> 14  10.8688444641493 6878.83777264691 7054.78971951673    171124.625622971
#> 15  3.95963894678306 8971.47199426603 9042.12247577848    253100.878716527
#> 16  1.43037225921892 20516.3243978681 20820.0253979127    502649.519651051
#> 17  9.16439641102569 6682.10000557057 7132.65001791483    164677.047759077
#> 18  6.05474352152988 9298.05467327189 9341.15242957057    261004.447739911
#> 19  1.17174843169044 19664.5396543259 19921.1143590431    474189.183139973
#> 20  10.3881325391718 6847.87566547399 7404.07192591667    166680.503342712
#> 21  6.10556292097839 9227.77898736901 9285.35430224044    262651.280604724
#>    ci95.integral.bt.up
#> 1     84958.6824486939
#> 2     44094.5301974676
#> 3     68403.6978373616
#> 4     155783.763553995
#> 5     59446.7580391792
#> 6     137407.341555532
#> 7     379501.794823417
#> 8     123019.553104818
#> 9     227338.876432743
#> 10    504845.951209984
#> 11    172491.043764281
#> 12    276291.278409715
#> 13    509711.888089085
#> 14    180981.205729662
#> 15    273567.096877329
#> 16    528495.430356374
#> 17    186086.848243142
#> 18    279710.257343987
#> 19    528303.534903712
#> 20    191576.501245957
#> 21    274532.329480078