Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.1534152747341  15.5162207295323
#> 2                 FALSE          0.75 32.3449186644616  31.5350603641179
#> 3                 FALSE          0.75 65.1528245477396  19.8983364813739
#> 4                 FALSE          0.75  241.47191946738  14.9459963330248
#> 5                 FALSE          0.75 65.2579684756377  22.0252096489106
#> 6                 FALSE          0.75 206.737731294993  12.3352271945353
#> 7                 FALSE          0.75 510.522995143348  7.14425288173965
#> 8                 FALSE          0.75 201.948560681191   13.313450938402
#> 9                 FALSE          0.75 289.797734898952  4.82805813578497
#> 10                FALSE          0.75 674.347309856515 0.545110071785605
#> 11                FALSE          0.75 327.145858021177  10.0157939111469
#> 12                FALSE          0.75 335.070709583987  3.91320468600242
#> 13                FALSE          0.75 730.695609486057  0.87488344919541
#> 14                FALSE          0.75 344.966962512242  9.77916639905641
#> 15                FALSE          0.75  319.06604324731  3.35110199224883
#> 16                FALSE          0.75 779.830256470821 0.492635766678903
#> 17                FALSE          0.75  330.55292899901  7.96225265301453
#> 18                FALSE          0.75 352.120688583883  4.96910565344955
#> 19                FALSE          0.75  791.89660945907 0.956578990717432
#> 20                FALSE          0.75 351.779199066751  9.59503982661588
#> 21                FALSE          0.75 336.362655579427  5.03488050675649
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3400.03617305629 2101.98115369973 81076.6879727539  3.0640227936844
#> 2  1466.92462038991 291.272924106783 42450.0366667479 1.96189146663846
#> 3  2195.32123680111 706.118768213796 65836.0652530741 1.21303776322355
#> 4  7555.97970853968 6038.01460014861 148745.167269114 2.22140428119282
#> 5  2360.21219126218 1135.14334711228 57131.3587601538 2.56844318997071
#> 6  4819.31945868541  3245.2953480107 131122.213643952 2.97013285371056
#> 7  17066.0738029932 15849.6710567397 362416.768493767 5.52891190827216
#> 8  5187.84099495322  3791.3634813467 118870.202475153 7.60516734827709
#> 9  7768.11201851572 6256.75701152513 220167.384252571 7.08082435632766
#> 10 20076.0452757119 18758.3045866466 486904.229321669 26.0772147737366
#> 11 6790.58267900797 5551.12824386118 168041.793616295 7.21001095962835
#> 12 9190.87462422325 7632.08530700695 267032.674154576 3.46228743604989
#> 13 20438.3800987647   19310.64371062 490774.099054713 28.3705823729744
#> 14 6966.92760050601 5665.67534304144 176868.187554376 13.4307547359821
#> 15 9007.01411262676 7395.30139396482 264647.679908623 5.39239361052728
#> 16 20700.3953881227 19410.4171992878  516272.09623318 27.2988041258425
#> 17 6910.43053925185 6019.68795370128 176311.519273428 14.9219110408331
#> 18 9319.91698306465  7685.5972317368 270525.113269076 6.74442078957946
#> 19  19803.072196877  18414.723810747 500833.534943581 19.6805293790782
#> 20 7114.10210523183 5773.76237581968 179538.904173946 18.9992568455841
#> 21 9253.75822449372 7497.19252629014 269579.433404156 9.01856070811643
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.609243587889167 50.0257476194947 71.5186050218929 2464.88644784921
#> 2   2.53192506432068 21.8908560918453 73.2450492647061 1018.42106221784
#> 3   1.85935837616896 4.80440209674299 116.253553650942 1317.23973213435
#> 4  0.282579893663601 70.7950484822489 97.7015740847215 3899.09747943166
#> 5    3.3634184051736 34.3991480602067  170.63302780822   1438.750186422
#> 6  0.433453304948272 12.4956326278463 90.3700736303639 2067.02040139747
#> 7  0.271405963200228 122.738508434488 133.321832129894 6677.84329227206
#> 8  0.637666845496711 48.0686483968347 132.017873774917 2406.69900778202
#> 9   0.40320731705026 17.7620419231738 70.5862464278351 4374.77975037455
#> 10 0.244426571282246 55.7424540737289 219.563685958281 6032.33703186544
#> 11 0.548395143143261 70.2638550476022 169.316378435345 3308.48640786337
#> 12 0.352780729787376 7.46745736126437 124.108707421295 5567.45981239919
#> 13 0.590689603217441 98.5204226597635 384.323343506223 9991.34009659342
#> 14  0.61664005685768  52.274717908773  175.73475053982 3155.62134522036
#> 15 0.520357710857724 12.1033275205321 135.190443902353 4486.33595340444
#> 16  0.43299453392463 97.5292508379344 275.683100487732 7382.73234566398
#> 17  0.74031881067422 60.6756205994311 124.787303986834 2365.92900102052
#> 18 0.318340783507651 15.7161237277608 50.6438988000927 4099.79297214481
#> 19 0.516504989809296 84.0557241648412 350.236887020301 10753.1637706609
#> 20 0.400215813281502 99.0419200227357 118.946451846317 3564.05629728925
#> 21 0.593438582435124 13.1850162086099 131.869595103736 2755.90170536558
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 86.1130977791233  96.193732770345    14.5140150274546
#> 2             TRUE 29.1176072018413 35.5722301270819    27.3700436333104
#> 3             TRUE 63.1573774272369 67.1482716682424    16.8396919525759
#> 4             TRUE 237.817709424818 245.126129509942    14.4811524079482
#> 5             TRUE 61.0328794281359 69.4830575231395       16.4923863724
#> 6             TRUE 201.851862750639 211.623599839347    11.6221965078954
#> 7             TRUE  501.42793505424 519.618055232456    6.69779007227528
#> 8             TRUE 189.438060393275 214.459060969107    12.2644889775599
#> 9             TRUE 278.149778832793 301.445690965111    4.16478209923729
#> 10            TRUE 631.450291553718 717.244328159312   0.143028362026311
#> 11            TRUE 315.285389992589 339.006326049766    9.11368390067621
#> 12            TRUE 329.375246751685  340.76617241629    3.33288038550218
#> 13            TRUE 684.026001482514   777.3652174896 -0.0968009480972812
#> 14            TRUE 322.873370971551 367.060554052932    8.76479350552553
#> 15            TRUE 310.195555757992 327.936530736627    2.49511355788787
#> 16            TRUE 734.923723683811 824.736789257832  -0.219640241627113
#> 17            TRUE  306.00638533684 355.099472661181    6.74442820945544
#> 18            TRUE 341.026116385025 363.215260782741    4.44543506457947
#> 19            TRUE 759.522138630487 824.271080287654   0.106928282481141
#> 20            TRUE 320.525421555765 383.032976577737    8.93668481376781
#> 21            TRUE 321.527123214575 351.198187944278    4.05867403865071
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1     16.51842643161 3317.74381822222 3482.32852789036    77021.9497660419
#> 2   35.7000770949254 1430.91416211882 1502.93507866099    40774.7340193995
#> 3   22.9569810101718 2187.41799535197 2203.22447825026    63669.2058937131
#> 4   15.4108402581014 7439.52185378638 7672.43756329298    142331.151915449
#> 5   27.5580329254212 2303.62559270314 2416.79878982122    54764.6147034896
#> 6   13.0482578811752 4798.76414301261 4839.87477435822    127721.965083653
#> 7   7.59071569120403 16864.1689566185  17267.978649368     351431.71627798
#> 8   14.3624128992441 5108.76806834042 5266.91392156601    114911.182607352
#> 9   5.49133417233265  7738.8934595521 7797.33057747934    212970.871563205
#> 10   0.9471917815449 19984.3489387606 20167.7416126632     476981.03490425
#> 11  10.9179039216175 6674.99863745466 6906.16672056127     162599.33347536
#> 12  4.49352898650265 9178.59065686397 9203.15859158253     257874.20276318
#> 13   1.8465678464881 20276.3140034894   20600.44619404    474338.344595817
#> 14  10.7935392925873 6880.93568954608 7052.91951146594    171677.190441489
#> 15  4.20709042660978 8987.10413885548 9026.92408639803    257267.657265272
#> 16  1.20491177498492 20539.9597704943 20860.8310057511    504127.501524563
#> 17  9.18007709657363 6810.61914336579 7010.24193513792    172419.566066749
#> 18  5.49277624231964 9294.06395953248 9345.77000659682    263780.953829898
#> 19  1.80622969895372 19664.8005306258 19941.3438631281    483144.580540844
#> 20  10.2533948394639 6951.17814679443 7277.02606366923    173676.031564905
#> 21  6.01108697486227 9232.06887283055 9275.44757615688     265045.97509883
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     85131.4261794659 85.1479305991127 97.1588999503556    14.3221032972696
#> 2     44125.3393140962 28.4996113898502  36.190225939073    26.5724872380494
#> 3     68002.9246124351 62.7752705318215 67.5303785636578    16.2539940640827
#> 4     155159.182622779 237.117967076242 245.825871858518    14.3921397414442
#> 5      59498.102816818 60.2238198232951 70.2921171279803    15.4329095747704
#> 6      134522.46220425  200.91627090172 212.559191688266    11.4856587168367
#> 7     373401.820709555 499.686327803134 521.359662483561    6.61229719386721
#> 8     122829.222342955 187.042432678568 216.854688683814    12.0636239212284
#> 9     227363.896941937  275.91931916055 303.676150637354    4.03777179436646
#> 10    496827.423739088 623.235968899991 725.458650813039  0.0660339920724031
#> 11     173484.25375723 313.014236540306 341.277479502049    8.94093943058608
#> 12    276191.145545973  328.28462620933 341.856792958645    3.22175445561916
#> 13    507209.853513609 675.089268035027 786.301950937087  -0.282868173110775
#> 14    182059.184667264 318.642683229717 371.291241794767    8.57055188761536
#> 15    272027.702551973 308.496951770676 329.635134723943    2.33120087896769
#> 16    528416.690941797  726.32460038417 833.335912557473  -0.356033519813372
#> 17    180203.472480107 301.305983358977 359.799874639043    6.51122778409306
#> 18    277269.272708254 338.901623836307 365.339753331459    4.34515771777456
#> 19    518522.489346318 753.322771876077 830.470447042063 -0.0557707893087875
#> 20    185401.776782987 314.540655649406 389.017742484096    8.81061683258413
#> 21    274112.891709483 318.686276591519 354.039034567335    3.87174088518365
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7103381617951 3301.98570772208  3498.0866383905    76245.5105349694
#> 2   36.4976334901864 1424.01854244989 1509.83069832993    40453.9313848009
#> 3    23.542678898665  2185.9046086915 2204.73786491073    63254.2753780908
#> 4   15.4998529246055 7417.22141351447 7694.73800356489    141102.936209428
#> 5   28.6175097230509 2292.78986106417 2427.63452146018    54311.4083947666
#> 6   13.1847956722339 4794.82801873484 4843.81089863599    127070.853657213
#> 7    7.6762085696121 16825.5063264616 17306.6412795248    349328.195640914
#> 8   14.5632779555755 5093.62644409542 5282.05554581101    114153.072419901
#> 9   5.61834447720348  7733.2984163463 7802.92562068514    211592.815941837
#> 10  1.02418615149881 19966.7900657274 20185.3004856964    475080.848739213
#> 11  11.0906483917077 6652.86552311466 6928.29983490127    161557.160256883
#> 12  4.60465491638567 9176.23840779517 9205.51084065133    256120.452922274
#> 13  2.03263507150159 20245.2800703515 20631.4801271778     471191.07246539
#> 14  10.9877809104975 6864.46915340481  7069.3860476072    170683.169717745
#> 15  4.37100310552997 8983.29159068652   9030.736634567     255854.46143995
#> 16  1.34130505317118 20509.2380564804 20891.5527197651    501801.940835678
#> 17    9.413277521936 6791.50632287697 7029.35475562674    171674.298431428
#> 18  5.59305358912455 9289.11338055824 9350.72058557106    262489.519043672
#> 19  1.96892877074365 19638.3229775139   19967.82141624    479757.333953085
#> 20  10.3794628206476 6919.97994198727 7308.22426847639    172553.353831259
#> 21  6.19802012832933 9227.91559272484 9279.60085626259     264177.86606164
#>    ci95.integral.bt.up
#> 1     85907.8654105383
#> 2     44446.1419486948
#> 3     68417.8551280575
#> 4       156387.3983288
#> 5     59951.3091255409
#> 6     135173.573630691
#> 7     375505.341346621
#> 8     123587.332530406
#> 9     228741.952563305
#> 10    498727.609904125
#> 11    174526.426975707
#> 12    277944.895386879
#> 13    510357.125644036
#> 14    183053.205391008
#> 15    273440.898377295
#> 16    530742.251630681
#> 17    180948.740115428
#> 18     278560.70749448
#> 19    521909.735934076
#> 20    186524.454516633
#> 21    274981.000746673