Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.3993620592587   15.555555244281
#> 2                 FALSE          0.75 32.6354528189659  31.8053052909097
#> 3                 FALSE          0.75 64.7927498966808  19.7126405206482
#> 4                 FALSE          0.75  242.23369235773  14.9334931070705
#> 5                 FALSE          0.75 65.7964846742832  23.0056146446345
#> 6                 FALSE          0.75 205.170146427801  12.5431491463969
#> 7                 FALSE          0.75 501.959269670485  6.81974160050214
#> 8                 FALSE          0.75 205.001730281438  13.4298810183923
#> 9                 FALSE          0.75 290.043746864585  4.70502101366119
#> 10                FALSE          0.75 676.659640132562 0.659625681846655
#> 11                FALSE          0.75  320.22515624944  9.57780221094315
#> 12                FALSE          0.75 334.233167552092  3.84796655711003
#> 13                FALSE          0.75 732.359774667243 0.758974937639325
#> 14                FALSE          0.75  339.08496302905  9.77756084902551
#> 15                FALSE          0.75  317.25743079533  3.26900450465424
#> 16                FALSE          0.75 775.655858206759 0.435881758713412
#> 17                FALSE          0.75 332.644145719468  7.97031594150477
#> 18                FALSE          0.75  351.81403747966  5.17115167916163
#> 19                FALSE          0.75 776.712058421716 0.668859724476424
#> 20                FALSE          0.75 347.714254540082  9.53248297486762
#> 21                FALSE          0.75  337.98160679238    5.037227426594
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3418.45291667897 2127.09955976131 81948.0944630514 1.97705202426615
#> 2  1467.49400131722 271.586419776121 42720.5214535967 1.99656463675015
#> 3  2195.79603948802  724.29590373249 66008.0616553096 1.84050578783992
#> 4  7561.10076039774 6055.18873880249 149100.340959081 2.98252254042572
#> 5  2375.78949211783 1113.62978832633 57546.3602337385 2.10501506438036
#> 6  4818.26001639261 3198.20126954103 130792.032386674 2.13658619626783
#> 7  17053.2098061686 15864.1243145979 363707.922534531 16.7902874868476
#> 8   5185.1537646419 3782.66548191674 119273.037391439 8.13158091446751
#> 9   7763.4572969935 6265.82021981337 221740.586790726 6.57580754667212
#> 10  20072.613874101 18696.5183641548  484217.40866667  37.316980747597
#> 11 6774.63857599122 5620.68967249518 167212.915932748 11.8147816120439
#> 12 9190.96509868964 7652.47112377895 267872.373188708 7.38724566972741
#> 13 20374.2956173183 19339.6202496543 484969.015220091 22.4242677551885
#> 14 6972.08116791619 5650.45090832868 176416.142487344  9.5266999129614
#> 15  8999.3037010788 7401.03139847799 264388.610983763 4.25898817985055
#> 16 20655.9729426748 19382.2972742554  513661.65463311 19.6943792897743
#> 17  6900.2906907524 6023.73719059112 176237.428426589 14.5400440091801
#> 18 9321.14656555943 7614.70556414291 271289.098507849 6.17789582298047
#> 19 19791.3200622712 18567.0924569081 497048.834612042 27.0019774864226
#> 20 7120.13492817205 5791.72364357361  179091.27547869 13.0805182484533
#> 21 9254.43264699139 7511.91605693988 269459.493885284 4.60109266281838
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.636498077211822 19.1736066815231  55.952805301194 1356.55450700158
#> 2    3.1170455803429 21.5204390173216 86.9155829457089 861.168451402526
#> 3   2.19198441845075 4.97289056669116 137.170759049262 817.620021112825
#> 4  0.300429064700758 69.9038925438956 103.166314799222 3386.69095147546
#> 5   1.85964987053643 24.6486865099695 129.867224229919 892.504745086585
#> 6  0.446021604108127 8.01078589281612 88.2451314969152 2821.15841629894
#> 7  0.563189216673519 172.660186071911 171.448122117013 8524.57026625656
#> 8  0.737602565892147 24.6606943644391 139.967838619858 1882.60653614612
#> 9   0.35047701542738 20.9384249268286 72.6747667276175 2915.66628686531
#> 10 0.470240428312426   81.96370901246 232.931383164856 11022.0751227619
#> 11 0.605933974467334 89.5998288570324 195.943744255985 2779.61472305896
#> 12 0.415172772242532 13.5756477104858 102.726748727699 2505.33995130809
#> 13 0.303126781267176 176.864667524227 250.590312195571 17296.4210811508
#> 14 0.479544380835254 34.0077203254319 164.783308028148 2438.49568930251
#> 15 0.332148691201582 14.7839473565304 109.407235511811 3332.51621804609
#> 16 0.238191437865131 90.6791879245787 165.756535681722 9021.33638932762
#> 17 0.550231827969717 66.5183337723049 197.440254278008 2592.09015635163
#> 18 0.369742181681232 13.4683291244734 110.786421097269  2430.2549047149
#> 19  0.34999754020847 71.2196516721521 177.126903564212 10713.5555360909
#> 20 0.335908344857505 125.418899636034 152.617509936131 4851.83328423279
#> 21 0.330609080522915 8.17626641793582 89.6706759949171 3806.01276977707
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.1471114793409 95.6516126391765   14.5085159072676
#> 2             TRUE 29.3511039915119 35.9198016464199   26.6777653112457
#> 3             TRUE 61.7651178756841 67.8203819176775   16.1068261522968
#> 4             TRUE  237.32744277873  247.13994193673   14.4392872956377
#> 5             TRUE 62.3337348933775 69.2592344551889   19.9464906076021
#> 6             TRUE  201.65546213494 208.684830720661   11.8094436076391
#> 7             TRUE 474.339246754621 529.579292586349    5.8932953390742
#> 8             TRUE 191.625279677139 218.378180885737   12.2165247974997
#> 9             TRUE 279.226543450309  300.86095027886   4.12848632328315
#> 10            TRUE 615.273206802764 738.046073462359 -0.113919822727285
#> 11            TRUE 300.789840497628 339.660472001252   8.58104082294438
#> 12            TRUE 322.081148425391 346.385186678794   3.16500734677106
#> 13            TRUE 695.471854209958 769.247695124528   0.26033138245482
#> 14            TRUE 323.413541672229 354.756384385872   8.98871034255152
#> 15            TRUE 310.251395239475 324.263466351184   2.72261990762764
#> 16            TRUE  743.25860427508 808.053112138438 0.0440568434252713
#> 17            TRUE 308.725773324367  356.56251811457   7.06518458449458
#> 18            TRUE 341.651398850858 361.976676108463     4.562925790296
#> 19            TRUE 732.293805456551 821.130311386881  0.093113770833491
#> 20            TRUE 326.196802021377 369.231707058788   8.97991374757703
#> 21            TRUE 330.412809362044 345.550404222716   4.49337548913381
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6025945812945 3386.91233368786 3449.99349967008    79716.5622990338
#> 2   36.9328452705738 1432.09287913372 1502.89512350071    41303.8993510395
#> 3   23.3184548889997 2187.61563450581 2203.97644447022     64663.076720579
#> 4   15.4276989185032 7446.10885716303 7676.09266363244    143529.234343904
#> 5   26.0647386816669 2335.24240280893 2416.33658142673    56078.1899280711
#> 6   13.2768546851548 4805.08227359893 4831.43775918629    126151.226791863
#> 7   7.74618786193008 16769.1838000803 17337.2358122569    349685.004446539
#> 8   14.6432372392849  5144.5869224124  5225.7206068714    116176.149639478
#> 9   5.28155570403923 7729.01358798887 7797.90100599814    216944.315748833
#> 10   1.4331711864206 19937.7835727755 20207.4441754265    466086.095089727
#> 11  10.5745635989419  6627.2468575214 6922.03029446104    162640.449713316
#> 12  4.53092576744899 9168.63315820589 9213.29703917339    263751.088968806
#> 13  1.25761849282383  20083.353239241 20665.2379953957    456516.402541598
#> 14  10.5664113554995 6916.13846798086 7028.02386785153    172404.817078442
#> 15  3.81538910168084 8974.98410767731  9023.6232944803    258906.621805078
#> 16 0.827706674001552 20506.8056785389 20805.1402068107    498821.556272666
#> 17  8.87544729851495 6790.86803169696 7009.71334980784     171973.44011939
#> 18  5.77937756802726 9298.99116414967 9343.30196696919    267291.329189593
#> 19  1.24460567811936 19674.1637352705 19908.4763892719    479425.035755172
#> 20  10.0850522021582 6913.82083827078 7326.44901807333    171110.009726127
#> 21   5.5810793640542 9240.98268873388 9267.88260524889    263198.602879001
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1      84179.626627069  88.524340091697 96.2743840268203    14.3080190129458
#> 2     44137.1435561538 28.7221861309357 36.5487195069962    25.6958959534376
#> 3     67353.0465900402 61.1853585525146  68.400141240847    15.4163510604848
#> 4     154671.447574258 236.387948178496 248.079436536964     14.344652140257
#> 5     59014.5305394059 61.6706551480977 69.9223142004687    19.3607008983831
#> 6     135432.837981486 200.982437483116 209.357855372485     11.668946802345
#> 7     377730.840622523 469.050306196264 534.868233144706    5.71589073582204
#> 8     122369.925143399 189.063831689081 220.939628873794    11.9841799892437
#> 9      226536.85783262 277.155164073107 302.932329656062    4.01808606342352
#> 10    502348.722243613 603.518357867271 749.800922397852  -0.262045557645699
#> 11     171785.38215218 297.068184289834 343.382128209046    8.39017162098717
#> 12    271993.657408609 319.754166039427 348.712169064758    3.03422792351466
#> 13    513421.627898584 688.408209867074 776.311339467413   0.164846446355659
#> 14    180427.467896247 320.412631199646 357.757294858454    8.83765386258841
#> 15    269870.600162449 308.909813962823 325.605047627837    2.61799306989914
#> 16    528501.752993554 737.054874798801 814.256841614717 -0.0309734595022449
#> 17    180501.416733787 304.145659461475 361.142631977461    6.89186155868412
#> 18    275286.867826105 339.705361666619 363.922713292702    4.44645700306641
#> 19    514672.633468912 723.788182548328 829.635934295104 -0.0171354543321769
#> 20    187072.541231253 322.076438773114 373.352070307051    8.87410261894691
#> 21    275720.384891568 328.963465173256 346.999748411504    4.38923362876909
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8030914756162 3380.87264758318 3456.03318577476    79289.2476293284
#> 2   37.9147146283818 1425.31394084327 1509.67406179117    41032.6312888477
#> 3   24.0089299808117  2186.0491739773 2205.54290499873    64405.5264139284
#> 4    15.522334073884  7424.0891310117 7698.11238978377    142462.426694189
#> 5   26.6505283908859 2327.47806655829 2424.10091767737    55797.0509333688
#> 6   13.4173514904488 4802.55887604269 4833.96115674253    125262.561890729
#> 7   7.92359246518223 16714.7958414677 17391.6237708696    346999.764812669
#> 8   14.8755820475409  5136.8188036876  5233.4887255962    115583.128580592
#> 9   5.39195596389885 7722.41798413692 7804.49660985009     216025.88086847
#> 10  1.58129692133901 19911.9650044366 20233.2627437654    462614.141426057
#> 11  10.7654328008991 6599.02291143143   6950.254240551    161764.871075552
#> 12  4.66170519070539 9164.35682917709 9217.57336820219    262961.906884144
#> 13  1.35310342892299 20027.6408689708 20720.9503656658    451068.029901036
#> 14  10.7174678354626 6905.42603607835 7038.73629975404    171636.690936311
#> 15  3.92001593940934    8970.32716426  9028.2802378976    257856.879196393
#> 16 0.902736976929068 20478.2417343426  20833.704151007    495979.835310028
#> 17  9.04877032432541 6769.91475655869 7030.66662494612    171156.931720139
#> 18  5.89584635525684 9294.74864047547  9347.5444906434    266525.798894608
#> 19  1.35485490328503 19651.7295449938 19930.9105795486    476050.265761304
#> 20  10.1908633307883 6874.31388488543 7365.95597145868    169581.682241593
#> 21  5.68522122441892 9238.40716481223 9270.45812917054    261999.708856521
#>    ci95.integral.bt.up
#> 1     84606.9412967745
#> 2     44408.4116183456
#> 3     67610.5968966907
#> 4     155738.255223973
#> 5     59295.6695341082
#> 6      136321.50288262
#> 7     380416.080256394
#> 8     122962.946202285
#> 9     227455.292712982
#> 10    505820.675907283
#> 11    172660.960789943
#> 12    272782.839493271
#> 13    518870.000539147
#> 14    181195.594038377
#> 15    270920.342771134
#> 16    531343.473956193
#> 17    181317.925133038
#> 18     276052.39812109
#> 19     518047.40346278
#> 20    188600.868715786
#> 21    276919.278914047