Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75   91.87300503848  15.4534902506906
#> 2                 FALSE          0.75 32.6226098327485  31.3514144266971
#> 3                 FALSE          0.75 64.9141824421031  20.2782269450116
#> 4                 FALSE          0.75  243.08836533655  15.0116523154879
#> 5                 FALSE          0.75 65.9806150010884  21.7345567151306
#> 6                 FALSE          0.75 204.642667183098  12.3242746712153
#> 7                 FALSE          0.75 504.532884874695  7.05025081830087
#> 8                 FALSE          0.75 201.959532811117  13.2709383535326
#> 9                 FALSE          0.75 288.596053921557  4.69712632855813
#> 10                FALSE          0.75 668.644766430061 0.553526393479354
#> 11                FALSE          0.75 326.035168542509   10.070211404636
#> 12                FALSE          0.75  333.39773768685  3.69590499344516
#> 13                FALSE          0.75 726.792407990441 0.693940703738291
#> 14                FALSE          0.75  343.59603543902  9.72923026966849
#> 15                FALSE          0.75 318.349916840344  3.28191993014878
#> 16                FALSE          0.75  775.87567262246  0.57068778458646
#> 17                FALSE          0.75 338.926387503801  8.22979422914625
#> 18                FALSE          0.75 349.421649199357  5.03321964594974
#> 19                FALSE          0.75 775.046296603649 0.597430169060363
#> 20                FALSE          0.75 341.208691475638   9.3669281754295
#> 21                FALSE          0.75 335.264604869737  5.26332610198026
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3409.06354254326 2122.43817571813 81566.6125171743 2.27333787278685
#> 2  1460.03754243353 276.474324681745 42730.2139511825 2.07960219537894
#> 3  2194.57109401215 675.970497583095 66145.4891675522 1.33248367629366
#> 4  7553.48025125673 6029.61739772271 149060.395335736 1.96388422482296
#> 5  2355.87128571786 1128.36291671179 56976.3830526515 2.77361678814867
#> 6  4817.97265625294 3240.27519522429 130474.220987391 2.67504736722575
#> 7  17042.4974574589 15797.6773236735 362307.313700801 14.0189751636412
#> 8  5176.63952980817 3789.81200063966 117879.880927373 7.88136873246956
#> 9  7766.67993863836 6265.13178490941 223393.102974726 6.78914691166934
#> 10  20046.831907871 18679.7725225194 484707.516250379 27.2400404605761
#> 11 6800.89706993532 5561.35012716997 168248.423866863 10.8711527181044
#> 12 9191.85628169153 7678.47681292164 267589.607683904 6.18265163540152
#> 13 20422.5723431617 19423.6718697881 489053.609007968  17.758749862951
#> 14 6931.94262030271 5656.65554025001 174456.917377375 13.6932583007119
#> 15  9008.8379373651 7421.60621375904 265118.011469361 5.33627832117659
#> 16 20662.2424178481   19290.07159289 512093.736199521 23.6589732707959
#> 17 6900.68921342873 5997.92925312849 175987.660258128 14.7835854824292
#> 18 9319.72627440105 7610.26878293287  271227.94088294 6.60942380626294
#> 19 19780.6583947055 18655.4146173141  503441.90458417 19.3273604045974
#> 20 7124.81132459548 5785.32983038648 180490.182529292 9.42395228427866
#> 21 9258.03979860703 7418.13356034215 267528.333617165 8.44393172511818
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.755900628961111 28.7676226206678 71.3296350580498 1765.26364560708
#> 2   2.60271600070266 19.6772263007822 87.2461150577363 946.129436399792
#> 3   2.14982484477323 4.89667965686141 140.215480027773 1035.36425638142
#> 4  0.208013355496984 49.9826642441912 73.9613957264284 2487.44790188973
#> 5   2.35037599995332 33.1563301393058 134.605776297392 1373.86063830758
#> 6  0.469506148724055 9.95019155652924 91.4750635170872 2292.60286279846
#> 7  0.604918680113291 178.015633720573 238.240809148225 9299.19814562705
#> 8  0.730288267722382 39.7458299321837 130.735186292637  2072.3630265207
#> 9  0.435974022139359 27.2411868417138 83.0693478399841 2024.97905856218
#> 10 0.373393551358714 81.2757346667592 268.836223833958 9246.28606798884
#> 11 0.581284218084221 75.1159748530803 168.901235652887 2782.00891586788
#> 12 0.308124784254406 10.1764432058047 66.8589554702448 2828.21476139246
#> 13 0.272124197669426  121.96100333306 223.792292374252 12716.6304101969
#> 14 0.586099713375341 89.8064473213607 190.368115124475 4537.81240456509
#> 15 0.374440495713854 12.7525942694739 86.6969517813069 3759.40936890439
#> 16 0.427899017544552 123.550210534093 305.530801376504 10355.3290197444
#> 17 0.516963562615852 79.6620247232137 143.709560681039 2755.12703613578
#> 18 0.479066642578088 7.92957071515066  132.06136792683 2806.21695865821
#> 19 0.158629502790027 73.5704933827964 162.957979723414 7398.94778092148
#> 20 0.462472391619269 111.796237864556 189.494104152545 4269.86926021959
#> 21 0.889324420794331 9.88604528432535 226.176796628979 3682.75858453802
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.1333642377456 95.6126458392144    14.2100337160496
#> 2             TRUE 29.2016642213501 36.0435554441469    27.0699466055413
#> 3             TRUE    62.7222467946 67.1061180896062    16.7417650753596
#> 4             TRUE 239.857775786716 246.318954886384    14.6694703456954
#> 5             TRUE 61.4180153845838 70.5432146175929    17.8681881952074
#> 6             TRUE 200.242214264012 209.043120102184    11.5519370565643
#> 7             TRUE 481.471670730505 527.594099018885     6.0551595895145
#> 8             TRUE 188.994681246205 214.924384376029    12.0696141531293
#> 9             TRUE 277.427907251861 299.764200591253    3.97994906213888
#> 10            TRUE 623.834899872413 713.454632987709 -0.0607059985057299
#> 11            TRUE 308.152122321227 343.918214763791    9.11399886588742
#> 12            TRUE 323.227275746614 343.568199627085    3.18903972334666
#> 13            TRUE 697.579264465886 756.005551514995   0.246296398572086
#> 14            TRUE 321.070625534349 366.121445343691    8.76509624116606
#> 15            TRUE 309.571739002009  327.12809467868    2.66596531469949
#> 16            TRUE 736.956661592001  814.79468365292  -0.133206099274329
#> 17            TRUE 314.607389385205 363.245385622397    7.37938916864317
#> 18            TRUE 338.549147038054 360.294151360659    4.24515501890879
#> 19            TRUE 743.252788738087 806.839804469212   0.336484636970768
#> 20            TRUE    325.706289968 356.711092983276     8.6061610912158
#> 21            TRUE 321.374337181918 349.154872557557    3.80038742977358
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6969467853317 3361.74080333226 3456.38628175426    78662.7538201507
#> 2    35.632882247853 1427.66850516874 1492.40657969831    41173.8310283048
#> 3   23.8146888146636 2186.51605597661 2202.62613204769    64442.3149658048
#> 4   15.3538342852804 7471.25876857504 7635.70173393843    144968.543537128
#> 5   25.6009252350538  2301.3291226387 2410.41344879701    54716.3823026355
#> 6   13.0966122858664 4801.60459114245 4834.34072136343    126702.889278087
#> 7   8.04534204708723 16749.6617399886 17335.3331749293    347010.132751244
#> 8    14.472262553936 5111.25763956972 5242.02142004661    114470.843748746
#> 9   5.41430359497737 7721.86818628374 7811.49169099298    220062.012423391
#> 10  1.16775878546444 19913.1333243442 20180.5304913978    469497.375668538
#> 11  11.0264239433845 6677.33129130201 6924.46284856864    163672.019200261
#> 12  4.20277026354366 9175.11603261798 9208.59653076507    262937.194401414
#> 13   1.1415850089045 20221.9464926788 20623.1981936446    468134.751983194
#> 14  10.6933642981709 6784.21101445907 7079.67422614635    166992.215971866
#> 15  3.89787454559807 8987.85991979181 9029.81595493838    258933.783057514
#> 16  1.27458166844725 20459.0023215195 20865.4825141767    495059.219962041
#> 17  9.08019928964933 6769.64518275904 7031.73324409842    171455.476283685
#> 18   5.8212842729907 9306.68213057463 9332.77041822748    266611.713985947
#> 19 0.858375701149958 19659.6349330908 19901.6818563202    491270.635484554
#> 20  10.1276952596432 6940.90651330829 7308.71613588268    173466.247596231
#> 21  6.72626477418693 9241.77725411432 9274.30234309975      261470.1957456
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1      84470.471214198 87.4172628078178 96.3287472691422   13.9719250179269
#> 2     44286.5968740601 28.5465895298058 36.6986301356912   26.2500910653199
#> 3     67848.6633692997 62.3025144365675 67.5258504476387   16.0645702492561
#> 4     153152.247134345 239.239152255897 246.937578417203   14.6039461387138
#> 5     59236.3838026674  60.544326096317 71.4169039058598   17.1278197552221
#> 6     134245.552696694 199.399574343336 209.885760022861   11.4040426197162
#> 7     377604.494650357 477.055693553958 532.010076195432   5.86461020527882
#> 8     121288.918105999 186.512050095477 217.407015526757   11.8395733487968
#> 9     226724.193526061 275.289325974685 301.902781868429   3.84261724516499
#> 10    499917.656832221 615.254287127332  722.03524573279 -0.178324967183725
#> 11    172824.828533466 304.727709215024 347.342627869994   8.93089433719089
#> 12    272242.020966395 321.279740481463 345.515734892237   3.09198041630653
#> 13    509972.466032742 691.985258259057 761.599557721825  0.160577276306217
#> 14    181921.618782885 316.757249169624 370.434821708415   8.58047483145282
#> 15    271302.239881209 307.890811330838  328.80902234985   2.54801655854963
#> 16       529128.252437   729.5040850117  822.24726023322 -0.267994289800863
#> 17    180519.844232571 309.950559958239 367.902215049362   7.21654564641918
#> 18    275844.167779933 336.467178539082 362.376119859632   4.09424902649669
#> 19    515613.173683786 737.164670210638  812.92792299666  0.286516343591909
#> 20    187514.117462353 322.737744998452 359.679637952824   8.46048228785573
#> 21     273586.47148873 318.714498688506 351.814711050969   3.52025023722337
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9350554834544 3352.67900220675 3465.44808287977    78106.6957717844
#> 2   36.4527377880743 1421.47017888399 1498.60490598306    40875.8002558389
#> 3   24.4918836407671  2184.9736018847  2204.1685861396    64116.1752250447
#> 4    15.419358492262 7455.51422933812 7651.44627317535    144184.997448032
#> 5   26.3412936750391 2290.88487864482  2420.8576927909    54283.6162015686
#> 6   13.2445067227145 4798.47028080215 4837.47503170374    125980.719376306
#> 7   8.23589143132292 16693.5868153666 17391.4080995513    344080.885335372
#> 8   14.7023033582685 5098.73770314109 5254.54135647525    113818.049395392
#> 9   5.55163541195127  7713.2872124286 7820.07266484812    219424.144019944
#> 10  1.28537775414243 19887.5314679241 20206.1323478178    466584.795557121
#> 11   11.209528472081 6653.66975922328 6948.12438064736    162795.686391762
#> 12   4.2998295705838 9171.91045300815  9211.8021103749    262046.306751575
#> 13  1.22730413117037 20183.5287766289 20661.6159096945    464129.013403982
#> 14  10.8779857078842 6755.92198355285 7107.96325705258    165562.805064428
#> 15  4.01582330174794 8983.84285259693 9033.83302213327    257749.569106309
#> 16  1.40936985897378 20420.0840052013 20904.4008304949    491797.291320822
#> 17  9.24304281187332 6744.55164497123 7056.82678188623    170587.611267302
#> 18   5.9721902654028 9304.18431579936 9335.26823300275     265727.75564397
#> 19 0.908343994528817 19636.4602276752 19924.8565617358    488939.966933564
#> 20  10.2733740630033 6905.69069838095 7343.93195081001    172121.238779262
#> 21  7.00640196673715 9238.66314984975 9277.41644736431    260310.126791471
#>    ci95.integral.bt.up
#> 1     85026.5292625642
#> 2     44584.6276465261
#> 3     68174.8031100598
#> 4      153935.79322344
#> 5     59669.1499037343
#> 6     134967.722598476
#> 7      380533.74206623
#> 8     121941.712459353
#> 9     227362.061929508
#> 10    502830.236943637
#> 11    173701.161341964
#> 12    273132.908616234
#> 13    513978.204611954
#> 14    183351.029690323
#> 15    272486.453832414
#> 16     532390.18107822
#> 17    181387.709248954
#> 18     276728.12612191
#> 19    517943.842234776
#> 20    188859.126279322
#> 21     274746.54044286