Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.2237095887713  15.6127134852773
#> 2                 FALSE          0.75 32.7439223459426  30.6181956053152
#> 3                 FALSE          0.75 65.3845274289502  20.6072264084456
#> 4                 FALSE          0.75 242.412365248409  14.9123757392403
#> 5                 FALSE          0.75 64.9238075680901  23.0430127493219
#> 6                 FALSE          0.75 204.915098405476  12.3460365464856
#> 7                 FALSE          0.75 511.624460520497  7.10849670870571
#> 8                 FALSE          0.75  200.78975530599  13.2688247235893
#> 9                 FALSE          0.75 290.975942140274  4.81946428062686
#> 10                FALSE          0.75 679.229539790354 0.625588218876424
#> 11                FALSE          0.75 319.653344839864  9.45697445898584
#> 12                FALSE          0.75 336.258738178882  3.87245310490003
#> 13                FALSE          0.75 729.239343046959 0.792620235960066
#> 14                FALSE          0.75 341.140364072276  9.66185509558254
#> 15                FALSE          0.75  316.89242553432  3.20714846180847
#> 16                FALSE          0.75 782.240656133607 0.532580830314485
#> 17                FALSE          0.75 332.173018687731  7.90503379552066
#> 18                FALSE          0.75 350.125898992691  5.03086692902109
#> 19                FALSE          0.75 781.240201893768 0.694672764274753
#> 20                FALSE          0.75  351.74815271942  9.51891600008735
#> 21                FALSE          0.75 334.539408977167  5.11152180245874
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.00602554549 2114.18739342252  81329.134187778  2.6855181942279
#> 2  1466.81206687251 320.787444192102 42045.0050830609 1.96692245004309
#> 3  2197.46167030239 655.396831946805 65673.6655853175  1.6556247023418
#> 4  7523.66337832872 6016.40779092297 147242.871023433 2.10588641213332
#> 5  2374.21811730701 1137.39196909916 57514.4532538402 2.14093541458381
#> 6  4820.95501546773 3234.84586158862 131408.011827704 1.73871784515316
#> 7  16989.4413982989 15802.3593074054 359860.593805456  13.830664626927
#> 8  5167.69100073184 3774.49833336488  118551.23738256 7.96984149199131
#> 9    7768.637966564 6263.14117005547 220296.160236821 6.41006016415824
#> 10 20048.0723476489 18704.6031370686 482194.862278378 27.2218889116365
#> 11 6806.75356599446 5700.84297710378 168034.444710529 10.4304467808605
#> 12 9184.50438257324 7657.55332146752 268212.409440777 5.57423551993297
#> 13 20468.3725028249 19402.4765401364 493176.329225579 27.0068500433416
#> 14 6936.93933253718 5656.40954204725 175189.271335208 11.4126554859842
#> 15 9009.22765124908 7423.72387467864  265475.95830433 5.53812773327654
#> 16 20654.5723476239  19334.975004375 513242.452196756 24.7018658363196
#> 17 6890.18984054835 6035.95239463359 175012.920233747 12.2725009443631
#> 18 9323.05394377187 7630.40859653903 269621.345771219 7.01926579422164
#> 19 19788.7128582491 18608.3227741386 503782.174581697 21.0966075870059
#> 20 7121.84230903389 5786.53228434036 179963.576976217 14.6259007914686
#> 21 9256.67795381879 7459.33341806222  266661.48441933 7.60267206618747
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.591592813030524 27.3001533811512 52.6206927573108 1958.75447857767
#> 2   3.19235866239605  19.051394498626 100.421248765635 1284.61328111838
#> 3    1.8749367419316 10.0709945918934 119.912065881614  2114.2805190779
#> 4  0.337365282947788 92.1494603344212 130.522794876419 4722.21268658468
#> 5   2.70876300460689 24.9153905755301 124.016054346848 1262.97312060615
#> 6  0.588772687907682  12.841213663887 114.805263291294 1575.56112586143
#> 7  0.575247992423512 200.501654055084 219.770871693232 9495.85506939854
#> 8  0.513771619267378 52.0693149397449 85.5747143556658 2584.01328468754
#> 9  0.342259238656697 17.3243840782888 61.8803742841371  4986.3801542423
#> 10 0.400746831859853 98.4279314304295 253.116971834104 12413.2830045541
#> 11 0.477044387590773 82.7381729177837 133.880132746337 2996.65821523106
#> 12 0.291938133778807 14.3486264607867 76.9706320802556 2601.97042905689
#> 13 0.425527160242919 63.8751021877174 266.777200866605 6656.29811927299
#> 14 0.512014254397824  59.529702771432 167.319990282375 3871.04101221334
#> 15 0.329657170181015 12.4075966571318 79.3690053547707  3288.5022906738
#> 16 0.457314394949694 76.5790557813292 320.410138216614 9304.19791454843
#> 17 0.422735236659713 91.4115626611215 136.729379622781 3870.90474826507
#> 18 0.538815793197701 9.43966215619082 137.468281885174 5097.38832033547
#> 19 0.271551629523719 84.8747452610978 195.149702873069 7111.82486634983
#> 20 0.444591310824587 98.6857235102213 108.627581865417 4510.55070505055
#> 21 0.582756300463737 9.09898339702969 181.399665903922 6902.36762619964
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.8060321592664 96.6413870182762    14.6395433078421
#> 2             TRUE 29.5083349156217 35.9795097762635    25.3667656056737
#> 3             TRUE 62.6610247935979 68.1080300643025    17.5229554679681
#> 4             TRUE  238.94818210045 245.876548396368    14.3574098487912
#> 5             TRUE 61.4019688110997 68.4456463250804    18.5870976067435
#> 6             TRUE 202.054907550199 207.775289260753    11.3775054748775
#> 7             TRUE 488.873017209202 534.375903831792    6.16221376116904
#> 8             TRUE 187.679366051665 213.900144560316    12.4236704098944
#> 9             TRUE 280.431393170233 301.520491110314    4.25644783303659
#> 10            TRUE 634.449532530712 724.009547049996 -0.0336403195330337
#> 11            TRUE 302.495259885348 336.811429794379    8.67223644139901
#> 12            TRUE 327.089120748592 345.428355609171    3.39221487483389
#> 13            TRUE 684.813074725662 773.665611368256   0.092628057360464
#> 14            TRUE 322.366545797832  359.91418234672    8.81959164709812
#> 15            TRUE  307.78220541308  326.00264565556     2.6648624168607
#> 16            TRUE 741.606086832861 822.875225434353  -0.219701349377763
#> 17            TRUE 311.984754634254 352.361282741209    7.20963433121543
#> 18            TRUE 338.579206761196 361.672591224185    4.14451494921088
#> 19            TRUE 746.536282413143 815.944121374392   0.247970333708235
#> 20            TRUE 327.688545917454 375.807759521386    8.78756329378091
#> 21            TRUE 322.033013428288 347.045804526045    4.15288768819589
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5858836627125  3366.0972732335 3455.91477785748    78106.9830705177
#> 2   35.8696256049567 1435.47252292227 1498.15161082275    39931.8162356212
#> 3   23.6914973489231 2180.89488419873 2214.02845640606    62195.6741314343
#> 4   15.4673416296894  7372.0775160786 7675.24924057885    139474.831154002
#> 5   27.4989278919002 2333.23229981027 2415.20393480376     55436.862470443
#> 6   13.3145676180938 4799.83121899063 4842.07881194482    128816.213775662
#> 7   8.05477965624239 16659.6161773783 17319.2666192195    344239.912216296
#> 8   14.1139790372841 5082.03697765596 5253.34502380772    114300.535529249
#> 9   5.38248072821713 7740.13935475521 7797.13657837279    212093.564883092
#> 10  1.28481675728588 19886.1584004458  20209.986294852    461775.011735887
#> 11  10.2417124765727 6670.64927154471 6942.85786044421    163104.941946473
#> 12  4.35269133496617 9160.90089204525 9208.10787310124    263932.168084978
#> 13  1.49261241455967 20363.2979597262 20573.4470459237    482226.718819375
#> 14   10.504118544067 6839.01297147818 7034.86569359619    168821.408870117
#> 15  3.74943450675624  8988.8171547481 9029.63814775006    260066.372036172
#> 16  1.28486301000673 20528.5998008636 20780.5448943842    497937.046627323
#> 17  8.60043325982589 6739.81781997081  7040.5618611259    168645.281922851
#> 18  5.91721890883131 9307.52569952494 9338.58218801881    261236.141984267
#> 19  1.14137519484127 19649.0939022946 19928.3318142036    492083.222676552
#> 20  10.2502687063938 6959.50429385958  7284.1803242082    172543.721066409
#> 21  6.07015591672159 9241.71012613067  9271.6457815069    255307.089674232
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84551.2853050383 86.9600939280846  97.487325249458    14.4531915717375
#> 2     44158.1939305007 28.8887543438582 36.5990903480271    24.3611726270189
#> 3     69151.6570392006 62.1395030123603 68.6295518455401    16.9323503942596
#> 4     155010.910892865 238.284827880628  246.53990261619    14.2511397846626
#> 5     59592.0440372373 60.7275741555058 69.1200409806743    17.7338372602924
#> 6     133999.809879746 201.507211428976 208.322985381976    11.1920420781866
#> 7     375481.275394617  484.51635785172 538.732563189274    5.98101064355563
#> 8     122801.939235871 185.168865981687 216.410644630293    12.2618323498252
#> 9      228498.75559055 278.412224218524 303.539660062024    4.14863617285973
#> 10     502614.71282087 625.874637523546 732.584442057161  -0.159875571568887
#> 11    172963.947474584 299.209669149377  340.09702053035    8.52196745930792
#> 12    272492.650796575 325.333236559813  347.18423979795    3.30025436269357
#> 13    504125.939631783  676.30591696201 782.172769131909 -0.0414129981160555
#> 14    181557.133800298 318.771559319747 363.509168824805     8.6583071569628
#> 15    270885.544572489 306.037695177098 327.747155891542    2.56102040825368
#> 16    528547.857766188 733.824999094421 830.656313172794  -0.363755383786916
#> 17    181380.558544643  308.11891683678 356.227120538683    7.07647273166762
#> 18    278006.549558171 336.368138036016 363.883659949365     3.9747879743536
#> 19    515481.126486843 739.890851023236 822.589552764299   0.162431570408264
#> 20    187383.432886026 323.081387168142 380.414918270699    8.64751703087116
#> 21    278015.879164429 319.638171727439 349.440646226894    3.96931945354982
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7722353988171 3357.49772491843 3464.51432617255    77489.9754097658
#> 2   36.8752185836114  1429.4713336552 1504.15280008981    39527.1630520689
#> 3   24.2821024226315 2177.72252090228  2217.2008197025    61529.6757679248
#> 4    15.573611693818 7343.05043607326 7704.27632058419    137987.334157728
#> 5   28.3521882383514 2325.38395177897 2423.05228283505    55039.0259374521
#> 6   13.5000310147847 4795.78623668651 4846.12379424894    128319.912021016
#> 7    8.2359827738558  16596.458156351 17382.4246402469    341248.717869435
#> 8   14.2758170973533 5065.63514344994 5269.74685801374    113486.571344573
#> 9   5.49029238839399 7734.68217377055 7802.59375935745    210522.855134506
#> 10  1.41105200932173 19855.1536020453 20240.9910932525    457864.827589452
#> 11  10.3919814586638  6644.5867470756 6968.92038491332    162160.994608676
#> 12  4.44465184710649  9156.3810747101 9212.62769043638    263112.547399825
#> 13  1.62665347003619  20343.177302537 20593.5677031129    480129.984911804
#> 14  10.6654030342023 6820.26111510518 7053.61754996919    167602.030951269
#> 15  3.85327651536326  8984.9087618011 9033.54654069706     259030.49381461
#> 16  1.42891704441589 20504.4773982925 20804.6672969553    495006.224284241
#> 17   8.7335948593737 6711.02317773255 7069.35650336415    167425.946927147
#> 18  6.08694588368859 9304.55220594574 9341.55568159801    259630.464663362
#> 19  1.22691395814124 19622.3583575374 19955.0673589609    489842.997843652
#> 20  10.3903149693035 6928.41829095386 7315.26632711392    171122.897594318
#> 21  6.25372415136767 9238.84394636061 9274.51196127696    253132.843871979
#>    ci95.integral.bt.up
#> 1     85168.2929657902
#> 2      44562.847114053
#> 3     69817.6554027102
#> 4     156498.407889139
#> 5     59989.8805702282
#> 6     134496.111634392
#> 7     378472.469741477
#> 8     123615.903420548
#> 9     230069.465339136
#> 10    506524.896967304
#> 11    173907.894812381
#> 12    273312.271481728
#> 13    506222.673539354
#> 14    182776.511719146
#> 15    271921.422794051
#> 16     531478.68010927
#> 17    182599.893540347
#> 18    279612.226879077
#> 19    517721.351319743
#> 20    188804.256358117
#> 21    280190.124966682