Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.3782690586278  15.4099506600506
#> 2                 FALSE          0.75 32.3262588603384  31.1440317661508
#> 3                 FALSE          0.75 65.1336550796553  21.0673567240934
#> 4                 FALSE          0.75 242.700122920558  14.8911590364237
#> 5                 FALSE          0.75 64.2394514749237  22.6522062162634
#> 6                 FALSE          0.75 206.232862026257  12.6068293410579
#> 7                 FALSE          0.75  503.44777511289  6.94355708647354
#> 8                 FALSE          0.75 202.277624237465  13.3562654000014
#> 9                 FALSE          0.75 289.245496559752  4.68371026009857
#> 10                FALSE          0.75 676.711295441277 0.593131892267887
#> 11                FALSE          0.75 328.479163962883  9.92209737245281
#> 12                FALSE          0.75 333.039767572173  3.83485743863575
#> 13                FALSE          0.75 729.710502994214  0.68174761128376
#> 14                FALSE          0.75 342.696356128699  9.82739366048692
#> 15                FALSE          0.75 318.910418023481  3.28206282084096
#> 16                FALSE          0.75 779.999452845811 0.384725132114353
#> 17                FALSE          0.75 330.257517589279  8.00614412887664
#> 18                FALSE          0.75 352.252029034445  5.12826183141167
#> 19                FALSE          0.75 781.180695137007 0.602599176202136
#> 20                FALSE          0.75  345.89770145781  9.41728963713977
#> 21                FALSE          0.75 336.479656292023  5.09715034731177
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.16390498336 2128.08812201137 81564.3952468808  2.4085842102303
#> 2  1464.37676777045 294.746250426526 42669.3696544183 2.93844816302799
#> 3  2193.95555041455 625.514216775583 66551.0588960202 1.45545509027837
#> 4  7563.73559034081 6067.08573344044 149431.295812983 2.57901980868825
#> 5   2357.9803782159  1109.2418274473 56952.8624530555 1.84804022713307
#> 6  4819.98883583427 3192.49700939163  130401.82140965 3.07049013565707
#> 7  17068.9841433648 15855.3211522168 364133.057885414 13.1600637692368
#> 8  5184.79361150126 3793.51364443003 119248.448699841 8.91418660487951
#> 9  7764.95372301726 6279.40570982672 220816.683568697 4.98459825993745
#> 10  20109.202157357 18788.2933008633 485796.908613733 33.8563428985761
#> 11  6774.4211785513 5566.26926786137 167424.546166121 13.1056456915337
#> 12 9189.61122784818 7635.44865251659 266325.545271431 5.85873720901773
#> 13 20467.1113352277  19491.308735842 492425.464330636 23.9005733859842
#> 14 6956.14022506891 5605.02714312806 176523.882816663 13.6011276237422
#> 15 9005.22253821843    7422.83505402 264469.021245224  5.0535142618598
#> 16 20666.0139285476 19462.9605449173 512951.287791022 21.5251722296139
#> 17 6935.06400418969  6009.3551219362 177445.003004222 15.3515068941625
#> 18 9320.31386734769 7625.31477821914 271045.794760161 7.42370499295066
#> 19 19784.5880278471 18659.0634570331 500200.591371123 23.4155684872936
#> 20 7140.93948592489 5812.04381327033 180670.884241709 14.4931529151092
#> 21 9253.81523372244 7487.56740798246 268708.619522402  7.0605756598908
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.507684794025138 24.0116891013286 59.0077349375275 1563.41694746122
#> 2   2.07569062278932 30.4910059081157 61.5400694284379  1094.6053682013
#> 3   1.11325019720059 3.89720180515927 63.5888013621922 911.692331545662
#> 4  0.320133768956994 51.5530752790057 106.912609687873 2878.95143770688
#> 5   2.36743928685374 29.7870378274693 145.550258166154 1067.96382360359
#> 6  0.366519219337634 9.28527433482444 74.5278221621401 2328.55219457224
#> 7  0.439178285426791 80.5110065394781 115.111110193865 4569.95312413815
#> 8  0.599507824994944 42.9202994652673 97.1914014337049 2223.07593256969
#> 9  0.310974154280808  21.072831000111 70.7326647653763 3396.67187347723
#> 10 0.405382728146274 81.6736040418657 262.835798989179 8564.79557088298
#> 11 0.532920314946425 92.2386032247859 182.350029385295 3272.81083988798
#> 12 0.386877824495688 11.4252208370782 97.4095864206442 4910.99803094112
#> 13 0.232097019888569  64.696611659271 147.296485307524 7450.43313716223
#> 14  0.59979467644003  54.186678090593 185.759709100407 3261.74788657202
#> 15 0.268535927912593 18.0362703691988 48.8885246526717  5061.2832043235
#> 16 0.288327800449398 90.5176702292485 201.994120155761 10176.5731486587
#> 17  0.74426727693105 53.7746281441278 147.980866324068 1983.99393954556
#> 18 0.433585641537759 7.48032748898191 87.2961040471652  2394.8896244546
#> 19 0.317148203897817 83.9406863480437 188.501739011467 14334.2768382617
#> 20 0.354419701341564 74.6465001410445 95.8905020551284 2517.41326446458
#> 21 0.520135612291683 9.11052162977428 117.308036499125 3869.30506653242
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE  87.416148032799 95.3403900844567    14.5748091738792
#> 2             TRUE 27.4925116321573 37.1600060885194    27.7295206916624
#> 3             TRUE 62.7394314561474 67.5278787031632    19.2360601496984
#> 4             TRUE 238.457635335266 246.942610505851    14.3645389864894
#> 5             TRUE 61.1994253012898 67.2794776485576     18.757768589389
#> 6             TRUE 201.181905753101 211.283818299413    12.0039052252475
#> 7             TRUE 481.799470212496 525.096080013285    6.22110880694647
#> 8             TRUE 187.613787272438 216.941461202492    12.3700750278847
#> 9             TRUE 281.045832422155  297.44516069735    4.17215777630664
#> 10            TRUE 621.017611373119 732.404979509435 -0.0737226955327331
#> 11            TRUE  306.92037680031 350.037951125455    9.04544345436594
#> 12            TRUE 323.402144863339 342.677390281007    3.19844341734034
#> 13            TRUE  690.39405977427 769.026946214158   0.299948013567064
#> 14            TRUE 320.322501187643 365.070211069755    8.84073141774307
#> 15            TRUE 310.597387062721  327.22344898424    2.84032121942474
#> 16            TRUE 744.590544528096 815.408361163526 -0.0895740996249071
#> 17            TRUE 305.004288748382 355.510746430176    6.78182445832506
#> 18            TRUE 340.040034321041 364.464023747849    4.41501345108206
#> 19            TRUE 742.662084975409 819.699305298605  0.0808903807902275
#> 20            TRUE 322.056464912455 369.738938003164     8.8342692284329
#> 21            TRUE 324.865009331503 348.094303252544    4.24152726509195
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2450921462219 3372.66467641168 3451.66313355505     78992.574368307
#> 2   34.5585428406392  1414.2190630516  1514.5344724893    40868.7438237272
#> 3   22.8986532984883 2187.54465344506 2200.36644738404    65051.3250106276
#> 4   15.4177790863579 7478.93078150685 7648.54039917478    144695.420697955
#> 5   26.5466438431378 2308.98070098971 2406.98005544208    55196.0619632276
#> 6   13.2097534568683 4804.71455955349 4835.26311211506    126571.353049579
#> 7   7.66600536600061 16936.5435376074 17201.4247491223    356615.484996206
#> 8    14.342455772118  5114.1897188809 5255.39750412163    115591.488790763
#> 9    5.1952627438905 7730.28891602208 7799.61853001244    215229.158336827
#> 10  1.25998648006851 19974.8490787082 20243.5552360059     471707.81989963
#> 11  10.7987512905397 6622.68867624653 6926.15368085608    162040.772334505
#> 12  4.47127145993115 9170.81673957119 9208.40571612518    258246.953510533
#> 13  1.06354720900046 20360.6854090482 20573.5372614072    480169.501820004
#> 14  10.8140559032308 6867.00313960988 7045.27731052793    171158.307543252
#> 15  3.72380442225717  8975.5528734611 9034.89220297576    256143.210374112
#> 16 0.859024363853613 20517.1123610205 20814.9154960747    496210.824961478
#> 17  9.23046379942821  6846.6047408926 7023.52326748678     174181.33297367
#> 18  5.84151021174129 9308.00872862831 9332.61900606706    267106.201327933
#> 19  1.12430797161405 19646.5055988045 19922.6704568896    476620.705972183
#> 20  10.0003100458466 7018.14599319287 7263.73297865691    176529.739421665
#> 21  5.95277342953159 9238.82842564146 9268.80204180342    262343.612687957
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84136.2161254545 86.6574440065764 96.0990941106792    14.4148884637613
#> 2     44469.9954851095 26.5669004608035 38.0856172598732    27.0756781454837
#> 3     68050.7927814128 62.2809631027097 67.9863470566009    18.8853863375802
#> 4     154167.170928011 237.645244095529 247.755001745587     14.263696849268
#> 5     58709.6629428834 60.6172926297429 67.8616103201046    18.0120252140301
#> 6     134232.289769722 200.214701360369 212.251022692145    11.8884516711561
#> 7     371650.630774621 477.654050125186 529.241500100594    6.08276764703703
#> 8     122905.408608918 184.805818491901 219.749429983029    12.1812300630113
#> 9     226404.208800567 279.475683970275  299.01530914923    4.07420091770818
#> 10    499885.997327836 610.352863360068 743.069727522486  -0.201418254898809
#> 11    172808.319997737 302.792098407477 354.166229518289    8.87757355515782
#> 12    274404.137032329 321.556642642498 344.522892501848     3.0765769026242
#> 13    504681.426841268 682.865379157685 776.555626830743   0.226837452302164
#> 14    181889.458090074 316.038145986164 369.354566271234    8.65179609466446
#> 15    272794.832116337 309.005530070235 328.815305976726    2.75573240213228
#> 16    529691.750620565 737.810115275768 822.188790415854  -0.180397356766467
#> 17    180708.673034775 300.168564076721 360.346471101838    6.54738026609178
#> 18    274985.388192389 337.701567248262 366.802490820628    4.27843397399767
#> 19    523780.476770064 735.286180901912 827.075209372102 -0.0190113034375848
#> 20    184812.029061753 317.491121744195 374.304281171424    8.72262702251031
#> 21    275073.626356848 322.640927998637 350.318384585409    4.07768454722007
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4050128563398 3365.10099434476 3459.22681562197    78500.0980298568
#> 2   35.2123853868178 1404.61439619055 1524.13913935036    40523.9431327438
#> 3   23.2493271106065 2186.31703487644 2201.59406595266    64764.1419261907
#> 4   15.5186212235794 7462.69156279396 7664.77961788766    143788.550995078
#> 5   27.2923872184968 2299.59778407406 2416.36297235774    54859.6533587925
#> 6   13.3252070109596 4801.78969813802 4838.18797353053    125837.859108289
#> 7   7.80434652591005 16911.1825705475 17226.7857161822    355175.949762103
#> 8   14.5313007369915 5100.66982454934 5268.91739845319    114891.219872004
#> 9   5.29321960248895 7723.65097425704 7806.25647177748    214159.206696682
#> 10  1.38768203943458  19949.121893435 20269.2824212791    469009.909294802
#> 11  10.9666211897478 6593.63351623072 6955.20884087188    161009.836919941
#> 12  4.59313797464729 9167.21779500751 9212.00466068886    256699.989130787
#> 13  1.13665777026536 20340.3059763755 20593.9166940799    477822.615381798
#> 14  11.0029912263094 6849.93433601134 7062.34611412647    170130.856958981
#> 15  3.80839323954964  8969.8714482948 9040.57362814206     254548.90616475
#> 16 0.949847620995173 20488.5992948982 20843.4285621969    493005.204419651
#> 17   9.4649079916615  6829.6657330272 7040.46227535218    173556.374882713
#> 18  5.97808968882568 9305.65242546928 9334.97530922609     266351.81109623
#> 19  1.22420965584186 19620.0642826049 19949.1117730892     472105.40876813
#> 20  10.1119522517692 6994.63234564844 7287.24662620134    175736.754243358
#> 21  6.11661614740347 9235.95861132808 9271.67185611679    261124.781591999
#>    ci95.integral.bt.up
#> 1     84628.6924639048
#> 2     44814.7961760929
#> 3     68337.9758658497
#> 4     155074.040630889
#> 5     59046.0715473186
#> 6     134965.783711012
#> 7     373090.166008725
#> 8     123605.677527677
#> 9     227474.160440713
#> 10    502583.907932664
#> 11    173839.255412301
#> 12    275951.101412076
#> 13    507028.313279474
#> 14    182916.908674344
#> 15    274389.136325698
#> 16    532897.371162393
#> 17    181333.631125731
#> 18    275739.778424092
#> 19    528295.773974116
#> 20     185605.01424006
#> 21    276292.457452806