Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.4838724755476   15.522133586646
#> 2                 FALSE          0.75 32.2657756900568  31.6172442385275
#> 3                 FALSE          0.75 64.9690315194853  20.1401358015857
#> 4                 FALSE          0.75 242.255936557539  14.8143973516624
#> 5                 FALSE          0.75 64.5121912550627  22.2209465669147
#> 6                 FALSE          0.75 205.857959731119    12.36275730914
#> 7                 FALSE          0.75 507.488800131037  7.13043376091631
#> 8                 FALSE          0.75 201.056455857591  13.4307567101668
#> 9                 FALSE          0.75 287.415673432613  4.61666075381622
#> 10                FALSE          0.75 670.702325720351 0.635131290687857
#> 11                FALSE          0.75 330.252315038261  9.95150904343145
#> 12                FALSE          0.75 335.789489704038  3.94427103967079
#> 13                FALSE          0.75 722.066502625714 0.668481064821459
#> 14                FALSE          0.75 344.376857417962  9.64841252105743
#> 15                FALSE          0.75 320.172311271619  3.57545681327022
#> 16                FALSE          0.75 762.506948394847 0.406231379269958
#> 17                FALSE          0.75 335.306926550345  8.23565630087212
#> 18                FALSE          0.75 352.046668198729  5.03530526402424
#> 19                FALSE          0.75 773.723783965488 0.637353238260386
#> 20                FALSE          0.75 350.985693531526  9.54731400045651
#> 21                FALSE          0.75 338.746829199473  5.16100176579387
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.66094470406 2116.65804258526 81591.7918585175  2.1777176244012
#> 2  1455.99524886197 267.407500792086 42468.1570817238 2.01018147270649
#> 3  2195.57905528461 684.031706689326 66027.0378546819 1.62040482247871
#> 4  7546.28120990641 6064.25924654058 148108.393985983  2.9053668316084
#> 5  2356.68292412346  1113.1735899912 56624.9142948927  2.6664342483458
#> 6   4817.8804591801 3234.75871717173  129952.90699326 2.63873744598917
#> 7  17071.3910740439 15841.5361968659 363362.256541645 11.6320428463372
#> 8  5181.91258028887 3765.06928007925 118847.180819396 9.01648716574134
#> 9  7776.75834375101 6299.20891950853 220866.150557086 5.50631315730135
#> 10 20098.5974430476 18706.3850776603  484640.28483959 22.6967592667298
#> 11 6786.58555969387 5536.48595556876  168215.36633254 15.7400597033961
#> 12 9188.10299691264 7642.70873923474 267511.854435829 7.69642910319388
#> 13 20449.3591568439 19431.2692778765 491585.768119173 27.3753694555103
#> 14 6961.48856851946  5714.0631501154 176411.915407619 16.1586533997238
#> 15 9006.25051752875 7343.23288593298 262490.707064566 4.07416251591154
#> 16 20675.9889446136 19390.3792937985 513092.467817669 17.7853837467917
#> 17 6888.47708992756 5969.11938701926 175743.886845211 9.70960548500226
#> 18 9321.52008445965 7650.07600334565 269424.299030373 6.89401840289447
#> 19 19804.7707554896 18606.4942179382 502330.271799398 20.2912946072814
#> 20 7131.45686101362 5791.58859922845 180166.748069058 16.0324559836637
#> 21 9255.52962121537 7499.12688017558 269383.929988277 5.57814169935054
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.620285310087869  29.658747283282 60.7790212568031 1718.14582991042
#> 2   2.59487354979838 22.5952231818039 91.2615877613792 989.353434450588
#> 3   1.74172156474974 5.41419238968042 113.847282782486  1077.0595405946
#> 4  0.246027074830549 69.7963600499976 85.0223591429269 3622.11754385074
#> 5   2.81441080441497 34.3320355991445  124.79855812129 1420.51830309369
#> 6  0.343390156982837 10.0856752732097 79.3605050933068 2552.03308918295
#> 7  0.435414007477459 104.657768223994 121.630407681042 5539.71684219342
#> 8  0.781758714071179 48.8257226072213 113.862976863467 2335.92115584196
#> 9  0.254440690556281 16.0224954536282 60.8666457885835 3231.66910182521
#> 10 0.394113657581138 95.3506056628428 208.224378269235 9556.54974033794
#> 11 0.675849771054671 67.5677911291572 172.198070177403 2640.60618975292
#> 12 0.391540123019319 16.8406633449662 81.7528083318228 2974.10826319283
#> 13 0.291264290342502 68.5562522010113 200.018434341502 8125.78409215182
#> 14 0.579873508137613 49.1900551038434  166.36367309954  3238.4107444672
#> 15 0.515785413752254 13.6725767174762 159.896424995718  7200.4702285552
#> 16 0.233747446352687 86.8326933358328 183.273062148942 9215.80203991677
#> 17 0.406606545950251 89.5841976869032 106.589326836892 3660.53790063432
#> 18 0.384981245158853 14.5981348565743 99.7330879278761 6764.95145638888
#> 19 0.236980960555564 98.9834895152442  214.69807049881 9850.21343820779
#> 20 0.418212847884959 67.8309643343025 107.092534821461  2581.3349784258
#> 21 0.432821244839671 13.7364361603344 120.074163959677 3030.11095959183
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.9015269834076 95.0662179676876    14.5017642515515
#> 2             TRUE 28.9590271674546 35.5725242126589    27.3486772491092
#> 3             TRUE 62.3034655865078 67.6345974524627    17.2750038275723
#> 4             TRUE 237.476608119543 247.035264995534    14.4096828135661
#> 5             TRUE 60.1259069165339 68.8984755935916    17.5912407936521
#> 6             TRUE 201.517236632467 210.198682829771    11.7978805009033
#> 7             TRUE 488.354089648812 526.623510613262    6.41417771861589
#> 8             TRUE 186.224334469947 215.888577245236    12.1447636255197
#> 9             TRUE 278.357788288853 296.473558576374    4.19810581785114
#> 10            TRUE 633.366156726581 708.038494714122 -0.0131856760331157
#> 11            TRUE 304.359916826175 356.144713250348    8.83973617004652
#> 12            TRUE 323.128863829284 348.450115578792    3.30018753730401
#> 13            TRUE   677.0340198714 767.098985380029   0.189351307208043
#> 14            TRUE 317.795872575417 370.957842260508    8.69452060017106
#> 15            TRUE 313.470313932944 326.874308610293    2.72698980764777
#> 16            TRUE 733.249992131374 791.763904658319  0.0217168300197878
#> 17            TRUE 319.334625527516 351.279227573174    7.56678853278396
#> 18            TRUE 340.706007925968 363.387328471491    4.40201111573793
#> 19            TRUE  740.34460433651 807.102963594466   0.247519558146484
#> 20            TRUE 324.612303438399 377.359083624653    8.85935386568575
#> 21            TRUE 329.570786104041 347.922872294904    4.44901081803261
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5425029217406 3363.87230542306 3461.44958398506    78765.4419683148
#> 2   35.8858112279459 1418.82610672791 1493.16439099604    40840.6706820526
#> 3    23.005267775599 2186.67270880359 2204.48540176564    64255.2749104038
#> 4   15.2191118897586 7431.46619762417 7661.09622218866    142150.010626348
#> 5   26.8506523401773 2300.20672556287 2413.15912268406    54288.1616863036
#> 6   12.9276341173768 4801.28952335567 4834.47139500453    125754.812561554
#> 7   7.84668980321673 16899.2290453155 17243.5531027724    354249.422336237
#> 8   14.7167497948139 5101.59426659999 5262.23089397775    115004.590518036
#> 9   5.03521568978131 7750.40133872979 7803.11534877223    215550.054884584
#> 10  1.28344825740883 19941.7456967322 20255.4491893629    468919.760516734
#> 11  11.0632819168164 6675.43654328641 6897.73457610134    163871.569150397
#> 12  4.58835454203757 9160.40010571018 9215.80588811511    262619.446342877
#> 13  1.14761082243487 20336.5841219732 20562.1341917146    478218.853287584
#> 14  10.6023044419438 6880.57092787363 7042.40620916528     171084.72973297
#> 15  4.42392381889268  8983.7591288285 9028.74190622899    250645.933538593
#> 16 0.790745928520128 20533.1491640762 20818.8287251511    497932.473462006
#> 17  8.90452406896028  6741.1110847326 7035.84309512251    169722.301998668
#> 18  5.66859941231055 9297.50615262058 9345.53401629871    258295.953884614
#> 19  1.02718691837429  19641.942915237 19967.5985957422    486126.670693546
#> 20  10.2352741352273 7019.87492468369 7243.03879734355    175920.452029548
#> 21  5.87299271355513 9232.93318373162 9278.12605869912    264399.397459748
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84418.1417487201 87.2155459317213  95.752199019374    14.3063743788738
#> 2      44095.643481395  28.325820003552 36.2057313765615    26.5312920809227
#> 3       67798.80079896  61.793038067427 68.1450249715435    16.7263615346762
#> 4     154066.777345617 236.561417567586 247.950455547491    14.3321842849945
#> 5     58961.6669034819  59.285980128305 69.7384023818205    16.7047013902613
#> 6     134151.001424966  200.68603433698 211.029885125257    11.6897126014537
#> 7     372475.090747053 484.689996152216 530.287604109858    6.27702230626049
#> 8     122689.771120756 183.384141012738 218.728770702444    11.8985096305873
#> 9     226182.246229589 276.623299644303 298.208047220924    4.11795700032591
#> 10    500360.809162446 626.216677557561 715.187973883142  -0.137331478171174
#> 11    172559.163514684 299.401798019605 361.102832056918    8.62684349216429
#> 12    272404.262528781 320.704488661778 350.874490746298    3.17685239855292
#> 13    504952.682950763 668.410778492914 775.722226758515  0.0976030557501547
#> 14    181739.101082267 312.705896754504 376.047818081421    8.51186044510771
#> 15    274335.480590539 312.186952740432 328.157669802805    2.56451740231581
#> 16    528252.462173332 727.647596251135 797.366300538558 -0.0519136155813086
#> 17    181765.471691755 316.276099799741  354.33775330095    7.43870747080963
#> 18    280552.644176133 338.534392129056 365.558944268402    4.28074202351289
#> 19    518533.872905249 733.952846535216 813.494721395759   0.172870555571481
#> 20    184413.044108569 319.562079803545 382.409307259507    8.72761681860199
#> 21    274368.462516805 327.813671468746   349.6799869302    4.31267212590811
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.7378927944183 3354.52980002883 3470.79208937929     78224.226031893
#> 2   36.7031963961323 1411.70861142564 1500.28188629831    40529.0243502006
#> 3   23.5539100684952 2184.96723820084 2206.19087236839    63916.0011551165
#> 4   15.2966104183303 7409.48034420842 7683.08207560441    141009.043600035
#> 5    27.737191743568 2289.39213434914 2423.97371389779    53840.6984208291
#> 6   13.0358020168264 4798.11253564461 4837.64838271559    124950.922138462
#> 7   7.98384521557213 16866.2618483249  17276.520299763    352504.411530946
#> 8   14.9630037897463 5086.21416397872 5277.61099659902    114268.775353946
#> 9   5.11536450730653  7745.3542526619 7808.16243484012    214532.079117509
#> 10  1.40759405954689 19911.7102559484 20285.4846301467    465909.447348527
#> 11  11.2761745946986 6654.15268908073 6919.01843030702    163039.778200624
#> 12  4.71168968078866 9155.09529675651 9221.11069706878    261682.602239971
#> 13  1.23935907389276 20314.9889025299 20583.7294111579    475659.231298556
#> 14  10.7849645970072 6865.07606051592 7057.90107652299    170064.630348463
#> 15  4.58639622422464 8979.45226716249   9033.048767895    248377.785416598
#> 16 0.864376374121225 20505.7968656754 20846.1810235519    495029.495819432
#> 17  9.03260513093461 6712.89206246123 7064.06211739389    168569.232559968
#> 18  5.78986850453559 9292.90774014076 9350.13242877854    256164.994175851
#> 19  1.10183592094929 19610.7631160397 19998.7783949395     483023.85346051
#> 20   10.367011182311 6998.50817091839 7264.40555110885    175107.331511344
#> 21  6.00933140567962 9228.60620634111 9282.45303608962    263444.912507477
#>    ci95.integral.bt.up
#> 1     84959.3576851419
#> 2     44407.2898132469
#> 3     68138.0745542473
#> 4      155207.74437193
#> 5     59409.1301689564
#> 6     134954.891848059
#> 7     374220.101552344
#> 8     123425.586284846
#> 9     227200.221996664
#> 10    503371.122330652
#> 11    173390.954464456
#> 12    273341.106631687
#> 13    507512.304939791
#> 14    182759.200466774
#> 15    276603.628712534
#> 16    531155.439815906
#> 17    182918.541130454
#> 18    282683.603884895
#> 19    521636.690138285
#> 20    185226.164626773
#> 21    275322.947469077