Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xls', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xls', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.5585755264355  15.8219290415476
#> 2                 FALSE          0.75 32.0816627213751  31.9916958431978
#> 3                 FALSE          0.75 65.1155524647443  20.5670769880084
#> 4                 FALSE          0.75 243.453526835174  15.0752321852233
#> 5                 FALSE          0.75 66.3330643081115   22.905911784185
#> 6                 FALSE          0.75 205.125698456448  12.2883198472092
#> 7                 FALSE          0.75 509.730721104181  7.07633264729713
#> 8                 FALSE          0.75 204.089606725552  13.4540825286527
#> 9                 FALSE          0.75 288.979690676426  4.60872346573046
#> 10                FALSE          0.75 671.943722444097 0.621307599347976
#> 11                FALSE          0.75 324.568940487174  9.83054416403538
#> 12                FALSE          0.75 334.054218397379   3.8435106019823
#> 13                FALSE          0.75 725.251959340483 0.730192142937895
#> 14                FALSE          0.75 341.628123274663   9.8006248139015
#> 15                FALSE          0.75 319.206753683041  3.30541122507671
#> 16                FALSE          0.75 775.367848937967 0.441082986115318
#> 17                FALSE          0.75 336.751627089425  8.18996990902176
#> 18                FALSE          0.75 348.408414244298  4.86672568652171
#> 19                FALSE          0.75 779.608263416418 0.674894543805976
#> 20                FALSE          0.75  350.37608709644  9.52944416353472
#> 21                FALSE          0.75 335.657009677391  5.12799259525195
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3419.02574818854 2101.19245503274 81731.4966647145 2.34278853224388
#> 2  1463.85742414989 266.022719919847 42459.3215835423 1.95669110786235
#> 3  2195.80514628473 655.891402612021 66256.1973125879 1.97502998807774
#> 4  7568.32667617774 6034.42766853388  149610.08481083 2.02158379263322
#> 5  2352.63130544154  1099.6982472787 56565.5950041096 2.42663073096227
#> 6  4818.76809892036 3254.01492132786 129698.515706525 2.77144184753983
#> 7  17034.4723804056 15841.9150662225 361377.492470305 12.0664248975686
#> 8  5176.53154453814 3770.81323405603 118286.531759118  7.0134358295044
#> 9  7762.39993626069 6297.51942537457 221152.332697753  5.0648559388091
#> 10 20041.4657451977 18628.1823260407 484418.391049097 33.7008715901412
#> 11 6800.31392873181 5605.39057454382 167999.852529548 13.8366337597263
#> 12 9186.41693103522 7632.93767041625 266733.168668153 6.00327490267064
#> 13 20444.1242779846 19430.4789019212  490531.87562063 21.0248356504732
#> 14 6979.71744223627 5687.79787007808  176961.93444357 10.8970128039511
#> 15 9008.41000389528 7410.41973994455 265573.874213143 6.24081602020162
#> 16 20704.3275409968 19454.4086852454 515382.392660031 23.1840439372895
#> 17 6880.96618281012  5968.7056630638 175232.666977655   13.77777571778
#> 18 9318.69776561437 7675.68778609466 271445.554615526 5.76821264923467
#> 19 19779.4141408392 18623.3237391805 502064.898084063 27.3808702927709
#> 20 7154.42312508729 5815.18472790692 180411.964787275 16.8764092267938
#> 21 9250.87071173989 7452.26414826461 268475.422758527 8.80192010594099
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.646346972485858  14.620385499574 57.7428566663375 1183.36354345484
#> 2   2.98104012493587  20.933174548918 84.2346582465396  905.90109305629
#> 3    1.6199684740961 6.67649585719493 108.321161903211 950.132745987964
#> 4  0.219945297989342 54.6051434640329 71.6782647977975  3299.8185490704
#> 5   3.15450622889706 48.9330245608975 146.005983795649 1941.80750864116
#> 6   0.63642686586618  12.787571344505 127.225750036976 3662.98564893764
#> 7  0.455038834938567 115.097087249495  156.71069575139 6489.26328429363
#> 8  0.640879845012014 56.5596090767506  137.89023325461 3365.87584173782
#> 9  0.310461972490081 22.0224762976826 65.8882112364582 2198.86476884546
#> 10 0.442740990177447 65.0977934295886 252.390788425261 6507.78570475446
#> 11 0.821667210919579  68.668495112258 236.166280612683 2365.88480684652
#> 12 0.573977285471472  13.766313593949 166.625218660595 6078.65479280829
#> 13 0.185250426095536 172.745735196307 205.877050940352 14711.2862517058
#> 14 0.631923138978102 48.1972320750201  159.51956487149 2396.85915560991
#> 15 0.381483820956965 18.3756529769925 79.8814121384096 3887.78091793016
#> 16 0.257941559822965 114.221403123381 173.653108617292 6715.77036613806
#> 17 0.513442585821074 95.9664319091942 170.514322370664 3740.65938018388
#> 18 0.255548021954769 8.24364828806918 51.4702566473505 3366.57072700698
#> 19 0.385689567394637 68.4295374722388  273.86814736757 8949.26515347508
#> 20 0.451096307262992 87.5445036917519  194.73967134264 3115.59794707002
#> 21  0.51674949502251 9.72664118897103 148.352460420405 4441.71199007199
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 88.7046883908944 96.4124626619767   14.7586882718083
#> 2             TRUE 28.8629058489415 35.3004195938086   27.0878848376782
#> 3             TRUE 61.8666281343564 68.3644767951322   17.9022288481203
#> 4             TRUE 240.128021496293 246.779032174056   14.7134221700308
#> 5             TRUE 62.3412567556786 70.3248718605445   17.7167490376494
#> 6             TRUE 200.566676617245 209.684720295651   11.2413976528594
#> 7             TRUE  489.88145214768 529.579990060681   6.32779376382319
#> 8             TRUE 192.552504786017 215.626708665087   12.3998351836079
#> 9             TRUE 280.648002657085 297.311378695767   4.09801352098427
#> 10            TRUE 616.505788678315 727.381656209879 -0.107001329493924
#> 11            TRUE 301.807677952425 347.330203021924   8.47890160207267
#> 12            TRUE 324.178831182486 343.929605612272   2.89931796738173
#> 13            TRUE 690.666104695454 759.837813985511  0.425455192010738
#> 14            TRUE 323.702537212163 359.553709337162   8.76111125028252
#> 15            TRUE 308.940611329809 329.472896036273    2.6778703396025
#> 16            TRUE 737.230096661126 813.505601214808 0.0167691202065408
#> 17            TRUE 314.087186033677 359.416068145173    7.3453568553461
#> 18            TRUE 338.919704436307 357.897124052289   4.44634919040611
#> 19            TRUE  734.56673178481 824.649795048026 0.0404352054417985
#> 20            TRUE 322.614393918364 378.137780274515    8.7873907380871
#> 21            TRUE 321.177851103118 350.136168251664   4.27793967593992
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.8851698112868 3394.97521404174 3443.07628233534    79784.8636357313
#> 2   36.8955068487173 1429.42235201692 1498.29249628286    40969.1142854647
#> 3   23.2319251278965 2184.82231059965 2206.78798196982    64693.2289454377
#> 4   15.4370422004157 7478.50121517941 7658.15213717608    144181.883297609
#> 5   28.0950745307207 2272.13648003887 2433.12613084422    53371.3216523949
#> 6   13.3352420415591 4797.73254405865 4839.80365378207    123672.904314022
#> 7   7.82487153077108 16845.1376718802  17223.807088931    350702.654367642
#> 8   14.5083298736974 5083.49098760688 5269.57210146939     112749.66599946
#> 9   5.11943341047664   7726.172962751 7798.62690977038    217535.200153002
#> 10  1.34961652818988  19934.379875006 20148.5516153893    473713.083564776
#> 11  11.1821867259981 6687.35425427214 6913.27360319147    164107.972022286
#> 12  4.78770323658287 9163.77134517318 9209.06251689727    256733.781533983
#> 13  1.03492909386505 20159.9575435867 20728.2910123825    466331.809736574
#> 14  10.8401383775205 6900.43299547286 7059.00188899968    173019.101132592
#> 15  3.93295211055092 8978.18205474813 9038.63795304243    259178.474603148
#> 16 0.865396852024095 20516.4333328588 20892.2217491347    504334.950407734
#> 17  9.03458296269743 6723.10140231949 7038.83096330074    169079.282297252
#> 18   5.2871021826373 9305.13696418049 9332.25856704824    265907.545769599
#> 19  1.30935388217015 19666.8475516974  19891.980729981    487343.356906597
#> 20  10.2714975889823 7010.41241651435 7298.43383366022    175286.806164344
#> 21  5.97804551456398 9234.87038698403 9266.87103649574    261168.806534859
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83678.1296936977 87.9667100032375 97.1504410496335    14.5550889754753
#> 2     43949.5288816199 28.2465481499649 35.9167772927853    26.1488571983235
#> 3     67819.1656797381  61.244493688112 68.9866112413767      17.39193877878
#> 4     155038.286324051 239.491222601613 247.415831068736    14.6441394011642
#> 5     59759.8683558243 61.5768680754255 71.0892605407976    16.7230795755468
#> 6     135724.127099027 199.693672435269 210.557724477626    11.0409231901115
#> 7     372052.330572968 486.080528304946 533.380913903415    6.18445653081754
#> 8     123823.397518777 190.343272499724 217.835940951381    12.1979580324291
#> 9     224769.465242504  279.05257303636 298.906808316492     4.0002179996499
#> 10    495123.698533418  605.89001412742 737.997430760774   -0.24646474139982
#> 11    171891.733036811 297.449138318111 351.688742656238    8.22007643063301
#> 12    276732.555802322 322.287799588144 345.820637206613    2.71851512245821
#> 13    514731.941504686 684.043281465555  766.46063721541   0.367101307790644
#> 14    180904.767754549 320.269978178918 362.986268370407    8.56205546150442
#> 15    271969.273823138 306.974754283446 331.438753082636    2.55770293600106
#> 16    526429.834912328 729.927122820879 820.808575055054 -0.0644824711376931
#> 17    181386.051658057 309.747186682576 363.756067496274    7.18362244081246
#> 18    276983.563461452 337.102717451798 359.714111036798    4.36585156349036
#> 19     516786.43926153 725.941757642587 833.274769190249 -0.0810570082875123
#> 20    185537.123410205 317.298325011924 383.453849180955    8.64529540129925
#> 21    275782.038982196 318.405246269747 352.908773085036    4.11516358500783
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0887691076198 3390.36979260937  3447.6817037677     79412.104119543
#> 2   37.8345344880721 1422.82840203401 1504.88644626577     40683.755441152
#> 3   23.7422151972367 2182.71921440463 2208.89107816483    64393.9371304515
#> 4   15.5063249692824 7461.30059498824 7675.35275736725    143142.440454652
#> 5   29.0887439928233 2256.72257730218  2448.5400335809     52759.652287173
#> 6   13.5357165043069 4793.70445908513 4843.83173875559    122519.063834607
#> 7   7.96820876377672 16808.8820893966 17260.0626714146    348658.536433089
#> 8   14.7102070248762 5065.67471074771 5287.38837832857    111689.415109312
#> 9   5.21722893181101 7719.23588271723 7805.56398980415    216842.557750816
#> 10  1.48907994009577 19913.8740700757 20169.0574203196    471663.131067778
#> 11  11.4410118974378 6665.72367831178 6934.90417915183    163362.718308129
#> 12  4.96850608150638 9159.43495639108 9213.39890567936    254819.005274248
#> 13  1.09328297808515 20105.5426369998 20782.7059189694    461697.754567286
#> 14  11.0391941662986 6885.25086736923 7074.18401710331    172264.090498575
#> 15  4.05311951415236 8972.39372406037 9044.42628373018       257953.823614
#> 16 0.946648443368329 20480.4535908749 20928.2014911186      502219.4827424
#> 17  9.19631737723107  6692.8719762681 7069.06038935214    167900.974592494
#> 18  5.36759980955306 9302.54021496975 9334.85531625898    264847.075990592
#> 19  1.43084609589947 19645.2922473936 19913.5360342848    484524.338383252
#> 20  10.4135929257702 6982.83589785145 7326.01035232312    174305.392811017
#> 21  6.14082160549607  9231.8064950095 9269.93492847027    259769.667257986
#>    ci95.integral.bt.up
#> 1      84050.889209886
#> 2     44234.8877259327
#> 3     68118.4574947243
#> 4     156077.729167008
#> 5     60371.5377210463
#> 6     136877.967578442
#> 7      374096.44850752
#> 8     124883.648408924
#> 9      225462.10764469
#> 10    497173.651030416
#> 11    172636.986750967
#> 12    278647.332062057
#> 13    519365.996673973
#> 14    181659.778388566
#> 15    273193.924812286
#> 16    528545.302577662
#> 17    182564.359362815
#> 18    278044.033240459
#> 19    519605.457784874
#> 20    186518.536763532
#> 21    277181.178259068