Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.3930332433692  15.3241277780322
#> 2                 FALSE          0.75 32.2866341716824  31.5303725080563
#> 3                 FALSE          0.75 64.8373378333158  19.6916708623537
#> 4                 FALSE          0.75 242.637079971855  15.0487600560996
#> 5                 FALSE          0.75   64.95106548938  21.5868584704188
#> 6                 FALSE          0.75 205.822764320878  12.3254420780833
#> 7                 FALSE          0.75 507.978321554699  7.03443204651058
#> 8                 FALSE          0.75 200.657102320898  13.3777324741951
#> 9                 FALSE          0.75 287.486396052872  4.71246667437551
#> 10                FALSE          0.75 667.877140554443 0.572438493202163
#> 11                FALSE          0.75 326.234620278321  9.77611614566224
#> 12                FALSE          0.75 336.830276426735  3.88214815609587
#> 13                FALSE          0.75 734.433544715111 0.719345105905239
#> 14                FALSE          0.75 344.061338478519  9.75789020478132
#> 15                FALSE          0.75 319.869058319815  3.39398639804888
#> 16                FALSE          0.75 773.326575626769 0.380455511516415
#> 17                FALSE          0.75 335.308773984538  8.30146972363884
#> 18                FALSE          0.75 350.116815114641  4.86343903937996
#> 19                FALSE          0.75 779.636383194756 0.718807748746892
#> 20                FALSE          0.75 350.280811313257  9.55149189298116
#> 21                FALSE          0.75 338.602572562524  5.01419953056022
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3421.99172656515 2150.39709112902 82153.0659765013 2.60079486974804
#> 2  1465.43957519503  282.44855817255 42646.5745915634 1.56846852302197
#> 3  2196.63255782962 722.229315213481 65407.9896559923 1.60156783293942
#> 4  7583.08875184848 6047.70313028534 150545.017999266 2.08841587850084
#> 5  2354.90153111144 1165.83503750633 57085.6389310688 3.51072087040066
#> 6  4819.53330109492 3247.60456709733 130559.264060975 2.48803861272027
#> 7  17046.3764319159 15846.8943481346 363161.412438296  15.959399838809
#> 8  5195.42381848998 3777.13045764706 119999.582317602  10.987608129101
#> 9  7765.54636439519 6260.69131088477 221385.312629011 5.78456325183911
#> 10 20081.2131360548 18699.6251062257 485681.093142159 20.2490255616646
#> 11 6773.73452960533  5622.8309768514 166933.093488323 7.97146639561214
#> 12 9191.56253444658 7669.02536568903 268316.909769389 6.07472034427134
#> 13 20469.5125641144 19482.4287049035 493552.901500029 27.5305136258429
#> 14 6965.37967712567 5654.26974418977 176002.254542658 12.9695953159488
#> 15 9007.61945481182 7396.41719615838 264427.210836024 6.16490117312707
#> 16 20718.5783022692 19474.0821837178 514507.612168367 18.2706542359389
#> 17 6899.12346062568 5946.35178070229 176182.511043098 13.2143662659216
#> 18 9321.08683840652  7681.4813418503 270010.443279266 7.53246436741343
#> 19 19790.0825786619 18593.4477087579 505016.017384577 28.5799023453399
#> 20 7086.76427682772 5749.02929496168 178413.359241484 14.7705538767951
#> 21 9254.99080293627 7532.63707943969 267603.136571033 4.71051086085941
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.545011369157071 11.8901235886375 47.0836119918846 864.180751791808
#> 2   2.34279121589336 23.0446593355137 68.7116475699733 688.430194944697
#> 3    2.2926373922006 4.71836482509638 152.385285882231 1769.49084042387
#> 4  0.292545592596362 34.8920847376557 65.1265573523876 2030.71480582391
#> 5    2.5887655831468 29.8143916087544 140.996652001955 1028.28939814138
#> 6   0.58592023493478 8.09441045356669 115.939614500966 2099.37944766668
#> 7  0.608221415167448 94.1535270431215 113.059756545118 5929.73867152376
#> 8  0.815694175220484 36.1823679843516 128.486350669352 2244.47172479389
#> 9   0.31500239254223 20.9116318136443 64.6676832942823 3212.63838076519
#> 10 0.355666173437185 73.2258314383093 233.533829629656 5844.02563894202
#> 11 0.522506762983132  86.875286943638  148.30901805104  3110.9969989418
#> 12 0.319010837686952 13.8773450711818 82.4923482368223 2412.35001672489
#> 13 0.271001529347293 73.5160273621443 224.153949747032 6818.39731872355
#> 14 0.733461508911461 66.2419362427048 204.907875516961 3830.45335170529
#> 15 0.339815622202614 17.1807448704306 81.7286871118657 4493.40000514642
#> 16 0.209602805372575 77.7928828459883 154.308720848615 8355.72227479618
#> 17 0.591851505621007 83.3236704078745 154.206696491743 3018.61129003803
#> 18 0.317522303549689  10.038836212363 42.0837085583153 3652.07264114065
#> 19 0.277809150123293 93.3998226454366 172.833335184068 4514.94793143261
#> 20 0.492981222757673 153.210258227956 185.599430368435 6485.40029550431
#> 21 0.374682474321223 8.51377727870514 85.5942424172713 5577.53347611161
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.1147256826336 96.6713408041047    14.4275840757689
#> 2             TRUE 29.7065034513112 34.8667648920535    27.6764809579117
#> 3             TRUE 62.2027587481305 67.4719169185012    15.9202823521837
#> 4             TRUE 239.201635851722 246.072524091989    14.5675225562786
#> 5             TRUE 59.1759296575709 70.7262013211891    17.3283390861423
#> 6             TRUE 201.729940802953 209.915587838803    11.3616032916156
#> 7             TRUE 481.725108819859  534.23153428954    6.03390781856013
#> 8             TRUE 182.582486948527 218.731717693269    12.0359155559574
#> 9             TRUE 277.970789503596 297.002002602147    4.19428773864354
#> 10            TRUE 634.567493505505 701.186787603381 -0.0126323621020062
#> 11            TRUE 313.121558057539 339.347682499103    8.91659252055499
#> 12            TRUE 326.837361460408 346.823191393061    3.35737532810083
#> 13            TRUE   689.1458498006 779.721239629623   0.273547590128942
#> 14            TRUE 322.726354183784 365.396322773255    8.55134602262197
#> 15            TRUE 309.727795890021 330.010320749609    2.83498969952558
#> 16            TRUE 743.271349408649 803.381801844888  0.0356588966785299
#> 17            TRUE 313.571141477097 357.046406491979    7.32787399689228
#> 18            TRUE 337.725911230246 362.507718999036    4.34111485004072
#> 19            TRUE 732.622443836672 826.650322552841   0.261811696794075
#> 20            TRUE 325.983250185929 374.578372440585    8.74053778154479
#> 21            TRUE  330.85378219641 346.351362928637    4.39784686030181
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2206714802956 3402.43247326184 3441.55097986846    80731.4886398038
#> 2   35.3842640582009 1427.53111058811 1503.34803980195    41514.1069208794
#> 3   23.4630593725237 2188.87084769234  2204.3942679669     62497.177223495
#> 4   15.5299975559207 7525.69127245503 7640.48623124192    147204.492143685
#> 5   25.8453778546953 2305.85685691503 2403.94620530784    55394.1028711262
#> 6    13.289280864551  4806.2179958988 4832.84860629104    127105.784869563
#> 7   8.03495627446104   16891.49387993 17201.2589839019    353406.992323639
#> 8   14.7195493924328 5135.90382315572 5254.94381382423    116307.426330316
#> 9   5.23064561010748 7731.14673006174 7799.94599872863    216100.522492652
#> 10  1.15750934850633 19960.7566433388 20201.6696287708    476067.670966099
#> 11  10.6356397707695 6630.82468258305 6916.64437662762    161815.503425063
#> 12   4.4069209840909 9168.73430180449 9214.39076708867    264348.593991876
#> 13  1.16514262168154 20348.5786991037 20590.4464291252    482336.637910729
#> 14  10.9644343869407 6856.41169200642 7074.34766224492    169701.158779103
#> 15  3.95298309657218 8979.35712949996 9035.88178012368    257035.567827558
#> 16 0.725252126354301 20590.6090099876 20846.5475945509    500762.449026327
#> 17  9.27506545038539 6762.05602280473 7036.19089844663    171216.895470986
#> 18   5.3857632287192 9304.57295283718 9337.60072397585    264002.783784589
#> 19  1.17580380069971 19636.4398704101 19943.7252869136     497588.92803737
#> 20  10.3624460044175 6834.73340204273 7338.79515161271    167744.875755379
#> 21  5.63055220081864  9240.9856393128 9268.99596655974    258428.094002829
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83574.6433131989  87.295475298663 97.4905911880753    14.2559054944844
#> 2     43779.0422622474 29.2124358665593 35.3608324768054    26.9385017249053
#> 3     68318.8020884895 61.6982648807545 67.9764107858771    15.1981015736405
#> 4     153885.543854846 238.543784849994 246.730375093717    14.4753706946108
#> 5     58777.1749910114 58.0700525833947 71.8320783953653    16.5128779274511
#> 6     134012.743252386 200.946208639946  210.69932000181    11.1770384176111
#> 7     372915.832552952 476.697897870634 539.258745238765    5.84231807278239
#> 8     123691.738304888  179.12139038786 222.192814253936    11.7789718907629
#> 9      226670.10276537 276.148652079267 298.824140026476    4.09506198499274
#> 10    495294.515318219  628.18905045358 707.565230655305   -0.12466720673472
#> 11    172050.683551582 310.610546142921 341.858694413721     8.7520028902153
#> 12    272285.225546901 324.923824551963 348.736728301507    3.25688691422944
#> 13    504769.165089329 680.473738008459 788.393351421763   0.188182108384545
#> 14    182303.350306213  318.64093165926 369.481745297779    8.32030564731486
#> 15     271818.85384449 307.785852020486 331.952264619144    2.72794777853175
#> 16    528252.775310407 737.516093324329 809.137057929209 -0.0303659870138311
#> 17    181148.126615211 309.408616103332 361.208931865744    7.14144077262166
#> 18    276018.102773942 335.353184954511 364.880445274771    4.24109532442257
#> 19    512443.106731783  723.61977459789 835.652991791623   0.174301814505238
#> 20    189081.842727589 321.330525714739 379.231096911775    8.58524869637612
#> 21    276778.179139236 329.369971275239 347.835173849808    4.27982188089063
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.3923500615801 3398.68708433142 3445.29636879888    80459.2717029894
#> 2   36.1222432912073 1420.27204289743 1510.60710749264    41297.2514094718
#> 3   24.1852401510669 2187.38456277243 2205.88055288681    61939.7876087615
#> 4   15.6221494175885 7514.70026576267 7651.47723793428    146564.816979851
#> 5   26.6608390133866 2296.46532355828 2413.33773866459    55070.1917107117
#> 6   13.4738457385554 4803.66825660593 4835.39834558391    126444.480343548
#> 7   8.22654602023878 16861.8355189114 17230.9173449205    351539.124642109
#> 8   14.9764930576272 5124.50637724065  5266.3412597393    115600.417737006
#> 9   5.32987136375828 7724.55956604045 7806.53316274993    215088.541402711
#> 10  1.26954419313905 19937.6905064357 20224.7357656739    474226.802889833
#> 11  10.8002294011092  6603.4589671958 6944.01009201486    160835.539370397
#> 12  4.50740939796229 9164.36293810706  9218.7621307861    263588.703736608
#> 13  1.25050810342593 20325.4211504846 20613.6039777442    480188.842755331
#> 14  11.1954747622478 6835.54548208997 7095.21387216138    168494.565973315
#> 15    4.060025017566 8973.94519486577 9041.29371475786    255620.146825937
#> 16 0.791277010046662 20566.1042518911 20871.0523526474    498130.396509767
#> 17  9.46149867465601 6735.80906662625 7062.43785462512    170266.032914624
#> 18  5.48578275433735 9301.41071943029 9340.76295738275     262852.38090263
#> 19  1.26331368298855 19607.0189262768 19973.1462310469    496166.719438969
#> 20  10.5177350895862 6786.47217070093 7387.05638295451    165701.974662296
#> 21  5.74857718022982    9238.30379947 9271.67780640253    256671.170957854
#>    ci95.integral.bt.up
#> 1     83846.8602500133
#> 2      43995.897773655
#> 3     68876.1917032231
#> 4     154525.219018681
#> 5     59101.0861514259
#> 6     134674.047778401
#> 7     374783.700234482
#> 8     124398.746898198
#> 9     227682.083855311
#> 10    497135.383394485
#> 11    173030.647606248
#> 12     273045.11580217
#> 13    506916.960244727
#> 14       183509.943112
#> 15    273234.274846111
#> 16    530884.827826968
#> 17    182098.989171573
#> 18    277168.505655901
#> 19    513865.315330185
#> 20    191124.743820672
#> 21    278535.102184211