Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.7166289206768   15.703837831056
#> 2                 FALSE          0.75 33.2487131814243  31.4830140613572
#> 3                 FALSE          0.75 65.1062802494435   20.259153964728
#> 4                 FALSE          0.75 241.967929638936  15.0247267166884
#> 5                 FALSE          0.75 65.4094737581693  22.6230066182713
#> 6                 FALSE          0.75 205.221527478942  12.4083104148702
#> 7                 FALSE          0.75 507.868369486985  6.94524076017509
#> 8                 FALSE          0.75 202.551153350899   13.398118295698
#> 9                 FALSE          0.75 290.816160399539  4.70749691128506
#> 10                FALSE          0.75 667.569820124192 0.672223449283225
#> 11                FALSE          0.75  323.32461573518  10.0689691187745
#> 12                FALSE          0.75 336.958816954008  3.89188334892195
#> 13                FALSE          0.75 722.332831573936 0.635089729150356
#> 14                FALSE          0.75  341.69607916002  9.95698289212606
#> 15                FALSE          0.75 316.204917442566  3.23043166486591
#> 16                FALSE          0.75 778.688227846384 0.524483302652297
#> 17                FALSE          0.75 335.295905836927  8.02111492525818
#> 18                FALSE          0.75 349.742684038523  5.11651750184913
#> 19                FALSE          0.75 780.363349419388  0.64683755606034
#> 20                FALSE          0.75  353.97035693641  9.57641540988689
#> 21                FALSE          0.75 342.732944799538  5.72840975800334
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3419.13238368982 2117.79715249564 81892.9927330668 1.40992264637315
#> 2  1461.13270962988 282.082853846115 42389.7621217251 2.39244946841294
#> 3  2194.46575283642 673.819185902294 65971.6027371339   1.714221799514
#> 4  7550.51606661654 6014.53668988923 148499.520703292 2.84710949050526
#> 5  2363.53760500093 1135.50968462091 57434.9285893303 2.59242744727063
#> 6  4821.24751280209 3227.28550943829 131215.278767334 3.03758914244817
#> 7  16993.7643170544 15829.9878826509 359904.795523942 16.8203566872769
#> 8   5190.0370631287 3799.26163527553 118700.322726895 10.0083131383138
#> 9  7763.88957343238  6275.4137979894 221100.981699963 6.93852983388707
#> 10 20087.7457905699 18685.3818160871 483843.715140037 29.0685493218744
#> 11 6789.76524393305  5512.0358490899 167712.605390356 11.5157538421943
#> 12 9193.48706384169 7661.42874378372 265301.451491533 7.17190742066723
#> 13 20456.6347111349 19456.0186241771  491149.37203189 17.4477010509677
#> 14 6983.68248697013 5594.19232581176 177440.309326178 11.6352093459495
#> 15 9004.99373225435 7384.34652243013 265556.050838879 5.19074603933362
#> 16 20681.8718602156 19347.5292380321 515488.976269278 23.4237724166469
#> 17 6895.22531626718  6029.2443850239 176011.673601255 13.3737756787312
#> 18  9316.5728357454  7595.0564975359 269652.180903921 6.78818483579151
#> 19 19787.7205300951 18626.9613542434 501001.674021463 22.2304906103641
#> 20 7117.66368374925  5782.8119400756 179371.953956139 14.2451338670731
#> 21 9257.37121180458 7348.07954613289 267934.658441966 8.72399615927604
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.617288070504899 17.3734498574926 62.9300609078365 1130.41697612493
#> 2   2.62679697958755 27.0271773412166 65.0100077862577 928.123917089892
#> 3   2.33146992772741 5.92789194125936 165.108000862796 1345.33525988182
#> 4  0.422312063748936 71.3508244654295 114.047298199023 3961.23821007489
#> 5   2.19615452750304 40.2350678677532 127.757513061439 1327.47793601611
#> 6  0.545140046391317 8.58359652268095 109.605487670858 1445.05498006619
#> 7    0.5716163270226 198.541354899836  215.12626968886 9530.00092563852
#> 8  0.513892866442058 37.9529831546048 109.459761761512 2437.34795617633
#> 9  0.302636406559703 26.6324574751625 67.5094081967905  3688.9840201879
#> 10  0.43017643137095 80.1901470953987 266.633294776649 8114.76595240084
#> 11 0.681498879716117 71.4339515874615 190.741974666276 2333.44628615569
#> 12 0.517552800052566 13.7688363828235 129.601480999086 4616.34180703806
#> 13 0.251359590633571 80.2555313552921 204.369704816633 8958.22515916859
#> 14 0.611422938964213 45.4412666685769 199.814709676816 1622.54319848754
#> 15 0.470077843401131 15.2494573376557 129.960138719776 3593.01949961884
#> 16  0.34285485884406 78.3507669766941 230.569347627078 6310.31619798342
#> 17 0.576906918139267 82.5956476581663 144.117807064944 2440.20953226566
#> 18 0.417830501138604 7.33926944369296 142.902086574626    4530.07337109
#> 19 0.201634024165016 51.3056856843256 160.540475511745 12268.4824075933
#> 20 0.377968930426601  109.84542120632 159.975635776186  4525.9415816312
#> 21 0.826265918470275 9.24648140625342  230.48518646458 4093.93269497674
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE  90.397306167393 95.0359516739607    14.6883989550754
#> 2             TRUE  29.313133805885 37.1842925569636    27.1619330299357
#> 3             TRUE  62.286385389243  67.926175109644    16.4238859336164
#> 4             TRUE 237.284434527055 246.651424750817    14.3300233718214
#> 5             TRUE 61.1449306074091 69.6740169089295    19.0103324205288
#> 6             TRUE 200.224693339615  210.21836161827    11.5115550385565
#> 7             TRUE 480.198882736414 535.537856237555    6.00493190222291
#> 8             TRUE 186.087478238373 219.014828463425    12.5527645304008
#> 9             TRUE 279.402278822795 302.230041976283    4.20966002249435
#> 10            TRUE 619.752056489708 715.387583758675 -0.0354167803219877
#> 11            TRUE  304.38120066477 342.268030805589    8.94790346164145
#> 12            TRUE 325.161029247011 348.756604661006    3.04050899283548
#> 13            TRUE 693.631363345094 751.034299802777   0.221603202558132
#> 14            TRUE 322.556159785933 360.835998534107    8.95119215752993
#> 15            TRUE 307.666140207862 324.743694677269    2.45715361247105
#> 16            TRUE    740.156122221 817.220333471768 -0.0395129401461818
#> 17            TRUE 313.296044845414  357.29576682844    7.07210304491908
#> 18            TRUE 338.576119983646   360.9092480934    4.42918632747613
#> 19            TRUE 743.794192365339 816.932506473437   0.315149586308889
#> 20            TRUE 330.537111725074 377.403602147745    8.95465651933513
#> 21            TRUE 328.381971117529 357.083918481547    4.36920232211974
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.7192767070365 3390.55305867425  3447.7117087054    80033.4568073413
#> 2   35.8040950927788 1416.67300290358 1505.59241635618    40862.9982781122
#> 3   24.0944219958396 2184.71437059304 2204.21713507979    63758.5262346283
#> 4   15.7194300615554  7433.1439603709 7667.88817286217    141983.283847719
#> 5   26.2356808160138 2297.35091835847 2429.72429164338    55251.2273845838
#> 6   13.3050657911839 4807.12749652227  4835.3675290819    128838.163325125
#> 7   7.88554961812726 16667.1637882442 17320.3648458647    344227.944001266
#> 8   14.2434720609951 5127.60440583937 5252.46972041802    114690.885338985
#> 9   5.20533380007577 7720.07918088573 7807.69996597902    215032.602986754
#> 10  1.37986367888844  19955.832998598 20219.6585825418    470494.925148338
#> 11  11.1900347759075 6672.25639357167 6907.27409429442     163874.08624963
#> 12  4.74325770500842 9170.83732799195 9216.13679969144    257707.569218956
#> 13  1.04857625574258 20324.6143620555 20588.6550602144    476413.091645057
#> 14  10.9627736267222 6908.93160330032 7058.43337063994    174771.225764666
#> 15  4.00370971726077  8979.9083749339 9030.07908957479    259645.533762006
#> 16  1.08847954545078  20552.984848539 20810.7588718923    505108.506123595
#> 17  8.97012680559727 6759.35547586949 7031.09515666486    171997.528920678
#> 18  5.80384867622213 9304.49973751052 9328.64593398027    262200.210208478
#> 19 0.978525525811791 19703.3226771443 19872.1183830458    480820.020460973
#> 20  10.1981743004387 6936.96796586485 7298.35940163364    171926.780054356
#> 21  7.08761719388694 9242.16074989129 9272.58167371787    261200.139158729
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83752.5286587923 89.9531805337855 95.4800773075682   14.4939532128664
#> 2     43916.5259653379 28.5595122233349 37.9379141395137   26.3344919813656
#> 3     68184.6792396395  61.746405522396 68.4661549764909   15.6894729063823
#> 4     155015.757558865 236.387595037545 247.548264240326   14.1969950717405
#> 5     59618.6297940768 60.3283159615189 70.4906315548197   18.3185437443653
#> 6     133592.394209543 199.267852759744 211.175202198141   11.3398359239432
#> 7     375581.647046617 474.900470379922 540.836268594048   5.82487275921079
#> 8     122709.760114805 182.934859599804 222.167447101994   12.3908882774715
#> 9     227169.360413172  277.21664192512 304.415678873958   4.11432955442804
#> 10    497192.505131737 610.595463453318 724.544176795065 -0.170922356203837
#> 11    171551.124531083 300.753738204479  345.89549326588   8.73323131453087
#> 12    272895.333764111 322.901878409501 351.015755498516   2.87747986081892
#> 13    505885.652418722 688.135337514039 756.530325633832  0.142424931508557
#> 14     180109.39288769 318.891068841959 364.501089478081    8.7585939317562
#> 15    271466.567915752 306.031055205472 326.378779679659   2.30907909179969
#> 16    525869.446414961 732.777633909756 824.598821783012 -0.147512220682061
#> 17    180025.818281832 309.083305506613  361.50850616724   6.89037736570521
#> 18    277104.151599364 336.437841760372 363.047526316674   4.29756971961747
#> 19    521183.327581954 736.791587823074 823.935111015701  0.251634868696909
#> 20    186817.127857922 326.049894556946 381.890819315873   8.83559630625075
#> 21    274669.177725203 325.633912327357 359.831977271719    4.1089285578016
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.9137224492456 3385.08042196914 3453.18434541051    79677.3754598619
#> 2   36.6315361413488 1408.15944204109 1514.10597721866    40570.6392442289
#> 3   24.8288350230737 2182.84708463155 2206.08442104128    63334.7456277656
#> 4   15.8524583616363 7410.66845066429 7690.36368256878    140735.493811545
#> 5   26.9274694921773 2284.67687198013 2442.39833802172    54833.0718347387
#> 6   13.4767849057971 4804.42366361763 4838.07136198654    128382.971006404
#> 7   8.06560876113938 16604.6232614507 17382.9053726581     341225.99370969
#> 8   14.4053483139244 5115.64921614567 5264.42491011172     113923.12073279
#> 9   5.30066426814208 7711.68995678106 7816.08919008369    213870.573020395
#> 10  1.51536925477029 19930.5731022629 20244.9184788769    467938.773873332
#> 11   11.404706923018 6649.75469882162 6929.77578904447    163139.050669491
#> 12  4.90628683702498 9166.50014453136 9220.47398315203    256253.421549739
#> 13  1.12775452679215 20299.3338696785 20613.9355525913    473591.250719919
#> 14  11.1553718524959 6894.61760429972 7072.74736964054    174260.124657142
#> 15  4.15178423793212 8975.10479587254 9034.88266863615    258513.732619626
#> 16  1.19647882598665 20528.3043569413 20835.4393634899     503120.75652123
#> 17  9.15185248481114 6733.33784685717 7057.11278567718    171228.862918014
#> 18  5.93546528408079 9302.18786763576 9330.95780385503    260773.237096585
#> 19  1.04204024342377 19687.1613861538 19888.2796740363    476955.448502581
#> 20   10.317234513523 6902.36665818486 7332.96070931363    170501.108456142
#> 21  7.34789095820508 9239.24810824832 9275.49431536084    259910.550359812
#>    ci95.integral.bt.up
#> 1     84108.6100062716
#> 2     44208.8849992212
#> 3     68608.4598465023
#> 4     156263.547595039
#> 5     60036.7853439219
#> 6     134047.586528263
#> 7     378583.597338193
#> 8     123477.524721001
#> 9     228331.390379531
#> 10    499748.656406743
#> 11    172286.160111222
#> 12    274349.481433328
#> 13     508707.49334386
#> 14    180620.493995214
#> 15    272598.369058132
#> 16    527857.196017325
#> 17    180794.484284496
#> 18    278531.124711258
#> 19    525047.899540346
#> 20    188242.799456136
#> 21    275958.766524121