Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.6669600407071  15.5295267512325
#> 2                 FALSE          0.75  31.972711584644  31.7784970848636
#> 3                 FALSE          0.75  64.711952464012  20.4337029287025
#> 4                 FALSE          0.75 242.794078052256  15.0273298955436
#> 5                 FALSE          0.75 66.0194948165392  21.9899099455649
#> 6                 FALSE          0.75 205.295802463759  12.3053862501361
#> 7                 FALSE          0.75 508.557168204133  7.12557220072836
#> 8                 FALSE          0.75 202.282860864054  13.4646899420982
#> 9                 FALSE          0.75 290.800496734338  4.73111848540437
#> 10                FALSE          0.75 667.998527796103 0.532544576955212
#> 11                FALSE          0.75 326.553101417462  9.81259438851663
#> 12                FALSE          0.75 333.108500257862  3.77209975025174
#> 13                FALSE          0.75  729.88876183408 0.832345232359606
#> 14                FALSE          0.75 338.935674628185   9.5926402060189
#> 15                FALSE          0.75 317.771248521092  3.43475450025638
#> 16                FALSE          0.75 778.484135291465 0.564480590899339
#> 17                FALSE          0.75 338.502075594007   8.2219100633471
#> 18                FALSE          0.75 351.758962186549  4.97141746956377
#> 19                FALSE          0.75    776.521007915 0.687734989638765
#> 20                FALSE          0.75 348.407015914795  9.54342889955301
#> 21                FALSE          0.75 338.169078907454  5.05652365144366
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3418.90021719497 2131.79880711302  81855.624441118 3.28628431300429
#> 2  1456.70337989523 264.833910431209 42532.0486044109 2.06073511363448
#> 3   2195.6412647534 664.557356051087 66515.5163035258 1.75052805073829
#> 4  7548.65262043532 6013.72229956157 148251.347354251 1.89603158923932
#> 5  2354.74786010375 1116.01529339527 56829.0366245378 2.65131443900881
#> 6  4821.49680483161 3248.97910291975 130821.925425558 2.64588597145648
#> 7  16996.6302171515 15752.9153533178 359526.832243571 11.1382345141215
#> 8    5187.271368389 3757.95495977602  118672.62754785 7.30378183991522
#> 9  7764.15989264975 6280.58969230514 221906.963840374 4.50896263136992
#> 10 20055.1308255388  18705.235838749  484368.98835289 22.2815982084122
#> 11 6787.84708038848 5601.78971146919 167503.265628359  12.830010064867
#> 12 9190.92584873513 7654.17921472641 267236.093115156 5.00936214773215
#> 13 20446.0451892263 19357.3357696213 490098.394821201 22.3208656356774
#> 14 6973.64969055885 5688.21552843527 177264.381912566 11.5986833055078
#> 15 9012.06832836398 7365.59070152811 265861.791977339 5.17966670649944
#> 16 20671.8265804128 19272.5272510843 513927.524013918 27.4472293969002
#> 17 6916.30380607482 6001.23834419127 176566.914982951 11.1977073022516
#> 18  9321.6908069573 7670.36202174201 271132.186491098 8.06897143453849
#> 19 19844.1655753262 18624.6078319489 505732.391762522  30.499665493188
#> 20 7147.87823194377 5808.86142744072 180305.722876475 13.8496723243894
#> 21 9257.22205200814 7534.73168868809 269370.852399856 7.02904490614747
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.855215820922832 19.6614163446561 70.0161511491086 1166.93866165698
#> 2   2.44298933599057 21.0805997250818 65.8618652486785 576.016260492041
#> 3   1.67572613814643 5.18961378013575 108.904628970508 958.047135900724
#> 4  0.194620576468699 72.3204961522268 99.3800175920803 3621.86322911172
#> 5   2.66145663885865 43.8962542150822 163.787049679407 2108.32134832842
#> 6   0.45441382434253 10.9589022911915 94.1675985936627 2319.59645147916
#> 7  0.371332168137192 173.126950424832 204.178714384006 9099.67848554005
#> 8  0.592582551784583 41.9914432244542 104.561486570602 2421.02207820755
#> 9  0.328152534770831 24.1254440008517 99.4095832518378 2555.18724898083
#> 10  0.32740066662768  44.955120696547 161.427552681825 7079.30107102038
#> 11 0.571849878931142 96.5373774662622 168.882277795696 3745.72557874449
#> 12  0.33006841146926 16.3186763258751 86.4001381182128 2944.99652524892
#> 13 0.394145865605523  89.550738051343 228.600130429156 10349.9313361926
#> 14 0.673191541567051 41.3155380526403 193.588852434139 2718.05436923142
#> 15 0.543062130072506 14.2338757465721 153.877106197337 2227.80764899569
#> 16 0.777813220087445 84.7349443469664 552.985179375001 9599.84317454864
#> 17 0.468451123896418 64.0562080096403 128.737073305538 2371.06462944723
#> 18 0.353933814318749 6.61892679354793 55.7556121882744  3780.3723499812
#> 19 0.405408798960187 67.4669201191506 249.105657909427  6252.6731196794
#> 20 0.391807851463881 85.7313656246752 88.1952019633676 3470.29237432555
#> 21 0.350605498843664 8.92036887675481  92.837729466519 3757.13516743534
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up    ci90.lambda.bt.lo
#> 1             TRUE  87.261022345815 98.0728977355991     14.1226967258144
#> 2             TRUE 28.5828023227153 35.3626208465727     27.7597796271591
#> 3             TRUE 61.8323338205475 67.5915711074765     17.6771334314516
#> 4             TRUE 239.675106087958 245.913050016555     14.7071790472526
#> 5             TRUE 61.6580825643697 70.3809070687087     17.6118137746424
#> 6             TRUE 200.943320040713 209.648284886805     11.5578755090927
#> 7             TRUE 490.234772428404 526.879563979863     6.51473078414268
#> 8             TRUE 190.268139737393 214.297581990714     12.4898916444125
#> 9             TRUE 283.383253205734 298.217740262941     4.19130756570635
#> 10            TRUE 631.345298743265 704.651756848942 -0.00602951964732157
#> 11            TRUE 305.447734860756 347.658467974169      8.8719013376749
#> 12            TRUE 324.868099524843 341.348900990882      3.2291372133848
#> 13            TRUE 693.170937863391  766.60658580477    0.183975283438522
#> 14            TRUE 319.855840590624 358.015508665745      8.4852401201411
#> 15            TRUE   309.2506967889 326.291800253283     2.54141729628711
#> 16            TRUE 733.333442933564 823.634827649366   -0.715022156144507
#> 17            TRUE 320.081847081803 356.922304106211     7.45130796453749
#> 18            TRUE 338.485504176733 365.032420196364     4.38919634500943
#> 19            TRUE 726.349058178706 826.692957651295   0.0208375153492564
#> 20            TRUE 325.624304941174 371.189726888415     8.89890498389492
#> 21            TRUE 326.606300036841 349.731857778066     4.47977760584583
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.9363567766505 3386.55718730801 3451.24324708193    79936.0103426922
#> 2   35.7972145425681 1422.02579334747 1491.38096644299    41584.5018559015
#> 3   23.1902724259534 2187.10435008507 2204.17817942172    64939.5287649691
#> 4   15.3474807438346 7429.68540426491 7667.61983660574    142293.382342362
#> 5   26.3680061164874 2282.53852191994 2426.95719828756    53360.8480065375
#> 6   13.0528969911796  4803.4694105626 4839.52419910062    127006.189262875
#> 7   7.73641361731404 16711.8363837026 17281.4240506003    344557.861134857
#> 8   14.4394882397838 5118.19544428477 5256.34729249322    114690.046229199
#> 9   5.27092940510238 7724.47353726835 7803.84624803115      217703.6808158
#> 10  1.07111867355774  19981.179651993 20129.0819990846    472723.538091062
#> 11  10.7532874393584 6629.04309445648 6946.65106632048    161341.547051324
#> 12  4.31506228711867 9164.08162617907  9217.7700712912    262391.573831121
#> 13  1.48071518128069 20298.7342251318 20593.3561533207    473072.757773165
#> 14  10.7000402918967 6905.68563046226 7041.61375065544     172793.18247518
#> 15  4.32809170422566 8988.65360276087  9035.4830539671    262197.048394741
#> 16  1.84398333794319 20532.4375969621 20811.2155638636    498135.781991785
#> 17   8.9925121621567 6810.93134389896 7021.67626825068    172666.513667511
#> 18  5.55363859411811 9310.80267238191 9332.57894153268    264913.473975379
#> 19  1.35463246392827 19733.1824917302 19955.1486589222    495446.744480649
#> 20  10.1879528152111 7006.85013549118 7288.90632839636    174597.091920709
#> 21  5.63326969704148 9242.54804520588  9271.8960588104    263190.365049425
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83775.2385395437 86.2258427872187 99.1080772941955   13.8533037422237
#> 2     43479.5953529203 27.9336707619204 36.0117524073676   26.9902379863221
#> 3     68091.5038420825  61.280917484565 68.1429874434591   17.1492796979355
#> 4     154209.312366139 239.077856137347 246.510299967165   14.6458735656649
#> 5      60297.225242538 60.8229185160819 71.2160711169965   16.7734549334019
#> 6     134637.661588241 200.109865959705 210.481738967814   11.4147351544248
#> 7     374495.803352284 486.726228556455 530.388107851812   6.39776115117946
#> 8     122655.208866502  187.96744845782 216.598273270287   12.3032281406004
#> 9     226110.246864947 281.962929976852 299.638063491823   4.08793951725354
#> 10    496014.438614719 624.326595307615 711.670460284591 -0.109160729635041
#> 11    173664.984205393 301.406281690323 351.699921144602   8.69176862581159
#> 12     272080.61239919 323.290150448307 342.926850067417   3.12516566377199
#> 13    507124.031869238 686.139865188152 773.637658480008 0.0598193357727819
#> 14    181735.581349951 316.202255349389  361.66909390698   8.27318478454748
#> 15    269526.535559937 307.619101776353  327.92339526583   2.37035272531427
#> 16     529719.26603605 724.687565673541  832.28070490939 -0.960033320472052
#> 17    180467.316298392 316.554569281594  360.44958190642   7.30374586051012
#> 18    277350.899006817 335.943778174853 367.574146198244   4.27770719349902
#> 19    516018.039044395 716.741663548352 836.300352281649 -0.106866256323203
#> 20     186014.35383224 321.261658158992 375.552373670598    8.7754855106838
#> 21    275551.339750287 324.392150891405 351.946006923503   4.36933687371008
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.2057497602412 3380.36384115944 3457.43659323049    79568.4246642703
#> 2   36.5667561834051 1415.38540443407 1498.02135535639    41403.0567338465
#> 3   23.7181261594695 2185.46962174433 2205.81290776246    64637.7439171604
#> 4   15.4087862254222 7406.90444797696 7690.40079289369    141152.495425192
#> 5   27.2063649577278 2268.71120184219 2440.78451836531    52696.7267818141
#> 6   13.1960373458475 4800.01735634087 4842.97625332234    126275.516380659
#> 7   7.85338325027725 16657.3013943188 17335.9590399841    341691.462411912
#> 8    14.626151743596 5104.96813966907 5269.57459710893    113927.424274563
#> 9   5.37429745355519 7716.87402240808 7811.44576289142    216898.796832371
#> 10  1.17424988354546 19967.0187889736  20143.242862104     470493.55825369
#> 11  10.9334201512217 6598.63382055461 6977.06034022235     160161.64349402
#> 12  4.41903383673149 9158.94124313642 9222.91045433385    261463.899925668
#> 13  1.60487112894643 20270.5257426456 20621.5646358069    469812.529402264
#> 14  10.9120956274903 6892.67123597567 7054.62814514202    171936.995348872
#> 15  4.49915627519849  8984.1699319007 9039.96672482727    261495.288985308
#> 16  2.08899450227073 20505.7460894928 20837.9070713329    495111.831391802
#> 17  9.14007426618407 6790.75363837592 7041.85397377371    171919.628309235
#> 18  5.66512774562852 9308.71771044194 9334.66390347265    263722.656685135
#> 19  1.48233623560073 19711.9304118926 19976.4007387597     493477.15244795
#> 20  10.3113722884222 6979.84475531941 7315.91170856814    173503.949822797
#> 21  5.74371042917724  9239.7381290097 9274.70597500658    262006.867471683
#>    ci95.integral.bt.up
#> 1     84142.8242179656
#> 2     43661.0404749753
#> 3     68393.2886898912
#> 4      155350.19928331
#> 5     60961.3464672615
#> 6     135368.334470457
#> 7     377362.202075229
#> 8     123417.830821137
#> 9     226915.130848376
#> 10     498244.41845209
#> 11    174844.887762698
#> 12    273008.286304644
#> 13    510384.260240139
#> 14    182591.768476259
#> 15    270228.294969371
#> 16    532743.216636033
#> 17    181214.201656668
#> 18    278541.716297061
#> 19    517987.631077094
#> 20    187107.495930153
#> 21     276734.83732803