Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8140444691318  15.7210357789876
#> 2                 FALSE          0.75 32.3289760109655  31.5795762748736
#> 3                 FALSE          0.75 64.9462334573597  20.3186445509193
#> 4                 FALSE          0.75 242.242567495008  14.9293689045021
#> 5                 FALSE          0.75 65.9180877741482  22.9306374071509
#> 6                 FALSE          0.75 204.960295488416  12.4231139494038
#> 7                 FALSE          0.75   505.9350404007  6.98972902019161
#> 8                 FALSE          0.75 206.073247834975  13.5264922673441
#> 9                 FALSE          0.75 289.625250961644  4.67843953174544
#> 10                FALSE          0.75 677.525085112975 0.774640186573646
#> 11                FALSE          0.75 323.761880448101  10.0023961867234
#> 12                FALSE          0.75 337.570164490903  3.96370914119192
#> 13                FALSE          0.75 731.973304554397 0.727167221890404
#> 14                FALSE          0.75 346.307817163444    9.885645242545
#> 15                FALSE          0.75 319.938605191357  3.44227200809006
#> 16                FALSE          0.75 777.552912240708 0.602233818646511
#> 17                FALSE          0.75 331.091658279811  8.14744828365822
#> 18                FALSE          0.75 348.430669941022  4.88719307855589
#> 19                FALSE          0.75 778.783414871687 0.654187366444813
#> 20                FALSE          0.75  349.15985354394  9.43300897129309
#> 21                FALSE          0.75 338.437781327944  5.29539863682022
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3412.74354022957  2109.3808799243 81046.0726723644  1.9985084483281
#> 2  1461.75678521582 270.856311211677 42407.9604880311 2.19617608749894
#> 3  2195.44935925783 682.088459357687 65872.1697129088  1.2782064620723
#> 4  7560.50733718682 6054.42470648681 149074.666972932 2.23192518354957
#> 5  2370.22524597707 1101.05417067752 57351.4171500025 3.01943694249125
#> 6  4816.73220848859 3220.00917523843 129962.561377266 2.17828056612577
#> 7  17016.4012885878 15799.8156494322 361015.778200194 15.9514276788436
#> 8  5177.17841528124 3749.88893886763 118422.458823094 9.27755048048836
#> 9  7762.30331975816 6270.84441111533 221441.068937682 5.94221662027194
#> 10 20096.2075621225 18632.2744659797 484433.211572032 34.4392293804881
#> 11 6753.05207975038 5503.78681485579 166305.573320953 15.7474022195941
#> 12 9188.18876333642 7647.16678472255 267510.489711263 6.17183325481727
#> 13 20461.7178546968  19440.999383439 492929.494159411 29.6306522972757
#> 14 6991.98463433639 5662.70468166531 177723.303032576 11.5258167416617
#> 15 8998.27492106709 7376.71326328104 262884.601596311 5.84888746623687
#> 16 20665.8976232515  19280.920546337  511769.48518948 28.7561401968746
#> 17 6904.19632265314 6009.56435104268 175849.530527957 15.4873085683247
#> 18 9319.27732209411  7665.8684351378 272562.799287383  5.9184720495546
#> 19 19799.0366414422 18601.2593100647 500078.754897893 22.8793209649449
#> 20 7084.47205607614 5780.45683467591 178652.624628616 15.6682349251413
#> 21  9256.3704218387 7443.40986039064  269099.68420941 8.09096363314245
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.72292015353439 22.2395884618412 68.7016154326243 1605.90032477564
#> 2   2.71158914253615 17.3604022673999 77.8858517837978 1115.93575700113
#> 3   1.88040804508331 4.42499780962129 118.296941140301 1124.43548491738
#> 4   0.25571090200386 61.3631640811392 113.366256772444  3179.6500756716
#> 5   2.07013754565205 33.7799952888274 159.228328030009 1375.39301593238
#> 6  0.416359631001983 12.3541634949349 82.4585517658334 4011.64166456742
#> 7  0.665677201925138 166.031048813336  193.73562791698 7780.85096778601
#> 8  0.682618923371497 48.5396581260346 141.647779053589 2336.52413504879
#> 9  0.410762322922752 20.8705141308367 65.2651663906247 4258.29650450576
#> 10 0.869090197896799 91.4794158790811 509.968028537681 9325.51843841256
#> 11 0.526864665461602 106.693994169564 162.961916478598 4335.00193288876
#> 12 0.469819614953762 8.15562236768416 125.563286209721 3056.75133514935
#> 13 0.426163962035092 86.9041692639234 232.668144738059 9243.33543944566
#> 14 0.551722841344477 35.7841304208426 172.197440673131 2129.66673638318
#> 15 0.485448866038481 17.4557420387277 131.859222715007 6597.10828504559
#> 16 0.491001986464538 94.1370264597167 343.384432373046 10389.0198479929
#> 17 0.400634694391175 66.4177565957027 126.312321556549 2579.75031732666
#> 18 0.342932236748438  7.8526761241314 69.7980459739249 2556.74295633747
#> 19 0.300855116188671 66.3900602026385 233.498851731634 12090.6822501377
#> 20 0.381797270261958 78.2874511507992 109.778046570158 3135.82981603845
#> 21  0.59293729247932 11.5287872703249 169.339061011156 3007.43814461975
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.5264980716321 96.1015908666315   14.5318321264235
#> 2             TRUE 28.7162663470297 35.9416856749013   27.1190121354016
#> 3             TRUE 62.8435838272507 67.0488830874686   17.2253733167573
#> 4             TRUE 238.571050568069 245.914084421947   14.5087244707058
#> 5             TRUE 60.9511140037501 70.8850615445463   19.5252611445532
#> 6             TRUE 201.377023957139 208.543567019693   11.7382023564055
#> 7             TRUE 479.694941869002 532.175138932398   5.89469002302476
#> 8             TRUE 190.811677294571 221.334818375378    12.403584138398
#> 9             TRUE 279.850304621297 299.400197301992   4.00273551053751
#> 10            TRUE 620.872552782073 734.177617443878 -0.655013188966588
#> 11            TRUE 297.857403796869 349.666357099333    9.1357038120391
#> 12            TRUE 327.417498786729 347.722830195078   3.19085587459299
#> 13            TRUE 683.230881525378 780.715727583416  0.026127504342677
#> 14            TRUE  327.34784862341 365.267785703477   8.97806116853334
#> 15            TRUE 310.317185309398 329.560025073317   2.64370862345676
#> 16            TRUE 730.249061616849 824.856762864567 -0.205464449087654
#> 17            TRUE 305.615035684917 356.568280874706   7.48840421138474
#> 18            TRUE 338.694783419504 358.166556462539   4.32306954910471
#> 19            TRUE 741.146931884353 816.419897859021   0.15928070031445
#> 20            TRUE 323.385607092082 374.934099995797   8.80495246171217
#> 21            TRUE 325.128146151424 351.747416504463   4.32001679069174
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.9102394315516 3376.15941720984  3449.3276632493    78404.3666381085
#> 2   36.0401404143455 1433.19892348595 1490.31464694569    40572.2461677643
#> 3   23.4119157850814   2188.170237861 2202.72848065466    64022.4733402197
#> 4   15.3500133382985 7459.56493227334 7661.44974210029    143844.142598453
#> 5   26.3360136697485 2314.65715372695 2425.79333822719    55088.8956387938
#> 6    13.108025542402 4796.40960953942 4837.05480743776    123363.410839053
#> 7   8.08476801735846 16743.2802132898 17289.5223638857    348216.278358186
#> 8   14.6494003962903 5097.33067766391 5257.02615289857    114578.876620939
#> 9   5.35414355295337 7727.97132401293 7796.63531550339     214436.17118777
#> 10  2.20429356211388 19945.7239230014 20246.6912012436    469092.733740844
#> 11  10.8690885614078 6577.54045934144 6928.56370015931    159174.495141351
#> 12  4.73656240779086 9174.77276454158 9201.60476213126    262482.133764942
#> 13  1.42820693943813 20318.7604962577  20604.675213136    477724.207361523
#> 14  10.7932293165567  6933.1197397941 7050.84952887868    174220.001251226
#> 15  4.24083539272336 8969.56022541338  9026.9896167208    252032.358467411
#> 16  1.40993208638068 20511.0422147253 20820.7530317777    494679.547539532
#> 17   8.8064923559317 6794.93911305321 7013.45353225307    171605.841255955
#> 18  5.45131660800707 9306.35966986992 9332.19497431831    268356.957124208
#> 19  1.14909403257518 19689.8249924089 19908.2482904755    480189.582596416
#> 20   10.061065480874 6955.68919893308 7213.25491321921    173494.184581233
#> 21   6.2707804829487 9237.40556677901 9275.33527689838    264152.448461511
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83687.7787066204 88.8969679104087 96.7311210278549   14.3041122780602
#> 2      44243.674808298 28.0244708794676 36.6334811424634   26.2648615555027
#> 3     67721.8660855978 62.4409487916979 67.4515181230214    16.633044782556
#> 4     154305.191347412  237.86799413525 246.617140854765   14.4281755365746
#> 5     59613.9386612113 59.9999913668653  71.836184181431   18.8731678176728
#> 6      136561.71191548 200.690865578809 209.229725398022   11.6070490726399
#> 7     373815.278042202 474.670242150167 537.199838651234   5.68500170441834
#> 8     122266.041025249 187.889248893218 224.257246776732    12.188559177536
#> 9     228445.966687594 277.978506385911 301.271995537377   3.87334537881685
#> 10    499773.689403221 610.024195527219 745.025974698732  -0.92877660130408
#> 11    173436.651500555 292.896972097696 354.626788798505    8.9697414424187
#> 12    272538.845657583 325.473371311462 349.666957670345   3.04286269588255
#> 13    508134.780957299 673.897226051736 790.049383057057 -0.108114143698377
#> 14    181226.604813927 323.717216349787 368.898417977101   8.80426847350983
#> 15    273736.844725211 308.474785757533 331.402424625182   2.49079223065464
#> 16    528859.422839428 721.190877454834 833.914947026582 -0.360130074823984
#> 17     180093.21979996 300.736533485895 361.446783073728   7.36220428265152
#> 18    276768.641450558 336.830464723895 360.030875158149   4.21504589452895
#> 19     519967.92719937 733.939945780395 823.626883962979  0.064511338715019
#> 20       183811.064676 318.450113090663 379.869593997217   8.68468632157965
#> 21     274046.91995731 322.579492606984 354.296070048903   4.13324154356075
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1    17.137959279915 3369.15394684436 3456.33313361478    77898.5080358042
#> 2   36.8942909942444 1427.73039677171 1495.78317365992    40220.7264043089
#> 3   24.0042443192826 2186.77636355097 2204.12235496469    63668.2761624707
#> 4   15.4305622724297 7440.23553558778 7680.77913878585    142842.552824616
#> 5   26.9881069966289 2304.01645521096 2436.43403674317    54655.6468387751
#> 6   13.2391788261677 4792.51804803852 4840.94636893866    122099.743714714
#> 7   8.29445633596488 16690.9804329136 17341.8221442619    345765.310303333
#> 8   14.8644253571523 5082.04068535421 5272.31614520827    113842.871518399
#> 9   5.48353368467403 7721.39711206172  7803.2095274546    213094.807788851
#> 10  2.47805697445137 19916.9079069995 20275.5072172455    466155.195432744
#> 11  11.0350509310282 6543.93185117803 6962.17230832272    157808.969532491
#> 12   4.8845555865013 9172.20374349576 9204.17378317708     261519.25709437
#> 13  1.56244858747918 20291.3856829395 20632.0500264541    474812.556698098
#> 14  10.9670220115802 6921.84773871154 7062.12152996124    173549.156229265
#> 15  4.39375178552548 8964.06166667118   9032.488175463    249954.269357622
#> 16    1.564597712117 20481.3890513904 20850.4061951125    491407.006287414
#> 17  8.93269228466492 6774.01751972556 7034.37512558072    170793.219905997
#> 18  5.55934026258283 9303.88607689082 9334.66856729741    267551.583092961
#> 19  1.24386339417461  19668.912123445 19929.1611594394    476381.017687623
#> 20  10.1813316210065 6931.02865182057 7237.91546033171    172506.398189181
#> 21  6.45755573007968 9233.77399878886 9278.96684488854    263205.105445956
#>    ci95.integral.bt.up
#> 1     84193.6373089247
#> 2     44595.1945717533
#> 3     68076.0632633468
#> 4     155306.781121249
#> 5       60047.18746123
#> 6     137825.379039818
#> 7     376266.246097054
#> 8      123002.04612779
#> 9     229787.330086514
#> 10    502711.227711321
#> 11    174802.177109415
#> 12    273501.722328155
#> 13    511046.431620725
#> 14    181897.449835887
#> 15    275814.933835001
#> 16    532131.964091546
#> 17    180905.841149917
#> 18    277574.015481804
#> 19    523776.492108163
#> 20    184798.851068052
#> 21    274994.262972865