Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for class 'flFit'
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP     1             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP     1             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP     1             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP     1          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP     1          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP     1          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP     1          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP     1          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP     1          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP     1           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP     1           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP     1           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP     1           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP     1           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP     1           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP     1           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP     1           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP     1           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP     1             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP     1             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP     1             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.5463966032594  15.5184813391066
#> 2                 FALSE          0.75 33.1667986504996  31.3733583984605
#> 3                 FALSE          0.75 65.3384435116991  20.1967658916363
#> 4                 FALSE          0.75 243.353571195125  15.0397312794848
#> 5                 FALSE          0.75 65.0042400426535  22.8229586942172
#> 6                 FALSE          0.75 204.957591825311  12.4470113682392
#> 7                 FALSE          0.75 508.211536462878  7.21689380487454
#> 8                 FALSE          0.75 201.875883564173  13.0950702401446
#> 9                 FALSE          0.75  290.82274845449  4.67926990592957
#> 10                FALSE          0.75 663.554165626676 0.664890468420249
#> 11                FALSE          0.75 320.642183551389  9.67781341339297
#> 12                FALSE          0.75 333.424137181004  3.77256913964949
#> 13                FALSE          0.75  712.02226482142 0.615126688418128
#> 14                FALSE          0.75 345.036994262351  9.89690649166494
#> 15                FALSE          0.75 321.772576150627  3.37410923502297
#> 16                FALSE          0.75 781.000544866005 0.573066239143825
#> 17                FALSE          0.75 337.176440408004  8.28812385251033
#> 18                FALSE          0.75 350.774677861543  5.12544674745676
#> 19                FALSE          0.75 777.740698177641 0.635698715266595
#> 20                FALSE          0.75  356.45891315571  9.71727738712474
#> 21                FALSE          0.75 337.657418299854   5.1858882488952
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3409.07249111648 2121.49734931896 81350.4308708412 2.12511889496982
#> 2  1461.25327821284 271.445643912595 41985.5387478764 1.72421110282744
#> 3  2196.05720391067 679.623827259535 65741.3713415894 1.76553797119635
#> 4  7568.70596520617 6042.50425611596 149665.654548745 2.22547038618906
#> 5  2361.02122570199 1074.49214058502 56913.5693380427 2.76112476210093
#> 6  4816.69534974059 3216.44220128049 129853.591824207 2.81108838547134
#> 7  17004.0693214553 15722.4370018421  360074.02487869 13.8977400720722
#> 8  5188.73821112898  3838.1734821177 118356.630990037 8.79504911496906
#> 9   7767.2496900824 6292.31934142077 221794.213686643 8.30579982382058
#> 10 20093.7600950006 18609.2221477527 487213.148334687 34.8550951622069
#> 11 6796.99773286669 5618.15121250895 167981.431519358 13.4041389792713
#> 12 9189.84706055838 7654.46652972893 268385.350329613 6.87377705846367
#> 13 20417.3486434626 19337.1807565019 489075.203612532 24.9478352371224
#> 14 6982.72792195301  5631.7533826393 177460.566640686 13.3686760367349
#> 15 9000.33446410188  7422.4468680733 264847.977208231  6.4370128598814
#> 16 20678.1734887479 19321.8357377871 513484.603175575 22.4946147509244
#> 17  6900.1898417116 5956.65456738436 176177.059934911 14.5822702609616
#> 18  9318.8627880924 7603.47004430968 270846.091282357 6.22699631981648
#> 19 19777.2891594565 18596.5945656397 503975.403230652 27.7962638681376
#> 20 7173.32117926389 5815.58964028959 181385.479373069 14.6023807040088
#> 21 9253.90374858231 7463.97319536472 267328.936022362 7.07266842048449
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.658391984866374 38.0992419163035  65.928649551731 2138.76971394585
#> 2   2.53290264035913 19.3747177572927 96.5779424158469 1022.26893003056
#> 3   2.03078471026394 4.49759922395865 139.800133422891 1728.03159041632
#> 4   0.25848247579349 47.1390246628431  84.189168371932 2568.22473989489
#> 5   1.45971663837761  43.439984846905 130.076515941912 1829.33867466439
#> 6  0.543604206970577 8.04004127297606 103.804494723949 3653.01760207908
#> 7  0.491052074164049   165.5278266322 241.157005479568 8619.38227227516
#> 8  0.722805474376428 42.0848813075889 134.538534066817 2349.19546576155
#> 9  0.462027958684981 24.1677186730615 60.3384401355005 3702.84359688155
#> 10 0.418190684846864 83.7252424823234 184.217766713335  5838.4071811725
#> 11 0.648910958516624 74.5224929227532 130.619568122207 2378.62679113595
#> 12 0.581410514800424 11.7567744506358 172.234124403663 2677.70377323893
#> 13 0.291287653032469 89.2019869780152 188.563749550384 9397.97513206387
#> 14 0.486747599105017 59.0771623712341 134.362131489858 3346.59554365211
#> 15 0.287004221556101 17.8612047225706 77.2473034733221 4137.77341071735
#> 16 0.578093391118959 78.5722009109086 413.782668883146 9943.66722436406
#> 17 0.714075575489119 83.6558244308235 194.381201848384 3385.39222791988
#> 18 0.522711696695231 9.95396246620455 171.861332582303 3430.80291661462
#> 19   0.3389468651055 74.5062839838655 208.213807843259 6058.96867198426
#> 20 0.381947863091274 62.6065691073238 122.067333692839 1570.77183416962
#> 21 0.506342270732548 11.2279694791583 140.984062497318  5357.6750737779
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 89.0505760210341 96.0422171854848    14.4354265240014
#> 2             TRUE 30.3304713863484 36.0031259146507    27.2067335550697
#> 3             TRUE 62.4341335490811 68.2427534743171    16.8561250432521
#> 4             TRUE 239.692672409844 247.014469980406    14.6145276068045
#> 5             TRUE 60.4621898089974 69.5462902763095     20.421724824086
#> 6             TRUE  200.33335143121 209.581832219411    11.5527824477726
#> 7             TRUE 485.349754044319 531.073318881437    6.40911314287468
#> 8             TRUE 187.408027770049 216.343739358297    11.9060552347954
#> 9             TRUE 277.159707744305 304.485789164675    3.91923391389278
#> 10            TRUE 606.217534084846 720.890797168507 -0.0230332081528435
#> 11            TRUE 298.592374930487  342.69199217229    8.61035488663313
#> 12            TRUE 322.116773919831 344.731500442176    2.81614884280279
#> 13            TRUE 670.983075856354 753.061453786486   0.135958499179717
#> 14            TRUE 323.045522181922  367.02846634278    9.09620669113718
#> 15            TRUE 311.183689996122 332.361462305132    2.90198729056319
#> 16            TRUE 743.996903600734 818.004186131275  -0.377897389246863
#> 17            TRUE 313.188605828722 361.164274987286    7.11346953083073
#> 18            TRUE 340.531268915445 361.018086807641    4.26558600639311
#> 19            TRUE 732.015844114555 823.465552240727  0.0781311221680463
#> 20            TRUE 332.437996897615 380.479829413804     9.0889731523396
#> 21            TRUE 326.022878748157 349.291957851552    4.35295521354016
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6015361542118 3346.39923816416  3471.7457440688    77832.1546914002
#> 2   35.5399832418513 1429.38186750209 1493.12468892358    40303.9063579761
#> 3   23.5374067400204 2188.65865318726 2203.45575463408    62898.7593753546
#> 4    15.464934952165  7491.1622696358 7646.24966077655    145440.924851618
#> 5   25.2241925643483 2289.56245062883 2432.48000077515    53904.3072182198
#> 6   13.3412402887058 4803.46948184655 4829.92121763464    123844.377868787
#> 7    8.0246744668744 16731.7760466453 17276.3625962652    345895.141040797
#> 8   14.2840852454938   5119.508581378 5257.96784087997    114492.204448859
#> 9   5.43930589796637 7727.49379286522 7807.00558729959    215703.035969773
#> 10  1.35281414499334 19956.0320711171  20231.488118884    477608.968521658
#> 11  10.7452719401528 6674.40823200876 6919.58723372462    164068.590447939
#> 12  4.72898943649619 9170.50716658709 9209.18695452968    263980.527622635
#> 13  1.09429487765654 20270.6113748838 20564.0859120414    473615.534520287
#> 14  10.6976062921927 6885.54598985233 7079.90985405369    171955.416971378
#> 15  3.84623117948276 8970.95278233325 9029.71614587051    258041.339947601
#> 16  1.52402986753451 20548.9222182494 20807.4247592463    497127.270591496
#> 17  9.46277817418993 6762.57601052289  7037.8036729003    170608.089719983
#> 18  5.98530748852042  9302.4885198355 9335.23705634931    265202.420484526
#> 19  1.19326630836514 19654.7263223031   19899.85199661    494008.399765238
#> 20  10.3455816219099 7070.33337308234 7276.30898544543     178801.55970586
#> 21  6.01882128425024 9235.43373878909 9272.37375837552    258515.560525997
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84868.7070502821 88.3811635691186 96.7116296374003    14.2280330487685
#> 2     43667.1711377766 29.7873448889578 36.5462524120414    26.4088692233566
#> 3     68583.9833078243 61.8779890881542 68.7988979352439     16.216427859519
#> 4     153890.384245872 238.991649238195 247.715493152056    14.5331056269295
#> 5     59922.8314578657 59.5924355089357 70.4160445763713     19.961914082997
#> 6     135862.805779628 199.447858589787 210.467325060835    11.3815471225768
#> 7     374252.908716583 480.971965921616  535.45110700414    6.25443173951301
#> 8     122221.057531215 184.637587298834 219.114179829512    11.6783715103668
#> 9     227885.391403513 274.543380799802 307.102116109178    3.77369510690701
#> 10    496817.328147716 595.238179108751 731.870152144602  -0.154763273879606
#> 11    171894.272590776 294.370071152017 346.914295950761    8.40594793470039
#> 12    272790.173036591 319.951534146415 346.896740215593    2.63300453064066
#> 13    504534.872704777  663.12450775666  760.92002188618  0.0442028884744889
#> 14    182965.716309994 318.834389230351 371.239599294352     8.9428811974191
#> 15    271654.614468861  309.15603094526 334.389121355995    2.81158096077302
#> 16    529841.935759654 736.911099954193 825.089989777816  -0.559996807449335
#> 17    181746.030149839 308.595190696519 365.757690119489    6.88853572455165
#> 18    276489.762080188 338.569765074703 362.979590648383    4.10093182193411
#> 19    513942.406696066 723.260020996091 832.221375359191 -0.0286371403401864
#> 20    183969.399040278 327.838246975852 385.079579335567    8.96865957546584
#> 21    276142.311518727 323.794988195705 351.519848404004    4.19345739825941
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8089296294447 3334.39797696053 3483.74700527244    77158.4422315073
#> 2   36.3378475735644 1423.27883140854 1499.22772501713    39981.8916450165
#> 3   24.1771039237536 2187.24190943171 2204.87249838963    62354.4294243734
#> 4     15.54635693204   7476.313476867 7661.09845354535    144631.934058551
#> 5   25.6840033054373 2275.87885540205 2446.16359600192    53328.0655357005
#> 6   13.5124756139015 4800.93686884556 4832.45383063563    122693.677324132
#> 7   8.17935587023608 16679.6347812561 17328.5038616544    343180.035625031
#> 8   14.5117689699224 5106.25184376611 5271.22457849186    113752.207877144
#> 9   5.58484470495213  7719.8809614832  7814.6184186816    214536.640236755
#> 10   1.4845442107201 19929.6586197352 20257.8615702659    475769.870259589
#> 11  10.9496788920856  6650.9336467381 6943.06181899529    163319.323008731
#> 12  4.91213374865832 9166.80378263513 9212.89033848163    263137.050934065
#> 13  1.18605048836177 20242.5127489857 20592.1845379395    470655.172353687
#> 14  10.8509317859108 6866.93668370539 7098.51916020063    170901.239375128
#> 15  3.93663750927293 8965.32650284564 9035.34242535812    256737.941323225
#> 16  1.70612928573699 20524.1719749625 20832.1750025333    493995.015415821
#> 17    9.687711980469 6736.22442582718 7064.15525759601    169541.691168188
#> 18  6.14996167297941 9299.35302165864 9338.37255452616    264121.717565793
#> 19  1.30003457087338 19631.2568428482 19923.3214760649    492099.824633563
#> 20  10.4658951987836 7050.61230381353 7296.03005471424    178306.766578097
#> 21    6.178319099531 9231.89692840316 9275.91056876146    256827.892877757
#>    ci95.integral.bt.up
#> 1      85542.419510175
#> 2     43989.1858507362
#> 3     69128.3132588054
#> 4     154699.375038939
#> 5     60499.0731403849
#> 6     137013.506324282
#> 7     376968.014132349
#> 8      122961.05410293
#> 9     229051.787136531
#> 10    498656.426409785
#> 11    172643.540029984
#> 12    273633.649725161
#> 13    507495.234871377
#> 14    184019.893906244
#> 15    272958.013093237
#> 16    532974.190935328
#> 17    182812.428701634
#> 18    277570.464998922
#> 19    515850.981827741
#> 20    184464.192168042
#> 21    277829.979166967