Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP     1             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP     1             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP     1             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP     1          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP     1          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP     1          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP     1          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP     1          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP     1          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP     1           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP     1           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP     1           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP     1           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP     1           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP     1           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP     1           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP     1           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP     1           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP     1             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP     1             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP     1             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917229051         20.5  15.5467166096248
#> 2                 FALSE 30.7727385498454           20   31.287226708742
#> 3                 FALSE 64.1757831646411           16   20.813124032719
#> 4                 FALSE 242.557704041571         23.5  14.9826287460188
#> 5                 FALSE  64.652909453684           19  21.6712998092168
#> 6                 FALSE 205.579719130336           18  12.5662904914223
#> 7                 FALSE 509.394721871847         21.5  6.95952698278093
#> 8                 FALSE 204.031623957399         17.5  13.4360564669405
#> 9                 FALSE 290.961299939289           13  4.67379381987306
#> 10                FALSE 670.985334314846            6 0.373248436529738
#> 11                FALSE 327.194544214716           17  9.86802427129762
#> 12                FALSE 335.243584785836         17.5  3.65080407905447
#> 13                FALSE 728.625260083536          6.5 0.553816578576483
#> 14                FALSE 346.268113254584         16.5  9.75821807365902
#> 15                FALSE 320.246598452298         17.5  3.21350390587511
#> 16                FALSE 774.053986420739          5.5 0.246732777851506
#> 17                FALSE 337.909403917616         16.5  8.08916482172275
#> 18                FALSE  351.64350166477           18  4.84484452251444
#> 19                FALSE 779.134078966834            6 0.462268482213733
#> 20                FALSE 353.185039298477         16.5  9.57178508598769
#> 21                FALSE 338.826806405314           18  5.01597033524087
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA> 1291.42428702462
#> 2               <NA>          <NA>           <NA> 1188.27831547728
#> 3               <NA>          <NA>           <NA> 1565.59217809665
#> 4               <NA>          <NA>           <NA> 1519.80363776283
#> 5               <NA>          <NA>           <NA> 1255.72130804323
#> 6               <NA>          <NA>           <NA> 1624.15289883828
#> 7               <NA>          <NA>           <NA> 1132.18524639265
#> 8               <NA>          <NA>           <NA>  1405.7165286464
#> 9               <NA>          <NA>           <NA>  1468.5243834002
#> 10              <NA>          <NA>           <NA> 1209.47395589003
#> 11              <NA>          <NA>           <NA> 1187.83879461862
#> 12              <NA>          <NA>           <NA> 1467.93981010251
#> 13              <NA>          <NA>           <NA> 891.795404946784
#> 14              <NA>          <NA>           <NA> 1282.77700477382
#> 15              <NA>          <NA>           <NA> 1542.64869773741
#> 16              <NA>          <NA>           <NA> 1132.44043950192
#> 17              <NA>          <NA>           <NA> 862.662992687371
#> 18              <NA>          <NA>           <NA> 1602.42470098299
#> 19              <NA>          <NA>           <NA> 1017.04362101505
#> 20              <NA>          <NA>           <NA> 1323.69594931063
#> 21              <NA>          <NA>           <NA> 1707.52183129238
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552422973 2135.70123720511 82881.4064741494                TRUE
#> 2  1470.99207749914 282.713762021855 43222.9360971347                TRUE
#> 3  2196.18856426069  630.59638616404 67033.6678589424                TRUE
#> 4  7592.80063750999 6072.99699974716 151370.675831998                TRUE
#> 5   2384.3037552124 1128.58244716917 58249.0814015683                TRUE
#> 6  4820.28715109272 3196.13425225444 132414.081549253                TRUE
#> 7  17121.2144379486 15989.0291915559  367233.85193445                TRUE
#> 8  5206.54317686562 3800.82664821922 120525.231431356                TRUE
#> 9  7768.18866992589 6299.66428652569 223919.428584961                TRUE
#> 10 20079.0095008424 18869.5355449523 490254.828661526                TRUE
#> 11 6840.37784262981 5652.53904801119  169914.30195731                TRUE
#> 12 9191.86346478577 7723.92365468326 269744.222628592                TRUE
#> 13 20481.0977680846 19589.3023631378 497122.375105331                TRUE
#> 14 7000.84532045939 5718.06831568557 179100.135589303                TRUE
#> 15 9006.21056404716 7463.56186630976 267724.576457351                TRUE
#> 16 20704.0909872784 19571.6505477765 520199.730690812                TRUE
#> 17 6952.68659445583 6090.02360176846 178120.140852721                TRUE
#> 18 9322.62835009351 7720.20364911053 273698.194401997                TRUE
#> 19  19774.702139555 18757.6585185399 508722.732093168                TRUE
#> 20 7189.59680354916 5865.90085423853   182547.7078057                TRUE
#> 21 9258.53106887516 7551.00923758278 271383.083788683                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.6977191840207  15.5229104813818
#> 2                 FALSE          0.75 32.9291935764017  31.5135758975589
#> 3                 FALSE          0.75 65.1833627898446  20.1786583417859
#> 4                 FALSE          0.75 243.329416346274  15.0288739758758
#> 5                 FALSE          0.75 65.0354838754715  23.2366175931258
#> 6                 FALSE          0.75 205.056033765441  12.4729490553261
#> 7                 FALSE          0.75 508.167335095328   7.1733094610108
#> 8                 FALSE          0.75 202.517473328267  13.1353450479821
#> 9                 FALSE          0.75 290.969002676228  4.67155750718633
#> 10                FALSE          0.75  661.07165232425 0.641744373229107
#> 11                FALSE          0.75 320.571235333184  9.64937932823517
#> 12                FALSE          0.75 333.840962236051  3.76681079957858
#> 13                FALSE          0.75  714.83532489765 0.627102453302372
#> 14                FALSE          0.75 344.777173565655  9.87830901636098
#> 15                FALSE          0.75 321.490597758924  3.35961875693895
#> 16                FALSE          0.75  778.65830081246 0.536773399218478
#> 17                FALSE          0.75 338.498298253309  8.34749160233859
#> 18                FALSE          0.75 350.860477431812  5.14963876492843
#> 19                FALSE          0.75 782.210556817047 0.709284121657018
#> 20                FALSE          0.75 356.099644687364  9.73682990877591
#> 21                FALSE          0.75  338.45200158842  5.13316344591979
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3411.24722550099 2124.03500276088 81453.7033540051 2.06281883164885
#> 2  1462.63797476845 272.127727704932 42097.4920325327 1.64730274160446
#> 3  2196.25294476574 679.691123035084 65604.3889889606 1.82080914455431
#> 4  7562.03209960662 6039.53742798539 149225.047542884 2.27317681755465
#> 5  2375.75520045355 1088.38492683147 57583.4209978505 2.77171861544739
#> 6  4817.52244642334 3212.09448206591 130404.379015732 3.12083461985601
#> 7   17013.236076021 15748.6007282464 360726.476733339 15.5445159980677
#> 8  5189.24692833734 3834.01846846686 118452.778458471 8.76600157651016
#> 9   7767.4264562342 6294.41368294446 221916.983074958 8.00614936147817
#> 10  20097.640214581  18616.505632629 485081.104891153 36.0412003374311
#> 11 6794.51752125626 5622.41905088835 168132.517403264 11.6893534993765
#> 12  9192.1820236026 7665.30377639916 267591.137938026 6.54295974369241
#> 13 20412.5917689538 19340.6534035032 489579.130599327 21.2048080571936
#> 14 6985.23876784523 5640.78512575051 177563.016921339 13.1937497437547
#> 15 9000.58118067132 7421.39150038558 264451.032300772 5.61830453744555
#> 16 20696.2176085327 19361.8998693845  515069.19336617 24.2394087405555
#> 17 6904.30306490606  5953.0291046068 176268.720090928 13.0372083306471
#> 18  9318.9197644709 7595.77818662494 271229.687142413 6.29204704227438
#> 19 19763.9443985546 18552.1180272094 504023.259843852 26.4434711495672
#> 20 7163.94827495834  5795.7150853801 181079.144570648 17.5875242232777
#> 21 9255.32155800189 7494.39889137575 267114.628175126 6.57639007145605
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1   0.61313051961135 28.6590984370647 65.9488425923293  1757.2101705871
#> 2   3.26269583399262 20.1635283655071 98.9444315756352 1099.56456035428
#> 3   2.05183876608946 4.06320205940095 139.481152958018 1746.36666711804
#> 4  0.203589574727544 68.6392192169424 82.3584259700981 3693.05005702616
#> 5   2.23657282936041  25.744377183223 110.494187286108  851.51992127552
#> 6  0.526475329536586 8.51061659527742 96.5561672574973 2325.56383053887
#> 7  0.537670855738735 140.171949926102 184.487014515698 7563.77262944335
#> 8  0.732025607275054 39.9875598885105 136.292272023664 2330.03988347029
#> 9   0.41596198398076 19.7785812899067  62.726117227625 2946.56374016766
#> 10 0.421430505734182 76.2582033688462  166.08430281283 7194.40255247781
#> 11 0.557867420764305 76.8271485631891  127.23435893916 2372.88804910499
#> 12 0.412013917619899 10.8317949256245  105.56763934093 3253.81755294292
#> 13 0.291030642525575 73.9419398165539 153.949601556199 8547.49913897322
#> 14 0.516883673443019 32.3210802456829 133.718993786485  2205.5823182875
#> 15 0.322321711600171 14.0988562664022 49.1134538401666 5987.79048599927
#> 16 0.426729084787974 75.1403049348616 309.059429394448  8182.9395993431
#> 17 0.751402438389577 77.0995417123516 210.094879582091 3270.60242784553
#> 18 0.625147660221864 9.73111602558074 186.373411031228 2901.33101619965
#> 19 0.372239019522593 58.1777672558584 201.797331159463  6346.6452390768
#> 20 0.476834299671405 69.4366706682624  120.17418594821  1582.3204670132
#> 21 0.389652239005256  9.8323502749559 107.802877604692 6538.73219586726
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 89.3043822059583  96.091056162083   14.5143107766211
#> 2             TRUE 30.2193805664624 35.6390065863411   26.1464412506411
#> 3             TRUE 62.1881317470528 68.1785938326365   16.8033835715687
#> 4             TRUE 239.590040481397 247.068792211151    14.693969125449
#> 5             TRUE 60.4760067530605 69.5949609978824   19.5574552888279
#> 6             TRUE 199.922260815778 210.189806715104   11.6068971382384
#> 7             TRUE 482.596606278507  533.73806391215   6.28884090332058
#> 8             TRUE 188.097400734908 216.937545921626   11.9311629240147
#> 9             TRUE 277.798886976596  304.13911837586   3.98730004353798
#> 10            TRUE 601.783877769176 720.359426879324 -0.051508808703622
#> 11            TRUE 301.342248826709 339.800221839658   8.73168742107788
#> 12            TRUE 323.077793457677 344.604131014425   3.08904790509384
#> 13            TRUE 679.953415643567 749.717234151734  0.148357046347801
#> 14            TRUE 323.073455237179 366.480891894132   9.02803537354722
#> 15            TRUE 312.248486794826 330.732708723022   2.82939954135667
#> 16            TRUE 738.784473434246 818.532128190673  -0.16519594525774
#> 17            TRUE 317.052090549395 359.944505957224   7.11143459118774
#> 18            TRUE  340.51006004727 361.210894816353   4.12127086386346
#> 19            TRUE 738.711046776009 825.710066858085  0.096950934542353
#> 20            TRUE 327.168167340072 385.031122034656   8.95243748581645
#> 21            TRUE 327.633839920875 349.270163255965   4.49218551275614
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.5315101861425 3364.10300857202 3458.39144242997    78563.0926233893
#> 2   36.8807105444768 1429.46897060719 1495.80697892971    40288.7083307499
#> 3    23.553933112003 2189.56897737803 2202.93691215346    62731.6158215514
#> 4   15.3637788263026 7449.12058399475 7674.94361521849    143149.980199076
#> 5   26.9157798974237 2333.40569998714 2418.10470091995    56182.6707273523
#> 6   13.3390009724137 4803.52248212411 4831.52241072258    126578.826514496
#> 7   8.05777801870102 16782.6532183926 17243.8189336494    348284.070757905
#> 8   14.3395271719496 5123.46739232074 5255.02646435394    114619.862850163
#> 9   5.35581497083468  7734.8906900123  7799.9622224561    217069.885722383
#> 10  1.33499755516184 19972.1954700393 20223.0849591228    473246.312692327
#> 11  10.5670712353924 6668.13686186981  6920.8981806427    164229.116562486
#> 12  4.44457369406331 9174.36372094994 9210.00032625525    262238.608063435
#> 13  1.10584786025694 20290.9572779556  20534.226259952    475518.494515716
#> 14  10.7285826591747 6932.07059084108 7038.40694484938    173934.834007756
#> 15  3.88983797252123 8977.38856211309 9023.77379922956    254601.116951303
#> 16   1.2387427436947 20572.6118069148 20819.8234101505    501608.257725251
#> 17  9.58354861348944 6777.47431878924 7031.13181102288    170888.579097122
#> 18   6.1780066659934 9302.91207860882 9334.92745033298    266456.997620765
#> 19  1.32161730877168 19668.2419714187 19859.6468256905     493583.02842557
#> 20  10.5212223317354 7049.72495170905 7278.17159820764    178476.227402412
#> 21  5.77414137908343 9239.14734179958 9271.49577420419    256358.413712925
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84344.3140846209 88.6545942739889 96.7408440940524    14.3211746629436
#> 2     43906.2757343154  29.700480202857 36.1579069499465    25.1186920629334
#> 3     68477.1621563698 61.6145768665182  68.752148713171    16.1570543602505
#> 4     155300.114886692 238.873989783867 247.784842908681    14.6298384094098
#> 5     58984.1712683487 59.6029153891946 70.4680523617483    18.8529348475794
#> 6     134229.931516968 198.939197910523 211.172869620359    11.4410574094343
#> 7     373168.882708773 477.700083739116 538.634586451541    6.11947458376288
#> 8      122285.69406678 185.336110238307 219.698836418227     11.700574857723
#> 9     226764.080427534 275.276949927731 306.661055424725    3.85627201858404
#> 10    496915.897089979 590.430899662886 731.712404985615  -0.184259418009889
#> 11    172035.918244042 297.660102474406 343.482368191962    8.55595918353713
#> 12    272943.667812617 321.016761138414 346.665163333688    2.95926352104358
#> 13    503639.766682938 673.273901105551  756.39674868975  0.0566823939522451
#> 14    181191.199834922 318.917424067896 370.636923063414    8.86521701641267
#> 15    274300.947650241 310.478720865531 332.502474652317    2.72786820220262
#> 16     528530.12900709 731.149059680971 826.167541943948  -0.299615606965952
#> 17    181648.861084734 312.945369925241 364.051226581378    6.87474282309502
#> 18    276002.376664062 338.528065228954  363.19288963467    3.92434935089358
#> 19    514463.491262133 730.381353363895 834.039760270198 -0.0203043566072636
#> 20    183682.061738885  321.62809720974 390.571192164988    8.80223468141995
#> 21    277870.842637328 325.562277048366 351.341726128474    4.36944505746949
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1     16.72464629982 3355.07539256435 3467.41905843764    78009.5714196544
#> 2   37.9084597321845 1423.11745917206 1502.15849036484    39942.3454942383
#> 3   24.2002623233212 2188.28906872932 2204.21682080217    62181.5103214093
#> 4   15.4279095423418 7427.49922994141 7696.56496927182    141986.669431113
#> 5   27.6203003386722 2325.29622117443 2426.21417973266    55914.4419521505
#> 6   13.5048407012178  4800.8416378966 4834.20325495009    125846.273907876
#> 7   8.22714433825872 16738.4990541658 17287.9730978762     345901.48237963
#> 8   14.5701152382412 5110.87131095586 5267.62254571882    113885.900286869
#> 9   5.48684299578861 7728.66043690598 7806.19247556242     216141.71814423
#> 10   1.4677481644681 19948.1741359781  20247.106293184    470980.075888296
#> 11  10.7427994729332 6643.93631007241 6945.09873244011    163481.656827018
#> 12  4.57435807811358 9170.95170554837 9213.41234165682    261213.655534258
#> 13   1.1975225126525 20267.6655669134 20557.5179709942    472826.032286939
#> 14  10.8914010163093 6921.88945056369 7048.58808512677    173240.075577495
#> 15  3.99136931167529 8972.94742238917 9028.21493895347    252714.962948214
#> 16  1.37316240540291 20548.9426108604  20843.492606205    499030.631751458
#> 17  9.82024038158216 6753.18796314985 7055.41816666227    169858.339332351
#> 18  6.37492817896328 9299.84677706076 9337.99275188104    265543.078350662
#> 19   1.4388725999213 19649.9159747331 19877.9728223761    491583.835175261
#> 20  10.6714251361319 7027.85240044855 7300.04414946814    177977.796455303
#> 21  5.89688183437009 9236.05015146297  9274.5929645408    254298.713071226
#>    ci95.integral.bt.up
#> 1     84897.8352883558
#> 2      44252.638570827
#> 3      69027.267656512
#> 4     156463.425654655
#> 5     59252.4000435505
#> 6     134962.484123588
#> 7     375551.471087048
#> 8     123019.656630073
#> 9     227692.248005687
#> 10     499182.13389401
#> 11     172783.37797951
#> 12    273968.620341794
#> 13    506332.228911714
#> 14    181885.958265182
#> 15    276187.101653331
#> 16    531107.754980883
#> 17    182679.100849505
#> 18    276916.295934165
#> 19    516462.684512442
#> 20    184180.492685994
#> 21    279930.543279026