Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xls', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xls', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.5231782286237  15.4702559323948
#> 2                 FALSE          0.75 33.4159657581744  30.2634705182304
#> 3                 FALSE          0.75 64.3076742349259  20.1352940100724
#> 4                 FALSE          0.75 242.721810020255  14.9240673024602
#> 5                 FALSE          0.75 64.6003939678496  22.3408269362902
#> 6                 FALSE          0.75 206.263641457007  12.3244807203037
#> 7                 FALSE          0.75 510.776321598088  7.07656189295079
#> 8                 FALSE          0.75 204.238033947513  13.3279575156638
#> 9                 FALSE          0.75 290.112808675987  4.76556729812286
#> 10                FALSE          0.75 667.621037526874 0.627993819368085
#> 11                FALSE          0.75 325.928244980635  9.94002141365937
#> 12                FALSE          0.75 334.420187994492  3.90442986798473
#> 13                FALSE          0.75 728.496257239213 0.712984589970736
#> 14                FALSE          0.75 346.471859684977  9.53947808718897
#> 15                FALSE          0.75 320.733408079708   3.4887074269994
#> 16                FALSE          0.75 772.515773409421 0.504687298442993
#> 17                FALSE          0.75 333.646650046631  8.16303571455677
#> 18                FALSE          0.75  351.29482780563  5.01785113824859
#> 19                FALSE          0.75 785.766661155426 0.777324560937443
#> 20                FALSE          0.75 352.725333668769  9.50431771534142
#> 21                FALSE          0.75 337.125847671856  5.30527639925368
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3407.28803610506 2126.69008101164 81244.8971148206 3.13534290027281
#> 2   1461.6569687073 305.676687934645 41926.5904588745 2.35428966083871
#> 3  2195.97038579215 691.275851173128 66313.4787853468 1.21249191976214
#> 4  7562.84605789615 6055.43715502814 149010.710613082 2.65564560029115
#> 5  2359.83870672309 1110.30406371592 57195.3598673881 2.37892638265049
#> 6  4820.89247142174 3250.03390326584 130877.083791332 2.47928164886027
#> 7  17057.7086361257 15881.9207194458 363738.931547158 9.24681051757336
#> 8  5167.23934306697 3790.15040881377 118242.244083302 7.67723119256935
#> 9  7758.74020096349 6270.60884884725 222361.759540849 6.19950839945385
#> 10 20090.3514187216 18687.4342639165 487574.878319984 28.4613813655906
#> 11 6800.53319599602 5584.42679934515 167892.153856092 11.7175926290875
#> 12 9196.40696196481 7628.87271179304 267030.161335654 6.56669918973931
#> 13 20414.0489512623 19373.1538906811 488146.493306998 23.2599919512964
#> 14 6983.17247432186 5746.00145673632 177271.229375369 13.9112796131087
#> 15 9005.58953833688 7385.11082184121 264700.749943678 7.41053327478354
#> 16 20636.3502469781 19276.9086791209 509556.490277711  22.860628220455
#> 17 6908.72430906172 5990.34978928897 176249.948761938 9.67462421427325
#> 18 9316.57747796697 7651.62949371108 269193.741786448 5.66101246310316
#> 19  19776.627192681 18530.6251781172 505181.910914863 25.3185435942612
#> 20 7109.75981965937  5810.6161763719   179594.4266691 18.2453496950564
#> 21 9255.98808831011 7428.32579511768  268095.91647092 8.04360249362252
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.483457480934868 30.7659426880701  54.203858127381 1611.90930642686
#> 2   2.82252575087127 22.3913824235478 85.2399696481976 1088.07105491363
#> 3   2.24166955155376 5.68936856879998 144.602265424025  986.57513574787
#> 4  0.248068856539834 55.8127769676367  95.147823584347 2943.90474000682
#> 5   2.64633698867503 39.2321244404308  120.56943976878 1399.92281352992
#> 6  0.555328161661889 10.4882410851095  111.21790139189 1422.75532505538
#> 7   0.28105349038902 121.342590260873 135.229099968671 6154.80490497672
#> 8  0.520550829797305 45.7850486363155 80.6068009400528 2891.67941672968
#> 9   0.36740332650329 23.6134684095445 81.5067366749891  2197.8229839317
#> 10 0.487092559283092 74.1735876410471 295.896869727574 5330.44801451903
#> 11 0.642356381919701 57.9944213095131 214.146895464597 2891.88763101945
#> 12 0.483902382023481 13.1336175819851  138.10464583714 3177.54905560228
#> 13 0.242062330432491 111.053329925586 189.612019365399 12698.6958285295
#> 14 0.520545799849185 42.9638444566809 147.895781459059 2427.33035377153
#> 15 0.494262819364232 14.6637960012778 120.973352532366 4337.48639467168
#> 16 0.359259895575946  183.10569678532 271.863959687041 18223.7738178745
#> 17 0.358072302097747  73.522457598849 115.262985476329 2496.72782517258
#> 18 0.466273017944954 12.2445631829993 145.323859207456 4296.41383118408
#> 19 0.537698315271572 48.4587001899404 334.798903996833  5695.8590011497
#> 20 0.447259711510777 84.9887817330315 121.001708429649 2824.64815229116
#> 21  0.63160782349451 11.5283488558757 189.327392269107  3348.0600890275
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 87.3655391576749 97.6808172995725     14.674968376257
#> 2             TRUE 29.5431592660948 37.2887722502541    25.6204156580472
#> 3             TRUE 62.3131250269172 66.3022234429346    16.4477475977665
#> 4             TRUE 238.353273007776 247.090347032734    14.5159940334522
#> 5             TRUE 60.6870600683895 68.5137278673097    17.9876025899198
#> 6             TRUE 202.185223144632 210.342059769382    11.4109658943699
#> 7             TRUE  495.56531829668 525.987324899497    6.61422890126085
#> 8             TRUE 191.608988635736 216.867079259289    12.4716514006473
#> 9             TRUE 279.914617358885 300.310999993089    4.16118882602495
#> 10            TRUE 620.802065180478 714.440009873271  -0.173273440652602
#> 11            TRUE 306.652805105786 345.203684855484    8.88334516540146
#> 12            TRUE 323.617967827371 345.222408161613    3.10841044955611
#> 13            TRUE  690.23357047933 766.758943999095   0.314792056409288
#> 14            TRUE 323.587804721413 369.355914648541    8.68318024643706
#> 15            TRUE 308.543080842689 332.923735316727    2.67564508914524
#> 16            TRUE 734.910039986773 810.121506832069 -0.0862952297794384
#> 17            TRUE 317.731893214152 349.561406879111    7.57400677760598
#> 18            TRUE 341.982462303825 360.607193307434    4.25083202372914
#> 19            TRUE 744.117656942866 827.415665367985  -0.107189167684293
#> 20            TRUE 322.711733420401 382.738933917137    8.76857548990619
#> 21            TRUE 323.894121569847 350.357573773865    4.26628152960522
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.2655434885327 3356.67806038319 3457.89801182694    78593.3063057484
#> 2   34.9065253784137 1424.82314462056 1498.49079279403    40136.7135735416
#> 3   23.8228404223784 2186.61137449648 2205.32939708783    64690.5626870416
#> 4   15.3321405714682 7471.03403978439 7654.65807600791     144167.98731577
#> 5   26.6940512826606 2295.30186201858  2424.3755514276    54892.4868391314
#> 6   13.2379955462375 4803.63931483673 4838.14562800674    128536.651281616
#> 7   7.53889488464073 16858.1000751466 17257.3171971048    353614.277478471
#> 8   14.1842636306804 5091.92293806023 5242.55574807371    113485.431442782
#> 9   5.36994577022078 7719.89604542979 7797.58435649719    218746.340732281
#> 10  1.42926107938877  19968.335867052 20212.3669703911      478806.2913361
#> 11  10.9966976619173 6705.13237294187 6895.93401905017    163134.998703065
#> 12  4.70044928641336 9174.80216104244 9218.01176288717    261803.093139188
#> 13  1.11117712353218 20231.3662235347 20596.7316789898    467257.138669067
#> 14  10.3957759279409 6912.49695019062  7053.8479984531    173278.270943415
#> 15  4.30176976485356 8981.46759391477 9029.71148275898    257565.584824443
#> 16  1.09566982666542 20335.1413757662 20937.5591181899    479578.382347308
#> 17  8.75206465150756 6787.77986631161 7029.66875181183    172142.831489529
#> 18  5.78487025276804 9296.43517153094 9336.71978440301     262126.14103415
#> 19  1.66183828955918 19696.9126308686 19856.3417544935    495812.222857971
#> 20  10.2400599407766 6969.95327370853 7249.56636561021    174947.880458581
#> 21  6.34427126890215 9237.02395444219 9274.95222217802    262588.357624469
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83896.4879238928  86.377906144089 98.6684503131584   14.5226792697625
#> 2     43716.4673442074 28.8015580229306 38.0303734934183   24.7313200465227
#> 3     67936.3948836521 61.9311900721921 66.6841583976597   15.7416216890271
#> 4     153853.433910393 237.516744643685 247.926875396826   14.4378523436421
#> 5     59498.2328956448 59.9376982578546 69.2630896778445   17.1540064384871
#> 6     133217.516301048 201.404249425241 211.123033488773   11.2360375234464
#> 7     373863.585615844 492.652572983645 528.900070212532   6.52569705178831
#> 8     122999.056723822 189.190660810077 219.285407084949   12.3076778892611
#> 9     225977.178349416 277.961772213057 302.263845138917   4.04545677817642
#> 10    496343.465303868 611.836730050317 723.405345003432 -0.326707596826776
#> 11    172649.309009119 302.961763427624 348.894726533647   8.68100290509676
#> 12     272257.22953212 321.549457582603 347.290918406381   2.95598119921871
#> 13    509035.847944929 682.906673014672 774.085841463754  0.238542422323053
#> 14    181264.187807323 319.205751643284  373.73796772667   8.51920831948457
#> 15    271835.915062912 306.208762861132 335.258053298284   2.51995230104551
#> 16    539534.598208115 727.708942097329 817.322604721513 -0.199462096885861
#> 17    180357.066034347 314.684386586656 352.608913506607   7.46121400244519
#> 18    276261.342538746 340.199243377947 362.390412233312   4.10395602307648
#> 19    514551.598971754 736.142315710674 835.391006600178 -0.276564136994839
#> 20    184240.972879619 316.964448266459  388.48621907108   8.62768868078029
#> 21     273603.47531737 321.360386784356 352.891308559356   4.06732506520444
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.4178325950272 3346.98678843645 3467.58928377368     78085.554874224
#> 2   35.7956209899381 1417.76985915714 1505.54407825745    39793.9711912438
#> 3   24.5289663311178 2184.81922339731   2207.121548187     64379.791519281
#> 4   15.4102822612783 7453.45301503958 7672.23910075272    143240.657322668
#> 5   27.5276474340933 2282.94374281985 2436.73367062633    54451.5111528694
#> 6    13.412923917161 4800.33551889492 4841.44942394855    128088.483354223
#> 7   7.62742673411327 16819.8771592144  17295.540113037    351675.513933403
#> 8   14.3482371420665 5077.50064773979 5256.97803839415    112574.552426512
#> 9   5.48567781806931 7712.45780288079  7805.0225990462    218054.026492343
#> 10  1.58269523556295 19944.9711869451  20235.731650498    477127.200211527
#> 11   11.199039922222 6686.86413022938 6914.20226176267    162224.054099293
#> 12  4.85287853675076 9170.66507150412  9222.1488524255    260802.165186674
#> 13  1.18742675761842 20196.3844246081 20631.7134779164    463257.049483081
#> 14  10.5597478548934 6898.96333918676 7067.38160945695    172513.661881976
#> 15   4.4574625529533 8976.84849817437 9034.33057849938    256199.276610121
#> 16  1.20883669377185 20277.4630812788 20995.2374126773    473837.893594677
#> 17  8.86485742666835 6764.62029216798 7052.82832595547      171356.3622246
#> 18   5.9317462534207 9292.57813412829 9340.57682180565    260772.770677327
#> 19  1.83121325886973 19681.6481403087 19871.6062450533    494018.027272609
#> 20  10.3809467499025 6943.18180746263 7276.33783185611    174058.116290609
#> 21  6.54322773330292 9233.39252455259 9278.58365206762    261533.718696426
#>    ci95.integral.bt.up
#> 1     84404.2393554173
#> 2     44059.2097265052
#> 3     68247.1660514127
#> 4     154780.763903495
#> 5     59939.2085819067
#> 6     133665.684228441
#> 7     375802.349160912
#> 8     123909.935740092
#> 9     226669.492589355
#> 10    498022.556428441
#> 11     173560.25361289
#> 12    273258.157484635
#> 13    513035.937130916
#> 14    182028.796868761
#> 15    273202.223277234
#> 16    545275.086960746
#> 17    181143.535299276
#> 18    277614.712895569
#> 19    516345.794557116
#> 20    185130.737047591
#> 21    274658.114245413