Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.8724405742403  15.8759854575979
#> 2                 FALSE          0.75 32.8201236479325  31.3420633684096
#> 3                 FALSE          0.75 65.2871400068065  20.3017092852331
#> 4                 FALSE          0.75 242.388943893982  15.0098760725499
#> 5                 FALSE          0.75 65.9970299200472  22.9540304203824
#> 6                 FALSE          0.75 204.239271047084  12.3248526808138
#> 7                 FALSE          0.75 508.062236553799  7.08706615080629
#> 8                 FALSE          0.75 200.360342476979  13.3882171235139
#> 9                 FALSE          0.75 290.011905837765  4.66371879138824
#> 10                FALSE          0.75 672.812020232521 0.551009841131205
#> 11                FALSE          0.75 322.743480902459  9.84198886522293
#> 12                FALSE          0.75 335.839438760847  3.89426172691887
#> 13                FALSE          0.75 730.284685980688 0.768313585323411
#> 14                FALSE          0.75 343.435409804173  9.83108506141913
#> 15                FALSE          0.75 318.829025358811  3.35685033966451
#> 16                FALSE          0.75  783.49259174694 0.582273754734896
#> 17                FALSE          0.75 332.760745370747  8.02276897719239
#> 18                FALSE          0.75 350.365935690726  4.86901296992606
#> 19                FALSE          0.75 780.266658308542 0.545022684257704
#> 20                FALSE          0.75 351.027338515928   9.6577746791906
#> 21                FALSE          0.75  339.52212473174  5.34303847771906
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3416.04221621051 2095.44642265021 81550.9522016247 1.91996674548426
#> 2  1468.34696063478  298.99152045355 42470.3606742624 2.20650778891898
#> 3  2194.95782005692 669.713586856349 66306.5652946901 1.42637597916928
#> 4  7529.02622081694 6001.09757992651  147723.29773401 2.69621363961547
#> 5   2371.7952853134  1107.5344474986 57331.3644650867 2.12167284735712
#> 6  4820.13670961258 3245.23084950612 130792.265011639 2.56993126686241
#> 7  17125.1821756862 15895.5825518165 366301.324352174 14.1833091693661
#> 8  5177.33607063806 3769.36626181086 117975.545324648 8.11313910578146
#> 9  7757.98177516077 6289.52605504158 221406.000395642 6.64528075848505
#> 10  20077.711146685 18743.0254119578 480557.973578731 28.6316958218146
#> 11 6744.07738663497 5553.15647425156  166240.28921926 11.7049352776566
#> 12 9191.76804846877 7647.35413444856 265666.280013716 5.77566865010654
#> 13 20495.9032345754 19453.2776285958 494239.822582505 26.5499867475687
#> 14 6958.47612665318 5637.17982871558 176790.307830184 15.7652480703195
#> 15  9002.4194592676 7391.56786537891 264779.309129344 5.80062623943707
#> 16 20686.7305558437  19322.661540689 511326.095745888 22.1920451618693
#> 17 6917.67919536389 6043.62026683729 176585.150319147 13.1202188182169
#> 18 9321.33086748085 7686.10683527027 270247.511612993 6.69410926218101
#> 19 19801.3537871366 18707.1633367731 504470.359242063 23.8906742842938
#> 20 7141.92003925734 5774.99841752567 180444.093267966 13.9204420610074
#> 21 9257.45085404311 7454.67127699353 266566.734088329 7.93910930518753
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.625565455792842 29.7324823498259 61.3871247371734 1734.51522131898
#> 2   2.25493626802141 23.1652603341349 68.6009919572175 759.806705602157
#> 3   2.24325359539554 5.55517007183236 147.036117715389 1158.62666816982
#> 4  0.370688885772261 87.6117388646837 95.3055190531728 4506.34219694604
#> 5   2.83775481295444 32.6388239756125 132.113771633703 1346.45977961372
#> 6  0.516374023010206 9.46601484968573 106.135093902549 2177.58909924224
#> 7  0.570677722586722 46.8840220491575 119.477513752339  2686.5196044753
#> 8  0.675734963519917 60.0943870211129 102.718831295523 3238.24997147977
#> 9  0.384905208886458 24.9469379486736 76.2303713134865 3012.33039669864
#> 10  0.34888271861188 98.0573546782808 271.130584745831 12811.7864282141
#> 11 0.510807006116429 126.483478430877  220.10692971086 5518.34234513669
#> 12 0.537774068357522 11.3539218720877 135.716038719937 5245.60813664373
#> 13 0.428809926066069 71.8177110966232 288.291754146163 5457.23102199689
#> 14 0.496620169322608 40.9186419368435 156.019661077028 2596.14097605407
#> 15 0.412549600983776 24.1061923998451 116.662940902373 4842.94753361308
#> 16 0.453623094842366 95.9859923793639 283.497660272407 9996.80642757274
#> 17 0.494956579833981 54.0810932882207 151.305397583392 1997.93102624724
#> 18 0.434928273157048  8.8049239741626 94.8680897984714 4346.59186859763
#> 19 0.270303516701088 82.7624765897174 180.957248189047 7712.81633422567
#> 20 0.417849578257133 98.0228643607649 128.984120683046 3604.94553303628
#> 21 0.565807660528896 11.3701262874762 156.143814201984 4859.61051298203
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 89.7140952779187 96.0307858705619    14.8469302828187
#> 2             TRUE 29.1904183351608 36.4498289607042    27.6326932075144
#> 3             TRUE  62.940751521073 67.6335284925399    16.6115571208074
#> 4             TRUE 237.953672456815  246.82421533115    14.4000928554546
#> 5             TRUE 62.5068780861447 69.4871817539497    18.2859237530723
#> 6             TRUE 200.011734113095 208.466807981072    11.4754174129621
#> 7             TRUE 484.730692970192 531.393780137407    6.14830129715113
#> 8             TRUE 187.014228647968 213.706456305989    12.2766331085236
#> 9             TRUE 279.080418990057 300.943392685472    4.03054972277002
#> 10            TRUE 625.712880605636 719.911159859406 -0.0229022309853383
#> 11            TRUE 303.488862370714 341.998099434204     9.0017113401614
#> 12            TRUE 326.338463831422 345.340413690272    3.00962338447075
#> 13            TRUE 686.609957780938 773.959414180439  0.0629212569447282
#> 14            TRUE 317.501576728498 369.369242879849    9.01414488288344
#> 15            TRUE 309.286995194937 328.371055522685     2.6782062460462
#> 16            TRUE 746.986677455665 819.998506038215  -0.163936236280797
#> 17            TRUE  311.17798541478 354.343505326714    7.20856540336549
#> 18            TRUE 339.354125954438 361.377745427013    4.15355596058272
#> 19            TRUE 740.966499110878 819.566817506205   0.100373399284414
#> 20            TRUE 328.128211325571 373.926465706285    8.97041212295761
#> 21            TRUE 326.462289924706 352.581959538773    4.41228487614903
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.9050406323771 3367.13228274505 3464.95214967598     78697.674662555
#> 2   35.0514335293049 1430.24010738513 1506.45381388443    41220.4786435469
#> 3   23.9918614496587 2185.81956528875 2204.09607482508    64400.6244255508
#> 4   15.6196592896453 7384.90491038454 7673.14753124935    140310.364820034
#> 5   27.6221370876924 2318.10441987351 2425.48615075328    55116.4381276221
#> 6   13.1742879486656 4804.56511518484 4835.70830404031    127210.130943386
#> 7   8.02583100446145 17048.0579594154 17202.3063919571    361881.999602812
#> 8   14.4998011385042 5078.48080398833 5276.19133728779    112648.624121563
#> 9   5.29688786000646  7716.9440622352 7799.01948808634    216450.716893073
#> 10  1.12492191324775 19916.4067982392 20239.0154951308    459482.584904319
#> 11  10.6822663902845 6536.01206461618 6952.14270865376     157162.61606151
#> 12  4.77890006936699 9173.09084698918 9210.44524994835    257037.254628937
#> 13  1.47370591370209 20377.7630998215 20614.0433693293     485262.67755132
#> 14  10.6480252399548 6891.16496066707 7025.78729263928    172519.655924575
#> 15  4.03549443328282 8962.76477276985 9042.07414576534     256812.66043655
#> 16  1.32848374575059 20528.8335983796 20844.6275133077    494881.349172531
#> 17  8.83697255101928 6828.71579690477 7006.64259382302     173298.55378097
#> 18  5.58446997926941 9306.84676754335 9335.81496741835     263097.36798915
#> 19 0.989671969230994 19665.2095131465 19937.4980611267    491782.776372262
#> 20  10.3451372354236 6980.67242738388  7303.1676511308    174513.957866122
#> 21   6.2737920792891 9238.74699630022 9276.15471178601    258572.674794473
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84404.2297406945 89.1093057530912 96.6355753953895    14.6498771642439
#> 2      43720.242704978 28.4953683816513 37.1448789142137    26.9223882830877
#> 3     68212.5061638295 62.4914430876347 68.0828369259783    15.9049322382578
#> 4     155136.230647986 237.104365160336 247.673522627628    14.2833258564363
#> 5     59546.2908025512 61.8385511392272 70.1555087008671    17.3920309869917
#> 6     134374.399079893 199.202205764033 209.276336330134    11.3127595957138
#> 7     370720.649101536 480.262950581842 535.861522525757    5.96853781453631
#> 8     123302.466527732 184.458589829647  216.26209512431    12.0637765950149
#> 9     226361.283898211 276.987155551134 303.036656124395    3.90930458197078
#> 10    501633.362253143 616.693896421765 728.930144043278   -0.13280028734808
#> 11    175317.962377009 299.801807758252 345.685154046666    8.84080713323473
#> 12    274295.305398495 324.519128206638 347.159749315056    2.84022455293813
#> 13     503216.96761369 678.246711955454 782.322660005923 -0.0721538697660834
#> 14    181060.959735793 312.535523586347    374.335296022    8.85770952954682
#> 15    272745.957822137 307.459797929514 330.198252788108    2.54825312173631
#> 16    527770.842319245 739.996183229676 826.989000264204  -0.306827511156142
#> 17    179871.746857323 307.045116487042 358.476374254452    7.05265408071778
#> 18    277397.655236836 337.245481536851   363.4863898446    4.01655355453825
#> 19    517157.942111864 733.440936711326 827.092379905758   0.015227791523571
#> 20    186374.228669811 323.743272076353 378.311404955502    8.83878950580662
#> 21    274560.793382184 323.961470493572 355.082778969908    4.23405546308243
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.1020937509519 3357.76655080486 3474.31788161617    78151.3023678395
#> 2   35.7617384537316 1422.94305037988 1513.75087088969    40981.1395312822
#> 3   24.6984863322083 2184.06968671612 2205.84595339771    64035.6570250773
#> 4   15.7364262886636 7357.30721264216 7700.74522899172    138890.867027996
#> 5   28.5160298537731  2307.8231903212  2435.7673803056    54692.3032970438
#> 6   13.3369457659138 4801.58332050719 4838.69009871796    126524.190377125
#> 7   8.20559448707627 17033.2894924699 17217.0748589026    361035.745927402
#> 8   14.7126576520129 5059.55107207668 5295.12106919944    111628.575380547
#> 9    5.4181330008057 7709.08577678137 7806.87777354017    215501.832818113
#> 10  1.23481996961049 19885.5187315156 20269.9035618544    455446.872179432
#> 11  10.8431705972111 6496.16976891045 6991.98500435949    155424.338222792
#> 12  4.94829890089961 9169.51436159948 9214.02173533806    255384.888065894
#> 13  1.60878104041291  20355.140520826 20636.6659483248    483543.649779391
#> 14  10.8044605932914 6878.27558845696 7038.67666484939    171701.871517118
#> 15  4.16544755759271  8955.1713221639 9049.66759637129    255287.131963462
#> 16  1.47137502062593 20498.5980107801 20874.8631009072    491732.355147845
#> 17  8.99288387366699 6811.68025251898 7023.67813820881    172669.205507702
#> 18  5.72147238531388 9304.07321649149 9338.58851847021    261728.191550542
#> 19  1.07481757699184 19639.1393330208 19963.5682412525    489353.239226981
#> 20  10.4767598525746 6949.79522511024 7334.04485340444    173378.400023215
#> 21   6.4520214923557 9235.16540651966 9279.73630156657    257041.897482884
#>    ci95.integral.bt.up
#> 1     84950.6020354099
#> 2     43959.5818172427
#> 3      68577.473564303
#> 4     156555.728440024
#> 5     59970.4256331296
#> 6     135060.339646154
#> 7     371566.902776946
#> 8     124322.515268748
#> 9     227310.167973171
#> 10    505669.074978031
#> 11    177056.240215727
#> 12    275947.671961537
#> 13    504935.995385619
#> 14     181878.74414325
#> 15    274271.486295225
#> 16     530919.83634393
#> 17    180501.095130591
#> 18    278766.831675445
#> 19    519587.479257145
#> 20    187509.786512718
#> 21    276091.570693774