Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.7571874621317  15.4951592846805
#> 2                 FALSE          0.75 33.6883675932537  30.0435321725322
#> 3                 FALSE          0.75 65.7003016331875  20.4375137287128
#> 4                 FALSE          0.75 242.804643660357   15.120666562135
#> 5                 FALSE          0.75 65.0978041358775   22.625580976285
#> 6                 FALSE          0.75 205.663389645023  12.4513605101749
#> 7                 FALSE          0.75 511.280601956544  7.16282090330433
#> 8                 FALSE          0.75 200.631649938206  13.3957926427753
#> 9                 FALSE          0.75 288.397209147271  4.72336098803318
#> 10                FALSE          0.75 680.079883766004 0.773277087930799
#> 11                FALSE          0.75 324.555540141284  9.82765163794777
#> 12                FALSE          0.75 335.394880416315  3.86990322944088
#> 13                FALSE          0.75 737.549989502854 0.698544460400932
#> 14                FALSE          0.75  347.95622227628  9.96840361856203
#> 15                FALSE          0.75 319.478458090499  3.33519176457119
#> 16                FALSE          0.75 772.712782102873 0.470223116259992
#> 17                FALSE          0.75 335.883636820561  8.05934706653639
#> 18                FALSE          0.75 349.135163042602  4.96775032673874
#> 19                FALSE          0.75 773.448941458843 0.723675466498743
#> 20                FALSE          0.75 349.278252288334    9.513730213528
#> 21                FALSE          0.75 338.048200830425  5.09698123035893
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3422.15075496419 2130.66940959556 82215.8029853642 2.28521781444277
#> 2  1460.38036023162 301.487215219054 42116.9898974438 2.76862686375399
#> 3  2193.90028078529 667.528301851681  66089.134443513 1.60099057732196
#> 4  7577.53895926798 6025.11799911221 150359.316792598  2.9290004408476
#> 5  2366.93025392241 1134.75960857041 57308.3258727914 2.67083481036196
#> 6  4820.49910414923 3224.31050626599   130750.4940858 2.32539691342454
#> 7  17093.5763886169 15896.3565267266 364795.204143048 10.8738345401162
#> 8   5202.6191736305 3797.53745496955  119679.09037532 8.52699859678305
#> 9  7767.09056363676 6248.50602200514 220041.794863022 6.35336062278117
#> 10 20079.0891987485 18625.5293818014 487147.469020559 24.8805021364351
#> 11 6776.32973312641 5560.82653009054 167278.913123103 16.0854748446625
#> 12 9183.10626943126 7633.76790538599 266372.755931256 5.96285455624501
#> 13 20437.3934811962 19449.5355468449 490747.509024971 28.5582021267503
#> 14 6962.11121487829 5619.92327961751  176226.69313968 12.3349402998471
#> 15 9001.80418640475  7411.6566978309 265651.991426469 5.33395037753728
#> 16 20657.3915896562 19325.2407691848  513025.54623458 26.1176899905668
#> 17 6866.13883150174 5986.25628779098 174853.017439626 14.0892969276682
#> 18 9322.39225530914 7633.65237395932 267261.410229311 6.57429433802513
#> 19 19777.1000524336 18526.1877629777 495223.347626551 32.1251009964011
#> 20 7135.81339906964 5796.37057598962 180536.617413617 14.0914388632565
#> 21 9254.72505407317 7493.65500248359 268119.257461069 8.42258794597612
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.705540771608567 13.8700474320123 66.4142350740106 993.801654301595
#> 2   2.72205965110038  21.992631050674 89.8034158881178 927.284450259652
#> 3   1.67802559749398 5.01830832458774 115.895872814242 1172.48452199526
#> 4  0.338989268801685 44.7443122727414 78.2749390194751 2566.18127652222
#> 5   2.96563853871121 28.2892824661541 170.649457047427 1091.23987039254
#> 6  0.395181394290108 12.6135117362313  79.467802753223 1894.28426545992
#> 7  0.311898627054443 79.4653740675939 94.9994229525587 4245.55799111226
#> 8  0.744008741502491 20.6082077846747 132.146430238172 1586.11694614267
#> 9  0.358062364664758 20.6018419941618 62.3974435949375 3193.13926154447
#> 10 0.353960326763955  64.283012096253 274.629745653646 6327.60796076254
#> 11 0.872160527891366 84.9628533799471 219.732555663291 3089.95017397695
#> 12 0.441474069831838 10.5378817565086 116.118303539726 6182.12654832107
#> 13 0.309790438388049 79.0487673399243 230.911377883117 8030.80634057796
#> 14 0.727775281624496   55.90727628134 225.423078447555  3546.1102469235
#> 15 0.400105826211087 23.5123520745434 106.174251828912  2648.9353270664
#> 16 0.352009080269672 117.449301216733 233.342550180491 12522.1213640306
#> 17 0.483877154860822 112.216324074938 150.997187996046 4589.36568016936
#> 18 0.435080405597402 9.39182171694711 131.921988006696 6525.02347807843
#> 19 0.361658047030798 78.8023121993293 202.261694319208 12827.2520091036
#> 20  0.41475641192724 85.6591488704806 120.794674975069 2736.77655577445
#> 21 0.635909606105566 9.85130440187095 160.193386684918 3526.90664324438
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.9980041573734 95.5163707668901   14.3345447153844
#> 2             TRUE 29.1339764023784  38.242758784129   25.5657440464721
#> 3             TRUE 63.0666721334928 68.3339311328821   17.6771616208352
#> 4             TRUE 237.986437935163 247.622849385551   14.5630292149562
#> 5             TRUE 60.7042808728321 69.4913273989229   17.7471055801051
#> 6             TRUE  201.83811172244 209.488667567606   11.8012871165676
#> 7             TRUE 493.393144138053 529.168059775035   6.64974766179977
#> 8             TRUE 186.604737246498 214.658562629914   12.1718982630037
#> 9             TRUE 277.945930922796 298.848487371746   4.13434839815966
#> 10            TRUE 639.151457751569  721.00830978044  0.191012350404093
#> 11            TRUE 298.094934021814 351.016146260754   8.39294756956647
#> 12            TRUE 325.585984671292 345.203776161338    3.1436783845675
#> 13            TRUE  690.57174700435 784.528232001358  0.188939189252591
#> 14            TRUE 327.665245483031 368.247199069528   8.77121328028973
#> 15            TRUE  310.70410971945 328.252806461548   2.67701768045395
#> 16            TRUE  729.74918206839 815.676382137355 -0.108831820783618
#> 17            TRUE 312.706743374547 359.060530266575   7.26336914679034
#> 18            TRUE  338.32044885655 359.949877228653   4.25204305953101
#> 19            TRUE 720.603150319763 826.294732597923   0.12874797913308
#> 20            TRUE 326.097835358277 372.458669218391   8.83145591590769
#> 21            TRUE 324.193043659294 351.903358001556   4.05090992831528
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6557738539766 3399.33452693853 3444.96698298985    80580.9992640381
#> 2   34.5213202985924 1424.20248215326 1496.55823830998    40591.6069767667
#> 3   23.1978658365904 2185.64516359134 2202.15539797924    64160.3974048308
#> 4   15.6783039093138 7503.93456557932 7651.14335295664    146137.948592719
#> 5    27.504056372465 2320.39438426558 2413.46612357923    55513.2362859957
#> 6   13.1014339037821 4799.74987734313 4841.24833095533    127634.396469118
#> 7   7.67589414480889 16962.8558482757 17224.2969289581    357811.261247668
#> 8   14.6196870225469 5168.71867182471 5236.51967543629    117069.927998915
#> 9   5.31237357790671 7733.20053355636 7800.98059371715    214789.080777781
#> 10  1.35554182545751 19973.3436438501 20184.8347536468    476738.553925104
#> 11  11.2623557063291  6636.5658393164 6916.09362693642    162195.945086911
#> 12  4.59612807431425  9165.7714539418 9200.44108492072    256203.157759267
#> 13  1.20814973154927  20307.358258922 20567.4287034704     477536.83259472
#> 14  11.1655939568343 6870.14374539549  7054.0786843611    170393.341783491
#> 15  3.99336584868842 8963.12636724212 9040.48200556737    261294.492813445
#> 16   1.0492780533036 20464.1874891546 20850.5956901577     492426.65659075
#> 17  8.85532498628244 6681.54297839847 7050.73468460501    167303.510895748
#> 18  5.68345759394646 9306.94270858477 9337.84180203352    256527.746607872
#> 19  1.31860295386441 19647.4702488657 19906.7298560015    474122.518071575
#> 20  10.1960045111483  6994.9040991777 7276.72269896158    176034.619979369
#> 21  6.14305253240259 9238.51965833209 9270.93044981425    262317.496032932
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83850.6067066903 87.2781605458239 96.2362143784395   14.1122993723277
#> 2     43642.3728181209 28.2618589402959 39.1148762462115   24.7082952563755
#> 3     68017.8714821952 62.5623601016364 68.8382431647385   17.1485835576246
#> 4     154580.684992478 237.063802796296 248.545484524418   14.4562475952837
#> 5     59103.4154595871 59.8629679075681  70.332640364187   16.8129294404111
#> 6     133866.591702481 201.105611694711 210.221167595335   11.6768049773663
#> 7     371779.147038428 489.967886257916 532.593317655172   6.55149959427762
#> 8     122288.252751725 183.918732688511 217.344567187901   11.9375355094304
#> 9     225294.508948262  275.94462232662 300.849795967922   4.02155875329026
#> 10    497556.384116013 631.314099578591 728.845667953417 0.0795148474734477
#> 11    172361.881159296 293.028009445746 356.083070836823   8.11821700328069
#> 12    276542.354103244 323.707685486075 347.082075346555   3.00461405257048
#> 13    503958.185455222 681.575913334423 793.524065671284  0.091355201160356
#> 14    182060.044495869 323.779739288579  372.13270526398   8.54196406657802
#> 15    270009.490039494 309.023915350526 329.933000830472   2.55098434519746
#> 16     533624.43587841 721.522109721362 823.903454484384 -0.219714681068565
#> 17    182402.523983505 308.268614842331  363.49865879879   7.11094784300918
#> 18     277995.07385075 336.249546140072 362.020779945131   4.11499273176783
#> 19    516324.177181526 710.483743505897 836.414139411789 0.0148256943183783
#> 20    185038.614847866 321.659032116351 376.897472460316   8.70080764615061
#> 21    273921.018889206 321.539928456312 354.556473204538   3.85059840239202
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8780191970333 3394.96546199744 3449.33604793093     80267.951742933
#> 2    35.378769088689  1417.2748033723 1503.48591709094    40299.5123749349
#> 3    23.726443899801  2184.0643964691 2203.73616510148    63791.0647804023
#> 4   15.7850855289863 7489.84010721341 7665.23781132255    145329.601490615
#> 5    28.438232512159 2311.48326028875 2422.37724755607     55169.495726822
#> 6   13.2259160429835 4795.77662114622 4845.22158715225    127037.696925498
#> 7   7.77414221233103 16937.8242554444 17249.3285217894    356473.910480468
#> 8   14.8540497761202 5162.22708637254 5243.01126088846     116570.30116088
#> 9   5.42516322277611  7726.7109533282 7807.47017394531    213783.241910395
#> 10  1.46703932838815 19953.0944950398 20205.0839024571    474745.357417464
#> 11  11.5370862726148 6609.80254050172 6942.85692575111    161222.610782109
#> 12  4.73519240631128  9162.4520211885 9203.76051767402    254255.787896546
#> 13  1.30573371964151   20282.45789721 20592.3290651825    475007.128597438
#> 14   11.394843170546 6852.53295336687 7071.68947638972     169276.31705571
#> 15  4.11939918394492 8955.71997633864 9047.88839647085    260460.078185419
#> 16  1.16016091358855 20427.1909592714  20887.592220041     488482.18836108
#> 17   9.0077462900636 6646.19483631486 7086.08282668862    165857.860706495
#> 18  5.82050792170965 9303.98428474393 9340.80022587436    254472.364212278
#> 19  1.43252523867911 19622.6475205229 19931.5525843443    470081.933688708
#> 20  10.3266527809054 6967.92146728349 7303.70533085578      175172.5353643
#> 21  6.34336405832584  9235.4164974455 9274.03361070084     261206.52044031
#>    ci95.integral.bt.up
#> 1     84163.6542277953
#> 2     43934.4674199527
#> 3     68387.2041066237
#> 4     155389.032094582
#> 5     59447.1560187608
#> 6     134463.291246101
#> 7     373116.497805628
#> 8      122787.87958976
#> 9     226300.347815649
#> 10    499549.580623653
#> 11    173335.215464098
#> 12    278489.723965965
#> 13    506487.889452504
#> 14     183177.06922365
#> 15     270843.90466752
#> 16     537568.90410808
#> 17    183848.174172758
#> 18    280050.456246345
#> 19    520364.761564394
#> 20    185900.699462935
#> 21    275031.994481828