Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75  92.775608237741  15.7409800718191
#> 2                 FALSE          0.75 33.4109000378685  31.5224631200234
#> 3                 FALSE          0.75 64.8841027376042  20.0942909626943
#> 4                 FALSE          0.75 242.259915434419  15.0172834597677
#> 5                 FALSE          0.75 64.8238882197752  23.3084589898613
#> 6                 FALSE          0.75 204.915127210045  12.3463530423007
#> 7                 FALSE          0.75 504.684780619311   6.8915200194643
#> 8                 FALSE          0.75 202.380062987056  13.1443762422786
#> 9                 FALSE          0.75 287.119443946493  4.60593782830531
#> 10                FALSE          0.75  673.30386707868 0.594518086376743
#> 11                FALSE          0.75 328.035656720339  9.98047014310213
#> 12                FALSE          0.75 334.224033791818   3.7764093357297
#> 13                FALSE          0.75 721.124052612582 0.767613949555473
#> 14                FALSE          0.75  347.32191792578  9.85103512614435
#> 15                FALSE          0.75 319.893055735271  3.28829579587035
#> 16                FALSE          0.75 770.701616044164 0.340761938044127
#> 17                FALSE          0.75 334.311816033485  8.05567907019844
#> 18                FALSE          0.75 349.574045542828  5.01138109079434
#> 19                FALSE          0.75 781.385481487026 0.626230484887226
#> 20                FALSE          0.75 355.017457046779  9.66662912836396
#> 21                FALSE          0.75 336.923317380976  5.32322375500246
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3404.02820402761 2097.61099398145 80805.8983080923 2.24864300381661
#> 2  1471.48837381416 282.540405724552 42511.9077489214 3.07772882752321
#> 3  2195.23004701917  690.17481839639 66244.2439446472  1.4688635023272
#> 4  7577.22663448244 6051.45949302852 149565.483025355 2.36217767150365
#> 5  2360.43323458045 1101.56891927972 57266.9645206182  2.1865322472999
#> 6  4819.61305796893 3234.17571737822 131290.916337014 2.71715820682141
#> 7  17056.0366666053  15869.158932479 363109.400668959 12.7495321119801
#> 8  5190.18437431067 3842.25842333391 118225.569315428 8.89415628695021
#> 9  7775.48093747066 6302.97158555718 222418.107436794  6.3907426775449
#> 10 20084.6017031325 18725.9465130564 480860.382104012 26.5863851322171
#> 11  6802.7004739893 5558.05772574648 168074.736862852 8.36042486283153
#> 12 9191.58941073957 7669.50650663473 266980.748138462 5.40833722563277
#> 13  20476.219871626 19372.5372832618 491869.030552138 18.8132018196067
#> 14 6956.68742056786 5658.37285303469 175727.471645151 16.8450956354979
#> 15 9006.63853140203 7433.22577588702 264693.493372192 6.00502371204517
#> 16 20709.1882366096 19470.1537897222 516389.402424642 27.6291778678133
#> 17 6923.13434888091 6021.59656559639 177017.435941247 12.5506799732082
#> 18 9321.79738339515  7617.6353946545 270385.144097188 6.04774255878661
#> 19 19791.7528298255 18613.6086203679 504173.724577811  18.990433197631
#> 20 7144.34186086953 5790.63460889243 180793.052336208 13.3802457658011
#> 21 9257.25341730089 7412.05605113868 268493.637201429 8.85169297292194
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.686412824243086 34.6938393511169 67.6731270636068 1915.13311146804
#> 2   2.61232593246013 22.0996975627923 75.9587266000701 746.615331375634
#> 3   1.64388425688616 4.79758849682065  109.61341591591 777.479079042819
#> 4  0.287540199831134 52.3543722561346 96.2582228416012 2696.30129864073
#> 5   2.46884584053533 35.6218950590698 121.166907692517 1352.49158983397
#> 6  0.451844765888268 10.9046708042443 87.1203918433939  1157.3388241889
#> 7  0.493123408804901 142.899992815484 176.150156047291 6635.59985886262
#> 8  0.554970990629416 39.8118074740798 126.238804616856 2025.26704241771
#> 9  0.395666502299448 18.4382447765825 72.3456971306257 1851.54130721193
#> 10 0.611362136291833 94.3540626229157 380.222780772685 10190.8384387346
#> 11 0.640365037649182  64.494949094281  220.48140181088 2533.66859513448
#> 12 0.317622100425022 9.31703174174003  88.507929799646 4955.20200144802
#> 13 0.324222423785099 83.5709280184523  236.33338817522 11466.1551823202
#> 14 0.497910693075275 83.9184593198462  197.20790822854 4542.82091117571
#> 15 0.364781674810635 16.1561374093712 71.4713619547049 4261.05824690402
#> 16 0.249521256982356 89.2798014244787 180.722142970574 6589.92578605468
#> 17 0.536225823939252 66.1585172449354 130.749337335797 2591.84787710516
#> 18 0.455725270200386 7.42490909885225 133.926266375443 4097.71005161048
#> 19 0.297817441747276 68.2750696222176 244.211094524917 7340.90411337356
#> 20 0.344877265996242 90.7763223248999 96.4963000600485 3124.37017733887
#> 21 0.741803824660792 9.06064380809204 205.141866664031 3205.63770422998
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 89.0765904964627 96.4746259790193    14.6118309759392
#> 2             TRUE 28.3480361165929 38.4737639591442    27.2251869611265
#> 3             TRUE  62.467822276276 67.3003831989325    17.3901013601166
#> 4             TRUE 238.374133164795 246.145697704042    14.5442798310455
#> 5             TRUE 61.2270426729668 68.4207337665835    19.2472075821807
#> 6             TRUE 200.445401959824 209.384852460266    11.6030684024145
#> 7             TRUE 483.711800295103 525.657760943518    6.08033201198024
#> 8             TRUE 187.749175895023 217.010950079089    12.2314489626932
#> 9             TRUE 276.606672241931 297.632215651054    3.95506643202272
#> 10            TRUE 629.569263536183 717.038470621177  -0.411172627823322
#> 11            TRUE 314.282757820982 341.788555619697    8.92706965616923
#> 12            TRUE 325.327319055653 343.120748527984    3.25392098053054
#> 13            TRUE 690.176335619329 752.071769605835   0.234268062428986
#> 14            TRUE 319.611735605386 375.032100246174    9.03197203603553
#> 15            TRUE 310.014791728956 329.771319741585    2.68822994080686
#> 16            TRUE 725.251618451611 816.151613636717 -0.0697005296918489
#> 17            TRUE 313.665947477557 354.957684589412    7.17358758981837
#> 18            TRUE 339.625509033624 359.522582052032    4.26171302131471
#> 19            TRUE 750.146218876923 812.624744097129   0.136320793212956
#> 20            TRUE 333.006952762036 377.027961331521    9.09930602580014
#> 21            TRUE 322.362282440519 351.484352321432    4.10295646343546
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.870129167699 3346.95683829502 3461.09956976019    77655.5043397274
#> 2   35.8197392789204 1435.13437132337 1507.84237630496    41283.7255288085
#> 3   22.7984805652721  2187.3380139419 2203.12208009644    64965.2908596218
#> 4   15.4902870884899  7491.1036921211 7663.34957684378    145130.067389091
#> 5   27.3697103975419 2301.83521720828 2419.03125195262    55042.1158553414
#> 6   13.0896376821869 4801.67487449595 4837.55124144191    129387.093971223
#> 7   7.70270802694836 16820.9661784238 17291.1071547867     352193.83890113
#> 8    14.057303521864 5124.69395101581 5255.67479760553     114894.00503065
#> 9    5.2568092245879 7745.15002481318 7805.81185012813     219372.32198643
#> 10  1.60020880057681 19929.3892701178 20239.8141361472    464096.452872294
#> 11   11.033870630035 6696.60628272921 6908.79466524939    163906.852023856
#> 12  4.29889769092886 9176.26289352441 9206.91592795473     258829.44084608
#> 13  1.30095983668196 20338.7456950356 20613.6940482163    473007.205277221
#> 14  10.6700982162532 6818.64155498671   7094.733286149    168254.531246267
#> 15  3.88836165093384 8980.06168536361 9033.21537744045    257684.052556035
#> 16 0.751224405780103 20562.3229632664 20856.0535099529    505548.974506582
#> 17  8.93777055057851 6814.30358801299 7031.96510974883    172753.846183409
#> 18  5.76104916027398 9309.58340792753 9334.01135886276    263644.411062289
#> 19  1.11614017656149 19679.4403402969  19904.065319354    492097.937311311
#> 20  10.2339522309278 6995.01481064507 7293.66891109399    175653.463394486
#> 21  6.54349104656947 9242.34865823658  9272.1581763652    263220.363177971
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up  ci95.lambda.bt.lo
#> 1     83956.2922764572 88.3682679502604 97.1829485252216   14.3956109363027
#> 2     43740.0899690343 27.3785515359231  39.443248539814   26.4023042924016
#> 3     67523.1970296727 62.0051302730429 67.7630752021655   16.8722778191975
#> 4     154000.898661619 237.630047198272 246.889783670566   14.4537046680987
#> 5     59491.8131858951 60.5382850150674  69.109491424483    18.469521142412
#> 6     133194.738702805 199.589497124675 210.240757295415   11.4607373011597
#> 7     374024.962436788 479.695697679829 529.673863558792    5.9249981382067
#> 8     121557.133600205 184.947516664634 219.812609309478    12.056633100645
#> 9     225463.892887158 274.593588298505 299.645299594481   3.83043148379839
#> 10     497624.31133573 621.194552219535 725.413181937826  -0.60375170075525
#> 11    172242.621701848  311.64922398919 344.422089451489   8.72535466930974
#> 12    275132.055430844 323.623692829578 344.824374754059   3.15387001889666
#> 13    510730.855827055 684.250177046153 757.997928179011  0.132137998936679
#> 14    183200.412044035 314.305530480204 380.338305371356   8.87513016771682
#> 15    271702.934188349 308.123209259662 331.662902210879   2.57332371324151
#> 16    527229.830342702  716.54842742325 824.854804665078 -0.148299725641291
#> 17    181281.025699085 309.712483285997 358.911148780973   7.00467645527751
#> 18    277125.877132088 337.720470127606  361.42762095805   4.11815956120159
#> 19     516249.51184431  744.16423241967 818.606730554383 0.0425082990625645
#> 20    185932.641277931 328.792175345808 381.242738747749   8.99066968701133
#> 21    273766.911224887 319.573999154049 354.272635607903   3.86928825866731
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   17.0863492073356 3336.02827889942  3472.0281291558    77052.2374096149
#> 2   36.6426219476453 1428.17296659109 1514.80378103724    41048.5416994252
#> 3   23.3163041061912  2185.8267735654 2204.63332047294    64720.3849497233
#> 4   15.5808622514367 7474.61206486042 7679.84120410447    144280.732480019
#> 5   28.1473968373105 2290.61432026467 2430.25214889622    54616.0810045437
#> 6   13.2319687834417 4798.23990319261 4840.98621274525    129022.532241604
#> 7   7.85804190072191 16775.9526806869 17336.1206525236    350103.624945588
#> 8   14.2321193839123 5112.15323166147 5268.21551695986    114256.045912289
#> 9   5.38144417281223 7739.34197770855 7811.61989723276    218789.086474659
#> 10  1.79278787350874 19899.6677403916 20269.5356658735    460886.338764092
#> 11  11.2355856168945 6676.29037376451 6929.11057421409    163108.746416388
#> 12  4.39894865256274 9173.32802852576 9209.85079295338    257268.552215624
#> 13  1.40308990017427 20312.4208527098 20640.0188905421    469395.366394791
#> 14  10.8269400845719 6792.20724030096 7121.16760083476    166823.542659247
#> 15  4.00326787849919 8974.97250207966  9038.3045607244     256341.81920826
#> 16 0.829823601729545 20534.1998258177 20884.1766474016    503473.147883974
#> 17  9.10668168511938 6793.46365508084 7052.80504268099    171937.414102121
#> 18   5.9046026203871 9307.24456156139  9336.3502052289    262353.632396032
#> 19  1.20995267071189 19657.9336933659  19925.571966285    489785.552515599
#> 20  10.3425885697166 6966.42026911273 7322.26345262633    174669.286788624
#> 21  6.77715925133761 9239.49455543703 9275.01227916475    262210.587301138
#>    ci95.integral.bt.up
#> 1     84559.5592065697
#> 2     43975.2737984177
#> 3     67768.1029395712
#> 4     154850.233570691
#> 5     59917.8480366928
#> 6     133559.300432424
#> 7      376115.17639233
#> 8     122195.092718566
#> 9      226047.12839893
#> 10    500834.425443932
#> 11    173040.727309315
#> 12      276692.9440613
#> 13    514342.694709486
#> 14    184631.400631056
#> 15    273045.167536124
#> 16    529305.656965309
#> 17    182097.457780373
#> 18    278416.655798345
#> 19    518561.896640023
#> 20    186916.817883793
#> 21     274776.68710172