Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file("lac_promoters_growth.txt", package = "QurvE"),
                   data.fl = system.file("lac_promoters_fluorescence.txt", package = "QurvE"),
                   csvsep = "\t",
                   csvsep.fl = "\t")
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rk-GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rk-GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rk-GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rk-GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rk-GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rk-GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r-GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rk-GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.9547301457728  15.4446239816409
#> 2                 FALSE          0.75 33.2100449823515  31.4383894113816
#> 3                 FALSE          0.75 64.6079974217452  19.8581176985423
#> 4                 FALSE          0.75 242.270153555516  14.8806599535177
#> 5                 FALSE          0.75 66.1576629625958  22.3826081372709
#> 6                 FALSE          0.75 205.330468114374  12.3718457098843
#> 7                 FALSE          0.75 507.710242779926  6.93176446939861
#> 8                 FALSE          0.75 203.153513334505  13.3342972011587
#> 9                 FALSE          0.75 291.599768023529  4.87725095654763
#> 10                FALSE          0.75 672.617299327552 0.538582072592386
#> 11                FALSE          0.75 326.160732529663  9.97626953597327
#> 12                FALSE          0.75 335.407245704557  3.74652611138123
#> 13                FALSE          0.75 724.799058071929 0.759739666376185
#> 14                FALSE          0.75 337.579791473194  9.72820646282341
#> 15                FALSE          0.75 320.555279726941  3.43897292346765
#> 16                FALSE          0.75 785.300540958897 0.724252772418189
#> 17                FALSE          0.75 335.762172428907  8.24450075051144
#> 18                FALSE          0.75 348.668232249144  4.82920845082114
#> 19                FALSE          0.75 775.499764286309 0.571259676924846
#> 20                FALSE          0.75 347.882571733967  9.45175174752618
#> 21                FALSE          0.75 337.509144536535  5.29655330393498
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3410.87573218531 2127.08370152489 81540.8361148032 2.42488417495137
#> 2  1477.71353993619 314.557488135576 42665.5052763078 1.99677770839287
#> 3    2196.583856573 698.985150782818 65607.0669263041 1.51671913848849
#> 4  7546.83636075398 6050.95293549896 148210.438096097 2.29346154020231
#> 5   2366.6303507943 1126.79835805512 57222.0903388738  2.4239876286759
#> 6  4821.27527020421 3237.95553259188 131348.124353308 2.17343417710303
#> 7  17031.3081475919 15855.9560393472 362525.560595604 15.7527495932239
#> 8  5185.17220122094 3787.83668344269 118523.872014882 8.09246489968296
#> 9  7767.89513956046 6253.65316411197 220284.159040283 6.86906975489785
#> 10 20028.2668080616  18722.136199752 483550.509357212 21.5514611509756
#> 11 6754.38819657901 5506.91700040219 167284.906160357 11.8264633144244
#> 12 9186.99976056693 7680.69486733045 266814.040603172 5.02845787048767
#> 13 20395.4876892797 19337.7267078197 487072.015675159 19.8598403759189
#> 14 6982.23560470082 5629.72643004192 177151.914009335 13.1373402630868
#> 15 9001.83423316654 7386.88172650795 264582.238395258 6.55208144051095
#> 16 20647.1456311729 19167.2885116512 509888.762211376  29.705086181049
#> 17 6892.83477900931  5969.5159648402 176157.285906804 12.0623042327625
#> 18 9320.52470744347 7676.95951168734 270751.591101538  7.7513153837361
#> 19 19806.5946843799 18701.4028398832 499579.820167386 21.0530888324321
#> 20 7146.40545418839 5820.48322432944 180837.711057833 13.8612450000368
#> 21 9256.30985696405 7429.29401838261 266877.122258328 6.92311876591421
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.419785534173295 29.0274867907298 46.1829177008419 1674.07913328735
#> 2   1.82032969695106 22.8722840279921 64.7095422092144 791.753195211497
#> 3   1.91869236303362 5.73187614077303 121.400313555171 1728.75418104653
#> 4   0.24486092178103 64.3796777652926 75.2231688396563 3217.50526925652
#> 5    2.6299472192271 39.9744975518874 140.884153388821 1536.13155256303
#> 6  0.408623649033518 10.6706637076763 84.8500648544016 1233.70568019593
#> 7  0.541080087397818 144.817222731747 157.812547929232  6865.0413829624
#> 8  0.613288568092561  38.654532241854 110.909429626438 2158.99111893423
#> 9  0.418913771830014 17.6055171736267 67.1662044561334 3636.45696829959
#> 10 0.287928960368203 97.8518940891949  219.12029643751 10920.5171773771
#> 11 0.571850743580134 101.900015768231 182.822880578782 3951.60177214249
#> 12 0.274502172775034 11.3950316255337 98.7229382778333 4224.97856780767
#> 13 0.255052074762561  98.067324886712 237.285394513815 11942.0943496554
#> 14 0.709259225110181 43.6482368835706  198.77192013317  2417.3189757907
#> 15 0.454896288391461 18.9592214249386 104.784030029805 3268.30987300805
#> 16 0.820267340798496 137.244727724794 582.733255802062 14849.2948403859
#> 17 0.528320359034685 92.0912684194286  189.87389585002 2766.10042964825
#> 18 0.425815576231116 8.51388148948459 46.1720084661038 4285.45110575703
#> 19 0.196216809053676 87.0038411046393 162.409093877755 9586.92059133592
#> 20  0.45919617596359 70.1146688610474 118.825188551214 2362.80244944049
#> 21 0.933527094926217 9.83325570679848 260.823121062165 5288.16632357403
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 87.9657956779778 95.9436646135678   14.7540767779259
#> 2             TRUE 29.9253456520452 36.4947443126578   28.4439470598971
#> 3             TRUE 62.1129944389317 67.1030004045588    16.701868761352
#> 4             TRUE 238.497409321883 246.042897789149   14.4778637371879
#> 5             TRUE 62.1702033134239 70.1451226117676   18.0563449616423
#> 6             TRUE 201.755168893039 208.905767335708   11.6996598072242
#> 7             TRUE 481.796969699073 533.623515860779    6.0416877256292
#> 8             TRUE 189.841408574527 216.465618094483   12.3254375066465
#> 9             TRUE 280.300148276722 302.899387770336   4.18813780188726
#> 10            TRUE 637.165145734197 708.069452920907 0.0649389327866924
#> 11            TRUE 306.706200377435 345.615264681891   9.03557506278395
#> 12            TRUE 327.135432507605  343.67905890151    3.2949700371663
#> 13            TRUE 692.129620653543 757.468495490316  0.340179003391772
#> 14            TRUE 315.968866740416 359.190716205972   8.56147503751716
#> 15            TRUE   309.7771057573 331.333453696581    2.6906685290637
#> 16            TRUE 736.435674191071 834.165407726722 -0.625087003195337
#> 17            TRUE 315.919681966013 355.604662891802   7.37541375989938
#> 18            TRUE 335.917318442898  361.41914605539   4.12874182792095
#> 19            TRUE 740.867433156958  810.13209541566  0.248483026031549
#> 20            TRUE 325.080823708907 370.684319759028   8.69637403806607
#> 21            TRUE 326.120614166606 348.897674906464   3.76090123278136
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1    16.135171185356 3363.12551641456 3458.62594795606    78786.9759405455
#> 2   34.4328317628661 1440.08863271015 1515.33844716224    41363.0712701849
#> 3   23.0143666357326 2187.15492032143 2206.01279282457    62763.2662984826
#> 4   15.2834561698475 7440.93179083007 7652.74093067789     142917.64192817
#> 5   26.7088713128995 2300.87230232145 2432.38839926716    54695.1539349076
#> 6   13.0440316125445 4803.72202840508 4838.82851200334    129318.678509385
#> 7   7.82184121316802 16793.0838161982 17269.5324789857    351232.567520631
#> 8    14.343156895671 5121.58549568309 5248.75890675879    114972.331624235
#> 9   5.56636411120801 7738.93406380984 7796.85621531107     214302.18732743
#> 10  1.01222521239808 19867.3004422849 20189.2331738383    465586.258600426
#> 11  10.9169640091626 6586.76267064027 6922.01372251775    160784.521245183
#> 12  4.19808218559616 9168.25493354292 9205.74458759093    259863.950859128
#> 13   1.1793003293606 20234.1669398411 20556.8084387184    467427.270469976
#> 14  10.8949378881297 6910.43425502734 7054.03695437429    173175.424294159
#> 15  4.18727731787161 8970.64631392251 9033.02215241056    259205.868654159
#> 16  2.07359254803172 20421.3780540656 20872.9132082801    485461.672198941
#> 17  9.11358774112349 6741.34464245935 7044.32491555927    171607.050700033
#> 18  5.52967507372133 9306.51937239327 9334.53004249367    263702.024032568
#> 19 0.894036327818143 19663.4733657627  19949.716002997    483809.335794638
#> 20  10.2071294569863 7031.06682391196 7261.74408446481    176950.901028504
#> 21  6.83220537508861 9240.13415132637 9272.48556260173    258178.088656048
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     84294.6962890608 87.2019571628681 96.7075031286774    14.6218443346613
#> 2     43967.9392824307 29.2963606739015 37.1237292908016    27.8705432053575
#> 3     68450.8675541257 61.6352279103078 67.5807669331826    16.0974806669964
#> 4     153503.234264024 237.774968936719 246.765338174312    14.4007325468268
#> 5       59749.02674284  61.406647210391 70.9086787148006    17.2279115875858
#> 6      133377.57019723 201.070537127252 209.590399101496    11.5709433577786
#> 7     373818.553670577 476.834853577207 538.585631982645    5.87124749809888
#> 8     122075.412405528 187.292282131126 219.014744537884    12.1322516076973
#> 9     226266.130753135 278.136391303929 305.063144743129    4.05617996376081
#> 10    501514.760113997  630.37643547164 714.858163183464 -0.0257586897292914
#> 11    173785.291075532 302.980864433391 349.340600625934    8.85544207855621
#> 12    273764.130347216 325.551468278401 345.263023130713    3.20850185274217
#> 13    506716.760880342 685.873770935128  763.72434520873   0.259837599841565
#> 14    181128.403724511 311.830604557544 363.328978388844    8.33805838160746
#> 15    269958.608136356 307.713200103539 333.397359350342    2.54737619822039
#> 16    534315.852223811 727.078572044041 843.522509873753  -0.883471215546863
#> 17    180707.521113576 312.120056132693 359.404288725122    7.20899284680345
#> 18    277801.158170509 333.475654097021 363.860810401267    3.99460992140815
#> 19    515350.304540133 734.235710174742 816.763818397876   0.186674731179641
#> 20    184724.521087163 320.714531533895  375.05061193404    8.55172724263754
#> 21    275576.155860607 323.939831755343 351.078457317727     3.4668401978796
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.2674036286206 3353.98185807548 3467.76960629514    78259.6410135599
#> 2   35.0062356174057 1432.88386324133 1522.54321663106    41113.6690136933
#> 3   23.6187547300882 2185.34937933709 2207.81833380892    62218.7087314529
#> 4   15.3605873602085 7420.65219233401 7673.02052917395    141904.127768354
#> 5    27.537304686956  2288.2803355926   2444.980365996    54211.2724958503
#> 6     13.17274806199 4800.36076933716 4842.18977107125    128930.061220124
#> 7   7.99228144069833 16747.4663910377 17315.1499041462    349070.079484998
#> 8   14.5363427946202 5109.40931802691 5260.93508441497    114292.249421771
#> 9   5.69832194933446 7733.38832590015 7802.40195322077    213156.703382415
#> 10  1.10292283491406 19836.4770956468 20220.0565204764    462146.295689553
#> 11  11.0970969933903 6554.66416567328 6954.11222748475    159539.766686958
#> 12   4.2845503700203 9164.66549858088 9209.33402255297    258533.082610269
#> 13  1.25964173291081 20203.2757325018 20587.6996460577    463665.510749834
#> 14  11.1183545440394 6896.68506040902 7067.78614899261    172413.968816785
#> 15  4.33056964871492 8964.67415917366 9038.99430715942    258176.351044162
#> 16  2.33197676038324 20378.1459648323 20916.1452975135    480784.144324219
#> 17  9.28000865421942 6712.33589290723 7073.33366511139    170735.729064694
#> 18  5.66380698023413 9303.83749972408 9337.21191516286    262352.106934254
#> 19 0.955844622670051 19636.0671558148  19977.122212945    480789.455808367
#> 20  10.3517762524148 7008.98070322073 7283.83020515604     176206.61825693
#> 21  7.12626640999037 9237.03667577873 9275.58303814937    256512.316264123
#>    ci95.integral.bt.up
#> 1     84822.0312160464
#> 2     44217.3415389224
#> 3     68995.4251211553
#> 4      154516.74842384
#> 5     60232.9081818974
#> 6     133766.187486492
#> 7     375981.041706211
#> 8     122755.494607993
#> 9      227411.61469815
#> 10    504954.723024871
#> 11    175030.045633756
#> 12    275094.998596075
#> 13    510478.520600483
#> 14    181889.859201885
#> 15    270988.125746353
#> 16    538993.380098532
#> 17    181578.842748915
#> 18    279151.075268822
#> 19    518370.184526404
#> 20    185468.803858737
#> 21    277241.928252533