Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228847         20.5  15.5467166096294
#> 2                 FALSE 30.7727385498294           20  31.2872267087641
#> 3                 FALSE 64.1757831646195           16  20.8131240327289
#> 4                 FALSE 242.557704041626         23.5   14.982628746023
#> 5                 FALSE 64.6529094536605           19  21.6712998092256
#> 6                 FALSE 205.579719130273           18  12.5662904914231
#> 7                 FALSE 509.394721871828         21.5  6.95952698278091
#> 8                 FALSE 204.031623957349         17.5  13.4360564669416
#> 9                 FALSE 290.961299939404           13  4.67379381987626
#> 10                FALSE 670.985334314906            6 0.373248436531125
#> 11                FALSE 327.194544214636           17  9.86802427129631
#> 12                FALSE  335.24358478568         17.5  3.65080407904796
#> 13                FALSE 728.625260083634          6.5 0.553816578577824
#> 14                FALSE 346.268113254507         16.5  9.75821807365783
#> 15                FALSE 320.246598452139         17.5  3.21350390586807
#> 16                FALSE  774.05398642078          5.5 0.246732777852533
#> 17                FALSE 337.909403917534         16.5  8.08916482172041
#> 18                FALSE 351.643501664612           18  4.84484452250892
#> 19                FALSE 779.134078966909            6 0.462268482214904
#> 20                FALSE 353.185039298395         16.5   9.5717850859864
#> 21                FALSE 338.826806405161           18  5.01597033523556
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589045
#> 11              <NA>          <NA>           <NA> 1187.83879461914
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950227
#> 17              <NA>          <NA>           <NA> 862.662992687727
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1   3427.1255242313  2135.7012372061  82881.406474157                TRUE
#> 2  1470.99207749982 282.713762021994 43222.9360971386                TRUE
#> 3  2196.18856426073 630.596386163366 67033.6678589485                TRUE
#> 4  7592.80063751346 6072.99699974995 151370.675832012                TRUE
#> 5  2384.30375521349 1128.58244716969 58249.0814015737                TRUE
#> 6  4820.28715109276 3196.13425225375 132414.081549265                TRUE
#> 7  17121.2144379564 15989.0291915634 367233.851934483                TRUE
#> 8    5206.543176868 3800.82664822097 120525.231431367                TRUE
#> 9  7768.18866992603 6299.66428652521  223919.42858498                TRUE
#> 10 20079.0095008516 18869.5355449612 490254.828661569                TRUE
#> 11 6840.37784263295 5652.53904801381 169914.301957325                TRUE
#> 12 9191.86346478591  7723.9236546828 269744.222628614                TRUE
#> 13  20481.097768094  19589.302363147 497122.375105375                TRUE
#> 14 7000.84532046262 5718.06831568824 179100.135589319                TRUE
#> 15 9006.21056404733 7463.56186630929 267724.576457373                TRUE
#> 16 20704.0909872879 19571.6505477857 520199.730690857                TRUE
#> 17 6952.68659445902 6090.02360177129 178120.140852737                TRUE
#> 18 9322.62835009364 7720.20364910999  273698.19440202                TRUE
#> 19  19774.702139564 18757.6585185487 508722.732093212                TRUE
#> 20 7189.59680355244 5865.90085424124 182547.707805716                TRUE
#> 21 9258.53106887529 7551.00923758219 271383.083788705                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 92.7359188268607  15.6685968289348
#> 2                 FALSE          0.75 32.0267254653404  31.7005402610264
#> 3                 FALSE          0.75 65.4694767799472  20.9719583319986
#> 4                 FALSE          0.75 242.263373760706  15.0456271592001
#> 5                 FALSE          0.75 64.7037557251294  22.7097173230871
#> 6                 FALSE          0.75 205.139469589183  12.3411132214952
#> 7                 FALSE          0.75 506.017033651349  7.03344541210291
#> 8                 FALSE          0.75 203.470957974187  13.2379753397931
#> 9                 FALSE          0.75 292.171152294959  4.85381026915179
#> 10                FALSE          0.75 672.053279960447 0.629489881328763
#> 11                FALSE          0.75 326.496654163389  9.76420738032473
#> 12                FALSE          0.75 334.243057056702  3.77811370299236
#> 13                FALSE          0.75 731.886155654271 0.994195921824282
#> 14                FALSE          0.75 345.054811468621  9.79402294778339
#> 15                FALSE          0.75 317.840868174456  3.28527472505629
#> 16                FALSE          0.75 766.264755740555 0.398186314436391
#> 17                FALSE          0.75 336.773464462501   8.2592697202227
#> 18                FALSE          0.75 351.213584221073  5.02655444473965
#> 19                FALSE          0.75 781.417215815415 0.611724957818543
#> 20                FALSE          0.75 352.436784199568  9.46941003690626
#> 21                FALSE          0.75 334.878955007481  5.03498559566819
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3418.43206738172 2123.22621986864 81703.2618755117 2.50856807662248
#> 2  1455.28572697177 262.471280068247  42838.838384032 1.54655966526759
#> 3  2192.03529414007 619.551355758232 66011.2466306588  1.0402577609487
#> 4  7565.06723194572 6030.33922446111 149136.369255657  2.4006345505462
#> 5  2361.75801972271 1098.25666605837 57414.4603745017 2.49256834547859
#> 6  4819.76804220503 3240.30641199425 130405.810217188 2.39257692771874
#> 7  16959.0462667255 15732.2234782854 358088.880807804  11.806078736447
#> 8  5178.76200873253 3792.90399332698 119376.340918897 9.56934883765721
#> 9  7756.72415260085 6261.92454181911 222791.134333084 5.65740957024177
#> 10 20057.7314646943 18673.8087088448 485305.489577055 21.5071031698718
#> 11 6765.17264914881 5585.49896246367 166667.868576767 12.8222767613844
#> 12 9193.24763748779 7670.31652359838 266896.771277373 5.15559366668165
#> 13 20433.2712575792 19219.2068661089 490395.609586825 20.8593654539132
#> 14 6971.50299987414 5648.53792110967 176363.291183881 12.4678002298246
#> 15 9007.20157075169  7409.3576969522 264447.743435529 6.61974935756379
#> 16  20675.650899115 19368.1422505412  513155.04188568 23.2952056599088
#> 17 6905.41091644748 5952.22160879324 176613.419254279 13.9476947813522
#> 18 9312.76712570899 7643.22192253092 271705.391860977 4.70888077672841
#> 19   19792.34869839 18665.9300931167 502851.173488692 25.5771642624616
#> 20 7115.59963898623  5800.5422104206 180039.551121245 16.0344192474398
#> 21 9257.56860043387 7494.84797710732 269058.806995181 7.87874037055349
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.626001170093277 20.0575814811436 61.6590360956449  1372.0399283887
#> 2   2.42551724635246 19.1851565536319  67.622339916385 1077.83151701695
#> 3  0.889032015022148 4.34935978711403 67.8895979554011 1235.79362286618
#> 4  0.325009365622393 59.8337080025088 81.1518081823901 3183.94401215714
#> 5   2.11006298905483 22.9933778162551 117.473301252896 847.977041968463
#> 6  0.627589592071861 10.9895484131482  129.27955526132 4518.09160883873
#> 7  0.409727234532852 237.916137492439 245.026321356808 13890.0514372424
#> 8  0.774051956145063 62.8843433582251 155.510789340482 3310.63488913461
#> 9  0.276916832518138 15.5749862915993 52.7328725851607 1630.90101406674
#> 10 0.338641251761028 64.9140531427361 211.314600537624 6799.19635881099
#> 11 0.832676145200789 111.823263322479 219.648676413029 4281.32327905937
#> 12 0.285560523867899 11.7912317269459 127.038556190947 2909.34616078363
#> 13 0.573063943423931 108.082988833477 382.034064521157 11624.5210979792
#> 14 0.569824662433662 82.3602254590914 196.914897210153 4507.41634752198
#> 15 0.433327150020346 12.9166020030286 89.2592372191904 4359.92444286444
#> 16 0.262485732488547 89.9020304002473 225.715990453245 10649.7490854089
#> 17 0.668867312164465 71.7363251680176 169.134582982696 2914.82075406441
#> 18 0.389773526847704 10.1238428653119  107.21691242248 2813.59841197616
#> 19 0.257474194862876 71.5686594473473 191.056504278745  8809.4557201562
#> 20 0.423381724210959 87.0477536112908 129.872109127806 2624.00116666477
#> 21 0.617745216277266 7.69496237447335  181.06101570867 3103.96066819174
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up   ci90.lambda.bt.lo
#> 1             TRUE 88.6093243408167 96.8625133129047    14.6388249041314
#> 2             TRUE 29.4826348159752 34.5708161147056    27.7105643907766
#> 3             TRUE 63.7582527631866 67.1807007967079    19.5095006672871
#> 4             TRUE 238.314329925057 246.212417596354    14.5109867527513
#> 5             TRUE 60.6034807968171 68.8040306534417    19.2386637060919
#> 6             TRUE 201.203680543085  209.07525863528     11.308728342537
#> 7             TRUE 486.596034129894 525.438033172805    6.35944411129637
#> 8             TRUE 187.729379136241 219.212536812133    11.9646598719345
#> 9             TRUE 282.864713551911 301.477591038007    4.39828207965945
#> 10            TRUE 636.674095246007 707.432464674886  0.0724250221818721
#> 11            TRUE 305.404008890912 347.589299435867    8.39445512146943
#> 12            TRUE  325.76210547501 342.724008638393    3.30836664122966
#> 13            TRUE 697.572499482584 766.199811825958  0.0515057348919152
#> 14            TRUE  324.54528009056 365.564342846683    8.85666137808002
#> 15            TRUE 306.951380481264 328.730355867649    2.57245156327282
#> 16            TRUE 727.944142430004 804.585369051105 -0.0336027155072691
#> 17            TRUE 313.829506547177 359.717422377826    7.15898299171216
#> 18            TRUE 343.467475343355 358.959693098792    4.38537699307518
#> 19            TRUE 739.342780603666 823.491651027164   0.188179907269112
#> 20            TRUE 326.060164537529 378.813403861606    8.77294710057924
#> 21            TRUE 321.918427097921 347.839482917042    4.01879471489209
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.6983687537383 3385.43734584524  3451.4267889182    79446.2561933123
#> 2   35.6905161312762 1423.72614444104 1486.84530950249    41065.8055385391
#> 3     22.43441599671 2184.88059729027 2199.18999098987    63978.3661210439
#> 4   15.5802675656489 7466.64078228159 7663.49368160984    143898.781355659
#> 5   26.1807709400823 2323.93391321497 2399.58212623045    56019.5381404636
#> 6   13.3734981004534 4801.69023506541 4837.84584934466    122973.549520648
#> 7   7.70744671290946 16567.6742205504 17350.4183129006     335239.74619354
#> 8   14.5112908076517 5075.31726390825 5282.20675355681    113930.346526271
#> 9   5.30933845864413 7731.10330015117 7782.34500505053    220108.302164945
#> 10  1.18655474047565 19950.9478472745 20164.5150821141    474120.811566811
#> 11    11.13395963918 6581.22338098333 6949.12191731429    159625.091782714
#> 12  4.24786076475505 9173.85106129696 9212.64421367862    262110.896842884
#> 13  1.93688610875665 20255.4747409482 20611.0677742103    471273.272380649
#> 14  10.7313845174868 6836.02042899394 7106.98557075435    168948.591292207
#> 15  3.99809788683976  8985.9537604567 9028.44938104667    257275.667727017
#> 16 0.829975344380051 20527.7620591066 20823.5397391234    495636.204640182
#> 17  9.35955644873325 6787.40466154609 7023.41717134887    171818.539113843
#> 18  5.66773189640412 9296.11340419555 9329.42084722243    267077.022473276
#> 19  1.03527000836797 19674.6182535991 19910.0791431809    488359.618829035
#> 20  10.1658729732333 6972.40608429566 7258.79319367681    175723.069202082
#> 21   6.0511764764443 9244.91038732786 9270.22681353988    263952.791696005
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83960.2675577111 87.8191253966806 97.6527122570407     14.441634535552
#> 2     44611.8712295249 28.9954685214159 35.0579824092649    26.9465264581756
#> 3     68044.1271402737 63.4305715684878 67.5083819914067    19.2294555825552
#> 4     154373.957155656 237.558130041635 246.968617479776    14.4086088025802
#> 5     58809.3826085399 59.8183217679914 69.5891896822674    18.5739938645396
#> 6     137838.070913728 200.450018810854 209.828920367512    11.1110376210344
#> 7     380938.015422068 482.877119327913 529.156947974786    6.23038003241852
#> 8     124822.335311523 184.715034252379 222.226881695995    11.7208335057488
#> 9     225473.966501224 281.082629537285 303.259675052633    4.31105327741624
#> 10    496490.167587299 629.899357747498 714.207202173395 -0.0342469721228517
#> 11     173710.64537082 301.364991711076 351.628316615703    8.13216213573119
#> 12    271682.645711862 324.138093470006 344.348020643398    3.21841507621127
#> 13    509517.946793001 691.001799364601 772.770511943941  -0.129009407286623
#> 14    183777.991075554 320.617923018165 369.491699919078    8.67716660941341
#> 15    271619.819144041 304.866159433631 330.815576915281    2.43595351101641
#> 16    530673.879131178 720.606152647133 811.923358833976  -0.116285721241161
#> 17    181408.299394715 309.435982691051 364.110946233952    6.94828978838035
#> 18    276333.761248678 341.984177898686 360.442990543461    4.26259833211815
#> 19    517342.728148349  731.28597386099  831.54845776984   0.107075535887307
#> 20    184356.033040409 321.009322474586  383.86424592455    8.63958185745279
#> 21    274164.822294356 319.436623881197 350.321286133766    3.82420497176475
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.8955591223177 3379.11920767868 3457.74492708476    79014.0636158699
#> 2   36.4545540638772 1417.68282012665 1492.88863381689    40726.2886106788
#> 3    22.714461081442 2183.51054895733 2200.56003932281    63589.0911298411
#> 4     15.68264551582  7447.7931642608 7682.34129963063    142895.838991829
#> 5   26.8454407816346 2316.69099920285 2406.82504024257    55752.4253722436
#> 6   13.5711888219561 4798.22852731526  4841.3075570948    121550.350663864
#> 7    7.8365107917873 16492.7306372403 17425.3618962107    330864.379990809
#> 8   14.7551171738374 5055.50869575041 5302.01532171466    112887.496536193
#> 9   5.39656726088734 7726.19717946931 7787.25112573238    219594.568345514
#> 10  1.29322673478038 19930.4999205345 20184.9630088541    471979.064713785
#> 11  11.3962526249183 6545.99905303675 6984.34624526087    158276.474949811
#> 12  4.33781232977344 9170.13682330298  9216.3584516726    261194.452802237
#> 13  2.11740125093519 20221.4285994656 20645.1139156929    467611.548234786
#> 14  10.9108792861534 6810.07695797432 7132.92904177396    167528.755142738
#> 15  4.13459593909617 8981.88503082575 9032.51811067762    255902.291527515
#> 16 0.912658350113943 20499.4429195305 20851.8588786995    492281.533678278
#> 17  9.57024965206505 6764.80771911816 7046.01411377679    170900.370576313
#> 18  5.79051055736115 9292.92439369298   9332.609857725    266190.738973504
#> 19  1.11637437974978 19652.0741258732 19932.6232709068    485584.640277186
#> 20  10.2992382163597  6944.9860419081 7286.21323606436    174896.508834582
#> 21  6.24576621957163  9242.4864741799 9272.65072668784    262975.044085525
#>    ci95.integral.bt.up
#> 1     84392.4601351536
#> 2     44951.3881573852
#> 3     68433.4021314765
#> 4     155376.899519485
#> 5     59076.4953767599
#> 6     139261.269770512
#> 7     385313.381624799
#> 8     125865.185301601
#> 9     225987.700320655
#> 10    498631.914440324
#> 11    175059.262203723
#> 12    272599.089752509
#> 13    513179.670938864
#> 14    185197.827225024
#> 15    272993.195343544
#> 16    534028.550093082
#> 17    182326.467932246
#> 18     277220.04474845
#> 19    520117.706700198
#> 20    185182.593407908
#> 21    275142.569904837