Generic summary function for flFit objects

Generic summary function for flFit objects

Usage

# S3 method for flFit
summary(object, ...)

Arguments

object

object of class flFit

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from all fits of a workflow.

Examples

# load example dataset
input <- read_data(data.growth = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   data.fl = system.file('lac_promoters.xlsx', package = 'QurvE'),
                   sheet.growth = 1,
                   sheet.fl = 2 )
#> Sample data are stored in columns. If they are stored in row format, please run read_data() with data.format = 'row'.

# Run curve fitting workflow
res <- flFit(fl_data = input$norm.fluorescence,
             time = input$time,
             parallelize = FALSE,
             control = fl.control(fit.opt = 's', suppress.messages = TRUE,
             x_type = 'time', norm_fl = TRUE, nboot.fl = 20))

summary(res)
#>            TestId AddId concentration reliability_tag log.x.spline log.y.spline
#> 1    pSEVA634∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 2   pSEVA634r∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 3  pSEVA634rκ∙GFP  <NA>             0            TRUE        FALSE        FALSE
#> 4    pSEVA634∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 5   pSEVA634r∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 6  pSEVA634rκ∙GFP  <NA>          0.01            TRUE        FALSE        FALSE
#> 7    pSEVA634∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 8   pSEVA634r∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 9  pSEVA634rκ∙GFP  <NA>          0.05            TRUE        FALSE        FALSE
#> 10   pSEVA634∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 11  pSEVA634r∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 12 pSEVA634rκ∙GFP  <NA>           0.1            TRUE        FALSE        FALSE
#> 13   pSEVA634∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 14  pSEVA634r∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 15 pSEVA634rκ∙GFP  <NA>           0.2            TRUE        FALSE        FALSE
#> 16   pSEVA634∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 17  pSEVA634r∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 18 pSEVA634rκ∙GFP  <NA>           0.5            TRUE        FALSE        FALSE
#> 19   pSEVA634∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 20  pSEVA634r∙GFP  <NA>             1            TRUE        FALSE        FALSE
#> 21 pSEVA634rκ∙GFP  <NA>             1            TRUE        FALSE        FALSE
#>    log.x.lin log.y.lin nboot.fl max_slope.linfit lambda.linfit dY.linfit
#> 1      FALSE     FALSE       20                0          <NA>      <NA>
#> 2      FALSE     FALSE       20                0          <NA>      <NA>
#> 3      FALSE     FALSE       20                0          <NA>      <NA>
#> 4      FALSE     FALSE       20                0          <NA>      <NA>
#> 5      FALSE     FALSE       20                0          <NA>      <NA>
#> 6      FALSE     FALSE       20                0          <NA>      <NA>
#> 7      FALSE     FALSE       20                0          <NA>      <NA>
#> 8      FALSE     FALSE       20                0          <NA>      <NA>
#> 9      FALSE     FALSE       20                0          <NA>      <NA>
#> 10     FALSE     FALSE       20                0          <NA>      <NA>
#> 11     FALSE     FALSE       20                0          <NA>      <NA>
#> 12     FALSE     FALSE       20                0          <NA>      <NA>
#> 13     FALSE     FALSE       20                0          <NA>      <NA>
#> 14     FALSE     FALSE       20                0          <NA>      <NA>
#> 15     FALSE     FALSE       20                0          <NA>      <NA>
#> 16     FALSE     FALSE       20                0          <NA>      <NA>
#> 17     FALSE     FALSE       20                0          <NA>      <NA>
#> 18     FALSE     FALSE       20                0          <NA>      <NA>
#> 19     FALSE     FALSE       20                0          <NA>      <NA>
#> 20     FALSE     FALSE       20                0          <NA>      <NA>
#> 21     FALSE     FALSE       20                0          <NA>      <NA>
#>    A.linfit x.mu.start.linfit x.mu.end.linfit r2mu.linfit reliable_fit.linfit
#> 1      <NA>              <NA>            <NA>        <NA>               FALSE
#> 2      <NA>              <NA>            <NA>        <NA>               FALSE
#> 3      <NA>              <NA>            <NA>        <NA>               FALSE
#> 4      <NA>              <NA>            <NA>        <NA>               FALSE
#> 5      <NA>              <NA>            <NA>        <NA>               FALSE
#> 6      <NA>              <NA>            <NA>        <NA>               FALSE
#> 7      <NA>              <NA>            <NA>        <NA>               FALSE
#> 8      <NA>              <NA>            <NA>        <NA>               FALSE
#> 9      <NA>              <NA>            <NA>        <NA>               FALSE
#> 10     <NA>              <NA>            <NA>        <NA>               FALSE
#> 11     <NA>              <NA>            <NA>        <NA>               FALSE
#> 12     <NA>              <NA>            <NA>        <NA>               FALSE
#> 13     <NA>              <NA>            <NA>        <NA>               FALSE
#> 14     <NA>              <NA>            <NA>        <NA>               FALSE
#> 15     <NA>              <NA>            <NA>        <NA>               FALSE
#> 16     <NA>              <NA>            <NA>        <NA>               FALSE
#> 17     <NA>              <NA>            <NA>        <NA>               FALSE
#> 18     <NA>              <NA>            <NA>        <NA>               FALSE
#> 19     <NA>              <NA>            <NA>        <NA>               FALSE
#> 20     <NA>              <NA>            <NA>        <NA>               FALSE
#> 21     <NA>              <NA>            <NA>        <NA>               FALSE
#>    max_slope2.linfit x.mu2.start.linfit x.mu2.end.linfit r2mu2.linfit
#> 1               <NA>               <NA>             <NA>         <NA>
#> 2               <NA>               <NA>             <NA>         <NA>
#> 3               <NA>               <NA>             <NA>         <NA>
#> 4               <NA>               <NA>             <NA>         <NA>
#> 5               <NA>               <NA>             <NA>         <NA>
#> 6               <NA>               <NA>             <NA>         <NA>
#> 7               <NA>               <NA>             <NA>         <NA>
#> 8               <NA>               <NA>             <NA>         <NA>
#> 9               <NA>               <NA>             <NA>         <NA>
#> 10              <NA>               <NA>             <NA>         <NA>
#> 11              <NA>               <NA>             <NA>         <NA>
#> 12              <NA>               <NA>             <NA>         <NA>
#> 13              <NA>               <NA>             <NA>         <NA>
#> 14              <NA>               <NA>             <NA>         <NA>
#> 15              <NA>               <NA>             <NA>         <NA>
#> 16              <NA>               <NA>             <NA>         <NA>
#> 17              <NA>               <NA>             <NA>         <NA>
#> 18              <NA>               <NA>             <NA>         <NA>
#> 19              <NA>               <NA>             <NA>         <NA>
#> 20              <NA>               <NA>             <NA>         <NA>
#> 21              <NA>               <NA>             <NA>         <NA>
#>    reliable_fit2.linfit max_slope.spline x.max.spline     lambda.spline
#> 1                 FALSE 92.2668917228958         20.5  15.5467166096303
#> 2                 FALSE 30.7727385498335           20  31.2872267087632
#> 3                 FALSE 64.1757831646139           16  20.8131240327297
#> 4                 FALSE 242.557704041646         23.5  14.9826287460236
#> 5                 FALSE 64.6529094536623           19  21.6712998092259
#> 6                 FALSE  205.57971913027           18  12.5662904914234
#> 7                 FALSE 509.394721871896         21.5  6.95952698278297
#> 8                 FALSE 204.031623957343         17.5  13.4360564669418
#> 9                 FALSE 290.961299939386           13  4.67379381987587
#> 10                FALSE 670.985334314907            6 0.373248436531114
#> 11                FALSE 327.194544214622           17  9.86802427129628
#> 12                FALSE 335.243584785665         17.5  3.65080407904781
#> 13                FALSE 728.625260083635          6.5  0.55381657857781
#> 14                FALSE  346.26811325449         16.5  9.75821807365777
#> 15                FALSE 320.246598452125         17.5  3.21350390586794
#> 16                FALSE 774.053986420781          5.5 0.246732777852532
#> 17                FALSE 337.909403917518         16.5  8.08916482172027
#> 18                FALSE 351.643501664605           18   4.8448445225091
#> 19                FALSE 779.134078966908            6 0.462268482214881
#> 20                FALSE 353.185039298378         16.5  9.57178508598632
#> 21                FALSE 338.826806405153           18  5.01597033523575
#>    max_slope2.spline x.max2.spline lambda2.spline        y0.spline
#> 1               <NA>          <NA>           <NA>  1291.4242870252
#> 2               <NA>          <NA>           <NA> 1188.27831547782
#> 3               <NA>          <NA>           <NA> 1565.59217809736
#> 4               <NA>          <NA>           <NA> 1519.80363776351
#> 5               <NA>          <NA>           <NA>  1255.7213080438
#> 6               <NA>          <NA>           <NA> 1624.15289883901
#> 7               <NA>          <NA>           <NA> 1132.18524639308
#> 8               <NA>          <NA>           <NA> 1405.71652864703
#> 9               <NA>          <NA>           <NA> 1468.52438340082
#> 10              <NA>          <NA>           <NA> 1209.47395589044
#> 11              <NA>          <NA>           <NA> 1187.83879461913
#> 12              <NA>          <NA>           <NA> 1467.93981010311
#> 13              <NA>          <NA>           <NA> 891.795404947046
#> 14              <NA>          <NA>           <NA> 1282.77700477438
#> 15              <NA>          <NA>           <NA> 1542.64869773804
#> 16              <NA>          <NA>           <NA> 1132.44043950228
#> 17              <NA>          <NA>           <NA> 862.662992687728
#> 18              <NA>          <NA>           <NA> 1602.42470098365
#> 19              <NA>          <NA>           <NA> 1017.04362101536
#> 20              <NA>          <NA>           <NA>  1323.6959493112
#> 21              <NA>          <NA>           <NA> 1707.52183129309
#>            A.spline        dY.spline  integral.spline reliable_fit.spline
#> 1  3427.12552423136 2135.70123720616 82881.4064741573                TRUE
#> 2  1470.99207749984 282.713762022016 43222.9360971387                TRUE
#> 3  2196.18856426073 630.596386163375 67033.6678589487                TRUE
#> 4   7592.8006375136 6072.99699975009 151370.675832013                TRUE
#> 5  2384.30375521353 1128.58244716973 58249.0814015739                TRUE
#> 6  4820.28715109277 3196.13425225376 132414.081549266                TRUE
#> 7  17121.2144379567 15989.0291915636 367233.851934484                TRUE
#> 8  5206.54317686809 3800.82664822106 120525.231431367                TRUE
#> 9   7768.1886699261 6299.66428652528 223919.428584981                TRUE
#> 10  20079.009500852 18869.5355449615 490254.828661571                TRUE
#> 11 6840.37784263307 5652.53904801394 169914.301957325                TRUE
#> 12 9191.86346478597 7723.92365468286 269744.222628615                TRUE
#> 13 20481.0977680944 19589.3023631473 497122.375105376                TRUE
#> 14 7000.84532046274 5718.06831568836 179100.135589319                TRUE
#> 15 9006.21056404742 7463.56186630938 267724.576457374                TRUE
#> 16 20704.0909872883  19571.650547786 520199.730690859                TRUE
#> 17 6952.68659445913  6090.0236017714 178120.140852738                TRUE
#> 18 9322.62835009371 7720.20364911006 273698.194402021                TRUE
#> 19 19774.7021395644  18757.658518549 508722.732093214                TRUE
#> 20 7189.59680355257 5865.90085424137 182547.707805717                TRUE
#> 21 9258.53106887535 7551.00923758226 271383.083788706                TRUE
#>    reliable_fit2.spline smooth.spline     max_slope.bt         lambda.bt
#> 1                 FALSE          0.75 91.5521842310919  15.2933292012148
#> 2                 FALSE          0.75 33.1129539390893  32.2982059983342
#> 3                 FALSE          0.75 64.7822292545393  20.6128350696835
#> 4                 FALSE          0.75 242.930944759427  15.0052156318778
#> 5                 FALSE          0.75 65.7433957891352  21.3188627997327
#> 6                 FALSE          0.75 205.223830727095  12.3434538961517
#> 7                 FALSE          0.75 504.752299019155  6.84388534243442
#> 8                 FALSE          0.75 203.055354530148  13.3499817664586
#> 9                 FALSE          0.75 289.709334976544   4.6534539630589
#> 10                FALSE          0.75 673.927469151644 0.803611718265811
#> 11                FALSE          0.75 326.094192282641  10.0589164397001
#> 12                FALSE          0.75 334.179703749999  3.78471846859065
#> 13                FALSE          0.75 725.642888582845 0.688482206537653
#> 14                FALSE          0.75 345.367411068064  10.0323896748843
#> 15                FALSE          0.75 318.705631796476   3.2925048040412
#> 16                FALSE          0.75 782.371991366201  0.41550270277776
#> 17                FALSE          0.75 330.979981262311  7.85354347595312
#> 18                FALSE          0.75 349.750127470517   4.9527989021906
#> 19                FALSE          0.75 779.819680065401 0.587692650348274
#> 20                FALSE          0.75 348.553582669521  9.43331495664432
#> 21                FALSE          0.75   339.6017920318  5.27989508036553
#>                A.bt            dY.bt      integral.bt  stdmax_slope.bt
#> 1  3417.91344248597  2142.9463098381 82086.9648302687 2.92069515003565
#> 2  1454.95973882802 233.570280154506 42344.8113348192 2.50011330256127
#> 3   2196.4435756092 649.764300924365 66294.6982997477 1.89446947200264
#> 4  7560.63542340536  6042.6134353849 148831.278582988 2.37737032885368
#> 5  2361.49042271293 1114.81078077038 57605.2841362148 3.10568627721772
#> 6  4819.00814792715 3237.95205242886  130872.01888982 2.48294398221698
#> 7  17008.2220472163 15861.6417163348 361059.535737513 12.1302328519299
#> 8  5181.83935860237 3781.55065280796  119071.10520396 9.79011593002924
#> 9  7765.73178896581 6286.19458490883 220769.135183024 7.86816519105217
#> 10  20077.499035107 18582.7120982276 484021.861907273 27.0281344790552
#> 11 6774.08557386629  5513.6137446459 167197.441807813 13.7348679264078
#> 12 9189.50966397542 7662.74202787226 265965.249365156 5.87349534008553
#> 13  20426.756941726 19417.6061027141 490279.073614687  17.333868820473
#> 14 6967.36587877186 5599.09990437462 177066.718890242 10.0715022942321
#> 15 9004.91429985397 7408.76949715418 266023.005909227  5.9846848406333
#> 16 20687.6564392255 19467.4498084732 514206.420577646 18.8236039483865
#> 17 6899.97764691862 6062.12660238947 175752.226101618 12.3840999012592
#> 18 9322.77294452902 7661.02921717363 271970.568771484 5.75875857191861
#> 19 19811.9156712124  18648.743928519 503401.225449604 23.6633954622078
#> 20 7156.26731666602 5855.04352049008 180622.218528133 9.51619398197009
#> 21 9257.45211700813 7465.44561119073 266397.999360663 10.0494929934631
#>         stdlambda.bt          stdA.bt         stddY.bt   stdintegral.bt
#> 1  0.707880616934917 21.2406981156916 71.1513278811156 1111.63276451473
#> 2   2.45192753531554 25.6014798927943 71.0367445648425 1056.02712761348
#> 3   1.87904979284727 5.43996665903485  130.28042989631 1084.83631075443
#> 4   0.35859234583303 71.3353115686751 115.477932179349 3731.18572953689
#> 5   5.20350193133708 37.9065964611186 139.744325314887 1502.67302527444
#> 6  0.420151977649233 9.92850348360669 87.5609110427022 2185.68295491808
#> 7  0.453893484100409 155.288840650196 190.391618345808 8668.08972969445
#> 8  0.659724831529737 37.9943655262486 106.391038369052 2545.89563656035
#> 9   0.45148996597513 16.6937518704668 78.5764982371677 4310.23352461345
#> 10 0.700596780685455 72.5946350706152 428.059036476288  8372.0584449638
#> 11 0.650332377365411 88.2272074823258 214.419233510097 3486.14003580075
#> 12 0.285398194140862 10.2809978744495 90.8553957695655 5310.65173711697
#> 13 0.273383304576085 119.691655525468 220.664201099221 13399.1396005786
#> 14 0.676740351386038 38.0596024640446 212.912464280311 2794.58575102352
#> 15 0.484981975707478 14.3259325499851 120.850784308825 3370.13321796165
#> 16 0.230853461689545 93.9418962174849 206.455466910484 7993.53408788605
#> 17 0.576884183462517 78.9454164439571 156.628560761467 2791.98397161068
#> 18 0.416241736650277 14.7551294189657 91.4252279039984 2504.12794590854
#> 19 0.251513414521535  76.508512114239 178.688474561371 8689.60630880782
#> 20 0.331720339423143 60.5574769154134 117.070343267349 2710.47875053289
#> 21 0.671823621773366 8.38012126058514 156.245837346092 5332.61665441915
#>    reliable_fit.bt    ci90.mu.bt.lo    ci90.mu.bt.up  ci90.lambda.bt.lo
#> 1             TRUE 86.7476407092833 96.3567277529006   14.1288655863569
#> 2             TRUE  29.000267556376 37.2256403218026   28.2647852027401
#> 3             TRUE  61.665826973095 67.8986315359836   17.5217981604497
#> 4             TRUE 239.020170568463 246.841718950392   14.4153312229824
#> 5             TRUE  60.634541863112 70.8522497151583   12.7591021226832
#> 6             TRUE 201.139387876348 209.308273577842   11.6523038929187
#> 7             TRUE  484.79806597773 524.706532060579   6.09723056108924
#> 8             TRUE  186.95061382525 219.160095235046   12.2647344185921
#> 9             TRUE 276.766203237263 302.652466715825   3.91075296902981
#> 10            TRUE 629.466187933598  718.38875036969 -0.348869985961762
#> 11            TRUE   303.5003345437 348.688050021582   8.98911967893398
#> 12            TRUE 324.517803915558  343.84160358444   3.31523843922894
#> 13            TRUE 697.128674373167 754.157102792523  0.238766670509994
#> 14            TRUE 328.799789794052 361.935032342076   8.91915179685425
#> 15            TRUE 308.860825233634 328.550438359318    2.4947094540024
#> 16            TRUE 751.407162871105 813.336819861297 0.0357487582984583
#> 17            TRUE 310.608136924739 351.351825599882   6.90456899415728
#> 18            TRUE 340.276969619711 359.223285321323   4.26808124540089
#> 19            TRUE 740.893394530069 818.745965600733  0.173953083460349
#> 20            TRUE  332.89944356918 364.207721769862   8.88763499829325
#> 21            TRUE 323.070376057553 356.133208006047   4.17474522254834
#>    ci90.lambda.bt.up     ci90.A.bt.lo     ci90.A.bt.up ci90.integral.bt.lo
#> 1   16.4577928160728 3382.97249408566 3452.85439088628    80258.3289326419
#> 2   36.3316267939283 1412.84530440437 1497.07417325166     40607.646709895
#> 3   23.7038719789172 2187.49483045509 2205.39232076331    64510.1425685567
#> 4   15.5951000407731 7443.28883587489 7677.98201093583      142693.4780579
#> 5   29.8786234767822 2299.13407153439 2423.84677389147    55133.3870096383
#> 6   13.0346038993847 4802.67575969662 4835.34053615768     127276.57042898
#> 7   7.59054012377959 16752.7719043467 17263.6721900859    346800.528132165
#> 8    14.435229114325 5119.33862731169 5244.34008989305    114883.106881819
#> 9   5.39615495708799  7738.2705671389 7793.19301079273    213678.801035034
#> 10  1.95609342249338 19958.0808604158 20196.9172097981    470249.825765308
#> 11  11.1287132004662 6628.95181755787 6919.21933017472    161462.741448921
#> 12  4.25419849795237 9172.59742247195 9206.42190547889    257229.227257599
#> 13  1.13819774256531 20229.8641683867 20623.6497150654    468237.488971735
#> 14  11.1456275529143  6904.7578327185 7029.97392482521    172469.625329808
#> 15     4.09030015408 8981.34814080924 9028.48045889869     260479.13676568
#> 16 0.795256647257061 20533.1220199477 20842.1908585032    501057.057003073
#> 17  8.80251795774896 6770.11243686831 7029.84285696893    171159.412468318
#> 18   5.6375165589803 9298.50075663482 9347.04513242322    267851.278300465
#> 19   1.0014322172362 19686.0591687845 19937.7721736403    489106.823071615
#> 20  9.97899491499539 7056.65026714016 7255.88436619187    176163.480983507
#> 21  6.38504493818271 9243.66681753447  9271.2374164818    257625.844964144
#>    ci90.integral.bt.up    ci95.mu.bt.lo    ci95.mu.bt.up   ci95.lambda.bt.lo
#> 1     83915.6007278954  85.827621737022 97.2767467251618    13.9058831920224
#> 2     44081.9759597434 28.2127318660692 38.0131760121094    27.4924280291157
#> 3     68079.2540309387 61.0690690894141 68.4953894196645    16.9298974757028
#> 4     154969.079108076 238.271298914874  247.59059060398     14.302374634045
#> 5     60077.1812627913 59.6562506857885 71.8305408924819     11.119999014312
#> 6     134467.467350661  200.35726052195  210.09040093224    11.5199560199592
#> 7      375318.54334286 480.977042629372 528.527555408937    5.95425411359761
#> 8     123259.103526102 183.866727307291 222.243981753005    12.0569210966603
#> 9     227859.469331013 274.287731202082 305.130938751006    3.76853362974765
#> 10    497793.898049238 620.952325572696 726.902612730592   -0.56955797187768
#> 11    172932.142166705 299.173851146881   353.0145334184    8.78426498006387
#> 12    274701.271472714 322.667652883431 345.691754616567    3.22533800807456
#> 13    512320.658257639 691.668505694718 759.617271470972   0.152650929568528
#> 14    181663.812450675 325.627266571369 365.107555564759    8.70597858616765
#> 15    271566.875052774 306.975649508835 330.435614084117    2.34194013165454
#> 16    527355.784152218 745.477727627364 819.266255105039 -0.0369700821337483
#> 17    180345.039734918 306.707145455843 355.252817068779    6.72285047636659
#> 18    276089.859242504 338.462960669557 361.037294271478    4.13696509835606
#> 19    517695.627827593 733.439424959474 826.199935171329  0.0947263578860659
#> 20     185080.95607276 329.901842464859 367.205322874182    8.78314309137496
#> 21    275170.153757183 319.904785764612 359.298798298988    3.96312078168973
#>    ci95.lambda.bt.up     ci95.A.bt.lo     ci95.A.bt.up ci95.integral.bt.lo
#> 1   16.6807752104073 3376.28167417921 3459.54521079273    79908.1646118198
#> 2   37.1039839675527 1404.78083823814  1505.1386394179    40274.9981646968
#> 3   24.2957726636641 2185.78124095749 2207.10591026091     64168.419130669
#> 4   15.7080566297105 7420.81821273076 7700.45263407996    141518.154553096
#> 5   31.5177265851534 2287.19349364914 2435.78735177672    54660.0450066769
#> 6   13.1669517723442 4799.54828109928 4838.46801475502    126588.080298181
#> 7   7.73351657127122 16703.8559195419 17312.5881748907    344070.079867312
#> 8   14.6430424362568 5107.37040217092 5256.30831503382    114081.149756302
#> 9   5.53837429637016  7733.0120352997 7798.45154263193    212321.077474781
#> 10   2.1767814084093 19935.2135503686 20219.7845198454    467612.627355144
#> 11  11.3335678993363 6601.16024720094 6947.01090053165    160364.607337643
#> 12  4.34409892910675  9169.3589081415 9209.66041980934    255556.371960407
#> 13  1.22431348350678 20192.1612968961  20661.352586556    464016.759997553
#> 14  11.3588007636009 6892.76905794233 7041.96269960138    171589.330818236
#> 15  4.24306947642786   8976.835472056 9032.99312765194    259417.544802022
#> 16 0.867975487689268 20503.5303226392 20871.7825558117    498539.093765389
#> 17  8.98423647553965 6745.24463068847 7054.71066314878    170279.937517261
#> 18  5.76863270602514 9293.85289086785 9351.69299819019    267062.477997504
#> 19  1.08065894281048 19661.9589874685 19961.8723549563    486369.597084341
#> 20  10.0834868219137 7037.57466191181 7274.95997142023    175309.680177089
#> 21  6.59666937904132 9241.02707933739 9273.87715467888    255946.070718002
#>    ci95.integral.bt.up
#> 1     84265.7650487175
#> 2     44414.6245049416
#> 3     68420.9774688264
#> 4      156144.40261288
#> 5     60550.5232657527
#> 6      135155.95748146
#> 7     378048.991607714
#> 8     124061.060651619
#> 9     229217.192891266
#> 10    500431.096459402
#> 11    174030.276277982
#> 12    276374.126769906
#> 13    516541.387231821
#> 14    182544.106962248
#> 15    272628.467016432
#> 16    529873.747389902
#> 17    181224.514685975
#> 18    276878.659545465
#> 19    520432.853814868
#> 20    185934.756879178
#> 21    276849.928003325