Generic summary function for gcBootSpline objects

Generic summary function for gcBootSpline objects

Usage

# S3 method for gcBootSpline
summary(object, ...)

Arguments

object

object of class gcBootSpline

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with statistical parameters extracted from the spline fit bootstrapping computation.

Examples

# Create random growth dataset
rnd.dataset <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')

# Extract time and growth data for single sample
time <- rnd.dataset$time[1,]
data <- rnd.dataset$data[1,-(1:3)] # Remove identifier columns

# Introduce some noise into the measurements
data <- data + stats::runif(97, -0.01, 0.09)

# Perform bootstrapping spline fit
TestFit <- growth.gcBootSpline(time, data, gcID = 'TestFit',
              control = growth.control(fit.opt = 's', nboot.gc = 50))

summary(TestFit)
#>               mu.bt        lambda.bt            A.bt            dY.bt
#> 1 0.642754193099722 5.40732893194279 6.2964955054614 6.19001281390171
#>        integral.bt           stdmu.bt      stdlambda.bt            stdA.bt
#> 1 56.7483227324953 0.0184557230552549 0.163080651612524 0.0176162703031347
#>             stddY.bt    stdintegral.bt reliable_fit.bt     ci90.mu.bt.lo
#> 1 0.0204804203006269 0.955315404470285            TRUE 0.612394528673828
#>       ci90.mu.bt.up ci90.lambda.bt.lo ci90.lambda.bt.up     ci90.A.bt.lo
#> 1 0.673113857525616  5.13906126004018  5.67559660384539 6.26751674081275
#>       ci90.A.bt.up ci90.integral.bt.lo ci90.integral.bt.up     ci95.mu.bt.lo
#> 1 6.32547427011006    55.1768288921417    58.3198165728489 0.606580975911422
#>       ci95.mu.bt.up ci95.lambda.bt.lo ci95.lambda.bt.up     ci95.A.bt.lo
#> 1 0.678927410288022  5.08769085478224  5.72696700910333 6.26196761566726
#>       ci95.A.bt.up ci95.integral.bt.lo ci95.integral.bt.up
#> 1 6.33102339525555    54.8759045397335     58.620740925257