Generic summary function for gcFitModel objects

Generic summary function for gcFitModel objects

Usage

# S3 method for gcFitModel
summary(object, ...)

Arguments

object

object of class gcFitModel

...

Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.

Value

A dataframe with parameters extracted from the growth model fit.

Examples

# Create random growth dataset
rnd.dataset <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')

# Extract time and growth data for single sample
time <- rnd.dataset$time[1,]
data <- rnd.dataset$data[1,-(1:3)] # Remove identifier columns

# Perform parametric fit
TestFit <- growth.gcFitModel(time, data, gcID = 'TestFit',
                 control = growth.control(fit.opt = 'm'))
#> --> Try to fit model logistic
#> ....... OK
#> --> Try to fit model richards
#> ....... OK
#> --> Try to fit model gompertz
#> ....... OK
#> --> Try to fit model gompertz.exp
#> ... ERROR in nls(). For further information see help(growth.gcFitModel)
#> --> Try to fit model huang
#> .......... OK
#> --> Try to fit model baranyi
#> ........ OK
#> 
#> Best fitting model: ~baranyi

summary(TestFit)
#>            mu.model          tD.model     lambda.model          A.model
#> 1 0.806383132163501 0.859575495707919 4.71765426019859 4.73666468368486
#>           dY.model     A.orig.model    dY.orig.model  integral.model
#> 1 4.76803089359124 6.22212522305327 6.42038274824604 76.957677263777
#>   parameter_nu.model parameter_alpha.model parameter_t_shift.model
#> 1               <NA>                  <NA>                    <NA>
#>    parameter_y0.model         stdmu.model    stdlambda.model
#> 1 -0.0313868257524894 0.00263355069468448 0.0172598891001026
#>            stdA.model reliable_fit.model  ci90.mu.model.lo  ci90.mu.model.up
#> 1 0.00234748186575986               TRUE 0.802050941270745 0.810715323056256
#>   ci90.lambda.model.lo ci90.lambda.model.up  ci90.A.model.lo  ci90.A.model.up
#> 1     4.68926174262892     4.74604677776826 4.73280307601568 4.74052629135403
#>    ci95.mu.model.lo  ci95.mu.model.up ci95.lambda.model.lo ci95.lambda.model.up
#> 1 0.801221372801919 0.811544891525082     4.68382487756239     4.75148364283479
#>    ci95.A.model.lo  ci95.A.model.up
#> 1 4.73206361922797 4.74126574814175