Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime            dY
#> 1      Test2 | 0 0.475 ± 0.008  1.46 ± 0.024 3.828 ± 0.073  4.17 ± 0.192
#> 2  Test2 | 0.039 0.389 ± 0.006 1.782 ± 0.026 3.906 ± 0.086 1.741 ± 0.105
#> 3  Test2 | 0.059 0.343 ± 0.004 2.022 ± 0.022 3.836 ± 0.205 1.499 ± 0.091
#> 4  Test2 | 0.088  0.292 ± 0.01 2.373 ± 0.079 3.646 ± 0.326 1.139 ± 0.083
#> 5   Test2 | 0.13 0.224 ± 0.015 3.107 ± 0.214 3.336 ± 0.288 0.842 ± 0.026
#> 6    Test2 | 0.2 0.111 ± 0.013   6.279 ± 0.7 2.059 ± 0.549  0.369 ± 0.05
#> 7    Test2 | 0.3                                                        
#> 8   Test2 | 0.44                                                        
#> 9   Test2 | 0.67                                                        
#> 10     Test2 | 1                                                        
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1  4.221 ± 0.197  7.083 ± 0.629      10 ± 0.25
#> 2  1.791 ± 0.105  7.083 ± 0.577     10 ± 0.433
#> 3  1.553 ± 0.096   6.75 ± 0.433 10.417 ± 0.764
#> 4  1.192 ± 0.086  7.417 ± 0.946  9.917 ± 0.878
#> 5  0.895 ± 0.022  7.417 ± 1.258   9.75 ± 1.521
#> 6   0.42 ± 0.053  8.833 ± 1.607 10.917 ± 1.893
#> 7                                             
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }