Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime            dY
#> 1      Test2 | 0 0.466 ± 0.009 1.489 ± 0.028 4.077 ± 0.098 4.066 ± 0.661
#> 2  Test2 | 0.039  0.39 ± 0.009 1.776 ± 0.041  4.003 ± 0.29  1.89 ± 0.268
#> 3  Test2 | 0.059 0.347 ± 0.006 1.999 ± 0.033 3.757 ± 0.073 1.467 ± 0.075
#> 4  Test2 | 0.088  0.29 ± 0.007  2.394 ± 0.06  3.74 ± 0.197   1.1 ± 0.051
#> 5   Test2 | 0.13 0.214 ± 0.007 3.234 ± 0.109 3.367 ± 0.307 0.761 ± 0.031
#> 6    Test2 | 0.2 0.116 ± 0.005 5.994 ± 0.281  2.211 ± 0.23 0.361 ± 0.012
#> 7    Test2 | 0.3 0.016 ± 0.027        14.622         8.008 0.017 ± 0.029
#> 8   Test2 | 0.44                                                        
#> 9   Test2 | 0.67                                                        
#> 10     Test2 | 1                                                        
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1   4.12 ± 0.667  6.833 ± 0.382          10.25
#> 2   1.945 ± 0.27    7.25 ± 0.25 10.083 ± 0.144
#> 3  1.521 ± 0.073  6.833 ± 0.289  9.917 ± 0.577
#> 4  1.153 ± 0.055  7.167 ± 0.144   9.667 ± 0.52
#> 5  0.811 ± 0.034  8.167 ± 0.629 10.583 ± 0.144
#> 6   0.412 ± 0.01  8.583 ± 3.263 10.667 ± 3.357
#> 7          0.099           20.5           22.5
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }