Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime            dY
#> 1      Test2 | 0 0.469 ± 0.006  1.479 ± 0.02 3.856 ± 0.201 3.935 ± 0.562
#> 2  Test2 | 0.039  0.39 ± 0.002 1.779 ± 0.009 4.042 ± 0.077 1.794 ± 0.133
#> 3  Test2 | 0.059 0.338 ± 0.013 2.055 ± 0.076  3.904 ± 0.15 1.433 ± 0.156
#> 4  Test2 | 0.088 0.293 ± 0.011 2.372 ± 0.088 3.769 ± 0.157 1.225 ± 0.101
#> 5   Test2 | 0.13 0.219 ± 0.004 3.168 ± 0.063 3.429 ± 0.191 0.874 ± 0.031
#> 6    Test2 | 0.2  0.12 ± 0.003 5.793 ± 0.132   2.709 ± 0.1 0.349 ± 0.005
#> 7    Test2 | 0.3                                                        
#> 8   Test2 | 0.44                                                        
#> 9   Test2 | 0.67                                                        
#> 10     Test2 | 1                                                        
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1  3.988 ± 0.566     6.5 ± 0.25      10 ± 0.25
#> 2  1.846 ± 0.136       7 ± 0.25   10.25 ± 0.25
#> 3  1.485 ± 0.159            6.5       9.75 ± 1
#> 4   1.28 ± 0.103    7.5 ± 0.661   10.75 ± 0.25
#> 5  0.927 ± 0.033  8.333 ± 0.722 11.083 ± 0.289
#> 6  0.399 ± 0.007   9.167 ± 1.01  11.25 ± 1.323
#> 7                                             
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }