Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime            dY
#> 1      Test2 | 0 0.472 ± 0.004 1.468 ± 0.011 4.023 ± 0.212  3.79 ± 0.209
#> 2  Test2 | 0.039 0.388 ± 0.002 1.789 ± 0.011  3.862 ± 0.11 1.774 ± 0.289
#> 3  Test2 | 0.059 0.344 ± 0.006 2.014 ± 0.037 3.661 ± 0.052  1.47 ± 0.209
#> 4  Test2 | 0.088 0.284 ± 0.005 2.439 ± 0.042 3.558 ± 0.102 1.054 ± 0.014
#> 5   Test2 | 0.13 0.218 ± 0.006 3.186 ± 0.083 3.543 ± 0.269 0.848 ± 0.066
#> 6    Test2 | 0.2  0.12 ± 0.004 5.781 ± 0.175 2.598 ± 0.391  0.388 ± 0.03
#> 7    Test2 | 0.3 0.014 ± 0.025        16.087         3.426 0.018 ± 0.032
#> 8   Test2 | 0.44                                                        
#> 9   Test2 | 0.67                                                        
#> 10     Test2 | 1                                                        
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1  3.839 ± 0.208  6.833 ± 0.382   9.75 ± 0.661
#> 2  1.825 ± 0.295       7 ± 0.25     10 ± 0.433
#> 3  1.522 ± 0.213  6.417 ± 0.144  10.25 ± 0.433
#> 4  1.104 ± 0.015  7.083 ± 0.577   10.25 ± 1.09
#> 5  0.904 ± 0.066  8.667 ± 1.258 10.917 ± 0.946
#> 6   0.441 ± 0.03  9.917 ± 3.055  12.25 ± 2.947
#> 7          0.105          10.75             13
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }