Generate a grouped results table for linear fits with average and standard deviations

Generate a grouped results table for linear fits with average and standard deviations

Usage

table_group_growth_linear(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped linear fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# \donttest{
# Create random growth data set
rnd.data <- rdm.data(d = 30, mu = 0.6, A = 4.5, label = "Test2")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "l",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_linear(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime            dY
#> 1      Test2 | 0 0.464 ± 0.003  1.495 ± 0.009 4.027 ± 0.108 4.012 ± 0.137
#> 2  Test2 | 0.039 0.386 ± 0.012  1.799 ± 0.057 4.031 ± 0.081 1.884 ± 0.183
#> 3  Test2 | 0.059 0.352 ± 0.009  1.971 ± 0.052 4.139 ± 0.109 1.429 ± 0.124
#> 4  Test2 | 0.088 0.297 ± 0.005  2.336 ± 0.039 3.876 ± 0.097  1.139 ± 0.02
#> 5   Test2 | 0.13 0.221 ± 0.003  3.137 ± 0.046 3.537 ± 0.355 0.882 ± 0.008
#> 6    Test2 | 0.2 0.119 ± 0.003  5.844 ± 0.169 2.746 ± 0.557 0.383 ± 0.039
#> 7    Test2 | 0.3 0.035 ± 0.031 13.335 ± 2.624 5.478 ± 2.972 0.052 ± 0.045
#> 8   Test2 | 0.44                                                         
#> 9   Test2 | 0.67                                                         
#> 10     Test2 | 1                                                         
#>            Y_max t_start(mumax)   t_end(mumax)
#> 1  4.066 ± 0.137    6.75 ± 0.25    10.5 ± 0.25
#> 2  1.938 ± 0.185  6.667 ± 0.144   10.5 ± 0.433
#> 3  1.479 ± 0.124        7 ± 0.5  10.333 ± 0.52
#> 4   1.19 ± 0.021  7.333 ± 0.382   9.917 ± 0.52
#> 5  0.935 ± 0.011  8.583 ± 0.382 11.333 ± 0.382
#> 6  0.438 ± 0.042      11 ± 1.25 12.833 ± 1.258
#> 7  0.128 ± 0.011 19.875 ± 0.884 21.625 ± 0.884
#> 8                                             
#> 9                                             
#> 10                                            

# with HTML formatting
DT::datatable(table_group_growth_linear(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities

# }