Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.803 ± 0.006  0.863 ± 0.006 5.123 ± 0.152 4.818 ± 0.058
#> 2  Test1 | 0.017 0.795 ± 0.004  0.872 ± 0.004 5.222 ± 0.106  4.208 ± 0.06
#> 3  Test1 | 0.026 0.779 ± 0.005   0.89 ± 0.006  5.27 ± 0.204  4.083 ± 0.01
#> 4  Test1 | 0.039 0.761 ± 0.011  0.911 ± 0.013  5.27 ± 0.098 3.962 ± 0.083
#> 5  Test1 | 0.059 0.736 ± 0.009  0.942 ± 0.011 5.465 ± 0.186 3.683 ± 0.035
#> 6  Test1 | 0.088  0.68 ± 0.003  1.019 ± 0.005 5.545 ± 0.146 3.479 ± 0.063
#> 7   Test1 | 0.13 0.589 ± 0.009  1.176 ± 0.019 5.734 ± 0.209 3.144 ± 0.049
#> 8    Test1 | 0.2 0.444 ± 0.003   1.56 ± 0.011 5.939 ± 0.212 2.759 ± 0.034
#> 9    Test1 | 0.3 0.276 ± 0.004  2.511 ± 0.039 6.323 ± 0.146 2.223 ± 0.056
#> 10  Test1 | 0.44  0.14 ± 0.012  4.967 ± 0.431 6.635 ± 0.031 1.679 ± 0.057
#> 11  Test1 | 0.67  0.07 ± 0.021 10.658 ± 3.741 4.884 ± 3.583 0.846 ± 0.026
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.831 ± 0.054
#> 2  4.219 ± 0.052
#> 3  4.083 ± 0.024
#> 4  3.956 ± 0.079
#> 5  3.673 ± 0.041
#> 6  3.468 ± 0.074
#> 7   3.14 ± 0.052
#> 8  2.755 ± 0.039
#> 9  2.153 ± 0.057
#> 10 1.371 ± 0.077
#> 11 0.544 ± 0.055
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities