Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.803 ± 0.001  0.863 ± 0.001 5.129 ± 0.175  4.78 ± 0.115
#> 2  Test1 | 0.017  0.79 ± 0.011  0.877 ± 0.013 5.223 ± 0.175 4.262 ± 0.085
#> 3  Test1 | 0.026 0.779 ± 0.005   0.89 ± 0.006 5.089 ± 0.234 4.115 ± 0.103
#> 4  Test1 | 0.039 0.764 ± 0.004  0.907 ± 0.005 5.372 ± 0.179 3.943 ± 0.098
#> 5  Test1 | 0.059 0.732 ± 0.003  0.947 ± 0.003 5.311 ± 0.161 3.672 ± 0.106
#> 6  Test1 | 0.088 0.683 ± 0.003  1.014 ± 0.005 5.677 ± 0.079 3.426 ± 0.025
#> 7   Test1 | 0.13 0.591 ± 0.003  1.174 ± 0.007  5.754 ± 0.25 3.166 ± 0.085
#> 8    Test1 | 0.2 0.439 ± 0.009  1.578 ± 0.034 5.978 ± 0.319  2.72 ± 0.055
#> 9    Test1 | 0.3  0.272 ± 0.01  2.549 ± 0.095 6.246 ± 0.243 2.205 ± 0.093
#> 10  Test1 | 0.44 0.141 ± 0.011   4.931 ± 0.39  6.783 ± 0.43 1.668 ± 0.097
#> 11  Test1 | 0.67  0.03 ± 0.003 23.466 ± 2.075 0.566 ± 0.358 1.103 ± 0.298
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.769 ± 0.104
#> 2  4.264 ± 0.076
#> 3  4.107 ± 0.105
#> 4  3.949 ± 0.086
#> 5  3.689 ± 0.102
#> 6  3.431 ± 0.032
#> 7  3.158 ± 0.088
#> 8  2.706 ± 0.042
#> 9  2.139 ± 0.084
#> 10 1.359 ± 0.013
#> 11 0.554 ± 0.051
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities