Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0         0.806           0.86 5.153 ± 0.197  4.811 ± 0.09
#> 2  Test1 | 0.017 0.791 ± 0.011  0.877 ± 0.012  5.23 ± 0.198  4.279 ± 0.08
#> 3  Test1 | 0.026 0.772 ± 0.008  0.898 ± 0.009  5.19 ± 0.203 4.138 ± 0.072
#> 4  Test1 | 0.039 0.765 ± 0.001  0.906 ± 0.001 5.234 ± 0.083 3.936 ± 0.021
#> 5  Test1 | 0.059 0.733 ± 0.006  0.945 ± 0.007 5.376 ± 0.198 3.764 ± 0.041
#> 6  Test1 | 0.088 0.672 ± 0.007  1.032 ± 0.011 5.591 ± 0.202 3.442 ± 0.075
#> 7   Test1 | 0.13 0.586 ± 0.007  1.182 ± 0.015 5.797 ± 0.217 3.109 ± 0.015
#> 8    Test1 | 0.2 0.448 ± 0.002  1.548 ± 0.008 6.092 ± 0.047 2.727 ± 0.091
#> 9    Test1 | 0.3 0.265 ± 0.005  2.614 ± 0.051 6.308 ± 0.221 2.248 ± 0.099
#> 10  Test1 | 0.44 0.144 ± 0.017   4.838 ± 0.52 7.052 ± 0.593  1.61 ± 0.148
#> 11  Test1 | 0.67 0.043 ± 0.015 17.786 ± 6.627 2.445 ± 3.297 0.994 ± 0.201
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.799 ± 0.083
#> 2  4.279 ± 0.081
#> 3  4.136 ± 0.057
#> 4  3.936 ± 0.036
#> 5  3.766 ± 0.058
#> 6  3.446 ± 0.061
#> 7  3.123 ± 0.021
#> 8  2.712 ± 0.087
#> 9   2.181 ± 0.09
#> 10 1.317 ± 0.053
#> 11 0.559 ± 0.025
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities