Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.798 ± 0.008  0.869 ± 0.008  4.991 ± 0.24 4.771 ± 0.031
#> 2  Test1 | 0.017         0.795          0.872 5.344 ± 0.133 4.239 ± 0.041
#> 3  Test1 | 0.026 0.777 ± 0.009  0.892 ± 0.011 5.237 ± 0.221 4.117 ± 0.082
#> 4  Test1 | 0.039 0.757 ± 0.008   0.915 ± 0.01 5.309 ± 0.205 3.965 ± 0.066
#> 5  Test1 | 0.059 0.732 ± 0.007  0.947 ± 0.008  5.335 ± 0.21 3.661 ± 0.044
#> 6  Test1 | 0.088 0.685 ± 0.004  1.011 ± 0.006 5.534 ± 0.175  3.438 ± 0.09
#> 7   Test1 | 0.13          0.59  1.174 ± 0.001   5.773 ± 0.2 3.135 ± 0.055
#> 8    Test1 | 0.2 0.441 ± 0.004  1.571 ± 0.013  6.048 ± 0.16  2.712 ± 0.07
#> 9    Test1 | 0.3 0.277 ± 0.004  2.504 ± 0.039  6.19 ± 0.251 2.155 ± 0.021
#> 10  Test1 | 0.44 0.137 ± 0.008  5.087 ± 0.292 6.699 ± 0.329 1.618 ± 0.118
#> 11  Test1 | 0.67 0.066 ± 0.024 11.533 ± 4.171  4.595 ± 2.96 0.987 ± 0.169
#> 12     Test1 | 1                                                         
#>               dY
#> 1   4.78 ± 0.042
#> 2  4.228 ± 0.042
#> 3  4.115 ± 0.062
#> 4   3.96 ± 0.052
#> 5  3.675 ± 0.045
#> 6  3.444 ± 0.083
#> 7  3.143 ± 0.069
#> 8  2.705 ± 0.071
#> 9  2.113 ± 0.022
#> 10 1.307 ± 0.059
#> 11  0.582 ± 0.06
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities