Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0   0.8 ± 0.004  0.866 ± 0.004 4.984 ± 0.191  4.818 ± 0.05
#> 2  Test1 | 0.017 0.781 ± 0.006  0.888 ± 0.006 5.356 ± 0.114 4.242 ± 0.091
#> 3  Test1 | 0.026 0.777 ± 0.007  0.892 ± 0.008 5.241 ± 0.233   4.1 ± 0.046
#> 4  Test1 | 0.039 0.762 ± 0.007  0.909 ± 0.009 5.444 ± 0.093 3.891 ± 0.025
#> 5  Test1 | 0.059 0.726 ± 0.004  0.955 ± 0.005  5.519 ± 0.16 3.738 ± 0.109
#> 6  Test1 | 0.088 0.673 ± 0.005   1.03 ± 0.007 5.496 ± 0.132 3.428 ± 0.091
#> 7   Test1 | 0.13 0.591 ± 0.003  1.172 ± 0.007 5.853 ± 0.267 3.156 ± 0.086
#> 8    Test1 | 0.2 0.443 ± 0.005  1.565 ± 0.018  5.921 ± 0.25 2.701 ± 0.059
#> 9    Test1 | 0.3 0.277 ± 0.006  2.499 ± 0.052 6.161 ± 0.162  2.184 ± 0.11
#> 10  Test1 | 0.44 0.134 ± 0.003  5.181 ± 0.132 6.742 ± 0.035 1.599 ± 0.035
#> 11  Test1 | 0.67 0.048 ± 0.026 16.984 ± 7.354 1.545 ± 6.008 1.193 ± 0.343
#> 12     Test1 | 1                                                         
#>               dY
#> 1   4.817 ± 0.05
#> 2  4.253 ± 0.104
#> 3  4.093 ± 0.042
#> 4  3.886 ± 0.035
#> 5  3.743 ± 0.096
#> 6  3.431 ± 0.073
#> 7  3.156 ± 0.078
#> 8  2.692 ± 0.085
#> 9  2.121 ± 0.092
#> 10 1.293 ± 0.041
#> 11 0.532 ± 0.011
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities