Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.799 ± 0.006  0.867 ± 0.006  5.09 ± 0.153  4.73 ± 0.026
#> 2  Test1 | 0.017 0.793 ± 0.003  0.874 ± 0.003   5.2 ± 0.269 4.267 ± 0.098
#> 3  Test1 | 0.026 0.778 ± 0.003  0.891 ± 0.004 5.209 ± 0.165 4.127 ± 0.069
#> 4  Test1 | 0.039 0.713 ± 0.089  0.983 ± 0.132 5.086 ± 0.579 3.903 ± 0.014
#> 5  Test1 | 0.059 0.731 ± 0.012  0.949 ± 0.015 5.366 ± 0.204 3.714 ± 0.085
#> 6  Test1 | 0.088  0.68 ± 0.005  1.019 ± 0.008 5.487 ± 0.136 3.444 ± 0.021
#> 7   Test1 | 0.13 0.597 ± 0.002  1.162 ± 0.003 5.646 ± 0.172 3.081 ± 0.013
#> 8    Test1 | 0.2 0.435 ± 0.005  1.594 ± 0.019 6.101 ± 0.167 2.713 ± 0.052
#> 9    Test1 | 0.3 0.276 ± 0.004   2.514 ± 0.04   6.3 ± 0.205 2.253 ± 0.033
#> 10  Test1 | 0.44 0.139 ± 0.003  4.985 ± 0.108 6.694 ± 0.163 1.662 ± 0.099
#> 11  Test1 | 0.67 0.075 ± 0.025 10.272 ± 4.343 6.224 ± 3.343 0.906 ± 0.193
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.722 ± 0.027
#> 2  4.288 ± 0.123
#> 3  4.126 ± 0.068
#> 4  3.903 ± 0.027
#> 5  3.711 ± 0.086
#> 6   3.428 ± 0.03
#> 7  3.093 ± 0.027
#> 8    2.7 ± 0.075
#> 9  2.198 ± 0.013
#> 10 1.344 ± 0.072
#> 11 0.559 ± 0.066
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities