Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.799 ± 0.008 0.868 ± 0.008 4.931 ± 0.177 4.784 ± 0.046
#> 2  Test1 | 0.017 0.785 ± 0.004 0.883 ± 0.005 4.942 ± 0.056 4.287 ± 0.109
#> 3  Test1 | 0.026  0.78 ± 0.009 0.889 ± 0.011 5.232 ± 0.167 4.143 ± 0.085
#> 4  Test1 | 0.039 0.764 ± 0.009 0.907 ± 0.011 5.229 ± 0.041 3.993 ± 0.015
#> 5  Test1 | 0.059  0.73 ± 0.011 0.949 ± 0.014 5.545 ± 0.217 3.693 ± 0.034
#> 6  Test1 | 0.088  0.68 ± 0.014  1.02 ± 0.021 5.756 ± 0.089 3.512 ± 0.039
#> 7   Test1 | 0.13 0.588 ± 0.003  1.18 ± 0.007 5.626 ± 0.159 3.169 ± 0.072
#> 8    Test1 | 0.2 0.446 ± 0.002 1.553 ± 0.008   5.9 ± 0.026 2.751 ± 0.043
#> 9    Test1 | 0.3 0.273 ± 0.009 2.539 ± 0.082 6.326 ± 0.271 2.187 ± 0.049
#> 10  Test1 | 0.44 0.142 ± 0.003 4.897 ± 0.095 6.962 ± 0.259 1.627 ± 0.084
#> 11  Test1 | 0.67 0.048 ± 0.003 14.59 ± 0.954 1.422 ± 0.174  1.03 ± 0.107
#> 12     Test1 | 1                                                        
#>               dY
#> 1  4.774 ± 0.029
#> 2  4.278 ± 0.109
#> 3  4.149 ± 0.071
#> 4  3.986 ± 0.017
#> 5   3.68 ± 0.041
#> 6  3.518 ± 0.032
#> 7  3.164 ± 0.054
#> 8  2.753 ± 0.044
#> 9  2.124 ± 0.057
#> 10 1.333 ± 0.067
#> 11 0.544 ± 0.029
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities