Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.797 ± 0.003 0.869 ± 0.003 4.869 ± 0.118   4.78 ± 0.08
#> 2  Test1 | 0.017  0.788 ± 0.01  0.88 ± 0.011 5.158 ± 0.065 4.329 ± 0.008
#> 3  Test1 | 0.026 0.787 ± 0.004 0.881 ± 0.004 5.218 ± 0.065 4.091 ± 0.016
#> 4  Test1 | 0.039 0.771 ± 0.007 0.899 ± 0.008   5.43 ± 0.19 3.964 ± 0.064
#> 5  Test1 | 0.059  0.73 ± 0.012 0.949 ± 0.016 5.522 ± 0.124 3.703 ± 0.066
#> 6  Test1 | 0.088 0.683 ± 0.009 1.015 ± 0.014 5.591 ± 0.188 3.443 ± 0.042
#> 7   Test1 | 0.13  0.59 ± 0.004 1.175 ± 0.009 5.708 ± 0.272 3.083 ± 0.019
#> 8    Test1 | 0.2 0.436 ± 0.003   1.59 ± 0.01 5.985 ± 0.216 2.716 ± 0.068
#> 9    Test1 | 0.3 0.278 ± 0.009 2.499 ± 0.085 6.405 ± 0.261 2.216 ± 0.059
#> 10  Test1 | 0.44 0.139 ± 0.007 4.982 ± 0.259 6.793 ± 0.243 1.607 ± 0.034
#> 11  Test1 | 0.67 0.075 ± 0.022 9.976 ± 3.478 5.332 ± 4.143 0.912 ± 0.144
#> 12     Test1 | 1                                                        
#>               dY
#> 1  4.775 ± 0.063
#> 2    4.31 ± 0.02
#> 3  4.084 ± 0.015
#> 4  3.971 ± 0.072
#> 5  3.691 ± 0.079
#> 6  3.437 ± 0.043
#> 7  3.086 ± 0.038
#> 8  2.699 ± 0.067
#> 9  2.152 ± 0.055
#> 10  1.298 ± 0.01
#> 11 0.588 ± 0.054
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities