Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.805 ± 0.009 0.861 ± 0.009 4.942 ± 0.206 4.766 ± 0.011
#> 2  Test1 | 0.017 0.788 ± 0.006  0.88 ± 0.006 5.454 ± 0.029 4.201 ± 0.009
#> 3  Test1 | 0.026 0.784 ± 0.002 0.884 ± 0.003 5.172 ± 0.199 4.098 ± 0.023
#> 4  Test1 | 0.039 0.761 ± 0.006  0.91 ± 0.007 5.341 ± 0.187  3.98 ± 0.057
#> 5  Test1 | 0.059 0.731 ± 0.009 0.949 ± 0.012 5.351 ± 0.236  3.75 ± 0.067
#> 6  Test1 | 0.088 0.681 ± 0.005 1.018 ± 0.008 5.632 ± 0.245  3.481 ± 0.05
#> 7   Test1 | 0.13 0.582 ± 0.012 1.191 ± 0.024 5.784 ± 0.133  3.17 ± 0.055
#> 8    Test1 | 0.2 0.437 ± 0.006 1.585 ± 0.023 6.052 ± 0.295 2.681 ± 0.071
#> 9    Test1 | 0.3 0.268 ± 0.002 2.582 ± 0.023 6.127 ± 0.198 2.254 ± 0.066
#> 10  Test1 | 0.44 0.145 ± 0.007 4.783 ± 0.238 6.757 ± 0.288 1.584 ± 0.062
#> 11  Test1 | 0.67 0.032 ± 0.004 21.749 ± 2.65  0.72 ± 0.754 1.039 ± 0.109
#> 12     Test1 | 1                                                        
#>               dY
#> 1  4.755 ± 0.038
#> 2  4.201 ± 0.025
#> 3   4.09 ± 0.021
#> 4  3.969 ± 0.047
#> 5   3.75 ± 0.075
#> 6  3.465 ± 0.044
#> 7  3.168 ± 0.051
#> 8  2.677 ± 0.067
#> 9  2.177 ± 0.071
#> 10 1.316 ± 0.031
#> 11 0.557 ± 0.068
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities