Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.804 ± 0.007  0.862 ± 0.007 4.923 ± 0.138  4.789 ± 0.06
#> 2  Test1 | 0.017 0.793 ± 0.003  0.874 ± 0.004   5.28 ± 0.06  4.33 ± 0.034
#> 3  Test1 | 0.026 0.782 ± 0.011  0.887 ± 0.013 5.248 ± 0.194 4.079 ± 0.015
#> 4  Test1 | 0.039 0.766 ± 0.001  0.905 ± 0.001 5.219 ± 0.037  3.968 ± 0.08
#> 5  Test1 | 0.059 0.733 ± 0.006  0.946 ± 0.008  5.499 ± 0.21 3.786 ± 0.041
#> 6  Test1 | 0.088 0.676 ± 0.008  1.026 ± 0.012 5.512 ± 0.203 3.454 ± 0.028
#> 7   Test1 | 0.13 0.528 ± 0.115  1.362 ± 0.338  5.367 ± 0.66 3.119 ± 0.067
#> 8    Test1 | 0.2 0.442 ± 0.006   1.569 ± 0.02 6.006 ± 0.102 2.801 ± 0.036
#> 9    Test1 | 0.3 0.272 ± 0.007  2.547 ± 0.066  6.316 ± 0.15 2.157 ± 0.043
#> 10  Test1 | 0.44 0.135 ± 0.003  5.155 ± 0.116  6.67 ± 0.143 1.604 ± 0.045
#> 11  Test1 | 0.67 0.056 ± 0.033 16.181 ± 9.876 3.086 ± 4.589 0.941 ± 0.119
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.787 ± 0.075
#> 2  4.327 ± 0.028
#> 3  4.081 ± 0.016
#> 4  3.974 ± 0.079
#> 5   3.78 ± 0.037
#> 6  3.447 ± 0.009
#> 7  3.115 ± 0.084
#> 8  2.776 ± 0.038
#> 9  2.085 ± 0.046
#> 10 1.293 ± 0.035
#> 11 0.561 ± 0.024
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities