Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.804 ± 0.005 0.862 ± 0.005 5.016 ± 0.161 4.802 ± 0.049
#> 2  Test1 | 0.017 0.782 ± 0.001 0.886 ± 0.002 5.215 ± 0.275 4.321 ± 0.083
#> 3  Test1 | 0.026 0.772 ± 0.002 0.898 ± 0.002 5.386 ± 0.184  4.052 ± 0.05
#> 4  Test1 | 0.039  0.76 ± 0.009 0.913 ± 0.011 5.368 ± 0.033 3.884 ± 0.013
#> 5  Test1 | 0.059  0.732 ± 0.01 0.947 ± 0.013 5.443 ± 0.164 3.666 ± 0.067
#> 6  Test1 | 0.088 0.682 ± 0.006 1.017 ± 0.009  5.62 ± 0.073 3.464 ± 0.077
#> 7   Test1 | 0.13 0.591 ± 0.007 1.174 ± 0.013 5.786 ± 0.279  3.172 ± 0.07
#> 8    Test1 | 0.2 0.444 ± 0.008 1.561 ± 0.029 5.822 ± 0.219 2.697 ± 0.029
#> 9    Test1 | 0.3 0.277 ± 0.006 2.502 ± 0.052 6.587 ± 0.083 2.209 ± 0.046
#> 10  Test1 | 0.44 0.136 ± 0.007 5.096 ± 0.287  6.74 ± 0.154   1.7 ± 0.066
#> 11  Test1 | 0.67 0.057 ± 0.025 14.39 ± 7.775 4.256 ± 3.877 0.935 ± 0.166
#> 12     Test1 | 1                                                        
#>               dY
#> 1  4.794 ± 0.056
#> 2   4.314 ± 0.07
#> 3  4.059 ± 0.026
#> 4  3.894 ± 0.016
#> 5  3.687 ± 0.049
#> 6  3.457 ± 0.071
#> 7   3.165 ± 0.07
#> 8  2.682 ± 0.045
#> 9   2.14 ± 0.044
#> 10 1.356 ± 0.049
#> 11 0.565 ± 0.045
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities