Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax              tD       lagtime         Y_max
#> 1      Test1 | 0 0.799 ± 0.004   0.868 ± 0.004 4.921 ± 0.156 4.742 ± 0.078
#> 2  Test1 | 0.017 0.791 ± 0.006   0.876 ± 0.006 5.187 ± 0.385 4.297 ± 0.073
#> 3  Test1 | 0.026 0.779 ± 0.007    0.89 ± 0.008  5.156 ± 0.21  4.186 ± 0.06
#> 4  Test1 | 0.039 0.759 ± 0.009    0.913 ± 0.01 5.308 ± 0.084 3.957 ± 0.086
#> 5  Test1 | 0.059 0.735 ± 0.003   0.944 ± 0.003   5.4 ± 0.223 3.661 ± 0.054
#> 6  Test1 | 0.088 0.672 ± 0.002   1.031 ± 0.003 5.555 ± 0.278 3.493 ± 0.078
#> 7   Test1 | 0.13 0.591 ± 0.008   1.173 ± 0.015 5.827 ± 0.137 3.183 ± 0.055
#> 8    Test1 | 0.2 0.446 ± 0.009   1.555 ± 0.031  6.01 ± 0.157 2.663 ± 0.033
#> 9    Test1 | 0.3 0.272 ± 0.006   2.552 ± 0.059 6.183 ± 0.127 2.179 ± 0.093
#> 10  Test1 | 0.44 0.138 ± 0.006   5.046 ± 0.223 6.592 ± 0.546 1.651 ± 0.134
#> 11  Test1 | 0.67 0.058 ± 0.029 15.286 ± 10.442 5.442 ± 4.259 0.863 ± 0.064
#> 12     Test1 | 1                                                          
#>               dY
#> 1  4.749 ± 0.069
#> 2   4.298 ± 0.08
#> 3  4.178 ± 0.046
#> 4  3.976 ± 0.091
#> 5  3.669 ± 0.066
#> 6  3.492 ± 0.064
#> 7  3.186 ± 0.071
#> 8  2.659 ± 0.039
#> 9   2.126 ± 0.08
#> 10 1.339 ± 0.091
#> 11 0.543 ± 0.036
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities