Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.798 ± 0.004  0.869 ± 0.004 4.902 ± 0.102 4.786 ± 0.072
#> 2  Test1 | 0.017         0.782          0.887 5.301 ± 0.249 4.308 ± 0.021
#> 3  Test1 | 0.026 0.782 ± 0.008  0.887 ± 0.009 5.167 ± 0.241 4.123 ± 0.049
#> 4  Test1 | 0.039 0.767 ± 0.005  0.904 ± 0.005 5.419 ± 0.233 3.971 ± 0.058
#> 5  Test1 | 0.059 0.726 ± 0.006  0.955 ± 0.008 5.454 ± 0.162 3.703 ± 0.077
#> 6  Test1 | 0.088  0.673 ± 0.01   1.03 ± 0.015 5.498 ± 0.167 3.411 ± 0.057
#> 7   Test1 | 0.13 0.594 ± 0.004  1.167 ± 0.008 5.658 ± 0.136 3.189 ± 0.059
#> 8    Test1 | 0.2         0.453  1.532 ± 0.001 5.898 ± 0.211 2.641 ± 0.043
#> 9    Test1 | 0.3 0.272 ± 0.006  2.546 ± 0.056 6.415 ± 0.204 2.226 ± 0.048
#> 10  Test1 | 0.44 0.142 ± 0.007   4.88 ± 0.245 6.726 ± 0.501 1.639 ± 0.063
#> 11  Test1 | 0.67 0.032 ± 0.007 22.375 ± 5.099  0.36 ± 0.533 0.936 ± 0.274
#> 12     Test1 | 1                                                         
#>               dY
#> 1  4.799 ± 0.066
#> 2  4.312 ± 0.021
#> 3  4.124 ± 0.045
#> 4  3.975 ± 0.061
#> 5  3.702 ± 0.072
#> 6  3.402 ± 0.049
#> 7  3.167 ± 0.047
#> 8   2.652 ± 0.04
#> 9  2.166 ± 0.055
#> 10 1.342 ± 0.036
#> 11  0.542 ± 0.09
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities