Generate a grouped results table for parametric fits with average and standard deviations

Generate a grouped results table for parametric fits with average and standard deviations

Usage

table_group_growth_model(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped model fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "m",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_model(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.777 ± 0.038  0.893 ± 0.045 4.895 ± 0.333 4.765 ± 0.081
#> 2  Test1 | 0.017 0.791 ± 0.008  0.876 ± 0.009 5.087 ± 0.129 4.274 ± 0.078
#> 3  Test1 | 0.026 0.779 ± 0.002   0.89 ± 0.002 5.236 ± 0.079 4.082 ± 0.037
#> 4  Test1 | 0.039 0.761 ± 0.009   0.911 ± 0.01 5.278 ± 0.265 3.961 ± 0.054
#> 5  Test1 | 0.059  0.68 ± 0.101  1.037 ± 0.168 5.283 ± 0.611 3.717 ± 0.085
#> 6  Test1 | 0.088 0.675 ± 0.007  1.027 ± 0.011 5.589 ± 0.108 3.505 ± 0.042
#> 7   Test1 | 0.13 0.589 ± 0.008  1.178 ± 0.016 5.643 ± 0.167 3.133 ± 0.092
#> 8    Test1 | 0.2 0.446 ± 0.004  1.553 ± 0.015 6.067 ± 0.049 2.736 ± 0.044
#> 9    Test1 | 0.3 0.277 ± 0.008  2.505 ± 0.071  6.382 ± 0.08  2.27 ± 0.035
#> 10  Test1 | 0.44 0.137 ± 0.009  5.068 ± 0.318 7.007 ± 0.205 1.668 ± 0.065
#> 11  Test1 | 0.67 0.045 ± 0.015 16.707 ± 5.802 2.508 ± 2.217 1.176 ± 0.076
#> 12     Test1 | 1                                                         
#>               dY
#> 1   4.761 ± 0.06
#> 2  4.285 ± 0.072
#> 3   4.08 ± 0.029
#> 4   3.967 ± 0.05
#> 5  3.717 ± 0.077
#> 6  3.497 ± 0.045
#> 7  3.126 ± 0.097
#> 8   2.73 ± 0.037
#> 9  2.199 ± 0.041
#> 10 1.333 ± 0.061
#> 11 0.623 ± 0.025
#> 12              

# with HTML formatting
DT::datatable(table_group_growth_model(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities