Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0  0.67 ± 0.003  1.035 ± 0.005 4.397 ± 0.309 6.422 ± 0.477
#> 2  Test1 | 0.017 0.617 ± 0.008  1.124 ± 0.015 4.542 ± 0.104 3.794 ± 0.489
#> 3  Test1 | 0.026 0.607 ± 0.007  1.141 ± 0.014 4.533 ± 0.169  3.24 ± 0.389
#> 4  Test1 | 0.039 0.582 ± 0.008   1.19 ± 0.016  4.48 ± 0.262 2.838 ± 0.226
#> 5  Test1 | 0.059  0.531 ± 0.01  1.307 ± 0.025 4.392 ± 0.151 2.162 ± 0.284
#> 6  Test1 | 0.088 0.469 ± 0.014  1.479 ± 0.045 4.481 ± 0.066 1.721 ± 0.196
#> 7   Test1 | 0.13 0.389 ± 0.002  1.781 ± 0.007 4.299 ± 0.158 1.125 ± 0.091
#> 8    Test1 | 0.2 0.271 ± 0.007  2.563 ± 0.062 3.863 ± 0.129 0.778 ± 0.013
#> 9    Test1 | 0.3 0.153 ± 0.005  4.532 ± 0.148  3.15 ± 0.276 0.451 ± 0.054
#> 10  Test1 | 0.44 0.069 ± 0.002  10.01 ± 0.291 1.759 ± 0.386 0.186 ± 0.011
#> 11  Test1 | 0.67 0.034 ± 0.003 20.617 ± 1.773 1.786 ± 1.908 0.089 ± 0.004
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.372 ± 0.475 8.083 ± 0.289
#> 2  3.741 ± 0.485 8.125 ± 0.177
#> 3  3.189 ± 0.387      8 ± 0.25
#> 4  2.787 ± 0.224      8 ± 0.25
#> 5  2.108 ± 0.279   7.75 ± 0.25
#> 6  1.668 ± 0.194 8.167 ± 0.144
#> 7  1.075 ± 0.089 8.333 ± 0.144
#> 8  0.728 ± 0.012 8.917 ± 0.804
#> 9     0.4 ± 0.05 9.417 ± 0.382
#> 10 0.136 ± 0.009 9.333 ± 0.804
#> 11 0.038 ± 0.004 7.083 ± 6.857
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities