Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.671 ± 0.006  1.034 ± 0.009   4.4 ± 0.233  6.617 ± 0.41
#> 2  Test1 | 0.017 0.624 ± 0.011    1.11 ± 0.02 4.516 ± 0.012  3.565 ± 0.46
#> 3  Test1 | 0.026 0.607 ± 0.008  1.142 ± 0.015 4.511 ± 0.124 3.321 ± 0.274
#> 4  Test1 | 0.039 0.573 ± 0.004  1.209 ± 0.008 4.507 ± 0.208 2.727 ± 0.142
#> 5  Test1 | 0.059 0.532 ± 0.007  1.302 ± 0.017 4.438 ± 0.211 2.188 ± 0.245
#> 6  Test1 | 0.088 0.471 ± 0.007  1.471 ± 0.021 4.297 ± 0.084  1.574 ± 0.12
#> 7   Test1 | 0.13 0.387 ± 0.001  1.792 ± 0.006 4.181 ± 0.167 1.231 ± 0.187
#> 8    Test1 | 0.2 0.263 ± 0.006  2.636 ± 0.063 3.831 ± 0.105 0.767 ± 0.061
#> 9    Test1 | 0.3  0.15 ± 0.004  4.619 ± 0.113 3.118 ± 0.141 0.488 ± 0.027
#> 10  Test1 | 0.44 0.075 ± 0.001  9.186 ± 0.174 2.354 ± 0.283 0.203 ± 0.013
#> 11  Test1 | 0.67 0.034 ± 0.005 20.503 ± 3.213  2.599 ± 1.87 0.091 ± 0.008
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.563 ± 0.406 7.917 ± 0.382
#> 2  3.516 ± 0.459          7.75
#> 3   3.267 ± 0.27 7.833 ± 0.144
#> 4   2.675 ± 0.14 7.917 ± 0.144
#> 5  2.136 ± 0.244 7.917 ± 0.289
#> 6  1.523 ± 0.117 8.083 ± 0.144
#> 7  1.176 ± 0.184      8 ± 0.25
#> 8  0.714 ± 0.057 8.583 ± 0.629
#> 9  0.432 ± 0.027  9.25 ± 0.901
#> 10  0.151 ± 0.01 10.917 ± 1.01
#> 11 0.039 ± 0.005   9.5 ± 7.653
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities