Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.676 ± 0.006   1.025 ± 0.01 4.484 ± 0.295  6.05 ± 0.322
#> 2  Test1 | 0.017 0.628 ± 0.004  1.105 ± 0.007  4.33 ± 0.054 4.164 ± 0.254
#> 3  Test1 | 0.026 0.602 ± 0.005   1.151 ± 0.01   4.48 ± 0.21 3.237 ± 0.129
#> 4  Test1 | 0.039 0.576 ± 0.005   1.202 ± 0.01 4.374 ± 0.177 2.856 ± 0.295
#> 5  Test1 | 0.059 0.532 ± 0.017  1.305 ± 0.041 4.461 ± 0.052 2.264 ± 0.121
#> 6  Test1 | 0.088  0.47 ± 0.003  1.476 ± 0.009  4.32 ± 0.171 1.589 ± 0.108
#> 7   Test1 | 0.13  0.392 ± 0.01  1.769 ± 0.044 4.408 ± 0.233 1.289 ± 0.151
#> 8    Test1 | 0.2 0.268 ± 0.003  2.583 ± 0.034   3.89 ± 0.27 0.844 ± 0.047
#> 9    Test1 | 0.3 0.149 ± 0.003  4.666 ± 0.107  3.07 ± 0.108 0.478 ± 0.019
#> 10  Test1 | 0.44  0.07 ± 0.003  9.945 ± 0.421 2.013 ± 0.255 0.193 ± 0.016
#> 11  Test1 | 0.67 0.035 ± 0.002 20.143 ± 1.278 1.238 ± 1.071 0.102 ± 0.001
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1      6 ± 0.323     8 ± 0.433
#> 2  4.109 ± 0.252 7.875 ± 0.177
#> 3  3.183 ± 0.129 7.833 ± 0.289
#> 4  2.802 ± 0.294   7.75 ± 0.25
#> 5   2.21 ± 0.121 7.917 ± 0.144
#> 6  1.538 ± 0.103      8 ± 0.25
#> 7  1.236 ± 0.148 8.417 ± 0.144
#> 8   0.79 ± 0.044 8.917 ± 0.289
#> 9  0.423 ± 0.019  10.5 ± 0.901
#> 10 0.142 ± 0.011    11 ± 3.041
#> 11 0.046 ± 0.002 5.833 ± 5.101
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities