Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.666 ± 0.009  1.041 ± 0.014 4.327 ± 0.123 6.465 ± 0.525
#> 2  Test1 | 0.017 0.627 ± 0.016  1.106 ± 0.028 4.345 ± 0.227 3.741 ± 0.021
#> 3  Test1 | 0.026  0.599 ± 0.01   1.158 ± 0.02 4.578 ± 0.108  3.373 ± 0.13
#> 4  Test1 | 0.039 0.585 ± 0.006  1.186 ± 0.013 4.258 ± 0.026 2.928 ± 0.406
#> 5  Test1 | 0.059 0.528 ± 0.006  1.313 ± 0.015 4.468 ± 0.266 2.202 ± 0.182
#> 6  Test1 | 0.088  0.464 ± 0.01  1.496 ± 0.033  4.428 ± 0.16 1.692 ± 0.162
#> 7   Test1 | 0.13  0.387 ± 0.01   1.79 ± 0.044 4.246 ± 0.341 1.207 ± 0.103
#> 8    Test1 | 0.2 0.276 ± 0.005   2.508 ± 0.05 3.893 ± 0.188 0.773 ± 0.023
#> 9    Test1 | 0.3 0.151 ± 0.005   4.592 ± 0.14 3.038 ± 0.309 0.447 ± 0.074
#> 10  Test1 | 0.44 0.076 ± 0.003  9.178 ± 0.412 1.941 ± 0.558 0.214 ± 0.028
#> 11  Test1 | 0.67 0.034 ± 0.003 20.719 ± 1.725 1.314 ± 0.881 0.096 ± 0.003
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.412 ± 0.521 7.833 ± 0.382
#> 2  3.689 ± 0.026 7.875 ± 0.177
#> 3  3.318 ± 0.131 8.083 ± 0.144
#> 4  2.875 ± 0.402 7.667 ± 0.144
#> 5  2.149 ± 0.179 8.083 ± 0.289
#> 6  1.638 ± 0.161 8.167 ± 0.289
#> 7  1.156 ± 0.101 8.333 ± 0.382
#> 8  0.725 ± 0.024 8.917 ± 0.577
#> 9   0.394 ± 0.07 9.833 ± 0.382
#> 10 0.161 ± 0.022   9.5 ± 3.041
#> 11 0.042 ± 0.003 6.083 ± 2.036
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities