Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.668 ± 0.008  1.038 ± 0.012 4.408 ± 0.136 6.174 ± 0.777
#> 2  Test1 | 0.017  0.61 ± 0.004  1.137 ± 0.007 4.432 ± 0.336 3.484 ± 0.305
#> 3  Test1 | 0.026 0.607 ± 0.009  1.143 ± 0.017 4.623 ± 0.079 3.253 ± 0.179
#> 4  Test1 | 0.039 0.582 ± 0.007  1.191 ± 0.013 4.348 ± 0.034 2.936 ± 0.162
#> 5  Test1 | 0.059 0.536 ± 0.005  1.293 ± 0.012 4.574 ± 0.146   2.13 ± 0.02
#> 6  Test1 | 0.088 0.466 ± 0.003  1.488 ± 0.011  4.314 ± 0.14 1.581 ± 0.183
#> 7   Test1 | 0.13 0.392 ± 0.003   1.77 ± 0.015 4.368 ± 0.154  1.249 ± 0.08
#> 8    Test1 | 0.2  0.27 ± 0.004  2.569 ± 0.038 4.063 ± 0.174 0.844 ± 0.071
#> 9    Test1 | 0.3 0.155 ± 0.006  4.489 ± 0.184 3.229 ± 0.279 0.488 ± 0.044
#> 10  Test1 | 0.44 0.071 ± 0.002  9.819 ± 0.215 1.922 ± 0.235 0.196 ± 0.016
#> 11  Test1 | 0.67 0.036 ± 0.002 19.399 ± 0.967 0.239 ± 0.418 0.094 ± 0.003
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.123 ± 0.775      8 ± 0.25
#> 2    3.433 ± 0.3     8 ± 0.354
#> 3  3.201 ± 0.177             8
#> 4  2.883 ± 0.161 7.833 ± 0.144
#> 5   2.079 ± 0.02 8.083 ± 0.144
#> 6  1.529 ± 0.178 7.917 ± 0.144
#> 7  1.198 ± 0.078 8.583 ± 0.289
#> 8   0.79 ± 0.066 9.167 ± 0.144
#> 9  0.433 ± 0.042            10
#> 10 0.144 ± 0.013  9.75 ± 2.385
#> 11  0.04 ± 0.004  1.75 ± 2.598
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities