Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.667 ± 0.007  1.039 ± 0.011 4.558 ± 0.117  6.491 ± 0.27
#> 2  Test1 | 0.017 0.631 ± 0.001  1.099 ± 0.001  4.34 ± 0.128 4.019 ± 0.425
#> 3  Test1 | 0.026 0.602 ± 0.014  1.153 ± 0.026   4.501 ± 0.1 3.313 ± 0.399
#> 4  Test1 | 0.039 0.581 ± 0.008  1.193 ± 0.017 4.405 ± 0.257 2.841 ± 0.099
#> 5  Test1 | 0.059 0.538 ± 0.012  1.289 ± 0.029 4.394 ± 0.124 2.217 ± 0.261
#> 6  Test1 | 0.088 0.468 ± 0.003  1.483 ± 0.011 4.392 ± 0.104  1.67 ± 0.111
#> 7   Test1 | 0.13 0.384 ± 0.003  1.803 ± 0.013 4.368 ± 0.238  1.161 ± 0.03
#> 8    Test1 | 0.2 0.267 ± 0.004  2.596 ± 0.044 4.037 ± 0.208 0.796 ± 0.095
#> 9    Test1 | 0.3 0.151 ± 0.004   4.584 ± 0.12  3.251 ± 0.08  0.47 ± 0.025
#> 10  Test1 | 0.44 0.071 ± 0.005  9.849 ± 0.679 2.147 ± 0.837 0.199 ± 0.015
#> 11  Test1 | 0.67 0.036 ± 0.002 19.485 ± 1.308 2.929 ± 2.992  0.09 ± 0.004
#> 12     Test1 | 1                                                         
#>               dY        t(mumax)
#> 1   6.436 ± 0.27   8.167 ± 0.144
#> 2  3.966 ± 0.423            7.75
#> 3  3.262 ± 0.396   8.083 ± 0.144
#> 4  2.789 ± 0.095   7.917 ± 0.289
#> 5  2.164 ± 0.257            7.75
#> 6  1.618 ± 0.111               8
#> 7   1.11 ± 0.028     8.5 ± 0.433
#> 8  0.743 ± 0.092   9.167 ± 0.382
#> 9  0.416 ± 0.022    9.75 ± 0.901
#> 10 0.146 ± 0.012  12.417 ± 2.428
#> 11 0.041 ± 0.003 11.167 ± 11.339
#> 12                              

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities