Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.673 ± 0.007   1.03 ± 0.011 4.432 ± 0.219 6.299 ± 0.588
#> 2  Test1 | 0.017 0.617 ± 0.007  1.123 ± 0.014 4.343 ± 0.237 3.765 ± 0.175
#> 3  Test1 | 0.026 0.605 ± 0.014  1.147 ± 0.027 4.401 ± 0.077   3.29 ± 0.43
#> 4  Test1 | 0.039  0.58 ± 0.004  1.194 ± 0.008 4.418 ± 0.213 2.912 ± 0.357
#> 5  Test1 | 0.059  0.54 ± 0.003  1.283 ± 0.006 4.443 ± 0.303 2.436 ± 0.147
#> 6  Test1 | 0.088 0.471 ± 0.003  1.471 ± 0.008 4.438 ± 0.139 1.719 ± 0.197
#> 7   Test1 | 0.13 0.388 ± 0.008  1.785 ± 0.037 4.456 ± 0.196 1.238 ± 0.171
#> 8    Test1 | 0.2 0.264 ± 0.002   2.621 ± 0.02 4.012 ± 0.288  0.75 ± 0.099
#> 9    Test1 | 0.3  0.15 ± 0.005  4.639 ± 0.158  3.07 ± 0.454 0.425 ± 0.046
#> 10  Test1 | 0.44 0.072 ± 0.005   9.66 ± 0.602 2.118 ± 0.458 0.196 ± 0.022
#> 11  Test1 | 0.67 0.038 ± 0.002 18.305 ± 1.036 4.222 ± 3.698 0.104 ± 0.003
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.247 ± 0.587      8 ± 0.25
#> 2  3.712 ± 0.176 7.625 ± 0.177
#> 3  3.237 ± 0.426 7.833 ± 0.144
#> 4  2.859 ± 0.352 7.917 ± 0.382
#> 5  2.382 ± 0.144      8 ± 0.25
#> 6  1.666 ± 0.193 8.167 ± 0.144
#> 7  1.184 ± 0.167 8.583 ± 0.144
#> 8    0.7 ± 0.094 9.083 ± 0.382
#> 9  0.374 ± 0.044 9.167 ± 0.289
#> 10 0.145 ± 0.021  10.5 ± 2.537
#> 11 0.047 ± 0.003 12.75 ± 9.972
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities