Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0  0.66 ± 0.011  1.051 ± 0.018 4.507 ± 0.213 6.437 ± 0.176
#> 2  Test1 | 0.017 0.625 ± 0.009  1.109 ± 0.016 4.546 ± 0.168 4.176 ± 0.347
#> 3  Test1 | 0.026 0.603 ± 0.006  1.149 ± 0.011  4.466 ± 0.16  3.37 ± 0.234
#> 4  Test1 | 0.039  0.573 ± 0.01    1.21 ± 0.02 4.379 ± 0.285 2.633 ± 0.353
#> 5  Test1 | 0.059 0.534 ± 0.006  1.298 ± 0.015  4.441 ± 0.25 2.326 ± 0.141
#> 6  Test1 | 0.088  0.471 ± 0.01  1.473 ± 0.032 4.393 ± 0.128 1.656 ± 0.178
#> 7   Test1 | 0.13 0.391 ± 0.008  1.775 ± 0.035 4.584 ± 0.131 1.339 ± 0.109
#> 8    Test1 | 0.2 0.265 ± 0.008  2.615 ± 0.074   3.8 ± 0.062 0.757 ± 0.078
#> 9    Test1 | 0.3 0.151 ± 0.003  4.589 ± 0.089 3.231 ± 0.085 0.441 ± 0.028
#> 10  Test1 | 0.44 0.076 ± 0.004  9.184 ± 0.483 2.293 ± 0.242 0.214 ± 0.011
#> 11  Test1 | 0.67 0.039 ± 0.004 17.865 ± 1.763  0.45 ± 0.411 0.093 ± 0.011
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.384 ± 0.179  8.083 ± 0.289
#> 2  4.122 ± 0.344      8 ± 0.354
#> 3   3.316 ± 0.23           7.75
#> 4   2.583 ± 0.35  7.833 ± 0.144
#> 5  2.272 ± 0.138  7.917 ± 0.382
#> 6  1.605 ± 0.176  8.167 ± 0.289
#> 7  1.283 ± 0.109  8.583 ± 0.382
#> 8  0.705 ± 0.072   8.25 ± 0.433
#> 9   0.39 ± 0.026  9.667 ± 0.289
#> 10 0.159 ± 0.009 10.417 ± 0.144
#> 11 0.042 ± 0.006      2 ± 1.639
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities