Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.672 ± 0.007  1.031 ± 0.011 4.229 ± 0.146  6.34 ± 0.789
#> 2  Test1 | 0.017 0.622 ± 0.008  1.115 ± 0.015 4.256 ± 0.105 3.835 ± 0.113
#> 3  Test1 | 0.026   0.6 ± 0.004  1.156 ± 0.008  4.457 ± 0.24 3.216 ± 0.239
#> 4  Test1 | 0.039 0.575 ± 0.007  1.205 ± 0.014 4.471 ± 0.236 2.818 ± 0.438
#> 5  Test1 | 0.059 0.529 ± 0.017  1.312 ± 0.042 4.643 ± 0.117 2.122 ± 0.088
#> 6  Test1 | 0.088 0.482 ± 0.002  1.438 ± 0.007 4.574 ± 0.224  1.85 ± 0.207
#> 7   Test1 | 0.13 0.389 ± 0.007  1.781 ± 0.034 4.342 ± 0.187 1.266 ± 0.063
#> 8    Test1 | 0.2 0.265 ± 0.006   2.613 ± 0.06 3.865 ± 0.201   0.8 ± 0.027
#> 9    Test1 | 0.3 0.143 ± 0.005  4.843 ± 0.157 2.902 ± 0.168 0.436 ± 0.007
#> 10  Test1 | 0.44  0.07 ± 0.001  9.944 ± 0.201 1.938 ± 0.414 0.196 ± 0.013
#> 11  Test1 | 0.67 0.035 ± 0.004 19.728 ± 1.959 0.951 ± 0.823 0.096 ± 0.005
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.289 ± 0.786          7.75
#> 2  3.783 ± 0.109          7.75
#> 3  3.166 ± 0.239     8 ± 0.433
#> 4  2.765 ± 0.434 7.917 ± 0.382
#> 5   2.07 ± 0.087 8.167 ± 0.144
#> 6  1.793 ± 0.205 8.083 ± 0.144
#> 7  1.212 ± 0.061    8.5 ± 0.25
#> 8  0.749 ± 0.029 9.333 ± 0.804
#> 9  0.381 ± 0.006 9.333 ± 1.233
#> 10 0.141 ± 0.011   8.5 ± 2.179
#> 11 0.043 ± 0.003 4.917 ± 4.544
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities