Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.672 ± 0.006   1.031 ± 0.01 4.583 ± 0.103  6.605 ± 0.57
#> 2  Test1 | 0.017  0.627 ± 0.01  1.106 ± 0.017 4.449 ± 0.223  4.02 ± 0.437
#> 3  Test1 | 0.026 0.606 ± 0.003  1.144 ± 0.005 4.592 ± 0.057 3.305 ± 0.228
#> 4  Test1 | 0.039 0.582 ± 0.007  1.191 ± 0.015 4.464 ± 0.176 2.806 ± 0.124
#> 5  Test1 | 0.059 0.524 ± 0.003  1.324 ± 0.006 4.492 ± 0.065 2.039 ± 0.256
#> 6  Test1 | 0.088 0.476 ± 0.007  1.457 ± 0.021  4.345 ± 0.32 1.681 ± 0.112
#> 7   Test1 | 0.13 0.387 ± 0.005  1.791 ± 0.023 4.187 ± 0.256 1.225 ± 0.161
#> 8    Test1 | 0.2 0.273 ± 0.013  2.545 ± 0.117 4.155 ± 0.366   0.8 ± 0.093
#> 9    Test1 | 0.3  0.15 ± 0.006  4.625 ± 0.186 2.948 ± 0.484 0.461 ± 0.022
#> 10  Test1 | 0.44  0.07 ± 0.005  9.883 ± 0.687 2.142 ± 0.664  0.19 ± 0.008
#> 11  Test1 | 0.67 0.036 ± 0.003 19.111 ± 1.506 0.513 ± 0.768 0.091 ± 0.004
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.552 ± 0.568           8.25
#> 2  3.966 ± 0.433   7.75 ± 0.354
#> 3  3.251 ± 0.227              8
#> 4  2.756 ± 0.123  7.833 ± 0.289
#> 5  1.988 ± 0.251       8 ± 0.25
#> 6   1.628 ± 0.11  7.917 ± 0.289
#> 7  1.172 ± 0.157       8 ± 0.25
#> 8  0.747 ± 0.089        9 ± 0.5
#> 9  0.408 ± 0.022    9.25 ± 1.25
#> 10 0.137 ± 0.003 10.583 ± 0.629
#> 11 0.039 ± 0.002  1.833 ± 1.942
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities