Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.673 ± 0.006   1.03 ± 0.009 4.565 ± 0.067  6.666 ± 0.45
#> 2  Test1 | 0.017  0.621 ± 0.01  1.116 ± 0.018 4.425 ± 0.259 3.844 ± 0.478
#> 3  Test1 | 0.026 0.602 ± 0.006  1.151 ± 0.012 4.514 ± 0.242 3.445 ± 0.182
#> 4  Test1 | 0.039 0.576 ± 0.012  1.204 ± 0.026 4.253 ± 0.084 2.754 ± 0.289
#> 5  Test1 | 0.059  0.528 ± 0.01  1.312 ± 0.026 4.642 ± 0.027 2.136 ± 0.065
#> 6  Test1 | 0.088 0.464 ± 0.003   1.494 ± 0.01 4.386 ± 0.231 1.498 ± 0.112
#> 7   Test1 | 0.13 0.386 ± 0.004  1.796 ± 0.016 4.342 ± 0.217 1.238 ± 0.161
#> 8    Test1 | 0.2  0.27 ± 0.004  2.564 ± 0.038  4.09 ± 0.035 0.842 ± 0.024
#> 9    Test1 | 0.3 0.154 ± 0.003  4.506 ± 0.102 3.108 ± 0.058 0.467 ± 0.034
#> 10  Test1 | 0.44 0.071 ± 0.002  9.814 ± 0.331 2.134 ± 0.349 0.194 ± 0.006
#> 11  Test1 | 0.67 0.034 ± 0.001 20.421 ± 0.634   3.2 ± 3.056 0.094 ± 0.009
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.613 ± 0.449           8.25
#> 2  3.794 ± 0.475  7.875 ± 0.177
#> 3  3.393 ± 0.182       8 ± 0.25
#> 4    2.7 ± 0.287  7.667 ± 0.144
#> 5  2.083 ± 0.062  8.167 ± 0.144
#> 6  1.447 ± 0.107  7.917 ± 0.289
#> 7  1.186 ± 0.157  8.417 ± 0.144
#> 8  0.788 ± 0.021    9.25 ± 0.25
#> 9  0.412 ± 0.033  9.583 ± 0.144
#> 10 0.141 ± 0.004 11.667 ± 0.804
#> 11 0.041 ± 0.007     11 ± 7.988
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities