Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.662 ± 0.004  1.047 ± 0.006 4.406 ± 0.143 6.267 ± 0.536
#> 2  Test1 | 0.017 0.617 ± 0.005  1.124 ± 0.009 4.254 ± 0.064 3.995 ± 0.191
#> 3  Test1 | 0.026   0.6 ± 0.008  1.156 ± 0.016 4.227 ± 0.118 3.089 ± 0.206
#> 4  Test1 | 0.039 0.571 ± 0.006  1.214 ± 0.014  4.315 ± 0.03 2.778 ± 0.178
#> 5  Test1 | 0.059 0.527 ± 0.007  1.315 ± 0.016 4.566 ± 0.143 2.105 ± 0.233
#> 6  Test1 | 0.088 0.474 ± 0.005  1.463 ± 0.015  4.605 ± 0.07 1.609 ± 0.132
#> 7   Test1 | 0.13   0.39 ± 0.01  1.778 ± 0.043 4.437 ± 0.179 1.262 ± 0.119
#> 8    Test1 | 0.2 0.275 ± 0.004   2.52 ± 0.037  4.11 ± 0.143 0.838 ± 0.053
#> 9    Test1 | 0.3 0.152 ± 0.004  4.576 ± 0.122 3.245 ± 0.031 0.445 ± 0.056
#> 10  Test1 | 0.44  0.07 ± 0.003   9.846 ± 0.43 2.181 ± 0.508 0.187 ± 0.022
#> 11  Test1 | 0.67 0.036 ± 0.004 19.586 ± 2.578 0.996 ± 1.037 0.095 ± 0.006
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.214 ± 0.532       8 ± 0.25
#> 2   3.94 ± 0.189  7.625 ± 0.177
#> 3  3.038 ± 0.204  7.583 ± 0.144
#> 4  2.724 ± 0.178           7.75
#> 5   2.054 ± 0.23              8
#> 6  1.559 ± 0.129  8.333 ± 0.144
#> 7  1.208 ± 0.119  8.333 ± 0.144
#> 8   0.786 ± 0.05  9.083 ± 0.382
#> 9  0.393 ± 0.053      10 ± 0.25
#> 10 0.137 ± 0.017 12.083 ± 3.215
#> 11  0.04 ± 0.003  4.333 ± 3.055
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities