Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0  0.67 ± 0.001  1.035 ± 0.002 4.426 ± 0.234 6.673 ± 0.607
#> 2  Test1 | 0.017 0.616 ± 0.002  1.126 ± 0.004 4.425 ± 0.219 3.561 ± 0.404
#> 3  Test1 | 0.026 0.596 ± 0.005   1.163 ± 0.01 4.584 ± 0.023 3.144 ± 0.177
#> 4  Test1 | 0.039 0.567 ± 0.011  1.223 ± 0.023 4.475 ± 0.299 2.621 ± 0.081
#> 5  Test1 | 0.059 0.535 ± 0.009  1.295 ± 0.021 4.301 ± 0.127 2.196 ± 0.152
#> 6  Test1 | 0.088  0.478 ± 0.01   1.452 ± 0.03 4.527 ± 0.197 1.655 ± 0.205
#> 7   Test1 | 0.13 0.384 ± 0.002  1.804 ± 0.009  4.36 ± 0.295  1.169 ± 0.13
#> 8    Test1 | 0.2 0.263 ± 0.004  2.631 ± 0.038  4.13 ± 0.143 0.808 ± 0.007
#> 9    Test1 | 0.3 0.152 ± 0.002  4.571 ± 0.058 3.187 ± 0.197 0.452 ± 0.007
#> 10  Test1 | 0.44 0.074 ± 0.002  9.348 ± 0.235 2.209 ± 0.337 0.209 ± 0.014
#> 11  Test1 | 0.67 0.038 ± 0.003 18.468 ± 1.204 2.082 ± 1.903 0.097 ± 0.007
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.618 ± 0.604       8 ± 0.25
#> 2     3.51 ± 0.4   7.75 ± 0.354
#> 3  3.093 ± 0.173              8
#> 4   2.57 ± 0.077  7.917 ± 0.144
#> 5  2.144 ± 0.149  7.833 ± 0.144
#> 6  1.604 ± 0.203  8.083 ± 0.144
#> 7  1.118 ± 0.126  8.417 ± 0.289
#> 8  0.752 ± 0.008  8.583 ± 0.144
#> 9  0.399 ± 0.009  9.083 ± 1.041
#> 10 0.156 ± 0.011 11.167 ± 0.382
#> 11 0.044 ± 0.004   8.083 ± 7.32
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities