Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.673 ± 0.012  1.031 ± 0.019 4.368 ± 0.213 6.772 ± 0.563
#> 2  Test1 | 0.017 0.618 ± 0.011   1.121 ± 0.02  4.62 ± 0.136 3.537 ± 0.312
#> 3  Test1 | 0.026    0.6 ± 0.01  1.156 ± 0.019 4.627 ± 0.097  3.144 ± 0.32
#> 4  Test1 | 0.039 0.577 ± 0.016  1.202 ± 0.033 4.517 ± 0.177 2.929 ± 0.222
#> 5  Test1 | 0.059 0.537 ± 0.001  1.291 ± 0.002 4.467 ± 0.133 2.331 ± 0.268
#> 6  Test1 | 0.088 0.465 ± 0.006   1.49 ± 0.018 4.462 ± 0.274 1.673 ± 0.207
#> 7   Test1 | 0.13  0.391 ± 0.01  1.772 ± 0.043  4.327 ± 0.12 1.143 ± 0.081
#> 8    Test1 | 0.2 0.267 ± 0.001    2.6 ± 0.013 4.051 ± 0.043 0.783 ± 0.014
#> 9    Test1 | 0.3 0.153 ± 0.003   4.528 ± 0.08 3.132 ± 0.191 0.435 ± 0.054
#> 10  Test1 | 0.44 0.071 ± 0.002  9.743 ± 0.268 1.591 ± 0.264 0.194 ± 0.008
#> 11  Test1 | 0.67 0.034 ± 0.005 20.403 ± 2.608 1.649 ± 0.357 0.092 ± 0.009
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)            AUC
#> 1  6.718 ± 0.561      8 ± 0.25 77.209 ± 1.605
#> 2  3.486 ± 0.307             8 67.605 ± 0.735
#> 3  3.091 ± 0.316 8.083 ± 0.144 64.825 ± 1.059
#> 4  2.874 ± 0.221 7.833 ± 0.144 63.022 ± 0.667
#> 5  2.278 ± 0.264      8 ± 0.25 60.316 ± 0.351
#> 6   1.62 ± 0.204 8.167 ± 0.382 54.615 ± 0.708
#> 7  1.095 ± 0.081    8.5 ± 0.25 49.167 ± 0.847
#> 8  0.729 ± 0.012 9.083 ± 0.629   39.55 ± 0.88
#> 9  0.384 ± 0.051      9 ± 1.09 29.021 ± 0.899
#> 10 0.144 ± 0.006 9.417 ± 1.756 17.344 ± 0.222
#> 11 0.039 ± 0.006         6 ± 1  6.991 ± 0.511
#> 12                                           

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities