Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.666 ± 0.007   1.04 ± 0.011 4.166 ± 0.099 6.488 ± 0.175
#> 2  Test1 | 0.017 0.624 ± 0.014  1.111 ± 0.024 4.379 ± 0.205  3.803 ± 0.73
#> 3  Test1 | 0.026 0.594 ± 0.002  1.166 ± 0.004 4.375 ± 0.102 3.331 ± 0.174
#> 4  Test1 | 0.039  0.574 ± 0.01   1.208 ± 0.02 4.369 ± 0.091 2.931 ± 0.278
#> 5  Test1 | 0.059 0.523 ± 0.005  1.325 ± 0.012 4.414 ± 0.244 2.032 ± 0.121
#> 6  Test1 | 0.088  0.47 ± 0.008  1.474 ± 0.024 4.522 ± 0.197  1.675 ± 0.21
#> 7   Test1 | 0.13 0.385 ± 0.007  1.802 ± 0.033 4.294 ± 0.246 1.319 ± 0.049
#> 8    Test1 | 0.2 0.266 ± 0.006  2.605 ± 0.064 3.865 ± 0.239 0.763 ± 0.062
#> 9    Test1 | 0.3 0.146 ± 0.006  4.738 ± 0.179 3.098 ± 0.118 0.422 ± 0.044
#> 10  Test1 | 0.44 0.071 ± 0.002  9.735 ± 0.239 2.172 ± 0.398 0.196 ± 0.006
#> 11  Test1 | 0.67 0.035 ± 0.003 19.867 ± 1.585  1.26 ± 1.392 0.097 ± 0.002
#> 12     Test1 | 1                                                         
#>               dY      t(mumax)
#> 1  6.436 ± 0.172   7.75 ± 0.25
#> 2  3.751 ± 0.724  7.75 ± 0.354
#> 3  3.277 ± 0.171 7.917 ± 0.144
#> 4  2.878 ± 0.276 7.833 ± 0.144
#> 5  1.981 ± 0.118 7.917 ± 0.289
#> 6  1.621 ± 0.206      8 ± 0.25
#> 7  1.263 ± 0.047  8.333 ± 0.52
#> 8  0.713 ± 0.062 8.417 ± 0.629
#> 9  0.371 ± 0.043    10 ± 0.866
#> 10 0.142 ± 0.007 10.25 ± 1.803
#> 11 0.043 ± 0.001 5.083 ± 4.446
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities