Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax            tD       lagtime         Y_max
#> 1      Test1 | 0 0.676 ± 0.003 1.025 ± 0.004 4.453 ± 0.243 6.805 ± 0.792
#> 2  Test1 | 0.017  0.63 ± 0.001   1.1 ± 0.003 4.489 ± 0.009 4.021 ± 0.188
#> 3  Test1 | 0.026  0.599 ± 0.01 1.158 ± 0.019 4.363 ± 0.121 3.199 ± 0.256
#> 4  Test1 | 0.039 0.575 ± 0.009 1.205 ± 0.019 4.393 ± 0.177 2.779 ± 0.163
#> 5  Test1 | 0.059 0.533 ± 0.004   1.3 ± 0.009 4.447 ± 0.202 2.031 ± 0.081
#> 6  Test1 | 0.088 0.475 ± 0.006 1.458 ± 0.019 4.478 ± 0.147 1.665 ± 0.171
#> 7   Test1 | 0.13  0.39 ± 0.005 1.779 ± 0.024  4.399 ± 0.27 1.245 ± 0.094
#> 8    Test1 | 0.2 0.269 ± 0.002   2.58 ± 0.02 4.173 ± 0.105 0.785 ± 0.018
#> 9    Test1 | 0.3  0.15 ± 0.004 4.632 ± 0.133 2.997 ± 0.463 0.472 ± 0.007
#> 10  Test1 | 0.44 0.071 ± 0.005 9.819 ± 0.665 1.824 ± 0.655 0.197 ± 0.007
#> 11  Test1 | 0.67 0.032 ± 0.004 21.87 ± 2.492 0.599 ± 0.383 0.091 ± 0.011
#> 12     Test1 | 1                                                        
#>               dY      t(mumax)
#> 1  6.752 ± 0.787 7.917 ± 0.382
#> 2  3.969 ± 0.184             8
#> 3  3.147 ± 0.255 7.667 ± 0.144
#> 4  2.726 ± 0.162 7.917 ± 0.144
#> 5  1.983 ± 0.081 7.917 ± 0.289
#> 6  1.615 ± 0.169 8.167 ± 0.144
#> 7  1.194 ± 0.093 8.583 ± 0.144
#> 8  0.732 ± 0.016  8.75 ± 0.433
#> 9  0.419 ± 0.006 9.833 ± 0.946
#> 10 0.144 ± 0.008   8.75 ± 2.25
#> 11 0.039 ± 0.009  3.75 ± 2.598
#> 12                            

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities