Generate a grouped results table for spline fits with average and standard deviations

Generate a grouped results table for spline fits with average and standard deviations

Usage

table_group_growth_spline(gcTable, html = FALSE)

Arguments

gcTable

An object of class gcTable

html

(Logical) Should column headers contain html formatting?

Value

A data frame with grouped spline fit results. Empty cells indicate that no reliable fit could be determined.

Examples

# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

table_group_growth_spline(res$gcFit$gcTable)
#>     Sample|Conc.         mumax             tD       lagtime         Y_max
#> 1      Test1 | 0 0.674 ± 0.004  1.029 ± 0.006 4.367 ± 0.118 6.558 ± 0.295
#> 2  Test1 | 0.017         0.636           1.09 4.398 ± 0.106  4.035 ± 0.55
#> 3  Test1 | 0.026 0.605 ± 0.011  1.146 ± 0.021 4.259 ± 0.039 3.503 ± 0.247
#> 4  Test1 | 0.039 0.578 ± 0.001  1.199 ± 0.002 4.628 ± 0.007 2.578 ± 0.148
#> 5  Test1 | 0.059 0.528 ± 0.007  1.312 ± 0.018 4.281 ± 0.089 2.074 ± 0.314
#> 6  Test1 | 0.088 0.475 ± 0.003   1.461 ± 0.01 4.407 ± 0.055 1.826 ± 0.083
#> 7   Test1 | 0.13 0.387 ± 0.005  1.792 ± 0.024 4.191 ± 0.118 1.278 ± 0.114
#> 8    Test1 | 0.2 0.265 ± 0.009  2.613 ± 0.087 3.846 ± 0.107 0.752 ± 0.036
#> 9    Test1 | 0.3 0.147 ± 0.004  4.719 ± 0.138  3.054 ± 0.12 0.422 ± 0.024
#> 10  Test1 | 0.44 0.072 ± 0.005   9.633 ± 0.65 2.409 ± 0.401 0.186 ± 0.014
#> 11  Test1 | 0.67 0.037 ± 0.004 18.931 ± 1.909 2.228 ± 0.626 0.093 ± 0.013
#> 12     Test1 | 1                                                         
#>               dY       t(mumax)
#> 1  6.505 ± 0.291  7.917 ± 0.144
#> 2  3.982 ± 0.545           7.75
#> 3  3.448 ± 0.244           7.75
#> 4  2.527 ± 0.146  7.917 ± 0.144
#> 5   2.024 ± 0.31  7.833 ± 0.144
#> 6  1.771 ± 0.085  8.083 ± 0.144
#> 7  1.224 ± 0.111  8.083 ± 0.289
#> 8  0.701 ± 0.037  8.917 ± 0.382
#> 9  0.372 ± 0.022 10.333 ± 1.507
#> 10 0.136 ± 0.012  10.75 ± 2.046
#> 11 0.042 ± 0.007   8.667 ± 4.26
#> 12                             

# with HTML formatting
DT::datatable(table_group_growth_spline(res$gcFit$gcTable, html = TRUE),
              escape = FALSE) # Do not escape HTML entities